Tor Arne Vestbø  <tavestbo@trolltech.com>
Fix QWebView::loadFinished isn't always emitted

Replaced loadDone() with loadFinished(bool) and moved the signals for progress
tracking (start, progres, and finish) to the page instead of the frame. This
ensures that we emit loadFinished even when a subframe started the actual load.

This causes a few regressions in the layout tests that we for now accept for
the sake of the correct API. Layout tests we can fix any time though, including
patch release, the API however we can't change anymore in patch releases.


git-svn-id: svn://svn.chromium.org/blink/trunk@32643 bbb929c8-8fbe-4397-9dbb-9b2b20218538