<rdar://problem/5956612> ibtool crashes during build
SCDynamicStoreCreate will return null if it cannot create a dynamic store object.  We need
to check for failure rather than continuing with a null store as this results in us
passing a null CFRunLoopSourceRef to CFRunLoopAddSource, which promptly crashes.

Reviewed by Tim Hatcher.

* platform/network/mac/NetworkStateNotifierMac.cpp:
(WebCore::NetworkStateNotifier::NetworkStateNotifier): Null check SCDynamicStoreCreate and
SCDynamicStoreCreateRunLoopSource as it is possible for them to fail.


git-svn-id: svn://svn.chromium.org/blink/trunk@34050 bbb929c8-8fbe-4397-9dbb-9b2b20218538