Bug 21054: Construction of certain DOM objects is heavily regressed by r36675 <https://bugs.webkit.org/show_bug.cgi?id=21054>
Reviewed by Maciej Stachowiak

This performance regression is actually just a symptom of a correctness
bug.  The constructor objects for a number of properties that have security
checks on access were returning new objects each time.  The most obvious
symptom of this bug is that window.Image != window.Image, etc.

The solution to this is to make sure we cache these constructors
in the same way as all the other DOM constructors.  To achieve this
without causing any refcount cycles it is necessary to replace the
refcounted document pointer in the Image, MessageChannel, Option,
XMLHttpRequest, and Audio constructor objects with a reference to
the document's JS wrapper.

Tests: fast/dom/constructors-cached-navigate.html
       fast/dom/constructors-cached.html


git-svn-id: svn://svn.chromium.org/blink/trunk@36954 bbb929c8-8fbe-4397-9dbb-9b2b20218538