<rdar://problem/6548446> TCMalloc_SystemRelease should use madvise rather than re-mmaping span of pages
* wtf/FastMalloc.cpp:
(WTF::mergeDecommittedStates): If either of the spans has been released to the system, release the other
span as well so that the flag in the merged span is accurate.
* wtf/Platform.h:
* wtf/TCSystemAlloc.cpp: Track decommitted spans when using MADV_FREE_REUSABLE / MADV_FREE_REUSE.
(TCMalloc_SystemRelease): Use madvise with MADV_FREE_REUSABLE when it is available.
(TCMalloc_SystemCommit): Use madvise with MADV_FREE_REUSE when it is available.
* wtf/TCSystemAlloc.h:

git-svn-id: svn://svn.chromium.org/blink/trunk@41660 bbb929c8-8fbe-4397-9dbb-9b2b20218538