Bug 25202: Improve performance of repeated callbacks into the VM
Reviewed by Cameron Zwarich

Add the concept of a CachedCall to native code for use in Array
prototype and similar functions where a single callback function
is called repeatedly with the same number of arguments.

Used Array.prototype.filter as the test function and got a 50% win
over a naive non-caching specialised version.  This makes the native
implementation of Array.prototype.filter faster than the JS one once
more.


git-svn-id: svn://svn.chromium.org/blink/trunk@42537 bbb929c8-8fbe-4397-9dbb-9b2b20218538