Add a limited literal parser for eval to handle object and array literals fired at eval
Reviewed by Gavin Barraclough and Darin Adler.

This is a simplified parser and lexer that we can throw at strings passed to eval
in case a site is using eval to parse JSON (eg. json2.js).  The lexer is intentionally
limited (in effect it's whitelisting a limited "common" subset of the JSON grammar)
as this decreases the likelihood of us wating time attempting to parse any significant
amount of non-JSON content.


git-svn-id: svn://svn.chromium.org/blink/trunk@43424 bbb929c8-8fbe-4397-9dbb-9b2b20218538