Fix <https://bugs.webkit.org/show_bug.cgi?id=25917> / <rdar://problem/6910066>. Bug 25917: REGRESSION (r43559?): Javascript debugger crashes when pausing page
Reviewed by Oliver Hunt.

The debugger currently retrieves the arguments object from an activation rather than pulling
it from a call frame.  This is unreliable to due to the recent optimization to lazily create
the arguments object.  In the long-term it should stop doing that (<rdar://problem/6911886>),
but for now we force eager creation of the arguments object when debugging.

* bytecompiler/BytecodeGenerator.cpp:
(JSC::BytecodeGenerator::BytecodeGenerator):

git-svn-id: svn://svn.chromium.org/blink/trunk@43976 bbb929c8-8fbe-4397-9dbb-9b2b20218538