Allow to explicitly choose a preferred plugin for a mimetype.
https://bugs.webkit.org/show_bug.cgi?id=27651

When the preferred plugin is set for a specific MIME type, it will be
always picked up, regardless of its version, quirks etc.

Client applications may want to use that API to resolve mimetype
ambiguity in a custom way, rather than in the default way that is
currently used in WebKit.

git-svn-id: svn://svn.chromium.org/blink/trunk@46586 bbb929c8-8fbe-4397-9dbb-9b2b20218538