Fix handling of QNetworkReply errors.
Patch by Kenneth Rohde Christiansen <kenneth@webkit.org> on 2009-09-29
Reviewed by Simon Hausmann.

In the QNetworkReplyHandler::finish() a response was sent even when
the reply contained an error. This resulted in a sendResponseIfNeeded()
calling didReceiveResponse on the client, leading to the destruction
of the m_resourceHandle, discontinuing further processing in finish(),
and thus not calling didFail on the client.

Instead it continued as everything went fine, and
FrameLoaderClientQt::dispatchDecidePolicyForMIMEType() changed the
policy to PolicyDownload due to not being able to show the non existing
MIMEType. As the download also obviously fails, it ended up with a
policy change error.

* platform/network/qt/QNetworkReplyHandler.cpp:
(WebCore::QNetworkReplyHandler::finish):

git-svn-id: svn://svn.chromium.org/blink/trunk@48900 bbb929c8-8fbe-4397-9dbb-9b2b20218538