Windowed Plugins: Don't crash when client is 0.
Client is 0 when we use QWebPage without a QWebView or QGraphicsWebView.
In addition, setFrameRect()/updatePluginWidget() is called even if the
plugin was not succesfully loaded. updatePluginWidget() updates the
window rect which is, in theory, useful to draw something that indicates
that we didn't load successfully.

So, a status check is added to setNPWindowIfNeeded.

git-svn-id: svn://svn.chromium.org/blink/trunk@49771 bbb929c8-8fbe-4397-9dbb-9b2b20218538