Unreviewed documentation fix from David Boddie (Qt Doc Team)
Patch by Kenneth Rohde Christiansen <kenneth@webkit.org> on 2009-10-26
Ensure that qdoc will always see the RenderHints property.

The property was only defined in the header file if the Q_OS_SYMBIAN
symbol was not defined, resulting in the property not showing up
in the Qt documentation just because one platform doesn't support it.

A follow up commit will improve the documentation for the property
and note that it is not supported on the Symbiam platform.

* Api/qwebview.h:

git-svn-id: svn://svn.chromium.org/blink/trunk@50096 bbb929c8-8fbe-4397-9dbb-9b2b20218538