WebCore: Implement window.navigator.registerProtocolHandler in webkit, https://bugs.webkit.org/b/29651
Patch by Bradley Green <brg@chromium.org> on 2009-11-03
Reviewed by Dmitry Titov.

Also implemented its sister API window.navigator.registerContentHandler.

These methods are as described in the HTML5 specification which can be
found here,
http://dev.w3.org/html5/spec/Overview.html#dom-navigator-registerprotocolhandler
http://dev.w3.org/html5/spec/Overview.html#dom-navigator-registercontenthandler

As specified in the document, the behavior of the browser is determined
by the current registered handler.  The state of a registered handler
can change at any time, with the user clearing a registered handler,
registering a different page as handler, or deferring the hander to the
OS.  If webkit was to track the state of the currently registered
handlers, it would need more APIs and complexity to keep in sync with
user actions reported to webkit from the UA.  For simplicity, the state
of protocol handlers should be kept isolated from webkit and webkit only
notifies the UA that a page has made the call.  The UA is then
responsible for correctly handling the registerProtocolHandler call and
the redirects which result from registration.

We do however follow the specification in insuring that the reserved
schemes and mimeTypes are not passed to the UA as custom handler
registration tests.  We also insure that the "%s" token is present as
required by the specification.

Updated test expectations for window.clientInformation and navigator
objects.

Tests: fast/dom/registerContentHandler.html
       fast/dom/registerProtocolHandler.html

* page/Chrome.cpp:
(WebCore::Chrome::registerProtocolHandler):
(WebCore::Chrome::registerContentHandler):
* page/Chrome.h:
* page/ChromeClient.h:
(WebCore::ChromeClient::registerProtocolHandler):
(WebCore::ChromeClient::registerContentHandler):
* page/Navigator.cpp:
(WebCore::verifyCustomHandlerURL):
(WebCore::verifyProtocolHandlerScheme):
(WebCore::Navigator::registerProtocolHandler):
(WebCore::verifyProtocolHandlerMimeType):
(WebCore::Navigator::registerContentHandler):
* page/Navigator.h:
* page/Navigator.idl:

LayoutTests: Updating test expectations for implementation of
window.navigator.registerProtocolHandler and
window.navigator.registerContentHandler.

Patch by Bradley Green <brg@chromium.org> on 2009-11-03
Reviewed by Dmitry Titov.

There is a default implimentation for these APIs, hence all platforms
with the navigator-detached-no-crash-expect and
window-properties-expected tests needed to be updated.  While
isolatedWorlds has a window-properties-expected test, it has no window
object and hence does not need to be updaetd.

We add two fast/dom tests to insure that the proper exceptions are
thrown when bad syntax or security violations are made by calling the
methods.

* fast/dom/Window/window-properties-expected.txt:
* fast/dom/navigator-detached-no-crash-expected.txt:
* fast/dom/navigator-detached-no-crash.html:
* fast/dom/registerContentHandler-expected.txt: Added.
* fast/dom/registerContentHandler.html: Added.
* fast/dom/registerProtocolHandler-expected.txt: Added.
* fast/dom/registerProtocolHandler.html: Added.
* platform/gtk/fast/dom/Window/window-properties-expected.txt:
* platform/gtk/fast/dom/navigator-detached-no-crash-expected.txt:
* platform/mac-leopard/fast/dom/navigator-detached-no-crash-expected.txt:

git-svn-id: svn://svn.chromium.org/blink/trunk@50477 bbb929c8-8fbe-4397-9dbb-9b2b20218538