[Gtk+] Avoid making a copy of the HTTPHeaderMap.
ResourceRequest::httpHeaderFields() is returning a const
reference but we are storing it as a simple value. This
means that we run either the copy constructor or the assignment
operator and make a copy of the RefPtrs.

This extra copy can be avoided easily by storing
const reference.

* platform/network/soup/ResourceRequestSoup.cpp:
(WebCore::ResourceRequest::toSoupMessage):
* platform/network/soup/ResourceResponseSoup.cpp:
(WebCore::ResourceResponseSoup::toSoupMessage):

git-svn-id: svn://svn.chromium.org/blink/trunk@56148 bbb929c8-8fbe-4397-9dbb-9b2b20218538