Autogenerate WebPage's message-receiving code and message types
This patch encompasses several changes that allow message types and
receiving code to be generated by a script, and add some type-safety
as a bonus. Messages are now represented by structs instead of an ID +
ArgumentCoder. The struct contains the arguments and the ID together,
and has a constructor that enforces the use of correct types.
Correspondingly, a new overload of WebProcessProxy::send that takes a
message struct (instead of a message ID and separate arguments) has
been added. Eventually all callers should use this overload and the
old one can be removed.

This patch only touches WebPage's messages. We should transition other
message receivers over to this new system eventually.

Fixes <http://webkit.org/b/43636> <rdar://problem/8282462> Add a
type-safe IPC mechanism to WebKit2

Reviewed by Anders Carlsson.

* DerivedSources.make: Added. Calls generate-message-receiver.py and
generate-messages-header.py for each message receiver it knows about
(just WebPage for now).

* Platform/CoreIPC/Arguments.h: Added First/Second/ThirdArgumentType
typedefs for use in handleMessage.

* Platform/CoreIPC/HandleMessage.h: Added.
(CoreIPC::handleMessage): This overloaded function template decodes
arguments and passes them along to the specified function.

* Scripts/generate-message-receiver.py: Added.
* Scripts/generate-messages-header.py: Added.
These scripts just wrap functionality in messages.py.

* Scripts/webkit2/__init__.py: Added. This just exists so that Python
will treat this directory as a package.

* Scripts/webkit2/messages.py: Added. Contains the code to parse
messages files and generate .cpp/.h files from them.
(MessageReceiver.__init__): This class represents a single receiver of
messages.
(MessageReceiver.iterparameters): Returns a generator that can be used
to iterate over all the parameters of all the messages of this
receiver.
(MessageReceiver.parse): Reads a messages file from a file-like object
and parses it into a MessageReceiver object.
(Message.__init__): This class represents a single message.
(Message.id): Returns the ID name for this message.
(Parameter.__init__): This class represents a single parameter for a
message.
(messages_header_filename): Returns the name of the header that
defines the messages for a given receiver.
(surround_in_condition): Surrounds the given string in #if/#endif if
there is an associated condition.
(messages_to_kind_enum): Returns a string that defines the Kind enum
for these messages.
(function_parameter_type): Returns the type that should be used when
passing a value of the given type as a parameter to a function.
(base_class): Returns the base class for a message struct.
(message_to_struct_declaration): Returns a string that declares the
struct for this message.
(forward_declarations_for_namespace): Returns a string that contains
forward-declarations for a set of types in a given namespace.
(forward_declarations): Returns a string that contains all the
forward-declarations needed in order to declare all the message
structs for this receiver.
(generate_header_file): Returns a string containing the messages
header file for this receiver.
(handler_function): Returns the name of the function that handles a
given message for a given receiver.
(case_statement): Returns a string containing a case statement for
handling the given message.
(header_for_type): Returns the header needed to define a given type,
enclosed in quotes or angle brackets as needed.
(generate_message_handler): Returns a string containing the contents
of a .cpp file that defines a didReceive*Message function.

* Scripts/webkit2/messages_unittest.py: Added. Contains tests for
messages.py.

* Shared/CoreIPCSupport/WebPageMessageKinds.h: Removed. This has been
replaced by a generated WebPageMessages.h header.

* UIProcess/WebEditCommandProxy.cpp:
* UIProcess/WebPageProxy.cpp:
Updated to use the new message structs and WebProcessProxy::send
overload.

* UIProcess/WebProcessProxy.h:
(WebKit::WebProcessProxy::send): Added this new overload that takes a
message struct.

* WebKit2.xcodeproj/project.pbxproj: Added a Derived Sources shell
script target that invokes DerivedSources.make. Added "Derived
Sources" and "Scripts" groups that contain the various new files.

* WebProcess/WebPage/WebPage.cpp:
(WebKit::WebPage::didReceivePolicyDecision):
(WebKit::WebPage::getSourceForFrame):
Changed these functions to take the raw IPC types and do the necessary
translation themselves. This keeps the generated code from need to
know how to perform the translation.

(WebKit::WebPage::didReceiveMessage): Replaced handling of WebPage
messages with a call to didReceiveWebPageMessage, whose implementation
is generated by the scripts.

* WebProcess/WebPage/WebPage.h: Added didReceiveWebPageMessage.

* WebProcess/WebPage/WebPage.messages.in: Added. This file declares
all of the messages that WebPage receives, roughly grouped by
functionality.

* win/WebKit2.vcproj: Added "Derived Sources" and "Scripts" filters
that contain the various new files. Let VS resort some other files.

* win/WebKit2Common.vsprops: Added
$(WebKitOutputDir)\obj\$(ProjectName)\DerivedSources to the include
path so that the generated messages header can be found.

* win/WebKit2.make: Copy the generated source files to $(DSTROOT).

* win/WebKit2Generated.make: Added a call to build-generated-files.sh.

* win/WebKit2Generated.vcproj: Added build-generated-files.sh.

* win/build-generated-files.sh: Added. Invokes DerivedSources.make.

git-svn-id: svn://svn.chromium.org/blink/trunk@68079 bbb929c8-8fbe-4397-9dbb-9b2b20218538