Add ability for WK2 to set domain relaxation forbidden for a URL scheme. https://bugs.webkit.org/show_bug.cgi?id=47562
Reviewed by Jon Honeycutt.

Add the schemes for which domain relaxation is forbidden to the
WebProcessCreationParameters.
* Shared/WebProcessCreationParameters.cpp:
(WebKit::WebProcessCreationParameters::encode):
(WebKit::WebProcessCreationParameters::decode):
* Shared/WebProcessCreationParameters.h:

* UIProcess/API/C/WKContext.cpp:
(WKContextSetDomainRelaxationForbiddenForURLScheme):
* UIProcess/API/C/WKContextPrivate.h:

* UIProcess/WebContext.cpp:
(WebKit::WebContext::ensureWebProcess):
Copy over the schemes for which domain relaxation is forbidden to
WebProcessCreationParameters.
(WebKit::WebContext::setDomainRelaxationForbiddenForURLScheme):
Only send a message to the WebProcess if it is valid.
* UIProcess/WebContext.h:

* WebProcess/WebProcess.cpp:
(WebKit::WebProcess::initializeWebProcess):
Set domain relaxation forbidden for the schemes in
WebProcesCreationParameters.urlSchemesForWhichDomainRelaxationIsForbidden.
(WebKit::WebProcess::setDomainRelaxationForbiddenForURLScheme):
* WebProcess/WebProcess.h:

* WebProcess/WebProcess.messages.in:
Add the SetDomainRelaxationFobiddenForURLScheme message.



git-svn-id: svn://svn.chromium.org/blink/trunk@69661 bbb929c8-8fbe-4397-9dbb-9b2b20218538