[Qt] Improve painting performance on 16-bit displays
Reviewed by Kenneth Christiansen.

When the default pixmap depth is 16-bit, then use RGB16 instead of
RGB32 for the chunks, to avoid unnecessary conversions and reduce
chunk size. The conversions happen as images as QPixmaps have 16-bit
depth and the tiles on the ui process are also backed by QPixmaps.

* Shared/qt/UpdateChunk.cpp:
(WebKit::UpdateChunk::size):
(WebKit::UpdateChunk::createImage):
* Shared/qt/UpdateChunk.h:

git-svn-id: svn://svn.chromium.org/blink/trunk@72185 bbb929c8-8fbe-4397-9dbb-9b2b20218538