https://bugs.webkit.org/show_bug.cgi?id=51328
Reviewed by Simon Fraser.

WebCore: 

Printing on a page with body { height: 100% } clips out the overflow contents.  This happened
because the computePageRects function was never correct.  It was using the layer's width and height
for the document's width and height, and the layer no longer includes overflow in its width and
height.  It was also incorrectly assuming the first page would begin at (0,0), and this is an invalid
assumption in RTL or vertical text environments.

Added printing/page-count-percentage-height.html

* page/PrintContext.cpp:
(WebCore::PrintContext::computePageRects):
(WebCore::PrintContext::computePageRectsWithPageSizeInternal):

LayoutTests: 

Add layout test for printing clipping bug.

* printing/page-count-percentage-height-expected.txt: Added.
* printing/page-count-percentage-height.html: Added.
* printing/script-tests/page-count-percentage-height.js: Added.
(test):



git-svn-id: svn://svn.chromium.org/blink/trunk@74908 bbb929c8-8fbe-4397-9dbb-9b2b20218538