Ensure $testResultsDirectory is an absolute path before setting up Windows crash log saving
Windows 7 (and maybe Vista?) launches the post-mortem debugger with a working directory that
is different from old-run-webkit-test's. Because of this, the path to the crash log file
(which is relative to $testResultsDirectory) must be an absolute path. See the bug for more
details.

Fixes <http://webkit.org/b/55318> Crash logs aren't saved on Windows 7 when a relative
--results-directory path is used (like on the test slaves)

Reviewed by Dan Bernstein.

* Scripts/old-run-webkit-tests: Moved the call to setUpWindowsCrashLogSaving after we've
made $testResultsDirectory an absolute path.

git-svn-id: svn://svn.chromium.org/blink/trunk@79819 bbb929c8-8fbe-4397-9dbb-9b2b20218538