Give up if a crash log for the web process is taking too long to be saved on Windows
This should keep the bots from getting stuck while waiting for a crash log, if writing a
crash log ever hangs or the UI process for some reason doesn't notice it's finished.

Fixes <http://webkit.org/b/55499> WebKitTestRunner can hang forever waiting for a crash log
to be saved for the web process

Reviewed by Steve Falkenburg.

* WebKitTestRunner/win/TestControllerWin.cpp:
(WTR::runRunLoopUntil): Moved code to run the run loop here here from platformRunUntil.
Generalized the code slightly to handle the optional object to wait on.
(WTR::TestController::platformRunUntil): Use the new runRunLoopUntil function, and also use
it (with a timeout) when waiting for a crash log for the web process to be saved. This will
prevent us from waiting forever.

git-svn-id: svn://svn.chromium.org/blink/trunk@80033 bbb929c8-8fbe-4397-9dbb-9b2b20218538