simplify gtest display now that we strip modifiers from the JSON https://bugs.webkit.org/show_bug.cgi?id=64990
Reviewed by Adam Barth.

Now that names are normalized, we can remove all the code that
handles gtest name changes (e.g. for adding modifiers like DISABLED_).
Instead, if you try to list a test with a modifier in it, we need
to strip the modifier so we get the normalized value.

We also get rid of the concept of extra/missing expectations for gtests.
In a patch soon, we'll stop showing extra/missing expectations from the UI
entirely and only leave it for the special updating test_expectations.txt
view of the dashboard, which doesn't apply to gtests.

* TestResultServer/static-dashboards/flakiness_dashboard.html:
* TestResultServer/static-dashboards/flakiness_dashboard_tests.js:

git-svn-id: svn://svn.chromium.org/blink/trunk@91541 bbb929c8-8fbe-4397-9dbb-9b2b20218538