Fix crash when mouse moves from <summary> element to parent <details> element. https://bugs.webkit.org/show_bug.cgi?id=66210
Reviewed by Dimitri Glazkov.

This is a regression caused by r92922, which wrongly assumes that
a shadow host always has a shadow root as an immediate child in
ancestors chain. This assumption does not apply to <details>
element. <details> element is implemented as a shadow host, but
may have a <summary> element as an immediate child element in
ancestors chain.

Source/WebCore:

Test: fast/dom/shadow/details-summary-mouseover.html

* dom/EventDispatcher.cpp:
(WebCore::EventDispatcher::adjustToShadowBoundaries):

LayoutTests:

* fast/dom/shadow/details-summary-mouseover-expected.txt: Added.
* fast/dom/shadow/details-summary-mouseover.html: Added.

git-svn-id: svn://svn.chromium.org/blink/trunk@93079 bbb929c8-8fbe-4397-9dbb-9b2b20218538