Update Windows Sputnik results for ICU 4.6
Windows now fails these tests due to changes in ICU. Mac fails these tests, too, though in a
different way. Alexey tells me that these tests are silly because they check compliance with
a particular version of Unicode. Given this, it doesn't seem that a bug is required to track
this change in behavior.

* platform/win/sputnik/Unicode/Unicode_510/S7.6_A3.1-expected.txt:
* platform/win/sputnik/Unicode/Unicode_510/S7.6_A3.2-expected.txt:
* platform/win/sputnik/Unicode/Unicode_510/S7.6_A5.3_T1-expected.txt:
* platform/win/sputnik/Unicode/Unicode_510/S7.6_A5.3_T2-expected.txt:

git-svn-id: svn://svn.chromium.org/blink/trunk@93103 bbb929c8-8fbe-4397-9dbb-9b2b20218538