<http://webkit.org/b/67983> CRASH under WebCore::ArchiveResourceCollection::addAllResources loading WebArchive
Reviewed by Darin Adler.

Source/WebCore:

Null mimetypes are okay for subresources, but as previously discovered
in bug 41082 they were required for main resources. Here we ensure
that they are required for main resources, but we allow null mimetypes
for subresources.

To fix a crash for bad input we never add a null ArchiveResource to the
Archive's subresource collection. It is useless and causes crashes whenever
someone iterates the Archive's subresource collection.

Test: webarchive/loading/test-loading-archive-subresource-null-mimetype.html

* loader/archive/cf/LegacyWebArchive.cpp:
(WebCore::LegacyWebArchive::createResource):
Allow a null mimetype, just type check from the dictionary.

(WebCore::LegacyWebArchive::extract):
Require a mimetype for the main resource. Do not add null
subresources to the subresource list.

LayoutTests:

Test a WebArchive that has been constructed to have a null
mimetype for a subresource.

* webarchive/loading/test-loading-archive-subresource-null-mimetype-expected.txt: Added.
* webarchive/loading/test-loading-archive-subresource-null-mimetype.html: Added.
* webarchive/loading/resources/subresource-null-mimetype.webarchive: Added.

git-svn-id: svn://svn.chromium.org/blink/trunk@95038 bbb929c8-8fbe-4397-9dbb-9b2b20218538