De-virtualize JSCell::getJSNumber https://bugs.webkit.org/show_bug.cgi?id=68651
Reviewed by Oliver Hunt.

Added a new JSType to check whether or not something is a 
NumberObject (which includes NumberPrototype) in TypeInfo::isNumberObject because there's not 
currently a better way to determine whether something is indeed a NumberObject.
Also de-virtualized JSCell::getJSNumber, having it check the TypeInfo 
for whether the object is a NumberObject or not.  This patch is part of 
the larger process of de-virtualizing JSCell.

* JavaScriptCore.exp:
* JavaScriptCore.vcproj/JavaScriptCore/JavaScriptCore.def:
* runtime/JSCell.cpp:
(JSC::JSCell::getJSNumber):
* runtime/JSCell.h:
(JSC::JSValue::getJSNumber):
* runtime/JSType.h:
* runtime/JSTypeInfo.h:
(JSC::TypeInfo::isNumberObject):
* runtime/JSValue.h:
* runtime/NumberObject.cpp:
(JSC::NumberObject::getJSNumber):
* runtime/NumberObject.h:
(JSC::NumberObject::createStructure):
* runtime/NumberPrototype.h:
(JSC::NumberPrototype::createStructure):


git-svn-id: svn://svn.chromium.org/blink/trunk@95893 bbb929c8-8fbe-4397-9dbb-9b2b20218538