Crash when PluginProxy object is destroyed inside PluginProxy::pluginScriptableNPObject https://bugs.webkit.org/show_bug.cgi?id=69559 <rdar://problem/9704066>
Reviewed by Maciej Stachowiak.

Protect the plug-in itself in the call to pluginScriptableNPObject. In some cases, sending
the synchronous message to the plug-in process can end up causing the web process to handle
an incoming message that will destroy t he plug-in.

Unfortunately, this is highly timing-dependent and can't be tested reliably.

* WebProcess/Plugins/PluginProxy.cpp:
(WebKit::PluginProxy::pluginScriptableNPObject):


git-svn-id: svn://svn.chromium.org/blink/trunk@96855 bbb929c8-8fbe-4397-9dbb-9b2b20218538