Add compile-time asserts for RenderStyle::(Non)InheritedFlags size. https://bugs.webkit.org/show_bug.cgi?id=69803
Reviewed by Darin Adler.

No new tests.

Adds COMPILE_ASSERTs for the size of RenderStyle::InheritedFlags and
RenderStyle::NonInheritedFlags to ensure that they are not accidentally
grown in future changes.
Change types of unsigned bitfields to unsigned char to ensure that the
above assertions remain true when building on windows.

* rendering/style/RenderStyle.cpp:
(WebCore::RenderStyle):
Adds assertions to ensure struct sizes do not change.
* rendering/style/RenderStyle.h:
Changes unsigned to unsigned char as outlined above.


git-svn-id: svn://svn.chromium.org/blink/trunk@97712 bbb929c8-8fbe-4397-9dbb-9b2b20218538