AX: buttons of number type <input> controls are not fully accessible https://bugs.webkit.org/show_bug.cgi?id=70241
Expose the spin button object to the AX hierarchy. Since there is no distinct
render object that backs this object, a mock object needs to be created for the entire
spin button and the actual increment and decrement buttons inside.

This will be a staged commit since the first time around Windows layout tests started crashing for an undeterminable reason.

Reviewed by Darin Adler.

* html/shadow/TextControlInnerElements.cpp:
(WebCore::SpinButtonElement::step):
(WebCore::SpinButtonElement::repeatingTimerFired):
* html/shadow/TextControlInnerElements.h:



git-svn-id: svn://svn.chromium.org/blink/trunk@97753 bbb929c8-8fbe-4397-9dbb-9b2b20218538