Caret can be positioned or repainted incorrectly in flipped-blocks multi-column blocks https://bugs.webkit.org/show_bug.cgi?id=70851
Reviewed by Dave Hyatt.

* manual-tests/caret-in-columns-flipped.html: Added.
* rendering/RenderBlock.cpp:
(WebCore::RenderBlock::adjustForColumns): Removed flipped-blocks writing mode considerations
from this function, since not all callers were expecting it to account for flippedness. This makes
it similar to adjustRectForColumns() in not adjusting for flippedness.
* rendering/RenderBox.cpp:
(WebCore::RenderBox::offsetFromContainer): In the columns case, account for flipped-blocks modes
by first mapping the point and the offset to non-flipped, non-columns space, then adjusting for
columns, then flipping.



git-svn-id: svn://svn.chromium.org/blink/trunk@98400 bbb929c8-8fbe-4397-9dbb-9b2b20218538