CMake: Add support to optionally install the built JavaScript shell https://bugs.webkit.org/show_bug.cgi?id=71062
Reviewed by Antonio Gomes.

.: 

Adds a CMake option, called SHOULD_INSTALL_JS_SHELL, that when
defined will generate an installation rule to install the built
JavaScript shell into /bin (with respect to the prefix path).

By default, we don't define this option and hence don't generate
an installation rule for the JavaScript shell.

* Source/CMakeLists.txt:

Source/JavaScriptCore: 

Generate an installation rule for installing the JavaScript shell in
/bin (with respect to the prefix path) when SHOULD_INSTALL_JS_SHELL
is defined.

* shell/CMakeLists.txt:


git-svn-id: svn://svn.chromium.org/blink/trunk@98681 bbb929c8-8fbe-4397-9dbb-9b2b20218538