[Qt] REGRESSION: 5 editing tests fail https://bugs.webkit.org/show_bug.cgi?id=71117
Reviewed by Andreas Kling.

Tools:

Make sure that the view is properly focused, in order to get the
didBeginEditing DRT calls.

This requires activating the window (the call will result in a FocusIn
event to the window as well as setting QGuiApplication::focusWindow())
as well as enabling the focus on the view item. The signature of
setFocus has changed to take a boolean.

* WebKitTestRunner/qt/PlatformWebViewQt.cpp:
(WTR::WrapperWindow::handleStatusChanged):

LayoutTests:

* platform/qt-5.0/Skipped: Unskip tests that now pass.

git-svn-id: svn://svn.chromium.org/blink/trunk@99081 bbb929c8-8fbe-4397-9dbb-9b2b20218538