run-leaks does not work on Lion? <http://webkit.org/b/71059> <rdar://problem/10428527>
Reviewed by Adam Roben.

The output of leaks(1) changed again in Lion to move the
"leaks Report Version: 2.0" line from the first line of the
output to just above the "Process " lines that run-leaks is
interested in parsing.  This required using a more generic
algorithm to find the start of the "Process " lines.

* Scripts/run-leaks:
(parseLeaksOutput): Make the code to skip headers more generic.
* Scripts/webkitperl/run-leaks_unittest/run-leaks-report-v1.0.pl: Added.
* Scripts/webkitperl/run-leaks_unittest/run-leaks-report-v2.0-new.pl: Added.
* Scripts/webkitperl/run-leaks_unittest/run-leaks-report-v2.0-old.pl: Added.


git-svn-id: svn://svn.chromium.org/blink/trunk@100174 bbb929c8-8fbe-4397-9dbb-9b2b20218538