Fix Aspect Ratio Property Inheritance And Make the Computed Value Equal the Specified Value https://bugs.webkit.org/show_bug.cgi?id=73038
Source/WebCore:

Reviewed by Ojan Vafai.

hasAspectRatio was not being set if the -webkit-aspect-ratio property was inherited from the parent.

Additionally, the computed value of -webkit-aspect-ratio was returning a number instead of a ratio.

Test: fast/css/aspect-ratio-inheritance.html

* css/CSSComputedStyleDeclaration.cpp:
(WebCore::CSSComputedStyleDeclaration::getPropertyCSSValue):
* css/CSSStyleApplyProperty.cpp:
(WebCore::ApplyPropertyAspectRatio::applyInheritValue):

LayoutTests:

Reviewed by Ojan Vafai.

Updated the parsing test to make sure that the -webkit-aspect-ratio property accepts inherit as a possible value.

Added an inheritance test that checks the computed value is correctly inherited.

* fast/css/aspect-ratio-inheritance-expected.txt: Added.
* fast/css/aspect-ratio-inheritance.html: Added.
* fast/css/aspect-ratio-parsing-tests-expected.txt:
* fast/css/aspect-ratio-parsing-tests.html:

git-svn-id: svn://svn.chromium.org/blink/trunk@101275 bbb929c8-8fbe-4397-9dbb-9b2b20218538