platform/mac/accessibility/search-when-element-starts-in-table.html is failing https://bugs.webkit.org/show_bug.cgi?id=73751
Source/WebCore: 

When encountering a table, the rows() and not the cells() should be queried (otherwise
we can end up finding the element we started with).

When searching in reverse, we also need to account for searching elements within the
parent hierarchy. because technically it is "behind" the start element.

Reviewed by Darin Adler.

* accessibility/AccessibilityObject.cpp:
(WebCore::appendChildrenToArray):
(WebCore::AccessibilityObject::objectMatchesSearchCriteriaWithResultLimit):
(WebCore::AccessibilityObject::findMatchingObjects):
* accessibility/AccessibilityObject.h:

LayoutTests: 

Updated tests to be more comprehensive.

Reviewed by Darin Adler.

* platform/mac/accessibility/search-when-element-starts-in-table-expected.txt:
* platform/mac/accessibility/search-when-element-starts-in-table.html:



git-svn-id: svn://svn.chromium.org/blink/trunk@102031 bbb929c8-8fbe-4397-9dbb-9b2b20218538