[Qt] Fix non-animated pinch-zoom scale commit. https://bugs.webkit.org/show_bug.cgi?id=74007
Reviewed by Kenneth Rohde Christiansen.

The update deferrer object was only destroyed at the end of the animation
and wouldn't be if the pinch-zoom was ended within legal bounds.
This patch also makes sure that the deferrer isn't destroyed and re-created
once the animation is started if it was already there.

* UIProcess/qt/QtViewportInteractionEngine.cpp:
(WebKit::QtViewportInteractionEngine::animateItemRectVisible):
(WebKit::QtViewportInteractionEngine::scaleAnimationStateChanged):
(WebKit::QtViewportInteractionEngine::ensureContentWithinViewportBoundary):
(WebKit::QtViewportInteractionEngine::pinchGestureEnded):
* UIProcess/qt/QtViewportInteractionEngine.h:

git-svn-id: svn://svn.chromium.org/blink/trunk@102338 bbb929c8-8fbe-4397-9dbb-9b2b20218538