chromium: visual studio projects are busted when build_webkit_exes_from_webkit_gyp==0 https://bugs.webkit.org/show_bug.cgi?id=74212
Reviewed by Tony Chang.

The change introduced in r102201 to split the executables out of
WebKit.gyp apparently doesn't work right on Windows, since
WebKit.gyp and Tools.gyp refer to dependent projects by
different relative paths (it appears gyp doesn't normalize the
paths). This change moves Tools.gyp to be in the same directory
as WebKit.gyp (and All.gyp), which should fix this and
conveniently will also solve the problem of how to add Tools.gyp
to the chromium DEPS file.

Source/WebKit/chromium:

* All.gyp:
* Tools.gyp: Renamed from ../../../Tools/Tools.gyp.

Tools:

* Tools.gyp: Renamed to ../Source/WebKit/chromium/Tools.gyp

git-svn-id: svn://svn.chromium.org/blink/trunk@102482 bbb929c8-8fbe-4397-9dbb-9b2b20218538