[Refactoring] Remove several global variables from prepare-ChangeLog https://bugs.webkit.org/show_bug.cgi?id=74389
Reviewed by David Kilzer.

We are planning to write unit-tests for prepare-ChangeLog in a run-leaks_unittest
manner. This bug is one of the incremental refactorings to remove all top-level
code and global variables from prepare-ChangeLog. In this patch,
we make the following global variables being used only through parameter passing.
    - %paths
    - $changedFiles
    - $conflictFiles
    - $functionLists
    - $bugURL

* Scripts/prepare-ChangeLog:
(findChangeLogs):
(generateNewChangeLogs):
(processPaths):
(generateFileList):
(firstDirectoryOrCwd):


git-svn-id: svn://svn.chromium.org/blink/trunk@102736 bbb929c8-8fbe-4397-9dbb-9b2b20218538