The HTML parser doesn't enforce the "Noah's Ark condition" from the HTML5 spec https://bugs.webkit.org/show_bug.cgi?id=74828
Reviewed by Darin Adler.

Source/WebCore:

This patch implement the "Noah's Ark condition" from the HTML5
specification.  This condition limits the number of identitical
elements that can be in the list of active formatting elements.  I'm not
entirely sure that enforcing this condition is worth the complexity,
but given that we've come this far in support of the HTML5 parsing
algorithm, we might as well finish it.

After this patch, we pass all but one of the html5lib parsing tests!

Tests: html5lib/runner.html

* html/parser/HTMLFormattingElementList.cpp:
(WebCore::attributeCount):
(WebCore::HTMLFormattingElementList::append):
(WebCore::HTMLFormattingElementList::tryToEnsureNoahsArkConditionQuickly):
(WebCore::HTMLFormattingElementList::ensureNoahsArkCondition):
* html/parser/HTMLFormattingElementList.h:

LayoutTests:

Show test progression.

* fast/parser/residual-style-dom-expected.txt:
    - This isn't a great test because it's hard to see how changes
      affect the output.  However, we have good coverage of these
      topics in the HTML5lib tests.
* html5lib/runner-expected.txt:
* platform/chromium/html5lib/runner-expected.txt:


git-svn-id: svn://svn.chromium.org/blink/trunk@103223 bbb929c8-8fbe-4397-9dbb-9b2b20218538