Fullscreen video controller can't be dragged the first time I enter fullscreen https://bugs.webkit.org/show_bug.cgi?id=75709
Reviewed by Eric Carlson.

Source/WebCore:

No new tests; updated video-controls-drag.html.

When the media controls are created, check to see if we are full screen, and pass
that information to the newly created controls.

* html/HTMLMediaElement.cpp:
(WebCore::HTMLMediaElement::createMediaControls):

LayoutTests:

Update the test to cover the case where the controls do not yet exist when entering
fullscreen.

* fullscreen/video-controls-drag.html:

git-svn-id: svn://svn.chromium.org/blink/trunk@104306 bbb929c8-8fbe-4397-9dbb-9b2b20218538