Unmatched transparency layer begin/end on a filtered element with an opacity ancestor https://bugs.webkit.org/show_bug.cgi?id=76329
Source/WebCore:

Reviewed by Dan Bernstein.

When doing a paint with painting disabled on the GraphicsContext, as we do
for updating control tints, or computing text rectangles on Find, do not
apply filter effects. This is both a performance gain, and fixes an issue
with mismatched begin/end transparency layers.

* rendering/RenderLayer.cpp:
(WebCore::RenderLayer::paintLayerContents):

Reviewed by Dan Bernstein.

Manual test with a filtered element as a descendant of an element with
opacity. Doing a Find is necessary to exercise the test.

* ManualTests/filters/opacity-above-filter.html: Added.

git-svn-id: svn://svn.chromium.org/blink/trunk@105019 bbb929c8-8fbe-4397-9dbb-9b2b20218538