REGRESSION (r104000): AdBlock extension fails to load/function https://bugs.webkit.org/show_bug.cgi?id=75554
Reviewed by Andy Estes.

Source/WebCore:

Previously, we would claim the documentElement was available before
we'd actually attached it to the DOM (which we now do via the
attachment queue).  This issue was noted in the code with a FIXME
comment, but there was no test coverage for the issue.  This patch
resolves the FIXME and adds a test.

Test: userscripts/document-element-available-at-start.html

* html/parser/HTMLConstructionSite.cpp:
(WebCore::HTMLConstructionSite::insertHTMLHtmlStartTagBeforeHTML):

LayoutTests:

This test checks that the documentElement is available when user
scripts are run at document start.

* userscripts/document-element-available-at-start-expected.txt: Added.
* userscripts/document-element-available-at-start.html: Added.


git-svn-id: svn://svn.chromium.org/blink/trunk@105321 bbb929c8-8fbe-4397-9dbb-9b2b20218538