In CodeGeneratorObjC.pm, overwrite the output .h/.mm only if the bytes differ https://bugs.webkit.org/show_bug.cgi?id=76874
Reviewed by Adam Barth.

This is one of steps to stop rebuilding .h/.cpp/.mm files
generated by unchanged IDLs (bug 76836).
This patch makes a change on CodeGeneratorObjC.pm so that
it overwrites the output .h/.mm only if the bytes differ.

No tests. No change in behavior.
I manually confirmed that when I add a new attribute to Element.idl,
the time-stamps of unrelated DOM*.h and DOM*.mm do not change.

* bindings/scripts/CodeGenerator.pm:
(UpdateFileIfChanged): Added. This method writes data to a file
only if the data is different from the data in the current file.
* bindings/scripts/CodeGeneratorObjC.pm:
(WriteData): Used UpdateFileIfChanged().


git-svn-id: svn://svn.chromium.org/blink/trunk@105697 bbb929c8-8fbe-4397-9dbb-9b2b20218538