r105674 broke check-webkit-style for chromium's test_expectations.
Unreviewed, build fix.

We now need to pass in a platform name of 'chromium' if we want
the chromium port; options.chromium is no longer recognized. I
will follow this patch up with a separate patch that has a test,
but this at least fixes the build.

* Scripts/webkitpy/style/checkers/test_expectations.py:
(TestExpectationsChecker._determine_port_from_exepectations_path):

git-svn-id: svn://svn.chromium.org/blink/trunk@105793 bbb929c8-8fbe-4397-9dbb-9b2b20218538