Source/WebCore: https://bugs.webkit.org/show_bug.cgi?id=78122
Add support for the "contains" value for line-grid-snap. This value centers the line box in between the
text-top and text-bottom of the minimum number of grid lines that enclose the line box. This is useful for
centering headers in a line grid.

Reviewed by Adam Roben.

Added a new test in fast/line-grid.

* rendering/RootInlineBox.cpp:
(WebCore::RootInlineBox::lineGridSnapAdjustment):

LayoutTests: https://bugs.webkit.org/show_bug.cgi?id=78122
        
Add a new test that shows the "contains" value in action.

Reviewed by Adam Roben.

* fast/line-grid/line-grid-contains-value.html: Added.
* platform/mac/fast/line-grid/line-grid-contains-value-expected.png: Added.
* platform/mac/fast/line-grid/line-grid-contains-value-expected.txt: Added.



git-svn-id: svn://svn.chromium.org/blink/trunk@107099 bbb929c8-8fbe-4397-9dbb-9b2b20218538