Source/WebCore: https://bugs.webkit.org/show_bug.cgi?id=79940 <rdar://problem/10080189>
Add support in WebKit for an intra-line character grid for Japanese text layout.
        
Patch logicalLeftOffsetForLine and logicalRightOffsetForLine in order to get the
basic edge snapping grid functionality up and running. See all the FIXMEs in the function for
some of the issues that still have to be dealt with for it to really work well.

Reviewed by Dan Bernstein.

Added new tests in fast/line-grid.

* rendering/RenderBlock.cpp:
(WebCore::RenderBlock::logicalLeftOffsetForLine):
(WebCore::RenderBlock::logicalRightOffsetForLine):

LayoutTests: https://bugs.webkit.org/show_bug.cgi?id=79940
<rdar://problem/10080189>
        
Add support in WebKit for an intra-line character grid for Japanese text layout.
        
New tests for basic left and right alignment of lines in the presence of
floating and positioned objects as well as objects with margins and borders.

Reviewed by Dan Bernstein.

* fast/line-grid/line-align-left-edges.html: Added.
* fast/line-grid/line-align-right-edges.html: Added.
* platform/mac/fast/line-grid/line-align-left-edges-expected.png: Added.
* platform/mac/fast/line-grid/line-align-left-edges-expected.txt: Added.
* platform/mac/fast/line-grid/line-align-right-edges-expected.png: Added.
* platform/mac/fast/line-grid/line-align-right-edges-expected.txt: Added.



git-svn-id: svn://svn.chromium.org/blink/trunk@109267 bbb929c8-8fbe-4397-9dbb-9b2b20218538