[Textmap] Disable driver based BGRA swizzling for OpenGL/ES
Reviewed by Noam Rosenthal.

The current way of detecting swizzling support (check for GL_EXT_texture_format_BGRA8888)
is unfortunately not reliable. We need a better way of detection that (filed bug 81103).
In the meantime this patch reverts back to doing the swizzling in the texture mapper to
get pixels back onto the screen.

* platform/graphics/texmap/TextureMapperGL.cpp:
(WebCore::driverSupportsBGRASwizzling):
(WebCore::BitmapTextureGL::updateContents):

git-svn-id: svn://svn.chromium.org/blink/trunk@110703 bbb929c8-8fbe-4397-9dbb-9b2b20218538