[TexMap] Reuse textures following the same rules as they do internally. https://bugs.webkit.org/show_bug.cgi?id=80843
Reviewed by Noam Rosenthal.

BitmapTextures were recently changed to be the same size as their contents.
This would obsolete the logic in acquireTextureFromPool which would
assume that a BitmapTexture has good chances of being reused if its
size is bigger or equal to the new size.

This asks the texture instead if it can be reused which now simply
check for an exact size match.

* platform/graphics/texmap/TextureMapper.cpp:
(WebCore::TextureMapper::acquireTextureFromPool):
* platform/graphics/texmap/TextureMapper.h:
(WebCore::BitmapTexture::canReuseWith):
(WebCore::BitmapTexture::reset):
(WebCore::BitmapTexture::didReset):
* platform/graphics/texmap/TextureMapperGL.cpp:
(WebCore::BitmapTextureGL::canReuseWith):
(WebCore):
* platform/graphics/texmap/TextureMapperGL.h:
(BitmapTextureGL):

git-svn-id: svn://svn.chromium.org/blink/trunk@110979 bbb929c8-8fbe-4397-9dbb-9b2b20218538