Unreviewed; fix webkitpy unit test failures after r111116.
It turns out that cmd_line in the patch in bug 80025 had the right
signature when it was submitted, but r109242 added a few more
required parameters, and since the method (and EflDriver) did not
exist at the time we ended up committing a broken method.

* Scripts/webkitpy/layout_tests/port/efl.py:
(EflDriver.cmd_line): Add the `pixel_tests' and `per_test_args'
arguments.

git-svn-id: svn://svn.chromium.org/blink/trunk@111122 bbb929c8-8fbe-4397-9dbb-9b2b20218538