REGRESSION (r110065-r110080): fast/forms/placeholder-set-attribute.html is failing intermittently because WebKit fails to repaint after setting the placeholder attribute https://bugs.webkit.org/show_bug.cgi?id=81802
Reviewed by Dan Bernstein.

Source/WebCore:

Covered by fast/forms/placeholder-set-attribute.html which should be less flaky.

Unfortunately no new test case as this bug requires a very specific set of conditions that I couldn't reproduce deterministically.

This is a regression from r110072: RenderTextControlSingleLine would rely on the placeholder's RenderLayer to properly repaint during
the first layout as the placeholder has overflow: hidden set. r110072 removed the layer in this case and thus we miss a repaint.

* rendering/RenderTextControlSingleLine.cpp:
(WebCore::RenderTextControlSingleLine::layout):
For our first layout, we need to make sure our placeholder is painted. layoutBlockChild has a very similar logic to force repaint
on a first layout. We don't hit this logic as the placeholder is explicitly skipped in RenderTextControl::layoutSpecialExcludedChild.

LayoutTests:

* platform/mac/Skipped: Unskip the test on Mac. The other platforms are either
skipping it for other reasons or did not have a specific entry for the flakiness.


git-svn-id: svn://svn.chromium.org/blink/trunk@112114 bbb929c8-8fbe-4397-9dbb-9b2b20218538