[jhbuild] Prepend jhbuildrc's directory to sys.path instead of appending. https://bugs.webkit.org/show_bug.cgi?id=82252
Reviewed by Gustavo Noronha Silva.

Appending the directory containing common.py may be problematic if
another path in sys.path also contains either a common.py or a
common/ module, so prepend to select our version instead. This is
the case with fail2ban, which has a common/ directory which may
end up being in Python's default path.

* efl/jhbuildrc:
* gtk/jhbuildrc:

git-svn-id: svn://svn.chromium.org/blink/trunk@112309 bbb929c8-8fbe-4397-9dbb-9b2b20218538