Fix a crash when closing a tab/window while the Web Inspector is stopped in the debugger.
https://webkit.org/b/82846
rdar://problem/8133494

Reviewed by Yury Semikhatsky.

* bindings/js/PageScriptDebugServer.cpp:
(WebCore::PageScriptDebugServer::didPause): Added assert for page and early return. Also assert that
m_pausedPage is null.
(WebCore::PageScriptDebugServer::didContinue): Page can be null if we are continuing because the Page closed.
So add a null check before using it. Added an assert that the page is null or m_pausedPage.
(WebCore::PageScriptDebugServer::didRemoveLastListener): Added assert for page.


git-svn-id: svn://svn.chromium.org/blink/trunk@112818 bbb929c8-8fbe-4397-9dbb-9b2b20218538