[Part 3] We should use CSSPropertyID rather than integers when manipulating CSS property ids. https://bugs.webkit.org/show_bug.cgi?id=83032
Reviewed by Kentaro Hara.

CSSPropertyID enum holds all the CSS property ids but many parts of WebKit treat the ids
as integers. While it's not incorrect it is nicer to use the enum as a parameter of
functions manipulating property ids, as we ensure that the value passed will be an
existing value. This patch clean up some remaining part of code.

No new tests : There should be no behavior change in this patch.

* bindings/js/JSCSSStyleDeclarationCustom.cpp:
(CSSPropertyInfo):
(WebCore::cssPropertyIDForJSCSSPropertyName):
* bindings/v8/custom/V8CSSStyleDeclarationCustom.cpp:
(CSSPropertyInfo):
(WebCore::cssPropertyInfo):
* css/CSSComputedStyleDeclaration.cpp:
(WebCore::getPositionOffsetValue):
(WebCore::counterToCSSValue):
* css/CSSParser.cpp:
(WebCore::CSSParser::parseValidPrimitive):
Remove the parameter's name from the h file as it doesn't follow the style and also because the
name was simply wrong we don't expect a propId here but an indentifier. Make it clear in the
cpp file too.
* css/CSSParser.h:
(CSSParser):
* page/animation/AnimationBase.cpp:
(WebCore::addPropertyWrapper):


git-svn-id: svn://svn.chromium.org/blink/trunk@113042 bbb929c8-8fbe-4397-9dbb-9b2b20218538