IndexedDB: Race condition in layout test logging from window and worker https://bugs.webkit.org/show_bug.cgi?id=83378
The sequencing of logging directly from window vs. logging via posted events
from a worker is racy, making this test flake when run under Chromium. Removed
the sensitive logging on the window side; the test logic is unaffected as the
blocked/success event sequence would not be present if the window handler was
not run.

Reviewed by Tony Chang.

* storage/indexeddb/dont-commit-on-blocked-expected.txt:
* storage/indexeddb/dont-commit-on-blocked.html:


git-svn-id: svn://svn.chromium.org/blink/trunk@113485 bbb929c8-8fbe-4397-9dbb-9b2b20218538