Remove unnecessary calls to RenderSVGShape::createShape() https://bugs.webkit.org/show_bug.cgi?id=83598
Reviewed by Rob Buis.

RenderSVGShape::strokeContains contains calls to fall back to path stroke hit detection
in several cases (dashed strokes, for example). One of these cases, a check for isStyled,
is not needed and is always true for Ellipses, Circles, and Rects (all styled elements).
This change removes the check for isStyled.

No new tests as this only affects performance, not test results.

* rendering/svg/RenderSVGShape.cpp:
(WebCore::RenderSVGShape::strokeContains):


git-svn-id: svn://svn.chromium.org/blink/trunk@113781 bbb929c8-8fbe-4397-9dbb-9b2b20218538