Avoid synchronously recalculating the nonFastScrollableRegion when the ScrollableArea set changes https://bugs.webkit.org/show_bug.cgi?id=84470
Patch by James Robinson <jamesr@chromium.org> on 2012-04-20
Reviewed by Anders Carlsson.

Source/WebCore:

When a scrollable area is added or removed, we shouldn't recalculate the nonFastScrollableRegion synchronously
since this is slow, the RenderObject tree might not be in an up-to-date state, and we'll typically do layout
soon anyway.

* page/FrameView.cpp:
(WebCore::FrameView::addScrollableArea):
(WebCore::FrameView::removeScrollableArea):
* page/scrolling/ScrollingCoordinator.cpp:
* page/scrolling/ScrollingCoordinator.h:
(ScrollingCoordinator):

Source/WebKit/chromium:

Trigger a layout when a plugin becomes or stops being scrollable so the nonFastScrollableRegion
can be updated.

* src/ScrollbarGroup.cpp:
(WebKit::ScrollbarGroup::scrollbarCreated):
(WebKit::ScrollbarGroup::scrollbarDestroyed):

git-svn-id: svn://svn.chromium.org/blink/trunk@114778 bbb929c8-8fbe-4397-9dbb-9b2b20218538