Regression(r117482): Run-in crashes relating to generated content and inline line box clearing. https://bugs.webkit.org/show_bug.cgi?id=86879
Reviewed by Julien Chaffraix.

Source/WebCore:

Tests: fast/runin/generated-content-crash.html
       fast/runin/move-run-in-original-position-crash.html

* rendering/RenderBlock.cpp:
(WebCore::RenderBlock::addChildIgnoringAnonymousColumnBlocks): Use the new helper
placeRunInIfNeeded. Do not place the run-in if it is a generated container since
the caller such as updateBeforeAfterContent might be keeping a reference to it
and adding children to it later.
(WebCore::destroyRunIn): Add ability to clear line box tree for inlines.
(WebCore::RenderBlock::placeRunInIfNeeded): Helper to place run-in. Add an
argument to not modify generated content during addChild, it should be moved
only at end of updateBeforeAfterContent.
(WebCore::RenderBlock::moveRunInUnderSiblingBlockIfNeeded):
(WebCore::RenderBlock::runInIsPlacedIntoSiblingBlock): helper to tell if this run-in
is actually placed into the next sibling block.
(WebCore::RenderBlock::moveRunInToOriginalPosition):
* rendering/RenderInline.cpp:
(WebCore::RenderInline::deleteLineBoxTree): like RenderBlock, add a helper
for deleteLineBoxTree. Virtualizing this might not be good, since this is
the only call site for inline line box tree clearing and RenderBlock::deleteLineBoxTree
is called a lot.
* rendering/RenderObjectChildList.cpp:
(WebCore::createRendererForBeforeAfterContent): fix a typo.
(WebCore::RenderObjectChildList::updateBeforeAfterContent): If insertBefore is equal
to the intruded run-in, then set it to next sibling so that new child will come after it. At
the end, place the generatedContainer if it is a run-in.

LayoutTests:

These tests need to show dump render tree to show the correct location of run-in placement. All
renderings match IE and Opera.

* fast/lists/parent-box-not-box-crash-expected.txt: Rebaselined, since now run-in functions correctly
by being the first child before list marker.
* fast/runin/generated-content-crash-expected.png: Added.
* fast/runin/generated-content-crash-expected.txt: Added.
* fast/runin/generated-content-crash.html: Added.
* fast/runin/move-run-in-original-position-crash-expected.png: Added.
* fast/runin/move-run-in-original-position-crash-expected.txt: Added.
* fast/runin/move-run-in-original-position-crash.html: Added.
* fast/runin/runin-between-list-marker-and-before-content-expected.png:
* fast/runin/runin-between-list-marker-and-before-content-expected.txt: Rebaselined, since now run-in
functions correctly by being the first child before list marker.

git-svn-id: svn://svn.chromium.org/blink/trunk@117865 bbb929c8-8fbe-4397-9dbb-9b2b20218538