&AElig doesn't get rendered as U+00C6 https://bugs.webkit.org/show_bug.cgi?id=87465
Reviewed by Adam Barth.

Source/WebCore:

The if-statement to check whether the first entry of a certain letter
already exists in the index used .get(). The very first alphabetical
entry is in position "0", which evaluates to false, causing the first
entry to be ignored. Instead, use a "x not in y" check here.

Also update WebCore.gyp to list the create-html-entity-table script as
an input for the action, to make sure the table will be recreated.

Test: html5lib/resources/entities02.dat

* WebCore.gyp/WebCore.gyp:
* html/parser/create-html-entity-table:

LayoutTests:

Test that the (current) first alphabetical entity in the named entity
table (&AElig, without a semi-colon) renders as expected.

* html5lib/resources/entities02.dat:


git-svn-id: svn://svn.chromium.org/blink/trunk@118672 bbb929c8-8fbe-4397-9dbb-9b2b20218538