[JSC] SerializedScriptValue.create() succeeds even if MessagePort object cannot be found in transferred ports https://bugs.webkit.org/show_bug.cgi?id=87118
Patch by Christophe Dumez <christophe.dumez@intel.com> on 2012-05-30
Reviewed by Adam Barth.

Source/WebCore:

Make SerializedScriptValue.create() throw an exception if one of the
MessagePort objects cannot be found in the transferred ports. This
matches the behavior of the V8 implementation.

Test: webintents/web-intents-obj-constructor.html

* bindings/js/SerializedScriptValue.cpp:
(WebCore::CloneSerializer::dumpIfTerminal):

LayoutTests:

Unskip webintents/web-intents-obj-constructor.html now that
SerializedScriptValue.create() throws an exception if one of the
MessagePort objects cannot be found in the transferred ports.

* platform/efl/test_expectations.txt:

git-svn-id: svn://svn.chromium.org/blink/trunk@119027 bbb929c8-8fbe-4397-9dbb-9b2b20218538