HTML parser should yield more to improve perceived page load time https://bugs.webkit.org/show_bug.cgi?id=86165
Source/WebCore:

Patch by Shrey Banga <banga@chromium.org> on 2012-06-12
Reviewed by Tony Gentilcore.

Test: fast/events/event-yield-timing.html

We want the parser to yield at least every 500ms to keep the page somewhat responsive and allow painting.
Since it would be too expensive to check the time after each token, the previous heuristic was to check every 4,096 tokens.
That works fine for most tokens, but a script may spend an arbitrary amount of time executing.

This patch fixes the issue by causing the parser to check the elapsed time immediately after executing a script.

* html/parser/HTMLParserScheduler.cpp:
(WebCore::HTMLParserScheduler::checkForYieldBeforeScript):
* html/parser/HTMLParserScheduler.h:
(WebCore::PumpSession::PumpSession):
(PumpSession):
(WebCore::HTMLParserScheduler::checkForYieldBeforeToken):

LayoutTests:

Added test for parser yield times after seeing a script

Patch by Shrey Banga <banga@chromium.org> on 2012-06-12
Reviewed by Tony Gentilcore.

* fast/parser/parser-yield-timing-expected.txt: Added.
* fast/parser/parser-yield-timing.html: Added.
* http/tests/loading/gmail-assert-on-load-expected.txt: The test was added
for an assert that failed on debug builds when an iframe removed itself
from its parent window after blocking for 1 second. In previous builds,
the destroyIt() function was called without yielding to the event loop,
which lead to didFailLoadWithError being called. The fix for bug 86165
ensures that we yield after executing the script, which leads to
didHandleOnloadEventsForFrame being called instead. This is the cause
of the change in output of the test. The test still successfully removes
the iframe and since no asserts fail in the debug build, we should consider
it as passing.

git-svn-id: svn://svn.chromium.org/blink/trunk@120108 bbb929c8-8fbe-4397-9dbb-9b2b20218538