Add test cases for header/footer elements' AXRoleDescription https://bugs.webkit.org/show_bug.cgi?id=88911
Header and footer elements change their accessibility roles depending
on the elements which contain them. Specifically, they are described
as "banner" or "footer" in the general case, and fall back to
generic "group" role descriptions when contained in either article or
section elements. This patch simply tests that existing behavior on
Mac.

Patch by Mike West <mkwst@chromium.org> on 2012-06-14
Reviewed by Chris Fleizach.

* platform/mac/accessibility/footer-expected.txt: Added.
* platform/mac/accessibility/footer.html: Added.
* platform/mac/accessibility/header-expected.txt: Added.
* platform/mac/accessibility/header.html: Added.

git-svn-id: svn://svn.chromium.org/blink/trunk@120383 bbb929c8-8fbe-4397-9dbb-9b2b20218538