-webkit-clip-path does not apply origin for polygon() https://bugs.webkit.org/show_bug.cgi?id=95656
Reviewed by Tim Horton.

Source/WebCore:

The polygon() shape function did not apply origin of bouding box on created path. The shape was
not moved to the correct position.

Tests: svg/clip-path/clip-path-shape-polygon-relative-expected.svg
       svg/clip-path/clip-path-shape-polygon-relative.svg

* rendering/style/BasicShapes.cpp:
(WebCore::BasicShapePolygon::path): Apply origin of bounding box.

LayoutTests:

Check that the origin of the bounding box gets applied to the clip path.

* svg/clip-path/clip-path-shape-polygon-relative-expected.svg: Added.
* svg/clip-path/clip-path-shape-polygon-relative.svg: Added.


git-svn-id: svn://svn.chromium.org/blink/trunk@127548 bbb929c8-8fbe-4397-9dbb-9b2b20218538