[Qt] GStreamer detection does not work when cross-compiling https://bugs.webkit.org/show_bug.cgi?id=96608
Patch by Simon Hausmann <simon.hausmann@nokia.com> on 2012-09-13
Reviewed by Kenneth Rohde Christiansen.

When cross-compiling it is not safe to just use system(pkg-config),
because there is a high probability that it will find packages in the
host system instead of in the target one. Instead use qmake's new
packagesExist() function that takes care of this delicate aspect of
cross compilation by setting PKG_CONFIG_SYSROOT_DIR and
PKG_CONFIG_LIBDIR accordingly.

* qmake/mkspecs/features/features.prf:

git-svn-id: svn://svn.chromium.org/blink/trunk@128422 bbb929c8-8fbe-4397-9dbb-9b2b20218538