REGRESSION (r127882): accessibility/spinbutton-value.html failing on GTK https://bugs.webkit.org/show_bug.cgi?id=96196
Patch by Joanmarie Diggs <jdiggs@igalia.com> on 2012-09-21
Reviewed by Martin Robinson.

The "regression" is that a new test was added but the support was missing
in the Gtk port for spin buttons.

Source/WebCore:

No new tests. Instead the new test which had been skipped was unskipped
as part of this fix.

* accessibility/gtk/WebKitAccessibleWrapperAtk.cpp:
(atkRole): Map SpinButtonRole to ATK_ROLE_SPIN_BUTTON
(getInterfaceMaskFromObject): Add SpinButtonRole to the roles implementing
the AtkValue interface.

Tools:

* DumpRenderTree/gtk/AccessibilityUIElementGtk.cpp:
(AccessibilityUIElement::valueDescription): Updated the FIXME comment to
indicate that this cannot be implemented until it is implemented in ATK.
URL of the newly-filed ATK bug included for reference.

LayoutTests:

* platform/gtk/TestExpectations: Unskip the new test.
* platform/gtk/accessibility/spinbutton-value-expected.txt: Added.

git-svn-id: svn://svn.chromium.org/blink/trunk@129247 bbb929c8-8fbe-4397-9dbb-9b2b20218538