[Chromium] Second part of Android's fast/ rebaselining.
Unreviewed rebaselining.

Second part of baselining in fast/. Equal to the first part (r129642),
most of the tests were failing because of the scrollbar of text selection
color. Bugs will be filed for all genuine failures, and every test has
been looked at and compared with Linux individually.

General info about the initial baselining is available in bug 96398.

* platform/chromium-android/fast/block/: 71 new results.
* platform/chromium-android/fast/multicol/: 93 new results.
* platform/chromium-android/fast/overflow/: 52 new results.
* platform/chromium-android/fast/repaint/: 67 new results.
* platform/chromium-android/fast/table/: 44 new results.
* platform/chromium-android/fast/text/: 22 new results.

git-svn-id: svn://svn.chromium.org/blink/trunk@129659 bbb929c8-8fbe-4397-9dbb-9b2b20218538