Follow-up to r129723 to once more allow parsing of scoped names in IDL files.
Patch by Takashi Sakamoto <tasak@google.com> on 2012-09-27
Reviewed by Kentaro Hara.

This functionality was supported by the old IDL parser but was lost in the rewrite of the parser
in r129723. It is being reinstated to unbreak clients that currently depend on it, but will likely
be removed in the future once those clients have a chance to adopt an approach that more closely
follows the WebIDL syntax.

* bindings/scripts/IDLParser.pm:
(parseDefinition):
(parseInheritance):
(parseImplementsStatement):
(parseExtendedAttribute):
(parseExtendedAttribute2):
(parseExtendedAttributeRest2):
(parseExtendedAttributeRest3):
(parseScopedNameListNoComma):
(parseNonAnyType):
(parseExceptionList):
(parseDefinitionOld):
(parseScopedName):
(parseAbsoluteScopedName):
(parseRelativeScopedName):
(parseScopedNameParts):
(parseScopedNameList):
(parseScopedNames):

git-svn-id: svn://svn.chromium.org/blink/trunk@129737 bbb929c8-8fbe-4397-9dbb-9b2b20218538