[EFL] WebContext::platformDefaultIconDatabasePath() should return a path to a file https://bugs.webkit.org/show_bug.cgi?id=100679
Patch by Thiago Marcos P. Santos <thiago.santos@intel.com> on 2012-10-31
Reviewed by Kenneth Rohde Christiansen.

This patch aligns the EFL implementation with other ports. Before
that, we had an inconsistence: the return value of the getter could
not be used on the setter (the setter expects a full path, not a
directory).

* UIProcess/API/efl/ewk_context.cpp:
(Ewk_Context::faviconDatabase):
* UIProcess/efl/WebContextEfl.cpp:
(WebKit::WebContext::platformDefaultIconDatabasePath):

git-svn-id: svn://svn.chromium.org/blink/trunk@133035 bbb929c8-8fbe-4397-9dbb-9b2b20218538