Unreviewed Qt/Windows build fix. https://bugs.webkit.org/show_bug.cgi?id=102787
Remove inclusion of QtQml, which drags in QtNetwork, which drags in qnetworksession.h, which
defines a member function called "interface". Since windows headers also define a pre-processor
macro called "interface", it has to do that. However we have now a dependency between combaseapi.h,
which #defines interface and is included earlier (before qnetworksession.h) and shlwapi.h, which _needs_
the interface macro and is included afterwards. Since this module-wide QtQml inclusion is unnecessary
and just slows down the build, this patch removes it.

* UIProcess/API/qt/qwebkittest_p.h:

git-svn-id: svn://svn.chromium.org/blink/trunk@135404 bbb929c8-8fbe-4397-9dbb-9b2b20218538