[CSS Regions] Crash when dispatching regionlayoutupdate https://bugs.webkit.org/show_bug.cgi?id=102944
Patch by Andrei Bucur <abucur@adobe.com> on 2012-11-27
Reviewed by Andreas Kling.

Source/WebCore:

The event dispatch function was incorrectly assuming the RenderNamedFlowThread is valid. Because the event is asynchronous it's
possible for the flow to be in the "NULL" state (the renderer is 0) while the regionlayoutupdate event is dispatched.

Test: fast/regions/webkit-named-flow-event-crash.html

* dom/WebKitNamedFlow.cpp:
(WebCore::WebKitNamedFlow::dispatchRegionLayoutUpdateEvent): Remove the ASSERT and don't dispatch the event if the flow is
    in the "NULL" state.

LayoutTests:

The test modifies the layout, forces a regionlayoutupdate event to be scheduled and then removes the content and the regions.
The event is dispatched on a named flow in the "NULL" state. Without this patch a crash occurs.

* fast/regions/webkit-named-flow-event-crash-expected.txt: Added.
* fast/regions/webkit-named-flow-event-crash.html: Added.

git-svn-id: svn://svn.chromium.org/blink/trunk@135853 bbb929c8-8fbe-4397-9dbb-9b2b20218538