[EFL][WK2] Duplicated WebPageGroup initialization https://bugs.webkit.org/show_bug.cgi?id=103229
Patch by Jinwoo Song <jinwoo7.song@samsung.com> on 2012-11-27
Reviewed by Gyuyoung Kim.

WebPageGroup does not need to be created in creating EwkView when the pageGroupRef is 0
because the default WebPageGroup(m_defaultPageGroup) is created in WebContext constructor.

As a side effect on the duplicated initialization, pageGroupID increases even when the EwkView
is created with default context.

* UIProcess/API/efl/ewk_view.cpp:
(createEwkView):

git-svn-id: svn://svn.chromium.org/blink/trunk@135973 bbb929c8-8fbe-4397-9dbb-9b2b20218538