[GTK] accessibility/placeholder.html is failing https://bugs.webkit.org/show_bug.cgi?id=98373
Patch by Joanmarie Diggs <jdiggs@igalia.com> on 2012-12-09
Reviewed by Martin Robinson.

The test was failing because the placeholder text was not supported and
AccessibilityUIElement::stringAttributeValue() was not implemented.

Source/WebCore:

No new tests; instead the failing test was unskipped.

* accessibility/atk/WebKitAccessibleWrapperAtk.cpp:
(webkitAccessibleGetAttributes): Add the placeholder text as an AtkAttribute
of the AtkObject, as is done in Gtk+ 3.

Tools:

* DumpRenderTree/gtk/AccessibilityUIElementGtk.cpp:
(coreAttributeToAtkAttribute): New convenience method to convert WebCore attribute
names into AtkObject attribute names
(AccessibilityUIElement::stringAttributeValue): implemented

LayoutTests:

* platform/gtk/TestExpectations: Unskip the failing test

git-svn-id: svn://svn.chromium.org/blink/trunk@137099 bbb929c8-8fbe-4397-9dbb-9b2b20218538