REGRESSION (r118735): svg/dom/complex-svgView-specification.html, svg/dom/SVGViewSpec.html, svg/dom/viewspec-parser.html failing on GTK Linux 64-bit Release https://bugs.webkit.org/show_bug.cgi?id=87734
Reviewed by Martin Robinson.

Source/WebCore: 

Don't modify ResourceRequest's URL to have the value of the (encoded)
URI of the SoupMessage. The only exception is in the case of cancelled
SoupMessage - the URL of the ResourceRequest should then be null.

No new tests - existing tests should show no regressions.

* platform/network/soup/ResourceRequestSoup.cpp:
(WebCore::ResourceRequest::updateFromSoupMessage):

Tools: 

When willSendRequest should return null in testing, cancel the SoupMessage
instead of setting the request's URI to about:blank.

* DumpRenderTree/gtk/DumpRenderTree.cpp:
(willSendRequestCallback):


git-svn-id: svn://svn.chromium.org/blink/trunk@137198 bbb929c8-8fbe-4397-9dbb-9b2b20218538