[CSS Exclusions] Floats should respect shape-inside on exclusions https://bugs.webkit.org/show_bug.cgi?id=89261
Patch by Bem Jones-Bey <bjonesbe@adobe.com> on 2012-12-17
Reviewed by David Hyatt.

Source/WebCore:

Position floats properly with respect to an exclusion shape. Note that
this will not attempt to position floats in a polygon that has
multiple segments. In the multiple segment case, the floats will be
positioned as if the exclusion did not exist.

Updated an existing test case to test for this.
Test: fast/exclusions/shape-inside/shape-inside-floats-simple.html

* rendering/RenderBlock.cpp:
(WebCore::RenderBlock::computeLogicalLocationForFloat): Adjust the
right and left offsets to take into account the right and left offset
contributed by the exclusion shape.

LayoutTests:

Update test to properly test floats with shape-inside.

* fast/exclusions/shape-inside/shape-inside-floats-simple-expected.html:
* fast/exclusions/shape-inside/shape-inside-floats-simple.html:

git-svn-id: svn://svn.chromium.org/blink/trunk@137920 bbb929c8-8fbe-4397-9dbb-9b2b20218538