[Qt] Remove misspelled ENABLE(SPEECH_INPUT) guard https://bugs.webkit.org/show_bug.cgi?id=105683
Patch by Laszlo Gombos <l.gombos@samsung.com> on 2013-01-29
Reviewed by Simon Hausmann.

The ENABLE(SPEECH_INPUT) guard is only used in one location
in the source tree and as such it is always 0.

I believe that this guard is a left over and is not needed any more
as the code guarded is not guarded anywhere else.

* WebProcess/qt/WebProcessQt.cpp:
(WebKit::WebProcess::platformInitializeWebProcess):

git-svn-id: svn://svn.chromium.org/blink/trunk@141092 bbb929c8-8fbe-4397-9dbb-9b2b20218538