Current error reporting method used by check-layout.js should not affect subsequent sub-tests using checking data-offset-y. https://bugs.webkit.org/show_bug.cgi?id=105407
Patch by Pravin D <pravind.2k4@gmail.com> on 2013-01-29
Reviewed by Tony Chang.

When a testcase is processed by check-layout.js, the result is written just after the node being processed. This causes
offsetTop of subsequent sub-testcases to altered. If however if we process the nodes in the reverse order in which they
appear in the DOM tree, the result of node being processed will not affect the offsetTop of susequent nodes.

* resources/check-layout.js:

git-svn-id: svn://svn.chromium.org/blink/trunk@141147 bbb929c8-8fbe-4397-9dbb-9b2b20218538