HRTFElevation segfault if a null AudioBus is returned by loadPlatformResource https://bugs.webkit.org/show_bug.cgi?id=108504
Patch by Jesus Sanchez-Palencia <jesus.palencia@openbossa.org> on 2013-02-01
Reviewed by Kenneth Russell.

Fixed a segfault by checking if the returned AudioBus is null or not.
If it is, we should early return in the same way it was being done already
in HRTFElevation::calculateKernelsForAzimuthElevation();

No new tests, covered by existing tests.

* platform/audio/HRTFElevation.cpp:
(WebCore::getConcatenatedImpulseResponsesForSubject):

git-svn-id: svn://svn.chromium.org/blink/trunk@141611 bbb929c8-8fbe-4397-9dbb-9b2b20218538