Web Inspector: fix "DOM Exception 8" when deleting lines containing decoration in DTE. https://bugs.webkit.org/show_bug.cgi?id=108689
Patch by Andrey Lushnikov <lushnikov@chromium.org> on 2013-02-01
Reviewed by Alexander Pavlov.

Add a check that decoration element is still added to the line before trying to delete it. This
won't be true if the whole line is deleted in contentEditable,
and this is the reason for the exception to be thrown.

No new tests.

* inspector/front-end/DefaultTextEditor.js:
(WebInspector.TextEditorMainChunk.prototype.removeAllDecorations):

git-svn-id: svn://svn.chromium.org/blink/trunk@141673 bbb929c8-8fbe-4397-9dbb-9b2b20218538