Scrollbars misplaced with accelerated compositing for overflow scroll https://bugs.webkit.org/show_bug.cgi?id=108625
Reviewed by Simon Fraser.

Source/WebCore:

Scrollbars require their own layer if overflow scroll is composited,
otherwise the scrollbars would be rendered on the content layer and
not fixed to the viewport.

* rendering/RenderLayerBacking.cpp:
(WebCore::RenderLayerBacking::requiresHorizontalScrollbarLayer):
(WebCore::RenderLayerBacking::requiresVerticalScrollbarLayer):
(WebCore::RenderLayerBacking::requiresScrollCornerLayer):

LayoutTests:

Update the results for the one test that explicitly set accelerated compositing for overflow scroll,
and used to have bad results for non-chromium. The new baselines are almost identical to chomium baseline.

* platform/mac/compositing/overflow/composited-scrolling-creates-a-stacking-container-expected.txt:
* platform/qt/compositing/overflow/composited-scrolling-creates-a-stacking-container-expected.txt:


git-svn-id: svn://svn.chromium.org/blink/trunk@142140 bbb929c8-8fbe-4397-9dbb-9b2b20218538