document.activeElement should not return a non-focusable element https://bugs.webkit.org/show_bug.cgi?id=86707
Reviewed by Hajime Morita.

Source/WebCore:

This is based on a patch by Arpita Bahuguna.

Test: fast/dom/HTMLDocument/set-focus-on-valid-element.html

* dom/Document.cpp:
(WebCore::Document::setFocusedNode):
Added check for verifying that the node to be focused is
focusable. However, this check should be skipped for HTMLPlugInElement
because it has special behavior.

LayoutTests:

* fast/dom/HTMLDocument/set-focus-on-valid-element-expected.txt: Added.
* fast/dom/HTMLDocument/set-focus-on-valid-element.html: Added.
This test verifies that document.activeElement does not return an
invalid or non-focusable element.


git-svn-id: svn://svn.chromium.org/blink/trunk@142234 bbb929c8-8fbe-4397-9dbb-9b2b20218538