[EFL][WKTR] Regression(r141836) fast/dom/Window/mozilla-focus-blur.html started failing https://bugs.webkit.org/show_bug.cgi?id=109438
Reviewed by Kenneth Rohde Christiansen.

Some refactoring in r141836 caused the view not to get focus if the focused
frame is not the main one. The idea of the code was to remove focus from the
view if the focused frame was not the main one, and then focus the view again.
However, after the refactoring, the second step never happened: Focus was
removed but not given again.

* WebKitTestRunner/efl/PlatformWebViewEfl.cpp:
(WTR::PlatformWebView::focus):

git-svn-id: svn://svn.chromium.org/blink/trunk@142477 bbb929c8-8fbe-4397-9dbb-9b2b20218538