Fix TileCache tile size when zoomed on slow-scrolling site https://bugs.webkit.org/show_bug.cgi?id=110289
Source/WebCore:

Reviewed by Tim Horton.

In slow scrolling mode, we use the tile coverage rect as the tile size
(so a single tile covers the viewport). When zoomed, the tile coverage rect
shrinks relative to the TileCache, so scale the tile size back up.

Test: platform/mac-wk2/tiled-drawing/tile-size-slow-zoomed.html

* platform/graphics/ca/mac/TileCache.mm:
(WebCore::TileCache::tileSizeForCoverageRect):

LayoutTests:

Reviewed by Tim Horton.

Test that dumps tile coverage in a slow-scrolling, zoomed page.

* platform/mac-wk2/tiled-drawing/tile-size-slow-zoomed-expected.txt: Added.
* platform/mac-wk2/tiled-drawing/tile-size-slow-zoomed.html: Added.

git-svn-id: svn://svn.chromium.org/blink/trunk@143423 bbb929c8-8fbe-4397-9dbb-9b2b20218538