[chromium] Re-enable the fixed version of freetype on linux https://bugs.webkit.org/show_bug.cgi?id=110542
Reviewed by Tony Chang.

Tools:

Re-land the change in r143794 to switch to building with a fixed
version of FreeType for just DumpRenderTree. This allows us to share
one version of pixel baselines for both Lucid and Precise, and run the
layout tests in a stock Precise install (instead of a Lucid chroot).
Going forward, this approach isolates us from os-level changes to the
version of FreeType. Note that Chromium itself still uses the system
FreeType and is unaffected by this change.

This setting can still be turned off (at compile/gyp-time) by
manually setting use_custom_freetype=0.

* DumpRenderTree/DumpRenderTree.gyp/DumpRenderTree.gyp:

LayoutTests:

Re-add entries for two tests that pass on Lucid but fail on Precise
even with the Freetype change. This isn't worth adding in
Lucid and Precise keywords for just these two tests.

* platform/chromium/TestExpectations:

git-svn-id: svn://svn.chromium.org/blink/trunk@144116 bbb929c8-8fbe-4397-9dbb-9b2b20218538