Disable some unsound DFG DCE https://bugs.webkit.org/show_bug.cgi?id=110948
Reviewed by Michael Saboff.
        
DCE of bitops is not sound since the bitops might call some variant of valueOf.
        
This used to work right because ValueToInt32 was MustGenerate. From the DFG IR
standpoint it feels weird to make ValueToInt32 be MustGenerate since that node is
implemented entirely as a pure conversion. If we ever gave the DFG the ability to
do effectful bitops, we would most likely implement them as special nodes not
related to the ValueToInt32 and bitop nodes we have now.
        
This change is performance neutral.

* dfg/DFGNodeType.h:
(DFG):



git-svn-id: svn://svn.chromium.org/blink/trunk@144219 bbb929c8-8fbe-4397-9dbb-9b2b20218538