REGRESSION(r140907): Incorrect baseline for cells with media content during load https://bugs.webkit.org/show_bug.cgi?id=108357
Reviewed by Julien Chaffraix.

Source/WebCore:

If a cell has replaced content, the intrinsic height of its content can change between layouts. If that's the case then the intrinsic padding we used
for layout (the padding required to push the contents of the cell down to the row's baseline) is included in the new height and baseline and makes both
of them wrong. So if a cell's content's intrinsic height has changed push the new content up into the intrinsic padding and relayout so that the rest of
table and row layout can use the correct baseline and height for this cell.

Tests: fast/css/vertical-align-baseline-rowspan-012.html
       http/tests/css/vertical-align-baseline-after-image-load-2.html
       http/tests/css/vertical-align-baseline-after-image-load-3.html
       http/tests/css/vertical-align-baseline-after-image-load.html

* rendering/RenderTableCell.cpp:
(WebCore::RenderTableCell::layout):
* rendering/RenderTableCell.h:
(WebCore::RenderTableCell::isBaselineAligned):
* rendering/RenderTableSection.cpp:
(WebCore::RenderTableSection::calcRowLogicalHeight):
(WebCore::RenderTableSection::layoutRows):

LayoutTests:

* fast/css/vertical-align-baseline-rowspan-012-expected.html: Added.
* fast/css/vertical-align-baseline-rowspan-012.html: Added.
* http/tests/css/vertical-align-baseline-after-image-load-2-expected.html: Added.
* http/tests/css/vertical-align-baseline-after-image-load-2.html: Added.
* http/tests/css/vertical-align-baseline-after-image-load-3-expected.html: Added.
* http/tests/css/vertical-align-baseline-after-image-load-3.html: Added.
* http/tests/css/vertical-align-baseline-after-image-load-expected.html: Added.
* http/tests/css/vertical-align-baseline-after-image-load.html: Added.

git-svn-id: svn://svn.chromium.org/blink/trunk@145305 bbb929c8-8fbe-4397-9dbb-9b2b20218538