Conformance Test 1.0.3 (Beta) function: bufferData undefined value failed. https://bugs.webkit.org/show_bug.cgi?id=111641
Patch by Jason Anderssen <janderssen@gmail.com> on 2013-03-10
Reviewed by Dean Jackson.

The WebGL specification requires that a size of 0 is not valid. In javascript, passing in undefined
as a parameter to a long long is the same as passing in 0, so we must check for this incorrect
value and fail.
The test suite in Kronos 1.0.3 failed, test to verify conformance is as follows:
https://www.khronos.org/registry/webgl/sdk/tests/conformance/more/functions/bufferDataBadArgs.html.

* html/canvas/WebGLRenderingContext.cpp:
(WebCore::WebGLRenderingContext::bufferData):
Synthesize error and returned if size is 0.

git-svn-id: svn://svn.chromium.org/blink/trunk@145334 bbb929c8-8fbe-4397-9dbb-9b2b20218538