[V8] IndexedDB: Exceptions thrown inconsistently for non-cloneable values https://bugs.webkit.org/show_bug.cgi?id=113091
Reviewed by Kentaro Hara.

Source/WebCore:

The exception thrown by SerializedScriptValue into the JS engine is not
observable by ScriptState. (We should fix that, but it appears non-trivial.)
Ask the SerializedScriptValue directly if it failed to clone. If so, don't
set an exception - one was already set so let that be processed normally.

Test: storage/indexeddb/clone-exception.html

* Modules/indexeddb/IDBObjectStore.cpp:
(WebCore::IDBObjectStore::put):

LayoutTests:

* storage/indexeddb/clone-exception-expected.txt: Added.
* storage/indexeddb/clone-exception.html: Added.


git-svn-id: svn://svn.chromium.org/blink/trunk@147241 bbb929c8-8fbe-4397-9dbb-9b2b20218538