Fold ScrollingCoordinatorChromium into ScrollingCoordinator and devirtualize
This folds the ScrollingCoordinatorChromium subclass into the base and devirtualizes it. This also removes
most (but not all) of the ScrollingCoordinator functionality that we don't use and
deletes many files in page/scrolling/ that we don't use. There are many more cleanups
possible here - this is just a minimal first cut that doesn't (intentionally) change
any behavior.

Review URL: https://codereview.chromium.org/13665002

git-svn-id: svn://svn.chromium.org/blink/trunk@147653 bbb929c8-8fbe-4397-9dbb-9b2b20218538