ASSERTION FAILED: resultAnimationElement->m_animatedType
Merge http://trac.webkit.org/changeset/147581

When the order of animation elements change within an animation
group, we can't assume that the resultElement is the same as it
was when the animation was frozen. Make sure to reset the
animation type for the resultElement before updating the
animation.

Test: svg/animations/animation-dependency-crash.html

R=pdr@chromium.org
BUG=141326

Review URL: https://codereview.chromium.org/13724009

git-svn-id: svn://svn.chromium.org/blink/trunk@147679 bbb929c8-8fbe-4397-9dbb-9b2b20218538