Don't recalc style for pseudo elements when it's not needed.
Previously we'd go through createPseudoElementIfNeeded for every Element
while walking down the tree during style recalc which is not needed. We
only need to updatePseudoElements if our style has changed.

This removes a tiny 0.1% on our animation benchmark, but is also
important for correctness. We should never call recalcStyle on
elements that don't actually need it.

R=eseidel@chromium.org,ojan@chromium.org

Review URL: https://codereview.chromium.org/13724017

git-svn-id: svn://svn.chromium.org/blink/trunk@147696 bbb929c8-8fbe-4397-9dbb-9b2b20218538