Make -webkit-line-break:normal correctly paint the selection when you have only selected a soft-wrap line break. This is in preparation for removing -webkit-line-break entirely.
All the paint code does the right thing, we just need to have
the InlineTextBox correctly think it is selected and to have the
correct selectionState.

Also, fix InlineTextBox::isSelected to use ints instead of LayoutUnits since they refer to number of characters. It was incorrect to ever make these LayoutUnits.

Unfortunately, the new test needs to be a pixel test because getting the right selection requires actual text, not just inline-block divs.

Review URL: https://codereview.chromium.org/13820027

git-svn-id: svn://svn.chromium.org/blink/trunk@148134 bbb929c8-8fbe-4397-9dbb-9b2b20218538