Simplify the code generated for allocations on non-Mac platforms.
This is the remainder from https://codereview.chromium.org/14006002, part of
which was not landed in the end due to a few classes in WTF and WebCore wanting
to mix "malloc" with "delete".

For some reason, I'm not seeing the original 4k saving in Release DumpRenderTree
size any more. But a Debug DumpRenderTree went from 107MB -> 104MB on Linux x64.

The problem of some classes wanting to mix "malloc" with "delete" was handled
by simply identifying them and tagging them specially, so that we override
operator new and operator delete even on non-Mac platforms.

R=eseidel@chromium.org
BUG=none
TEST=Tests still pass in debug build.

Review URL: https://codereview.chromium.org/14163003

git-svn-id: svn://svn.chromium.org/blink/trunk@148503 bbb929c8-8fbe-4397-9dbb-9b2b20218538