CSP: Allow injection of inline script from isolated world.
Extensions' content scripts, for better or worse, love to inject inline script
in order to interact with a page's main world. If that page is protected by
Content Security Policy, the injection fails, and the extension's author and
users are sad.

This patch allows script running in an isolated world to bypass a page's CSP
in order to inject an inline script.

BUG=181602

Review URL: https://codereview.chromium.org/13986007

git-svn-id: svn://svn.chromium.org/blink/trunk@148805 bbb929c8-8fbe-4397-9dbb-9b2b20218538