DevTools: [OverviewGrid] Refactor overview scaling code.
The idea of the patch is to avoid conversion back and forth from percents to pixels coords.
The overview window positions left and right borders in percents.
Also it uses windowLeft and windowRight values for interacting with the embedder.
So it would be simpler to use windowLeft/windowRight among all the methods
and convert pixelLeft/pixelRight into windowLeft/windowRight as earlier as possible.

BUG=none
TEST=none
R=pfeldman, alph, yurys@chromium.org

Review URL: https://codereview.chromium.org/14404010

git-svn-id: svn://svn.chromium.org/blink/trunk@148845 bbb929c8-8fbe-4397-9dbb-9b2b20218538