Add an intermediate function to generate 2 paint-order lists.
The first represents the list before opting in to composited scrolling, and the
second one shows what the state would be if we decided to opt in. These two
lists will be necessary for our decision of whether or not to opt in, because we
have to verify that this decision will not change the paint order. This also
includes some plumbing and a layout test to ensure that the paint-order lists
generated in this function conform to the stacking-order as verified by existing
hit-testing code.

R=jchaffraix

BUG=222016

Review URL: https://codereview.chromium.org/13467028

git-svn-id: svn://svn.chromium.org/blink/trunk@149028 bbb929c8-8fbe-4397-9dbb-9b2b20218538