Fix button baseline calculations for Mac with dynamic changes
Also convert a button test to a reftest and add a new button test.
    
The new test is to test baselines, which is useful because in the past I
have relied on button-generated-content to verify baseline behavior.
    
The conversion to a reftest is what found the bug that I fixed in
RenderButton.cpp -- the button would otherwise be positioned wrong
after the dynamic changes.
    
R=ojan@chromium.org
BUG=

Review URL: https://chromiumcodereview.appspot.com/14717002

git-svn-id: svn://svn.chromium.org/blink/trunk@149908 bbb929c8-8fbe-4397-9dbb-9b2b20218538