Go back to having full_results.json actually list every test.
Also, include whether a failure is unexpected in the summarized results
and use that value in buildbot_results.print_unexpected_results. This way
we avoid duplicating the logic of whether a failure is unexpected.

This lets the tools know what happened with each test on each run.
Specifically, it allows the flakiness dashboard to list slow tests
even if they always pass.

Also, have result_was_expected always return true if the result is Skip.
When we do batched running of tests (e.g. on the debug bots) the result
is Skip even though it's not [ Skip ] in TestExpectations.

BUG=238159

Review URL: https://chromiumcodereview.appspot.com/14837003

git-svn-id: svn://svn.chromium.org/blink/trunk@150000 bbb929c8-8fbe-4397-9dbb-9b2b20218538