Remove dead ContextMenu code from Blink core
Most of ContextMenuController was devoted to building up context menus
and handling actions for those menus it had built up, but in Blink/Chromium,
those menus are almost entirely dropped on the floor by the embedder.
Only those menu items tagged as "custom" are preserved. And these are added
elsewhere, by ContextMenuProvider instances; the only currently-existing
one belongs to the web inspector.

As part of this cleanup, renamed ContextMenuClient::customizeMenu() to
showContextMenu(), since it's a call to that method from the controller
which actually causes a context menu to appear. Also tightened up that
method, which has no reason to transfer ownership (or even allow mutation)
of its ContextMenu to the embedder.

R=darin

Review URL: https://chromiumcodereview.appspot.com/14795011

git-svn-id: svn://svn.chromium.org/blink/trunk@150375 bbb929c8-8fbe-4397-9dbb-9b2b20218538