Clear SVGInlineTextBox fragments when the text changes.
This patch modifies SVGInlineTextBox::dirtyLineBoxes to clear all
following text boxes when invoked. Typically this method is called
when the underlying text string changes, and that change needs to
be propagated to all the boxes that use the text beyond the point
where the text is first modified.

Also cleans up virtual, OVERRIDE and FINAL for SVGRootInlineBox, which was all messed up.

R=inferno@chromium.org,leviw@chromium.org
BUG=233848

Review URL: https://chromiumcodereview.appspot.com/15183002

git-svn-id: svn://svn.chromium.org/blink/trunk@150456 bbb929c8-8fbe-4397-9dbb-9b2b20218538