remove DriverProxy._running_drivers from webkitpy.layout_tests.port.
The DriverProxy class had the concept of keeping a pool of started
DRTs around (indexed by command line) so that we didn't necessarily
need to restart DRT when the command line for a test changed (e.g., when
starting to run tests in a virtual suite). However, the pool was per-worker
(so not shared across the machine) and we didn't cap the number of processes
started, so we would end up with a whole bunch of DRTs (3x-4x the number of
workers started). If anything, this seemed to make the tests slightly slower
and probably added to memory pressure on the machine.

This change removes the pool code; now we'll simply restart DRT when the
command line changes. With this change, we can actually fold the restarting
code into Driver itself (and get rid of DriverProxy), but I'll leave that
for a subsequent patch just to make things easier to follow.

BUG=none
R=ojan@chromium.org

Review URL: https://codereview.chromium.org/15001046

git-svn-id: svn://svn.chromium.org/blink/trunk@150466 bbb929c8-8fbe-4397-9dbb-9b2b20218538