Remove frame flattening from rendering code.
Chromium never enables frame flattening; we can simplify frame rendering by
removing the various paths that only execute if flattening is enabled.

Currently, this patch leaves RenderFrameBase as an empty shell. We will
remove that entirely in a subsequent patch.

This is a reland of a reland of a reland. It's totally going to work this time.
Carrying over Ojan's LGTM from the original patch: https://codereview.chromium.org/13877010

BUG=231902
TBR=ojan

Review URL: https://chromiumcodereview.appspot.com/14614007

git-svn-id: svn://svn.chromium.org/blink/trunk@150495 bbb929c8-8fbe-4397-9dbb-9b2b20218538