Use full_results.json to popuplate results.json files.
-Just for layout test for now since we don't generate full_results.json files
for gtests.
-Convert full_results.json into the incremental results format and then continue
to merge it in as normal.
-Get rid of the "incremental" post parameter. We've long not used it.
-For full_results.json, even when we merge it, we still want to upload the raw file
as that raw data is used by other tools.
-Update the django version to 1.3. This doesn't affect anything in practice, it
just makes the unittests work with the latest appengine SDK.
-Include a char to failure map in the results json so that the flakiness dashboard
can use it and we can avoid having this mapping copy-pasted everywhere.
-Include the expected results for each listed test. We should change run-webkit-tests
to include any test that has an expected result other than pass. Then this can be used
to show the expectations for a test in the flakiness dashboard instead of the current code
that duplicates all the python-side logic.
-Simplify _is_directory now that the corrupted results.json files are gone.
-Move _normalize_results call down into merge so that it always gets called.
-Have update_file always write out the results of the merge call. That way
we can know that the passed in JSON was properly normalized.
-Get rid of the code that converts the JSON from version 3 to version 4.
Everything's been at version 4 for a long time now.

R=dpranke@chromium.org

Review URL: https://codereview.chromium.org/15094016

git-svn-id: svn://svn.chromium.org/blink/trunk@150597 bbb929c8-8fbe-4397-9dbb-9b2b20218538