Dirty the first RootInlineBox, always
The current code does not dirty the first RootInlineBox in a text
run if the change affects the first InlineTextBox. It is unclear
why it was not considered necessary. However, several problems with
rtl text editing have revealed the need to dirty the root box
regardless of whether we have already dirtued the line. I believe
this is because rtl text layout creates line boxes where they would
otherwise not exist.

This patch also modifies the fix for an existing SVG rtl bug, which
avoids duplciating the same clear operations. Apparently this also
fixes a layout test that previously had a failing result.

BUG=230726

Review URL: https://chromiumcodereview.appspot.com/15271005

git-svn-id: svn://svn.chromium.org/blink/trunk@150718 bbb929c8-8fbe-4397-9dbb-9b2b20218538