Make FrameLoader::reload always use the active document loader.
FrameLoader::reload used the m_documentLoader, even if the loader was in the
FrameStateProvisional state.  This means that if the FrameLoader is sent a
reload call while it is loading it's first document, the reload will fail
and instead load "about:blank" since m_documentLoader is empty.  This change
makes reload use the activeDocumentLoader, which means reload works correctly
if the frame is in the FrameStateProvisional state.

This bug exposes itself on Chrome for Android where the browser process can
restore a swapped-out tab, then immediatly call reload on this tab before the
FrameLoader has finished performing the previous load (that was due to the tab
restored from the swapped-out state).

BUG=chromium:115842
TEST=Ran layout webkit layout tests. Built and ran on Chrome for Android.

Review URL: https://chromiumcodereview.appspot.com/15288006

git-svn-id: svn://svn.chromium.org/blink/trunk@150935 bbb929c8-8fbe-4397-9dbb-9b2b20218538