Merge part of r200655: Fix handling of data type throttling
This branch is susceptible to only the second of the two issues that
this patch tries to fix.  This merge contains only the changes required
to fix that issue.

Original commit message for r200655 follows.

=============================================

This change attempts to fix two semi-related issues in data type
throttling that were introduced in r194766. Data type throttling has
always had a few rough edge cases, but that commit introduced some
issues that must be fixed.

1:
Prior to r194766, data type throttling would block sync cycles only if
the source was LOCAL and all types were throttled. This meant that a
refresh request or notification could override data type throttling, if
it arrived at the right time.

That change made this behaviour much more deterministic. It would fail
to perform any sync cycles as long as the only local nudges that were
currently tracked were for throttled types. Even notificiations and
local refresh requests for these types will not be sufficient to force a
sync cycle.

This change tries to find a better balance. It will allow sync cycles
in the presence of unthrottled data types if one of the following is
true:
- There is an unserviced invalidation for any type.
- There is an unservice refresh request for any type.
- There is a local nudge for any non-throttled type.

2:
The other problem has to do with exponential backoff in the presence of
per-type throttling. In r194766, we made the assumption that we can
early-exit from attempted sync cycle only if there was some other reason
to perform a sync cycle later. Common causes of early-exit include
throttling, exponential backoff, and network connectivity issues, all of
which should be able to get back into a normal state on their own.

However, one of those early-exit causes is data type throttling. In
that case, there is no timer or notification that will restart sync once
the issue is resolved. This is a problem because DoNudgeSyncSessionJob
assumes that it can return without scheduling a retry in the "if
(!CanRunNudgeJobNow())" case, while the code that manages the
exponential backoff assumes the function that invokes
DoNudgeSyncSessionJob, TryCanaryJob(), will always either clear or
restart the wait interval.

To fix this issue, this change adds a small function to invoke the
exponential backoff retry callback, and gives it responsibility for
making sure the WaitInterval is always managed correctly.

BUG=239693

Review URL: https://chromiumcodereview.appspot.com/14710006

R=tim@chromium.org

Review URL: https://codereview.chromium.org/15518003

git-svn-id: http://src.chromium.org/svn/branches/1500/src@201825 4ff67af0-8c30-449e-8e8b-ad334ec8d88c