Refactor viewport initialization logic out of WebViewImpl.
(Second commit attempt)

This centralizes and brings some structure to viewport initialization,
expressing it as a cascade of constraints expressed as
ViewportAttributes objects.  When any of the constraints changes, a
dirty bit is set, and the page scale factor and fixed-layout size is
recalculated on the next layout.

This refactor mostly maintains existing behavior, but also includes two bugfixes
(with new tests for them):
1. Page scale is now reset to minimum when the content width increases,
even if it was reset in the past.
2. setIgnoreViewportTagMaximumScale now also ignores the minimum (I'll
need to rename this WebKit API).

NOTRY=true
BUG=176998,161695,165688

Review URL: https://chromiumcodereview.appspot.com/14813025

git-svn-id: svn://svn.chromium.org/blink/trunk@151134 bbb929c8-8fbe-4397-9dbb-9b2b20218538