Use failures_by_type and failure_map in the flakiness dashboard code.
-Get rid of all the code that knows about the short version of the failure type
(e.g. 'T'-->TIMEOUT). This means that the short-version are purely implementation
detail of the results json. I plan to change the short versions to ints soon.
-Use failures_by_type for all the failure counts. fixableCount, allFixableCount
and fixableCounts will soon be removed from the JSON.
-Add getTotalTestCounts to dashboard_base so that it can be reused between
timeline_explorer and aggregate_results. It's a bit unfortunate to dump this here,
but I couldn't think of a better place. Down the road, we should probably
have a class that abstracts away the results json format and this function
should go there.
-Add a showOutliers flag to aggregate_results. We've always had this functionality.
This just lets you toggle it on/off.
-Simplify the rawValues view of aggregate_results to have a single table per builder.
The multiple tables only made sense when we tracked deferred/wontfix tests.
-Get rid of expectationsMap now that we use the failure_map from the json.
-Start using constants for the failure types (e.g. PASS instead of 'PASS').
We should eventually migrate all the code over to these constants.
-Remove slowestNonTimeoutCrashTime. This hasn't been used for a while.
-Allow any valid name string for "result" query parameters. The list of
valid result types is in the JSON. We don't really need to validate this
other than as a saftey agaist potential XSS.

R=jparent@chromium.org

Review URL: https://codereview.chromium.org/16032005

git-svn-id: svn://svn.chromium.org/blink/trunk@151148 bbb929c8-8fbe-4397-9dbb-9b2b20218538