Do not pass security origin to V8 when collecting CPU profiles
Now that console.profiles was removed there is no need to pass
security origin to V8 so that it filter out callstacks based
on the security origin. Instead full CPU profile should be returned
and it is up to the DevTools client to analyze it and perform
filtering.

BUG=None
R=alph@chromium.org

Review URL: https://codereview.chromium.org/15766004

git-svn-id: svn://svn.chromium.org/blink/trunk@151196 bbb929c8-8fbe-4397-9dbb-9b2b20218538