Add wtf_unittests target that links unit tests outside DLL
This moves the wtf unittests out of the webkit_unit_tests target and into
a dedicated gtest executable. This target builds and links much faster since
it's tiny and moves the tests outside the DLL. This target is in all_webkit,
but annoyingly we'll have to configure the bots to run this step.

R=abarth@chromium.org, darin@chromium.org

Review URL: https://codereview.chromium.org/16066003

git-svn-id: svn://svn.chromium.org/blink/trunk@151305 bbb929c8-8fbe-4397-9dbb-9b2b20218538