Clean up the way layout tests force elements to opt in/out of composited scrolling.
Currently, we're toggling a global setting, but we should really be affecting
the opt in decision for individual elements. This patch cleans that up.

This is part of a larger plan to finish the opt-in checks. Here's the aggregate patch (it's getting stale, so YMMV): https://codereview.chromium.org/14741004/

The reason this patch is necessary, is that in subsequent patches, we will be reducing the number of times we update composited scrolling state; explicitly forcing a particular layer (and only one layer) to opt in or out will allow us to guarantee that the layer's state (in particular, its z-order lists) will be updated, even if we've optimized/reduced the frequency with which we update the composited scrolling state.

R=hartmanng@chromium.org, jchaffraix@chromium.org
BUG=None

Review URL: https://chromiumcodereview.appspot.com/14858004

git-svn-id: svn://svn.chromium.org/blink/trunk@151433 bbb929c8-8fbe-4397-9dbb-9b2b20218538