[CSSRegions] Constrain auto-height region computation in a different way
Since the region is a render block, we can use RenderBox::constrainContentBoxLogicalHeightByMinMax
instead of RenderBox::computeReplacedLogicalHeightRespectingMinMaxHeight to better reflect the fact
that the region is not a replaced element.
Covered by existing regions tests, since when replaced-based the region had an intrinsic height of 0 and now, block-based,
it does not have any children.

BUG=243651
NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/16033003

git-svn-id: svn://svn.chromium.org/blink/trunk@151442 bbb929c8-8fbe-4397-9dbb-9b2b20218538