Remove CaptionUserPreferences class from PageGroup
This class was almost completely unused in Blink. Most of what was
there was test scaffolding that itself was unused in Blink (because
the tests that exercised it either didn't depend on their use of it
or because those tests always failed or always timed out).

In the few cases where CaptionUserPreferences was used, those pieces
of functionality have been inlined into their callers.

This change also removes the aforementioned tests that either had
always-failing expectations or were marked as 'Timeout'. These
tests were exercising AppleWebKit-specific behavior.

BUG=237595
R=vcarbune@chromium.org

Review URL: https://codereview.chromium.org/15740032

git-svn-id: svn://svn.chromium.org/blink/trunk@151570 bbb929c8-8fbe-4397-9dbb-9b2b20218538