Fix RenderLayer::collectLayers logic bug.
In collectLayers, if a layer is normal flow only, we will not add it
to the z-order lists. The trouble is that
RenderLayer::shouldBeNormalFlowOnly depends on 
RenderLayer::needsCompositedScrolling. This means that the result of
collectLayers depends on the opt-in decision of other layers,
something which must not happen: when we are determining opt-in, we
must never depend on the opt-in decision for another layer. This CL
makes this function opt-in agnostic when it needs to be.

R=jchaffraix

BUG=238282

Review URL: https://chromiumcodereview.appspot.com/14999005

git-svn-id: svn://svn.chromium.org/blink/trunk@151665 bbb929c8-8fbe-4397-9dbb-9b2b20218538