Mobile Facebook should use 913 kB less memory
CachedScript has two representations for scripts: an encoded representation,
which is a copy of the data we received over the network, and a decoded
representation, which is a String. Prior to this CL, we could eagerly clear out
the String representation to avoid retaining two copies.

Unfortunately, retaining the encoded copy of the data isn't the right choice
because V8 needs to retain the String anyway to back
Function.prototype.toString. If we retain the String copy in the MemoryCache,
then we are able to share that data with V8.

Retaining the encoded version is actually pathologically bad in the case where
we get hits in the MemoryCache. For each hit, we'd create a new decoded String
and V8 would retain that string. That meant that after N hits, we'd end up with
N+1 copies of the data!

This CL saves 913 kB of memory on Mobile Facebook because we no longer need to
retain 913 kB of SharedBuffers for CachedScripts.

== Before ==

JavaScript
----------
Count: 13
Size: 1863554
DecodedSize: 912575

== After ==

JavaScript
----------
Count: 13
Size: 949923
DecodedSize: 0

Note: In these stats we now represent the String representation as the "encoded
size" of the resource because we no longer have the underlying SharedBuffer.
I've also hacked the stats for the "before" case to count the String
representation in the DecodedSize because the memory is still retained, just
not by the MemoryCache directly.

R=eseidel,japhet

Review URL: https://chromiumcodereview.appspot.com/16339003

git-svn-id: svn://svn.chromium.org/blink/trunk@151779 bbb929c8-8fbe-4397-9dbb-9b2b20218538