Extend WebBlobRegistry API to support Stream.
This will provide a bridge between the Streams implementation in content/ and the W3C Stream API.

registerBlobURL(const WebURL&, const WebURL& srcURL) and unregisterBlobURL which are applicable for both Blob and Stream could be renamed to ones without "Blob" or with "Stream". Left FIXME to do that later.

Blob is planned to be a subclass of Stream but it's still under discussion. See the thread "Intent to Implement: Streams API" at blink-dev@ for more info. 

BUG=169957

Review URL: https://chromiumcodereview.appspot.com/14628004

git-svn-id: svn://svn.chromium.org/blink/trunk@152217 bbb929c8-8fbe-4397-9dbb-9b2b20218538