Consistently autosize list item markers (e.g. bullets and numbers).
This patch implements correct autosizing of list item markers when the list item is not a direct parent of an autosized text node, which happens when the list items contain marked up text such as e.g. links, bold and italic text or autosized text as part of an anonymous render block.

BUG=135854,169484

Review URL: https://chromiumcodereview.appspot.com/16714002

git-svn-id: svn://svn.chromium.org/blink/trunk@152391 bbb929c8-8fbe-4397-9dbb-9b2b20218538