Add a histogram to track the number of webfonts in a page
This patch adds WebFont.WebFontsInPage histogram that counts number of
webfonts used in a page. This is recorded when the first layout is done,
and so will not count webfonts dynamically loaded by scripts.

The number of fonts is tracked by FontLoader, so FontLoader object will
be created even if the runtime flag is off, but not visible from user.

BUG=237060

Review URL: https://chromiumcodereview.appspot.com/17301007

git-svn-id: svn://svn.chromium.org/blink/trunk@152680 bbb929c8-8fbe-4397-9dbb-9b2b20218538