Put slow test shards at the beginning of the unlocked shards list.
This is a bit of a hack, but many of the the virtual test suites are extremely slow
and they happen to run at the end because of the alphabetical ordering.
Instead, run them first.

This takes the time to run the full test suite from 4 minutes to 3 minutes
on my Linux Z620 and looks like it might save ~45 seconds off the Linux
buildbot.

Review URL: https://chromiumcodereview.appspot.com/17395009

git-svn-id: svn://svn.chromium.org/blink/trunk@152697 bbb929c8-8fbe-4397-9dbb-9b2b20218538