Get rid of dependency from DocumentParser to DocumentWriter
This change removes dependency from DocumentParser to DocumentWriter.
DocumentWriter is called only from a subclas DecodedDataDocumentParser,
where the parser requests a decoder object to the writer.
This change let DocumentWriter prepare the decoder object for the parser
so that the parser can retrieve the decoder from Document when it is needed.

This change also introduces TextResourceDecoderBuilder for clarifying
the decoder setup proces.

TEST=none
BUG=none
R=abarth@chromium.org,eseidel@chromium.org,tkent@chromium.org

Review URL: https://chromiumcodereview.appspot.com/17073005

git-svn-id: svn://svn.chromium.org/blink/trunk@152864 bbb929c8-8fbe-4397-9dbb-9b2b20218538