CSS clip should not be considered when clamping for overflow clip.
Previous patch in webkit rev 152919 partially fixed this, but skipped one code
path that still needed fixing. These code paths are intended to clamp the
backgroundRect to the borderBox, or a slightly inflated borderBox.  It is
incorrect to trigger those code paths when there is only CSS clip and no
overflow clip.  Fixing this bug also results in a tiny refactor where the CSS
clip if-statement is separated from the overflow clip chunks of code.

BUG=246728
R=esprehn@chromium.org

Review URL: https://codereview.chromium.org/17653002

git-svn-id: svn://svn.chromium.org/blink/trunk@152980 bbb929c8-8fbe-4397-9dbb-9b2b20218538