Add the revisions by builder to garden-o-matic.
This changes replaces the link to the oldest revision in the builders w/ a details element containing a table listing the revisions by builder. We also modify the details element so that if you hover over the summary, you get the table as a popup.

R=abarth@chromium.org

Review URL: https://chromiumcodereview.appspot.com/17784002

git-svn-id: svn://svn.chromium.org/blink/trunk@153189 bbb929c8-8fbe-4397-9dbb-9b2b20218538