Remove calls to updateMarqueePosition from RenderLayer
updateMarqueePosition only requires layout information for its
computations, which is why we can remove the 2 calls to the function
in RenderLayer:
- updateLayerPositions is called after layout so it's strictly
equivalent to calling it during layout. Also the current logic in
RenderLayer was working around <marquee> triggering spurious
invalidation (see e.g. the ASSERT fix for
https://bugs.webkit.org/show_bug.cgi?id=92464).
- the call in updateLayerPositionsAfterScroll was unneeded as
updateLayerPositions doesn't rely on the scroll offset.

The rebaselines correspond to timing difference in the tests as we
start the <marquee> during layout and not just after.

R=leviw@chromium.org

Review URL: https://codereview.chromium.org/18257002

git-svn-id: svn://svn.chromium.org/blink/trunk@153425 bbb929c8-8fbe-4397-9dbb-9b2b20218538