Fix a crash in Range::processContents().
This crash can be initiated by calling Range.detach() while deleteContents()
is processing the same range. Range::processContents() should save the state
of the range since mutation observers can change the state of the range.

This change affects one existing test (rtl-delete-within-line-crash.html)
that was apparently generated by fuzzers. Though it's hard to understand
the correct behavior of this test, the goal of this test seems to be
"no crash", and actually it does not crash with this change applied.
The new baseline is included in this change list.

A new layout test is added to test the scenario that causes the actual crash.
Interestingly, this test does not work on Windows trybots (crash on debug bots,
timeout on release bots), but does work on local Windows workstation. I used
a whole day to research this issue, but I got no clue about what's happening
on the bots. I have updated TestExpectations for the time being, hoping that
the main builders give me better idea about this.

BUG=112427
R=arv@chromium.org

Review URL: https://chromiumcodereview.appspot.com/18329003

git-svn-id: svn://svn.chromium.org/blink/trunk@153483 bbb929c8-8fbe-4397-9dbb-9b2b20218538