SmartReplaceICU shouldn't use bloatedCharacters()
This code only executes twice, so it's not overly important to remove the extra
memcpy. We could store the data 16 bits wide in the binary, but that would cost
binary space. The approach I've used in this CL is to keep the memcpy but to
copy out into a temporary Vector<UChar> rather than using the upconverting
mechanism in String.

R=eseidel
BUG=254708

Review URL: https://chromiumcodereview.appspot.com/18477005

git-svn-id: svn://svn.chromium.org/blink/trunk@153531 bbb929c8-8fbe-4397-9dbb-9b2b20218538