Add support for 'order' on grid items
The property is used to modify painting order and the order in which
we consider grid items for the auto-flow algorithm.

The code is build on the existing OrderIterator and matches what
flexbox does. Also to avoid adding an extra child walk, refactored
the explicit grid initialization so that we do also initialize the
OrderIterator's Vector. This refactoring saves a child walk as we
now compute both explicit grid's sizes in a single child walk.

BUG=257078
TESTS=fast/css-grid-layout/grid-item-order-auto-flow-resolution.html
      fast/css-grid-layout/grid-item-order-paint-order-expected.html

Review URL: https://chromiumcodereview.appspot.com/18516007

git-svn-id: svn://svn.chromium.org/blink/trunk@153536 bbb929c8-8fbe-4397-9dbb-9b2b20218538