PreserveWindow handles focus of smart_object.
smart_object that PreserveWindow creates is not elm_object, so focus of this object can
not be controlled by container elm_object that embedder creates.
So embedder gives container elm_object to PreserveWindow, and PreserveWindow
hooks focus events of the container elm_object, and PreserveWindow handles focus
of smart_object.