Build a shared or static library depending on `component'.
Do like the rest of the libraries in the project and use the `component'
variable (ie. -Dcomponent=shared for gyp) to decide whether to build
libefl_webview as a shared or static library.

So far a shared library was always being built regardless of the value of
`component'. xwalk_export.h checks for COMPONENT_BUILD, which is only set if
shared libraries are being used, and thus a static build would fail because
no symbols were being exported.