Element::setContainsFullScreenElement should schedule Forced style recalcs
Fullscreen has a bunch of pseudo classes that need updating on all ancestors when an
element goes into fullscreen mode (or out of it) so we mark all ancestors as needing a
SubtreeStyleChange all the way the tree to the documentElement. Doing this is the
same as just scheduling a Forced style recalc on the whole document so we can instead
do that.

I also cleaned up the SelectorChecker fullscreen pseudo element code.

Review URL: https://chromiumcodereview.appspot.com/18572013

git-svn-id: svn://svn.chromium.org/blink/trunk@153603 bbb929c8-8fbe-4397-9dbb-9b2b20218538