Refactoring: Decouple ScriptElement from HTMLScriptElement and SVGScriptElement.
This change pulls ScriptElement out from Element inheritance chain.
This give us an opportunity to decouple the lifetime of the element and
script loading and execution. Various states which is realted to script loading
won't be needed and can be discarded once it is feed to the script engine.
Such refactoring will follow.

Also in a coming change, ScriptElement will be renamed to ElementScriptLoader.

TEST=none
BUG=240592
R=eseidel,dglazkov,abarth

Review URL: https://chromiumcodereview.appspot.com/18676002

git-svn-id: svn://svn.chromium.org/blink/trunk@153748 bbb929c8-8fbe-4397-9dbb-9b2b20218538