Make PerformanceTests/Layout/line-layout.html 2% faster
I was looking at this performance test in pprof or an unrelated reason and
noticed that we were spending 2% of our time clearing a Vector. Rather than
clearing the Vector, we can just set its size to 0 and avoid calling malloc and
free for each iteration of the loop. The Vector is held in a local variable, so
we'll reclaim its storage when the function returns.

R=leviw@chromium.org

Review URL: https://chromiumcodereview.appspot.com/18868007

git-svn-id: svn://svn.chromium.org/blink/trunk@153751 bbb929c8-8fbe-4397-9dbb-9b2b20218538