HTML Media Capture: Update implementation, part 2.
The most recent version of the HTML Media Capture spec changes the `capture'
attribute from an enum to a boolean (the values that used to be passed to
the enum are now supposed to be extracted from the `accept' attribute).

Now that the Chromium side has landed, make the `capture' attribute in
HTMLInputElement a boolean and adjust the layout tests accordingly.

R=abarth@chromium.org,eseidel@chromium.org,jochen@chromium.org
BUG=240252

Review URL: https://chromiumcodereview.appspot.com/18332015

git-svn-id: svn://svn.chromium.org/blink/trunk@153804 bbb929c8-8fbe-4397-9dbb-9b2b20218538