Refactoring: Get rid of mediasource dependency from HTMLMediaElement
This change introduces MediaSourceable interface to decouple
MediaSourceBase class from HTLMediaElement.
MediaSourceable defines a minimum set of APIs for HTMLMediaElement
which uses the API to interact with mediasource module behind the it.

This is a part of an effort to eliminate non-core dependency from core module.

BUG=240988
TEST=none
R=abarth,acolwell

Review URL: https://chromiumcodereview.appspot.com/18753004

git-svn-id: svn://svn.chromium.org/blink/trunk@153807 bbb929c8-8fbe-4397-9dbb-9b2b20218538