Remove HTMLImageElement subclass from ImageDocument
The only reason for this subclass is apparently to clear the pointer the ImageDocument is holding
to the <img> that shows the image, pressumably in case someone reaches into an image document and
moves the <img> into a new document. We can instead use Document::dispose and an ownerDocument
check combined with a RefPtr to get the same effect and remove this subclass. This lets us
make HTMLImageElement FINAL.

I also tightened up the types in ImageDocument::createDocumentStructure() where it was using
Document::createElement instead of just creating the elements itself. This removes one cast
and lets the compiler devirtualize all the method calls we later make on the elements since
all the element types we create are FINAL.

R=abarth@chromium.org

Review URL: https://codereview.chromium.org/18886003

git-svn-id: svn://svn.chromium.org/blink/trunk@153834 bbb929c8-8fbe-4397-9dbb-9b2b20218538