Teach ShadowData about the create() pattern and PassOwnPtr
I noticed manual adoptPtr calls in SVGCSSStyleSelector
one of which I think was a double adopt in the inherit
case (and thus prossibly a double-free?) and decided to
fix ShadowData to never need a manual adoptPtr call.

I also fixed percentageOrNumberToFloat to include the
primativeValue null-check so that callers did not have to
(sorry for mixing that into this CL).

BUG=237268
R=dglazkov@chromium.org

Review URL: https://codereview.chromium.org/18689007

git-svn-id: svn://svn.chromium.org/blink/trunk@153930 bbb929c8-8fbe-4397-9dbb-9b2b20218538