Remove IDBNotFoundError ExceptionCode
Instead of passing an ExceptionCode (int) around to signify
an exception, we now pass an ExceptionState object. This is
so that we can pass more data out than just a single int.

The ExceptionState object has methods for throwing an exception with an optional error message.

There are also temporary converters so that functions can
still use ExceptionCode until the entire code base has
transitioned. Once that is done the ExcetionState code can
be simplified further.

BUG=252233

Review URL: https://chromiumcodereview.appspot.com/18398002

git-svn-id: svn://svn.chromium.org/blink/trunk@153938 bbb929c8-8fbe-4397-9dbb-9b2b20218538