Avoid N^2 walk placing renderers when building the render tree
Blink resolves style by walking each element from first children to last, but
when we go to insert their renderer, we look for the next renderer to insert
before. On the initial style resolve, this means we'll walk the entire DOM
tree trying to find the right renderer to insert before leading to an N^2
loop over the DOM on load.

This could be fixed by changing the semantics by which we insert renderers
(insert after instead of insert before). Instead, here I reverse the order
we resolve style. This should ensure that in the common case, we'll find
the renderer to insert before immediately.

While looking at this, I also found we have an N^2 loop for resolving
Nth last child selectors, and reversing the style resolve loop would have
caused the same issue for Nth child selectors. To fix prevent this regression,
I'm piping the child index down to the style resolver in the common case.
This ensures both Nth and Nth last should usually be O(1).

The previous version of this patch resulted in a perf regression due to
the extra in-order loop in Element::recalcStyle. This loop is now gated
on the presence of child or sibling selectors. The common case should
be faster.

Previous code review: https://chromiumcodereview.appspot.com/15871005
BUG=245478

Review URL: https://chromiumcodereview.appspot.com/18276003

git-svn-id: svn://svn.chromium.org/blink/trunk@154053 bbb929c8-8fbe-4397-9dbb-9b2b20218538