Remove FixedArray, which had only a single instantiation
This class used to be used in various JSC code. But in Blink, it only
had a single user, with a fixed size and without any calls to its member
functions. This patch changes GlyphMetrics to simply use an array directly,
and makes that code a little safer by using a proper class instead of a struct.

R=eseidel

Review URL: https://chromiumcodereview.appspot.com/18691004

git-svn-id: svn://svn.chromium.org/blink/trunk@154070 bbb929c8-8fbe-4397-9dbb-9b2b20218538