Refactoring: Extract DocumentInit for capture Document construction parameter.
This change replaces a set of parameters for the Document constructor
with DocumentInit class.

The motivation here is to pass a HTMLImport to the constructor.
In a coming change, passed import object will affect how the
security context of the document is decided.

For that purpose, DocumentInit encapsulates which Frame is
being used to initialize the context. The coming change will extend it
so taht HTML Imports can load external scripts.

Since DocumentInit has a default constructor, call sites where create
empty documents become cleaner.

BUG=240592
TEST=none
R=abarth@chromium.org

Review URL: https://codereview.chromium.org/18808004

git-svn-id: svn://svn.chromium.org/blink/trunk@154278 bbb929c8-8fbe-4397-9dbb-9b2b20218538