Introduce FontBuilder.
Once it grows up, the FontBuilder is a builder-style class that will vend a Font object after letting you set up all font parameters.

As a first baby step, we start it as a chunk of font-building code that is functionally equivalent to what it is now.

BUG=259085
R=eseidel@chromium.org

Review URL: https://codereview.chromium.org/19307002

git-svn-id: svn://svn.chromium.org/blink/trunk@154345 bbb929c8-8fbe-4397-9dbb-9b2b20218538