Separate WebIDBKey type into separate header
WebIDBKey has exported symbols that we don't want to call from any common or browser
code as well as a type enum that we want to reuse everywhere. This separates the
type enum into a standalone WebKit::WebIDBKeyType enum in its own header and
normalizes the names to standard blink public API conventions. This way we can
use DEPS to enforce that we're only using the types in common and browser code.

BUG=237267

Review URL: https://chromiumcodereview.appspot.com/19316002

git-svn-id: svn://svn.chromium.org/blink/trunk@154358 bbb929c8-8fbe-4397-9dbb-9b2b20218538