Workaround DocumentLoader assertions hit by layout tests.
The following three tests hit an assertion in DocumentLoader::responseReceived
because they 'alert()' after loading a 'data:' URL in a frame or popup. This
apparently breaks the assumptions in responseReceived.

I've updated the bug with a minimal reproduction; this patch simply adjusts the
tests to avoid 'alert()' in order to regain a bit of test coverage. The actual
issue will have to be fixed in a followup patch.

BUG=243300

Review URL: https://chromiumcodereview.appspot.com/19506003

git-svn-id: svn://svn.chromium.org/blink/trunk@154420 bbb929c8-8fbe-4397-9dbb-9b2b20218538