Populate cross-origin location access SecurityError messages.
We currently add a message to the console each time a Location object
is accessed cross-origin, in addition to throwing a SecurityError exception.
We should allow developers to suppress the console message by catching
and properly handling any exception that's generated.

This patch adds the ability to pipe a message through 'setDOMException()'
such that it's applied as the exception's message property. When we deny
access to a Location object, we can use this new method, and avoid
writing an unsuppressable message out to the console.

BUG=17325
R=abarth@chromium.org

Review URL: https://codereview.chromium.org/19564003

git-svn-id: svn://svn.chromium.org/blink/trunk@154564 bbb929c8-8fbe-4397-9dbb-9b2b20218538