Pass column as 4th argument to WorkerGlobalScope.onerror handler
As per the latest WebWorker specification, the WorkerGlobalScope.onerror
event handler should be given the column as fourth argument:
http://www.whatwg.org/specs/web-apps/current-work/multipage/workers.html#workerglobalscope
http://www.whatwg.org/specs/web-apps/current-work/multipage/webappapis.html#onerroreventhandler

The behavior is consistent with IE10.

Note that ErrorEvent.column is not exposed yet because no other browser
implements it except IE10 which uses a different name ("colno").

BUG=262126
R=abarth@chromium.org, arv@chromium.org

Review URL: https://codereview.chromium.org/19693006

git-svn-id: svn://svn.chromium.org/blink/trunk@154630 bbb929c8-8fbe-4397-9dbb-9b2b20218538