Put generated parsetab.py in output directory.
The new Python parser generates a file, parsetab.py.
(This is expensive, so it is cached.)
The path defaults to the source directory, so it makes a mess in the user's
source files, and needs to be deleted.

This CL fixes this by putting it in the generated directory.

Longer-term, we should be able to generated this table as a build target,
and then run the parser in optimized mode, which should be faster,
but that's more involved (do later).

Review URL: https://chromiumcodereview.appspot.com/19576009

git-svn-id: svn://svn.chromium.org/blink/trunk@154746 bbb929c8-8fbe-4397-9dbb-9b2b20218538