Separate the scrolling code out of RenderLayer
RenderLayer was a ScrollableArea even though it couldn't scroll (e.g.
if it had overflow: visible). This change is the mechanical addition
of RenderLayerScrollableArea. For now, the new class is a shim that
forwards all calls to RenderLayer but more logic will be moved to
the class in follow-up patches.

The current logic checks the style for allocating the ScrollableArea,
which is the safest but could end up allocating more than needed
(e.g. for overflow: auto).

BUG=260899
R=abarth@chromium.org, esprehn@chromium.org, jamesr@chromium.org, vollick@chromium.org

Review URL: https://codereview.chromium.org/19218007

git-svn-id: svn://svn.chromium.org/blink/trunk@154778 bbb929c8-8fbe-4397-9dbb-9b2b20218538