Make StylePropertyMetadata a 16 bits object.
This is a new patch after the rollout of 153173.

Profiling the benchmark with Intel Vtune to find out the performance regression
showed that copying uint16_t, so StylePropertyMetadata, is not
a fast operation in term of assembly code and has a big penalty on Windows 
MSVC over unsigned for example. The latter produces a much faster code 
when using unsigned over uint16_t (45% difference in the benchmark).

The patch fixes the regression by avoiding the copies of StylePropertyMetada
when applicable (by using const ref).

The second part of the fix is avoiding the conversion from an int (enum) to
an uint16_t in a tight loop such as StylePropertySet::findPropertyIndex 
(which is the hotspot of the benchmark).

On my Windows 7 64 bits Core i5 machine CSSPropertySetterGetter results are :
- avg : 2714 runs/s with the patch
- avg : 2696 runs/s without the patch

Apple internal MemBuster reports ~1.8 Mb of improvements (loading the top 50 Alexa websites).

Review URL: https://chromiumcodereview.appspot.com/19846004

git-svn-id: svn://svn.chromium.org/blink/trunk@154892 bbb929c8-8fbe-4397-9dbb-9b2b20218538