Various crashers in RenderLayer due to missing a ScrollableArea
When splitting ScrollableArea out of RenderLayer, the scrollbar
teardown wasn't taken into account and thus we would try de-allocate
or not have a ScrollableArea when it was required to hang the
scrollbar on.

RenderLayerScrollableArea and Scrollbars have fairly entangled
bring-up and tear-down code which makes it hard to fix correctly in
the current situation. However the scrollbars should be moved to
RenderLayerScrollableArea soon which will solve this problem in a
better way.

BUG=264213,264287,264171,263923
TESTS=fast/layers/crash-no-scrollable-area-scrollbars.html
      fast/layers/crash-remove-scrollbar-no-scrollable-area.html
      fast/layers/crash-shadowdom-scrollbar-no-scrollable-area.html

R=vollick@chromium.org

Review URL: https://codereview.chromium.org/20422002

Patch from Julien Chaffraix <jchaffraix@chromium.org>.

git-svn-id: svn://svn.chromium.org/blink/trunk@154938 bbb929c8-8fbe-4397-9dbb-9b2b20218538