Update toNativeArray() / toRefPtrNativeArray() do not match Web IDL specification
Update toNativeArray() / toRefPtrNativeArray() do not match the latest Web IDL
specification:
- http://dev.w3.org/2006/webapi/WebIDL/#es-array
- http://dev.w3.org/2006/webapi/WebIDL/#es-sequence

In particular, we no longer generate an empty Vector if the JavaScript value is not
an array but can still be converted to an IDL sequence. If the JavaScript value is
an object (Other than a Date or a RegExp), we now call [[Get]] on it with a property
name of "length" and call ToUint32() on the result. If this succeeds, we then
retrieve each item by calling [[Get]] with a property name corresponding to the
index.

This essentially means that IDL operations taking an array or a sequence in argument
will now be more permissive and convert non v8::Array by converting them to sequences.

BUG=263758

Review URL: https://chromiumcodereview.appspot.com/19969004

git-svn-id: svn://svn.chromium.org/blink/trunk@154940 bbb929c8-8fbe-4397-9dbb-9b2b20218538