Get rid of the distinction between modifiers and expectations.
Now there are just the platform specifiers and the expectations
and they match the syntax of the expectations file, i.e.
specifiers are in the brackets before the test and
expectations are in the brackets after the tests.

In addition to cleaning up the code, this makes it so that we
don't artificially add a PASS expectation to SLOW, WONTFIX, etc. tests.
This, in turn, fixes a bug where merging two expectations line
would mark the test as flaky, e.g.:
"foo.html [ Slow ]" merged with "foo.html [ Failure ]" would
become "foo.html [ Slow Pass Failure ]" before this patch.

R=dpranke@chromium.org

Review URL: https://codereview.chromium.org/20830003

git-svn-id: svn://svn.chromium.org/blink/trunk@155028 bbb929c8-8fbe-4397-9dbb-9b2b20218538