TechCrunch should use 600 kB less memory
Before this CL, we could end up with duplicate copies of JavaScript code if the
CachedScript was evicted from the MemoryCache but V8 retained the underlying
String and we loaded the script again. Similarly, we could end up with
duplicates if a page loads an identical script from two distinct URLs.

This CL changes the backing store for CachedScript to an AtomicString to
coalesce these copies. We already coalesce inline scripts for the same reason.

R=eseidel

Review URL: https://chromiumcodereview.appspot.com/21133009

git-svn-id: svn://svn.chromium.org/blink/trunk@155200 bbb929c8-8fbe-4397-9dbb-9b2b20218538