Replace RenderArena with PartitionAlloc.
We believe this is memory positive, a very minor binary size saving, and perf neutral; see the bug for measurements. It's definitely a code clean-up. Perf bots will be carefully monitored after the land.

(This CL is mostly a subset of previously-reviewed https://codereview.chromium.org/16896019/, which we tried to land in a few pieces, to the limited extent possible.)

BUG=252057
R=abarth@chromium.org

Review URL: https://codereview.chromium.org/20231002

git-svn-id: svn://svn.chromium.org/blink/trunk@155202 bbb929c8-8fbe-4397-9dbb-9b2b20218538