Bold and Italic style get stuck when applied on a selection having content with mixed editability
When bold/italic style is applied to a content with mixed editability,
style is applied appropriately for the first time. But after this,
style toggling does not work any more. Same behavior can be observed
when the content on which style is applied contains text node
without renderer (Ex: text node corresponding to tab/space etc).

When bold/italic style is applied on a selection, we actually apply the
style only for text nodes that have renderers and also are
contenteditable. Similarly during style toggling, to check whether a
style has been already applied on a selection, we should be considering
only those text nodes that have renderers and are contenteditable.

Review URL: https://chromiumcodereview.appspot.com/19888007

git-svn-id: svn://svn.chromium.org/blink/trunk@155246 bbb929c8-8fbe-4397-9dbb-9b2b20218538