Fix baseline position when it is outside the element's box
Specifically, we shouldn't force the baseline to be inside
the element. IE and FF don't do this, and it's incompatible
with the CSS spec:

"The baseline of an 'inline-block' is the baseline of its last line box in the normal flow, unless it has either no in-flow line boxes or if its 'overflow' property has a computed value other than 'visible', in which case the baseline is the bottom margin edge."
(http://www.w3.org/TR/CSS21/visudet.html#leading)

It doesn't have a special case for "baseline is outside of the element's margin box".

This affects some tests in the obvious way which I marked as needing
rebaseline.

BUG=258650
R=leviw@chromium.org

Review URL: https://codereview.chromium.org/21124008

git-svn-id: svn://svn.chromium.org/blink/trunk@155253 bbb929c8-8fbe-4397-9dbb-9b2b20218538