Reset the resumeTime and accumulated time on svg.setCurrentTime
This patch fixes the bug in the recent fix, crrev.com/18374004
Earlier(after the above patch) we were not resetting the m_resumeTime
and m_accumulatedActiveTime because of which if we use setCurrentTime
while the animation is paused that time was getting accumulated if we
unpaused the animation later. This bug is represented by the test
included in the patch.

BUG=254067
TEST=LayoutTests/svg/animations/getCurrentTime-pause-unpause-setCurrentTime.html

Review URL: https://chromiumcodereview.appspot.com/20795005

git-svn-id: svn://svn.chromium.org/blink/trunk@155289 bbb929c8-8fbe-4397-9dbb-9b2b20218538