Drop [IsIndex] attribute from CharacterData API
The DOM level3 specification used to require the following CharacterData
operations to throw an IndexSizeError if the offset was negative:
substringData(), deleteData(), insertData() and replaceData().

This is however not the case anymore in DOM4:
http://dom.spec.whatwg.org/#concept-cd-substring
http://dom.spec.whatwg.org/#concept-cd-replace

We should now only throw an IndexSizeError if the offset is greater than
the data length. We can therefore drop the [IsIndex] attribute and let
the bindings convert the number to an unsigned long as per the Web IDL
specification:
http://dev.w3.org/2006/webapi/WebIDL/#es-unsigned-long

This behavior is consistent with both IE10 and Firefox.

BUG=237739

Review URL: https://chromiumcodereview.appspot.com/21581003

git-svn-id: svn://svn.chromium.org/blink/trunk@155423 bbb929c8-8fbe-4397-9dbb-9b2b20218538