Remove StringImpl::tryCreateUninitialized.
Create or do not create. There is no try.

The only client of tryCreateUninitialized is tryMakeString and the only client of _that_ does a RELEASE_ASSERT on the results anyway. So the cleanup cascades a little.

It's also worth noting that our underlying system allocator is tcmalloc, which we have hard wired to crash on a failed allocation anyway.

BUG=none

Review URL: https://chromiumcodereview.appspot.com/21961002

git-svn-id: svn://svn.chromium.org/blink/trunk@155455 bbb929c8-8fbe-4397-9dbb-9b2b20218538