Start the removal of DumpRenderTree.
The plan is to move everything from Tools/DumpRenderTree/ to Source/testing/runner.

DumpRenderTree is a test harness, but we don't anymore. Instead, we use
content_shell with the --dump-render-tree flag.
We still use most of the code in that directory, but now it's a library that's
used by content_shell.

BUG=None
TEST=still builds.
R=abarth@chromium.org

Review URL: https://chromiumcodereview.appspot.com/22050002

git-svn-id: svn://svn.chromium.org/blink/trunk@155475 bbb929c8-8fbe-4397-9dbb-9b2b20218538