Fix touch support issue on Tizen 2.1
On Tizen 2.1, the multi-touch support is different from the XInput2.2
approach. It extends the older Input2.1 spec to use multiple touch
devices for multiple touch points, one touch device for one touch
points. The device type name for this purpose is MULTITOUCHSCREEN which
is not defined by XInput extension.

This fix is to make Crosswalk be able to handle the new extension to
XInput2.1 by introducing a new build option 'enable_xi21_mt'. It should
be turned on for Tizen 2.1, and is exclusive to 'use_xi2_mt' option.