DevTools: Use rjsmin instead of jsmin for js minification
This patch adds rjsmin instead of jsmin for js minification. The reason
for this change is that both bundled jsmin implementation and its updated
version seem to be failing on some tricky regexps. This results in
invalid javascript. That was the reason why minification was not enabled for CodeMirrorTextEditor (this patch enables it)

BUG=

Review URL: https://chromiumcodereview.appspot.com/22393005

git-svn-id: svn://svn.chromium.org/blink/trunk@155695 bbb929c8-8fbe-4397-9dbb-9b2b20218538