Don't clear element style state when using context.resolvedStyle in reattach
If we're going to reattach and use the resolvedStyle from inside recalcStyle we
should not clear the style state (affectedByEmpty and childIndex) since we're not
going the run the SelectorChecker again.

I couldn't construct a test for childIndex (ex. nth-child), but it looks like such
a bug should exist.

BUG=266407

Review URL: https://chromiumcodereview.appspot.com/22584004

git-svn-id: svn://svn.chromium.org/blink/trunk@155877 bbb929c8-8fbe-4397-9dbb-9b2b20218538