Replace style sharing cousin list search with LRU
The current implementation of style sharing uses a very confusing set of methods to
search a node's cousins in an attempt to find other nodes to share style with. This
patch replaces this confusing method with a simple LRU that allows us to find
nodes to share with even if they're not our cousins.

Using this method results in more Style Sharing, but it also results in spending
more time looking for nodes to share with. The end result is a different performance
profile. Some tests will be slower, some will be faster. I used an internal set of
25 "typical" web pages, a set of websites, and our perf tests to help refine the
size of the LRU to maximize for performance.

Review URL: https://chromiumcodereview.appspot.com/20602002

git-svn-id: svn://svn.chromium.org/blink/trunk@156070 bbb929c8-8fbe-4397-9dbb-9b2b20218538