Change new float offset methods to take fixedOffset parameter
When moving these methods, I attempted to reduce the number of
arguments. However, this seems to have caused some test failures. This
patch reduces the scope of the change by keeping the same number of
arguments on the moved methods as the ones they replaced. Attempting to
reduce the number of arguments will have to wait for a future patch.

BUG=271774
R=eseidel@chromium.org

Review URL: https://codereview.chromium.org/23224002

git-svn-id: svn://svn.chromium.org/blink/trunk@156121 bbb929c8-8fbe-4397-9dbb-9b2b20218538