frameContentAsPlainText should call updateLayout not check needsLayout
frameContentAsPlainText was checking needsLayout() and then calling layout()
manually but this doesn't work if the frame didn't need a layout but did
need a style recalc which could have triggered a layout.

Instead of manually checking for needsLayout() (which is almost always wrong)
we now just call document->updateLayout() which will make sure things are in
the correct state.

Review URL: https://chromiumcodereview.appspot.com/23099005

git-svn-id: svn://svn.chromium.org/blink/trunk@156128 bbb929c8-8fbe-4397-9dbb-9b2b20218538