Don't fire Custom Element callbacks while the created callback is running.
Custom Element callbacks used to be fired as the DOM-modifying method
that caused them is returning to script. When a callback modified DOM,
this implied recursively invoking callbacks.

However the created callback is deemed to be special; callbacks should
be delayed until the created callback has run to completion. This is
based on a proposal described here:
<https://www.w3.org/Bugs/Public/show_bug.cgi?id=22966>

BUG=272561

Review URL: https://chromiumcodereview.appspot.com/22999013

git-svn-id: svn://svn.chromium.org/blink/trunk@156185 bbb929c8-8fbe-4397-9dbb-9b2b20218538