Fix editing/input/editable-container-with-word-wrap-normal.html on Mac
Use locally installed Ahem font instead of loading it as WebFont,
because WebFont may not be available at the point the test code executes.
Until WebFont is available, local font with the same name used to be used
as fallback, but that behavior has changed in r156182.

BUG=261029

Review URL: https://chromiumcodereview.appspot.com/23026008

git-svn-id: svn://svn.chromium.org/blink/trunk@156195 bbb929c8-8fbe-4397-9dbb-9b2b20218538