Make configurationForViewport match production code
While the production code has used initial viewport width and height for
device-width and device-height, the configurationForViewport API in test builds
still used separate values for the device size. This CL removes the redundant
deviceSize in the production code and removes the device parameters from
configurationForViewport so that the tests give the same results as the
production code would do.

BUG=273649

Review URL: https://chromiumcodereview.appspot.com/23101004

git-svn-id: svn://svn.chromium.org/blink/trunk@156302 bbb929c8-8fbe-4397-9dbb-9b2b20218538