Added UMA instrumentation for site-instrumented user timing marks and measures
Added UMA histograms for each performance.mark (PLT.UserTiming_Mark) and merformance.measure (PLT.UserTiming_MeasureDuration) that were instrumented by pages.

Entry points were also exposed that measure the last mark on a page as well as if a page was instrumented for user timing so that the Chrome side of the code can report UMA's to track those.  The last mark (PLT.UserTiming_LastMark) is useful in cases where sites instrument a bunch of things they want to track and it lets you know when the last one of those happened (like all of the above-the-fold images finished loading).  Knowing if a page was instrumented (PLT.UserTiming_Instrumented) with user timing will let us track adoption over time.

The Chrome-side of the code will be uploaded after the blink patch lands and is included in the blink milestone used by Chrome.

BUG=262567

Review URL: https://chromiumcodereview.appspot.com/23245004

git-svn-id: svn://svn.chromium.org/blink/trunk@156481 bbb929c8-8fbe-4397-9dbb-9b2b20218538