Add animated WebP image that uses 'no blend' option to Layout tests
This is to test the case when an animation frame is supposed to
overwrite the previous frame (instead of blending with previous frame).

This was added to libwebp in Chromium here:
https://codereview.chromium.org/22802020/

Also, this patch is in preparation for the Blink patch:
https://codereview.chromium.org/23068027/, where the
corresponding unit tests are also added.

BUG=
R=noel@chromium.org

Review URL: https://codereview.chromium.org/23129024

git-svn-id: svn://svn.chromium.org/blink/trunk@156614 bbb929c8-8fbe-4397-9dbb-9b2b20218538