Change WindowNameCollectin to include both SVG and HTML elements
This is a merge of http://http://trac.webkit.org/changeset/154441
by Rob Buis <rwlbuis@webkit.org>.

The assert is hit in Debug mode because the DocumentOrderedMap in
HTMLDocument::m_windowNamedItem includes matched SVG elements, but
the WindowNameCollection used to collect the elements does not. This
means the HTMLCollection stripped of SVG elements could end up
hitting the empty or single item assertion, which the testcase verifies.

To fix this, this patch changes WindowNameCollection to include both
SVG and HTML elements so it matches DocumentOrderedMap.

BUG=280406

Review URL: https://chromiumcodereview.appspot.com/23463005

git-svn-id: svn://svn.chromium.org/blink/trunk@156875 bbb929c8-8fbe-4397-9dbb-9b2b20218538