Add a public Blink function to canonicalize CSS selectors and restrict their type.
This adds the ability for Blink embedders to check that a selector is
appropriate in the <content select> attribute or the planned
WebDocument::watchSelectors() API, which is currently impossible in
Javascript.

This also adds a C++ equivalent of writing a rule into a stylesheet
and then retrieving its selector from CSSStyleRule.selectorText, which
is also useful for the WebDocument::watchSelectors() API, which
reports results using the selectorText string.

BUG=172011

Review URL: https://chromiumcodereview.appspot.com/23659002

git-svn-id: svn://svn.chromium.org/blink/trunk@156880 bbb929c8-8fbe-4397-9dbb-9b2b20218538