CORS: Update the redirection status in Inspector Network tab for CORS requests.
In the Inspector Network tab, 'Status' column shows '(cancelled)' and
'Type' shows 'pending' when CORS request is redirected. The status
is not updated when redirection is successful or fails due to access
control checks.

Added changes to update the response in Inspector when the CORS request
is redirected for both success and failure cases. Made changes to
log appropriate message to console when redirection fails
due to access control checks.

BUG=279203

R=abarth@chromium.org, pfeldman@chromium.org, bbudge@google.com

Review URL: https://chromiumcodereview.appspot.com/23582002

git-svn-id: svn://svn.chromium.org/blink/trunk@156915 bbb929c8-8fbe-4397-9dbb-9b2b20218538