Make createPattern synchronously snapshot images
This patch changes createPattern to synchronously snapshot an image
instead of lazily using the image at draw time. For animating images, we
now use the frame available when createPattern is called and not the
frame available when the canvas draw operation is performed.

After this change, we now match the canvas pattern spec:
"Modifying the image used when creating a CanvasPattern object after
calling the createPattern() method must not affect the pattern(s)
rendered by the CanvasPattern object."
http://www.whatwg.org/specs/web-apps/current-work/multipage/the-canvas-element.html#canvaspattern

This patch also removes the unused tiledImage() function on Pattern.

BUG=279445

Review URL: https://chromiumcodereview.appspot.com/23472003

git-svn-id: svn://svn.chromium.org/blink/trunk@156997 bbb929c8-8fbe-4397-9dbb-9b2b20218538