Modify StepsTimingFunction to record whether it was created from a preset CSS Value
The CSS3 Animations spec [1] requires that the value returned by
getComputedStyle() for the animation-timing-function property is the specified
value.

Currently, we follow this for all values other than 'step-start' and
'step-end'. In these cases, we represent the values using the 'step()' syntax.
This change fixes these two cases so that getComputedStyle() always return the
specified value.

Note that Firefox does not follow the spec and always returns the values using
the 'step()' or 'cubic-bezier()' syntax.

Also cache a singleton StepsTimingFunction for each of 'step-start' and
'step-end'.

[1] http://dev.w3.org/csswg/css-animations/#animation-timing-function

R=dstockwell

Review URL: https://chromiumcodereview.appspot.com/23442009

git-svn-id: svn://svn.chromium.org/blink/trunk@157064 bbb929c8-8fbe-4397-9dbb-9b2b20218538