Avoid collapsing anonymous block children already being destroyed
When normalizing a block with anonymous blocks for first-letter,
before content, and contained text, a collapsing anonymous block
cascade is triggered that attempts to collapse the contained
text's anonymous block within its destruction method. To avoid
this, adding logic to bail early if we're already destroying
the child anonymous block we're trying to collapse.

While in the function, doing a little cleanup to make it more
obvious that it only operates on RenderBlocks, not RenderBoxes,
and updating some of the comments to hopefully be more useful.

BUG=282088

Review URL: https://chromiumcodereview.appspot.com/23463021

git-svn-id: svn://svn.chromium.org/blink/trunk@157392 bbb929c8-8fbe-4397-9dbb-9b2b20218538