DevTools: Get rid of BreakpointManager.reset method.
BreakpointManager.reset is a complicated method with unclear semantics that is only used in tests.
The only bit that is missing in BreakpointManager's public API and is needed in tests is an ability to 
remove all provisional breakpoints.
This patch added removeProvisionalBreakpoints() method and updates test to use it instead of .reset() method.

R=pfeldman

Review URL: https://chromiumcodereview.appspot.com/23437034

git-svn-id: svn://svn.chromium.org/blink/trunk@157756 bbb929c8-8fbe-4397-9dbb-9b2b20218538