[CSS Grid Layout] Use CSS_FR for flexible lengths instead of CSS_DIMENSION
Although 'fr' units are dimensions we're using a custom
CSSPrimitiveValue:UnitType which matches how px, em, ... are parsed. That's
why we should use CSS_FR instead of CSS_DIMENSION for the computed style.

This also fixes a bug in the style builder which was converting the fractional
primitive value to a float instead of a double causing rounding issues in
some tests.

BUG=235258

Review URL: https://chromiumcodereview.appspot.com/23537037

git-svn-id: svn://svn.chromium.org/blink/trunk@157820 bbb929c8-8fbe-4397-9dbb-9b2b20218538