Get rid of custom code for Document::createTouchList()
Get rid of custom code for Document::createTouchList() now that the bindings
generator supports variadic methods.

One behavior change is that we now throw a TypeError if one of the given arguments
is not a Touch, instead of treating it as null. This behavior is however consistent
with Firefox Mobile and the specification:
https://dvcs.w3.org/hg/webevents/raw-file/v1/touchevents.html#idl-def-Document

R=haraken,arv
BUG=

Review URL: https://chromiumcodereview.appspot.com/23522051

git-svn-id: svn://svn.chromium.org/blink/trunk@157885 bbb929c8-8fbe-4397-9dbb-9b2b20218538