Add histograms to ResourceFetcher to measure memory cache hit rate.
This will drive possible improvements that could be made in the cache to
decrease memory usage (at least on Android).

Note that the metrics are in the ResourceFetcher rather than in the MemoryCache
itself to only report events that really matter in practice (e.g. the resource
has to be fetched from the browser process vs. the resource was served directly
from the in-memory cache).

In addition to the overall hit rate, an important metric that this CL is adding
is whether a resource is live or dead upon a cache hit. This will help us
understand whether memory usage improvements can be made in the cache since
only dead resources can be evicted (live resources are used by clients thus
can't be deleted by definition).

BUG=290193

Review URL: https://chromiumcodereview.appspot.com/23734006

git-svn-id: svn://svn.chromium.org/blink/trunk@157890 bbb929c8-8fbe-4397-9dbb-9b2b20218538