Fix for link highlighting of <img> links
Temporary fix for <img> inside <a>. I changed the node that gets passed to the LinkHighlight object. Previously, when a node was tapped, we'd go up the tree until we find a node that causes a "hand cursor" and use that. Now, we do the same but we stop ascending up the tree if a node isn't contained by its parent's box.

In the long term, the LinkHighlight should be changed to use the entire hit testing region of the node, rather than the line box.

BUG=170673

Review URL: https://chromiumcodereview.appspot.com/24110002

git-svn-id: svn://svn.chromium.org/blink/trunk@157936 bbb929c8-8fbe-4397-9dbb-9b2b20218538