WTF::notFound looks too much like a local variable.
I discovered this as part of attempting to deploy -Wshadow.
There was one place (HashMap.h) where we had a notFound local.

Changing notFound to kNotFound also made it look pretty in my editor.

This sorta seems like a silly change, but I do believe it will help
code readability.  I think it was a mistake to have a global constant
beginging with a lowercase letter in the first place.

BUG=294205

Review URL: https://chromiumcodereview.appspot.com/23464095

git-svn-id: svn://svn.chromium.org/blink/trunk@158058 bbb929c8-8fbe-4397-9dbb-9b2b20218538