Reduce PartitionAlloc memory waste for infrequently used buckets.
Let's suppose that a renderer process is showing a single web page with a
single Document / Frame. In this case, the 2528 byte sized Document object
goes into a relatively uncommon allocation bucket. Nothing else goes into this
bucket. But the bucket page size is 16KB so there is a lot of waste.

The solution is to hand out bucket page freelists in smaller batches, say 4KB.

Saves about 300KB of system memory on the NTP (which previously would have been
unused memory, but "dirty" in the view of the system on account of writing the
freelist pointers to the system pages). Full logs in the bug.

Also fix wtf_unittests to always initialize the random subsystem, fix a comment,
and enhance partitionDumpStats to better report on fragmentation.

BUG=294507

Review URL: https://chromiumcodereview.appspot.com/24152010

git-svn-id: svn://svn.chromium.org/blink/trunk@158081 bbb929c8-8fbe-4397-9dbb-9b2b20218538