MediaControls shouldn't call confusingAndOftenMisusedAttached
The media controls were calling confusingAndOftenMisusedAttached to determine
how to respond to input events. What they're really trying to determine is
whether they're part of a document that's being displayed to the user. The
canonical way to check that condition is to check whether they're inDocument
and whether that document has an associated renderer.

R=esprehn
BUG=299011

Review URL: https://chromiumcodereview.appspot.com/24974002

git-svn-id: svn://svn.chromium.org/blink/trunk@158459 bbb929c8-8fbe-4397-9dbb-9b2b20218538