Have all skia platforms share FontPlatformData::hash() implementation
Update Windows to use a SkTypeface based hash value for FontPlatformData
when the non-GDI path is used. Move the existing implementation from
FontPlatformDataHarfBuzz to FontPlatformDataSkia to allow and to ensure
that all skia platforms share the same implementation.

Also, rename m_size to m_textSiz for FontPlatformDataChromiumWin to be
consistent with the other platforms.

R=bungeman@chromium.org
BUG=249099

Review URL: https://chromiumcodereview.appspot.com/24874002

git-svn-id: svn://svn.chromium.org/blink/trunk@158495 bbb929c8-8fbe-4397-9dbb-9b2b20218538