ImageLoader shouldn't call confusingAndOftenMisusedAttached
What the ImageLoader really wants to know is whether it's appropriate to
fire events on its associated image elements. That's safe when the document is
active in a frame, which we can check more directly by asking whether the
document has an associated frame.

R=esprehn
BUG=299011

Review URL: https://chromiumcodereview.appspot.com/24972002

git-svn-id: svn://svn.chromium.org/blink/trunk@158517 bbb929c8-8fbe-4397-9dbb-9b2b20218538