Add WebDocument::watchCssSelectors(selectors) which requests callbacks to track whether elements matching |selectors| exist within the document.  Blink calls the embedder through WebFrameClient::didMatchCss(frame, newlyMatching, stoppedMatching) to report changes.
This patch only watches for the binary set "selector matches anything on the page" vs "selector matches nothing on the page". There are some use cases for getting called back when the number of elements matching the selector changes, but I'm not suggesting an interface for that yet. It might also be useful to publish this to the web through MutationObservers. I'm also not proposing that. :)

This continues https://bugs.webkit.org/show_bug.cgi?id=110524.

BUG=172011

Review URL: https://codereview.chromium.org/18371008

git-svn-id: svn://svn.chromium.org/blink/trunk@158582 bbb929c8-8fbe-4397-9dbb-9b2b20218538