[CSS Regions] Layout Test for selection direction
Patch already landed in WebKit. See http://trac.webkit.org/changeset/155974

Selection follows DOM tree, so the highlighted text in this test case is
the opposite to what the user is actually selecting. This was not
working properly from r139197 to r158274, so adding a Layout Test to
prevent future breaks. It is a reftest comparing regions and absolute
positions behavior.

Test: fast/regions/selection-direction.html

Review URL: https://codereview.chromium.org/25220002

git-svn-id: svn://svn.chromium.org/blink/trunk@158588 bbb929c8-8fbe-4397-9dbb-9b2b20218538