[blink] Store InCustomSpellcheckDictionary marker in document
This patch changes Blink to store InCustomSpellcheckDictionary marker in the
document after the client (the browser) returns spellcheck results that include
such a marker. Previously, Blink stored only Spelling and Grammar markers in the
document.

Blink keeps track of the marker locations in the document and sends them to the
browser along with the text to spellcheck. The browser uses the marker
information to improve spellcheck suggestions.

This patch should be the last of several for issue 295190 - "Add
InCustomSpellcheckDictionary marker to Blink for spellcheck feedback."

BUG=295190

Review URL: https://codereview.chromium.org/25144003

git-svn-id: svn://svn.chromium.org/blink/trunk@158648 bbb929c8-8fbe-4397-9dbb-9b2b20218538