Adding a function to request media element to enter fullscreen
For encrypted DRM video, a secured surface is required to playback the media.
On android, this requires the media element to enter fullscreen mode.
Adding a call from WebMediaPlayerClient to HTMLMediaElement so that this can happen.
Also fixed a problem that mediaFullscreenRequiresUserGesture is not honored when setting it to false.

BUG=297934,288678

Review URL: https://codereview.chromium.org/24482003

git-svn-id: svn://svn.chromium.org/blink/trunk@158694 bbb929c8-8fbe-4397-9dbb-9b2b20218538