Let a Document be disposed multiple times
Apparently a document can be disposed multiple times, which triggers an ASSERT
in the DocumentLifecycle. We appear to have been doing this for a while, so
I've relaxed the new ASSERT for now, but we should keep this bug open to track
this oddness.

BUG=301668

Review URL: https://codereview.chromium.org/25736003

git-svn-id: svn://svn.chromium.org/blink/trunk@158758 bbb929c8-8fbe-4397-9dbb-9b2b20218538