De-assign UseCounter sample IDs for -internal-* CSS properties.
As UseCounter does not count the usage of internal CSS properties, it makes no
sense to keep the sample IDs (IDs used in histograms) for these properties.

This refactoring intends to prevent unnecessary and unintended addition to
UseCounter sample IDs in the future.

This CL is a simple refactoring and does not intend to change any behavior.

BUG=none
R=jchaffraix@chromium.org

Review URL: https://codereview.chromium.org/25216002

git-svn-id: svn://svn.chromium.org/blink/trunk@158791 bbb929c8-8fbe-4397-9dbb-9b2b20218538