IDL compiler: [CustomElementCallbacks] for getter + cleanup values
[CustomElementCallbacks] is only ever called as [CustomElementCallbacks=Enable],
so the value is redundant.
This changes it to not take an argument.
Also the Perl CG has cruft from when 'Disable' was called 'None';
this decrufts it.

This also adds a test case to TestObjectPython.idl, though since this
extended attribute doesn't affect getters, no Python CG changes are needed;
will do when we do setters.

Doc:
https://sites.google.com/a/chromium.org/dev/blink/webidl/blink-idl-extended-attributes#TOC-CustomElementCallbacks-m-a-

BUG=239771

Review URL: https://codereview.chromium.org/26466003

git-svn-id: svn://svn.chromium.org/blink/trunk@159218 bbb929c8-8fbe-4397-9dbb-9b2b20218538