Refine the GlobalEventHandlers test
http://lists.whatwg.org/htdig.cgi/whatwg-whatwg.org/2013-October/thread.html#40980

The conclusion of the WHATWG "Should onfoo event handler properties be
on Element or HTMLElement?" thread was that not letting Element have
these properties is more sane, because reflecting the corresponding
content attributes for Element would be odd.

Update the test to assert that Element does not have the properties,
but that HTMLElement and SVGElement do.

Also split the enumerate tests so that individual results can be seen.

BUG=305112

Review URL: https://codereview.chromium.org/26478004

git-svn-id: svn://svn.chromium.org/blink/trunk@159478 bbb929c8-8fbe-4397-9dbb-9b2b20218538