Remove support for animating -webkit-text-fill-color
This patch removes support for the animation of -webkit-text-fill-color,
from both the existing and Web Animations implementations.

This property causes an issue with 'transition: all' which we work
around in the existing implementation by not animating unset colors,
which is wrong given that unset colors should be currentColor and
transitions happen on computed styles. For the new implementation,
we actually use computed styles for the animation of colors, so we can't
use the same hack here, and we want to move towards more correct
behaviour anyway.

Given that this property does not appear to be on track to be specified
in any CSS spec, is prefixed, and is extremely rarely animated (it is
in the same range as orphans and widows!), I would like to remove
support for animating this property.

BUG=

Review URL: https://codereview.chromium.org/26591004

git-svn-id: svn://svn.chromium.org/blink/trunk@159512 bbb929c8-8fbe-4397-9dbb-9b2b20218538