Do not layout on no-op list item :hover effects
A RenderListItem can contain a marker which depends on the list item
style. During style recalc, if the list item's style is unchanged, we
do not need to update the marker's style.

This patch fixes a bug on Wikipedia where a global :hover selector
would cause a new style to be set on markers which would in turn
cause an unnecessary layout.

BUG=306716

Review URL: https://codereview.chromium.org/27013008

git-svn-id: svn://svn.chromium.org/blink/trunk@159751 bbb929c8-8fbe-4397-9dbb-9b2b20218538