Length interpolation does not need to truncate progress to a float
The blending function for lengths in CSSPropertyAnimation.cpp was
truncating the progress argument from a double to a float, then expanding
back to a double in the actual interpolation (Length::blend). This
resulted in some results being off by very small decimal amounts,
particularly when moving before 0 as part of an easing function.

BUG=308251

Review URL: https://codereview.chromium.org/27513011

git-svn-id: svn://svn.chromium.org/blink/trunk@159852 bbb929c8-8fbe-4397-9dbb-9b2b20218538