Fixing bad usage patterns of WTF::RefPtr with SkRefCnt-derived objects
Before we can trun on run-time debug checks for RefPtr+SkRefCnt
adoption, we have to fix the cases where the checks are tripping.
None of the corrections in this patch fix actual memory leakis, but they
do fix fragile patterns in the code, like mixing OwnPtr with RefPtr on
the same object, and explicit deref() calls instead of adoptions.

BUG=304265

Review URL: https://codereview.chromium.org/27418002

git-svn-id: svn://svn.chromium.org/blink/trunk@159923 bbb929c8-8fbe-4397-9dbb-9b2b20218538