Make Document::renderer() private
Callers should use renderView() instead. It means in many cases virtual
calls to the RenderView will be devirtualized and it's also more clear
in the code what the caller is attempting to do.

I also took this as an opportunity to switch some callers to isActive()
that didn't really need to look at the renderView() and also clean up
some callers getting the renderView() in very round about ways.

Review URL: https://codereview.chromium.org/27310003

git-svn-id: svn://svn.chromium.org/blink/trunk@159956 bbb929c8-8fbe-4397-9dbb-9b2b20218538