Make *AutoResizeMode functions pure virtual.
WebTestDelegate should really only have pure virtual functions, and in cases where it doesn't, it's probably because they were a part of a three-sided patch: one in Blink to expose the API as an empty function, one in Chromium to hook it up, and one in Blink to replace empty function with a pure virtual one. This patch simply completes the last step in the process.

R=eseidel, tony
BUG=

Review URL: https://codereview.chromium.org/36803005

git-svn-id: svn://svn.chromium.org/blink/trunk@160392 bbb929c8-8fbe-4397-9dbb-9b2b20218538