Deflake 3 SVG tests that (tried to) use Ahem
This is an experimental patch to de-flake the following tests:
  svg/custom/repaint-shadow.svg
  svg/zoom/page/zoom-zoom-coords.xhtml
  svg/custom/font-face-cascade-order.svg

Each of these tests references Ahem and is flaky on OSX. In this
patch I've explicitly included Ahem via CSS so we don't rely on
the OSX font cache.

svg/custom/repaint-shadow.svg now requires a rebaseline because
the font now loads and causes some additional repainting.
The actual test is unaffected.

svg/zoom/page/zoom-zoom-coords.xhtml requires a rebaseline
because the Ahem font was not used at all before.

If this deflake works, an additional 5 SVG tests that reference
Ahem can also be deflaked.

Follow along the deflakification at
http://test-results.appspot.com/dashboards/flakiness_dashboard.html#tests=svg/custom/repaint-shadow.svg,svg/zoom/page/zoom-zoom-coords.xhtml,svg/custom/font-face-cascade-order.svg

BUG=311469

Review URL: https://codereview.chromium.org/43273002

git-svn-id: svn://svn.chromium.org/blink/trunk@160551 bbb929c8-8fbe-4397-9dbb-9b2b20218538