IDL compiler: remove unnecessary return from generated methods
Currently the Perl CG appends an unnecessary |return;| at the end of
generated methods (in GenerateFunctionCallString).
This is clear in the Python template, and thus we're removing it.
However, the return *is* necessary in GenerateParametersCheck,
where it's not in tail position, so this CL moves the return statement
into that function.

BUG=239771

Review URL: https://codereview.chromium.org/47513005

git-svn-id: svn://svn.chromium.org/blink/trunk@160705 bbb929c8-8fbe-4397-9dbb-9b2b20218538