ElementShadow should not check confusingAndOftenMisusedAttached() on ShadowRoot in the destructor
The only time we destroy ElementShadow objects is inside ~Element, and if we're running
an Element's destructor all of it's ShadowRoots have already been detached and all
element subtrees as well. There's no reason to check this, the code can never be reached
and we have a RELEASE_ASSERT in ~Node to ensure that renderers have already been
torn down.

Review URL: https://codereview.chromium.org/47923006

git-svn-id: svn://svn.chromium.org/blink/trunk@160773 bbb929c8-8fbe-4397-9dbb-9b2b20218538