Refactoring: Move Document::m_eventQueue to DOMWindow
This renames DocumentEventQueue to DOMWindowEventQueue,
and moves it from Document to DOMWindow.

This move is a preparation for Document-ExecutionContext
decoupling. This also migrates related member functions
from Document to DOMWindow.

Note that Document and DOMWindow have slightly different
lifetimes. This change keeps the queue lifetime same:
It matches one of the Document.

BUG=305497
R=haraken,abarth,dglazkov

Review URL: https://codereview.chromium.org/48433003

git-svn-id: svn://svn.chromium.org/blink/trunk@160793 bbb929c8-8fbe-4397-9dbb-9b2b20218538