Update outdated XHR layout test.
The explanation in the HTML file says the expected Content-Type is
application/xml, but it actually is text/plain. This is correct
according to the XHR spec, and application/xml was legacy WebKit behavior.

The HTML file also references a WebKit bug that specifies the old
behavior, and will be updated to reference a Chromium bug for the
correct behavior.

The test passed before and will still pass after this test. However,
running it manually in the browser will be less confusing.

BUG=172802

Review URL: https://codereview.chromium.org/48113012

git-svn-id: svn://svn.chromium.org/blink/trunk@160839 bbb929c8-8fbe-4397-9dbb-9b2b20218538