Fixed-position compositing code needs to compute layer bounds as if the layer is composited
When trying to determine whether a fixed-position element is offscreen,
we need to actually compute the layer bounds as if the layer would be
composited.

Before this patch, the layer bounds were being computed with whatever
the layer's current compositing state was, which caused fixed-position
layer with a transform to flip compositing state on every compositing
update.

BUG=293926

Review URL: https://codereview.chromium.org/50363003

git-svn-id: svn://svn.chromium.org/blink/trunk@160885 bbb929c8-8fbe-4397-9dbb-9b2b20218538