Remove Internals.attached and related test
This test and the method doesn't make any sense since any node in the main
document is attached (except in really special cases in mutation events).
If this test was trying to cover that ShadowRoots are correctly rendered
we have tons of tests for that now too, so this test is not needed or this
suspect attached() check.

Review URL: https://codereview.chromium.org/50263006

git-svn-id: svn://svn.chromium.org/blink/trunk@160887 bbb929c8-8fbe-4397-9dbb-9b2b20218538