ShadowRoot should not use confusingAndOftenMisusedAttached()
This check is not needed, instead we always mark a ShadowRoot as attached
if a recalcStyle reaches it. In the normal case it's not possible for a
recalcStyle to reach a ShadowRoot that is not attached but apparently it
can happen inside of ChildNodeInsertionNotifier before we call lazyAttach
on the tree with some kinds of events (which events is not clear).

Once the concept of attach is implicit this can all go away.

Review URL: https://codereview.chromium.org/45993007

git-svn-id: svn://svn.chromium.org/blink/trunk@160905 bbb929c8-8fbe-4397-9dbb-9b2b20218538