[M31] Fix crash in DoesRawIdMatchGuid
The given |device_guid| is provided from JavaScript, so it is not
guaranteed to be valid. This CL checks to make sure the hex to bytes
conversion succeeds before reading the value of the converted guid.

Note that this CL was already reviewed in
https://codereview.chromium.org/33743009/, so this is just
landing it in M31.

BUG=309098

Review URL: https://codereview.chromium.org/46573006

git-svn-id: svn://svn.chromium.org/chrome/branches/1650/src@231907 0039d316-1c4b-4281-b951-d872f2087c98