Add Frame::isMainFrame and deploy it
I originally started this to remove Page.h includes
but found that it just made the code simpler to read in most cases.

In an out of process iframes enabled Blink it will not be
possible to access the Page object from core/, so Frame will
have to know itself if it is (or thinks it is) the main frame.

There are quite a few uses of Page::mainFrame() in core/inspector/
code but it's not obvious that the frame being compared always has
the same page as the page its being compared with, so I didn't
change those for now.

BUG=305811

Review URL: https://codereview.chromium.org/59573002

git-svn-id: svn://svn.chromium.org/blink/trunk@161367 bbb929c8-8fbe-4397-9dbb-9b2b20218538