Experimental viewport meta tag support for desktop, Blink-side.
Added flag to allow desktop builds to experimentally turn
on support for the viewport meta tag. The --enable-viewport
flag now turns on only support for @viewport.
--enable-viewport-meta turns on viewport meta and @viewport
support.

On Android, the viewport gets rescaled when the view is
resized, i.e. phone is rotated between landscape and
portrait. On desktop, enabling the viewport causes the page
to zoom out when the window is shrunk, rather than adding
scrollbars. To support viewport on both devices, I've added
a --scale-viewport-on-resize flag that should be set on
Android and not desktop. This prevents the viewport scaling
code in WebViewImpl::resize().

Enabling viewport on desktop means the page scale can
change. This causes a known issue with scrollbars where
mouse events are scaled with the page, making mouse
interaction with scrollbars broken.

This is the Blink side of a 2-side patch.

BUG=232102

Review URL: https://codereview.chromium.org/40423003

git-svn-id: svn://svn.chromium.org/blink/trunk@161378 bbb929c8-8fbe-4397-9dbb-9b2b20218538