[Tizen] Change scrollbar animator
Based on patch from: Kenneth Rohde Christiansen <kenneth.r.christiansen@intel.com>

Since we have enabled overlay scrollbars we now get the desktop behavior, which means that a thin scroll indicator is always visible, which immediately grows when scrolling and shrinks (animated) back to the small size when scrolling is over.

On Tizen we want a different behavior:
a) no scroll indicator when not scrolling
b) a think scroll indicator which is slightly offset from the right viewport edge
c) the scroll indicator should fade out (animated) when the scroll is over

This patch is changing the scrollbar animator to 'fade out' and fixes 'a)' and 'c)' from the list above.