Don't force a style recalc just to clear the selection.
clearRenderViewSelection was a very silly function which
was copy-pasta'd into two files.  In both cases it took
a Position even though it only needed the Document*
and then null-checked RenderView (even though an Active
Document always has a RenderView) and then would update
style (unconditionally) only to clear the selection.

RenderView::setSelection won't update the rendering tree
so it is nice to pass it an up-to-date rendering tree
but in the case of clearing selection we don't care.

This changed one test result as a bi-product of removing
this synchronous style recalc.  My understanding is that
this test was triggering the insertion of empty text
nodes as a result of synchronous style recalc during parsing
(which has been the case in many tests in the past).
It's also possible that this empty text node was the result of
inserting the caret and forcing the creation of a text node
to hold the caret.

In either case, by deferring this unecessary recalc we have
now made this unecessary empty text node go away too.

R=esprehn,ojan
BUG=315854

Review URL: https://codereview.chromium.org/65683004

git-svn-id: svn://svn.chromium.org/blink/trunk@161608 bbb929c8-8fbe-4397-9dbb-9b2b20218538