Regression(r161367): Inspector crashes inspecting inactive document.
I incorrectly changed behavior of Docuemnt::updateViewportDescription
which caused this crash.

The inspector code is unfortunately both dangerous (using
internal style classes, like StyleResolver) and wrong
(resolving style on a non-active document).  (It's not the
inspector's fault, we haven't given them anything better
to use!)

I've both fixed the unintentional behavior change, as well
as fixed the inspector to not resolve style when the
document is inactive.

I've added an ASSERT(document->isActive()), which I suspect
we'll hit in other places, but those places are also wrong.

R=esprehn,pfeldman
BUG=316162

Review URL: https://codereview.chromium.org/62953009

git-svn-id: svn://svn.chromium.org/blink/trunk@161611 bbb929c8-8fbe-4397-9dbb-9b2b20218538