[svg] TearOffs dynamically hold on to the current contextElement().
Prior to this patch, TearOffs held on to contextElement() resolved at creation time. This was on based on an assumption that contextElement() never change. However, this assumption is wrong, as animatedProperty and its contextElement() can be dynamically assigned if an existing (detached) TearOff is made an item of SVG*ListProperty afterwards.

This patch fixes the issue by dynamically holding on to the current contextElement() by using SetReference inside resolveWrapperReachability callback.

BUG=317097

Review URL: https://codereview.chromium.org/64683006

git-svn-id: svn://svn.chromium.org/blink/trunk@161752 bbb929c8-8fbe-4397-9dbb-9b2b20218538