Add ASSERTs to Element::recalcOwnStyle
newStyle is never null, add an ASSERT about it. If oldStyle was null
then we should always go down the Reattach code path, so add an
ASSERT that oldStyls is not null if we go down the non-reattach path.

I then removed the null check for newStyle and oldStyle.

Review URL: https://codereview.chromium.org/60093004

git-svn-id: svn://svn.chromium.org/blink/trunk@161783 bbb929c8-8fbe-4397-9dbb-9b2b20218538