Fix FontPlatformData::operator ==
The equals operator in FontPlatformData incorrectly compares the m_font
field regardless of the GDI_FONTS_ON_WINDOWS. This is incorrect as it
should use the m_typeface field instead if GDI fonts are not used.

This works today as the m_font field is populated but breaks if it
isn't (which is the direction we are going).

Originally landed as r161467 but reverted due to breakage.

This version uses SkTypeface::Equal instead.

R=eseidel@chromium.org
BUG=

Review URL: https://codereview.chromium.org/25788002

git-svn-id: svn://svn.chromium.org/blink/trunk@161867 bbb929c8-8fbe-4397-9dbb-9b2b20218538