Use a boolean hasCompositedLayerMapping() accessor instead of the pointer
In order to add the ability to paint multiple RenderLayers into one
composited layer backing, we need to complexity the possible "states"
that a RenderLayer identifies as its compositingState. There are many
places in code that still use compositedLayerMapping() as a boolean
accessor to indicate compositingState, but doing so now is already
inadequate, and with layer squashing, it may even be wrong in many
places.

To gradually evolve the codebase away from the compositedLayerMapping()
accessor as a boolean, it is helpful to create a boolean accessor,
hasCompositedLayerMapping(). The purpose of doing this is to more easily
distinguish between places in code where compositedLayerMapping() is
actually used as an object versus places where it is used as a boolean.

BUG=261605

Review URL: https://codereview.chromium.org/58543002

git-svn-id: svn://svn.chromium.org/blink/trunk@161973 bbb929c8-8fbe-4397-9dbb-9b2b20218538