Web Animations: Use legacy cascade behaviour and the MatchedPropertiesCache
This patch changes the behaviour of the Web Animations backed animations
and transitions implementation to apply animations / transitions at the
last step of the cascade. This is the same behaviour as the legacy code
but not correct according to the spec[1]. The current behaviour is a bit
strange, it attempts to follow the spec but property dependencies (such
as using ems) do not work correctly as low priority properties will only
be re-applied if they are !important and we are running animations on a
low priority property.

As a result of moving animations / transitions to the end of the cascade
we can easily use the MatchedPropertiesCache for all running animations
and transitions. Once we correct the cascade order we'll need to disable
the cache in several cases.

[1] http://www.w3.org/TR/css3-cascade/#cascading

BUG=232273,314952

Review URL: https://codereview.chromium.org/71083002

git-svn-id: svn://svn.chromium.org/blink/trunk@161985 bbb929c8-8fbe-4397-9dbb-9b2b20218538