Remove TreatNullAs=NullString for <frame>, <frameset> and <iframe>
http://www.whatwg.org/specs/web-apps/current-work/multipage/obsolete.html#htmlframeelement
http://www.whatwg.org/specs/web-apps/current-work/multipage/obsolete.html#htmlframesetelement
http://www.whatwg.org/specs/web-apps/current-work/multipage/the-iframe-element.html#htmliframeelement
http://www.whatwg.org/specs/web-apps/current-work/multipage/obsolete.html#HTMLIFrameElement-partial

Per spec the longDesc attribute on HTMLFrameElement and
HTMLIFrameElement should be reflected as a URL, so make it so.

HTMLIFrameElement.sandbox is readonly DOMSettableTokenList in the spec,
so leave it with [TreatNullAs=NullString] until that is fixed.

The tests for the updated properties pass in Firefox Nightly, IE11
Release Preview and Opera 12.16, with these exceptions:

IE11 throws exceptions for HTMLFrameElement.scrolling,
HTMLIFrameElement.align and HTMLIFrameElement.scrolling, reflects
HTMLIFrameElement.width/height as the empty string and does not support
HTMLIFrameElement.srcdoc.

Opera does not support HTMLIFrameElement.srcdoc.

Note that no other browser supports the location attribute, but that has
not been touched here.

BUG=310298

Review URL: https://codereview.chromium.org/68503013

git-svn-id: svn://svn.chromium.org/blink/trunk@162147 bbb929c8-8fbe-4397-9dbb-9b2b20218538