Remove unused <webkitShadowContent> element from HTMLTagNames.in
In times gone by it used by <details> in place of the <content> element
because it was an "implementation detail". But no more. This is just
barely web-exposed: the only difference in behavior is that <webkitshadowcontent>
after this change will be an HTMLUnknownElement instead of an HTMLElement.

R=dglazkov

Review URL: https://codereview.chromium.org/76593002

git-svn-id: svn://svn.chromium.org/blink/trunk@162320 bbb929c8-8fbe-4397-9dbb-9b2b20218538