IndexedDB: Add support for "binary" keys in public API
Extend public and internal enums and types with support for "binary"
keys, i.e. using Uint8Array() as a key type. This will be exposed
to script behind an experimental features flag, but the Chromium
back-end support must be landed first.

(This is part 2 of a 4 part landing sequence.)

BUG=297304
R=alecflett@chromium.org

Review URL: https://codereview.chromium.org/25091005

git-svn-id: svn://svn.chromium.org/blink/trunk@162342 bbb929c8-8fbe-4397-9dbb-9b2b20218538