XHR: reset error flag on aborting request that restarts.
Close a small window in internal abort logic that takes care of
repeated uses of the XHR object upon cancellation.

The possibility existed that such a repeated use might reset the error
flag, so that upon return to the outer request (the one that is
aborting), it was left as 'false'. Spot this and (re) set the error
flag to 'true'.

Not a common use of XHR; added test coverage.

R=
BUG=315488

Review URL: https://codereview.chromium.org/76133002

git-svn-id: svn://svn.chromium.org/blink/trunk@162346 bbb929c8-8fbe-4397-9dbb-9b2b20218538