Web Animations CSS: Simplify interaction between CompositorAnimations and KeyframeAnimationEffect
This patch implements some methods in KeyframeAnimationEffect that gives CompositorAnimations direct access to things like the set of properties an animation controls. It reduces the surface area of CompositorAnimationsKeyframeEffectHelper to one method.

I suspect we might be able to remove this method too in a future patch.

Review URL: https://codereview.chromium.org/74883002

git-svn-id: svn://svn.chromium.org/blink/trunk@162441 bbb929c8-8fbe-4397-9dbb-9b2b20218538