[Linux] Enabling Hardware Acceleration with Libva in linux-x86.
This commit is not upstreamable.

Currently, VAAPI only enabled for ChromeOS and Windows on
upstream. Although the ChromeOS uses same libva API as
common linux world, upstream refuesed to enable it for linux
common with maintainence concern. Refer to
"https://codereview.chromium.org/16430003/" for more details.

Upstream confirm VAVDA will continue to be restricted to
CrOS/X86 for dev & testing only and not for chromium road map.
So we maintain it for Crosswalk and that will benifit
linux with VAAPI installed and includes Tizen.