IndexedDB: Remove IDBDatabaseBackendInterface/Proxy
Continued Source/web/*IDB* purging. This lets Source/modules
directly use the blink::WebIDBDatabase provided by the embedder.

The one subtle bit is a contract between Chromium and Blink
passing the WebIDBDatabase instance. In an upgrade scenario, 
onUpgradeNeeded(db) and onSuccess(db) are both called with
the same db object; obviously, the latter call is not an
ownership transfer. This was previously handled by re-using
the same refcounted proxy; now a flag is used and the pointer
is ignored. Once this lands we can update Chromium to pass null
in this case which will simplify both sides.

BUG=234277

Review URL: https://codereview.chromium.org/67503002

git-svn-id: svn://svn.chromium.org/blink/trunk@162569 bbb929c8-8fbe-4397-9dbb-9b2b20218538