[Temp] Export GpuProcessHost::AddFilter().
This patch was originally written for ozone-wayland
(https://github.com/01org/ozone-wayland) and is required in order to make
the build with -Dcomponent=shared_library work when building Crosswalk with
Ozone-Wayland support.

    commit 2e68f5c50e375551ac765ade0e7fda3184190f64
    Author: Kondapally Kalyan <kalyan.kondapally@intel.com>
    Date:   Fri Dec 20 17:43:37 2013 +0200

    Chromium patch to fix shared_library build.

    This patch adds needed changes in Chromium side to fix
    component=shared_library build for Ozone-Wayland.

According to the patch's author, this patch was not upstreamed because the
ozone-wayland code base is undergoing some refactoring that will make it
unneeded in the future.