Correctly apply per-keyframe timing functions
Currently, when a per-keyframe timing function is applied to
an element which is targeted by multiple animations, the
timing function may be applied to the wrong CSSAnimationData
object. This causes the per-keyframe timing function to be
ignored.

This patch fixes the problem by making sure the relevant
CSSAnimationData object is always first in the keyframe's
RenderStyle.

R=dstockwell
BUG=288540, 322680

Review URL: https://codereview.chromium.org/24085002

git-svn-id: svn://svn.chromium.org/blink/trunk@162672 bbb929c8-8fbe-4397-9dbb-9b2b20218538