Wrap CSS length conversion arguments in an object
This patch introduces a class CSSToLengthConversionData to wrap the data
required to convert CSS lengths to Lengths. This simplifies the plumbing
that goes on whenever we need to resolve CSS lengths and makes it easier
to update the arguments needed for resolving these (in particular adding
a RenderView for resolving viewport units at style recalc time; removing
the computingFontSize bool also appears possible).

Note that the zoom argument, which was previously a float in some places
and a double in others is now a float.

BUG=322365

Review URL: https://codereview.chromium.org/64293008

git-svn-id: svn://svn.chromium.org/blink/trunk@162881 bbb929c8-8fbe-4397-9dbb-9b2b20218538