Remove TreatNullAs=NullString for HTMLObjectElement
http://whatwg.org/html#htmlobjectelement
http://whatwg.org/html#HTMLObjectElement-partial

Make the codebase attribute reflect as a URL, as per spec.

The border attribute has [TreatNullAs=EmptyString] in the spec.

The tests for the updated properties pass in Firefox Nightly, IE11
Release Preview and Opera 12.16, with these exceptions:

Firefox reflect code as a URL.

IE throws an exception for align, doesn't reflect codeBase as a URL,
and reflects width/height as the empty string.

BUG=310298

Review URL: https://codereview.chromium.org/99233002

git-svn-id: svn://svn.chromium.org/blink/trunk@163037 bbb929c8-8fbe-4397-9dbb-9b2b20218538