Use downloadToFile option when XHR downloads a Blob
Leveraging downloadToFile when XHR downloads a Blob. As a result, a Blob
is downloaded in the browser. It eliminates unnecessary memory copies
between the renderer and the browser.

When downloadToFile options is specified, didDownloadData is called
instead of didReceiveData. So not to break the Inspector, we add the
inspector instrumentations to it.

BUG=269055

Review URL: https://codereview.chromium.org/23444058

git-svn-id: svn://svn.chromium.org/blink/trunk@163141 bbb929c8-8fbe-4397-9dbb-9b2b20218538