Prevent ASSERTION FAILED !(len & kPageAllocationGranularityOffsetMask) on win.
This can only happen on windows, where kPageAllocationGranularity is 64K. To
ensure this isn't violated, we can't ask for N-1 additonal system pages when
the alignement fails (although that is sufficient from a space perspective to
cover the desired address).  So just ask for N.

This assertion is preventing a webkit roll.
TBR=cevans@chromium.org

Review URL: https://codereview.chromium.org/107133002

git-svn-id: svn://svn.chromium.org/blink/trunk@163295 bbb929c8-8fbe-4397-9dbb-9b2b20218538