Revert "Make arguments to Element methods non-optional"
This reverts commit cebc60bc4d928b4b5217a162c51b3150966d3455.

This patch seems to break a number of tests in Chromium on
linux_chromeos. I'd guess that they're written in such a way
as to rely on the current (incorrect) behavior, and will need
to be tweaked to provide all the arguments they're supposed to
provide.

See [1] for an example.

[1]: http://build.chromium.org/p/tryserver.chromium/builders/linux_chromeos/builds/182031

TBR=philipj@opera.com, arv@chromium.org
BUG=325922
NOTRY=true

Review URL: https://codereview.chromium.org/106493004

git-svn-id: svn://svn.chromium.org/blink/trunk@163431 bbb929c8-8fbe-4397-9dbb-9b2b20218538