See also webkit.org/b/113415
Include the logical bottom margin in the layout overflow rect. In the most common use
case this ensures we can scroll down to the margin that sits after any element that has it.
The only cases we need to avoid are where the margin is present but doesn't contribute height.
These are (1) where the element has a margin quirk (i.e. the margin is
defined with quirky em values ('__qems')) and so is disregarded in margin collapsing and (2) when
the element is self-collapsing so the entire element adds no height of its own.

This fix revealed a separate bug where collapsed margins from the content of a table caption
were not contributing to the height of the caption and its parent. This is tested in
fast/forms/form-hides-table.html.

BUG=320743

Committed: https://src.chromium.org/viewvc/blink?view=rev&revision=163733

Review URL: https://codereview.chromium.org/75593002

git-svn-id: svn://svn.chromium.org/blink/trunk@163807 bbb929c8-8fbe-4397-9dbb-9b2b20218538