Add support for Mac 10.9 (Mavericks) to run-webkit-tests.
This change adds support for 10.9; it recognizes the operating system
version properly, updates the fallback paths for baselines to
recognize the new 'platform/mac-mountainlion' directory, and adds support for a 'Mavericks' keyword in the TestExpectations syntax.

This change does not support 10.9 on Retina displays yet; I'll probably
make that work in a separate change, although it's unclear if we really
need to support Retina displays on both 10.9 and 10.8 or if we should
just support 10.9 and call that good enough.

This change also does not add any new baselines or TestExpectations,
so actually running the tests on 10.9 will produce ~750 failures (nearly
all due to minor pixel diffs in the native scroll bars ... sigh ...).

I will upload new baselines for that separately, shortly after this lands.

R=abarth@chromium.org, ojan@chromium.org
BUG=322749

Review URL: https://codereview.chromium.org/99513006

git-svn-id: svn://svn.chromium.org/blink/trunk@163814 bbb929c8-8fbe-4397-9dbb-9b2b20218538