Crash reparenting a floating inline list item without a list
When a list item isn't in a list, without adjusting the style we instruct
the renderer to behave as if its marker is inside. Floats are layed out
late in line layout, and if the list item is moved and the marker already
created, updateMarkerLocation would mark the chain through its containing
block as needing layout towards the end of that block's layout, and
therefor with stale pointers.

There's no reason to defer calling updateMarkerLocation in this case.
This change gets the marker's renderer properly positioned during
attach.

BUG=301196

Review URL: https://codereview.chromium.org/75683002

git-svn-id: svn://svn.chromium.org/blink/trunk@164210 bbb929c8-8fbe-4397-9dbb-9b2b20218538