DevTools: network log bars use CSS instead of pre-rendered images.
This change removes the pre-rendered PNG images used by the bars shown
in the timeline in the Network panel, and relies on CSS to render the
bars. The colors in the CSS rules are picked from the PNG files that
they replace.

The immediate benefit of this change is that the Network view looks more
crisp on Retina displays and other high-resolution displays. A secondary
benefit is that adding new resource types (and thus new bar colors)
should be quite a bit easier.

BUG=330659
TEST=Browsed to a few sites in Chromium with the Network panel open.

Review URL: https://codereview.chromium.org/99923009

git-svn-id: svn://svn.chromium.org/blink/trunk@164362 bbb929c8-8fbe-4397-9dbb-9b2b20218538