[SysApps] Expose ffmpeg function signatures needed by Device Capabilities API
WARNING: This is a follow-up patch fixing Windows build. Squash this with the
previous patch with the same name on the next rebase (and remove this warning).

We need these functions in to get the list of Codecs available in the
system for getAVCodecs(). If we ever fork ffmpeg, we can just remove
this patch and add the signatures directly at the ffmpeg's signature
list.