Transition Between Shape Properties
The transition code checks to see if the start and end values are equal
before blending the two. The ShapeValue code must provide a more accurate
comparision to make this possible, as the previous comparision was only
based on the ShapeValue's type. Supplying a test required some minor fixup
work to animation-test-helpers, as the syntax for shapes has changed.

BUG=304916

Review URL: https://codereview.chromium.org/124363002

git-svn-id: svn://svn.chromium.org/blink/trunk@164533 bbb929c8-8fbe-4397-9dbb-9b2b20218538