Avoid layout/full-repaint on view height change if possible
Previously we always layout/full-repaint of the whole view when the height
of the RenderView changes. In many cases this is unnecessary. Especially,
this causes unnecessary cost on Android when the top control shows or
hides.

Avoid repaint if the RenderView doesn't needs full layout.

BUG=258219
BUG=327815(about RenderBlock::percentHeightDescendants)

TEST=WebFrameTest.heightChangeRepaint
TEST=all layout tests still pass

Review URL: https://codereview.chromium.org/99663004

git-svn-id: svn://svn.chromium.org/blink/trunk@164645 bbb929c8-8fbe-4397-9dbb-9b2b20218538