Web Animations: Animate pseudo elements where the animation starts immediately
In order to determine whether a pseudo element needs to be created, we
determine the style of the potential pseudo element and check various
properties of this style. This means that during this style recalc, the pseudo
element does not yet exist.

This causes problems for Web Animations, which generates animations during
style recalc. This process requires the pseudo element, both to resolve
animation keyframes and as the owner of the new animation.

This change fixes the problem by splitting up the process of creating a pseudo
element to allow Web Animations to calculate the animation update and store it
on the newly created element.

BUG=258896,328074

Review URL: https://codereview.chromium.org/125723002

git-svn-id: svn://svn.chromium.org/blink/trunk@164667 bbb929c8-8fbe-4397-9dbb-9b2b20218538