[SVG] Refactoring the logic to handle eventBase dependency
Our problem here is we want the dependent element to get notified whenever a change occurs on eventBase.
The current logic to handle this scenario is very inefficient as it is re-connecting the conditions on every loop.
Instead, the same can be handled using SVGDocumentExtensions pattern to improve efficiency.

syncBases and eventBases are functionally different hence separating their code paths to handle them properly.

R=pdr@chromium.org, schenney@chromium.org, dschulze@chromium.org, fmalita@chromium.org, rob.buis@samsung.com
BUG=323915

Review URL: https://codereview.chromium.org/102353003

git-svn-id: svn://svn.chromium.org/blink/trunk@164697 bbb929c8-8fbe-4397-9dbb-9b2b20218538