Verify document active before doing layout.
This fixes up an issue where when a document contains a focused pepper plugin,
when the document is destroyed it is possible to trigger a layout on that
document. Since the document is being torn down some pieces, such as the
renderer, may already have been removed. When we attempt to do layout this will
cause issue as we attempt to access the renderer.

BUG=331889

Review URL: https://codereview.chromium.org/129283002

git-svn-id: svn://svn.chromium.org/blink/trunk@164771 bbb929c8-8fbe-4397-9dbb-9b2b20218538