Don't cycle through subsequent siblings in clearFloatsIfNeeded()
When we encountered a self-collapsing block with clearance we were
cycling through subsequent siblings to determine if they were all self-collapsing.
That being the case we set a bit in MarginInfo to prevent the parent collapsing
its margins with the last child. It's better to set a bit when we encounter
a self-collapsing block with clearance and then reset it if we encounter a
block with height that allows the parent to collapse margins again.

This will allow us to ASSERT(!selfNeedsLayout()) in RenderBlock::isSelfCollapsingBlock().

BUG=

Review URL: https://codereview.chromium.org/124483002

git-svn-id: svn://svn.chromium.org/blink/trunk@164819 bbb929c8-8fbe-4397-9dbb-9b2b20218538