Implement Node.lookupPrefix per DOM4.
The spec for this is here:
<http://dom.spec.whatwg.org/#dom-node-lookupprefix>

This simplifies lookupNamespacePrefix, renames it to
locateNamespacePrefix to match the language in the DOM4 spec, and
pushes it down to Element which is the only kind of node for which it
makes sense.

This also removes Node::prefix. Callers that want to know prefixes
should be aware when they're dealing with something with a name (that
is, an element or an attribute.)

BUG=325922

Review URL: https://codereview.chromium.org/126713005

git-svn-id: svn://svn.chromium.org/blink/trunk@164853 bbb929c8-8fbe-4397-9dbb-9b2b20218538