Improve a search logic to find a SVGFontFaceSrcElement
In SVGFontFaceElement::rebuildFontFace(), it tries to find a SVGFontFaceSrcElement. However, current search logic looks inefficient. The search logic starts to find SVGFontFaceSrcElement from a first child node, then it replaces a discovered element whenever finding next SVGFontFaceSrcElement. Thus, it would be good if we find SVGFontFaceSrcElement from a last node. Then, we go out of the searching loop as soon as we find a first SVGFontFaceSrcElement. It will help to reduce iteration number in the search loop.

Additionally the search loop is moved to the place which uses a SVGFontFaceSrcElement actually.

BUG=N/A

Review URL: https://codereview.chromium.org/133493002

git-svn-id: svn://svn.chromium.org/blink/trunk@164881 bbb929c8-8fbe-4397-9dbb-9b2b20218538