Regularize extended attributes on partial interfaces
This is cleanup CL, as part of implementing partial interfaces in the
Python IDL compiler.
Primarily it moves overall [RuntimeEnabled] extended attributes (and
[Conditional] and [PerContextEnabled]) to the partial interface itself,
rather than individual members, so:

[
　　RuntimeEnabled=FeatureName,
] partial interface Foo {
　　void a();
　　void b();
};

instead of:

partial interface Foo {
　　[RuntimeEnabled=FeatureName] void a();
　　[RuntimeEnabled=FeatureName] void b();
};

This is for consistency: these extended attributes can be applied
to the partial interface itself, which means they are applied to
all members, so we can switch it on or off with one line.

Currently some IDL files put this on the partial interface, while
others put it on all members; this makes it consistent and clarifies intent.

(We could remove all of these, simplifying merging a bit, at the cost
of putting these on all the individual member lines.)

Docs:
https://sites.google.com/a/chromium.org/dev/blink/webidl/blink-idl-extended-attributes#TOC-Partial-interfaces

No changes to generated code.

R=haraken

Review URL: https://codereview.chromium.org/130943004

git-svn-id: svn://svn.chromium.org/blink/trunk@165049 bbb929c8-8fbe-4397-9dbb-9b2b20218538