PartitionAlloc: wait just a little while before actually freeing empty pages.
We hold a small ring-list cache of recently-emptied pages, and only free the
oldest entry in the ring list when the cache is full. This makes us hold on to
a little bit of unused memory, but in return we do now eventually free the last
slot span in any given bucket, which would have been held indefinitely before.

This patch is a sufficient generic replacement for the bandaid added in
https://code.google.com/p/chromium/issues/detail?id=330798; accordingly, the
bandaid is removed. In fact, we now seem to be +4% faster than _before_ the
original regression.

BUG=332282

Review URL: https://codereview.chromium.org/133863006

git-svn-id: svn://svn.chromium.org/blink/trunk@165112 bbb929c8-8fbe-4397-9dbb-9b2b20218538