Revert of Copy WebFrameImpl.{cpp,h} to TextFinder.{cpp,h} (https://codereview.chromium.org/119113003/)
Reason for revert:
The follow-up patch was reverted.

I guess it would be easier to do the copy and refactoring in one CL, so we don't end up with a copy of WebFrameImpl.* that gets stale before the refactoring is finished.

Original issue's description:
> Copy WebFrameImpl.{cpp,h} to TextFinder.{cpp,h}
> 
> This patch copies Source/web/WebFrameImpl.{cpp,h} into Source/web/TextFinder.{cpp,h} for decoupling WebFrameImpl and text finder.
> 
> Another patch will complete decoupling.
> 
> BUG=
> 
> Committed: https://src.chromium.org/viewvc/blink?view=rev&revision=164490

TBR=tkent@chromium.org,iceman@yandex-team.ru
NOTREECHECKS=true
NOTRY=true
BUG=

Review URL: https://codereview.chromium.org/139383002

git-svn-id: svn://svn.chromium.org/blink/trunk@165120 bbb929c8-8fbe-4397-9dbb-9b2b20218538