Change string version of constructTextRun to compute text direction
Change the string version of RenderBlockFlow::constructTextRun to either
take a text direction parameter or to compute the direction based on the
string if no explicit direction was specified. This in turn allows form
controls, ellipsis boxes and other users of this API to get accurate
metrics for strings where the width is dependent on the directionality.

TEST=fast/text/international/text-overflow-ellipsis-bidi.html
BUG=333004
R=eseidel@chromium.org

Review URL: https://codereview.chromium.org/138143008

git-svn-id: svn://svn.chromium.org/blink/trunk@165174 bbb929c8-8fbe-4397-9dbb-9b2b20218538