Revert the feature of the <shadow> element as 'a function call' to the previous behavior.
The spec bug is:
https://www.w3.org/Bugs/Public/show_bug.cgi?id=24288

The change of the spec is:
https://github.com/w3c/webcomponents/commit/61745e0b40c8d609ad1070e7babc7915fca637ea

Regarding with https://www.w3.org/Bugs/Public/show_bug.cgi?id=22344, it turned
out that there is a kind of a technical difficulty to implement this feature
correctly [1].

It might be better to revert it to the previous behavior until we find a reasonable
idea of the implementation.

We will also remove fallback ability of <shadow> element. Child elements of
<shadow> elements will be simply ignored. We preserve the usage of child
elements of the <shadow> elements for future use, such as enabling the new
behavior again.

[1]: https://codereview.chromium.org/77863002/

BUG=334152

Review URL: https://codereview.chromium.org/137993003

git-svn-id: svn://svn.chromium.org/blink/trunk@165194 bbb929c8-8fbe-4397-9dbb-9b2b20218538