Mark private method layoutBlockFlow as inline
Mark private method layoutBlockFlow as inline since it is only used in
the implementation file and this appears to be performance critical code.  This
is an attempt to fix a regression introduced by r165254.  This also reduces the
binary size on clang by 40 bytes.

BUG=33574
R=esprehn@chromium.org, ch.dumez@samsung.com

Review URL: https://codereview.chromium.org/140323004

git-svn-id: svn://svn.chromium.org/blink/trunk@165419 bbb929c8-8fbe-4397-9dbb-9b2b20218538