[webcrypto] Remove code supporting Key.algorithm.
Now the reflected Key.algorithm has ONLY a "name" property and nothing else.

The issue is that the spec [1] doesn't define what interfaces should be used to reflect back the Key's algorithm, and there are several interpretations.

Previously the Blink code was reflecting the same Algorithm as was used to import the key, or generate the key. This is awkward since it gives different parameters depending on whether the key was generated or imported (imported keys will not indicate RSA key sizes, DH groups, etc).

Rather than try to invent new interfaces for this purpose, this changelist limits Key.algorithm to JUST having a "name" property and no other parameters.

The intention is to add back a richer Key.algorithm once this has been added to the spec (so there is clarity on what the interface names should be, what normalizations to perform, and what properties to expose).

Until then, I am deleting the ambiguous code to make this more maintainable, and avoid exposing inconsistent key.algorithms to consumers of WebCrypto.

[1] https://dvcs.w3.org/hg/webcrypto-api/raw-file/tip/spec/Overview.html#key-interface

BUG=245025

Review URL: https://codereview.chromium.org/144093009

git-svn-id: svn://svn.chromium.org/blink/trunk@165519 bbb929c8-8fbe-4397-9dbb-9b2b20218538