[Android] Support app scheme in media player
This a hot fix on the bug that the audio/video can not be loaded
with app scheme.
The resource loading of media player on Android is executed by
Android system media player(setDataSource), which goes a total
different way with other ports(Chromium's own network stack). The
interaction with system media player is in Chromium, so we must do
the modification in Chromium, and can not in xwalk instead.

BUG=https://crosswalk-project.org/jira/browse/XWALK-880