Remove the harmful description('xxx') in LayoutTests/fast/dom/shadow/resources/event-dispatching.js.
event-dispatchig.js, which is intented to be included from other layout tests, contains the statement of "description('xxx')".
That is harmful because it may create <div id='console'> unintentionally as the body's first child element by just including this JS.
That will cause the tests which depends on element's position behave wrongly.

Removed it and rebased the affected tests.

TBR=tasak,mkwst

Review URL: https://codereview.chromium.org/130983004

git-svn-id: svn://svn.chromium.org/blink/trunk@165830 bbb929c8-8fbe-4397-9dbb-9b2b20218538