Avoid doing malloc in a loop in RenderFlexibleBox::layoutFlexItems
When profiling the silk infinite_scroll benchmark we spend about 4ms in
layout on the flex boxes. Looking at the profile on my mac 2% of that was
inside malloc on a Vector<LayoutUnit>.

This patch adds some stack allocated storage for the Vector and also makes
the Vector be sized and reused so we don't need to malloc it over and
over inside the loop.

Review URL: https://codereview.chromium.org/148763003

git-svn-id: svn://svn.chromium.org/blink/trunk@165929 bbb929c8-8fbe-4397-9dbb-9b2b20218538