Remove SVGDocument binding
The SVG2 specification removes the SVGDocument interface and instead suggests the SVG document
to be a partial implementation on Document. This patch does that except it does a partial implement
of XMLDocument, since an SVG document is also an XML Document according to DOM4.
The createEvent specialization for SVG is now handled by Document. The rootElement method is still
on the native SVGDocument for now.
The GetSVGDocument interface now returns a Document instead of an SVGDocument as specified in SVG2 spec.

BUG=238366

Review URL: https://codereview.chromium.org/146193003

git-svn-id: svn://svn.chromium.org/blink/trunk@166179 bbb929c8-8fbe-4397-9dbb-9b2b20218538