Disable text-autosizing when screen size is unavailable.
Currently, in Android Chrome, TextAutosizer always computes inf multiplier
for every page on initial rendering because screen dimension is not available at
this point, though the result is not actually displayed until relayout happens.

I'm not sure it is the case too in other platforms (win / linux / mac) that
screen dimension is not available. I discussed this with yutak@ and he pointed
out that, even though we're not sure if this is Android-specific or not,
rendering code should be expected to work even if main frame dimension is zero
because:
- FrameView::create() creates a frame with zero size by default
https://code.google.com/p/chromium/codesearch#chromium/src/third_party/WebKit/Source/core/frame/FrameView.cpp&sq=package:chromium&l=175&type=cs
- Some callers, at least SVGImage, creates a frame with default size
https://code.google.com/p/chromium/codesearch#chromium/src/third_party/WebKit/Source/core/svg/graphics/SVGImage.cpp&sq=package:chromium&l=397&rcl=1391483731&type=cs

Review URL: https://codereview.chromium.org/150073003

git-svn-id: svn://svn.chromium.org/blink/trunk@166476 bbb929c8-8fbe-4397-9dbb-9b2b20218538