Avoid double-painting the viewport background.
RenderView has code to paint its background when the
root renderer doesn't cover the viewport. RenderBox
later had code added to fix a case when the root is
a table (webkit.org/b/47147). This later code
obsoleted the code in RenderView for CSS boxes and
causes us to unnecessarily double-paint.

Cleaning up the code (and incorrect comments) in
RenderView, and only considering the size of the
root renderer if it's SVG.

BUG=266925

Review URL: https://codereview.chromium.org/149393007

git-svn-id: svn://svn.chromium.org/blink/trunk@166582 bbb929c8-8fbe-4397-9dbb-9b2b20218538