Remove input.type = ""; quirk.
Do not treat setting input.type to the empty string as a special case
(i.e., as removing the attribute), just update the underlying
attribute as for any other value.

This aligns behavior with Gecko and Presto (and spec); Trident treats
such as a no-op, which in any case was incompatible with the previous
behavior here of removing the attribute. i.e., the test added passes
for all but Trident.

R=tkent@chromium.org
BUG=
TEST=fast/forms/input-type-change3

Review URL: https://codereview.chromium.org/141703028

git-svn-id: svn://svn.chromium.org/blink/trunk@166593 bbb929c8-8fbe-4397-9dbb-9b2b20218538