Make the cookie-flood fit in 256KB.
The new WebSocket implementation ignores headers which are > 256KB, so
for the cookie flood test to be meaningful the headers need to fit.

Reduce the number of Set-Cookie headers output to fit in 256KB. It
should still be enough to overflow recursion-based implementations, at
least on 32-bit platforms.

BUG=339373

Review URL: https://codereview.chromium.org/154113005

git-svn-id: svn://svn.chromium.org/blink/trunk@166607 bbb929c8-8fbe-4397-9dbb-9b2b20218538