Fix WorkerGlobalScopePerformance thread state mismatch asserts.
A handful of worker tests (in webexposed/ and http/tests/w3c/webperf/)
currently assert and fail.

The Persistent reference to WorkerPerformance from
WorkerGlobalScopePerformance requires that the worker thread looks up
and uses the current thread state rather than that of the main thread.
That state is associated with the Persistent object when created.

R=
BUG=340522

Review URL: https://codereview.chromium.org/158513002

git-svn-id: svn://svn.chromium.org/blink/trunk@166791 bbb929c8-8fbe-4397-9dbb-9b2b20218538