Avoid requiring a recalc to update compositing state for iframes
r164013[1] first exposed this issue when it stopped issuing
StyleChangeFromRenderer in some cases.

The solution here is:
1. iframes always generate RenderLayers. This allows the
compositing update to consider the iframe in the same pass.
2. When the compositing update detects a change it requests
a requirements update in the enclosing frame. In the full
compositing update this will happen in the same pass.

[1] https://src.chromium.org/viewvc/blink?revision=164013&view=revision

BUG=332024

Review URL: https://codereview.chromium.org/149653003

git-svn-id: svn://svn.chromium.org/blink/trunk@166807 bbb929c8-8fbe-4397-9dbb-9b2b20218538