[FastTextAutosizer] Cleanup FastTextAutosizer::enabled() checks.
This patch is a followup to [1] to check enabled() on all public
functions of FastTextAutosizer. In [1] I lost an enabled() check by
mistake so I have put it back (in LayoutScope) and added asserts in
beginLayout/endLayout. I've also filed crbug.com/342260 to reduce the
number of calls to enabled().

No change in behavior as this was only observable while hacking.

[1] https://src.chromium.org/viewvc/blink?view=rev&revision=166602

BUG=342252

Review URL: https://codereview.chromium.org/158463004

git-svn-id: svn://svn.chromium.org/blink/trunk@166850 bbb929c8-8fbe-4397-9dbb-9b2b20218538