[New multicol] Avoid 1px tall multicol if percent height is unresolvable.
If a box has a percentage height, but the computed height of its
containing block is auto, the computed height of the box also becomes
auto. computeContentLogicalHeight() returns -1 if the height isn't
resolvable, and we need to make sure that such a value doesn't
constrain the height of the multicol container.

This is a merge of https://bugs.webkit.org/show_bug.cgi?id=122826
Reviewed in WebKit by David Hyatt.

BUG=342779

Review URL: https://codereview.chromium.org/144463015

git-svn-id: svn://svn.chromium.org/blink/trunk@166974 bbb929c8-8fbe-4397-9dbb-9b2b20218538