Refactor CSSDefaultStyleSheets from a statics-only class into a singleton.
It used to host statics of primitive pointer type but for OilPan we need those statics to become persistent roots and thus we can't simply have them in static variables anymore.

This CL also cleans up some FIXMEs related to CSSDefaultStyleSheets object: loading of default styles is moved into constructor which is invoked lazily when instance is accessed for the first time and loadDefaultStylesheetIfNecessary is killed.

BUG=

Review URL: https://codereview.chromium.org/133083005

git-svn-id: svn://svn.chromium.org/blink/trunk@166995 bbb929c8-8fbe-4397-9dbb-9b2b20218538