Fixup SVGSVGElement.createSVGTransformFromMatrix IDL definition
The optional SVGMatrix with a default of undefined is nonsensical, and
is not in line with the spec (where the argument is not optional).

Remove the 'optional' qualifier and the default value. This aligns
behavior with Gecko. (The exception thrown now is about arity, and not
about type mismatch as previously.)

As a result, the move of the [StrictTypeChecking] attribute to the
interface can be completed.

Review URL: https://codereview.chromium.org/149603007

git-svn-id: svn://svn.chromium.org/blink/trunk@167014 bbb929c8-8fbe-4397-9dbb-9b2b20218538