Repaint-after-layout ASSERTs with nested layouts
What would happen is that:
- the inner layout would call repaintTree, thus clearing everything.
- the outer layout would call repaintTree on a cleared tree, thus
exploding.

While this shouldn't have caused user visible issue, it was
inefficent. The fix is to ensure that we only ever call repaintTree
for the outer-most layout, this requires to refactor the existing
code to properly handle nesteds layout instead of generating some
extra repaints.

BUG=320139

Review URL: https://codereview.chromium.org/157033006

git-svn-id: svn://svn.chromium.org/blink/trunk@167044 bbb929c8-8fbe-4397-9dbb-9b2b20218538