Support invalidation sets for simple class selectors.
Invalidation sets were supported for selectors like '.a .b' with an empty
invalidation set for 'b' and a LocalStyleChange for elements for which 'b'
was added or removed. However, not for simple class selectors like '.b' for
which an empty invalidation set would be added, but also marked with
wholeSubtreeInvalid yielding a SubtreeStyleChange instead.

Now, start using invalidation sets for class selectors in the rightmost
compound selector even when there are no other class selectors present.

R=chrishtr@chromium.org,ojan@chromium.org
BUG=335247

Review URL: https://codereview.chromium.org/171513006

git-svn-id: svn://svn.chromium.org/blink/trunk@167470 bbb929c8-8fbe-4397-9dbb-9b2b20218538