Add frame color option to packaged app windows.
If app windows are created with a frame color option, that color is used
for the frame. If no color is supplied a native style look is used
instead. This is currently implemented on linux aura and windows only.

This change also removes the --apps-use-native-frame flag on windows.

BUG=339558

Review URL: https://codereview.chromium.org/166443004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@252548 0039d316-1c4b-4281-b951-d872f2087c98