The purpose of this cl is to lazy instantiate webrtc/libjingle MediaStream and video tracks that are created locally. The end goal is to not use libjingle interfaces until a MediaStream is added to a PeerConnection.
This cl renames MediaStreamTrackExtraData and MediaStreamExtraData to MediaStreamTrack and MediaStream.
Both MediaStreams and MediaStreamTracks gets a pointer to 
the one and only MediaStreamDependencyFactory. 
The instantiation of all webrtc MediaStream classes are done in MediaStreamDependencyFactory.  
Instantiation of the webrtc / libjingle MediaStreams happen 
when MediaStream::GetAdapter or MediaStreamVideoTrack::GetAdapter is called the first time.

MediaStreamCenter is updated to not call directly into the libjingle classes but instead use the chrome representations of streams and tracks.

BUG=323223

Review URL: https://codereview.chromium.org/139343014

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@252562 0039d316-1c4b-4281-b951-d872f2087c98