[Sync] Fix handling of dropped favicons to respect type
Previously, we would locally drop both the image and tracking part of a favicon
if it was unsynced. But, the notion of unsynced is per-type, so this can be
incorrect if only part of a favicon is available. Now, drop logic is per type
as well.

BUG=335493

Review URL: https://codereview.chromium.org/171403006

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@252702 0039d316-1c4b-4281-b951-d872f2087c98