sync: Unregister SyncInternalsMessageHandler
Unregisters SyncInternalsMessageHandler as an observer of
ProfileSyncService events during the object's destruction.

Fixes a serious bug that caused a lot of crashes on canary.  If we did
not unregister, then it would be possible to trigger a crash by opening
about:sync, closing it, then doing something that triggers a sync cycle
(eg. navigating to a new page or adding a bookmark).

NOTRY=true
BUG=346244

Review URL: https://codereview.chromium.org/178003005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@253009 0039d316-1c4b-4281-b951-d872f2087c98