Clean up WebContentses in the PrerenderManager on profile shutdown.
This avoids hitting a DCHECK on shutdown; by the time the PrerenderManager is
destroyed, there should already be no more WebContentses for normal profiles.
Also fix match-complete logic to not create match-completes for profile
shutdown; those will happen equally in the control group.

This, along with a fix to watch resource requests instead of title changes,
unbreaks PrerenderUnload. (content::TitleWatcher cannot outlive the
WebContents it observes.)

Add a new test which forces the case which used to flakily fail to assert it
works now.

BUG=128986

Review URL: https://codereview.chromium.org/165983008

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@253090 0039d316-1c4b-4281-b951-d872f2087c98