Consider text alignment and direction when computing the left offset for horizontal writing modes.
To properly layout a positioned box, we need to consider special cases where the direction of
the text for that object flows towards the opposite direction the text inside that box would be
aligned to (e.g. right aligned text for LTR boxes and left aligned text for RTL boxes).

Not doing this this ways yields to undesired effects, such as having the absolutely positioned
text inside a table cell with relative positioned being wrongly rendered outside of the cell.

BUG=157539
TEST=Try rendering absolutely positioned text inside table cells with relative positioning.
R=eae@chromium.org

Review URL: https://codereview.chromium.org/150403003

git-svn-id: svn://svn.chromium.org/blink/trunk@167732 bbb929c8-8fbe-4397-9dbb-9b2b20218538