Move LiveNodeListBase class to its own header
Move LiveNodeListBase class to its own header instead of having it in the
LiveNodeList header. HTMLCollection also inherits from LiveNodeListBase
and is not interested in the LiveNodeList class.

There is still LiveNodeListBase code in HTMLCollection.cpp, this will be
moved in a follow-up patch.

R=esprehn, adamk

Review URL: https://codereview.chromium.org/177213010

git-svn-id: svn://svn.chromium.org/blink/trunk@167821 bbb929c8-8fbe-4397-9dbb-9b2b20218538