sync: null-check SyncedTabDelegate in NotificationServiceSessionsRouter
In every case but one we were null-checking the SyncedTabDelegate returned
from the NotificationDetails, except for one, and that case is crashing on
a null pointer.  There's no guarantee that the tab delegate be there, so
this is a speculative but relatively obvious fix.

BUG=347220
R=maniscalco@chromium.org, zea@chromium.org

Review URL: https://codereview.chromium.org/179833007

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@253681 0039d316-1c4b-4281-b951-d872f2087c98