Resubmit of refactoring ImageWriterPrivate into the utility process.
I had a few errors in the CL I submitted last week.  I need to resubmit, but to
do that in a new CL I need new approvals (I'm not a committer).  The changes
should be minor.  The first PS is the original CL, the second PS is the changes.

Summary of changes:

* Didn't properly clean up some objects when testing.  In the long run I'd like
to see if I could limit the refcounting some more, but in the mean time I have
to be diligent about cleanup.  Cleanup functions are now automatically added
when the operation is set up for testing.
* When writing we write the whole buffer and not just the part read (due to
Windows alignment requirements). This extra was uninitialized but is now zeroed.
* When validating, if less data was read from the disk than was in the image,
then the image was compared against uninitialized memory.  This was fixed by
failing immediately if fewer bytes are read.

See original CL: https://codereview.chromium.org/61643015/

BUG=345659

TBR=jorgelo@chromium.org
TBR=mpcomplete@chromium.org
TBR=thestig@chromium.org
TBR=jln@chromium.org

Review URL: https://codereview.chromium.org/176103002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@253773 0039d316-1c4b-4281-b951-d872f2087c98