Remove the strange scaling by 2 of the forward FFT.
The FFT using FFMPEG on Linux and Windows has a strange scaling by a
factor of two. This is a carry over from the strange scaling on Mac's
veclib. This is removed here so that the FFT produces the correct
scaling. This effect shows up in the realtimeanalyser-fft-scaling
test, which has been updated with new thresholds. These thresholds
more closely match what would be expected from computing the windowed
FFT using matlab.

Android does not do this scaling.

BUG=342530

Review URL: https://codereview.chromium.org/132313018

git-svn-id: svn://svn.chromium.org/blink/trunk@168060 bbb929c8-8fbe-4397-9dbb-9b2b20218538