Speculative fix for crash in run-webkit-tests on 10.6.
It looks like there might be a case where we get a crash
dump that contains utf8-encoded data, but attempt to write
it out with an ascii encoding. This fix attempts to force
the write back out to utf8 to see if it solves the problem.

TBR=pkasting@chromium.org
BUG=332840

Review URL: https://codereview.chromium.org/180093007

git-svn-id: svn://svn.chromium.org/blink/trunk@168077 bbb929c8-8fbe-4397-9dbb-9b2b20218538