Fire accessible selection change event on the proper element.
Previously sometimes it was firing on the body element
even when it was an element like an <input> that had focus.
This fixes it. It should only fire on the body element
when it's a static text selection that's changing.

Trivial 1-line change but a comprehensive set of tests.

BUG=253462

Review URL: https://codereview.chromium.org/177973006

git-svn-id: svn://svn.chromium.org/blink/trunk@168080 bbb929c8-8fbe-4397-9dbb-9b2b20218538