Cast: Create separate CastEnvironments for sender/receiver in End2EndTest.
This is done to avoid a loopback of logging events described in bug 347305.
Now we can have slightly stronger assertions in End2EndTest.

Because CastEnvironment takes ownership of the passed in clock, we now have
to create an extra clock for the receiver CastEnvironment. We also have to
keep the two clocks in sync. This isn't too bad since the clock is only
advanced in one place, and only End2EndTest has this interesting
sender+receiver setup.


BUG=347305

Review URL: https://codereview.chromium.org/180093004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@254094 0039d316-1c4b-4281-b951-d872f2087c98