DevTools: support configurable category filter for trace events callback
DevTools callback is currently invoked for all callbacks which is a waste
of time as we only need a few. This will become more critical once we start
processing ConvertableToTraceFormat arguments.
We'd also like to select categories based on the different options enabled,
so let's move the filter to the DevTools back-end.

BUG=

Review URL: https://codereview.chromium.org/184283003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@254143 0039d316-1c4b-4281-b951-d872f2087c98