Hide error message before switching to other overlays.
When having a visible error bubble and switching to
another overlay, the error bubble needs to be hidden,
because otherwise it will shortly be visible when
reopening the overlay. Exception: the learn more overlay,
which is displayed on top of the createProfile overlay,
so the createProfile overlay keeps its state.

BUG=347937
TEST=See bug.

Review URL: https://codereview.chromium.org/184463002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@254149 0039d316-1c4b-4281-b951-d872f2087c98