Don't setStyle the RenderView every style recalc
Prior to this CL, we were setting the style of the RenderView on every style
recalc because we were comparing the RenderView's overflow style with the wrong
value. This CL corrects the bug.

This CL doesn't have much effect yet, but it will become important in a future
CL when we use style modification on RenderLayers to scope compositing updates.

Review URL: https://codereview.chromium.org/185263002

git-svn-id: svn://svn.chromium.org/blink/trunk@168254 bbb929c8-8fbe-4397-9dbb-9b2b20218538