Make text visible when font loading takes longer than 3 seconds
When a font resource takes a long time to download, currently Blink
keeps the text invisible until the font is ready. After this patch,
Blink behaves similar to FireFox; keeps the text invisible until 3
seconds, then fallback to the next available font in the fallback
list. When the webfont finally loads, change the text to use it.

- Add a Timer to FontResource that fires after 3 seconds since load start.
- fontLoadWaitLimitExceeded() callback invalidates the fallback font and
  triggers style recalc.
- Fallback fonts have m_skipDrawing boolean in CustomFontData, that is
  false if the font load exceeds the wait limit.
- FontFallbackList::shouldSkipDrawing() returns true if it contains at
  least one FontData with m_skipDrawing == true.

BUG=235303
TEST=http/tests/webfont/slow-loading.html

Review URL: https://codereview.chromium.org/171823002

git-svn-id: svn://svn.chromium.org/blink/trunk@168258 bbb929c8-8fbe-4397-9dbb-9b2b20218538