re-land: Use browser compositor rendering stats in smoothness
PERF SHERRIFS: This patch can change the metrics of smoothness up or down.
The new values should be more accurate than what we had before. If there are
large changes on a platform, we should still sanity check that everything works
as expected.

Switch smoothness benchmark to use rendering stats from the top level
compositor; i.e. if data from a browser compositor is available, use that.
Otherwise use the data from the renderer compositor.

The re-land fixes the problem where a browser compositor was present, but
didn't record rendering stats. That was the case with the reference
builds (that didn't have the patch that enabled recording) and on Android.
The new version checks if the browser compositor rendering stats events
actually have frames in them. If not, the render compositor stats will be used.
There is also a fix on the way to enable the recording on Android
(https://codereview.chromium.org/168193004/).

R=tonyg@chromium.org,nduca@chromium.org
BUG=340753

Review URL: https://codereview.chromium.org/174663008

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@254811 0039d316-1c4b-4281-b951-d872f2087c98