Bluetooth: show notification when adapter is discoverable
Certain white-listed apps will be allowed to make the adapter
discoverable to other nearby devices, we may also permit the user to
make the adapter discoverable as well.

For users security, inform them if the adapter is in this state when
they login, or when the adapter moves into this state.

BUG=239801
TEST=using bt_console, enter 'discoverable on'

Review URL: https://codereview.chromium.org/184783004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@254881 0039d316-1c4b-4281-b951-d872f2087c98