Password manager internals page: wire page updates to the client
This is step 2, following https://codereview.chromium.org/184413002. It has 1 main part and 2 tests:

1) Main part
The internals page now notifies the PasswordManagerClients when it gets created and destroyed.
The clients expose a SendLogToInternalsPage method for adding logs to the internals page. This method will be used by password managers to log progress in the next CL. Currently it is only used in tests.

2) Internals page browser test
It navigates to chrome://password-manager-internals. Then it grabs the WebUI controller of that page, and adds some logs. It checks through JavaScript that the logs are there. My main concerns are whether I'm waiting correctly after logging and before checking to avoid flakiness, but I see neither an issue, nor problems during running.

3) ChromePasswordManagerClient unit test
It checks that SendLogToInternalsPage works when attached to a controller, and is a no-op without a controller.
To do this, I had to mock out the controller, and introduce a factory to inject it, so the test is bigger than expected.

BUG=347927
TBR=gcasto@chromium.org

Review URL: https://codereview.chromium.org/185423007

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@254945 0039d316-1c4b-4281-b951-d872f2087c98