Remove dependency on element in applying color properties
This patch removes the state.element().isLink() calls in applying color
properties for visited links. This has a small behavioural change where
the -webkit-link color keyword will now be the visited link color for a
visited link descendent (bug 82498). I don't think we can currently test
visited link colors in a layout test.

This also fixes a crash when trying to set color in a pseudo-element
animation on a link element (or descendent), since we go through the
StyleBuilder with a null element. Note that the other places in the
StyleBuilder where we access the element already check for null.

BUG=82498,343048

Review URL: https://codereview.chromium.org/179363005

git-svn-id: svn://svn.chromium.org/blink/trunk@168435 bbb929c8-8fbe-4397-9dbb-9b2b20218538