GraphicsLayer::removeAllChildren shouldn't be n^2
This CL changes GraphicsLayer::removeAllChildren to be linear in the number of
children instead of n^2. Previously, we were removing the zero-th element of
the vector in a loop, which causes a shift of all the elements of the vector.
Now we walk the vector from the end, which means remove is constant cost.

This code is called during a compositing update if anything changes about the
GraphicsLayer's children. We didn't actually see this code on our profile
because we're not actually changing the GraphicsLayer tree, but it seemed worth
fixing.

R=ojan@chromium.org

Review URL: https://codereview.chromium.org/187603011

git-svn-id: svn://svn.chromium.org/blink/trunk@168598 bbb929c8-8fbe-4397-9dbb-9b2b20218538