Update AshmemRegion::highest_allocated_chunk_ when merging free chunks.
This is a follow up of r254755 that updated the pointer to the chunk with the
highest address in the region when the highest chunk in the region was being
split (during an allocation reusing a free chunk).

While this was enough to fix the bug specified below, it was only partly
addressing the problem since this pointer also needs to be updated when free
chunks are merged. This CL fixes this issue and adds an extra DCHECK() exposing
the problem with the existing unit tests.

BUG=347919

Review URL: https://codereview.chromium.org/183763037

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@255691 0039d316-1c4b-4281-b951-d872f2087c98