Cleanup: have common HttpResponseHeaders routine to update with range
Currently we do this header-fixup in two places: PartialData and AppCache,
and I'm planning to do the very same header-fixup in yet another module,
ServiceWorker.  I want to have a common utility in net/http/http_util
to do this so that all 3 modules can share it.

BUG=349319
TEST=HttpResponseHeadersTest.UpdateWithNewRange

Review URL: https://codereview.chromium.org/187583002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@255905 0039d316-1c4b-4281-b951-d872f2087c98