Remove HistoryItem::targetFrameID(), make it a HistoryController impl detail
targetFrameID() is generated from a counter that is guaranteed to be unique for
the life of the current renderer process, but is not guaranteed to be unique
across restart. Still use frame ids internally in HistoryController, but don't
serialize them or try to reuse them when navigating back/forward.

No test because it's very difficult to get a renderer process reliably into
this state, especially after removing the accessors that were causing the problem


BUG=341425
TEST=manually on http://my.opera.com/community

Review URL: https://codereview.chromium.org/171713012

git-svn-id: svn://svn.chromium.org/blink/trunk@168850 bbb929c8-8fbe-4397-9dbb-9b2b20218538