Oilpan: Use weak pointers in StyleSheetContents caches.
This fixes use-after-free that could occur when StyleSheetContents
dies when an allocation happens during insertion of a new
StyleSheetContents. Using the built-in weak processing this will
not happen because the iterator used during insertion will keep
the StyleSheetContents alive.

This is relanding the change and depends on thread-local
weak processing to not crash when performing weak processing
which can deallocate an AtomicStringImpl.

R=erik.corry@gmail.com, oilpan-reviews@chromium.org, vegorov@chromium.org, haraken@chromium.org

Review URL: https://codereview.chromium.org/183763048

git-svn-id: svn://svn.chromium.org/blink/trunk@168860 bbb929c8-8fbe-4397-9dbb-9b2b20218538