cc: Fix damage when the HUD is enabled.
When the HUD is on, and partial swaps are enabled, the damage rects
emitted by the renderer do not include the HUD, so the HUD is not
updated correctly on screen.

Previously we disabled partial swap in the renderer compositor to
force the hud to display, but in ubercomp, partial swap is never
used in the renderer, and the damage rect is given to the browser
instead. So we need to update the renderer's damage rect in this
case.

R=enne@chromium.org, piman@chromium.org, caseq, enne
BUG=350495

Review URL: https://codereview.chromium.org/191433002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@256052 0039d316-1c4b-4281-b951-d872f2087c98