Ignore window.{testRunner/layoutTestController} at cached properties test
Now testRunner (and layoutTestController, which is an alias of testRunner)
is moved from CppBoundClass to gin::Wrappable (crbug/331301),
they are shown in the for-each loop of window properties.
testRunner and layoutTestController are only for testing,
so we don't have to check the cached states of them.

BUG=331301, 351195

Review URL: https://codereview.chromium.org/193713004

git-svn-id: svn://svn.chromium.org/blink/trunk@168893 bbb929c8-8fbe-4397-9dbb-9b2b20218538