content_shell: Add event filtering to PluginObject
For testing input event filtering in plugins, this change adds a
"alwaysFilterEvents" argument to PluginObject. That argument causes the plugin
to always return a non-zero value from HandleEvent, which requests the event be
"filtered" to stop propagating it to other handlers on the page.

BUG=324351

Review URL: https://codereview.chromium.org/191383005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@256163 0039d316-1c4b-4281-b951-d872f2087c98