Bookmark manager: Correctly focus on the URL field after invalid URL alert.
We focus on a URL input after showing invalid URL alert. However it failed if a
user removed focus from the URL input by TAB key. This CL fixes it.

Behavior before this CL:
1. urlInput looses focus by Tab key
 The page lost focus as well because the urlInput was the last focusable
 element.
2. The alert is shown in respond to 'blur' event for urlInput.
3. urlInput.focus() is called.
 It sets document.activeElement to urlInput, but it doesn't accept user input
 because the page has no focus.

To resolve this issue, we have to avoid to remove the page focus. So we do
preventDefault for Tab key if the urlInput has an invalid URL.

BUG=343502

Review URL: https://codereview.chromium.org/187703005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@256194 0039d316-1c4b-4281-b951-d872f2087c98