[NaCl SDK] Fix create_nmf test failure on mac builder.
The /var directory on a mac is actually a symlink to /private/var. Normally
this isn't relevant.

One create_nmf_test changes to this directory, and calls os.path.abspath(''),
which returns '/private/var/...', and later uses os.path.relpath to find the
relative path to a file in that directory tree.

This fails to do the right thing, because relpath only uses the path name (not
the filesystem) to determine relative paths.

The fix here is to use os.path.realpath to create a canonical path whenever we
want a relative path. I've also removed other calls to os.path.abspath that
really should be os.path.realpath.

BUG=none
R=bradnelson@google.com, sbc@chromium.org, bradnelson@chromium.org
TEST=create_nmf_test.py

Review URL: https://codereview.chromium.org/195493003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@256307 0039d316-1c4b-4281-b951-d872f2087c98