Rename V8BindingTestScope to V8ExecutionScope and use it in V8GCController and ScriptRegexp
Currently V8GCController and ScriptRegexp execute V8 things without having any associated world. This is not causing any issue at the moment because they don't execute V8 things that require to look up the world. However this is fragile, so this CL creates a world for them. After this CL, it is guaranteed that all V8 binding code is executed with some context associated with some world.

The essence of this CL is:

- Rename V8BindingTestScope to V8ExecutionScope
- Use V8ExecutionScope in V8GCController and ScriptRegexp

BUG=341032

Review URL: https://codereview.chromium.org/196583005

git-svn-id: svn://svn.chromium.org/blink/trunk@169008 bbb929c8-8fbe-4397-9dbb-9b2b20218538