[CSS Grid Layout] Initial span position not correctly resolved sometimes
We're using std::lower_bound to find the first line before the opposite
resolved position. The returned iterator might point to a resolved position
which is actually greater than the opposite position (since lower_bound
returns the first non less than). In those cases we need to move the
iterator to the previous line so that the span can be correctly computed.

Review URL: https://codereview.chromium.org/194613003

git-svn-id: svn://svn.chromium.org/blink/trunk@169031 bbb929c8-8fbe-4397-9dbb-9b2b20218538