Pause the timeline on load in the SVGAnimationTestCase.js framework
Most tests using this framework drive the "clock" / currentTime "manually"
by setCurrentTime(t), and add the instance time of a click event to it.
This means that there's a risk of rounding errors creeping in and causing
the actual elapsed time to be off by a fraction - large enough to cause
failures - at least for discrete animations.
(Example: svglengthlist-animation-3.html)
By pausing the timeline at 'load', the instance time for the click event
will be '0', which should eliminate the risk of rounding issues.
For the cases where the test actually need time to pass the timeline can
be unpaused explicitly. (Currently only one testcase seemed to require
that treatment.)

BUG=350827

Review URL: https://codereview.chromium.org/194943002

git-svn-id: svn://svn.chromium.org/blink/trunk@169033 bbb929c8-8fbe-4397-9dbb-9b2b20218538