Adds -Wno-deprecated-register to the build settings for iOS.
Xcode 5.1 ships with a clang that deprecates the register keyword by default.
This warning flag is needed in order to compile for iOS with Xcode 5.1.

However, the old version of Xcode (5.0) does not understand this flag.  In order
to temporarily support compiling on both versions of Xcode, this CL also adds
-Wno-unknown-warning-option to suppress errors about unknown compiler flags.
This flag will be removed as soon as all of the bots have been updated with
Xcode 5.1, within a week.

BUG=255186,337611
TEST=None

Review URL: https://codereview.chromium.org/194803006

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@256592 0039d316-1c4b-4281-b951-d872f2087c98