CoreAnimation: Be robust to frames coming when the NSView has no window
The bug here was that, for some reason (which no longer reproduces), a
frame came in while the RWHVMac was not hidden, but had no window, and
was not being captured. The result was that an IOSurface wasn't created,
the frame was acknowledged, and when the view did land in the hierarchy,
there was nothing available to draw.

Fix this to always populate the IOSurface, its context, and its layer, so that
if we ever get into this situation, we can draw the frame that arrived. This
was tested when the bug could reproduce, but re-syncing made the repro
case go away.

While doing this, re-organize the code that handles the situation where a
tab that is not in a window's view but is not hidden receives a frame.

Unify the code that is responsible for setting up the IOSurface and sending
the frame to the frame subscriber. There's no downside to doing this, and
it makes it so that there are fewer states that need to be thought about.

When there is no window, move the early-out to be just before the contents
are drawn.

BUG=350871

Review URL: https://codereview.chromium.org/194713005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@256606 0039d316-1c4b-4281-b951-d872f2087c98