Linux Aura: Move tab strip background image up 15px in condensed mode.
Affects windows that are maximized or using system title bars on Linux.

For unthemed browser windows, this effectively removes the gradient,
which looks much nicer when used with a native title bar. For themed
browser windows, this aligns the bottom of the tab strip theme image
with the bottom of the tab strip (rather than aligning the tops).

This matches the behaviour of Linux GTK in all cases.

BUG=339418

Review URL: https://codereview.chromium.org/163123002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@256770 0039d316-1c4b-4281-b951-d872f2087c98