[MemUsage][BinarySize] More perfect hashes for CSS strings.
gperf generates some "perfect" hashes for CSS properties and CSS
values. Those are not very compact but can be more compact if gperf
is asked to try harder. This will cost 1 s build time and reduce binary
size by about 12 KB.

The flags changed are:

'-m 50' - Try 50 times more initial values to the search algorithm. This
will make the call 50 times slower but for the current input size that is
no problem.

Removal of ‘-s 2’ - -s 2 told gperf to generate a twice as large data table
to make it easier (faster) to find a solution.

Also '-D' moved around but was there both before and after so no addition
or removal.

In the longer run we should move away from gperf to a better algorithm (they
exist at the research level but not as easily installed programs)

BUG=351774

Review URL: https://codereview.chromium.org/196413006

git-svn-id: svn://svn.chromium.org/blink/trunk@169111 bbb929c8-8fbe-4397-9dbb-9b2b20218538