[CSS Grid Layout] getComputedStyle() must return specified values
Positioning properties, grid-{column|row}-{start|end}, must return the
specified value for getComputedStyle(). We were incorrectly changing some of
those properties in the StyleAdjuster in order to handle some errors in grid
placement. Now we do it in the RenderGrid so that the style is left
untouched.

The test expectations were accordingly updated and a new test to check the
behavior of a grid positioning property with 2 spans was added, since we can
no longer check it with the computed style.

This CL adds a new test to check that we return the right values even for
named grid lines and named grid areas which might also require changes to
the specified values.

Review URL: https://codereview.chromium.org/195123003

git-svn-id: svn://svn.chromium.org/blink/trunk@169123 bbb929c8-8fbe-4397-9dbb-9b2b20218538