Reland "[chromeos/about:power] Collect cpuidle and cpufreq stats"
Original: https://chromiumcodereview.appspot.com/149973002
Revert: https://codereview.chromium.org/196613002/

The difference between this and the original CL is that the sysfs is
read on the blocking pool via a non-member function. This avoids passing
the weak ptr to blocking pool from the UI thread (as was the case in the
original CL).

This reverts commit 92e40737ec9a1cf806761b4c67caf688d3a1ef5e.

BUG=335816
TBR=arv

Review URL: https://codereview.chromium.org/197813006

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@256828 0039d316-1c4b-4281-b951-d872f2087c98