Fix Autofill popup hover behavior on GTK.
Hovering over an entry in the popup should result in previewing the suggestion,
not in filling it into the form.  This was accidentally regressed in a catch-up
CL to fix up the GTK build after a period when only the Aura build was kept
up-to-date.

BUG=351001
TEST=(see bug)
R=estade@chromium.org
NOTRY=true

Review URL: https://codereview.chromium.org/199533002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@257049 0039d316-1c4b-4281-b951-d872f2087c98