Disable ServiceWorkerDispatcherHostTest.Enabled
Since we've changed the code which checks command-line flag to cache
the result in a static variable (to make it as fast as possible since
we could hit the code in every request interception), changing the
switch dynamically in a unit test no longer works in a deterministic
way.

This CL simply disables such test because I think now we have
other tests that can cover the same code in browsertests.

BUG=352581
TBR=michaeln
NOTRY=true

Review URL: https://codereview.chromium.org/197313009

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@257077 0039d316-1c4b-4281-b951-d872f2087c98