Add "previewing on hover" support for single-field autocomplete input
In terms of previewing on hover, treat single-field autocomplete
the same as Autofill. Functions are introduced to handle when an
AutoComplete selection is hovered over and clear once preview is done.

This fix only handles autocomplete field. Preview support for password
autofill support mentioned in the bug descriptions will be handled in a
separate commit.

BUG=63421
R=estade@chromium.org, isherman@chromium.org

Review URL: https://codereview.chromium.org/148413002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@257274 0039d316-1c4b-4281-b951-d872f2087c98