[rac] Improve libaddressinput suggestion performance with tries.
The patch changes suggestion algorithm from a for loop that compared
prefixes to a trie. The performance impact on a debug build:

Old (for loop that compared prefixes):
 - Build rulesets = 200 ms
 - Each character typed = 500 ms

New (trie):
 - Build rulesets = 200 ms
 - First character typed = 220 ms (200 ms due to building trie)
 - Each subsequent character typed = 20 ms

The performance for each character improves significantly. Building
rulesets and the trie still have room for optimization.

BUG=347383

Review URL: https://codereview.chromium.org/183793003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@257277 0039d316-1c4b-4281-b951-d872f2087c98