Redoing Issue 36073011: Allowing file:/// in Instant Extended's Most Visited links.
The previous change was rolled back because it prevented the NTP ping, which was needed for UMA. 3 main parts of this CL are:

(1) Redo Issue 36073011, adjusted to code changes. This fixes the file:/// issue, and changes NTP outgoing page transition to AUTO_BOOKMARK (instead of LINK).
(2) Adding logNavigation(pos, provider), which is called on click.
(3) Deprecate ping and chrome://log.html pages.

Caveat: keyboard navigation is not properly handled, since it involves the NTP rather than the iframe.  This will be fixed in a follow-up CL.

For (1):
- Unlike Issue 36073011, we're not enabling LocalNTPTest.LocalNTPJavascriptTest, because recent cleanups causes more failure modes of this test.
- Using IDR_MOST_VISITED_WINDOW_DISPOSITION_UTIL_JS instead of IDR_CLICK_UTIL_JS.
- In local_ntp.js: init(): skipping change to the "recentTabsText" block, since it got deleted.

BUG=306137

Review URL: https://codereview.chromium.org/178253008

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@257389 0039d316-1c4b-4281-b951-d872f2087c98