linux-aura: Keep track of whether the drag-client is alive after the drag.
It is possible for the DesktopWindowTreeHost (and its drag-client) to be
destroyed during the drag (or after the drag ends in OnXdndFinished(), but
the control returns in StartDragAndDrop()). So keep a weak-ptr around for
the drag session, and check whether it is still valid after the drag
before resetting internal state.

BUG=351504
R=erg@chromium.org

Review URL: https://codereview.chromium.org/202083004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@257493 0039d316-1c4b-4281-b951-d872f2087c98