Ignore WM_MOUSEACTIVATE messages which occur in the context of touch events which result in bubbles like bookmarks, avatar button, being displayed.
When we touch the button which brings up the bubble, a WM_MOUSEACTIVATE message comes in much later for the main
window. This results in the window being activated which in turn hides the bubble which occurs due to an activation
change.

Fix is to ignore WM_MOUSEACTIVATE messages in the context of touch downs if we have the flag which indicates that
inactive rendering is to be faked. This does not cause activation issues when we touch/click elsewhere due to
calls to SetFocus which occur in the context of these events from the focus controller code.

The touch down context flag is reset to false in a delayed task.

The other change is to ignore WM_MOUSEACTIVATE messages in the LegacyRenderWidgetHostHWND class if the
kIgnoreTouchMouseActivateForWindow property is set on it. This property is set by the MenuController code to
prevent activation loss when the menu options are selected via touch. This is mimicing similar code in
HWNDMessageHandler.

BUG=329854,326128
R=sky@chromium.org, sky

Review URL: https://codereview.chromium.org/199593004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@257522 0039d316-1c4b-4281-b951-d872f2087c98