[chromeos/about:power] Do not crash if cpufreq stats are not available
CPU idle and/or CPU freq stats might not be available on VMs even
if they are running a kernel >= 2.6. This definitely is the case for
the marked bug. To handle such cases, instead of crashing, this CL
modifies to just log an error message and ignore collecting freq
samples. Collection of idle samples is already setup that way.

BUG=353043

Review URL: https://codereview.chromium.org/202073005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@257557 0039d316-1c4b-4281-b951-d872f2087c98