cc: Attempt to de-flake test by preventing commit to start before draw.
The draw code checks that only 2 textures are used. However if a
BeginMainFrame occurs before draw and we upload textures for the next
frame, it would mess up the counting by the time we draw. Instead of
posting a commit for the next frame after activate, do it after we
have checked the number of textures used and drawn.

R=brianderson@chromium.org, enne@chromium.org, enne
BUG=352730

Review URL: https://codereview.chromium.org/198173004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@257712 0039d316-1c4b-4281-b951-d872f2087c98