Fix PNaCl debug commandline flags, and use the pexe with debug info in test.
Now that the PNaCl translator which supports reading
llvm-formatted bitcode files is DEPS'ed in, we can change
the debug test to actually use a pexe with debug info. In
the process, fix the way the commandline flags were set
up (avoid <<'ing a string that starts with a null byte).

Also try to turn on the test for 64-bit windows instead,
since there should only be a problem with 32-bit windows.

BUG= http://code.google.com/p/nativeclient/issues/detail?id=3765
R=bradnelson@chromium.org, bradnelson@google.com

Review URL: https://codereview.chromium.org/203103002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@257729 0039d316-1c4b-4281-b951-d872f2087c98