evdev: dri: Add CursorFactoryOzone implementation for DRI platform
This connects cursor loading & movement to DriSurfaceFactory. A loaded
cursor for DRI is an SkBitmap together with a gfx::Point for the hotspot.
Introduce BitmapCursorOzone to represent such a cursor, as well as
BitmapCursorFactoryOzone for managing them.

Drop SetCursorImage & MoveCursorTo from the SurfaceFactoryOzone interface;
they do not belong there as they are not used from common code.

BUG=252315
TBR=ben

Review URL: https://codereview.chromium.org/193893002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@257772 0039d316-1c4b-4281-b951-d872f2087c98