Remove ErrorConsole handling from ExtensionHost
Code to handle the IPC for detailed console messages already exists in
ExtensionWebContentsObserver. The version in ExtensionHost mostly differed
in that it could provide an "associated" web contents as the context for
the message. This is less desirable than the first version -- for example,
for an infobar error message it's more useful to have the infobar's HTML page
as context rather than the tab's URL.

ExtensionWebContentsObserver needs minor changes to ensure it gets the correct
extension ID.

This change is a step toward moving ExtensionHost out of src/chrome and into
src/extensions for the app_shell project.

BUG=321341
TEST=browser_tests ErrorConsoleBrowserTest

Review URL: https://codereview.chromium.org/200763010

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@257781 0039d316-1c4b-4281-b951-d872f2087c98