Panel callout widgets should not accept events
Disallow panel callout widgets (the arrows between panels docked to
the shelf and the shelf itself) from accepting events. Attempted
taps on shelf buttons are often targeted to these callout widgets,
and since the callout widgets do not handle the events in any way,
they should just not be allowed to accept them in the first place.
This allows for taps to be targeted to either the shelf button
or the frame of the panel immediately above it.

BUG=351348
TEST=none

Review URL: https://codereview.chromium.org/201243004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@257801 0039d316-1c4b-4281-b951-d872f2087c98