Fix flakyness in ScreenOrientationProviderTests.
There was two causes. The first was that mObserver is initialized
asynchrounously so the tests were flaky because the initial value might
or might not have been the expected one. The second cause was related to
pre-API 17 were we can't reliably listen to 180 degrees orientation
changes.

BUG=353500,353322

Review URL: https://codereview.chromium.org/197213025

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@257908 0039d316-1c4b-4281-b951-d872f2087c98