Restore editor text upon AdapterInputConnection is recreated.
As Editable is created in BaseInputConnection which is the
parent of AdapterInputConnection and no one owns it,
once a new AdapterInputConnection is created (e.g. when user
switches keyboard), the old AdapterInputConnection object
gets lost along with the Editable object and losts internal
Editor state (especially the text content).

Still Blink holds the text content, but upon such events as
keyboard switch, Blink doesn't update the Editor and there will
be inconsistency between what is held in Blink and what Android
keyboard can see in the Editable.

Hold the previous Editable and restore the text content
upon a new AdapterInputConnection is created.

BUG=344683
TEST=ContentShellTest -f ContentCoreInputConnectionTest.*

Review URL: https://codereview.chromium.org/172273002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@257998 0039d316-1c4b-4281-b951-d872f2087c98