Crash fix: WebView does not attempt to alter view hierarchy during WebContents destruction.
The problem is that WebView was attempting to detach the WebContents native view during a call to WebContentsDestroyed().  When the native view (an aura::Window) was detached, the window was hidden, and this in turn caused WebContentsImpl to attempt to background a non-existant render process.

TBR=sky@chromium.org

Review URL: https://codereview.chromium.org/205263002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@258107 0039d316-1c4b-4281-b951-d872f2087c98