Remove redundant clusterfuzz test
The test was added at r150718.  https://src.chromium.org/viewvc/blink?revision=150718

Currently It doesn't cause a crash even if we revert the change. It means the test is useless to keep stability.
Additionally the test looks less sophisticated and it is hard to get what it does.

BUG=

Review URL: https://codereview.chromium.org/206073002

git-svn-id: svn://svn.chromium.org/blink/trunk@169628 bbb929c8-8fbe-4397-9dbb-9b2b20218538