Always keep browser restore state when closing a window on Mac.
This is largely a revert of cl#362016.

The code was not working correctly anyway.  If multiple profiles were open then it would save & restore tabs but if only one profile was open then it would not.  The original intent was to mimic the behavior of Safari on Mac but it makes Mac Chrome inconsistent with other editions of Chrome and (possibly) Win8 browsers.  It also causes problem with Lock on Mac which is expected to always restore tabs but fails to do so.

BUG=352172

Review URL: https://codereview.chromium.org/200373002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@258324 0039d316-1c4b-4281-b951-d872f2087c98