Do not clip inside OcclusionTracker: Contributing Surfaces Edition.
This is a follow-up for https://codereview.chromium.org/23708021 where
we stopped clipping layers in the occlusion tracker. In this CL we stop
clipping contributing RenderSurfaces as well. The contents inside the
render surface are already clipped, and this is applied to the content
rect.

While we might be able to clip further in target space, we have to
blow the result back up into the content space again, so this should
produce about the same results.

While we're at it, fix a bug in QuadCuller where it always used the
surface draw transform to cull RenderPassDrawQuads even if the quad
was for the replica, which could lead to us culling the replica of
a surface when the surface itself is occluded. This was covered in
occlusion tracker tests, but not in QuadCuller tests.

R=enne
BUG=344962,276725

Review URL: https://codereview.chromium.org/205443002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@258431 0039d316-1c4b-4281-b951-d872f2087c98