cc: Add layer raster tile iterator
This patch continues to build on top of tiling
raster tile iterators and introduces a layer
raster tile iterator that uses the tiling one.

Comes with a unittest and a perftest. The next
step after this should be a raster iterator
in the tile manager

R=enne@chromium.org

Review URL: https://codereview.chromium.org/195583002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@258437 0039d316-1c4b-4281-b951-d872f2087c98