[invalidations] Added internal debugging information for push and GCM
This patch builds up from SVN r257555.

This patch adds debugging information from the GCMNetworkChannel
and for the PushClientChannel. It changes a bit how the
RequestDetailedStatus work. Now the prefered way to do debugging is,
for each object, assemblying a DictionaryValue with all the debugging
information, running the callback with that dictionary and passing
the callback to other objects that might want to log something.
The pro of doing it this way is that is very easy to propagate this
debugging mechanism to any interesting object. All the dictionaries
are merged in the JS, making it easy to keep adding more information
if needed.

Among others, about:invalidations detailed status can now display:
-Underlying Network Channel in use
-Sent messages / Received messages count
-Chrome flags used to set the jingle servers
-GCM: Last response status, if it was an empty echo packet,
	timestamp of the last received message, registration result,
	hash of the registration_id.

BUG=263863

Review URL: https://codereview.chromium.org/202863003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@258493 0039d316-1c4b-4281-b951-d872f2087c98