Chrome metro mode win7 emulation.
It seems possible that we can run Ash (chrome metro) in Windows7 since
we are using the GPU-direct path in Aura. In fact, we had something
close to this that bitrotted because only the output (display) had
common code with what we shipped but no commonality otherwise in two
important regards

- Input
- Focus & Activation.

Basically what we had it was a frankenstein between Ash and Desktop Aura
that had the worst bugs from both and nobody cared to fix because a fix
there would be mostly orthogonal from the two modes that we ship.

This is an effort to address these issues. It uses COM / WinRT magic to
try to reuse as much code as possible from the metro_viewer side, while
trying to keep the browser side identical.

In other to have that reuse, we need to emulate the host (microsoft) side
of the metro environment. This patch is a start on that direction namely
an emulation of a WinRT object known as
"Windows.ApplicationModel.Core.CoreApplication"

Which exposes a bunch of interfaces but only two we try to emulate.


BUG=none

Review URL: https://codereview.chromium.org/199843004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@258530 0039d316-1c4b-4281-b951-d872f2087c98