Step 1 in unifying the task manager's WebContents-based ResourceProviders.
Motivation.

The primary motivation here is to centralize the tracking of WebContents, so
that the upcoming logic that shows cross-process iframes in the task manager
does not have to be copied five times. A secondary motivations is to eliminate
uses of some (but not all) deprecated notification service notifications,
switching to WebContentsObserver instead.

Design.

A new class, WebContentsResourceProvider, will provides resources to the task
manager on behalf of a chrome service that owns WebContentses. The chrome
services will be parameterized via another new interface,
WebContentsInformation, which provides a list of WebContentses to track and
also acts as a RendererResource* factory.

The WebContentsInformation is responsible for notifying the
WebContentsResourceProvider of the existence of new WebContents* instances
only (not deletion/updates). This is because for the rest of the tracking,
WebContentsResourceProvider creates a WebContentsObserver. The
WebContentsInformation impls will derive from, as an implementation detail, an
abstract base class called NotificationObservingWebContentsInformation. This
base class takes care of listening for NOTIFICATION_WEB_CONTENTS_CONNECTED.

In this CL, serving as an example, only the Panel resource provider is
replaced: PanelResourceProvider just becomes a much smaller PanelInformation,
and the TaskManager instantiates a WebContentsResourceProvider, passing
instance of PanelInformation as its |info_| parameter.

Direction.

In subsequent CLs this new mechanism will subsume the commonly implemented
parts of Extension, Panel, TabContents, and Background resource providers.
This promises to be a net deletion of code, on account of how much duplicate
functionality existed between these classes.

BUG=348836
TEST=browser_tests

Review URL: https://codereview.chromium.org/196203003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@258545 0039d316-1c4b-4281-b951-d872f2087c98