Disk cache: Some cleanup for the skeleton of the backend worker.
This adds the backend worker to net.gyp and deletes/disables code so that
the file compiles but does nothing else than construct/destroy an object.

This is the last skeleton needed to start building the infrastructure to create
and destroy the backend. The actual description of the class will be added
when some functionality is added: this class will be in charge of proxying
work to the cache thread.

BUG=241277

Review URL: https://codereview.chromium.org/205623003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@259107 0039d316-1c4b-4281-b951-d872f2087c98