Use QuadSink::Append for debug borders.
Debug borders do not get partially culled anyway, as we want to draw
them where the whole bounds of the layer are, not where the visible
parts of the layer are. So just change to Append, and avoid doing any
occlusion computation for debug borders.

R=enne@chromium.org
BUG=344962

Review URL: https://codereview.chromium.org/211393004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@259376 0039d316-1c4b-4281-b951-d872f2087c98