Introduce InfoBarService::Observer
This Observer class is intended to replace the current use of Notifications.
This is necessary to achieve the layered componentization of infobars, and also
results in a cleaner code as the observer pattern is generrally preferred to
notifications.

This CL also updates InfoBarContainer to use the new Observer class, as a first
concrete use of the new code.
Other listeners of the notifications will be converted in the future.

BUG=354379, 354380

Review URL: https://codereview.chromium.org/203963006

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@259556 0039d316-1c4b-4281-b951-d872f2087c98