Adds suppression for leak
The objects showing up as being leaked (CallbackHolder) are embedded
in the v8 heap. This tests seems to be shutting down in a way that
doesn't cause the finalizers for these objects to run and we
leak. Temporarily supressing until we sort out the issue.

BUG=356785
TEST=none
R=jochen@chromium.org
TBR=jochen@chromium.org

Review URL: https://codereview.chromium.org/213723003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@259732 0039d316-1c4b-4281-b951-d872f2087c98