Cast: UDP proxy library and binary
This udp prody lets you simulate packet delays and drops for UDP packets.
The code is separated into a library and a binary, the library is intended to
be usable directly from tests, while the binary is intended to be used for manual
testing. Library is modular and will probably be extended with more ways to do
evail/bad/naughty things to the packets in the future.

Review URL: https://codereview.chromium.org/210553002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@259750 0039d316-1c4b-4281-b951-d872f2087c98