Add CSSUnknownRule.idl to |core_idl_files|
There's currently a comment in core.gypi that reads:
"""
We should probably add CSSUnknownRule.idl to this list,
but it currently causes a compile error.
"""
Now that we've switched to Python (and proper IDL processing),
this no longer causes a compile error, so I think we can switch it on.

Changes:
This adds CSSUnknownRule to the global constructors.

This dates to April 2013:
Introduce bindings.gyp
https://codereview.chromium.org/14110002

Review URL: https://codereview.chromium.org/212653006

git-svn-id: svn://svn.chromium.org/blink/trunk@170144 bbb929c8-8fbe-4397-9dbb-9b2b20218538