Make it possible to read CLD data from a file.
Note that this change DOES NOT "opt-in" any platform at this time, it merely
makes it possible to read the CLD data in from a file if cld_version=2 and
cld_dynamic=1. Subsequent changes will be required for platforms to "opt-in"
to using this approach; e.g., each platform will have to generate their data
file and package it appropriately before turning this feature on in compile
flags.

IMPORTANT: Can't be enabled on on Android until the SIGBUS in CLD is fixed:
https://code.google.com/p/cld2/issues/detail?id=11

BUG=326023

Review URL: https://codereview.chromium.org/187393005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@259900 0039d316-1c4b-4281-b951-d872f2087c98