RefCounted::derefBase() should do the decrement from 1 to 0
Currently derefBase() does not do the decrement from 1 to 0 because an object that has 0 refcount is going to die immediately (there is no need to do the decrement).

However, in the oilpan world, it's possible that an object that has 0 refcount is resurrected, because the object can still be retained through oilpan handles. Thus we need to manage the refcount correctly even when the refcount is going to be 0.

I'm making this change separately from oilpan changes, since this change might regress performance of something. derefBase() is performance-sensitive.

BUG=340522

Review URL: https://codereview.chromium.org/214453005

git-svn-id: svn://svn.chromium.org/blink/trunk@170256 bbb929c8-8fbe-4397-9dbb-9b2b20218538