Add support for midstream audio configuration changes.
This introduces a new class, AudioBufferConverter, which takes AudioBuffers of arbitrary formats and converts them to AudioBuffers of a common format.

To support config changes seamlessly we use the AudioBufferConverter to convert incoming AudioBuffers to the hardware output format as we get them from the AudioBufferStream. This way the rest of the audio pipeline doesn't need to handle config changes.

We only enable this conversion step for DemuxerStreams that support config changes (namely ChunkDemuxerStream).

BUG=347270
TEST=AudioBufferConverterTest
NOTRY=true
R=dalecurtis@chromium.org

Review URL: https://codereview.chromium.org/177333003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@260071 0039d316-1c4b-4281-b951-d872f2087c98