[TIZEN] Tag the audio stream as "browser" in pulseaudio.
Tag the audio stream as "browser" in pulseaudio so that all the streams
coming from the web will be considered by Murphy as coming from the same
source and preempt them for example if there's a phone call coming.

At the moment the full Murphy integration is not in place due to some
integration issues with the main loop. But, once we have all the fixes
in place, we will directly connect to the Murphy's resource engine and
start receiving notifications when any another stream preempting the
streams tagged as 'browser'. This will make the Crosswalk fully
resource-aware.