[OriginChip] Show the URL on Esc key pressed
This change makes it so hitting the Esc while the omnibox has focus shows the URL unless search term replacement is being performed, in which case we revert to the original search terms.

This holds for both the hide-on-mouse-release and hide-on-input options.  For the latter, this means that a click in the omnibox will also hide the origin chip.  This supports the case where the user is trying to get at the URL and forgets to click on the origin chip or misses it due to existing muscle memory.  In this case, an Esc will show the URL as users are accustomed to.

In no case will hitting Esc cause the origin chip to reappear.

BUG=349497

Review URL: https://codereview.chromium.org/214613003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@260173 0039d316-1c4b-4281-b951-d872f2087c98