Resetting metrics ids on clump detection.
When a cloned install is detected the metrics client id and low entropy source are reset on the next start of Chrome. This is done by setting a pref in local state. The next time either the client id or low entropy source are loaded from the local state, they are first reset.

BUG=343273
NOTRY=true

Review URL: https://codereview.chromium.org/213363004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@260348 0039d316-1c4b-4281-b951-d872f2087c98