V4L2VDA: Post StartDevicePoll instead of calling it directly from Initialize
StartDevicePoll() will fail the DCHECK for being called on decoder_thread_
when called from Initialize(), which is run on the ChildThread.

It's a good idea to keep that DCHECK, we don't really want to start polling
if decoder thread is not running. Also, the if that surrounds the DCHECK
is currently always true.

BUG=None
TEST=video playback on Debug build

Review URL: https://codereview.chromium.org/208313013

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@260509 0039d316-1c4b-4281-b951-d872f2087c98