Step 5 in unifying the task manager's WebContents-based ResourceProviders.
In this change, PrintingInformation, which plugs into
WebContentsResourceProvider, is introduced and the code that tracks
background printing pages is removed from TabContentsResourceProvider.

( WebContentsResourceProvider is introduced in an earlier patch:
https://codereview.chromium.org/196203003/ )

The new PrintingInformation tracks not only WebContents that are being
background printed, but also WebContents that are foreground print
preview dialog. Prior to this change these processes did not appear
in the task manager. Fixing this is what bug 345903 is about. A new
browsertest is added for this case.

BUG=348836, 345903
TEST=browser_tests including a new one

Review URL: https://codereview.chromium.org/201183005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@260801 0039d316-1c4b-4281-b951-d872f2087c98