CSS Transforms: Implement transform-origin
Implements transform-origin as defined in CSS Transforms[1]. Behavior is
similar to -webkit-transform-origin, however transform-origin is not a 
shorthand. New parsing and serialization (mostly matching Gecko) has been 
introduced.

The animation system is modified to avoid running animations and 
transitions on both the prefixed and unprefixed variants at the same 
time. Internally, animations are mapped to the unprefixed variant. 
As animations operate directly on the RenderStyle storage the result 
of specifying an animation on either -webkit-transform-origin
or transform-origin is the same.

However, support for independent animation/transition of 
-webkit-transform-origin-* is removed. The X, Y and Z coordinates now
always transition together.

Note, this property and changes in behavior are guarded by the
CSSTransformsUnprefixed runtime feature (enabled only for testing).

[1] http://www.w3.org/TR/css3-transforms/

BUG=154772

Review URL: https://codereview.chromium.org/212483003

git-svn-id: svn://svn.chromium.org/blink/trunk@170537 bbb929c8-8fbe-4397-9dbb-9b2b20218538