Update pad-gesture-fling test to accommodate short-circuited fling
The pad-gesture-fling layout tests assumes the fling curve will generate a
minimum number of events.  This may not always hold, depending on the
fling curve parameters, the fling curve update frequency and whether the main
thread is slow.  Instead, impose a maximum limit on the number of fling updates
it takes to reach the minimum offset target, and use preventDefault() on the
mousewheel event to prevent the fling from terminating early.

Review URL: https://codereview.chromium.org/219243010

git-svn-id: svn://svn.chromium.org/blink/trunk@170597 bbb929c8-8fbe-4397-9dbb-9b2b20218538