MSE: Allow requesting coded frame processor: legacy or new
Updates WebMediaSource to have an addSourceBuffer() version that
allows caller to request which of LegacyFrameProcessor or a (not yet
landed in Chromium) new frame processor to use to process frames parsed
during appends to the source buffer.

The choice of which frame processor to request is based on
--enable-experimental-web-platform-features, as exposed through
RuntimeEnabledFeatures::mediaSourceExperimentalEnabled(): if enabled,
the new frame processor is requested. Otherwise, LegacyFrameProcessor is
requested.

Until the updated Chromium side WebMediaSourceImpl lands and (later)
the new Chromium frame processor lands, LegacyFrameProcessor will be
used regardless of the requested processor.

R=acolwell@chromium.org,abarth@chromium.org
BUG=249422
TEST=All MediaSource layout tests pass both before and after Chromium
side WMSI change is applied locally, and MediaSourceBase logging shows
frameProcessorChoice reflects usage (or lack) of cmdline switch
--enable-experimental-web-platform-features

Review URL: https://codereview.chromium.org/220593010

git-svn-id: svn://svn.chromium.org/blink/trunk@170678 bbb929c8-8fbe-4397-9dbb-9b2b20218538