Remove VideoTrack runtime feature since we always support TextTracks.
We haven't built Chrome without TextTrack support for a long
time so I'm removing the VideoTrack feature flag and removing
all conditional code that depended on it. It is part of the
HTML5 spec and should be on by default for any compliant
implementation that supports media.

Review URL: https://codereview.chromium.org/205583009

git-svn-id: svn://svn.chromium.org/blink/trunk@170680 bbb929c8-8fbe-4397-9dbb-9b2b20218538