Oilpan: Make TreeShared GarbageCollectedFinalized.
This is the smallest possible patch to move the Node and SVGElementInstance
trees into the oilpan heap. Everything is still using the TreeShared
algorithms for deletion. Therefore, no RefPtrs can be replaced with
tracing at this point. I'll work on that next.

I initially did a patch that actually got rid of TreeShared in favor
of RefCountedGarbageCollected. However, that was too big a mouthful
and I lost track. I'm therefore going back to the approach that we
used on the branch. Move it a little at a time while keeping it
working.

R=haraken@chromium.org, kouhei@chromium.org, oilpan-reviews@chromium.org, wibling@chromium.org
BUG=357163

Review URL: https://codereview.chromium.org/213143002

git-svn-id: svn://svn.chromium.org/blink/trunk@170759 bbb929c8-8fbe-4397-9dbb-9b2b20218538