Make DistillerPageWebContentsTest a real test that uses the distiller.
Previously, the test used a hardcoded distiller template which didn't
help test the distiller. Updates the test to use the real distiller and
adds tests for relativization of images and anchors.

In the process, attempts to handle a few failure modes of the distiller.

BUG=353349,354737
TBR=blundell

Review URL: https://codereview.chromium.org/211493008

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@261463 0039d316-1c4b-4281-b951-d872f2087c98