Add an extra delay for policy invalidations to be available.
The cloud policy refresh scheduler no longer waits for policy invalidations to become available before issuing the first policy fetch. We want to ensure that any policy change that results from this initial policy fetch is counted as a change without invalidations enabled for metrics purposes.

Currently we determine if invalidations are enabled based on whether the invalidation service has reported that it is on. However, there may be a delay between the invalidation service being enabled and the invalidation service actually delivering a pending invalidation. Additionally, on Android the invalidation service always reports that it is enabled. To account for this, add a "grace period" after the invalidation service is enabled during which we still consider invalidations disabled for metrics reporting.

BUG=355884

Review URL: https://codereview.chromium.org/213743014

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@261506 0039d316-1c4b-4281-b951-d872f2087c98