More work to identify sw version at main resource load time, this change builds on r260672.
This adds a ServiceWorkerHostMsg_SetVersionId message handler that finds the live version, checks that its in the STARTING state, and associates it with the ProviderHost that's hooked up to observe requests from the service worker.

Added maps of 'live' versions and registrations to the context core, key'd by int64 id. (this is analogous to the 'workingset' class in appcache)

Made our browser-side object graph more navigable by giving weakptrs to the contextcore to more objects.

BUG=285976

Review URL: https://codereview.chromium.org/221473003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@261534 0039d316-1c4b-4281-b951-d872f2087c98