[FastTextAutosizer] Only create one LayoutScope for RenderTables
This patch fixes a bug were we would create two FastTextAutosizer
LayoutScopes when going down the simplifiedLayout codepath of
RenderTable. This would result in popping the cluster twice in
endLayout, which caused an assertion in clusterStack().

This is not the first bug of this flavor so I've changed the
currentCluster assert into a release assert to catch any more
instances of this bug.

BUG=359387

Review URL: https://codereview.chromium.org/224253007

git-svn-id: svn://svn.chromium.org/blink/trunk@170808 bbb929c8-8fbe-4397-9dbb-9b2b20218538