Change v8DateOrNull() to v8DateOrNaN() to allow NaN dates to be returned.
This change the double to Date conversion to map all non-finite values
to NaN instead of null. This can cause a script visible change in
behavior if non-finite values happened to get exposed by code that
provides data for the Filesystem API Mediadata.modificationTime or
WebRTC's RTCStatsReport.timestamp. In both of these cases valid dates
are always supposed to be returned according to the spec so null should
not have been getting returned unless there was a bug. Neither of these
attributes are marked as nullable so they shouldn't have been returning
null in this case anyways. The only other instance of Date being returned
is HTMLInputElement.valueAsDate and it is marked nullable so this
conversion won't even be used if a null needs to be returned.

BUG=312699
TEST=All LayoutTests still pass.

Review URL: https://codereview.chromium.org/224563003

git-svn-id: svn://svn.chromium.org/blink/trunk@170822 bbb929c8-8fbe-4397-9dbb-9b2b20218538