NodeFilter should accept a Node if the filter wrapper is already gone
This is a monkey patch to fix crashes of bug 358858 and 345486. This CL makes acceptNode() return FILTER_ACCEPT when m_filter is 0.

However, I don't think this is a right fix because m_filter should not be 0:

- We set the NodeFilter wrapper on a hidden property of a NodeIterator wrapper.

- Thus it should be guaranteed that NodeFilter is kept alive as long as the NodeIterator is alive.

- However, the stack trace of the crash indicates that the NodeFilter can die before the NodeIterator dies.

We need to investigate what revision caused the regression. However, I cannot reproduce the crash locally and it will take time to fix the crash. So I'd like to land this monkey patch for the time being.

BUG=358858,345486
TEST=Node (I cannot reproduce the crash)

Review URL: https://codereview.chromium.org/225173004

git-svn-id: svn://svn.chromium.org/blink/trunk@170833 bbb929c8-8fbe-4397-9dbb-9b2b20218538