Stop doing unnecessary initialization in non-SFI
We do not need to call NaClChromeMainInit for non-SFI NaCl.
It seems the log module is initialized in other places, and
SRPC module is initialized in nonsfi_main.

Also set uses_irt=false for non-SFI NaCl. We do not need to
send the unnecessary FD of IRT.

With this patch, we will not call modify_ldt in non-SFI mode.

BUG=https://code.google.com/p/nativeclient/issues/detail?id=3734
BUG=350487
TEST=out/Release/browser_tests --gtest_filter='NaCl*'
TEST=Our app works without modify_ldt

Review URL: https://codereview.chromium.org/218633011

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@261691 0039d316-1c4b-4281-b951-d872f2087c98