Fix for CSS escaped identifier assert
Escaped identifiers are stored as 16-bit string so using the 8-bit
string storage for them causes unexpected errors in tokenToLowerCase(). 
In debug there was an ASSERT that crashed the browser. In release
any non-ascii chars in the string would have been partly garbled.

This also removes the "const" from an argument because the function
does mutate the argument so calling it const was not very honest.

BUG=358750
R=eseidel

Review URL: https://codereview.chromium.org/213743018

git-svn-id: svn://svn.chromium.org/blink/trunk@170841 bbb929c8-8fbe-4397-9dbb-9b2b20218538