Tab Capture: Hook up VideoFrameDeliveryLog to report captures per second
In r190339 the mechanism with which capturing was managed was
changed to a push model. This change reuses the orignal
VideoFrameDeliveryLog and modifies it hook up the the capture oracle.

The reported FPS represents actual captures per second, if the source is
updating at a rate that is lower than the desired capture rate this
number will report that lower rate but this does not mean that we are
dropping frames. If the source is updating at a rate equal to or greater
than the capture rate and this reported number is less than the desired
captured rate then frames are being dropped.

BUG=223600
TEST=Run chrome with Tab Casting and observe frame rate being reported

Review URL: https://codereview.chromium.org/222603003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@261806 0039d316-1c4b-4281-b951-d872f2087c98