metrics: Add UpdateEngine.Check.* metrics.
We've rewritten the metrics reported by the ChromeOS update_engine.
It's done in a way so the new metrics (UpdateEngine.*) will live
alongside the with the existing metrics (Installer.*). We plan to
eventually remove the old metrics.

This CL adds metrics in the UpdateEngine.Check.* namespace. These
metrics do not have any counterparts in the existing Installer.*
namespace.

BUG=chromium:355745
TEST=Validated the xml, pretty print succeeded.
R=asvitkine@chromium.org, avakulenko@chromium.org, deymo@chromium.org, isherman@chromium.org, jar@chromium.org, sosa@chromium.org
NOTRY=true

Review URL: https://codereview.chromium.org/220723018

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@261904 0039d316-1c4b-4281-b951-d872f2087c98