Fix a couple of bugs in webkitpy discovered during the latest w3c import.
First, the import script (update-w3c-deps) would get upset and bail out
because we thought we were checking if the blink checkout was clean
when we were still in the web-platfom-tests checkout (which would never
be clean at the time).

Second, run-webkit-tests would fail to recognize a testharness based test
properly if it raised console errors prior to logging "This is a
testharness-based test)." I reworked the code so that any unexpected
output would be treated as a test failure assuming the header and
footer appeared *somewhere* during the test, even if not at the
beginnig or end.

R=abarth@chromium.org
BUG=242329

Review URL: https://codereview.chromium.org/225283010

git-svn-id: svn://svn.chromium.org/blink/trunk@170920 bbb929c8-8fbe-4397-9dbb-9b2b20218538