Add testRunner.displayAsync(Then)
Some tests actually need to invoke paint to test what they need, but
having a synchronous call to do painting is problematic.  Since paints
are driven by the compositor when compositing, it's not great to do
this on the stack synchronously.

These functions don't actually do anything different, but it forces the
callers to behave as if paint was asynchronous so that the underlying
implementation can be changed to be always composited without breaking
these tests.

Once layout tests are converted to use these new functions, then
testRunner.display can be removed.

R=abarth@chromium.org
BUG=345027

Review URL: https://codereview.chromium.org/224743002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@261957 0039d316-1c4b-4281-b951-d872f2087c98