Reland: Fix several double-tap gesture related issues
Always dispatch GestureTapCancel if a GestureShowPress fired before a
GestureDoubleTap, properly clearing up the active state. Also fix an
issue where a secondary pointer is pressed while a double-tap drag zoom gesture
is active, as well as a potential crash when double-tap is enabled during
a touch sequence.

This patch originally landed in r262322, but was reverted in r262335 because of
failures in TouchDispositionGestureFilterTest.  The source of the failure was
an uninitialized member variable controlling dispatch of showpress events,
which was not exposed by any of the CQ test bots.

BUG=332418

Review URL: https://codereview.chromium.org/226423003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@262435 0039d316-1c4b-4281-b951-d872f2087c98