Google+ image view flickers
The image view in Google+ has a transition set on an invisible child element.
When the transition ends, we remove the decendant clipping layer, which
previously did not trigger a tree rebuild. This CL checks the return value of
updateGraphicsLayerConfiguration and sets a bit to trigger a tree rebuild when
the configuration changes the structure of the tree.

R=esprehn@chromium.org
BUG=361108

Review URL: https://codereview.chromium.org/229793002

git-svn-id: svn://svn.chromium.org/blink/trunk@171106 bbb929c8-8fbe-4397-9dbb-9b2b20218538