Remove special case for changing from text to file types.
The m_hasType bit in <input> enforced the strange behavior that if you
changed the type of an input you could never change it to a type=file
after. Firefox doesn't do this though so it doesn't seem to be a web
compat issue.

The check also doesn't make sense since you can set the value of an input
after creating it (when the m_hasType bit is false) and then set the type
to file which would be the same as the behavior the comment in the code
is claiming to prevent.

Note that actually creating the InputType inside ::attach() would hit the
ASSERT in incDOMTreeVersion() since you're not allowed to mutate the DOM
inside recalcStyle. In practice it was impossible to hit that case since
if you never changed the type we'd early return when the default type of
text matched the if (m_inputType->formControlType() == newTypeName) check.
If you did change the type then we would have synchronously updated the
type as well so the m_hasType bit would have been true and we wouldn't
have even called it in ::attach().

Review URL: https://codereview.chromium.org/212603022

git-svn-id: svn://svn.chromium.org/blink/trunk@171111 bbb929c8-8fbe-4397-9dbb-9b2b20218538