Type extensions should apply similar checks to custom tags
When creating a Custom Element with a custom tag, the element must be
created by a document with a Custom Element registration context and
have a tag name that conforms to being a valid Custom Element type
name.

Nobody was enforcing a similar check for Custom Elements with type
extensions (for example <button is="panic-button">.) Do that.

BUG=360429
TEST=fast/dom/custom/invalid-type-extension-crash.html
R=haraken@chromium.org

Review URL: https://codereview.chromium.org/230293002

git-svn-id: svn://svn.chromium.org/blink/trunk@171127 bbb929c8-8fbe-4397-9dbb-9b2b20218538