Changing between multicol and regular block shouldn't recreate all renderers.
Make the new (flowthread-based) multicol implementation cope without
having to reattach renderers for all nodes in the subtree.

Up until recently (last year) a reattach was only done for the new
flowthread-based multicol implementation, while it wasn't done for the
old implementation (ColumnInfo), but due to some refactoring of the
code that determines whether or not to reattach on style changes, any
changes to column-width or column-count would result in a reattach,
regardless of which implementation was used.

Now that both implementations can handle it, there's no need to force
a reattach anymore.

Review URL: https://codereview.chromium.org/218663004

git-svn-id: svn://svn.chromium.org/blink/trunk@171192 bbb929c8-8fbe-4397-9dbb-9b2b20218538