Split the invisible-descendants-should-not-affect-opt-in test up.
The test does a for loop over 3 cases, checking if composited is enabled
at all to determine if we've promoted layers. When ForceCompositingMode
is on, we can not use this test to verify promotion. So split the
test's 3 cases into separate tests, and each one dumps the layer tree
so that we can verify promotion/not correctly.

R=enne@chromium.org, vollick@chromium.org
BUG=361729

Review URL: https://codereview.chromium.org/232043002

git-svn-id: svn://svn.chromium.org/blink/trunk@171210 bbb929c8-8fbe-4397-9dbb-9b2b20218538