Add MessagePumpObserver for MessagePumpOzone.
A MessagePumpObserver is an object that receives global
notifications from the UI MessageLoop. It contains two APIs
WillProcessEvent() and DidProcessEvent(), so that observer could
do customized job before or after processing event in MessagePump.

Ozone-wayland implementation maps HW input as XKB key event and
supports to dispatch the event in Ozone message queue, while Crosswalk
doesn’t support fetching the event atm. A solution is to implement
the MessagePumpObserver APIs for OzoneMessagePump, so that Crosswalk
Runtime will be able to retrieve the HW key event.