Web Animations API: Simplify and remove finished players and animations
This patch makes us discard any players in forwards fill with an effect that
cannot ever be applied due to being replaced by another player. We also simplify
interpolations of forwards filling animations that don't meet the requirements
for being discarded.

https://docs.google.com/a/chromium.org/document/d/1OXCmPnwZMICs9ErZIj54E18_k6EhTVJ5OmAh8dSq2ao/edit
(The second approach for simplifying players is implemented here.)

I've also tweaked the logic in setting the finished flag on AnimationPlayer so
that if we don't have any listeners, we'll set the flag on an update not from
an animation frame (for purposes of the unit test).

TBR=dstockwell
BUG=360473

Review URL: https://codereview.chromium.org/221553006

git-svn-id: svn://svn.chromium.org/blink/trunk@171256 bbb929c8-8fbe-4397-9dbb-9b2b20218538