ozone: Rename XKeyboard to ImeKeyboard & use fake under ozone
This fixes some build errors for the "chromeos=1 use_ozone=1" build. We
need to generalize this X11 interface for controlling the system keyboard
for use under ozone.

This currently forces ozone builds to use the fake implementation, which
merely fixes the build. We still need to implement an ozone version.

BUG=361341
TEST=chrome --ozone-platform=test
TBR=derat, sky

Review URL: https://codereview.chromium.org/232333002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@263219 0039d316-1c4b-4281-b951-d872f2087c98