Modify supercluster multiplier logic.
Previously the multiplier was based on the first root touched by the current
layout pass (for table cells) or the deepest common ancestor of all the roots
(for other superclusters).

We now base the multiplier on the max known width of any root.  This is still
not great since not all widths are known, but it gives us consistent results
on hacker news.

BUG=361749
R=pdr@chromium.org

Review URL: https://codereview.chromium.org/231663006

git-svn-id: svn://svn.chromium.org/blink/trunk@171364 bbb929c8-8fbe-4397-9dbb-9b2b20218538