Get rid of code duplication in LinkRelAttribute
There was a "fast path" that just compared the strings and then another
that split and then compared. The two paths were slightly different for
no obvious reasons which doesn't really follow the spec where the tokens
can be all mixed with each other (except for "shortcut icon" apparently).

Allocating the separate strings for "alternate stylesheet" should not be
an issue.

I also removed the extra constructor, we can just use a default argument
value.

The only behavior change is that now we allow things like "alternate import"
(or dns-prefetch) and just ignore the alternate part when deciding what
to do with the link.

Review URL: https://codereview.chromium.org/235063003

git-svn-id: svn://svn.chromium.org/blink/trunk@171380 bbb929c8-8fbe-4397-9dbb-9b2b20218538