Set track_start on solid color layers to the margin.
The scrollbar margin on the overlay scrollbar theme is used to distance
the thumb from the edge of the screen on all sides, not just the
widthwise side.  This was already computed properly in the theme's
TrackRect, but it wasn't communicated to the WebScrollbarLayer.

Also, shrink the margin to 3 DIPs on Android, which on closer inspection
better matches the system theme.

Chromium-side change at https://codereview.chromium.org/234603002/
NOTRY=true
BUG=361923

Review URL: https://codereview.chromium.org/234613002

git-svn-id: svn://svn.chromium.org/blink/trunk@171404 bbb929c8-8fbe-4397-9dbb-9b2b20218538