[Android] Address Context leak in ResourceExtractor
The leak is fixed by commit "Make the ResourceExtractor
use the App context"(0f1465d5f), which is reverted
because xwalk shared mode needs to keep original Context.

In this commit, the static instance of ResourceExtractor
will be released after the background job is done. The
ResourceExtractor should not be used any more from that point.