Fix table repaint after r171183
My previous patch brought LayoutState-optimized repaint logic to paint
codepaths that don't use a null repaintContainer to indicate that
RenderView is the repaintContainer. Unfortunately, there was still
some code in RenderTableCell with that assumption, which could lead
to us missing the inclusion of margins.

The fix is easy, the test is a bit ugly to ensure that we don't trigger
any invalidations in the tile that fails to repaint. This isn't strictly
necessary for the test, but failing to do so means the issue won't
visually reproduce due to the compositor's repainting algorithm.

BUG=363411

Review URL: https://codereview.chromium.org/238773004

git-svn-id: svn://svn.chromium.org/blink/trunk@171721 bbb929c8-8fbe-4397-9dbb-9b2b20218538