clang plugin: Fix tests when running on OS X 10.9
On 10.9, trunk clang defaults to targeting libc++. Clang expects to bundle
the libc++ headers to be able to use libc++, but we don't currently do this.
Instead, tell the plugin tests to use libstdc++ and pass a sysroot flag.

(The blink oilpan plugin doesn't use and standard library includes, so it
doesn't need this change at the moment.)

BUG=none
TEST=tools/clang/plugins/tests/test.sh third_party/llvm-build/Release+Asserts
NOTRY=true

Review URL: https://codereview.chromium.org/239753005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@264089 0039d316-1c4b-4281-b951-d872f2087c98