Preserve Page::openedByDOM state across process swaps.
Every Page in Blink has an openedByDOM bit which partially determines whether
window.close() may be called. This bit is equivalent to whether the WebContents
was created with an opener. Move this bit up to WebContents and forward it to
new renderers to be preserved across process swaps. Add a suite of tests to
ensure it's preserved correctly.

This also fixes a bug where the bit was not preserved for blocked-then-opened
popups. Add a test for this case.

BUG=357579

Review URL: https://codereview.chromium.org/212703005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@264291 0039d316-1c4b-4281-b951-d872f2087c98