requestAutocomplete should not unlatinize administrative region names.
requestAutocomplete canonicalizes administrative region names before
sending an address to Wallet. For example, "California" becomes "CA".
"California" is a "name" and "CA" is a "key". Similar logic cannot be
used for switching from "lname" to "key", however. For example, "Tokyo"
should not become "東京都", because we should avoid mixed-language
addresses. This patch prevents unlatinizing of administrative region
names when canonicalizing an address to be sent to Wallet.

BUG=362691

Review URL: https://codereview.chromium.org/237693002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@264435 0039d316-1c4b-4281-b951-d872f2087c98