Add IPC Channel for new ManifestService.
This CL adds a new IPC Channel between NaCl plugin and the renderer process
with introducing ManifestService (in the plugin) and ManifestServiceChannel
(in the renderer) as its end points.
Currently, ManifestService is just an empty service. Its functions will be
added in following CLs. The service will be used only for non-SFI mode
as a first step. On other platforms, IPC Channel will not be created.

TEST=Ran trybots.
BUG=358431

Review URL: https://codereview.chromium.org/231793003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@264477 0039d316-1c4b-4281-b951-d872f2087c98