Ignore clip-rect in the chromoting client plugin
In some cases Chrome passes incorrect clip-rect to the plugin, and
so the plugin doesn't render content correctly in such cases. Fixed
the plugin to ignore clip-rect from the browser.

This may degrade performance in some cases, but not by much and only 
until we switch to using MediaSource API for rendering.

BUG=360240

Review URL: https://codereview.chromium.org/240833008

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@264727 0039d316-1c4b-4281-b951-d872f2087c98