Reland of Only create webrtc MediaStreams when connected to a PC.
This reverts commit d5158f4cec5dc31309ef93b66a8f8b46ae708fc0.
https://src.chromium.org/viewvc/chrome?view=rev&revision=264300

Furthermore it adds a MediaStreamObserver (patch set 2) that can be used to get notifications when a track is removed or added from a MediaStream. It is needed to allow modifying a MediaStream that is already beeing sent on a PeerConnection and fixes the problem showed in:
http://build.chromium.org/p/chromium.webrtc/builders/WinXP%20Tester/builds/8185
http://build.chromium.org/p/chromium.webrtc/builders/Win8%20Tester/builds/1202
http://build.chromium.org/p/chromium.webrtc/builders/Mac%20Tester/builds/21768

Original cl description:
This cl adds a new class WebRtcMediaStreamAdapter. The responsibility of the class is to create and own a representation of a webrtc MediaStream that can be added and removed from a PeerConnection.
An instance of WebRtcMediaStreamAdapter is created when a MediaStream is added to RTCPeerConnection object.

The purpose is to clean up the the webrtc specific use of MediaStreams.

The original cl was commited here:  https://src.chromium.org/viewvc/chrome?view=rev&revision=262395

BUG=323223
R=xians@chromium.org

Review URL: https://codereview.chromium.org/240723004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@265212 0039d316-1c4b-4281-b951-d872f2087c98