Merge updateRenderTreeForNodeIfNeeded checks into needsRenderTreeUpdate
The checks in this function were brittle since they had to be kept in
sync with the ones in needsRenderTreeUpdate(). Instead we can merge the
checks together.

This patch introduces needsFullRenderTreeUpdate() which means the Document
level state is dirty and we need to run a render tree update, allowing
updateRenderTreeForNodeIfNeeded to only check the other bits. It also moves
the style invalidation bit checks into the the ancestor loop.

Finally I added a FIXME to the loop. It needs to be over the composed tree.
I have a test case and will fix the bug with walking over the non-composed
ancestors in the next patch.

Review URL: https://codereview.chromium.org/246003002

git-svn-id: svn://svn.chromium.org/blink/trunk@172142 bbb929c8-8fbe-4397-9dbb-9b2b20218538