Convert notifications to use RenderFrame instead of RenderView.
ContentBrowserClient::CancelDesktopNotification is removed in favor of having ShowDesktopNotification return a callback to cancel the notification, which seems simpler.

DesktopNotificationDelegate is a new interface passed to ShowDesktopNotification which hides the implementation details of sending IPCs to the right RenderFrame, and the notification_id parameter.

BUG=304341
R=nasko@chromium.org

Review URL: https://codereview.chromium.org/241553002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@265295 0039d316-1c4b-4281-b951-d872f2087c98