Fix notification crash when a page removes it using notification.close().
The problem was introduced in r265295, because the string holding the id gets deallocated when the callback is removed from a map, but it used later. This is tested by NotificationsTest.TestCancelNotification, but the notification tests were accidentally removed in r262988, so readding them.

BUG=366090
R=stevenjb@chromium.org

Review URL: https://codereview.chromium.org/250043002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@265758 0039d316-1c4b-4281-b951-d872f2087c98