cc: Use the onscreen context to perform composited ganesh filters.
This avoids the need for an offscreen context at all. Since the GPU
rasterization path is already doing similar stuff, we need to support
both cc and ganesh using the same context. This is a bit trickier
since it occurs in the middle of drawing a frame instead of between
frames, but as long as we restore all state that GLRenderer ever sets,
it works.

R=enne@chromium.org, senorblanco@chromium.org
BUG=366132

Review URL: https://codereview.chromium.org/250083002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@265810 0039d316-1c4b-4281-b951-d872f2087c98