Implement syncing of bookmark apps.
This CL adds syncing of bookmark apps. This involves syncing two extra fields,
which are the bookmark app url and description. These are not populated for
non-bookmark apps. The bookmark app name is obtained from the extension sync
data's name field.

Bookmark apps that are installled from sync will not have any icon. Updating
the icon and possibly fetching it on sync will be implemented in a future CL.
Bookmark apps that are updated with a new name will keep their icons.

This CL also adds a GetWebApplicationInfoFromBookmarkApp() function that
returns a ready-to-install WebApplicationInfo that is populated with the app's
details and icons.

BUG=318607

Review URL: https://codereview.chromium.org/229553003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@265887 0039d316-1c4b-4281-b951-d872f2087c98