Linux App Launcher has smarter positioning based on OS shelf position.
The heuristic is to snap to the edge of the display with the largest
area missing from the work area (this is likely to be the desktop
environment's shelf or launcher).

There is a special case for Ubuntu Unity (always snap to the left edge).

It is a known issue that on Gnome Shell, it will snap to the top edge
when it should snap to the left (this is a minor issue; it still feels
okay).

BUG=365563
TEST=On Ubuntu Unity, launch App Launcher from shelf (set to auto-hide).
Should snap to left edge.
TEST=On Gnome Classic, launch App Launcher from Applications menu.
Should snap to top edge.

Review URL: https://codereview.chromium.org/255473004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@265907 0039d316-1c4b-4281-b951-d872f2087c98