Revert 172495 "Make window.screen property non replaceable"
This broke the Maps pixel test on the GPU bots, because it is setting
window.screen to null to work around a bug in the (somewhat old) version of
the Maps team's code. We will work with the Maps team to get an updated
version of the test ASAP, but in the meantime to prevent these bots from
being red for a day or more I am reverting this change. Sorry.

http://build.chromium.org/p/chromium.webkit/builders/GPU%20Mac10.7%20%28dbg%29/builds/24997

> Make window.screen property non replaceable
> 
> Make window.screen property non replaceable for consistency with other
> browsers (tested with Firefox 28 and IE11).
> 
> This CL also adds an internals.setOverrideScreenData() utility function for
> Layout tests as the fast/forms/page-popup/page-popup-adjust-rect.html test
> needs to be able to override the screen object.
> 
> BUG=364937
> TEST=fast/dom/Window/get-set-properties.html
> 
> Committed: https://src.chromium.org/viewvc/blink?view=rev&revision=172134
> 
> Review URL: https://codereview.chromium.org/243173004

TBR=ch.dumez@samsung.com

Review URL: https://codereview.chromium.org/259523005

git-svn-id: svn://svn.chromium.org/blink/trunk@172524 bbb929c8-8fbe-4397-9dbb-9b2b20218538