Preserve identical values for background-size and -webkit-mask-size.
This patch fixes a bug where background-size/-webkit-mask-size would be
incorrectly serialized to text when the values in the pair are identical.
Setting 'background-size: 100% 100%', for example, would produce a text
representation of 'background-size: 100%', which is treated as
'background-position: 100% auto'.

For -webkit-background-size, '100%' really does mean '100% 100%', so we
maintain a policy of 'DropIdenticalValues' for this property.

BUG=364962
TEST=fast/backgrounds/size/parsing-background-sizes-values.html
     fast/masking/parsing-mask.html

Review URL: https://codereview.chromium.org/249803002

git-svn-id: svn://svn.chromium.org/blink/trunk@172548 bbb929c8-8fbe-4397-9dbb-9b2b20218538