Remove case DocumentLifecycle::canAdvanceTo that allowed post layout tasks to recalcStyle
Even with the patch to turn on force compositing mode in layout tests I
can't hit a recalcStyle from inside post layout tasks in the updateAppearance()
logic. I don't think we need this transition now.

I also removed a transition that was for partial layout which has been removed.

It's still probably possible for a plugin to make us recalcStyle from inside
of post layout tasks, but unless tests trip that we should leave the ASSERT
such that new code can't be added that does it.

BUG=363844

Review URL: https://codereview.chromium.org/251373010

git-svn-id: svn://svn.chromium.org/blink/trunk@172565 bbb929c8-8fbe-4397-9dbb-9b2b20218538