about:sync: Fix refresh crash and broken node browser
The refresh crash was getting in the way of debugging and reproducing the node
browser issue.  This CL fixes both at once.

The refresh crash happened because the SyncInternalsMessageHandler will retain
its state across a refresh, but the JavaScript code it talks to will not.  The
message handler will receive a second message from the JavaScript code to
inform it that it's time to register with the ProfileSyncService.  This CL adds
a check to ensure that this double-request will not result in
a double-registration.

The node browser issue was caused by a mismatch between the C++ and JavaScript
APIs.  The C++ side passed an array of arguments, but the JavaScript code was
expecting the two arguments to be passed separately.

BUG=363300,360197

Review URL: https://codereview.chromium.org/248153003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@266079 0039d316-1c4b-4281-b951-d872f2087c98