Show drive-related settings on gear menu only when Ctrl key is pressed.
Listen to Ctrl key in FileManager, and when gear menu is opened while Ctrl key is pressed, the gear menu is shown as secret mode.
The drive-related setting items will be shown only when gear menu is in secret mode and drive volume is focused.
Drive-related settings are now implemented as Comands to be consistent with other menu items.

BUG=239365

Review URL: https://codereview.chromium.org/243013009

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@266098 0039d316-1c4b-4281-b951-d872f2087c98