Introduces 'highlighted' property to candidate view.
With my previous 'cleanup', the candidate view becomes a CustomButton
and the 'PRESSED' button status indicates the highlighted. I thought
this is good enough, but hovering also changes the button state.

Using a button state is a bad idea, so this CL adds a boolean flag
for highlighted status.

BUG=365959
R=oshima@chromium.org
TEST=manually, ash_unittests

Review URL: https://codereview.chromium.org/258463002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@266123 0039d316-1c4b-4281-b951-d872f2087c98