Don't check isActive inside setCompositingLayersNeedRebuild.
This was needed because we'd call setCompositingLayersNeedRebuild
on a document that was being destroyed because RenderLayer
calls setCompositingLayersNeedRebuild when a layer is removed.
Then trying to set the document lifecycle state to LayoutClean
would assert since we can't go from Stopping to LayoutClean.

With https://src.chromium.org/viewvc/blink?view=rev&revision=172591,
we no longer call setCompositingLayersNeedRebuild once we've
entered the Stopping state, so we don't need this early return.

Review URL: https://codereview.chromium.org/257843005

git-svn-id: svn://svn.chromium.org/blink/trunk@172674 bbb929c8-8fbe-4397-9dbb-9b2b20218538