Move DefaultActivationClient to wm/core
Several tests are using AuraTestHelper and were relying on it to set up the DefaultActivationClient, so I had to have them instantiate the DAC directly after calling SetUp() on the helper. I also had to make the DAC delete itself after the root window it's bound to is destroyed.

R=sky@chromium.org
http://crbug.com/308710

Review URL: https://codereview.chromium.org/250113002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@266616 0039d316-1c4b-4281-b951-d872f2087c98