Reland "GestureRecognizerTest tests both Aura and Unified Gesture Detectors."
Reland "GestureRecognizerTest tests both Aura and Unified Gesture Detectors."

This is in preparation of landing the unified GR
(see https://codereview.chromium.org/251543003/).

There are some cases where behavior differs in minor ways.

This is sometimes dealt with by increasing the test's margin of error
(for instance, slop handling is slightly different, so we now
exceed the slop more thoroughly).

This is sometimes dealt with by ignoring the test if the test is
run with the unified GR.

BUG=332418

Review URL: https://codereview.chromium.org/252833005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@266652 0039d316-1c4b-4281-b951-d872f2087c98