Linux GPU sandbox: only allocate broker policy in the broker.
The GPU broker policy was allocated in the main GPU process and then used in
the broker process. We switch the logic so that the broker policy is only ever
allocated in the broker process itself.

Besides fixing a small memory leak (in the GPU process), this makes sure that a
policy is only ever used in the process that allocated it. This will allow to
bind policies with properties such as "which processes does this policy allow
to send signal to".

BUG=367986
R=jorgelo@chromium.org

Review URL: https://codereview.chromium.org/251183004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@266726 0039d316-1c4b-4281-b951-d872f2087c98