[FastTextAutosizer] Enable Fast Text Autosizer by default (1 of 3)
This patch enables the Fast Text Autosizer by default.

The following WebFrameTests have been modified to not depend on the
internal details of the old autosizer:
1) ChangeInFixedLayoutResetsTextAutosizingMultipliers
This test checks that the autosizing multipliers are reset after calling
updatePageDefinedViewportConstraints. This correctly exercises
FastTextAutosizer::resetMultipliers but because the page doesn't actually
autosize we end up not resetting the multipliers due to an optimization in
the fast text autosizer. I've changed this test so that the page autosizes.
2) SetFrameRectInvalidatesTextAutosizingMultipliers
This is a new test that verifies that we call
FastTextAutosizer::setAllTextNeedsLayout on frame rect changes.

The virtual fasttextautosizing directory has been moved and all relevant
tests have been put in fast/text-autosizing. I have also filed
crbug.com/367493 to cleanup the virtual test directory expectations.

For reference, the second patch is queued up:
https://codereview.chromium.org/249573003

This patch requires that we land
https://codereview.chromium.org/258823009 first.

BUG=365911

Review URL: https://codereview.chromium.org/249943002

git-svn-id: svn://svn.chromium.org/blink/trunk@172846 bbb929c8-8fbe-4397-9dbb-9b2b20218538