Convert KeyboardDrivenEventRewriter to a ui::EventRewriter.
KeyboardDrivenEventRewriter was previously called at the start of
chromeos::EventRewriter, and suppressed further rewriting if it made
any changes. Since EventSource follows that model of breaking out of
the rewriter list, it makes sense to have this be a ui::EventRewriter
itself.

EventRewriterController will also be used for chromeos::EventRewriter
when that is converted to a ui::EventRewriter.

BUG=354034
TEST=unit_test
R=derat@chromium.org, sadrul@chromium.org

Review URL: https://codereview.chromium.org/240173007

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@266969 0039d316-1c4b-4281-b951-d872f2087c98