Allow compiling Android with notifications=1, add stubbed UI manager
This patch makes it possible to compile Android's chrome_shell_apk target
whilst setting notifications=1 in the GYP_DEFINES. It also adds a stubbed
version of the NotificationUIManagerAndroid, which will be responsible
for displaying notifcations (through the Android framework).

Chrome for Android will not be using the new Message Center, because
displaying notifications will be done by the platform directly.

BUG=90795

Review URL: https://codereview.chromium.org/247943003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@267041 0039d316-1c4b-4281-b951-d872f2087c98