[New Multicolumn] Make offsetLeft, offsetTop and getClientRects() behave.
Get rid of mapFromFlowToRegion() and mapLocalToContainer() in
RenderFlowThread, and reimplement columnOffset() in
RenderMultiColumnFlowThread instead. This method is already used a lot
in the engine, but it has up until now only worked with the old multicol
implementation. Need to add a call from
RenderBox::offsetFromContainer(), since the one that's already there
is for old multicol only and does special stuff.

This incidentally also gets rid of assertion failures in
RenderGeometryMap when compositing is turned on (what used to be in
RenderFlowThread::mapLocalToContainer() did the wrong thing if the
target was that flow thread itself). Re-enable tests that used to
fail on that assertion.

BUG=367688,364960

Review URL: https://codereview.chromium.org/256743006

git-svn-id: svn://svn.chromium.org/blink/trunk@172961 bbb929c8-8fbe-4397-9dbb-9b2b20218538