Move methods from chrome.bluetooth to chrome.bluetoothSocket
The updateSocket(), setSocketPaused(), send(), getSocket() and
getSockets() are moved to chrome.bluetoothSocket, renaming as necessary,
with the old implementation replaced by a stub that returns an error.
When un-paused sockets generate chrome.bluetoothSocket.onReceive and
chrome.bluetoothSocket.onReceiveError events instead of the old events
which are now unused.

Also implements chrome.bluetoothSocket.create() and close() to complete
the lifetime of the socket.

Since BluetoothSocket is self-contained, and does not need references to
the adapter or observer methods, BluetoothSocketEventDispatcher changes
from being a refcounted object to BrowserContextKeyedAPI, matching the
sockets API.

BUG=349475,364581

Review URL: https://codereview.chromium.org/252673004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@267344 0039d316-1c4b-4281-b951-d872f2087c98