Don't early out in PreCalculateMetaInformation
Currently, we bail early in this function if we encounter a layer with
a singular transform. This causes problems because we depend on a full
tree walk in order to clear DrawProperties::sorted_for_recursion. This
cl prevents this early out.

(Note: this does not affect other optimizations to prevent the drawing
of subtrees rooted at layers with singular transforms.)

R=ajuma@chromium.org,danakj@chromium.org
BUG=366233

Review URL: https://codereview.chromium.org/263643007

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@267438 0039d316-1c4b-4281-b951-d872f2087c98