Remove get/set currentPath in canvas
currentPath is not the right way to apply a Path2D object to a context.
Main argument is that currentPath creates a copy of the Path2D object.
Instead we pass Path2D as argument to fill, stroke and clip now.
(The way isn't defined in the w3c specification as well)

Intent to Deprecate and Remove:
https://groups.google.com/a/chromium.org/forum/#!topic/blink-dev/6l01Mv3VND0

BUG=362962

Review URL: https://codereview.chromium.org/226953004

git-svn-id: svn://svn.chromium.org/blink/trunk@173073 bbb929c8-8fbe-4397-9dbb-9b2b20218538