Fixed ordering assumption bug in AppInfoPermissionsTabTest
Fixed AppInfoPermissionsTabTest::RetainedFilePermissionsObtainedCorrectly test
to no longer assume that retained files are returned in a particular order.

Since the files are read as keys of a dictionary in
SavedFilesService::GetAllFileEntries(), there are no guarantees on the ordering
of the files, and this is now reflected in the test.

BUG=NONE

Review URL: https://codereview.chromium.org/261753003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@267509 0039d316-1c4b-4281-b951-d872f2087c98