Slice off DelegatedFrameHost from RenderWidgetHostViewAura
The DelegatedFrameHost is responsible for providing the interface 
between the delegated frames coming in to a RenderWidgetHostView and
the a ui::Compositor being used to display them.

The functionality in DelegatedFrameHost will be used by the browser
compositor in RenderWidgetHostView.

This patch is done in such a way as to minimize the functional changes
made, in an effort to separate them from code reorganization changes.
The two places where this will need to be revisited (the complication
of the testing interface, and the redundant device scale functions) are
noted with TODOs.

BUG=314190

Review URL: https://codereview.chromium.org/258263002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@267539 0039d316-1c4b-4281-b951-d872f2087c98