DontPromptWhenEnableAutomaticPasswordSavingSwitchIsSet: Do not click on save.
PasswordManagerBrowserTest.DontPromptWhenEnableAutomaticPasswordSavingSwitchIsSet needs to check in official builds, that passwords are not auto-saved. The default behaviour of the browser test is, that even if the save-password-prompt is shown, its "Save" button is clicked. As a result, the passwords get saved in the browser test even if auto-saving is off. This CL adds an override of that behaviour for the particular test for auto-saving.

The CL introducing the failing test was: https://codereview.chromium.org/256003003

BUG=368778,367690
TBR=rchtara@chromium.org,gcasto@chromium.org

Review URL: https://codereview.chromium.org/269573005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@267556 0039d316-1c4b-4281-b951-d872f2087c98