cc: Reraster and fix the raster source scale factor when it changes.
When a web page changes the scale of elements in the page outside of an
animation, we should have that cause us to make a new tiling at the new
scale.

However, we don't want to do this every main thread frame. So when the
first change happens, we set raster_source_scale_is_fixed_ to true
and change the source scale to 1, ignoring any css transform scales
when rastering the layer from then on.

R=enne
BUG=368201

Review URL: https://codereview.chromium.org/262613005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@267570 0039d316-1c4b-4281-b951-d872f2087c98