Add the means to disable GraphicsContext.
For performance measurements we would like to isolate
Blink paint calls from the graphics back-end that implements
the GraphicsContext API. This patch adds a flag to GraphicsContext
to disable most functionality, with the exception of graphics state
get and set with a single state container.

The flag is plumbed through from the WebContentLayerClient, enabling
it to be accessed from Chromium.

The flag itself should have negligable overhead, as we replace the
existing painting disabled flag with a common flag, so no additional
cost to existing API methods.

R=abarth@chromium.org,fmalita@chromium.org
BUG=350684

Review URL: https://codereview.chromium.org/260123004

git-svn-id: svn://svn.chromium.org/blink/trunk@173096 bbb929c8-8fbe-4397-9dbb-9b2b20218538