Fix Amazon page in key_mobile_sites
On the mobile Amazon page, document.body.getBoundingClientRect() returns a
height of 0 and top of -1. So instead of scrolling on the page, use the main
element of that page for scrolling. We need to specify the scroll distance
explicitly though, because the element itself isn't scrollable (clientHeight == scrollHeight).

BUG=368947

Review URL: https://codereview.chromium.org/267713002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@267620 0039d316-1c4b-4281-b951-d872f2087c98