Make same-document history navigations commit with the proper type.
There is a quirk in HistoryController where we commit a history navigation immediately upon starting a same-document navigation. If a same-document history navigation is committed with a HistoryCommitType of BackForwardCommit instead of HistoryInertCommit, the commit logic does the right thing and we don't need that quirk.

BUG=

Review URL: https://codereview.chromium.org/252323003

git-svn-id: svn://svn.chromium.org/blink/trunk@173142 bbb929c8-8fbe-4397-9dbb-9b2b20218538