Directly schedule compositing updates for <canvas>
Instead of using scheduleLayerUpdate() to trigger the compositing state
change for the <canvas>, we can directly set the bits on the
RenderLayerCompositor and our RenderLayer.

This removes several usages of scheduleLayerUpdate(), which we want to
remove since it requires calling into setStyle() with identical styles
and is the source of chicken and egg issues with the compositing update.

Review URL: https://codereview.chromium.org/265833007

git-svn-id: svn://svn.chromium.org/blink/trunk@173232 bbb929c8-8fbe-4397-9dbb-9b2b20218538