Don't clobber equivalent paint info when changing squashing assignment.
This CL makes two changes:

1) If the new and old GraphicsLayerPaintInfo objects are
equivalent for squashing, there's no need to overwrite the
other fields in the struct.

2) When we add or remove a RenderLayer from a squashing
layer, we now mark the CompositedLayerMapping for a graphics
update. This change ensures that we'll program the data about
the newly squashed layers into the compositor.

This CL is based on https://codereview.chromium.org/265083002
by vollick@

BUG=368504

Review URL: https://codereview.chromium.org/261983006

git-svn-id: svn://svn.chromium.org/blink/trunk@173237 bbb929c8-8fbe-4397-9dbb-9b2b20218538