net: Avoid recursion in URLRequestJob::ReadFilteredData.
There's no need to use recursion when getting the data from a filter.
This version uses a loop instead.

The issue with recursion is that it opens the door to stack exhaustion
if there is no way to limit its depth. In this case, that would be a
hard thing to do given that it basically depends on interactions outside
of our control.

BUG=366149

Review URL: https://codereview.chromium.org/267793002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@268030 0039d316-1c4b-4281-b951-d872f2087c98