Password bubble: Create subviews for ManagePasswordsBubbleView.
Same idea as we've done for ManagePasswordItemView: separate subview
classes make it simpler to change the content and to understand where
the content is coming from.

After this CL, it will be a simpler addition to shoehorn in a blacklist
confirmation screen in the same way that we've done undo for the item
views.

-----------------------------------------------------------------------
This is a reland of r267213, which was reverted in order to revert a
different patch (r266859) due to LSAN errors. The original review is
https://codereview.chromium.org/255903006/
-----------------------------------------------------------------------

BUG=261628

Review URL: https://codereview.chromium.org/266973002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@268179 0039d316-1c4b-4281-b951-d872f2087c98