Password bubble: Give users the option of unblacklisting a site.
This CL adds a new view to the bubble, to be displayed to the user when
she clicks on the icon for a site where she's previously elected to
"Never save passwords". From this view, she may elect to unblacklist the
site, enabling passwords to be saved going forward.

To do so, a PasswordFormMap is piped through the password manager client
when sending a BlacklistBlockedAutofill notifiation. That contains
the information we'll need to pass into the PasswordStore to remove the
blacklist.

BUG=357597

Review URL: https://codereview.chromium.org/259153004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@268202 0039d316-1c4b-4281-b951-d872f2087c98