aw: Fix invalidate getting lost
An invalidate could be lost in this case:
* SetNeedsContinuousInvalidate starts out as true
* Functor happens on render thread, and SetNeedsContinuousInvalidate
goes false, but because we are on render thread, this is posted back
to UI thread
* Before the task runs on UI, compositor SetNeedsContinuousInvalidate
to true again, since the value is already true, this is a no-op, but
we should really be invalidating now
* The posted task runs and SetNeedsContinuousInvalidate to false,
and we just lost an invalidate

Fix by updating the value with a lock and only post for the invalidate.
This avoids the race causing early out in SetNeedsContinuousInvalidate.

For internal b/13979752
BUG=

Review URL: https://codereview.chromium.org/269913002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@268258 0039d316-1c4b-4281-b951-d872f2087c98