Mark when we may have been invalidated to early out on repaintTreeAfterLayout.
This CL adds in a flag to mark if we may have been invalidated. If the flag is
set we will propagate the flag up to our parents. This allows us to early
out of the repaintTreeAfterLayout walk if at any point the node does not
need to be invalidated.

This improves the performance of the
large-table-with-collapsed-borders-and-colspans-wider-than-table.html test
by ~17% and removes some over-invalidations that were output.

I visually verified the items in TestExpectations to make sure they matched.

BUG=320139, 368739

Review URL: https://codereview.chromium.org/264963004

git-svn-id: svn://svn.chromium.org/blink/trunk@173348 bbb929c8-8fbe-4397-9dbb-9b2b20218538