Android: Make ContentViewCore's layer a SolidColorLayer
Change ContentViewCore's container layer to be a SolidColorLayer
that uses the view's background color and physical dimensions.
But only make it drawable as long as there is no RenderView or
we don't have a frame yet to avoid any potential draw (or drawcalc)
overhead.

NOTRY=True

Review URL: https://codereview.chromium.org/264113003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@268530 0039d316-1c4b-4281-b951-d872f2087c98