sync: Plumb per-type counter observers to UI
Implements the plumbing to get events from the
DirectoryTypeDebugInfoEmitters in the ModelTypeRegistry, through the
SyncManager, SyncBackendHostCore, and SyncBackendHost to the
ProfileSyncService.

At the moment, there are no observers register and no events
are emitted.  Future patches will include logic to update the
counters and to expose their values on the about:sync page.

BUG=328606

Review URL: https://codereview.chromium.org/269543020

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@268747 0039d316-1c4b-4281-b951-d872f2087c98