PNaCl translator: teach Chrome to send CPU features to the compiler
The PNaCl compiler used to assume SSE2 and no more, this is very
suboptimial when dealing with SIMD, which is newly supported by PNaCl.

This implementation only detects the basic CPU features that Chrome's
base/cpu.h knows about, and passes along that information to the PNaCl
translator so that it may compile using the features. It also doesn't
pass AVX and later as features since those aren't currently supported by
the NaCl sandbox.

The extra flags are also passed to the cache, so that a user can swap in
a different CPU and get the expected retranslation.

This work could also teach PNaCl about other CPU features such as popcnt
or idiv/udiv on ARM, but is restricted to x86 SIMD for now since that's
what base/cpu.h supports.

R= dschuff@chromium.org, jvoung@chromium.org, nfullagar@chromium.org
TEST= ./out/Release/browser_tests --gtest_filter=NaClBrowserTestPnacl.*:PPAPINaClPNaClTest.*:NaClBrowserTestPnacl.*
BUG= https://code.google.com/p/nativeclient/issues/detail?id=2205

Review URL: https://codereview.chromium.org/265933008

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@268761 0039d316-1c4b-4281-b951-d872f2087c98