RFHM teardown issue: RenderFrameHostManager::RenderViewDeleted should not need to clean anything up. Nowadays RenderViewHosts are owned by the RenderFrameHosts that reference them, and the RenderFrameHosts are owned (in a few different ways) by the RenderFrameHostManager. Thus, it shouldn't need to delegate back to RFHM to clean up a view deletion, because it's the RFHM that's initiating the deletion of the view in the first place, and that cleanup should occur higher on the stack.
In the case of the proxy_hosts_ cleanup code in RenderViewDeleted, the host is already being deleted higher on the stack, and deleting here actually could mess with the iterators kept at the higher level, which are also trying to remove from the map.

In the case of the pending_render_frame_host_ cleanup code in RenderViewDeleted, it seems backwards and hence impossible. If RFH's own a refcount on their RenderViewHost, then the deletion of a RVH should never occur with pending_render_frame_host_ as an unaware bystander.

As a result, the solution here seems to be to just get rid of RFHM::RenderViewDeleted.

BUG=364974,23411

Review URL: https://codereview.chromium.org/269573003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@268957 0039d316-1c4b-4281-b951-d872f2087c98