A speculative fix for a MediaValues related crash
According to the stack at crbug.com/367210 it is believed that the crash is a result of null RenderView,
since a plugin is running a script that adds a new HTMLImageElement during the document teardown.
This CL adds a null check for such a case, as well as a fixme, so that null check can be removed once the underlying issue is resolved.

BUG=367210, 371084

Review URL: https://codereview.chromium.org/271613004

git-svn-id: svn://svn.chromium.org/blink/trunk@173563 bbb929c8-8fbe-4397-9dbb-9b2b20218538