Oilpan: Prepare to support garbage-collected Node in WebNode hierarchy.
That is to say, WebNode::m_private should be based on Persistent<Node>.
Because all of WebNode subclasses shares it, we need to change the
classes so that they can be convertible to RawPtr instead of PassRefPtr.

We need some #if ENABLE(OILPAN) because some WebCore functions
still return PassRefPtrs.


BUG=357163

Review URL: https://codereview.chromium.org/270573008

git-svn-id: svn://svn.chromium.org/blink/trunk@173622 bbb929c8-8fbe-4397-9dbb-9b2b20218538