Fix Telemetry to calculate bounding visible rect correctly.
Fixes the parts of bounding client rects which are to the left or to the
top of the visible area not to visible (in addition to the parts which
are to the right or to the bottom of the visible area).

Enhances the corresponding unittest to test these cases, too.

BUG=369479

Review URL: https://codereview.chromium.org/265933006

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@269116 0039d316-1c4b-4281-b951-d872f2087c98