Non-SFI NaCl: Avoid exposing validation caching key to Non-SFI nexes
This per-profile key is an extra input into the hashing that NaCl's
validation cache does.

Hiding this key means that even if we did accidentally expose
SetKnownToValidate() to Non-SFI nexes, the nexe wouldn't be able to
create a hole in the SFI NaCl sandbox.

We should hide this key to reduce the attacker's ability to generate
collisions in hash-based caching.

Also hide some other startup parameters that aren't relevant to Non-SFI
mode.

BUG=367263
TEST=browser_tests
R=jln@chromium.org

Review URL: https://codereview.chromium.org/275513002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@269270 0039d316-1c4b-4281-b951-d872f2087c98