Pepper: Clean up ProgressEvent logic.
This simplifies NexeLoadManager, which is getting pretty big. This also removes
an unused field from ProgressEvent, and adds a convenience function for posting
a task to dispatch a progress event from the main thread, which is often
necessary to ensure that progress events are posted to the DOM after other
state changes.

Review URL: https://codereview.chromium.org/270453004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@269362 0039d316-1c4b-4281-b951-d872f2087c98