Update the task manager with the debug stub port chosen by nacl.
On mac/linux the debug stub port is allocated chrome side and provided to
Native Client. On Windows (where socket handles cannot be passed between
processes), the debug port must be allocated by the nacl process (which can
only happen when the outer sandbox is disable).

Propagate the debug port selected by the nacl side to the task manager to make
it available to the user and debugging extensions.

Depends on this nacl side change:
https://codereview.chromium.org/206493005

BUG=328714
TEST=None
R=mseaborn@chromium.org,jschuh@chromium.org

Review URL: https://codereview.chromium.org/198083006

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@269499 0039d316-1c4b-4281-b951-d872f2087c98