Resource Timing: Use original Timing-Allow-Origin for cache validating
For 304 responses, we need to check the Timing-Allow-Origin header of the
original response. Even though we keep the original response during
revalidation, it isn't available at the time passesTimingAllowCheck() evaluates
the final response. This CL makes ResourceTimingInfo keep the original value
of the Timing-Allow-Origin header so that passesTimingAllowCheck() can
refer the original value.

BUG=368966

Review URL: https://codereview.chromium.org/271083002

git-svn-id: svn://svn.chromium.org/blink/trunk@173813 bbb929c8-8fbe-4397-9dbb-9b2b20218538