[Ozone-DRI] Move framebuffer initialization into the buffer
Surface/buffer lifetimes are no longer bound by the controller. In addition,
it is possible for multiple controllers to be linked against the same surface,
making it impossible for a controller to know if the framebuffers need to be
unregistered. Thus move framebuffer initialization into the buffer object.

NOTRY=true

Review URL: https://codereview.chromium.org/276753003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@269531 0039d316-1c4b-4281-b951-d872f2087c98