Domain Reliability: Don't clear "upload pending" bit until upload starts
Someone noticed that occasionally the uploader sent an empty upload. I
looked into the scheduler and found that it was clearing the "upload
pending" bit early, when the upload was scheduled. This meant that if a
beacon arrived between scheduling and starting the upload, the bit was
still set on its behalf, despite it being included in the upload. Once
the upload finished, the scheduler would see the bit still set and
schedule another.

This patch adds a couple of unittests for interesting interleavings
(beacons that arrive between scheduling and starting the upload, and
between starting and finishing the upload), and clears the upload
pending bit in the right place.

BUG=356791

Review URL: https://codereview.chromium.org/272773004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@269592 0039d316-1c4b-4281-b951-d872f2087c98