cc: fix pre-painting trigger for GPU rasterization
CUTOFF_ALLOW_NICE_TO_HAVE was only triggered for --force-gpu-rasterization.
This patch triggers it when the active or the pending tree uses GPU
rasterization. In case of a transition between CPU and GPU rasterization it
is preferrable to pre-paint less on the CPU rasterized tree (which may
checkerboard) than it is to pre-paint more on the GPU rasterized tree (which
may stall the browser compositor).

This patch depends on https://codereview.chromium.org/270823003/

R=enne@chromium.org, alokp@chromium.org
BUG=368884

Review URL: https://codereview.chromium.org/272033002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@269682 0039d316-1c4b-4281-b951-d872f2087c98