Don't use webmediaplayer to inform webview about fullscreen change
Currently we rely on webmediaplayer to inform webview about fullscreen change.
With https://codereview.chromium.org/277823002/, we no longer need to do so.
When entering fullscreen, the page will be resized to fullscreen, and that will call Will/DidEnterFullscreen for us.
When leaving the fullscreen, ToggleFullscreenForTab() will request the tab to exit fullscreen. That should trigger the Resize() call and send Will/DidExitFullscreen for us.

BUG=367346

Review URL: https://codereview.chromium.org/275163002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@270005 0039d316-1c4b-4281-b951-d872f2087c98