Chromoting iOS client
This is a sub-set of:
https://codereview.chromium.org/186733007/

It doesn't include the .xcodeproj files which were causing problems with "git
cl" due to being in the root .gitignore. They won't be necessary after gyp
integration.

This is just a code drop. Nothing is being built at the moment.

NOTRY=true
TBR=sergeyu@chromium.org
BUG=331356

Review URL: https://codereview.chromium.org/278863003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@270143 0039d316-1c4b-4281-b951-d872f2087c98