[Ozone-DRI] Move display management code to ScreenManager
ScreenManager will handle all the display configuration logic and ownership.
This CL moves the logic in the ScreenManager. Before moving its ownership to
NativeDisplayDelegateDri there needs to be a way to create/destory on
content_shell builds. For that a simpler NDD is required (part of a future CL).

BUG=333413
NOTRY=true

Review URL: https://codereview.chromium.org/275213005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@270149 0039d316-1c4b-4281-b951-d872f2087c98