Move first-party cookie URL logic to browser process.
Rather than have the renderer decide the value and inform the browser, the
browser decides and informs the render.

This removes one piece of navigation request redirect policy that needs to come
from the renderer. Redundant logic in Blink's DocumentLoader::willSendRequest to
be removed in a follow-up.

BUG=368813

Review URL: https://codereview.chromium.org/264613006

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@270463 0039d316-1c4b-4281-b951-d872f2087c98