IDL: last 2 method overloads!
This finishes method overloads!
(I'll do constructor overloads in a followup.)

This turns on method overloads for all methods,
and cleans up cruft and unused tests.

The remaining 2 cases are, unfortunately, invalid overloads:
we would need to look at *2* arguments to distinguish them.
(These should be *union* types, but we don't support union types for
 method arguments yet, so they're being hacked around with overloads.)

This can't be done without adding *major* hacks to the CG.

This is because the order of tests depends on the argument type,
so we can't get the order right if there are 2 argument types.
Thus we'd need to add custom code to the CG to handle these.

Instead, this adds custom bindings to handle the union types;
we're already doing this for other union types.

In the first Patch Set I've just put the auto-generated code from
the existing CG.
I'll clean this up so it becomes legible and reviewable.

R=haraken
BUG=293561

Review URL: https://codereview.chromium.org/283083002

git-svn-id: svn://svn.chromium.org/blink/trunk@174121 bbb929c8-8fbe-4397-9dbb-9b2b20218538