Fix a data race in BrowserGpuChannelHostFactory Initialization
It's not safe for a RefCounted object to PostTask in its constructor.
Refcounts start at zero, so if the task completes before the constructor
returns, the closure will see the refcount going 0->1->0 and think the
object can be deleted.

This change moves the PostTask in ..::EstablishRequest() to a separate
Create() static function so it can be called after assignment into a
scoped_refptr wrapper, which ensures the refcount is incremented before
posting the task.

BUG=116100

Review URL: https://codereview.chromium.org/286983002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@270938 0039d316-1c4b-4281-b951-d872f2087c98