Hide context menu on 'touchstart' event.
When a user do non-Tap (e.g. LongPress for drag, LongTap for context menu) operations on touch panel, 'mousedown' is not dispatched immediately.
In these cases, context menu should be hided because user doesn't focus on the context menu anymore.
This CL hide the menu on 'touchstart' in the same manner as 'mousedown'.

BUG=267211

Review URL: https://codereview.chromium.org/283363002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@270996 0039d316-1c4b-4281-b951-d872f2087c98