Pepper: Nexe downloading out of the trusted plugin.
This moves the logic for nexe downloading outside the trusted plugin. This is
part of a series of changes to remove FileDownloader and its associated logic.

I've tried to make this change small so it will be easy to review. A followup
change will clean up some of the accounting logic post-download to make this
path simpler and easier to follow.

BUG=370556
R=bbudge@chromium.org

Review URL: https://codereview.chromium.org/276423003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@271090 0039d316-1c4b-4281-b951-d872f2087c98