tcmalloc: Honor HEAPPROFILE envvar only in profiling=1 builds.
Several systems in chromium (about:tracing, some code in blink) use
tcmalloc's heap profiler, so keep most of it around.  Do disable support
for the HEAPPROFILE envvar, and for the exit-time dump to "Exiting" (unless in
profiling=1 builds), as that causes a static initializer.

BUG=94925
R=willchan@chromium.org

Review URL: https://codereview.chromium.org/295583002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@271415 0039d316-1c4b-4281-b951-d872f2087c98