Unmounts volumes before writing to a drive.
When writing to a drive with mounted volumes, Chrome OS will recreate the partition table if it is observe to be missing and the volumes are accessed after writing.  This only occurs when erasing just the partition table because enough of the volume information is still present.

We also expand the amount of data erased when destroying partitions to 8k to account for 4k block sizes.  This has the side effect of overwriting the LBA for 512-byte-block drives, which only increases the effect of destroying partitions.

BUG=284834
BUG=335390

Review URL: https://codereview.chromium.org/282853003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@271551 0039d316-1c4b-4281-b951-d872f2087c98