[Oilpan]: Move DescendantInvalidationSet to the oilpan heap using transition types.
This also moves RuleFeatureSet and SelectRuleFeatureSet as well since they transitively embed DescendantInvalidationSet.

In order to support off-heap tracing the PseudoTypeInvalidationSetMap hash map I had to change the OffHeapCollectionTraceTrait to use the HashMap's KeyType instead of the passed in Key. Otherwise I get a conversion error from the enum to unsigned int. This is in sync with how we trace with oilpan enabled. It is a bit of a hack since ideally the two types should be identical, but in this case the hashmap is instantiated with the UnsignedWithZeroKeyHashTraits<unsigned> which collides with the enums type in the key. Suggestions are welcome.
Btw. for symmetry I also changed the second call to use HashMap::MappedType instead of the passed in Value.

R=ager@chromium.org, erik.corry@gmail.com, haraken@chromium.org, oilpan-reviews@chromium.org, tkent@chromium.org, vegorov@chromium.org, zerny@chromium.org
BUG=341815

Review URL: https://codereview.chromium.org/274083002

git-svn-id: svn://svn.chromium.org/blink/trunk@174365 bbb929c8-8fbe-4397-9dbb-9b2b20218538