<webview>: Move NewWindow API to chrome
This CL moves the New Window API and navigation to chrome.

This CL removes a lot of unnecessary content APIs for BrowserPlugin now that permissions no longer live in content.

Up next, in a separate CL will be a huge cleanup that removes all WebContentsDelegate code out of BrowserPluginGuest and into WebViewGuest. This should significantly reduce the size of the BrowserPlugin content API.

BUG=364141, 330264
TBR=kenrb@chromium.org for trivial browser_plugin_messages.h change.

Review URL: https://codereview.chromium.org/272573005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@271703 0039d316-1c4b-4281-b951-d872f2087c98