Remove GetCurrentContext from V8Window::toV8
The code removed in this CL is now just for performance optimization (Previously it might have been not, but now it is just for performance optimization). Now that I refactored toV8Context(), we can always get a correct context by calling toV8Context(). Thus the removed code is not needed for correctness.

Another interesting fact is that the removed code is not contributing to performance optimization at all. This is because:

- In a case where we can get the context from the GetCurrentContext(), it means that the context is already initialized.
- Then we can simply call toV8Context() to get the context, because it is guaranteed that the toV8Context() doesn't hit the heavy operation of context initialization.

In conclusion, the removed code is now no help. Thus this CL removes it.

BUG=357144

Review URL: https://codereview.chromium.org/293003007

git-svn-id: svn://svn.chromium.org/blink/trunk@174499 bbb929c8-8fbe-4397-9dbb-9b2b20218538