Implement actions for desktop automation.
This cl adds the plumbing to support desktop automation actions and implements the doDefault and focus actions for underlying views::View.

Actions flow from the client extension, through automation internal's performAction binding, to the browser process. There, it makes its way to either a rwh or an AutomationManagerAsh; an AutomationActionTarget class assists in routing the action appropriately.

For the desktop tree, the action routes to the AutomationManagerAsh singleton and crosses over out of the extension system into ash. The manager forwards the action to the backing AXTree(SourceAsh) which then allows the underlying Aura/Ash view type to handle the action.

Currently, support has been implemented for doDefault and focus actions on views::View. Other actions for Aura/Ash views will be implemented in subsequent cl's as appropriate for those views.

Additionally, enforcement of the 'interact' attribute has been added for tabs; this is done extension/js side and slightly changes the action portion of the API to take an optional callback.

BUG=309681
TEST=unit_tests --gtest_filter=AXTreeSourceAsh.* and browser_tests --gtest_filter=AutomationApiTest.DesktopActions

Review URL: https://codereview.chromium.org/292513005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@272128 0039d316-1c4b-4281-b951-d872f2087c98