Remove override of View::HitTestRect() from BookmarkBarView
The override of View::HitTestRect() in BookmarkBarView checks
for the state of the omnibox popup to determine whether or not the
bookmark bar can accept events. Since this is unrelated to location,
it is better to move this logic into an override of
EventTarget::CanAcceptEvent() instead.

BUG=374340
TEST=none

Review URL: https://codereview.chromium.org/286313003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@272133 0039d316-1c4b-4281-b951-d872f2087c98