Change MediaStreamVideoSource to output different resolutions do different tracks depending on the track constraints.
This cl introduce a new class VideoTrackAdapter. The adapter is responsible for forwarding frames on the IO-thread from a MediaStreamVideoSource to a the tracks connected to the source. It has the ability to wrap media::VideoFrames with new visible_rect and natural_size to match max width and height resolutions as well as min and max aspect ratio per track. 
Note that this does not yet work for textures.

It also changes the WebrtcVideoCaptureAdapter to use libyuv::Scale instead of libyuv::I420Copy. This is needed to allow the video frames sent on a PC to be both cropped and scaled.

BUG= 346616

Review URL: https://codereview.chromium.org/246433006

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@272214 0039d316-1c4b-4281-b951-d872f2087c98