Implement sync in the NonBlockingTypeProcessor
Introduces the NonBlockingTypeProcessor's sync logic.  When combined
with the NonBlockingTypeProcessorCore's sync logic (which will be
introduced in a follow-up commit), this will be an alternative to the
existing sync engine implemented with DirectoryUpdateHandler and
DirectoryCommitContributor.

Adds non_blocking_sync_common.h, which defines structs to be used to
pass messages between the processor and processor core.

Adds DataTypeState as a parameter to the processor to processor core
connection methods.  Eventually this will be used to initialize the
processor core with state that the processor loaded from disk.

Adds a lot of unit tests and unit test framework intrastructure.  The
NonBlockingTypeProcessor and NonBlockingTypeProcessorCore's
communications with each other will be very racy.  These tests are
intended to help manage the complexity this will cause by allowing us to
test all the possible race conditions individually.

BUG=351005

Review URL: https://codereview.chromium.org/280983002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@272329 0039d316-1c4b-4281-b951-d872f2087c98