[FastTextAutosizer] Use firstChild for performance
This patch fixes a performance bug where nextInPreOrder was used when
firstChild was sufficient. In particular, in the cases when we have
a RenderBlock, firstChild is significantly faster than the more
generic slowFirstChild which is called by nextInPreOrder.

large-table-with-collapsed-borders-and-colspans.html performance test:
Without patch: 8008.3ms :(
With patch: 606.8ms :)

BUG=376190

Review URL: https://codereview.chromium.org/299083002

git-svn-id: svn://svn.chromium.org/blink/trunk@174625 bbb929c8-8fbe-4397-9dbb-9b2b20218538