Fix bad cast in TextAutosizer.cpp after Blink r174275
This patch fixes an assert in the old text autosizer that was
introduced in [1]. The code previously had the following line:
... !isIndependentDescendant(toRenderBlock(child)))
which makes it seem like toRenderBlock(child) is safe. Unfortunately
that is not always the case and this patch reverts to the previous
behavior.

This was not caught in tests because we've switched to the fast text
autosizer and the old Text Autosizer is untested (and will be removed
once M36 is out the door).

[1] https://src.chromium.org/viewvc/blink?view=rev&revision=174275

TEST=manual, load any autosized page with --disable-fast-text-autosizing

Review URL: https://codereview.chromium.org/296183005

git-svn-id: svn://svn.chromium.org/blink/trunk@174652 bbb929c8-8fbe-4397-9dbb-9b2b20218538