cc: Compositor without MessageLoopProxy
Support using single-threaded compositor with a single
DelegatedRendererLayer on a thread that does not have a MessageLoop.

First, Allow BlockingTaskRunner to be used on a thread that does not
have a MessageLoopProxy. For these threads, all PostTask calls must
be wrapped in CapturePostTasks. Use PlatformThreadId rather than
MessageLoopProxy::BelongsToCurrent thread to check for current.

Then fix DCHECK failures in Proxy to allow MessageLoopProxy to be
NULL.

Add tests to verify these code paths are indeed working.

BUG=344087

Review URL: https://codereview.chromium.org/292493006

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@272473 0039d316-1c4b-4281-b951-d872f2087c98