sync: Fix crash on backend initialize failure
Fixes a crash in ungraceful sync backend shutdown.  The problem was that
the sync backend host core might try to unregister itself as a listener
of some information generated in the ModelTypeRegistry, but the
ModelTypeRegistry did not exist if backend initialization failed.

The fix used here is to track whether or not we were registered for
these notifications in the first place, and only unregister if we
had previously registered successfully.  In the failure shutdown
case, the flag is unset, so we will not attempt to unregister.

BUG=376863

Review URL: https://codereview.chromium.org/294023017

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@272697 0039d316-1c4b-4281-b951-d872f2087c98