More reliable keep-alive video packets.
Previously the host was not sending empty keep alive messages when
video stream is paused or when it's blocked on the encoder, this
triggers reconnect too often.
Refactored keep-alive logic in VideoScheduler so now it always
sends keep-alive messages when there is no other activity on the
stream.

BUG=376528,375568

Review URL: https://codereview.chromium.org/292373002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@272790 0039d316-1c4b-4281-b951-d872f2087c98