[New Multicolumn] Cleanup: Remove redundant members.
RenderMultiColumnSet doesn't need to store column width and
height. These are readily available from the flow thread.

And the flow thread doesn't need an extra member to store column
width, since it's the same as the logical width. Calculate it on
the fly when we need to determine whether the flow thread will
get a new width, and when actually updating the logical
width. It's not expensive to calculate, and this is not a hot
path. Duplicated storage can only cause confusion and do harm,
anyway.

Also use the CSS 2.1 terms "used" and "actual" when we need to
disambiguate, rather than "computed" and
nothing. Remove "computed" from some members where it didn't add
any value. Those were not CSS 2.1 "computed" values, so I found
it somewhat confusing and misleading.

No need for columnWidth() and columnHeight() in
RenderMultiColumnSet.  Use pageLogicalWidth() and
pageLogicalHeight() instead.

No need for columnWidth() in RenderMultiColumnFlowThread either.

Review URL: https://codereview.chromium.org/299213002

git-svn-id: svn://svn.chromium.org/blink/trunk@174870 bbb929c8-8fbe-4397-9dbb-9b2b20218538