Remove SVGSymbolElement::selfHasRelativeLengths
Since <symbol> is cloned into the shadow tree I see no reason why it should have this method. But for this to work <use> has to delegate its selfHasRelativeLengths to the shadow tree element instead of the corresponding element, so change that behavior too.

Review URL: https://codereview.chromium.org/294783002

git-svn-id: svn://svn.chromium.org/blink/trunk@174881 bbb929c8-8fbe-4397-9dbb-9b2b20218538