Change a flaky expectation in ChromeRenderWidgetHostViewMacHistorySwiperTest.
The flakiness of this expectation in this test reveals that the current swipe
gesture recognition is broken. Depending on the timing of the IPCs between
Chrome and the renderer, a non-trivial portion of vertical motion that should
scroll content is being lost.

The previous expectation was that no more than 15% of the vertical motion is
lost. I've updated the expectation to check that there's been more than zero
vertical motion.

BUG=378158

Review URL: https://codereview.chromium.org/303853002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@273644 0039d316-1c4b-4281-b951-d872f2087c98