Node.textContent setter should treat undefined the same way as null
Node.textContent setter should treat undefined the same way as null.
Previously, undefined would get stringified to "undefined" in the
setter, instead of being treated as "". This is inconsistent with
what happens when assigning null and with the DOM specification:
http://dom.spec.whatwg.org/#dom-node-textcontent

Since the textContent attribute is nullable, both null and undefined
input should be treated as null as per the Web IDL specification:
http://heycam.github.io/webidl/#es-nullable-type

The new behavior is consistent with Firefox 29. However, IE11 returns
"undefined" in this case.

R=arv@chromium.org, tkent@chromium.org
BUG=378454
TEST=fast/dom/Node/textContent-set-undefined.html

Review URL: https://codereview.chromium.org/303163002

git-svn-id: svn://svn.chromium.org/blink/trunk@175103 bbb929c8-8fbe-4397-9dbb-9b2b20218538