cc: Add HasPendingNotification to delayed unique notifier.
It would be nice to know if a notification is pending, so this patch
adds this ability. This would be used in RenewTreePriority to detect
whether we should be in smoothness mode or same priority mode. That is,
if we have an expiration from smoothness pending, then we should still
be in smoothness mode.

R=reveman

Review URL: https://codereview.chromium.org/302003004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@273780 0039d316-1c4b-4281-b951-d872f2087c98