Simplify animation/transition parsing slightly
This patch simplifies parsing CSS animations and transitions slightly by
making us only return CSSValueLists from the parser, including when we
only have a single value in the list. This was probably done differently
in the past (with single values not being wrapped in a CSSValueList) as
premature optimisation, which adds needless complexity to the code. Note
this changes the foo.style.getPropertyCSSValue output slightly.

I've also made BisonCSSParser::parseAnimationTimingFunctionValue return
a single (non-valuelist-wrapped) CSSValue which isn't initial or inherit
to better match the intention of the callers.

R=alancutter,dstockwell

Review URL: https://codereview.chromium.org/293113007

git-svn-id: svn://svn.chromium.org/blink/trunk@175153 bbb929c8-8fbe-4397-9dbb-9b2b20218538