gpu testing: Handle non-integral device pixel ratio.
Device pixel ratio is used to scale test rect and sampling locations.
It can be non-integral on some devices. For example it is 1.325 on
Nexus 7. The scaled quantities - dimensions of screenshot and sampled
pixel locations - need to be converted to integers before they can
be used.

BUG=368495

Review URL: https://codereview.chromium.org/303113004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@273951 0039d316-1c4b-4281-b951-d872f2087c98