Introduce View::CanProcessEventsWithinSubtree()
Introduce the virtual function 
View::CanProcessEventsWithinSubtree(), which returns 
true if the view and any of its descendants are 
permitted to be the target of an event. This will be 
called as part of event targeting and replaces the 
overrides of View::HitTestRect() in cases where the 
rect parameter is unused. One such replacement
(for ImageView) is made in this CL.

BUG=374325, 378530
TEST=ViewTest.CanProcessEventsWithinSubtree, ViewTargeterTest.CanProcessEventsWithinSubtree

Review URL: https://codereview.chromium.org/297963006

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@274011 0039d316-1c4b-4281-b951-d872f2087c98