aura: Reland r266681 to debug WindowObserver crashes.
Add code to make the aura::WindowObserver interface crash
when it's destroyed without having unregistered its interest
in all aura::Windows that it was observing. This will
hopefully help track down the source of crashes that are
seen when closing video windows on Google Play.

https://codereview.chromium.org/303163003/ hopefully fixes
the test crashes that resulted in this change being reverted
earlier.

BUG=365364

Review URL: https://codereview.chromium.org/309573004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@274086 0039d316-1c4b-4281-b951-d872f2087c98