[Telemetry] Retire JSGameBench.
As of May 14, 2014, this benchmark is no longer maintained.
https://github.com/facebookarchive/jsgamebench/commit/3938f9893d8713c47032d8d98bac41c32f4545ec

I think it is time we thank it for all the fish (e.g. http://crbug.com/161828)
and turn our eyes towards new benchmarks.

My motivation for pursuing this is that it has a long cycle time and is flaky in many places.

BUG=
NOTRY=True

Review URL: https://codereview.chromium.org/307263002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@274272 0039d316-1c4b-4281-b951-d872f2087c98