New animation for the origin chip URL showing/hiding.
When the chip is hidden, we instantly show the entire omnibox text, aligned so
that the hostname is in the same place it was in the chip, and clipped to just
have the hostname visible.  Then we animate moving the alignment over to the
normal omnibox text position (i.e. just after the location icon) while
expanding the clip to gradually reveal the whole text.  We also crossfade the
selection colors from the chip colors to the standard omnibox ones.

This also:
* Animates correctly when the hostname is not at the beginning of the origin
  chip (i.e. for EV certs)
* Properly reverses the animation if the user changes state in the middle (hard
  to see with quick animations, but looks nicer if we slow them down or with
  very large animation distances)
* Fixes dropping characters during animation (because the real omnibox is
  visible and focused, and can respond correctly to keypresses)
* Fixes animations improperly running across tab switches
* Cleans up some unused origin chip code

Also tested in RTL mode.

BUG=363681,365818,374251
TEST=With the origin chip on, showing and hiding the URL should animate much more quickly and cleanly

Review URL: https://codereview.chromium.org/302453002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@274305 0039d316-1c4b-4281-b951-d872f2087c98