Resubmit: Refactor external_install_ui
Original CL: https://codereview.chromium.org/300853008/
Original Description:
Refactors ExternalInstallUI to have determinate lifetimes.
This is likely the first of multiple patches targeting this area of code; the
next one will concentrate on pulling the related logic out of ExtensionService.

----------------------------------------------------

The original patch contained a cycle in that when an external install warning
was clicked, it would accept or remove the extension, and then remove the error.
Since the manager listened for the extension to be loaded or unloaded, this
resulted in a double-removal.  Fix is straightforward.

We also didn't have tests for actually interacting with the dialog, so two have
been added.

TBR=yoz@chromium.org
BUG=378042

Review URL: https://codereview.chromium.org/309643007

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@274325 0039d316-1c4b-4281-b951-d872f2087c98