Add display_unittests target
As a first step to getting a separate target for the display unittests, this CL is adding a simple target with one unittest. All display_unittests currently running under ui_unittests will be migrated under this target when display_unittests is enabled on the main waterfall.

BUG=333413

Review URL: https://codereview.chromium.org/304183010

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@274332 0039d316-1c4b-4281-b951-d872f2087c98