Don't add extra 8 byte to DOM wrappers in non-oilpan builds
Imagine a DOM object that is RefCountedWillBeGarbageCollected. In oilpan builds, we need extra 8 byte on its DOM wrapper to store a persistent handle to the DOM object. On the other hand, in non-oilpan builds, we don't need the 8 byte. This CL removes the 8 byte from the DOM wrapper.

Review URL: https://codereview.chromium.org/314603003

git-svn-id: svn://svn.chromium.org/blink/trunk@175361 bbb929c8-8fbe-4397-9dbb-9b2b20218538