sync: Populate entity counts in about:sync tab
Populates the count of entities per type in about:sync.

Includes C++ code to count and emit the set of deleted and non-deleted
counts, though for space reasons only the non-deleted count is displayed
on the about:sync page.

This calculation is somewhat expensive.  It could be made cheaper by
having an index in the Directory to keep track of which entities belong
to which type, but that doesn't exist at the moment.  For now, we just
avoid calculating these counts unless the about:sync page is open.

BUG=349301

Review URL: https://codereview.chromium.org/302283007

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@274563 0039d316-1c4b-4281-b951-d872f2087c98