Fix renderer process host backgrounding test on Windows.
Now that process backgrounding is controlled from within the renderer
process on Windows, we need to wait for the process to handle the IPC
message before the background state changes.

This fix changes the test to wait for the result of a script from
the renderer process before checking background status.  Doing so
ensures the renderer will process the background IPC before the
check is performed.

BUG=362294
TEST=Ran unittest on Windows.

Review URL: https://codereview.chromium.org/305373004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@274778 0039d316-1c4b-4281-b951-d872f2087c98