Gate disabling nacl code on proper gyp flags, etc.
Background: I am currently trying to test NaCl in clank by setting disable_nacl=0 and enable_plugins=1 in gyp.  Some of the code required to get this to work is guarded under overly-specific OS==android and similar constructs.  This first CL enables a coherent build with the flags set to gyp, but does not yet turn the functionality on completely.  A sampling of issues fixed:

1) std:: prefix is more strictly observed on android
2) define NACL_LINUX, etc., is needed for nacl code
3) properly guard flag-setting code

BUG= https://code.google.com/p/nativeclient/issues/detail?id=3032
R=bradnelson@google.com, darin@chromium.org

Review URL: https://codereview.chromium.org/299703016

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@274839 0039d316-1c4b-4281-b951-d872f2087c98