Return early in SelectorChecker::checkOne() if selector.isAttributeSelector() is true
Return early in SelectorChecker::checkOne() if selector.isAttributeSelector()
is true and the selector matched. Previously, we were only returning early if
the selector did not match and we were falling through if it did. This means
that we ended up doing the 2 if() checks below in the method unnecessarily:
- selector.m_match == CSSSelector::PseudoClass
- selector.m_match == CSSSelector::PseudoElement

Review URL: https://codereview.chromium.org/319513005

git-svn-id: svn://svn.chromium.org/blink/trunk@175518 bbb929c8-8fbe-4397-9dbb-9b2b20218538