Make cygprofile order functions by process and thread ID.
This is used to make the order of functions not depend on thread-scheduling
which can be greatly impacted when profiling is done with cygprofile.

As a result each thread has its own contiguous segment of code (ordered by
timestamp) and processes also have their code isolated (i.e. not interleaved).

BUG=372323
R=pasko@chromium.org

Review URL: https://codereview.chromium.org/281093002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@275076 0039d316-1c4b-4281-b951-d872f2087c98