Do not try to verify focus placement in ClearBrowserDataBrowserTest.
The test was flaky because of a more general bug in the WebUI framework (crbug.com/380737). This latter bug exhibits itself when, during opening a WebUI overlay, the BrowserWindow loses focus -- which happens a lot on build bots running multiple browser tests in parallel.

BUG=379574

Review URL: https://codereview.chromium.org/313843002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@275128 0039d316-1c4b-4281-b951-d872f2087c98