Make the Bookmark Manager aware of managed bookmarks.
The managed bookmarks are a new node at the root that are configured via
policy and can't be modified directly by the user. This change makes the
Bookmark Manager aware of those nodes by disabling the mutating options
when working with nodes from the managed tree.

BUG=49598
R=arv@chromium.org

Review URL: https://codereview.chromium.org/308003013

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@275388 0039d316-1c4b-4281-b951-d872f2087c98