Supports fake key events for setComposition/commitText by on-screen keyboards.
IMF now assumes the IME composing happens with a physical key being pressed.
That assumption is not true for on-screen keyboard scenario.
For on-screen keyboard, when user touch a key button on the screen, the JS
code will call IME API like chrome.input.ime.setComposition(). IMF then should
generate events as:
  keydown(229), compositionstart, compositionupdate, input, keyup.
The events compositionstart, compositionupdate and input are generated by
RenderWidgetHostViewAura::SetCompositionText(). So we need to send fake key
events to generate keydown(229) and keyup.

The other part of this change is to make InputMethodEventFilter not depend on
physical key events. Originally, InputMethodEventFilter relies on key events
to find the root window and then get the EventProcessor. It will break
on-screen scenario, because without a physical key being pressed,
target_dispatcher_ will be NULL.
This cl is to let InputMethodEventFilter to always get the EventProcessor from
input_method_->GetTextInputClient()->GetAttachedWindow()->GetRootWindow()...

BUG=376183
TEST=Verified on Pixel device.

Review URL: https://codereview.chromium.org/298893003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@275422 0039d316-1c4b-4281-b951-d872f2087c98