ozone: Implement PlatformObject<T> for platform-specific objects
This paves the way for having more than one platform-specific object,
which is necessary for the componentization of ozone.

The plumbing to build objects that depend on the platform is shared
using the PlatformObject<T> template.

Code that wants to build such an object need simply do

  scoped_ptr<MyPlatformSpecificType> my_platform_object =
      PlatformObject<MyPlatformSpecificType>::Create();

We have to generate a couple of source files to make this work:

  - a header file with an enumeration of platforms
  - a (per-component) source file with a static const list of
    constructors for each platform type

This converts the previous object (OzonePlatform) instantiation
to use PlatformObject<>. This should be functionally identical.

BUG=380884
TEST=build aura_builder with "use_ozone=1 chromeos=1", ran egltest
NOTRY=true

Review URL: https://codereview.chromium.org/313963007

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@275460 0039d316-1c4b-4281-b951-d872f2087c98