Don't assume output dir + "../.." == src/ in ui/gl/generate_bindings.py
With gn, it'll probably be common for folks to generate their out directory
in places other than out/{Debug|Release}. This is technically possible (and
busted) in gyp as well. This avoids the relative path dependency from the
script and simplifies it some more by adding the src/ui/gl and src/gpu
directories in the build files (where they can be relativized properly)
instead of trying to do so in the script.

R=brettw@chromium.org

Review URL: https://codereview.chromium.org/317813002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@275520 0039d316-1c4b-4281-b951-d872f2087c98