Files.app: Explicitly add o+r permission to mounted zip archives under Downloads.
After the change r274593, downloaded files started to have permission 0640
rather than 0644. This CL explicitly adds read-permission before mounting,
provided the archive file is located under the local Downloads folder.

The modification is not applied to zip files on external storage because (1) the
silent change of permission may not be what the user want, (2) it should be rare
that download destination is set to removable devices, and (3) according to UMA
stat 99% of mounted filesystems are those we mount ignoring permissions anyhow
(FAT, NTFS, ISO9660, ...).

BUG=381144

Review URL: https://codereview.chromium.org/323783002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@275777 0039d316-1c4b-4281-b951-d872f2087c98