Exclude test types that don't upload from the flakiness dashboard UI.
Moves some logic from the JS code in to the code that generates
builders.jsonp for steps with test in the name that don't actually
run tests. Also expands this list a bit.

Also, creates a new list for test steps that do run tests, but
don't upload to the dashboard. In a future patch, I'll expand this
list and give it a nice UI in the flakiness dashboard that points
people to a way to get the test step to upload results.

Review URL: https://codereview.chromium.org/309783002

git-svn-id: svn://svn.chromium.org/blink/trunk@175829 bbb929c8-8fbe-4397-9dbb-9b2b20218538