Improve error messages and reporting in GN
Update help for cflags to reference new escaping behavior. We now escape spaces. The main reason it didn't do this before was pkg-config integration, but we now have specialized scripts to wrap pkgconfig.

Improve error message when a label doesn't match. This is a common one to run into.

Add the origin when a load fails. Previously if you mistyped a label in a dependency you would get a message about the file not loading but no location, which made it very hard to track down. This plumbs through the original location to the actual load.

R=scottmg@chromium.org

Review URL: https://codereview.chromium.org/318383003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@276064 0039d316-1c4b-4281-b951-d872f2087c98