Delete RenderLayerCompositor::layerWillBeRemoved
This function doesn't do anything. The repaint it generates is redundant and
marking the RenderLayerCompositor as needing a rebuild is redundant because
RenderLayerStackingNode will already mark the same dirty bit. It's more
sensible for RenderLayerStackingNode to be responsible for setting this dirty
bit because the reason it needs to be set is that the stacking tree has changed.

Review URL: https://codereview.chromium.org/325063003

git-svn-id: svn://svn.chromium.org/blink/trunk@175941 bbb929c8-8fbe-4397-9dbb-9b2b20218538