[fsp] Remove remaining space indicator for provided volumes.
Fetching remaining space is not implemented yet, hence we should not show the
indicator in Files app for provided file systems.

Note, that since provided file systems are read only, there is no actual need
for this feature.

This is temporary. Once the R/W support is added, we should bring back the
indicator.

TEST=Tested manually with experimental providing extensions.
BUG=375020

Review URL: https://codereview.chromium.org/324253003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@276433 0039d316-1c4b-4281-b951-d872f2087c98