Add date_synced to PasswordForm.
We need it for the "Remote unsync" project. If user accidentally signed-in in the library then he should be able to clean his passwords. date_synced designates the time the password was downloaded from the sync server. All passwords with (date_synced >= sign-in time) are to be deleted.

BUG=362679

Review URL: https://codereview.chromium.org/316243002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@276445 0039d316-1c4b-4281-b951-d872f2087c98