Files.app: Specify the size of file type icons explicitly.
For some devices, we does not include 2x theme assets.  But the devices can be
connected HiDPI displays. In that case, 2x assets are routed to fallback 1x
assets. To show the fallbacks in correct size, we need to specify the icon size
explicitly.

BUG=382971
TEST=remove chrome_200_percent.pak and launch chrome with '--force-device-scale-factor=2'
R=fukino@chromium.org

Review URL: https://codereview.chromium.org/335563002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@276569 0039d316-1c4b-4281-b951-d872f2087c98