Cast: Updated congestion control
This tries to update the bandwidth used to what is available.
The algorithm is essentially that we estimate how much bandwidth we've been sending recently.
Then we check how full our buffers are. We set a target for having our buffers be 90% empty.
If the buffer is more than 90% empty, we use more bandwidth than the resently sent, if
the buffer is less than 90% empty, we use less.

This change should be relatively safe since we normally set our min and max bandwidth to a
very limited range, and this change respects those limits.

Review URL: https://codereview.chromium.org/326783002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@276611 0039d316-1c4b-4281-b951-d872f2087c98