cc: Refactor how picture layers are exposed to the tile manager.
This moves the vector of registered picture layers from
the tile manager to LTHI. Allows layers to be registered
in ctor and unregistered in dtor instead of having to
unregister them from the tile manager and then re-register
them as needed.

The picture layer vector is exposed to the tile manager
using a client function and this vector will include all
existing PictureLayerImpl instances. Whether it makes
sense for the tile manager to use a layer or not depends
on the return value of
PictureLayerImpl::HasValidTilePriorities().

BUG=

Review URL: https://codereview.chromium.org/322443002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@276665 0039d316-1c4b-4281-b951-d872f2087c98