aura: Allow held touch-events to contribute to gesture events.
When a touch-move event is 'held', i.e. it is not dispatched, the event does
produce any gesture events. However, once this 'held' event is dispatched
(either because a touch-release event was encountered, or because
ReleasePointerMoves() was called on the dispatcher), it should be used to
generate the appropriate gesture events.

BUG=none
R=sky@chromium.org

Review URL: https://codereview.chromium.org/332703003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@276781 0039d316-1c4b-4281-b951-d872f2087c98