Issue repaints when offsetFromRenderer changes for any squashed layer.
This is necessary because the coordinate system of the squashing layer may change
w.r.t. a squashed layer, even if the squashed layer's layout position did not change.
For example, a new layer may be added to a squashing layer that expands its bouds to the top
or left.

BUG=383779

Review URL: https://codereview.chromium.org/331693004

git-svn-id: svn://svn.chromium.org/blink/trunk@176087 bbb929c8-8fbe-4397-9dbb-9b2b20218538