Fix BrowserMediaPlayerManager::RequestFullscreen().
Previously we send an IPC message to the renderer side when RequestFullscreen()
is called. This IPC message sending code was removed in r273600. Now user has to
click the fullscreen button to start playback in L1 mode, which isn't intuitive.
This CL fixes this issue so that when user click play, the video automatically
enters fullscreen mode and starts playback. This was the behavior before we had
the --disable-gesture-requirement-for-media-fullscreen flag.

Note that MediaSourceDelegate guarantees that we only call Start() after the
SetCdm() is called if any stream is encrytped.

BUG=383080
TEST=L1 video starts to play in fullscreen mode automatically.

Review URL: https://codereview.chromium.org/327323010

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@276947 0039d316-1c4b-4281-b951-d872f2087c98