Web MIDI: MidiManager crashes if a session is ended while initializing
If a MidiManagerClient was destructed while an asynchronous MidiManager
initialization performs, the browser process crashes when the
initilization is finished. On EndSession(), MidiManager should remove
the client even from |pending_clients_|.

Also, this change fixes a problem that CompleteStartSession() may not be
invoked on a MidiManagerClient if its |client_id| is confclicting with
another MidiManagerClient.

BUG=382522
TEST=media_unittests --gtest_filter='Midi*'

Review URL: https://codereview.chromium.org/323323002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@276987 0039d316-1c4b-4281-b951-d872f2087c98