Fix rebaseline-o-matic logging
readline blocks, so reading stdout, then stderr doesn't make any
sense because we'll block on each on. So if we print only to one
of them, we'll just wait until the process completes.

Also, simplify the readline loop in RebaselineOMatic to just
use readline instead of poll. That way it avoids missing any
output between the last poll and when the process finishes.

Review URL: https://codereview.chromium.org/329873005

git-svn-id: svn://svn.chromium.org/blink/trunk@176146 bbb929c8-8fbe-4397-9dbb-9b2b20218538