Simplify Writing-Mode Related Calculation of Margins
Remove the very complicated logic from computeLogicalHeight() and
computeLogicalWidth() whose sole purpose was to use the writing mode of
the child vs the containing block when appropriate. The complexity was
unnecessary and the same result can be achieved by ensuring we use the
child's writing mode when computing the size of the box and the parent's
writing mode when computing the margins for position (e.g. for margin
collapsing).

Review URL: https://codereview.chromium.org/298563002

git-svn-id: svn://svn.chromium.org/blink/trunk@176154 bbb929c8-8fbe-4397-9dbb-9b2b20218538