Transition networkState back to NETWORK_IDLE on fetch if preload=none
When a media element is marked with the preload hint 'none', a load is not
started until data is explicitly requested. The progress timer is however
still started, which means that a 'stalled' event will be fired after ~3s.
This is confusing, since loading never actually started in the first place.
Fix this by implementing (the "optional") step 3 from the 'resource fetch
algorithm'.

BUG=382505
BUG=284413

Review URL: https://codereview.chromium.org/327553002

git-svn-id: svn://svn.chromium.org/blink/trunk@176199 bbb929c8-8fbe-4397-9dbb-9b2b20218538