Add presubmit script to statically verify ChromeVox using the Closure compiler.
In addition to enforcing that ChromeVox compiles when uploading a change
(on Linux only), this provides a script that can be used to manually
compile some or all of the scripts, optionally based on a set
of modified files.

This adds the necessary externs for compiling ChromeVox.

BUG=371692
R=dtseng@chromium.org

Review URL: https://codereview.chromium.org/323273002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@277585 0039d316-1c4b-4281-b951-d872f2087c98