Always get latest metadata to check offline availability on Drive.
After a file is pinned offline and synced, the file's cached metadata about offline availability is not necessarily updated.
Because of this outdated metadata cache, the file couldn't be opened immediately after going offline.
This CL invalidates the cache to get latest metadata when a user open a file on Drive in offline environment .

BUG=383377
TEST=go through steps in 383377
R=yoshiki@chromium.org

Review URL: https://codereview.chromium.org/332163002

git-svn-id: svn://svn.chromium.org/chrome/branches/1985/src@277678 0039d316-1c4b-4281-b951-d872f2087c98