Make LoadAndLaunchPlatformApp harder to make flaky tests with.
LoadAndLaunchPlatformApp will send a launched event to the app it is
launching, which will often end up with windows being opened. If the
test does not wait for the window to stop opening, the test will be
flaky.

This change updates LoadAndLaunchPlatformApp to wait for a signal to
come back from the app after it is launched before returning.

TBR=sky
BUG=383260

Review URL: https://codereview.chromium.org/324333002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@277688 0039d316-1c4b-4281-b951-d872f2087c98