MSE: Start removing Blink-side support for LegacyFrameProcessor
This is the first of a three-sided change to remove LegacyFrameProcessor
from Blink and Chromium now that the new FrameProcessor is stabilized.
Once Chromium no longer implements the FrameProcessorChoice version of
WMSI::addSourceBuffer(), a third change will remove the default Blink
implemenetations of addSourceBuffer() and the FrameProcessorChoice enum.

R=acolwell@chromium.org,abarth@chromium.org
BUG=249422
TEST=No MSE layout test regression locally on Linux

Review URL: https://codereview.chromium.org/310733002

git-svn-id: svn://svn.chromium.org/blink/trunk@176379 bbb929c8-8fbe-4397-9dbb-9b2b20218538