Make the migrated ChromeVox work.
This sets the COMPILED flag in base.js to true when compressing the
javascript sources. ChromeVox makes assumptions about what content
script files to load (which it could get from the manifest) and also
does initialization differently depending on thsi flag.  Leaving cleanup
of that to later CLs if we want to do that in the interest of changing
as little as possible during the migration.

BUG=371692
R=dtseng@chromium.org

Review URL: https://codereview.chromium.org/333503002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@277956 0039d316-1c4b-4281-b951-d872f2087c98