Make iterator for Element's attributes more lightweight
Make iterator for Element's attributes more lightweight. Kill the
AttributeConstIterator class and make the iterator a simple pointer instead
(similarly to the Vector iterator). Also rename the AttributeIteratorAccessor
class to AttributeCollection as I think it looks better. This CL adds a typedef
so that the iterator is now used as AttributeCollection::const_iterator.

Also rename the attributesIterator() method to attributes() for simplicity.

R=esprehn@chromium.org

Review URL: https://codereview.chromium.org/337753005

git-svn-id: svn://svn.chromium.org/blink/trunk@176430 bbb929c8-8fbe-4397-9dbb-9b2b20218538