WorkspaceLayoutManager backdrop should fill entire region.
Changes to the visible state of the shelf causes the WorkspaceLayoutManager to change the size of its children. This includes changing the size of the backdrop. In order for the backdrop to cover the entire workspace it needs to be notified of changes to the available area. Such as when the shelf causes the inset regions to update.

TEST=WorkspaceLayoutManagerBackdropTest.ShelfVisibilityChangesBounds
BUG=383383

Review URL: https://codereview.chromium.org/337193002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@278106 0039d316-1c4b-4281-b951-d872f2087c98