Round in the range [-0.5, 0.5) when computing offsetFromRenderer and subpixelAccumulation for squashed layers.
Also remove the assert that the subpixel accumulation matches
the subpixel accumulation if the layer was separately
composited. This doesn't make sense, since the squashing
layer is positioned relative to the compositing ancestor
of m_owningLayer, not the squashing layer, which can have
different subpixel accumulations since subpixel
accumulation is relative to the compositing container of
the graphics layer.

BUG=369526

Review URL: https://codereview.chromium.org/343543005

git-svn-id: svn://svn.chromium.org/blink/trunk@176441 bbb929c8-8fbe-4397-9dbb-9b2b20218538