Map light-source oBB-relative coordinates to the user-space of the filter
The 'fePointLight' and 'feSpotLight' light-source elements for
'fe{Diffuse,Specular}Lighting' defines positions, and in the case of
the latter a 'look-at' point (implicit direction). When an SVG filter is
declared to use primitiveUnits=objectBoundingBox, these positions/points
needs to be mapped to the user-space of the filter.
This is achieved by adding a new method resolve3dPoint() to Filter, and
perform the coordinate space mapping when creating the various
LightSources.
Open-code SVGFELightElement::findLightSource() in it's sources to avoid
passing the Filter-instance around more than necessary.

BUG=176419

Review URL: https://codereview.chromium.org/337903003

git-svn-id: svn://svn.chromium.org/blink/trunk@176499 bbb929c8-8fbe-4397-9dbb-9b2b20218538