Revert of Reland Window Control Animations for TouchView. (https://codereview.chromium.org/316693002/)
Reason for revert:
In Immersice Fullscreen, when a mouse hovers near the top of the screen the header animates in. This is done by animating TopContainerView. However this view does not contain all of the title bar, and forces a paint to its canvas. With FrameCaptionButtonContainerView painting to a layer it no longer paints to the canvas of TopContainerView. Due to z-order of the views FrameCaptionButtonContainer exists below the TopContainerView as well as below the ClientView of the browser.

Reverting the use of layers in M37.
For M38 the use of layers can be rexamined, however changes to the header views, or how immersive mode renders, will be needed before relanding this.

Original issue's description:
> Reland Window Control Animations for TouchView
> 
> The original change contained a use-after-free bug. This was caused by window crossfade animation during
> maximizing, combined with input event processing. The crossfade would take the current layer and delete it. 
> However the next input event to the window tree would attempt to access the now deleted layer.

> Original Code Review: https://codereview.chromium.org/271913002/
> Revert Review: https://codereview.chromium.org/309973002/
> Revert "Revert of Animate window control changes in TouchView (https://codereview.chromium.org/271913002/)"
> This reverts commit 8863218a5e97c4914a5a03df59e9d5d17f53a2cb.
> 
> TBR=jamescook@chromium.org
> TEST=FrameCaptionButtonContainerViewTest.AnimationUpdatesLayerTree
> BUG=363717
> Committed: https://src.chromium.org/viewvc/chrome?view=rev&revision=276782

TBR=jamescook@chromium.org
BUG=384612, 363717

Review URL: https://codereview.chromium.org/343433009

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@278520 0039d316-1c4b-4281-b951-d872f2087c98