Increase "log too large" threshold in MetricsService.
Current value is at 50k, which is actually the average log
size on a lot of platforms. Increase it to 100k for a more
reasonable default that should cover the majority of cases
where logs are discarded due to this size limit according
to UMA data.

BUG=380282

Review URL: https://codereview.chromium.org/313613003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@278612 0039d316-1c4b-4281-b951-d872f2087c98