Increase the child process connection timeout for browser tests.
Normally a child process kills itself if it can't connect to the browser in 15 seconds. This is to prevent a child process from staying around if the browser crashed right after launching it but before connecting to it.

When browser tests are sharded, sometimes tests run slower because of resource starvation which is why we have longer timeouts. So increase the child process connection timeout to match this.

BUG=TODO
R=phajdan.jr@chromium.org

Review URL: https://codereview.chromium.org/343543002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@278812 0039d316-1c4b-4281-b951-d872f2087c98