Using a mock LevelDBTransaction for corruption tests.
Was previously flushing the actual LevelDB table on disk and then corrupting.
This approach is more loosely coupled with LevelDB disk format and lays the
groundwork for future tests (and refactoring) to not rely on an actual
corrupted db.

Also reduced the number of test objects written to the db to 5.

Review URL: https://codereview.chromium.org/334303002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@278838 0039d316-1c4b-4281-b951-d872f2087c98