Improve our detection of which line an object with culled lineboxes is on.
When figuring out which linebox an object belongs to we walk through its
previous siblings looking for the first object with a linebox and use that.
If the object is preceded by a sequence of objects that have had their lineboxes
culled (but can still act as line-break objects) then we can end up using
a linebox that has nothing to do with our object. So when we select this linebox
we need to check that we didn't pass a linebreak object on both it or the one after
when walking back through our object's previous siblings. If we did then
that means our object is not sitting on the linebox we originally detected and
we need to dirty the one after it as well.


BUG=372134

Review URL: https://codereview.chromium.org/338703002

git-svn-id: svn://svn.chromium.org/blink/trunk@176663 bbb929c8-8fbe-4397-9dbb-9b2b20218538