[fsp] [reland] Cleanup handling errors for operation requests.
This is a reland of https://codereview.chromium.org/335753004/, after reverting
because of breaking ASAN bots. Includes a trivial fix. The original description
is below.

This CL removes redundant internal functions which are identical, and replaces
them with one.

Moreover, the new general purpose error API function, passes now the entire
value to the request manager. This is because, very soon, we are going to
pass extra information with every error and success responses and we want to
have access to them in the RequestManager.

TBR=benwells@chromium.org, isherman@chromium.org, kinaba@chromium.org
TEST=unit_tests, browser_tests: *FileSystemProvider*
BUG=373165, 384201

Review URL: https://codereview.chromium.org/342893002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@279023 0039d316-1c4b-4281-b951-d872f2087c98