IDL parser: implement Stringifier production
Emit a node with class 'Stringifier' for all stringifier variants.
If the stringifier is an attribute, this node has a child node that
represents the attribute. If the stringifier is an operation (named
or unnamed) the node has a child node that represents the operation.

BUG=306606

Review URL: https://codereview.chromium.org/329163002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@279097 0039d316-1c4b-4281-b951-d872f2087c98