Skip occluded tiles during rasterization.
This causes the LayerRasterTileIterator to skip occluded tiles by
tracking the occlusion for each layer and then determining
whether each tile is occluded or not. The occlusion information
is stored on the tile. Another CL will be created to store the
occlusion information on the TilePriority instead so that the
same tile can have different occlusion on each tree.

This patch depends on https://codereview.chromium.org/246673005 or
similar changes that cause the raster iterator to be used.

BUG=178971

Review URL: https://codereview.chromium.org/316093002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@279195 0039d316-1c4b-4281-b951-d872f2087c98