Fix the do-not-repaint-if-scrolling-composited-layers layout tests.
With https://codereview.chromium.org/303253004/, Glenn fixed the issue
that had prevented these layout tests from succeeding: that has-
block-selection-gaps-bounds depended on a paint in order to be valid.

With his CL, we know if we have selection gaps based solely on style
information. The test expectation had to be modified slightly due to a
harmless error related to contiguous selection (the selection in the
test is, in fact, contiguous).

BUG=353827

Review URL: https://codereview.chromium.org/353483003

git-svn-id: svn://svn.chromium.org/blink/trunk@176882 bbb929c8-8fbe-4397-9dbb-9b2b20218538