Fix DCHECK crash in RtcVideoEncoder.
webrtc::VideoEncoder is synchronized by webrtc::VideoSender.
The methods are called on different threads but that is OK.
SetRates and SetChannelParameters are called on ProcessThread
or libjingle worker thread. Encode is called on VieCaptureThread.
Others are called on libjingle worker thread.
    
BUG=chromium:350111
TESTS=webrtc call on CrOS.

Review URL: https://codereview.chromium.org/347073002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@279669 0039d316-1c4b-4281-b951-d872f2087c98