Update 4 autosizing tests for the new autosizer
This patch updates 4 tests for the new autosizer:
1) nested-child.html
The old autosizer would group blocks with similar widths which caused
this test to fail with the new autosizer. We now rely on fingerprinting
for this purpose. I've kept this test around since it does test
nested children of different widths.
2) narrow-descendants-combined.html
This test has been removed because narrow descendants are no longer
combined.
3) narrow-iframe.html
This test has been updated to not depend on the iframe's width but
instead autosize based on the block's width.
4) clusters-insufficient-text.html
Table autosizing has changed significantly with the new text
autosizer so I've removed the table from this test. The test still
provides good coverage for blocks with insufficient text.

Review URL: https://codereview.chromium.org/351143002

git-svn-id: svn://svn.chromium.org/blink/trunk@177020 bbb929c8-8fbe-4397-9dbb-9b2b20218538