Overview Mode Close Buttons Not Rendering on 2nd Display
The bounds of a layer are already mapped to the given display coordinate space.
This makes ScreenUtil::ConvertRectFromScreen redundant on the primary display.
However for other displays this incorrectly changes the coordinate space.

This change switches to use the correct layer bounds for placing the close button.

TEST=WindowSelectorTest.MultipleDisplays
BUG=388848

Review URL: https://codereview.chromium.org/358703005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@280175 0039d316-1c4b-4281-b951-d872f2087c98