Move animation properties back into LowPriorityProperties
Previously we've attempted to have animations and transitions apply as
part of the cascade instead of always as the last step, which required
applying animation-related properties before high priority properties.
This hasn't ended up working and we have no plans to fix the cascade,
so we no longer need the separate animation properties pass.

BUG=232273

Review URL: https://codereview.chromium.org/351553006

git-svn-id: svn://svn.chromium.org/blink/trunk@177034 bbb929c8-8fbe-4397-9dbb-9b2b20218538