Drop early-outs for "no boxes" in SVGTextQuery entrypoints
All entrypoints in SVGTextQuery start with check if the text content
element has any boxes. If it doesn't, a value representing "empty" is
returned. In all cases this value is the same as the default value of the
corresponding query-data. SVGTextQuery::executeQuery itself has no
problems handling this case (and will return 'false' as expected.)
Removing these early-outs means a few more cycles is needed in these
"degenerate" cases, although that seems like a small (and hopefully
reasonable) price to pay for less code (and duplication/exit-paths).

Review URL: https://codereview.chromium.org/358033005

git-svn-id: svn://svn.chromium.org/blink/trunk@177149 bbb929c8-8fbe-4397-9dbb-9b2b20218538