Add Bluetooth LE enumeration to adapter polling.
Initial change: During adapter polling, the list of Bluetooth LE devices is enumerated as the same time as the Classic devices. They are all reported in a single collection which ends up being exposed to the JavaScript API. For BLE devices, we currently only initialize the "name" and "address" fields of "DeviceInfo".

BUG=388016

Review URL: https://codereview.chromium.org/349373002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@280470 0039d316-1c4b-4281-b951-d872f2087c98