Don't send accessibility notifications on wrong HWND.
In r276950 (cl 329053004), we made it so that a
BrowserAccessibilityManagerWin can be created without a
LegacyRenderWidgetHostHWND, to fix a race condition.
However, we should suppress all notifications from firing
until the LegacyRenderWidgetHostHWND is ready because they'd
just go to the wrong HWND.

As soon as the LegacyRenderWidgetHostHWND is passed in,
we'll fire a focus event on the root and all events from
then on will be fine.

BUG=368549

Review URL: https://codereview.chromium.org/359063004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@280702 0039d316-1c4b-4281-b951-d872f2087c98