Use DirectWrite to populate list of avalible fonts
Use the DirectWrite API to populate the list of avalible fonts in
settings when DirectWrite is enabled. False back on GDI on failure.
As GDI and DirectWrite handle font family variants and alternate names
differently this ensures that all fonts presented in the list are valid
options that can be rendered.

R=scottmg@chromium.org, cevans@chromium.org
BUG=367702, 373494

Review URL: https://codereview.chromium.org/354023007

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@280719 0039d316-1c4b-4281-b951-d872f2087c98