Specifies the popup alignment from Ash.
MessagePopupCollection estimates the popup alignment by itself,
checking the work area and display bounds. However, this approach
may misunderstand the reserved area of docked windows as the
taskbar, which causes weird effect as is described in the bug.

Actually, this estimation is not necessary at all in Ash, because
Ash knows the exact taskbar alignment. This CL allows Ash to
specify the popup alignment, and ignores the estimation code.

BUG=389656
R=stevenjb@chromium.org, dimich@chromium.org
TEST=manually

Review URL: https://codereview.chromium.org/353243002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@280721 0039d316-1c4b-4281-b951-d872f2087c98