Do not zero handed_out_sockets in FlushWithError
WebSocketTransportClientSocketPool::FlushWithError() was zeroing the
count of handed out sockets. Since these sockets belong to the various
ClientSocketHandles and not to the pool, we cannot make them go away in
FlushWithError(). Instead they are deleted when ReleaseSocket() is
called.

Also add a test to verify this behaviour.

TBR=jgraettinger@chromium.org
BUG=389116
TEST=net_unittests

Review URL: https://codereview.chromium.org/351393003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@280776 0039d316-1c4b-4281-b951-d872f2087c98