Replace 'ResourceRequest::TargetType' with 'ResourceRequest::RequestContext'.
Fetch defines a variety of "request contexts", which Blink more or less
already models as either 'ResourceRequest::TargetType' or 'Resource::Type'.
These two enums more or less cover the same ground, but inconsistently
(even amongst themselves).

This patch is the first step towards a higher fidelity implementation of
Fetch's semantics. It replaces the 'TargetType' enum with a 'RequestContext'
enum, and sets up a temporary mapping in WebURLRequest which will be removed
once Chromium is updated with the new semantics.

A future patch will come up with some reasonable way of aligning
'Resource::Type' with the list of contexts.

BUG=390497

Review URL: https://codereview.chromium.org/360233005

git-svn-id: svn://svn.chromium.org/blink/trunk@177371 bbb929c8-8fbe-4397-9dbb-9b2b20218538