[Hotword] Remove excessive special casing for EXTERNAL_COMPONENTS in the extenion system.
Special casing was handling for EXTERNAL_COMPONENTs to allow them to be disabled by users. However, in reality we just wanted the ability to disable them from Chrome without making it possible for users to do so (which can lead to consistency issues between the extension and the preference).

This CL adds the check for EXTERNAL_COMPONENTs only where it's needed -- when disabling -- and removes it from the other cases. Now EXTERNAL_COMPONENTs are handled like regular components with respect to user modification and the admin policy.

BUG=353266

Review URL: https://codereview.chromium.org/362103004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@281251 0039d316-1c4b-4281-b951-d872f2087c98