Mac AVFoundation: Blackmagic camera blacklisting update
crrev.com/255933006 landed blacklisting for BlackMagic
cameras in Mac: QTKit would be used instead of AVFoundation
for these devices.

The code was based on the assumption that these cameras
have a USB vendor ID but this is not correct since they
use a Thunderbolt connection, i.e. PCIe, and another
crrev.com/366593003 is going to prevent GetModel().

This CL changes to using a characteristic substring of
the uniqueId.

TEST: Plug in a Blackmagic, f.i. an UltraStudio Mini
Recorder (doesn't need any camera connected to it) and
check enumerate devices, there should be a multiplicity
of devices with a "QTKit..." prefix, indicating there was
a blacklisting process.

BUG=347371

Review URL: https://codereview.chromium.org/362323003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@281707 0039d316-1c4b-4281-b951-d872f2087c98