Improve YouCompleteMe handling of Blink header without source files.
In this case, simply giving the Chromium src root as the include path
is unlikely to be helpful. We can get a pretty close approximation by
getting the flags for an arbitrary Blink source file, which will have
more reasonable include paths and defines for a Blink header.

Review URL: https://codereview.chromium.org/368313003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@281759 0039d316-1c4b-4281-b951-d872f2087c98