Remove DispositionChangePhase.
The problem with having a single observer method with a phase param is that it's cumbersome and error prone to use for most clients that just want to know when something *has* changed. I think it's appropriate then to create separate methods for each phase and have the client override the one (or both) that is relevant to them.

TBR=sky@chromium.org
BUG=none

Review URL: https://codereview.chromium.org/374933003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@281874 0039d316-1c4b-4281-b951-d872f2087c98