sync: Inject sync/'s dependency on invalidations
Defines an InvalidationInterface class and uses it to break any direct
dependencies from sync code on syncer::Invalidation.  Despite its name,
syncer::Invalidation should belong solely to the invalidations
component, which code in the sync/ directory should not depend on.

Changes the interface in the sync engine from copying
syncer::Invalidation to managing scoped_ptr<InvalidationInterface>.
This change in memory management was required to support the use of an
abstract interface.

Removes the DroppedInvaldiationTracker.  This class was previously only
used by sync.  The small benefit provided by this class is outweighed by
the amount of glue code it would take to maintain it.

Changes tests to conform to the new interface.  Adds some test-only
implementations of InvalidationInterface and some associated classes.

BUG=259559

Review URL: https://codereview.chromium.org/322333004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@281884 0039d316-1c4b-4281-b951-d872f2087c98