ScriptController::callFunction should report a recursion error message.
v8::Function::Call reports a message when there is a v8::TryCatch with
is_verbose_ set. On the other hand, when the recursion level is too deep,
ScriptController::callFunction doesn't call v8::Function::Call and simply
throws a RangeError exception.
That is a problem when we don't want to rethrow the exception.
This CL makes V8ScriptRunner::callFunction and V8ScriptRunner::runCompiledScript
call a v8 function throwing a stack overflow exception when the recursion level
is too deep.

This is a preparation to make v8::TryCatch nestable.

BUG=362388

Review URL: https://codereview.chromium.org/358073003

git-svn-id: svn://svn.chromium.org/blink/trunk@177715 bbb929c8-8fbe-4397-9dbb-9b2b20218538