Implement midi permissions on top of the new common permission classes
- move all the generic permission settings from gcm to content_settings
- delete the midi implementation in favour of this one.

Some important changes in midi.

Clicking Yes/No on Bubbles did not save the permission for future uses,
while when using infobars it was saved. Now it is saved in both cases

The existing midi implementation had some renderer crashes that have now
been fixed. It also had a browser race when tring to run the result
callback. Both issues are fixed in this implementation.

The midi cancel closure is not implemented, it does not seem to be used
anywhere and the paradigm was a bit convoluted
(passing it as an out pointer in RequestMidiSysExPermission)

The old midi infobar used to say Allow/Deny but the bubble said Allow/Block,
now it says Allow/Block in both cases.

BUG=392145

Review URL: https://codereview.chromium.org/365123003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@282003 0039d316-1c4b-4281-b951-d872f2087c98