Show zoom bubble for everything but an empty host.
Changes committed in r281468 were overly restrictive about when the
zoom bubble can be shown. This CL restores the old behaviour, noting
that the usual cases where the zoom bubble is supressed, namely
for changes initiated from the wrench menu and for changes in the
dev tools window, are done elsewhere in the browser.

BUG=391680

Review URL: https://codereview.chromium.org/372813003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@282073 0039d316-1c4b-4281-b951-d872f2087c98