In iossim, fix default device value for Xcode 6.
The default was "iPhone", but that's not supported in Xcode 6, so use
"iPhone 5" in that case.

Also remove an unnecessary call to |connect| when printing supported
devices on Xcode 6. This call is not needed and just hangs for a few
seconds so remove it (all that's really needed is to load the
CoreSimulation framework.

BUG=390581
TEST=With Xcode 6 installed, build iossim then run iossim with no
optional paramters (e.g. out/Debug-iphonesimulator/iossim path/to/my.app
TBR=rohitrao@chromium.org

Review URL: https://codereview.chromium.org/379863002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@282075 0039d316-1c4b-4281-b951-d872f2087c98