Fixes for re-enabling more MSVC level 4 warnings: pdf/ edition
This contains fixes for the following sorts of issues:
* Signedness mismatch

This relies on https://codereview.chromium.org/376003003 to make FPDF_FillRect()
take a 32-bit value for the color in order to make the changes here as simple
and clean as possible.

This also contains a few other cleanups to make code simpler or more consistent.

BUG=81439
TEST=none

Review URL: https://codereview.chromium.org/372273005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@282146 0039d316-1c4b-4281-b951-d872f2087c98