Refine UMA quality for warm-start app list cases
Currently "warm start" app list cases (i.e. shown via the process
singleton) have noisy UMA data. Whether the profile to show is loaded or
not has a big impact on the metric. We are only interested in the case 
that the app list view is "warm", and can be immediately shown. 
Recording cases that first must load a profile from disk, or switch the 
app list view to a different profile, count as noise and skew the 
metric. 

This change filters out the noise, fixing an incorrect assumption in 
r270615 which resulted in only the noise being recorded (and a failed 
DCHECK otherwise). Before r270615, the UMA included cases requiring a 
profile load as well as those that didn't. After r270615, only cases 
that required a profile load were being recorded. 

Note that ShowAppListInteractiveTest.ShowAppListFlag was updated in
r281213 to include both cold and warm codepaths via
StartupBrowserCreator, to cover the aforementioned DCHECK.

BUG=382793

Review URL: https://codereview.chromium.org/366233002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@282245 0039d316-1c4b-4281-b951-d872f2087c98