Remove the WM_SYSCOMMAND handling from the LegacyRenderWidgetHostHWND class.
This handler was added to forward the message to the parent window for cases where the size of the
parent and the LegacyRenderWidgetHostHWND window were the same.

We can directly call DefWindowProc on the parent window for non client mouse messages which generates
WM_SYSCOMMAND for the parent window.

This simplifies the LegacyRenderWidgetHostHWND class and also fixes bugs with WM_SYSCOMMAND messages not
getting generated correctly due to missing styles on the LegacyRenderWidgetHostHWND hwnd which does not have
the same styles as the parent.

BUG=377367

Review URL: https://codereview.chromium.org/374283002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@282248 0039d316-1c4b-4281-b951-d872f2087c98