Use [Reflect=value] for HTMLInputElement.defaultValue
The HTML specification describes the defaultValue as reflecting the value
content attribute. This was also how it was implemented, but unlike all
other HTMLInputElement attributes that reflect content attributes,
[Reflect] wasn't used for defaultValue.

The C++ getter HTMLInputElement::defaultValue() is kept since it is used
by other C++ code, and because it's name is a good documentation of what
it returns (even if it really just reads the value content attribute.)

This does not affect behavior in any way.

BUG=392471

Review URL: https://codereview.chromium.org/387493003

git-svn-id: svn://svn.chromium.org/blink/trunk@177838 bbb929c8-8fbe-4397-9dbb-9b2b20218538