Invalid invalidation on www.flipkart.com
The bug was solved by https://codereview.chromium.org/371913005/.
The issue is that computeRepaintRectsIncludingNonCompositingDescendants
updates the renderers' position before invalidations are generated.
That means that the repaint-after-layout logic never sees a potential
bounds' position shift as it is taken care of by the previous logic.

Landing the test to make sure we don't regress.

BUG=391648

Review URL: https://codereview.chromium.org/382003002

git-svn-id: svn://svn.chromium.org/blink/trunk@177977 bbb929c8-8fbe-4397-9dbb-9b2b20218538