Split targets with external dependencies out of mojo_public.gypi
The non-testing code in src/mojo/public should be buildable without
depending on anything outside of that directory. It's fine for test
code to depend on base/ and gtest/, though. Due to the way GYP works,
having targets for both production and test code in the same .gypi
file winds up trying to resolve the dependencies for both kinds of
targets, breaking the ability to build the public mojo code
hermetically.

Splitting the targets out will enable consumers of the mojo code to
build the public code standalone using GYP.

BUG=chromium:388412
TEST=build and run all mojo unittests

Review URL: https://codereview.chromium.org/385603003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@282699 0039d316-1c4b-4281-b951-d872f2087c98