ServiceWorker: Stop a worker thread on unregistration
ServiceWorker is designed to delay stopping the worker thread (30 sec) when all
controlled documents are closed so that the next navigation can save the startup
time.

In the current code, this improvement also applies to the unregistration with no
controllees, but in that case we should immediately stop the thread because the
worker doesn't work anymore.


BUG=392735,393127
TEST=content_unittests --gtest_filter=ServiceWorker*

Review URL: https://codereview.chromium.org/385143002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@282922 0039d316-1c4b-4281-b951-d872f2087c98