Enable resize and layout for frames with RemoteFrame ancestors
In order to support frame tree replication across processes, frames
must be able to render correctly even when they have ancestor frames
out of process. This will require a refactoring of WebViewImpl, but
to enable immediate progress on this goal we add the assumption to
WebViewImpl that the first LocalFrame found in a frame tree traversal
should be treated as a local root, and used instead of the actual
main frame in some cases.

BUG=346764

Review URL: https://codereview.chromium.org/380303003

git-svn-id: svn://svn.chromium.org/blink/trunk@178066 bbb929c8-8fbe-4397-9dbb-9b2b20218538