mac: Fix tab dragging visual bug in Yosemite.
In OSX 10.10+, all views must be added to the NSWindow's contentView. Some
views (like the tab strip and the profile icon) are placed on top of the title
bar and require special treatment. All other views are added as subviews of
'chromeContentView' in TabWindowController.  This allows tab dragging and
fullscreen logic to easily move the views that don't need special treatment.

This CL also removes the instances where a VersionIndependentWindow's
contentView gets replaced by setContentView:. Instead, the 'chromeContentView'
gets passed around as a subview. This allows VersionIndependentWindow to remove
another of its internal hacks.

BUG=392239

Review URL: https://codereview.chromium.org/379293003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@283166 0039d316-1c4b-4281-b951-d872f2087c98