First pass at a page set for the Polymer widgets.
Note that this deliberately used the full embedding of the Polymer Paper
Sampler, rather than individual pages, because the DOM is larger & more
complex which brings out more of our perf issues. Each widget page is still
an individual page in the page set, though.

The approach is to mostly-automatically interact with every widget on each
subpage in the sampler. Still todo are the input subpage and the shadows
subpage, but I'd like to get a basic version sanity checked and submitted
first.

BUG=333078

Review URL: https://codereview.chromium.org/376613002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@283217 0039d316-1c4b-4281-b951-d872f2087c98