Add a FrameHostMsg_BeginNavigation IPC
As part of the project PlzNavigate, this CL introduces a new IPC,
FrameHostMsg_BeginNavigation that should be sent from the renderer to the
browser on renderer-initiated navigation, or in response to a
FrameMsg_RequestNavigation (that will be introduced in a later CL) for browser
initiated ones.

Upon reception of FrameHostMsg_BeginNavigation, the UI thread will forward the relevant information to the IO thread (to the ResourceDispatcherHost). 

BUG=376014

Review URL: https://codereview.chromium.org/367653002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@283346 0039d316-1c4b-4281-b951-d872f2087c98