MacViews: Refactor ShowConfirmBubble to also take a gfx::NativeWindow
Currently there's a view/window type mismatch that Aura excuses because
they are the same type (aura::Window*). This doesn't work on MacViews.

For toolkit-views, CreateBrowserModalDialogViews(..) actually takes a
gfx::NativeWindow. However, for Cocoa, the confirm bubble is added as a
subview of the render view.

This changes chrome::ShowConfirmBubble() to accept both a NativeView and
a NativeWindow. Views ignores the view and Cocoa ignores the window but
it allows a toolkit-views build on Mac to call the Cocoa implementation
which is what we need to get started.

BUG=390755

Review URL: https://codereview.chromium.org/393763002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@283373 0039d316-1c4b-4281-b951-d872f2087c98