Make drag and drop work for a node in a shadow tree.
We need to use Node::containIncludingShadowDOM() instead of Node::contains() when checking ancestor/descendant relationships for dragging.

This is a quick fix and should be better than the current broken drag-and-drop behavior. In the future, we might want to use a composed-tree based checking.

BUG=264983
TEST=A new layout test + manual test for the case filed in the bug

Review URL: https://codereview.chromium.org/397583002

git-svn-id: svn://svn.chromium.org/blink/trunk@178285 bbb929c8-8fbe-4397-9dbb-9b2b20218538