athena: Add a minimized state for the home-card.
The home-card can be in a minimized state. In this state, only a narrow line
is displayed at the bottom of the screen, and dragging it up shows the overview
mode. When a window is selected in the overview mode, the home-card is minimized
again.

BUG=381224
R=mukai@chromium.org, oshima@chromium.org

Review URL: https://codereview.chromium.org/394043002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@283539 0039d316-1c4b-4281-b951-d872f2087c98