ozone: Port WindowTreeHostOzone on top of PlatformWindow
Most of the WindowTreeHost functionality is actually implemented by the
platform window, so there's not much in WindowTreeHostOzone aside from
forwarding calls to and from the platform part.
    
Should be no functional change for in-tree platforms. To accomplish
that, this adds a transitional class PlatformWindowCompat that
implements PlatformWindow functions as WindowTreeHostOzone did before.
As a followup, we'll convert each platform to use the new PlatformWindow
natively & then remove PlatformWindowCompat.

BUG=392280
TEST=built with chromeos==1 use_ozone==1 & ran all platforms.
NOTRY=true
TBR=ben@chromium.org

Review URL: https://codereview.chromium.org/375053002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@283573 0039d316-1c4b-4281-b951-d872f2087c98