Associate internal touchscreen to internal display
Currently we associate touchscreen to display based on matching
resolution. But there could be cases the internal touchscreen
reports a resolution that does not match any display resolution,
or the internal display and external display have the same resolution
so we don't know which display to match the touchscreen to.

This patch adds a logic that if we have internal touchscreen and
internal display, we should match them first regardless of the
resolution.

We already have the information about which display is internal display.
For deciding internal touchscreen, we query X to get the touchscreen's
/dev/input/eventX  node, and call a script is_touchscreen_internal
(which checks if the device is a I2C device) to decide whether the
touchscreen is internal or not.

BUG=chrome-os-partner:29398
TEST=touch position is correctly mapped on Blaze in extending mode, with
     external monitor having 1920x1280 resolution.

Review URL: https://codereview.chromium.org/394063004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@283641 0039d316-1c4b-4281-b951-d872f2087c98