Add support in mojo to check if a MojoHandle identifies a message pipe.
Updated mojo spy to use the newly added function MojoIsMessagePipe to check if the handle
coming back from the ReadMessageRaw call is a message pipe. If yes we go ahead and intercept it.

Removed the hacks added in the previous patch which assumed that the first handle coming back from
the vendor is a message pipe.

BUG=360188

Review URL: https://codereview.chromium.org/391233002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@283680 0039d316-1c4b-4281-b951-d872f2087c98