Have mojo_shell run in its custom thread on android.
The native viewport service on android must run on the main thread
of the application. Because it is expected at the moment to run on
its own thread, this CL move the shell thread on android to a
secondary background thread and use the main thread of the
application for the native viewport service.

Review URL: https://codereview.chromium.org/399663002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@283764 0039d316-1c4b-4281-b951-d872f2087c98