Use audio thread for the fake input audio stream.
Previously the fake input stream is creating a new thread to pump the callbacks, this is unnecessary and FakeAudioInputStream::Stop() might take time to stop the thread. I suspect that this is one of the potential causes for the timeout in WebRtc content browser tests on Android bots.

This patch changes the code to use AudioManager audio thread instead, hopefully this can reduce the flakiness on Android tests.
Also it changes |beep_lock| from a global lock to a lock used only by BeepContext privately.

BUG=387895
TEST=bots

Review URL: https://codereview.chromium.org/399623008

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@283822 0039d316-1c4b-4281-b951-d872f2087c98