Do not wrap the bugs column on the flakiness dashboard
The flakiness dashboard forces different whites-space handling for the
bugs column to handle long bugs. When the dashboard is narrow due to
the screen being narrow or the flakiness runs being longer than the
page (common after [1]), the wrapped "File new bug" text causes every
row to be tall and ugly (see: http://pr.gg/longbug.png). This patch
changes the new bug text to be nowrap.

The original code for longbug handling was added before
TestExpectations was as clean as it is today.

[1] https://codereview.chromium.org/359533004/

Review URL: https://codereview.chromium.org/402643002

git-svn-id: svn://svn.chromium.org/blink/trunk@178396 bbb929c8-8fbe-4397-9dbb-9b2b20218538