Experimentally disable termination on heap corruption in order to measure the contribution of this feature to missing crash reports.
Because this feature is configured very early in the process lifetime it cannot be directly controlled by a field-trial. Rather, we query the status during a given execution, store that status in the registry, and then query the registry during startup. This means the experiment will only take effect the 2nd time it is executed.

BUG=394842

Review URL: https://codereview.chromium.org/402723002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@284100 0039d316-1c4b-4281-b951-d872f2087c98