sync: Support non-blocking initial sync in proto
Adds support for non-blocking initial sync to the sync protocol.  This
allows some types to request an initial sync without having the
scheduler enter configure mode.  The protocol changes are necessary to
keep the server informed that the request is due to an initial sync
request.

The remainder of the changes in this CL were required to plumb
information about the cause of the sync cycle through the SyncScheduler.
This CL also includes some new tests.

At the moment, there are no types in use that support this feature.
This CL will not affect behavior in any way.

BUG=351005

Review URL: https://codereview.chromium.org/387983004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@284159 0039d316-1c4b-4281-b951-d872f2087c98