Use a C++11 feature in a single file, to discover if this works everywhere.
In theory, as of r284372 all our bots should support using C++11 language
features in target builds. (Build targets that are needed in host binaries still
can't use it yet, and library features can't be used yet either, see
https://groups.google.com/a/chromium.org/forum/#!topic/chromium-dev/2RzwrIxnvqM ).

To test this theory, use a C++11 feature in a file that should be used on most
platforms. If this causes problems in many places, this is easy to revert.
If your bot can't deal with this, please let me know!

BUG=none

Review URL: https://codereview.chromium.org/385743002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@284444 0039d316-1c4b-4281-b951-d872f2087c98