Remove 2 calls to paintInvalidationForWholeRenderer
These calls were needed as we can make the body or the
document element's background bleed into the canvas (ie
replacing the FrameView's background). However the logic
was working around the fact that
clippedOverflowRectForPaintInvalidation didn't account for
the previous cases.

The whole logic works as the 2 renderers where this happens
share the same coordinate space as the viewport. The tests
are all invalidating as much, we just got rid of some stray
invalidations.

Review URL: https://codereview.chromium.org/396073002

git-svn-id: svn://svn.chromium.org/blink/trunk@178605 bbb929c8-8fbe-4397-9dbb-9b2b20218538