mojo: make NetworkServiceImpl clean up after itself
Each URLLoaderImpl created by NetworkServiceImpl can allocate a net::URLRequest.
When shutting down the app, destruction of the NetworkContext and the
URLRequestContext within it expects that all net::URLRequest resources have been
freed by now (it CHECKs). NetworkService is an app like any other and can't make
any assumptions about when it is being destroyed relative to other apps in the
system that may have been using it, so the safest thing to do is to have it
gracefully tear-down in-progress URLLoaders.

Note: the only reason we're not seeing the CHECK happen is because things
shut down (the process terminates) before we even get to ~URLRequestContext.
I'll be repairing shutdown in an upcoming CL.

BUG=394477

Review URL: https://codereview.chromium.org/407593002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@284777 0039d316-1c4b-4281-b951-d872f2087c98