Remove disable-pnacl flag.
This was used as part of the PNaCl rollout but is no longer necessary. I came
across this while investigating what code in chrome needs to be updated once
the file for the trusted plugin library is gone.

Usage statistics say that this flag has been used 1850 times in the past 7
days by 197 users.

TBR=ncbray
BUG=354867

Review URL: https://codereview.chromium.org/392023002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@284997 0039d316-1c4b-4281-b951-d872f2087c98