Remove the blue tool-bar for apps v2.
Its visual style is not really in keeping with the rest of the v2 app,
and the form-factor is not conducive to an easily extensible UI. For now,
a menu will suffice; if we need anything more complex, it can easily be
replaced with an options dialog.

This CL also simplifies the MenuButton class to use stopPropagation instead of timers to prevent the close handler being triggered by the same click that opens the menu.

BUG=134213

Review URL: https://codereview.chromium.org/339613003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@285136 0039d316-1c4b-4281-b951-d872f2087c98