Correctly update the bounds of objects in the accessibility tree.
Fixes regression that happened when moving to the new AXTree code.
Improves on it by only sending location changes when there's an
AXLayoutChanged event, since that will always be fired when there's a
layout and objects might have moved.

To make it testable, adds a new feature to DumpAccessibilityTree tests
where you can make it wait until a certain magic string appears in the
dump before the dump is diffed with the expectation. In this test, we
start a CSS transition that animates changing a button's size, and make
sure the final dump reflects the final size.

BUG=268296

Review URL: https://codereview.chromium.org/401643003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@285174 0039d316-1c4b-4281-b951-d872f2087c98