Telemetry: Initialize smoothness and thread_times metrics before page load, not after.
Currently both smoothness and thread_times initialize their metrics after page
load. This leads to a gap between the page load and the page actions starting.
This patch moves the initialization before the page load, thus eliminating the
gap. It is safe to do so because smoothness and thread_times require a page load
before each run (even when repeating the same page).

Note to perf sheriffs:
This patch is likely to cause performance numbers to worsen, because the browser is often more busy immediately after page load. If you get alerts related to this CL, feel free to ignore them. If there is a reference build for the bot, the number for the reference build should also change accordingly.

BUG=392895

Review URL: https://codereview.chromium.org/392613002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@285297 0039d316-1c4b-4281-b951-d872f2087c98