GN foreach should mark the list identifier as used.
Previously an identifier used as the list in a feoreach statement would not count as "using" the identifier, which would give unused variable errors. Added a unit test.

Checks for unused identifiers in component(). Previously these were not checked. This checking turned up a misspelling in the cc build. Added a unit test.

Fixes for the ozone media build (turned up when testing).

R=dpranke@chromium.org

Review URL: https://codereview.chromium.org/420443003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@285367 0039d316-1c4b-4281-b951-d872f2087c98