Never attempt to update the render tree on a dead document.
In https://src.chromium.org/viewvc/blink?revision=167137&view=revision
we added a check for change != Force before the check to
needsRenderTreeUpdate() update in updateRenderTree, but this means that
now if the change is Force you can skip the view() and isActive() checks
that are inside needsRenderTreeUpdate and then incorrectly attempt to
update the render tree of a dead document which will crash.

The ClusterFuzz test case in super hard to minimize, but this is clearly
what happens. This manifests as a crash in the release assert for
entering updateRenderTree from inside layout:

RELEASE_ASSERT(!view()->isInPerformLayout());

but what really happened is that view() is null.

BUG=377724

Review URL: https://codereview.chromium.org/415973003

git-svn-id: svn://svn.chromium.org/blink/trunk@178875 bbb929c8-8fbe-4397-9dbb-9b2b20218538