cc: Fix tiling raster iterator to process all tiles correctly.
This is a fix for the following situation: If we have no skewport,
then in operator++ we call advance phase after we process all visible
tiles, however because the skewport yields no tiles, we immediately
skip to the eventually bin, since the soon iterator construction
was in operator++ instead of advance phase.

The reason this passed unittests is that there was a second bug in
which we would process the eventually rect without ignoring the soon
rect, which caused the same tiles to show up.

This patch fixes both bugs. The first one is fixed by moving tiling
data iterator construction to be in advancephase in all cases. 
The second bug changes the iterator to iterate around the soon
rect instead of the visible rect.

R=reveman

Review URL: https://codereview.chromium.org/413053004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@285448 0039d316-1c4b-4281-b951-d872f2087c98