Componentize ComponentUpdater: Remove erroneous task runner DCHECK in DefaultComponentInstaller.
Removes an erroneous DCHECK added in this patch: https://codereview.chromium.org/385013002/

This DCHECK is erroneous, because the Install() function in DefaultComponentInstaller is actually called on a different task runner than the member task runner.

It's called from ComponentUnpacker::Install. ComponentUnpacker has a different task runner with a different sequence token, so it's erroneous to DCHECK it against the member task runner of DefaultComponentInstaller.

BUG=397264

Review URL: https://codereview.chromium.org/419103002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@285526 0039d316-1c4b-4281-b951-d872f2087c98