Attempt to deflake service-worker-gc.html
Two failure modes I've seen are:
- one of the "assert wasCollected is false" is missing
- the final "assert wasCollected is false" is actually true

I suspect the setTimeout in onRegister() is causing the 
flakes. The failures are explained if that task can execute 
after unregister or onUnregister.

This patch simplifies the test to only check for gc
in two places: immediately before unregister and
after the worker reaches the redundant state.

BUG=383972,393721

Review URL: https://codereview.chromium.org/419463003

git-svn-id: svn://svn.chromium.org/blink/trunk@178917 bbb929c8-8fbe-4397-9dbb-9b2b20218538