extensions: Create extensions_pak target.
Currently this was just an action directly into extensions_unittests.

But we want to share the same pak file between extensions_unittests and app_shell.

So in order to do this we create a new target called "extensions_pak" and make
 extensions_unittests depend on it.

After this lands, app_shell will be refactored to use extensions.pak too.

BUG=397250
TEST=extensions_unittests
R=jamescook@chromium.org
TBR=yoz@chromium.org

Review URL: https://codereview.chromium.org/413293003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@285597 0039d316-1c4b-4281-b951-d872f2087c98