Telemetry: Fix inbox_app.html?slide_drawer of key_silk_cases.
This page toggles the drawer as part of its navigate steps. It then waits until
the navigation drawer is active before continuing. The navigation drawer is
active, however, as soon as it starts moving, so it doesn't mean that the drawer
is in its final position.
This CL adds an additional check to make sure that the drawer is actually fully
out before continuing.

BUG=

Review URL: https://codereview.chromium.org/418133008

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@285608 0039d316-1c4b-4281-b951-d872f2087c98