Identity API: make sure account ID is set correctly after sign in
Previously login access token requests were always made for the
primary account ID of the profile. Starting in r275173, the ID is
computed and saved in a member variable. This change makes sure the
member is updated after the user signs in, which allows getAuthToken
requests that trigger a sign-in to succeed.

BUG=396883

Review URL: https://codereview.chromium.org/418393003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@285654 0039d316-1c4b-4281-b951-d872f2087c98