Limit the window of time a promo can be visible after first view.
Notification promos currently have start, end times, and a number of maximum
views. However, the length of time a promo can be visible was not capped.
This change adds a flag to cap the length of time a promo is visible by
max_seconds_.

BUG=None
Test=None

Review URL: https://codereview.chromium.org/419033007

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@285803 0039d316-1c4b-4281-b951-d872f2087c98