[Temp] Roll ANGLE to 5ac2ae3ea1f311be61955b2d554e8510474cf4f0.
Manually advance the ANGLE checkout one commit ahead in its
chromium/2062 to grab the fix its bug 669 ("Build started depending on
.git/").

This is required for a build from a tarball to work, and allows us to
remove our manual workaround from the Tizen packaging.

The roll upstream will be part of more recent M37 releases.