Make IPC macros work with -Wunused-local-typedefs.
The typedef can be used by the IPC_MESSAGE_HANDLER macro, but that macro isn't
always used when IPC_BEGIN_MESSAGE_MAP is used. The alternatives are to either
pass ClassName::HanderFunction to IPC_MESSAGE_HANDLER (like what one has to do
to use IPC_MESSAGE_FORWARD), or to mark the typedef in IPC_BEGIN_MESSAGE_MAP
as ALLOW_UNUSED so that it isn't warned on.

BUG=321833
TBR=brettw

Review URL: https://codereview.chromium.org/422083002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@286115 0039d316-1c4b-4281-b951-d872f2087c98