GN: Fix up gpu targets in component build
The gles2_cmd_helper files should only be linked into the gpu component
when using the component build and can be linked standalone in the
static build. This corrects the _IMPLEMENTATION define for these files,
hides the actual source list in a visibility-restricted source_set, and
adds a group called gles2_cmd_helper that expands out to the expected
thing for clients who just want to use this functionality without caring
about which component or library provides it.

Review URL: https://codereview.chromium.org/419053003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@286136 0039d316-1c4b-4281-b951-d872f2087c98