Make resize (very) smooth in ÜC
Make the ui::Compositor used by BrowserCompositorViewMac use the
TaskRunner provided by RenderWidgetResizeHelper.

Bracket the call to RenderWidgetResizeHelper that pumps the nested task
runner in a BeginPumpingFrames/EndPumpingFrames block to instruct the
view to ack frames immediately and defer draws until the resize is 
finished. Plumb this through to the CompositingIOSurfaceLayerHelper 
class where the logic is implemented.

Update RenderWidgetHostViewMac::HasAcceleratedSurface to compare with 
the size of the frame drawn by BrowserCompositorViewMac.

BUG=392031, 390071, 388005

Review URL: https://codereview.chromium.org/419103005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@286155 0039d316-1c4b-4281-b951-d872f2087c98