DevTools: highlight page regions as paths, not quads
We used to pass a typed set of quads plus the config that specified
color for each type of quad. This is not flexible enough to cover
the needs of highlighting of different types of render objects.
This switches to a more flexible interface passing an array of paths
with fill & outline colors specified for each path.

BUG=

Review URL: https://codereview.chromium.org/430463002

git-svn-id: svn://svn.chromium.org/blink/trunk@179150 bbb929c8-8fbe-4397-9dbb-9b2b20218538