Modified visited link shared section to read-only for renderers.
This change modifies the behaviour of the visited link shared section
so that it's read-only inside the renderers. This is to prevent a
compromised renderer causing a reliability issue by corrupting the copy
of the shared hash-table so that all new renderers crash during
initialization. The renderers should not need to modify the hash table.

BUG=397682

Review URL: https://codereview.chromium.org/421653004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@286354 0039d316-1c4b-4281-b951-d872f2087c98