MacViews: Accessibility bridge
Implement basic accessibility hierarchy support for MacViews, bridging the
View hierarchy to a hierarchy of Objective-C objects that implement the
NSAccessibility informal protocol.
AXPlatformNode is introduced as the accessibility abstraction that will be
used for views and browser content, for both Mac and Windows. This initial
patch includes the basic views on Mac implementation.

BUG=396137

Review URL: https://codereview.chromium.org/420653003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@286768 0039d316-1c4b-4281-b951-d872f2087c98