Make event sender supply a good default area for tap gesture events
In most real-world scenarios, tap gestures will come with an area and so
trigger a very different hit testing path in blink (for touch adjustment).
We've had a few cases of missing test coverage for tests that didn't
explicitly supply an area.  So now default to using a reasonable area if
none is provided.

BUG=None

Review URL: https://codereview.chromium.org/423593003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@286873 0039d316-1c4b-4281-b951-d872f2087c98