Fix DelegatedFrameHost::WasShown
When we enabled delegated rendering, we didn't update the logic in
(then) RWHVA::WasShown that would skip waiting on the renderer if we
already have a good frame. Adding this logic now means faster tab
switch (and the frames kept around are not wasted any more).

BUG=399071

Review URL: https://codereview.chromium.org/428933005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@286905 0039d316-1c4b-4281-b951-d872f2087c98