All tasks on worker thread should be posted through WorkerThread.
In the initial refactor to move WorkerThread to use WebThread, I removed the usage of WebWorkerRunLoop, which was the API to post tasks to the worker thread. This causes a problem since content/ posts plain Task objects to the worker message loop. Since logic to skip tasks during cleanup is in WorkerThreadTask, this doesn't apply to content/ posted tasks.
This CL restores the usage of WebWorkerRunLoop and leaves WorkerThread::postTask as the only API to put tasks on a worker thread. This ensures that in the future there will be no plain tasks to get us in the same bad spot.

BUG=301515

Review URL: https://codereview.chromium.org/420683004

git-svn-id: svn://svn.chromium.org/blink/trunk@179423 bbb929c8-8fbe-4397-9dbb-9b2b20218538