[New Tab Page] Add Field Trial support to the Local NTP
This will allow development of the new design in parallel with the current implementation. The plan is to put new styles in local-ntp.css, which will be selected according to the class of the containing div.

This CL also adds a command-line flag, which will be tied to the Field trial from the server.

BUG=399388
TEST=using force-fieldtrials, confirmed that it works.

Review URL: https://codereview.chromium.org/435723002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@287131 0039d316-1c4b-4281-b951-d872f2087c98