ash: Fix a screen-locking crash.
LockStateController starts its 350 ms pre-lock animation
when the screen is locked via the system tray. When the
animation completes, the controller was incorrectly starting
the lock-fail timer. If the lock screen had already become
ready before this point (possible on a fast system), the
timer would crash the process eight seconds later.

Change LockStateController to only start the timer when
actually requesting that the screen be locked.

BUG=349083

Review URL: https://codereview.chromium.org/435083002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@287187 0039d316-1c4b-4281-b951-d872f2087c98