Worker threads should be registered to task observers
r179309 switched the implementation of worker threads to
web threads. This caused timeout in a lot of worker tests
in oilpan builds. This is because the worker threads
were not registered to task observers, and thus the
worker threads didn't enter safe points at the end of
each task execution.

This CL registers worker threads to task observers
and fixes the issue. 

BUG=399762

Review URL: https://codereview.chromium.org/441653002

git-svn-id: svn://svn.chromium.org/blink/trunk@179463 bbb929c8-8fbe-4397-9dbb-9b2b20218538