[win] Set the drag item visible when externally cancelling a drag in the app launcher
When the nested message loop running a synchronous drag event on windows
is cancelled due to an external event, it does not currently set the
item visible again.

Now it does.

BUG=393519
TEST=Open a signed-in profile on two computers. On one computer,
uninstall an app from the app launcher. Then, Quickly switch to the
other computer and start dragging around a different app. When the sync
event to uninstall the app comes in, the drag should cancel, with the
item being dragged reappearing.

Review URL: https://codereview.chromium.org/435303005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@287499 0039d316-1c4b-4281-b951-d872f2087c98