oilpan: Removes unnecessary hack for CSSValue's size.
I misunderstood the problem and workaround, and it turned out that a workaround introduced in http://crrev.com/423383003 didn't work.  The problem is multiple inheritance and we have to use single chain inheritance to avoid the issue of VC++.  However, single chain inheritance doesn't work at this case.  Please read the article, especially the last part, for details.
http://social.msdn.microsoft.com/forums/vstudio/en-US/504c6598-6076-4acf-96b6-e6acb475d302/vc-multiple-inheritance-empty-base-classes-bloats-object-size

Thus, we'll leave the problem as is until VC++'s issue gets fixed or we introduce CSSOMValue class.

Review URL: https://codereview.chromium.org/436223002

git-svn-id: svn://svn.chromium.org/blink/trunk@179605 bbb929c8-8fbe-4397-9dbb-9b2b20218538