Fix widget visibility accounting when crashing in background.
RenderWidgetHostImpl tracks its visibility using the |is_hidden_| field.
This value should be not be reset when the renderer crashes, otherwise a
widget that crashes in backround will think it's already visible when
put in foreground, messing up process visibility accounting.

This patch fixes process visibility for processes that crashed in
background and are respawned upon foregrounding by preserving the
|is_hidden_| field through renderer crashes.

BUG=399521
R=aelias@chromium.org

Review URL: https://codereview.chromium.org/443133002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@287864 0039d316-1c4b-4281-b951-d872f2087c98