Create ExtensionActionView class
Create an ExtensionActionView class which represents the core of either a
PageActionView or a BrowserActionView. The goal here is to move all logic for
executing the action itself into ExtensionActionView, so that
PageActionImageView and BrowserActionButton (the current implementation classes)
become nothing more than a view wrapper for the ExtensionActionView.

This should allow us to show page actions in the now-incorrectly-named
BrowserActionsContainer and have them render in the same fashion as the current
browser actions.

BUG=397259

Review URL: https://codereview.chromium.org/431173002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@287951 0039d316-1c4b-4281-b951-d872f2087c98