Add the whispernet proxy.
To communicate with whispernet (which needs to be checked in as a pNacl binary),
the only way possible is for us to go through an extension. This code
implements the ChromeWhispernetClient, which uses the whispernet proxy compoent
extension. This extension is implemented on top of a private API that allows it
to communicate back and forth with the ChromeWhispernetClient.

The tests for this code exist only for the ChromeWhispernetClient since they
excersize the full flow back and forth from sending a request to encode a token
to receiving the samples, and then decoding samples and receiving the correct
token.

This is our temporary measure for M38 till we can find a better way to
communicate with the whispernet library.


Owners reviews requested:
xiyuan@ - For the overall review + the whispernet proxy code in
          c/b/r/whispernet_proxy
jochen@ - For adding components/copresence to c/b/DEPS
kalman@ - For the copresence_private API
ben@    - For adding the c/b/copresence directory and the OWNERS file + the
          addition to c/b/browser_resources.grd

R=ben@chromium.org, jochen@chromium.org, kalman@chromium.org, xiyuan@chromium.org
BUG=365493

Review URL: https://codereview.chromium.org/438513002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@287976 0039d316-1c4b-4281-b951-d872f2087c98