Creates a dummy view to occupy the logo area of athena home card.
As is discussed in the bug, it is important to occupy certain
area for logo mark in the home card, otherwise the area for
the search results is restricted.

Actual experimental app-list uses start_page webview for this,
which embeds the logo from our resource data. I think it's better
to create a dummy view with flat background instead of loading
the resource image to emphasize this is still work-in-progress.

BUG=400153
R=oshima@chromium.org
TEST=none

Review URL: https://codereview.chromium.org/441123002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@288199 0039d316-1c4b-4281-b951-d872f2087c98