Implement auto-retry of failed tests for telemetry{_perf}_unittests.
This implements the same basic logic we use in GTest-based tests to
retry tests that fail during a run. By default, we will not retry
anything (this matches what gtests do when run not in botmode).

If --retry-limit is passed on the command line, we will retry each
failing test up to that number of times. So, to enable retries on
the bots by default they will have to pass this arg.

R=dtu@chromium.org, tonyg@chromium.org
BUG=398027

Review URL: https://codereview.chromium.org/426123002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@288265 0039d316-1c4b-4281-b951-d872f2087c98