Refactor SyncManager::Init to receive args via a struct.
SyncManager::Init is implemented in several places, takes a whole lot of
args, and in the near future is going to get a few more.

In some cases, many of the args are optional and it is tedious to pass a
bunch of NULLs or empty strings.  Use a struct instead of an arg list.

BUG=

Review URL: https://codereview.chromium.org/451743002

Cr-Commit-Position: refs/heads/master@{#288280}
git-svn-id: svn://svn.chromium.org/chrome/trunk/src@288280 0039d316-1c4b-4281-b951-d872f2087c98