Disable Screen Orientation API on Win8.
This is a temporary measure until:
- we can disable the API only on Win8 Metro OR,
- we can make it work on Win8 Metro.

The API is currently broken on Win8 Metro and it is not trivial to
disable it on that mode only. Given the discussion with the Blink API
owners, disabling it on Win8 would be better than not shipping because
of the bug on Win8 Metro.

BUG=400846

Review URL: https://codereview.chromium.org/456623003

Cr-Commit-Position: refs/heads/master@{#288371}
git-svn-id: svn://svn.chromium.org/chrome/trunk/src@288371 0039d316-1c4b-4281-b951-d872f2087c98