Change the title of the desktop User Manager widget.
The title of the window widget should be the product name (Chrome/Chromium).

Added a "close" button to make the last bubble in the tutorial dismissible.

BUG=399698,399697
TEST=Start Chrome with --new-avatar-menu. Choose "Switch person" from the
avatar menu of a brand new profile. Once you go through all the slides in
the tutorial, the last bubble which points to the "Add Person" button
should have a little x button in the top right corner. Also, the title of
the window should be "Chrome".

Review URL: https://codereview.chromium.org/440713005

Cr-Commit-Position: refs/heads/master@{#288385}
git-svn-id: svn://svn.chromium.org/chrome/trunk/src@288385 0039d316-1c4b-4281-b951-d872f2087c98