Recommitting: https://codereview.chromium.org/406043003/ (originally reverted because the buildbots weren't pointed to verifier_test_dll.dll by unit_tests.isolate)
=========================================================
Adding the VerifyModule function (and helpers) to safe browsing.

This function allows us to compare a module in memory with 
its original on disk, and identify any differences (after 
accounting for relocations).  The state of the module 
(MODULE_UNKNOWN, MODULE_UNMODIFIED or MODULE_MODIFIED) is 
returned along with a list of the possibly modified exported 
functions of the module.


BUG=401854
TBR= noelutz@google.com, robertshield@chromium.org, csharp@chromium.org, asvitkine@chromium.org, grt@chromium.org
NOTRY=True

Review URL: https://codereview.chromium.org/451893003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@288417 0039d316-1c4b-4281-b951-d872f2087c98