Break labels apart in preloaded list.
Since the hosts are stored in DNS wire format (with length-prefixed labels),
any labels that start with a digit cause an escape like "\0071..." (that's
length=7, first byte of the label = "1"). But this causes a warning in Visual
Studio and breaks the build.

This change splits the length bytes from the labels so that this warning
doesn't happen.

BUG=none
R=palmer@chromium.org

Review URL: https://codereview.chromium.org/456853002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@288443 0039d316-1c4b-4281-b951-d872f2087c98