Stopping Intermediate Load of Page when Distilling Page.
The CopyStateFrom method in the NavigationController followed by
SwapTabContents triggers a reload which is then cancelled in
StartNavigationToDistillerView. However, TabObservers and WebContentsObservers
cannot distinguish between this reload and the user hitting the back button.
By switching the order, the reload is setup in CopyStateFrom and then promptly
cancelled in StartNavigationToDistillerView, so the reload never happens.

BUG=383630

Review URL: https://codereview.chromium.org/448383002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@288516 0039d316-1c4b-4281-b951-d872f2087c98