Add temporary instrumentation to debug ui::Compositor crashes
Multiple crashes are being implicating problems with the ui::Compositor
observer list. Inspection has not found any issues, so add explicit
CHECKs to ensure that
- observers be unregistered before their destruction
- no NULL observers be added

This instrumentation is to be removed once one or two Canary builds
have identified or failed to identify the source of the problems.

BUG=401630

Review URL: https://codereview.chromium.org/449233002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@288553 0039d316-1c4b-4281-b951-d872f2087c98