Add PepperTryCatch and V8ObjectVar classes
This adds PepperTryCatch and V8ObjectVar classes which will be used in the
refactoring which replaces the NPAPI bindings with gin bindings for pepper.

V8ObjectVar is a fairly straightforward replacement for NPObjectVar.

PepperTryCatch (and subclasses) are used to conveniently catch and report
errors eminating from either pepper or V8. It also includes facilities for
conveniently converting between v8 objects and vars. It is essentially the
equivalent of some of the utilities in npapi_glue.h.

This also adds conversion for object vars to the V8VarConverter and some
unimplemented functions to the HostVarTracker to allow these new files to 
compile.

BUG=351636

Review URL: https://codereview.chromium.org/421963008

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@288637 0039d316-1c4b-4281-b951-d872f2087c98