Add test support for transition from NPObject to gin bindings for Pepper
This adds a function to PPB_Testing to run the garbage collector. This is
necessary when using gin-backed v8 objects so that we can test that
destruction works correctly when the gin-backed objects are no longer
referenced. It also modifies the tests to run garbage collection and also
to be careful with scoping such that any handles on the stack to v8 objects
are destroyed before running the GC.

BUG=351636

Review URL: https://codereview.chromium.org/445813002

Cr-Commit-Position: refs/heads/master@{#288647}
git-svn-id: svn://svn.chromium.org/chrome/trunk/src@288647 0039d316-1c4b-4281-b951-d872f2087c98