Custom launcher pages: right-click menu is now the platform apps menu.
Added new extensions::ViewType, VIEW_TYPE_LAUNCHER_PAGE, which is used
to identify custom launcher page views and give them the platform app
menu. Note that currently some of the menu items have no effect.

Right-clicking on a custom launcher page that is not actually a platform
app results in a DCHECK failure (the --custom-launcher-page switch is a
use-at-your-own-risk switch; this situation will not be possible once we
switch to manifests).

BUG=399123

Review URL: https://codereview.chromium.org/442653002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@288670 0039d316-1c4b-4281-b951-d872f2087c98