Oilpan: tweak heuristics for forcing conservative GC.
Keep track of collection rate. If we collect less than 50% of objects,
use a factor of 4 instead of 2 allocation to force conservative GCs.

This will only have an effect when you do not get back to the event
loop in a timely manner. That happens a lot in benchmarks and almost
never in reality.

R=erik.corry@gmail.com, zerny@chromium.org

Review URL: https://codereview.chromium.org/460593002

git-svn-id: svn://svn.chromium.org/blink/trunk@179949 bbb929c8-8fbe-4397-9dbb-9b2b20218538