Set SO_REUSEADDR on layout test server sockets on non-Windows systems
Code to set the flag was removed as part of debugging crbug.com/378444
but this causes ports to report as "in use" for some time after the
test server stops, meaning that re-running layout tests with only a
short delay may fail.

Restore setting the flag, but scoped to non-Windows systems.

BUG=378444
R=dpranke@chromium.org

Review URL: https://codereview.chromium.org/464543003

git-svn-id: svn://svn.chromium.org/blink/trunk@180001 bbb929c8-8fbe-4397-9dbb-9b2b20218538