cc: Fix Raster/Eviction order comparators to return stable values.
For make_heap et al, it is important that we return consistent values
representing strict ordering. For instance, if we Comp(a, b) returns
true, then it must be the case that Comp(b, a) returns false.

Currently, because of IsEmpty checks we can return true in both cases,
which makes stl throw an exception.

This fix, along with added tests fixes this issue.

R=reveman

Review URL: https://codereview.chromium.org/448363002

Cr-Commit-Position: refs/heads/master@{#288818}
git-svn-id: svn://svn.chromium.org/chrome/trunk/src@288818 0039d316-1c4b-4281-b951-d872f2087c98