Fix SetShape (SetAlphaShape) to allow Null regions (+ tests).
SetAlphaShape uses a filter to enforce min/max transparency on the window shape specified by the user. This transparency clamping is required by security to ship the SetShape feature.

We added the call to SetAlphaShape earlier, but we needed to temporarily remove it because of bugs with HiDPI devices and NULL shapes.

The issue with HiDPI devices has been fixed in crrev.com/394193003.
This cl fixes Null shape and re-adds the call to SetAlphaShape so that we can enable this feature on Stable.

BUG=324071

Review URL: https://codereview.chromium.org/456943004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@288889 0039d316-1c4b-4281-b951-d872f2087c98