Refactor PrivateScriptRunner.js to reduce the number of globally exposed variables
It's not a good idea to use global variables in PrivateScriptRunner.js
because they are exposed to all window objects of the private scripts.
This CL reduces the number of global variables used in PrivateScriptRunner.js.

Also this CL removes a |global| parameter from installClass()
since the |global| is always equal to the window object of the JS file.

BUG=341031

Review URL: https://codereview.chromium.org/466623002

git-svn-id: svn://svn.chromium.org/blink/trunk@180053 bbb929c8-8fbe-4397-9dbb-9b2b20218538