ExtensionUpdater: Abstract ExtensionDownloader creation
ExtensionUpdater likes to create and destroy ExtensionDownloader
instances as needed. This is fine, but it means duplicating efforts
any time new dependencies need to be injected into ExtensionDownloader,
because then they also need to be injected into ExtensionUpdater.

This CL adds a factory callback to ExtensionUpdater's constructor
so that ExtensionUpdater creators can opaquely inform the instance
about how it should construct new ExtensionDownloaders.

This also reverts the change to ExtensionDownloader's constructor
which took an (optionally NULL) IdentityProvider; this can now
instead be attached to new ExtensionDownloader instances by the
factory responsible for their construction via the
SetWebstoreIdentityProvider method. ExtensionService
provides an example of this.

BUG=398671

Review URL: https://codereview.chromium.org/456063002

Cr-Commit-Position: refs/heads/master@{#288964}
git-svn-id: svn://svn.chromium.org/chrome/trunk/src@288964 0039d316-1c4b-4281-b951-d872f2087c98