Merge NamedNodesCollection and StaticNodeList classes
Merge NamedNodesCollection and StaticNodeList classes into a single
StaticNodeTypeList templated class as these are identical besides the type
of Node they hold. NamedNodesCollection was holding Elements so it is now
named StaticElementList which is a typedef to StaticNodeTypeList<Element>.
StaticNodeList holds Node and is now a typedef to StaticNodeTypeList<Node>.

This avoids code duplication and allows for tighter typing when using a
StaticNodeList. As a side effect, memory used by StaticElementList (formely
NamedNodesCollection is now reported to V8 as StaticNodeList code was doing
so).

R=adamk@chromium.org, jochen@chromium.org

Review URL: https://codereview.chromium.org/465483002

git-svn-id: svn://svn.chromium.org/blink/trunk@180084 bbb929c8-8fbe-4397-9dbb-9b2b20218538