Export URLRequestContextBuilder::SchemeFactory
The inner class needs to be exported in case any uses in inline methods
or vtable of the outer class causes it to potentially be referenced across
a shared library boundary.

On non-Windows, the visibility attribute is already inherited by nested
classes, but on Windows, they must be marked explicitly.

BUG=82385

Review URL: https://codereview.chromium.org/464563004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@289093 0039d316-1c4b-4281-b951-d872f2087c98