Fix GN action target writing with no sources list.
Properly write outputs of actions that have no sources list. Previously this would lead to empty output lists in a build rule because the writer code was trying to apply the sources list to the output substitutions list. However, for actions, the output is the literal file list and should have no substitutions.

R=jamesr@chromium.org

Review URL: https://codereview.chromium.org/461153002

Cr-Commit-Position: refs/heads/master@{#289109}
git-svn-id: svn://svn.chromium.org/chrome/trunk/src@289109 0039d316-1c4b-4281-b951-d872f2087c98