Fix failing (flaky) LayerTreeHostTestLCDNotification test.
This test failed with impl-side painting because PictureLayer did not
skip commits caused by invalidating the layer during Update.

Meanwhile, this CL has some changes that changed the flaky failures
into always-failures. The change is to make the
ThreadProxy::CommitPendingForTesting check not only if a main frame is
in progress, but also if one will happen in the future. The failure was
flaky because the commit would be requested but not happen immediately
when impl-side painting was on due to activation (if the machine was
suitably loaded at the time). I renamed CommitPendingForTesting to
MainFrameWillHappenForTesting because "CommitPending" is a specific
notion in the public API of the scheduler and I didn't want to confuse
these two.

R=ajuma, brianderson, enne
BUG=402449, 397120

Review URL: https://codereview.chromium.org/462803002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@289165 0039d316-1c4b-4281-b951-d872f2087c98