Change threading in ServiceWorkerCacheStorage
ServiceWorkerCacheStorage currently expects to have its public
functions run by a SequencedTaskRunner.

This CL changes the functions to run on the calling thread and only
uses the TaskRunner for blocking operations.  This way, when calling
the cache and blob storage, we're on the calling thread (IO).

Also added some comments along the way.

BUG=392621

Review URL: https://codereview.chromium.org/460683002

Cr-Commit-Position: refs/heads/master@{#289186}
git-svn-id: svn://svn.chromium.org/chrome/trunk/src@289186 0039d316-1c4b-4281-b951-d872f2087c98