Fix crasher in the pdf tear down code path.
The crash occurs because of a timer dereferencing a null engine pointer.
The crash appears to have occurred due to this patch https://codereview.chromium.org/427583003/

Fix is to cancel all timers when the pdf instance is being torn down before
destroying the engine.

BUG=403036

Review URL: https://codereview.chromium.org/462283002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@289218 0039d316-1c4b-4281-b951-d872f2087c98