Call OnNativeWindowChanged after maximizing and restoring.
OnNativeWindowChanged is normally called when a window's bounds change.
However, if an app maximizes via the AppWindow API, but its bounds are
the same as that of a maximized window, OnNativeWindowChanged does not
end up being called.

This CL ensures that OnNativeWindowChanged is called whenever the app
maximizes or restores.

BUG=395900

Review URL: https://codereview.chromium.org/434343004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@289246 0039d316-1c4b-4281-b951-d872f2087c98