telemetry: Pass systrace categories properly
Pass systrace categories to chrome_profiler as a comma separated single
argument instead of one argument per category. This fixes a bug where
only the first category was actually enabled.

We also set the output format to JSON because it's not necessary to wrap
the resulting systrace into a HTML container as we end up zipping the
result anyway.

TEST=tools/perf/run_benchmark smoothness.key_mobile_sites --browser=android-chrome --page-filter=linked --profiler=android-systrace

Review URL: https://codereview.chromium.org/457823004

Cr-Commit-Position: refs/heads/master@{#289277}
git-svn-id: svn://svn.chromium.org/chrome/trunk/src@289277 0039d316-1c4b-4281-b951-d872f2087c98