Correct browser side audio effect state for mediastream requests.
The problem was that MediaStreamDispatcher::IsAudioDuckingActive could
return true incorrectly (i.e. when an input device was open but it was
not actually opened with ducking enabled - as specified by the application).

The reason for this is that the audio effect state as known by
MediaStreamDispatcher reflects what the browser side delivers when an
audio track has been opened but the browser side had not taken into
account the stream options of the request.

BUG=391414

Review URL: https://codereview.chromium.org/464923002

Cr-Commit-Position: refs/heads/master@{#289286}
git-svn-id: svn://svn.chromium.org/chrome/trunk/src@289286 0039d316-1c4b-4281-b951-d872f2087c98