Remove serial test support from test-webkitpy.
Previously test-webkitpy had a way to annotate tests that could not
be run in parallel with other tests. This was useful in rare occasions,
but such tests usually are flaky and slow, so they were gradually removed
and/or reworked. Ultimately this feature wasn't really being used and
was just a source of needless complexity, so I'm killing it.

R=eseidel@chromium.org
BUG=402172

Review URL: https://codereview.chromium.org/458643003

git-svn-id: svn://svn.chromium.org/blink/trunk@180217 bbb929c8-8fbe-4397-9dbb-9b2b20218538