AudioNode::breakConnectionWithLock() doesn't need to check 'm_normalRefCount > 1' in non-oilpan builds
This check was introduced in https://codereview.chromium.org/386403002/.
I agree that it's OK to have the check, but the check is not needed.
If we have the check, we can avoid the overhead of calling disableOutputsIfNecessary()
when the AudioNode is about to die. However, having the check causes an issue
when rewriting the reference counting system (see full CL: https://codereview.chromium.org/438293003/),
I remove the check in this CL in preparation for the rewrite.

BUG=340522

Review URL: https://codereview.chromium.org/466763002

git-svn-id: svn://svn.chromium.org/blink/trunk@180233 bbb929c8-8fbe-4397-9dbb-9b2b20218538