Disable uploading telemetry results temporarily.
For some reason the test-results server is crashing when processing
the uploaded JSON file. I'm not sure if the file format is invalid,
or if we're hitting some other problem on the server. However, if
we get a 500 back from the server, we fail the test run, and that is bad,
so for now, we'll disable the upload.

TBR=ojan@chromium.org, tonyg@chromium.org
BUG=403663

Review URL: https://codereview.chromium.org/470823003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@289471 0039d316-1c4b-4281-b951-d872f2087c98