[Resources] Use the correct orientation for avatars in the new-avatar-menu world.
We are planning on turning on --enable-new-avatar-menu by default in M38. In this
brave new world, the avatar button is always in the top right corner of the browser,
for all platforms. Previously, on Windows this button lived on the left side.

To make sure that the new world is the one that looks the best, we should always
use the correct orientation for the avatars. This does mean that for people
with the flag turned off, that are using the old style avatar menu on Windows,
they will see the avatars mirror.

BUG=380611
TEST=Start Chrome on Windows with --enable-new-avatar-menu. The avatar in the
avatar bubble should have the same orientation as in chrome://settings.

Review URL: https://codereview.chromium.org/464603002

Cr-Commit-Position: refs/heads/master@{#289522}
git-svn-id: svn://svn.chromium.org/chrome/trunk/src@289522 0039d316-1c4b-4281-b951-d872f2087c98