Accept invalid Schemas for the managed storage API.
Extensions won't get installed if they declare a managed storage schema
and the schema is invalid. But if the schema is modified or deleted
after the extension is installed then this CHECK will be hit whenever
Chrome restarts.

If the schema is modified or becomes invalid then the managed storage
API won't pickup the right policies for the extension, but that's
better than crashing the browser.

BUG=399208

Review URL: https://codereview.chromium.org/464893002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@289623 0039d316-1c4b-4281-b951-d872f2087c98