Check the Argument to the Catch Function
Due to promise chaining, the rejecting promise may not necessarily be the
topmost method returning the promise.

In this specific case, getAuth() is chained in by requestFromServer, and
getAuth can reject with no arguments.

The fix here is to ensure the argument is always an object.

BUG=402729

Review URL: https://codereview.chromium.org/469943002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@289668 0039d316-1c4b-4281-b951-d872f2087c98