Remove dependency on NavigationEntry to get a SiteInstance
Remove the dependency on a NavigationEntry in
RenderFrameHostManager::ShouldSwapBrowsingInstancesForNavigation and
RenderFrameHostManager::GetSiteInstanceForEntry. This is needed by the
PlzNavigate project, as the SiteInstance to use is determined when the
navigation is about to commit. At this time, there is no convinient access to a
SNavigationEntry.

BUG=376082

Review URL: https://codereview.chromium.org/471603002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@289681 0039d316-1c4b-4281-b951-d872f2087c98