sync: Attempt to fix sync scheduler test flakiness
Increases the delay duration in several sync scheduler unit tests.
According to [1], the timer resolution on Windows may be under 1ms.  We
suspect the low timer resolution compbined with very short delays could
cause problems in some unit tests.  This will make the tests slower, but
hopefully it will make them more reliable, too.

Changes a timestmap comparison in nudge_tracker.cc from < to <=.  We
suspect that a lower timer resolution might allow two Time::Now() calls
that should be separated by a >1ms gap of time might still return the
same value in both calls.  If this did happen, it could be the cause of
misbehaving unit tests.

It is believed that these issues only affect tests.  The delay lengths
used in the real world should be much longer than 1ms.  The timer
resolution should not be an issue for delays of that length.

[1] http://www.chromium.org/developers/design-documents/threading

BUG=402212

Review URL: https://codereview.chromium.org/463563005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@289731 0039d316-1c4b-4281-b951-d872f2087c98