Fix Mac sandbox meta data access (reland)
We currently allow all metadata access due to
https://codereview.chromium.org/10539009/ made the for loop comparison
in Sandbox::AllowMetadataForPath() always false, when we actually only
want to allow access to the path and all its parent path until root.

Turn the for loop to a do/while loop instead as it's a better fit, also
add a test case for Sandbox::AllowMetadataForPath().

We also need file read meta data access to all the .dylibs we linked to.

It should only affect component builds on OS X 10.6 and utility process
as no other process is using this mechanism.

BUG=403801

Review URL: https://codereview.chromium.org/469383002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@289738 0039d316-1c4b-4281-b951-d872f2087c98