Apply the text autosizing multiplier in FontBuilder::updateComputedSize.
When we resolve style for an element with an explicit font size, we build a new
FontDescription but we inherit the text autosizing multiplier from the parent's
RenderStyle (see http://crbug.com/380903).

We need to respect this multiplier when we initialize the computed size in the
FontDescription, or else they are out of sync with each other.  They remain out
of sync even after we layout the element, because TextAutosizer sees that the
element already has the correct multiplier and doesn't touch it.

BUG=403132

Review URL: https://codereview.chromium.org/472893004

git-svn-id: svn://svn.chromium.org/blink/trunk@180322 bbb929c8-8fbe-4397-9dbb-9b2b20218538