Use named variables for arguments to SyncManager::ConfigureSyncer.
Evaluate arguments to ConfigureSyncer outside of the function call to
simplify stack trace so that when this call appears in crash dumps, it's
clear that the crash is related to the call and not argument evaluation
(or vice versa).

Add DCHECKs to verify assumptions about arguments and object state.

BUG=402336

Review URL: https://codereview.chromium.org/469283003

Cr-Commit-Position: refs/heads/master@{#289767}
git-svn-id: svn://svn.chromium.org/chrome/trunk/src@289767 0039d316-1c4b-4281-b951-d872f2087c98