Hack to ensure that menu separators display correctly on Windows with fractional scales.
For more context please refer to the following bug https://code.google.com/p/chromium/issues/detail?id=403955

Reason for this bug is our familiar friend scaling as in views not landing on physical pixel boundaries
when the display scale on the machine is a fractional scale like 1.25 / 1.5 etc.

Hack for now is to move the separator 1 pixel down if the scale is fractional to ensure that it plays nice
with the clip rect which is rounded up.

BUG=402818

Review URL: https://codereview.chromium.org/473943003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@289806 0039d316-1c4b-4281-b951-d872f2087c98