gn yasm: use root_build_dir instead of . for rebase_path
I think these should be relative to the root_build_dir not the BUILD.gn
dir (which is what "." is right?). It was working before if you
happen to use (say) out\Default as the GN directory, but I've been using
out_gn so there was an extra .. in the path.

Also, a missing size_t->int warning disable on Windows.

R=brettw@chromium.org
BUG=354261

Review URL: https://codereview.chromium.org/477483002

Cr-Commit-Position: refs/heads/master@{#289813}
git-svn-id: svn://svn.chromium.org/chrome/trunk/src@289813 0039d316-1c4b-4281-b951-d872f2087c98