Allow app list tiles to show search results in the experimental app list.
This CL makes app results show as tiles in the experimental app list by
making TileItemViews take SearchResults instead of AppListItems and by
introducing a DisplayType for SearchResults which determines which
surface the result will display on.

Only 4 tiles appear at any one time. This will be fixed in a future
patch.

This CL also has the side effect of fixing a focus bug which was
happening because StartPageView was listening to the AppListModel.

BUG=398801, 349727

Review URL: https://codereview.chromium.org/439703002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@289826 0039d316-1c4b-4281-b951-d872f2087c98