w3c deps roll for layout tests: copy stale expectations downstream.
This copies any TestExpectations that refer to tests that no longer exist
or no longer fail at tip-of-tree in the w3c repos from Blink to Chromium,
so that we can safely roll to the new revisions.

This is patch 2 of 6 in the roll.

TBR=phoglund@chromium.org
BUG=404022

Review URL: https://codereview.chromium.org/469993004

Cr-Commit-Position: refs/heads/master@{#289885}
git-svn-id: svn://svn.chromium.org/chrome/trunk/src@289885 0039d316-1c4b-4281-b951-d872f2087c98