Show no information on the preview panel when no item is selected.
In the updated UI, path information is moved from footer to toolbar, and only information about selected entries is displayed on the footer.
In open dialog, there is a case that the preview panel is visible and no item is selected.
This patch hide everything (thumbnail and entry name) from the footer when no item is selected.

BUG=402821
TEST=manually

Review URL: https://codereview.chromium.org/468913005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@289934 0039d316-1c4b-4281-b951-d872f2087c98