Log NTP loading time separately for local, Google and other NTPs.
We currently use the same UMA metric to log the loading time of different kinds of NTP:
- The local NTP
- The Google app-cached NTP
- NTP from other providers

This CL separates these in three UMA metric using histogram suffixes.

In addition, this CL brings back the Tab.NewTab* UMA metrics to the public histograms.xml in accordance with the practice that metrics in the public chromium repository should be in the public histograms.xml.

Review URL: https://codereview.chromium.org/474053002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@289946 0039d316-1c4b-4281-b951-d872f2087c98