cc: Add more eviction categories to picture layer impl.
Since we use priority_for_tree_priority to get the priority_bin,
we might end up in situations where we have a non-now priority_bin
with a tile that is required for activation (this is because
required for activation property is read off the pending priority).

In particular in smoothness takes priority, if a tile is only
available on the pending tree, then it will have EVENTUALLY
priority_bin with inf distance. However, at the same time
it can be required for activation (NOW on the active tree).

We need to ensure that this is handled explicitly and correctly.
This patch adds the extra required categories, as well as
makes an array instead of separate variables for each of
the categories.

R=reveman

Review URL: https://codereview.chromium.org/471833002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@290011 0039d316-1c4b-4281-b951-d872f2087c98