Use common converter for -webkit-column-[gap, width].
This patch adds support for the 'converter' attribute for a group of
semi-custom CSS properties handled by the 'apply_auto' macro. This gets
rid of the redundant 'compute_length=true/false' argument, plus the code
related to that.

It also makes it easier to add other things to the apply_auto-list which
can't be converted with computeLength or a primitive value mapping.

Review URL: https://codereview.chromium.org/454133004

git-svn-id: svn://svn.chromium.org/blink/trunk@180437 bbb929c8-8fbe-4397-9dbb-9b2b20218538