Fixes WebSettingDialog addDescriptionSection assertion causing crash.
This addresses two problems: a crash due to an assertion failure in
addDescriptionSection in WebSettingDialog for Android and two nearly identical
description sections appearing in the page info on Android. The root cause of
this was a bad rebase while working on
https://codereview.chromium.org/473643002/ where the double call to
addDescriptionSection was added. This CL simply removes the redundant/wrong
second call to addDecriptionSection.

BUG=404293
TBR=ianwen@chromium.org,tedchoc@chromium.org

Review URL: https://codereview.chromium.org/482063002

Cr-Commit-Position: refs/heads/master@{#290356}
git-svn-id: svn://svn.chromium.org/chrome/trunk/src@290356 0039d316-1c4b-4281-b951-d872f2087c98