Remove inaccurate comment from isSizeAvailable
The comment attached to ImageDecoder::isSizeAvailable() tends to confuse
readers (see the bug). No lazy-decoding is performed so words about that
are incorrect. The refactor suggested was from another time, since today
some decoder's decode(), others parse(), while others updateDemuxer(): a
lesson in comments falling out-of-date. It is not clear if the suggested
refactoring buys us anything but confusion, so ditch the FIXME.

BUG=404544

Review URL: https://codereview.chromium.org/484983002

git-svn-id: svn://svn.chromium.org/blink/trunk@180531 bbb929c8-8fbe-4397-9dbb-9b2b20218538