[Telemetry] Add a shorter timeout for Request().
This improves telemetry_unittests cycle time by 15% (7min -> 6min)
on my windows z600.

The issue here had to do with retries. If the browser wasn't up yet, the
socket wouldn't try to reconnect for 1s. So instead, we set a low timeout
so that we'll keep retrying the connection instead of waiting for that
long delay.

BUG=388256

Review URL: https://codereview.chromium.org/485103004

Cr-Commit-Position: refs/heads/master@{#290634}
git-svn-id: svn://svn.chromium.org/chrome/trunk/src@290634 0039d316-1c4b-4281-b951-d872f2087c98