Improve testing for SDCH.
Added some end-to-end browser tests, and added a bit of extra checking
for the filter.

The extra files for encoding dictionaries from sdch/open-vcdiff were
added so that the tests could create an SDCH dictionary when run.  
IMO, this is more flexible for future tests and clearer to read and 
maintain than just hard-coding a dictionary inline in the tests.

BUG=None
R=mef@chromium.org

Review URL: https://codereview.chromium.org/380003002

Cr-Commit-Position: refs/heads/master@{#290638}
git-svn-id: svn://svn.chromium.org/chrome/trunk/src@290638 0039d316-1c4b-4281-b951-d872f2087c98