Page info not closing on non-Mac desktop when close button icon clicked.
This fixes the page info close icon so when it is clicked, it closes the dialog
properly. This was a regression introduced by
https://codereview.chromium.org/473643002/ where the widget's Close() method was
accidentally moved inside of a conditional in the button handler when it really
should have been executed on all button presses. This moves the Close() method
outside the conditional so it properly closes the dialog.

BUG=404963
TBR=pkasting@chromium.org

Review URL: https://codereview.chromium.org/485723004

Cr-Commit-Position: refs/heads/master@{#290659}
git-svn-id: svn://svn.chromium.org/chrome/trunk/src@290659 0039d316-1c4b-4281-b951-d872f2087c98