NaCl: Send fatal log messages via shared memory.
nacl_helper reports the last few log messages to the renderer when a LOG_FATAL
message is logged in the plugin. This currently uses the "bootstrap" channel,
which we would like to get rid of.

This uses a shared memory segment instead of IPC, since something bad may have
happened in the plugin process, so we'd like to keep the behavior for handling
that case as simple as possible. For that reason, the shared memory segment is
mapped before the nexe is loaded.

I've tested this locally by calling RemoteLog(LOG_FATAL, ...) after StartModule
in ServiceRuntime, though I'd like to find some better way to test this as
there's a decent amount of logic here.

BUG=391039
R=jschuh@chromium.org, mseaborn@chromium.org

Review URL: https://codereview.chromium.org/469423002

Cr-Commit-Position: refs/heads/master@{#290910}
git-svn-id: svn://svn.chromium.org/chrome/trunk/src@290910 0039d316-1c4b-4281-b951-d872f2087c98