cc: Avoid redraw for missing tile outside visible rect
Tiles outside of the visible rect for tile priority but
inside the draw viewport should be drawn on a best effort
basis. There is no need to redraw or block activations on
missing or incomplete tiles in this region.

Rename tile priority rect/matrix to activation rect, and
use this to control activation as well. Add new counts for
missing/incomplete tiles in the activation rect, and only
these tiles will lead to redraws.

BUG=403982
NOTRY=true

Review URL: https://codereview.chromium.org/475233002

Cr-Commit-Position: refs/heads/master@{#290937}
git-svn-id: svn://svn.chromium.org/chrome/trunk/src@290937 0039d316-1c4b-4281-b951-d872f2087c98