Add Chart JSON processing to Telemetry.
This patch implements Chart JSON for Telemetry, which represents a JSON format specifically for conveying data to Buildbot to be sent to the perf dashboard in charted/summarized form. 
Consult go/telemetry-json for further specs on this format.
This is an 0.1 since the overall metadata and description portions of the format are still in flux.

BUG=399970

Review URL: https://codereview.chromium.org/439613003

Cr-Commit-Position: refs/heads/master@{#291185}
git-svn-id: svn://svn.chromium.org/chrome/trunk/src@291185 0039d316-1c4b-4281-b951-d872f2087c98