Pepper: Make pending_sync_msgs_ local to IO thread.
Adding support for PpapiHostMessage_OpenResource to SFI mode is tricky.
Validating the file tokens requires that a message to be sent to the browser
(!) to validate that information in the case of a compromised renderer.

This change is one step to allowing processing for the reply of the
PpapiHostMessage_OpenResource message to send a message to the browser for
constructing a NaClDesc properly.

This makes the encapsulation of NaClMessageScanner a bit worse, but also
reduces the amount of code for which the critical section under lock_ must be
held.

BUG=394130

Review URL: https://codereview.chromium.org/472073003

Cr-Commit-Position: refs/heads/master@{#291234}
git-svn-id: svn://svn.chromium.org/chrome/trunk/src@291234 0039d316-1c4b-4281-b951-d872f2087c98