Fix the image decode telemetry perf unittest on impl side painting.
This test was enabling a lot of categories that it should not have.
When using impl side painting, we can use the trace events from
blink.ImageFrameGenerator::decode to measure the time. On legacy
compositing, this doesn't exist and we have to fall back to the
"timeline.Decode Image" event still. Left FIXMEs to remove this.

BUG=405244
NOTRY=true

Review URL: https://codereview.chromium.org/480713004

Cr-Commit-Position: refs/heads/master@{#291304}
git-svn-id: svn://svn.chromium.org/chrome/trunk/src@291304 0039d316-1c4b-4281-b951-d872f2087c98