Hide PopupZoomer when the container view or window loses focus
Currently there is a method which hides all the popups but call to 
hide PopupZoomer is missing. Added the call to hide PopupZoomer inside 
hidePopups and hiding the zoomer when the view loses focus or the 
window loses focus. Also added unit test code.

BUG=405477

Review URL: https://codereview.chromium.org/470193005

Cr-Commit-Position: refs/heads/master@{#291359}
git-svn-id: svn://svn.chromium.org/chrome/trunk/src@291359 0039d316-1c4b-4281-b951-d872f2087c98