Files.app: Update the list partially when changed
This patch makes the list redraw partially instead of full-redraw when some contents are changed in the directory. This should improve the performance of redraw.

In addition, as the result of the fix, the cache is not flashed on partial redraw, and the issue 374713 has gone.

BUG=374713
TEST=manaul tested

Review URL: https://codereview.chromium.org/486783006

Cr-Commit-Position: refs/heads/master@{#291366}
git-svn-id: svn://svn.chromium.org/chrome/trunk/src@291366 0039d316-1c4b-4281-b951-d872f2087c98