Expose a pair of points for selection bound edge coordinates
Using a rect to indicate the logical selection bound edge is confusing,
particular if the selection bound is not axis aligned with respect to its
compositing layer. Instead, expose a top and bottom coordinate pair for the
logical edge of the selection bound.

The corresponding Chromium change is https://codereview.chromium.org/494823002.

BUG=405666

Review URL: https://codereview.chromium.org/495673003

git-svn-id: svn://svn.chromium.org/blink/trunk@180773 bbb929c8-8fbe-4397-9dbb-9b2b20218538