cc: Don't infinitely throttle large raster tasks
Previously, if a raster task had a resource size that was larger than
the throttling limit for the raster worker, that task would never get
scheduled.  This leads to starvation for that task, preventing tree
activation, and causing freezes.

The fix is to always allow such large tasks if it is the first task.

R=reveman@chromium.org
BUG=403446

Review URL: https://codereview.chromium.org/489293002

Cr-Commit-Position: refs/heads/master@{#291484}
git-svn-id: svn://svn.chromium.org/chrome/trunk/src@291484 0039d316-1c4b-4281-b951-d872f2087c98