Componentize component_updater: Copy over test data with executable bit.
As part of making component_updater a component, we are now moving over all the tests.

Before we do that, we're copying the test data over. We are deleting it from the original spot in a separate CL.

This patch will need to be dcommitted, since component1.dll file is executable bit, and CQ chokes on that.

Review of this change is here: https://codereview.chromium.org/497803002/

BUG=371463
R=waffles@chromium.org

Review URL: https://codereview.chromium.org/499633002

Cr-Commit-Position: refs/heads/master@{#291490}
git-svn-id: svn://svn.chromium.org/chrome/trunk/src@291490 0039d316-1c4b-4281-b951-d872f2087c98