About page should not automatically trigger OS updates
Currently, if the user opens the Help/About page on Chrome OS it will
automatically check for an update and begin downloading/installing it.
Users should be able to decide when their update is applied, e.g., in
case they are connected to a tethered or bandwidth-limited network.

This disables the auto-update check on Chrome OS, replacing it with a
button the user can press to check for updates. (This has no effect on
the scheduled auto-updating service.)

Screenshots: http://imgur.com/a/k2s6y#0

BUG=394226
R=estade@chromium.org, deymo@chromium.org

Review URL: https://codereview.chromium.org/476983002

Cr-Commit-Position: refs/heads/master@{#291580}