Specify clip rects when drawing Mac native widgets
Add the clip rect as an argument to LocalCurrentGraphicsContext,
and specify the rect for all callers. Add the rect argument for the one
direct caller to SkiaBitLocker (the argument will be made a requirement
once this lands).

De-duplicate a bunch of rectangle-inflating code and move it into
ThemeMac.

For the selected widgets that draw their own focus rings (most don't),
simply specify the maximum inflation of the clip rect as 16 pixels on
each side. This appears to be very generous. The correct API call to
compute this, focusRingMaskBoundsForFrame, was introduced in the
10.7 API, and the widgets that draw their own focus rings wouldn't
need to do so if written against the the 10.7 API.

BUG=247716

Review URL: https://codereview.chromium.org/488353003

git-svn-id: svn://svn.chromium.org/blink/trunk@180822 bbb929c8-8fbe-4397-9dbb-9b2b20218538