Drop 64 character limit in DeviceManagementRequest |agent| parameter
The HTTP POST |agent| parameter sent with a DeviceManagementRequest is
limited to 64 characters. After the git migration, Chrome includes its
git revision in the |agent|, making it longer than 64 characters.

This CL removes the arbitrary 64 character limit. This solution should be
considered temporary, until we have verified that the DMServer can handle
|agents| longer than 64 characters. If not, we will restore this limit and
change the construction of the string so that it never exceeds 64
characters.

BUG=406948
TEST=Browser tests no longer fail when the |agent| string is too long

Review URL: https://codereview.chromium.org/500013002

Cr-Commit-Position: refs/heads/master@{#291672}