Mojo C++ bindings: better log messages for some validation errors at the receiver side.
The CL provides more information in the log messages for the following errors:
- Fixed-size array has wrong number of elements.
- non-nullable field is set to null/invalid handle.

The reason to have better log messages for these errors is that they are
triggered when users (of the bindings) provide invalid arguments. The users
probably want to find out what they have done wrong. The other validation errors
are rarer, caused by bugs in the bindings or malicious senders. Providing better
log messages for them is of slightly lower priority.

BUG=324170
TEST=None

Review URL: https://codereview.chromium.org/494943004

Cr-Commit-Position: refs/heads/master@{#291704}