Separate menu & disconnect buttons from the window controls and clean up CSS.
One piece of feedback we've received is that users want an
explicit full-screen control rather than tying it to
maximize. Simply adding a full-screen button means that the
window has six icons of varying crypticness (crypticality?)

This CL prepares the ground for adding a full-screen button
while reducing clutter. It moves the Options and Disconnect
buttons to the left-hand side of the window, while leaving
them adjacent to the window controls in full-screen mode.

There are two main changes to the CSS. The first is that we
no longer use "display: table"; I thought this was needed to
get rid of the implicit spaces between the icons, but
"display: flex" does the job just as well. The second is to
do with how the etched borders are rendered. Previously, a
button was responsible for rending both the light and dark
highlights to the left and bottom. This worked because they
were all right-aligned, but with the new layout, each button
renders all of its borders.

Review URL: https://codereview.chromium.org/498813003

Cr-Commit-Position: refs/heads/master@{#291745}