Add an explicit full-screen button.
We've recevied feedback that conflating maximized and full-screen is not what
users expect. This CL builds on https://codereview.chromium.org/498813003/ to
remove this functionality and instead add an explicit full-screen button to
the window frame.

Note that the icon is a placeholder until we can get UX resources to design a
better one.

BUG=404693

Review URL: https://codereview.chromium.org/491873004

Cr-Commit-Position: refs/heads/master@{#291811}