Open embedded extension options from the extension action context menu
This CL allows the extension action context menu to use the new embedded
extension options feature. If the enable-embedded-extension-options flag
is enabled, instead of opening an extension's options page in it's own
tab, it will open chrome://extensions, scroll down to the correct
extension, and then open the embedded options page popup.

The context menu opens a link to chrome://extensions/ with the query
string ?options=|extensionId|. This query string is appended and
removed whenever the embedded extension options popup is opened or
closed.

BUG=386842

Review URL: https://codereview.chromium.org/484033003

Cr-Commit-Position: refs/heads/master@{#291826}