Support middle-clicking on the back button to open in a new tab.
Check for left mouse button in onBack(), and let the browser take care
of middle clicks.

This also has the effect of properly showing a hover icon over "Go back"
(due to the presence of the href attribute on the <a> tag).

R=ojan@chromium.org
BUG=400414
NOTRY=true

Review URL: https://codereview.chromium.org/513543002

git-svn-id: svn://svn.chromium.org/blink/trunk@180926 bbb929c8-8fbe-4397-9dbb-9b2b20218538