BrowserPlugin: Simplify guest access
Now that it's impossible for content to create guests and it's impossible
for a compromised embedder to access a guest it does not have permission to
access from content, BrowserPluginGuestManager::MaybeGetGuestByInstanceIDOrKill
seems a bit heavy-handed. This CL simplifies the API a bit.

This is also a prerequiste to support accessibility in BrowserPlugin.

BUG=226145

Review URL: https://codereview.chromium.org/511643002

Cr-Commit-Position: refs/heads/master@{#292060}