cc: Report only on active tiles in tracing.
This patch changes which tiles we report on. Previously, we logged
all of the available tiles. Now, we only log tiles that are active.
That is, only tiles that are on either active or pending trees are
logged.

The way this is done is as follows:
- Walk the tree, layers, tilings
- Insert all tiles from a tiling into a set
- Stuff those into a trace.

BUG=403094

Review URL: https://codereview.chromium.org/493543002

Cr-Commit-Position: refs/heads/master@{#292183}