Properly recompute SVG font size under zoom
This patch fixes a regression from [1] where an optimization in
RenderSVGInlineText::computeNewScaledFontForStyle would skip updating
the font size on lowdpi devices while zooming. The computed font size
calculation depends on zoom and the font scaling factor depends on the
device's scaling factor. Previously, we would skip updating the font
when the scaling factor was 1 which would fail to account for zoom.
This patch updates the optimization to only be used when both the
scaling factor and zoom are 1.

[1] http://src.chromium.org/viewvc/blink?revision=153683&view=revision

BUG=405468

Review URL: https://codereview.chromium.org/508823002

git-svn-id: svn://svn.chromium.org/blink/trunk@180995 bbb929c8-8fbe-4397-9dbb-9b2b20218538