Updated media queries boolean context evaluation to match spec change.
The wording of the spec was recently changed to make evaluation of
features that can be true for multiple values more intuitive in a
boolean context. In practice, this means that the any-hover and
any-pointer features will evaluate to true as long as there is a non-
none available pointer/hover. The previous wording meant that if 'none'
was in the available list it would evaluate to false.

See http://dev.w3.org/csswg/mediaqueries-4/#mq-boolean-context and
http://lists.w3.org/Archives/Public/www-style/2014Aug/0190.html for
details.

BUG=136119

Review URL: https://codereview.chromium.org/508413002

git-svn-id: svn://svn.chromium.org/blink/trunk@181035 bbb929c8-8fbe-4397-9dbb-9b2b20218538