Clean up setting accessibility focus.
Consolidate the code to set accessibility focus to one place.
Fix it so we never set accessibility focus to the same id twice in a row,
as that puts things in a weird state (because we then get a "clear"
accessibility focus action on the current focused view).

In a couple of places where we want to explicitly update the focus,
set focus away and then back again so Android can't ignore it.

BUG=407871

Review URL: https://codereview.chromium.org/513483002

Cr-Commit-Position: refs/heads/master@{#292386}