Mac: Update sheet position when bookmark bar is shown or hidden.
In r272421, I changed sheet repositioning on window resize to be based on
NSWindowDidResizeNotification instead of NSViewFrameDidChangeNotification
because it tracks window resizing much more smoothly.

That caused this problem where if the browser window is vertically maximized
(or is in fullscreen), the window size does not change when the bookmark bar
is shown or hidden.

The old code has a different bug however. If the browser window is not
vertically maximized, showing or hiding the bookmark bar does not change the
browser content view's frame. It moves up or down to follow the window resize
but its frame relative to its superview remains the same.

I settled for adding code to explicitly update the sheet's position when the
bookmark bar is shown or hidden.

BUG=407509

Review URL: https://codereview.chromium.org/502413004

Cr-Commit-Position: refs/heads/master@{#292421}