Remove BrowserThread dependency from usb_service.
Instead of explicitly depending on specific browser threads the USB
service can assume that it is instantiated on BrowserThread::FILE (or
equivalent) and save a TaskRunner reference from this instantiation for
later use when called from other threads.

To reach BrowserThread::UI (required for DBus on Chrome OS) a reference
to the appropriate TaskRunner must be provided when calling
UsbService::GetInstance().

BUG=

Review URL: https://codereview.chromium.org/507503002

Cr-Commit-Position: refs/heads/master@{#292546}