[NaCl SDK] Allow ppapi_simple executables to run in both sel_ldr and in chrome.
Override the main() function from libppapi_stub, and rather than
simply failing when PPAPI is missing, assume that we are running
in sel_ldr and jump directly to the user's main function.

As a nice side effect this removes the needs to SEL_LDR=1 builds.

Review URL: https://codereview.chromium.org/412083002

Cr-Commit-Position: refs/heads/master@{#292710}