[fsp] Add support for providing thumbnails.
Previously, thumbnails were generated from full size images, which in case of
cloud providers would invoke downloading the entire file.

This CL adds an ability to provide a thumbnail as a data URI, so:
(1) We can have thumbnails for other file types than images.
(2) Thumbnails are shown fast.
(3) Resource consumption is significantly reduced.

Note, that the thumbnails are not wired to Files app yet. That will be done in
a separate patch.

TEST=unit_tests: *FileSystemProvider*GetMetadata*
BUG=407954

Review URL: https://codereview.chromium.org/513683002

Cr-Commit-Position: refs/heads/master@{#292840}