Tell synthetic attribute nodes inside shadow DOM about move to new Document
Node::didMoveToNewDocument() is called to notify a Node that it has
been moved under a new Document. This patch fixes a bug in
TreeScopeAdopter where this notification was not done for synthetic
attribute nodes inside a shadow DOM. The bug caused an issue where an
event handler registered for an attribute node was not correctly
unregistered when the parent element was moved to a new document.

BUG=407477

Review URL: https://codereview.chromium.org/519003003

git-svn-id: svn://svn.chromium.org/blink/trunk@181302 bbb929c8-8fbe-4397-9dbb-9b2b20218538