Uncouple ActiveScriptController from LocationBarController
Since the indication of an active script running won't always be tied to the
location bar, it doesn't make sense for ActiveScriptController to be owned by
LocationBarController.

Since the "generated" actions were also only used for display
in the location bar, move that logic to the LocationBarController.

And, since we're there, also add a bit more testing for LocationBarController.

BUG=408676
TBR=avi@chromium.org (for method name change in cocoa)

Review URL: https://codereview.chromium.org/527963003

Cr-Commit-Position: refs/heads/master@{#293153}