Raw SocketDescriptor variant of UnixDomainServerSocket::Accept
The Mojo code on CrOS needs to accept inbound connections
on a unix domain socket, and then 'promote' the resulting
sockets to Mojo MessagePipes. This really requires access
to the underying file descriptor, so provide a mechanism
to accept a connection and get back a SocketDescriptor.

BUG=407782
TEST=UnixDomain*SocketTest
R=mmenke@chromium.org

Review URL: https://codereview.chromium.org/509133002

Cr-Commit-Position: refs/heads/master@{#293172}