Oilpan: Prompt freeing and coalescing of collection backings.
A promptly-freed object is finalized immediately and marked as
free/dead.  However, it is not added the memory freelist and its memory
is still considered to be used, and thus the prompt freeing will not
immediately have an impact on when the next GC occurs. This means that
prompt freeing cannot by itself result in a higher memory pressure.

When an allocation occurs which would otherwise result in allocating a
new memory page, we try to coalesce promptly-freed entries and use the
reclaimed space instead. Only once we have coalesced the entries of a
page will the space be added to the freelist and considered free.

Coalescing a page is not cheap, so we use various heuristics to avoid
unsuccessful coalescing:

 - Only coalesce if a certain number of objects have been promptly
   freed (currently 256).

 - Only coalesce a specific page if it contains a reasonable about of
   promptly freed space (currently 3 times the required entry size).

 - Only consider the most recently added pages because they are assumed
   more "active" (currently the first 100 pages).

R=ager@chromium.org
BUG=

Review URL: https://codereview.chromium.org/535523002

git-svn-id: svn://svn.chromium.org/blink/trunk@181337 bbb929c8-8fbe-4397-9dbb-9b2b20218538