Domain Reliability: Don't create Service if metrics reporting disabled
Given that the metrics reporting pref is labelled in the UI as requiring
a restart to apply, this is a much simpler way of gating Domain
Reliability uploads on it.

We simply don't create the Service (so we don't collect data and upload
it) if the pref is false.

This will need to be revisited when we also start using the Monitor for
JavaScript hooks (e.g. the Navigation Error Logging API).

BUG=407170

Review URL: https://codereview.chromium.org/525303004

Cr-Commit-Position: refs/heads/master@{#293226}