Rename HTMLMediaElement::closeCaptionTracksChanged to textTracksChanged
The name closeCaptionTracksChanged() is slightly misleading since this
method is called for all kinds of (text) tracks, and not just 'caption'
tracks (and the 'closed' needn't necessarily apply either...)
Rename to textTracksChanged() since that more accurately describe what it
does.
Also replace a few "manual" instances of textTracksChanged() with calls
to the method. In the case of HTMLMediaElement::didAddTrackElement, it
can be dropped altogether because addTextTrack already calls it (and
scheduleDelayedAction(LoadTextTrackResource) does not have side-effects
that would affect the result.)
Also do the same name-change for the closedCaptionTracksChanged() notifier
on MediaControls (same semantic meaning.)

Review URL: https://codereview.chromium.org/540853003

git-svn-id: svn://svn.chromium.org/blink/trunk@181401 bbb929c8-8fbe-4397-9dbb-9b2b20218538