cc: Be less aggressive about scheduling for scroll handlers
A previous patch changed the scheduling policy so that scroll handlers
get more of a chance of completing in the same frame as the scroll
update ("best effort sync scrolling"). It turns out this went a little
too far by fully disabling prefer smoothness mode in the presense of
scroll handlers. For instance, this caused us to start synchronously
waiting for asynchronous texture upload completion, leading to skipped
frames.

This patch adjusts the logic so that scroll handlers are still given an
opportunity to complete within the same frame (i.e., the deadline isn't
triggered early), but the tree activation policy is to prefer existing
content over new content.

We also rename the smoothness mode in the scheduler to "impl latency
takes priority" to make its purpose clearer.

BUG=404267

Review URL: https://codereview.chromium.org/522903002

Cr-Commit-Position: refs/heads/master@{#293369}