Do not paste clipboard contents when middle click on a bookmark in Bookmark Manager.
Middle click on a bookmark in Bookmark Manager should open the bookmark in new
tab without pasting the contents of paste buffer. In Linux, since middle click
is used as 'paste', prevent it from pasting if the clicked item is a bookmark.
If clicked elsewhere in the Bookmark Manager window, do the default behaviour.

BUG=44811

R=bauerb@chromium.org

TEST=
1) Add few bookmarks and open Bookmark Manager.
2) Copy one or more bookmark item(s).
3) Middle click on a bookmark item and observe.
4) The clicked bookmark should open in new tab without pasting the copied bookmark
   inside the window.

Review URL: https://codereview.chromium.org/527923002

Cr-Commit-Position: refs/heads/master@{#293382}