Respect window transparency when applying window shape.
Currently, the code clamps alpha up to to 100% within the window shape and
down to 0% outside the shape. This means that windows with shape *and*
transparency enabled effectively lose their transparency.

This change makes it so that the alpha is clamped up to 0% within the window
shape, which preserves the window's transparency.

BUG=405751

Review URL: https://codereview.chromium.org/499503002

Cr-Commit-Position: refs/heads/master@{#293433}