Fix copying of SAML IdP cookies on subsequent logins
The code that copies cookies set by a SAML IdP on subsequent logins is
wrong as it uses a method which modifies the cookie being copied by
prepending a dot to the domain name. This CL fixes the problem by
switching to the method that is used during initial login. The method
is also renamed to make it clearer that it may be used on subsequent
logins too.

BUG=410416
TEST=unit_tests updated to detect this problem

TBR=dbeam (wallet_signin_helper_unittest.cc)

Review URL: https://codereview.chromium.org/538543003

Cr-Commit-Position: refs/heads/master@{#293524}