Speculative fix for DelegatedFrameHost destructor crashes
It appears that RenderWidgetHostViewMac::Destroy is never getting
called in some instances, and that is resulting in crashes when we end
up in its destructor, which is called by Objective C's runtime.

I suspect that this is a leak somewhere, and that on other platforms we
simply never call the destructor.

Call the necessary bits of RenderWidgetHostViewMac::Destroy to avoid
crash when a RenderWidgetHostImpl reaches its destructor before its
view_ has been reset.

Also, turn the weak pointers to RenderWidgetHostViewBase in
RenderWidgetHostImpl and RenderWidgetHostViewGuest
to use base::WeakPtr to prevent the many dangling instances.

BUG=404828

Review URL: https://codereview.chromium.org/538883002

Cr-Commit-Position: refs/heads/master@{#293573}