[ServiceWorker] Don't pass OPTIONS request to the ServiceWorker
This is work around to avoid hijacking CORS preflight.
In current implementation, CORS preflight request is also sent to the ServiceWorker.
This is a bug in the implementation of HTTP fetch in DocumentThreadableLoader.
But it takes time to fix this.
So I want to introduce this work around because this is a security bug.

BUG=409454

Review URL: https://codereview.chromium.org/548773002

Cr-Commit-Position: refs/heads/master@{#293697}