Fix maximize on some window managers
On some window managers the Maximize hints are not respected when they
are set on an unmapped window. There was an existing workaround for this
in the code, but the workaround failed whenever ShowWindowWithState
didn't happen to be called with the parameters needed to activate the
workaround. This meant that if a window was first made visible and later
activated, it would not get maximized at all.

This fix saves whether the maximize hints need to be sent to the window,
and makes sure it gets done after mapping the window.

BUG=

Review URL: https://codereview.chromium.org/543663003

Cr-Commit-Position: refs/heads/master@{#293698}