Defer changes to the channel count mode to the pre or post-render phase
In some situations, the channel count mode is updated just before the
audio rendering starts. During audio processing, the channel count
mode is used to set the number of channels for processing.  However,
for AudioBasicProcessor's, the number of kernels has not yet been
updated to the correct number of channels, so the incorrect number of
channels is accessed.

We delay the setting of the channel count mode until the pre or post
rendering phase of the audio thread where it is safe to change things
since the audio graph is not being pulled.

BUG=407489

Review URL: https://codereview.chromium.org/543073004

git-svn-id: svn://svn.chromium.org/blink/trunk@181573 bbb929c8-8fbe-4397-9dbb-9b2b20218538