Don't try to access a NULL plugin element or document in PepperPluginInstanceImpl
There appear to be cases where the WebElement or Document associated with a plugin
have been destroyed but the PepperPluginInstance isn't yet destroyed and there is
some attempt to script into the plugin which tries to access the plugin context.
In these case we should just return an empty handle to the context.

TBR=dmichael@chromium.org
BUG=411778

Review URL: https://codereview.chromium.org/558513002

Cr-Commit-Position: refs/heads/master@{#293873}