ozone: Plumb animated cursors from BitmapCursorFactoryOzone to DriSurfaceFactory
Update BitmapCursorOzone type to store a vector of bitmaps, a hotspot,
and a frame delay. Plumb the vector through to the GPU process, where
we'll update the cursor on a timer. This is plumbing only; it does not
change the user-visible behavior of showing only the first cursor frame.

R=dnicoara, kenrb
BUG=343245
TEST=viewed css "wait" cursor on cros link_freon

Review URL: https://codereview.chromium.org/543643003

Cr-Commit-Position: refs/heads/master@{#294045}