Remove NOTREACHED verifying no previous exceptions are set in PepperTryCatch
When setting an exception we check that no other exceptions have been set.
However, sometimes it is preferable not to bail out of the
codepath immediately because the code becomes messier. In particular, if
the context is destroyed, an exception will be set which can happen upon
construction of the try/catch. Keeping this NOTREACHED means that we have to check
for an exception after the construction of every try/catch. I can't see a
problem with not checking the exception immediately - other operations will
fail and the result of the try-catch will eventually be set.

BUG=303491

Review URL: https://codereview.chromium.org/554173002

Cr-Commit-Position: refs/heads/master@{#294046}