Chrome OS Touchpad Settings: Always update natural scrolling pref.
Natural (aka australian) scrolling occurs outside Chrome in the X11
touchpad driver. Chrome dictates via an inputcontrol script what the
value (true/false) should be. The script notes (in the user's homedir)
what the last set value is so that added devices get it applied.

Unfortunately, what was happening was Chrome would update the pref
before the user logged in, so the inputcontrol script couldn't make a
note. Later, when a user was logged in and a new device was attached,
the script wouldn't know if australian scrolling should be on or off.

To fix, we always set australian scrolling whenever we consider
sending touchpad value changes to the inputcontrol script.

BUG=chromium:406480
TEST=Tested that when user logs in, chrome always sends the current
value of natural scrolling to the inputcontrol script. Also tested
that changing non-mouse/tp settings doesn't trigger calls to
inputcontrol script.

Review URL: https://codereview.chromium.org/553153004

Cr-Commit-Position: refs/heads/master@{#294058}