Stop using an atom to store plugin name/version on Windows.
On the swarming bots we are seeing flakes in the plugin tests. Setting the atom fails because the atom table fills up. I tracked this down to the fact that sometimes the atom destruction path doesn't get hit depending on how the plugin shuts down. Rather than fix it, I removed this code and used a simpler mechanism to get the title & version in the browser process.

BUG=412042

Review URL: https://codereview.chromium.org/557893003

Cr-Commit-Position: refs/heads/master@{#294088}