cc: Don't make tiles for mask layers that are too big for a texture.
If a mask layer's width or height are larger than the max texture size,
we do not use the contents of the mask layer. So if we're not going to
use the contents, there's no reason to make tiles and allocate memory
for them.

This changes PictureLayerImpl to return an empty tile size for masks
that are too large to have tiles. Then changes PictureLayerTiling to
use the empty tile size as a signal to make the whole tiling an empty
size.

Last, we use the empty tile size to also signal not making a low res
tiling, so that masks continue to not get a low res tiling if they are
large.

TBR=enne@chromium.org, vmpstr@chromium.org
BUG=409984

Review URL: https://codereview.chromium.org/547463002

Cr-Commit-Position: refs/heads/master@{#293656}
(cherry picked from commit 132fdfe17dd03c0066e27fe876602122c522aa8b)

Review URL: https://codereview.chromium.org/557083003

Cr-Commit-Position: refs/branch-heads/2125@{#304}
Cr-Branched-From: b68026d94bda36dd106a3d91a098719f952a9477-refs/heads/master@{#290040}