Files.app: Remove getDriveFiles function from the private API.
The API is for ensuring the files are downloaded in local storage, and wad used in two places in Files.app.

The first is mounting ZIP. We don't need to care about it. We can mount ZIP
without downloading it in advance.

The second is selecting files in the file select dialog. We can download files
when fileManagerPrivate.selectFiles(s), and UI shows the progress correctly. But
previously, if downloading is cancelled, the API is regarded as success and it
passes the selected file inforamtion to the browser. The CL handles the case and
has the API report an error if the downloading is cancelled.

BUG=326082
TEST=delete cache, open the file in the file select dialog. Also cancel uploading in the dialog.

Review URL: https://codereview.chromium.org/557223003

Cr-Commit-Position: refs/heads/master@{#294320}