Get rid of FileManagerPrivateAPI class
The whole purpose of FileManagerPrivateAPI class was to hold an
EventRouter instance. It's simpler to just make EventRouter a
KeyedService and get rid of FileManagerPrivateAPI.

Note that BrowserContextKeyedAPI could not be used for EventRouter as
EventRouter needs a Profile rather than a BrowserContext.

In the next patch, EventRouter will be renamed to something like
FileManagerService.

BUG=412643
TEST=Files.app works as before

Review URL: https://codereview.chromium.org/553393002

Cr-Commit-Position: refs/heads/master@{#294351}