Move ServiceWorkerCache backend creation to a lazy init function.
Previously the ServiceWorkerCacheStorage called ServiceWorkerCache::CreateBackend on Get() and Create().  Now the ServiceWorkerCache does it itself lazily.  This is simpler for the ServiceWorkerCache's caller and should improve performance as backends are created only when needed.  This makes the downstream refcounting CL simpler.

Related CLs:
1. https://crrev.com/542703002: Change ownership of the parameters to ServiceWorkerCache:: Put and Match.
* 2. https://crrev.com/545533002: Move ServiceWorkerCache backend creation to a lazy init function.
3. https://crrev.com/548533002: Make ServiceWorkerCacheStorage::CacheLoader::LoadCache synchronous
4. https://crrev.com/549493002: Expose ServiceWorkerCache objects to ServiceWorkerCacheStorageManager clients.

BUG=392621

Review URL: https://codereview.chromium.org/545533002

Cr-Commit-Position: refs/heads/master@{#294372}