PPAPI: Make CallWhileUnlocked more permissive.
Prior to this patch, CallWhileUnlocked required its parameters' types to
be perfect matches to the types in the function pointer it's meant to
invoke. This is counter to programmers' usual expectations for function
calls, since the compiler will usually do safe casts like add const or
promoting integral types. This just adds template params to CallWhileUnlocked
so that any call with the appropriate number of parameters will be a template
match. The compiler will still decide if there is a valid implicit conversion,
so this is still safe in the sense that mismatched params still won't
compile.

BUG=160925

Review URL: https://codereview.chromium.org/552423003

Cr-Commit-Position: refs/heads/master@{#294405}