Link to stdio output for compile failures.
Some compile failures on Sheriff-O-Matic link to the wrong page today, causing
"No results to display" to be shown under the information for a host, instead
of stdio for that compile failure.

This change checks if the step is "compile" and always displays stdio in that
case.

I haven't added a test case for this; I'd like to clean up the routines in this
class to make it easier to test. I tested this manually while waiting for a
compile failure to happen that didn't display well.

NOTRY=true
BUG=404927

Review URL: https://codereview.chromium.org/549313005

git-svn-id: svn://svn.chromium.org/blink/trunk@181841 bbb929c8-8fbe-4397-9dbb-9b2b20218538