Mac: Fix profile button overlaps fullscreen button on new windows.
This was recently broken by https://codereview.chromium.org/476313003/.
The problem is that when opening a new window, the fullscreen button was not
enabled yet by the time we layout the profile button.
We need to call layoutSubviews after the fullscreen button is enabled because
the avatar button's position depends on it. But with the old avatar button,
we need to call layoutSubviews first because the fullscreen button's position
depends on it.

BUG=412018

Review URL: https://codereview.chromium.org/567713002

Cr-Commit-Position: refs/heads/master@{#294450}