Ensure app install is aborted when the install dialog is closed
If the user initiates an app install via the app launcher, it is
possible to close the install dialog by clicking on the background.
The app launcher window and modal dialog are destroyed without
cancelling the dialog, therefore keeping the install active.

This patch ensures that ExtensionInstallPrompt::Delegate::
InstallUIAbort() is called in this scenario to allow the delegate
(WebstoreStandaloneInstaller) to abort the install.

BUG=409616
TEST=browser_tests

Review URL: https://codereview.chromium.org/547253002

Cr-Commit-Position: refs/heads/master@{#294505}