Make host pairing screen the first screen for shark requisition.
With --show-host-pairing-demo and remora requisition, we start on the host pairing screen instead of the network screen.

There's an issue with the host pairing screen not being ready to accept context changes.

This is a temporary fix to get the end to end flow working. dzhioev@ will land a sequence of patches that are better long term.

BUG=405150
TEST=manual

Review URL: https://codereview.chromium.org/562923002

Cr-Commit-Position: refs/heads/master@{#294513}