Don't clip invalidation of composited scrolling contents
The compositor expects that the whole composited scrolling layer to be
invalidated in time and won't update the tiles that are newly visible
when scrolled if the tiles are up-to-date from the compositor side.

If we clip invalidations, the compositor will think the layer areas
outside of the clip are all valid and won't repaint them.

BUG=412516
TEST=compositing/repaint/should-not-clip-composited-viewport-scrolling-layer.html
TEST=compositing/repaint/should-not-clip-composited-overflow-scrolling-layer.html

Review URL: https://codereview.chromium.org/566433002

git-svn-id: svn://svn.chromium.org/blink/trunk@181873 bbb929c8-8fbe-4397-9dbb-9b2b20218538