Move transient/temporary RenderSVGResource state out of SVGInlineTextBox
The SVGInlineTextBox member variable |m_paintingResource| is only set
temporarily within paintTextWithShadows and paintDecorationWithStyle, and
is only accessed from the methods that do setup and teardown of it. This
makes it possible to remove it from SVGInlineTextBox and instead manage it
in through a variable on the stack.
The lifetime management is kept manual/the same as previously. The
interaction with TextRun::RenderingContext is changed to be passed as a
parameter to the added "scope" object in order to simplify the interface
a bit.

Review URL: https://codereview.chromium.org/568723002

git-svn-id: svn://svn.chromium.org/blink/trunk@181904 bbb929c8-8fbe-4397-9dbb-9b2b20218538