Dispose ApplicationCacheHost on DocumentLoader finalization.
To cater to Oilpan, introduce an explicit dispose() for
ApplicationCacheHost, which the DocumentLoader will call upon when
being finalized.

Without it, a DocumentLoader might clear the Persistent reference to
its ApplicationCacheHost, but it'll still be accessible to embedder
notifications via its WebApplicationCacheHost until the next Oilpan GC
strikes. If so, the callbacks might access a now-dead DocumentLoader.

R=ager
BUG=

Review URL: https://codereview.chromium.org/571003002

git-svn-id: svn://svn.chromium.org/blink/trunk@181973 bbb929c8-8fbe-4397-9dbb-9b2b20218538