Properly initialize top controls content offset on LayerTreeHost.
https://codereview.chromium.org/511253003 added top control content offset
to LayerTreeHost and LayerTreeImpl but didn't sync the two properly on
startup, causing the top controls to hide on navigations. This patch makes
the initial value of the top controls such that the top controls are hidden
(and rely on the initial UpdateTopControlsState from the browser). At the
same time, it also changes LayerTreeImpl's offset to be the content offset,
rather than top controls offset, to match LayerTreeHost.

BUG=412918

Review URL: https://codereview.chromium.org/565673003

Cr-Commit-Position: refs/heads/master@{#294824}