Adding overview / layer framework to Activities so that unloaded / sleeping activities can be shown with as little resources as possible
The new |ContentProxy| object is a placeholder for now which only creates a single color view. It can use the |web_view| to generate the image preview in the next step and will work for WebContents as well as App's.

So more work is necessary, this patch is mainly to get the overview mode working again when running out of memory (and avoiding that it gets too big).

Adding sky for the DEPS inclusion.

BUG=408837
TEST=unittests and visual
TBR=sky@chromium.org

Review URL: https://codereview.chromium.org/548633005

Cr-Commit-Position: refs/heads/master@{#294837}