Re-activate touch selection after fling
Currently, when a scroll starts touch selection handles are hidden and
then re-shown after scroll ends. But, if scroll ends in a fling, handles
are not re-shown, because there hadn't been a way to determine end of
the fling. Now that there is an IPC message notifying end of flings, we
can re-show handles after a fling.

BUG=413927

Review URL: https://codereview.chromium.org/569863003

Cr-Commit-Position: refs/heads/master@{#294927}