Remove guava dependency for multiple_proguards test.
The multiple_proguards_test_apk target used guava as an example for
something that needed to be proguarded out to ensure no proguard
warnings.

Since guava is going away, this CL moves the bad code into chromium
directly so there is no need for guava.

It does bad things which produce warnings, so they are silenced by
telling GYP that it is not chromium code.

BUG=412673

Review URL: https://codereview.chromium.org/571203003

Cr-Commit-Position: refs/heads/master@{#294950}