Roll Clang 214024:216630 (+216684) and switch to CMake
This updates Chromium's clang version to r216630 with
r216684 cherry-picked to fix an ASan issue.

It also changes the build script for Clang to use CMake
instead of Autoconf. The ASan team say this configuration
is better tested, and it also makes us consistent with
the Windows Clang build which already uses CMake.

BUG=400849
R=thakis@chromium.org

Review URL: https://codereview.chromium.org/453513004

Cr-Commit-Position: refs/heads/master@{#294953}