[DirectWrite] Add support for font-stretch suffixes
Add support for font family names with a font-stretch suffix (such as
Condensed, UltaExpanded, Narrow etc). If the full name fails to match an
installed font check for a font-stretch suffix and if so strip it and
set the stretch value according to the suffix.

The current implementation matches IE when it comes to picking or
overriding the stretch value based on the suffix.

The coresponding logic for font-weight suffix matching was implemented
in r181493.

TEST=fast/text/font-stretch-variant.html
BUG=410940
R=cpu@chromium.org, dglazkov@chromium.org

Review URL: https://codereview.chromium.org/550093005

git-svn-id: svn://svn.chromium.org/blink/trunk@182061 bbb929c8-8fbe-4397-9dbb-9b2b20218538