Fix sizing issue of svg as html image
When a html img's src changes from one svg to another, previously the
inner content rect is not correctly updated causing the second svg is
not correctly sized. This is a regression caused by
http://src.chromium.org/viewvc/blink?view=rev&revision=180015 which
call updateInnerContentRect() only if replacedContentRect changes,
missing the case of src change.

Now always call updateInnerContentRect() in RenderImage::layout()
(as before the regression).

The issue only reproduced if the svg is loaded with http protocol.

BUG=412444
TEST=http/tests/svg/image-size-on-src-change.html

Review URL: https://codereview.chromium.org/580453002

git-svn-id: svn://svn.chromium.org/blink/trunk@182079 bbb929c8-8fbe-4397-9dbb-9b2b20218538