chromeos: power: Refactor RendererFreezer and add tests
Refactor the RendererFreezer class so that untestable code is moved into
a delegate and add unit tests.  Modify FakePowerManagerClient so that it
provides more information for tests.

Also use a CancelableClosure for the asynchronous suspend readiness
callback in case we get a SuspendDone before we've had a chance to run
OnReadyForSuspend().

BUG=364339,414396

Signed-off-by: Chirantan Ekbote <chirantan@chromium.org>

Review URL: https://codereview.chromium.org/543303002

Cr-Commit-Position: refs/heads/master@{#295111}