Avoid unnecessary visibility changes to BookmarkBarView
I came across this while investigating slowness. Without this we would
end up unnecessarily toggling visibility a couple of times. While not
the end of the world, toggling visibility for the bookmarkbar can take
a while (mostly because it can end up containing a ton of views).

BUG=none
TEST=covered by test now
R=jamescook@chromium.org

Review URL: https://codereview.chromium.org/573893003

Cr-Commit-Position: refs/heads/master@{#295116}