SQLite's recover.c should be valid C code
Move two variable declarations to the top of their function, so
Chromium's version of SQLite compiles on Windows. SQLite is used
by NSS on Dart's standalone Windows executable.

NOTRY=true
This is not based on src/, and is not tested until DEPS are updated.

This patch is not added to recover.patch, because the last 3 patches,
including the one this fixes, are not there, and recover.patch says:
    Since recover.c is in somewhat active development, it is possible that
    the patch below will not reliably re-create the file.

Review URL: https://codereview.chromium.org/552673003

Cr-Commit-Position: refs/heads/master@{#295260}