AX: Calendar Picker: Support of input[type=week].
Add accessibility attributes to WeekNumberCell.  This is similar to what we did
for DayCell. However we need to add a helper function to PagePopupController in
order to get a descriptive string for a week, and needs DateComponents::setWeek.

We'll generate a string like "Week 38, 2014, starting on September 15, 2014"

This CL doesn't contain a test. We'll add a test after landing a
Chromium CL for the string.

BUG=123896

Review URL: https://codereview.chromium.org/554383004

git-svn-id: svn://svn.chromium.org/blink/trunk@182213 bbb929c8-8fbe-4397-9dbb-9b2b20218538