rendering_helper - Schedule the RenderContent on a constant rate.
RepeatingTimer uses a constant delay to schedule the next task. If
the current task is postponed for any reason, the delay would be
accumulated.

This is not the behavior we want for playing video (to sync with
the audio). In this CL, we schedule the task by ourselves and try
to catch up once we fall behind.

BUG=411123
TEST=Run the test on Peach Pit.

Review URL: https://codereview.chromium.org/571743003

Cr-Commit-Position: refs/heads/master@{#295407}