Move ContentWindow from BrowserPlugin To GuestView
This CL plumbs out the swapped out RenderView's routing ID out to the content embedder which then uses it to expose its contentWindow to an optional callback in the AttachGuest API Method.

This patch also makes it fairly trivial to support a contentWindow in other GuestViews.

BUG=330264

Review URL: https://codereview.chromium.org/564973004

Cr-Commit-Position: refs/heads/master@{#295477}