[Ash] Only snap windows that can maximize.
Previously, windows that have only one of maxWidth or maxHeight defined
could not be maximized by the user, but they would still snap when
dragged to the edge of the screen.

This changes WindowState::CanSnap to use CanMaximize which is false
when there is a maxWidth or maxHeight, or when the window is not
resizable.

BUG=412610, 408737

Review URL: https://codereview.chromium.org/557693002

Cr-Commit-Position: refs/heads/master@{#295657}