Gallery: Fix a beginning position gap of cropping.
The gap occures when a user moves the mouse cursor fast at the beginning of
cropping. The reason the app uses the first mousemove event's coordinate as the
beginning position of cropping.  The CL uses the mousedown event's coordinate
instead of the first mousemove's one.

BUG=412237
TEST=manually

Review URL: https://codereview.chromium.org/590073002

Cr-Commit-Position: refs/heads/master@{#295945}