Script streaming enhancements: suppress streaming for small scripts.
It doesn't make sense to start the streaming for scripts which are so small that
all data will come in one chunk. As a heuristic, if the first data chunk is
small, we assume there won't be another one (getting this information for real
requires a considerable amount of refactoring).

In addition, when the streaming is suppressed (for whatever reason), we don't
need to wait for the background task (if any) to complete, but we can resume the
non-streaming code path right when the loading finishes.

R=haraken@chromium.org
BUG=

Review URL: https://codereview.chromium.org/587243002

git-svn-id: svn://svn.chromium.org/blink/trunk@182420 bbb929c8-8fbe-4397-9dbb-9b2b20218538