Add utility script for GPU Pixel Wranglers.
This script attempts to automate a lot of the waterfall-staring
that pixel wranglers need to do. A version is already published
and documented at
http://www.chromium.org/developers/how-tos/gpu-wrangling/check_gpu_bots-script.
This script is intended to be a temporary measure to help out
pixel wranglers until such functionality (or better) is
incorporated into tools like the sheriff-o-matic.

BUG=

Review URL: https://codereview.chromium.org/588603003

Cr-Commit-Position: refs/heads/master@{#296032}