This is a follow up patch for "Introduce WebRequestEventRouterDelegate.".
In this CL, we avoid leaking the void* pointer of profile outside of
web_request_api.cc. The main idea behind is void* is used in webrequest API to
avoid accidently access to profile on IO thread. When we are certain that a
function is called on UI thread, we can cast the void* to BrowserContext*.

A refactor is made to
WebRequestEventRouterDelegate::LogExtensionActivity(...):
-Change its first parameter from void* to BrowserContext*; So we won't leak void* pointer outside of web_request_api.cc.
-This function only keeps the logic that is run on UI thread, the logic for IO thread is moved back to web_request_api.cc.

BUG=352293

Review URL: https://codereview.chromium.org/589153002

Cr-Commit-Position: refs/heads/master@{#296040}