Merge Cr-Commit-Position: refs/heads/master@{#294425}
commit f29ee62eeb9057b81d4e6f690409b209df367c04
Author: rch <rch@chromium.org>
Date: Thu Sep 11 18:13:26 2014

Fix a bug when a QUIC session is closed during the call to InitializeSession. This seems to be able to happen if too much time elapses between the call to the constructor and the call to InitializeSession.

Also adds a histogram to monitor the frequency of this situation.

BUG=399147

Review URL: https://codereview.chromium.org/560913002

Cr-Commit-Position: refs/heads/master@{#294425}

R=rtenneti@chromium.org

Review URL: https://codereview.chromium.org/586393004

Cr-Commit-Position: refs/branch-heads/2125@{#443}
Cr-Branched-From: b68026d94bda36dd106a3d91a098719f952a9477-refs/heads/master@{#290040}