Keep reference view pressed while extension actions have a popup
The reference view for an extension popup is either the
extension action button itself (if the extension action is
visible on the main bar) or the wrench menu/chevron (if the
action is overflowed). When a popup is opened, the reference
view should remain pressed for the duration of the popup.

BUG=414944
BUG=415171

Review URL: https://codereview.chromium.org/547303003

Cr-Commit-Position: refs/heads/master@{#296057}