Merge text from all changes to the same live region into one utterance.
The problem was that if there were two objects that both changed in the
same live region, like 'apple' and 'orange', they should really be
spoken in sequence, but there was no way to enforce that - another
utterance can be inserted in-between them. Until we have a way to enforce it,
combining the text is a better user experience.

To match other screen readers, read text and control values, but not
roles and states.

BUG=408821

Review URL: https://codereview.chromium.org/522433002

Cr-Commit-Position: refs/heads/master@{#296070}