Remove StartPageView from being a AppListViewDelegateObserver
It observes OnProfilesChanged, but the set of profiles changing doesn't
affect the start page. However, it probably once helped avoid crashes
when deleting profiles since the AppListViewDelegate would hold on to
the Profile* until the app list was shown again. But that was fixed in
http://crrev.com/291852 (now AppListServiceImpl::DestroyAppList() will
reliably tear down the AppListView if the profile is destroyed, without
having to wait for a ShowAppList().

When regular profile switching occurs (i.e. not when the profile is
being destroyed), the profile currently being used by the
StartPageView is also replaced. That is handled by
AppListMainView::ModelChanged() calling into
ContentsView::InitNamedPages() which makes an entirely new
StartPageView().

BUG=403647

Review URL: https://codereview.chromium.org/589333002

Cr-Commit-Position: refs/heads/master@{#296175}