Add a document describing which C++11 features we allow.
For now, this allows only a very limited set.  Also describes the process we're
going to follow to allow more features.  The idea is that we're gradually going
to allow more features over time.

Also included is the source of a tiny appengine app running at
chromium-cpp.appspot.com that will show the contents of the document in this
CL once it's submitted.

BUG=360096
NOTRY=true
R=ben@chromium.org, inferno@chromium.org, jamesr@chromium.org

Review URL: https://codereview.chromium.org/589413003

Cr-Commit-Position: refs/heads/master@{#296214}