Change ScriptInjection to work with WebLocalFrames.
This is to help unblock some of the Blink work to move local frame only
APIs off WebFrame. Strictly speaking, all that was required to unblock
this bug was to manually coerce the WebFrame to a WebLocalFrame for the
call to setIsolatedWorldHumanReadableName.

However, it never makes sense to inject script on a remote frame, so
I've updated ScriptInjection to just work with local frames where
possible. Code that uses ScriptInjection now does the type coercion at
the class boundary. As extensions code is gradually updated to
understand out-of-process iframes, these manual type coercions should
disappear.

BUG=416659

Review URL: https://codereview.chromium.org/594043002

Cr-Commit-Position: refs/heads/master@{#296345}