Get rid of dataToExamine from sheriff-o-matic.
It doesn't serve any purpose. We should just pass the CTFailureGroup
directly to the ct-results-panel. In addition to simplifying the code
this fixes a bug introduced in r182172 where examine pages wouldn't
work.

Store the failure group key in the pushState so that we have an easy
way to grab the key out and find the right CTFailureGroup on popState.

Added a new test to ct-results-panel-tests for the url results panel case.
Also, cleaned up this test to get rid of the unnecessary use of beforeEach.
beforeEach was making it hard to write the url test because it used global
state that didn't apply.

NOTRY=true
R=leviw@chromium.org, michaelpg@chromium.org

Review URL: https://codereview.chromium.org/590003002

git-svn-id: svn://svn.chromium.org/blink/trunk@182557 bbb929c8-8fbe-4397-9dbb-9b2b20218538