cc: Stop using transfer buffer limit for 1-copy rasterizer.
It doesn't make sense for 1-copy rasterizer where we can
control the exact amount of memory used to be affected
by the transfer buffer limit.

This sets the upper bound to 32 resources, which is reduced
to 0 when invisible. A simple constant like this is likely
what we want long term.

BUG=406404

Review URL: https://codereview.chromium.org/597833002

Cr-Commit-Position: refs/heads/master@{#296423}