Check multi-crx path for force update based on manifest entries
The previous CL (https://codereview.chromium.org/540673002) was not fully
correct.  It used a path based on <nacl_arch>, but the path can be arbitrary
based on entries in the extension's manifest.

If the extension used non-nacl_arch paths, it would continually download and
install the same copy of the extension, eventually filling up the user's disk if
left long enough without restarting Chrome.

BUG=414156

Review URL: https://codereview.chromium.org/596193002

Cr-Commit-Position: refs/heads/master@{#296514}