[NaCl SDK AppEngine] Fix PNaCl demos after html5 onpopstate change.
window.onpopstate used to be called with a null state when a page was
loaded for the first time. This was a bug, and inconsistent with other
browsers, but we relied on it in the PNaCl demo page.

Without it, direct links to a demo page were not working properly. Even
after fixing this, I ran into issues with history when creating/destroying
a new iframe element whenever the page switched. Using
iframe.contentDocument.location.replace seems to have fixed these issues.

No trybots because this only affects the AppEngine demo page.

BUG=none
R=sbc@chromium.org
NOTRY=true

Review URL: https://codereview.chromium.org/600653002

Cr-Commit-Position: refs/heads/master@{#296526}