When changing a background image, invalidate the whole document, not the visible viewport.
This fixes a bug introduced by crrev.com/497233005.
The code prior to that CL did it correctly, by having one-off code that reached
into FrameView::contentsWidth() and FrameView::contentsHeight(). This CL does the
same thing without a one-off, by taking advantage of the fact that the current
renderer is either document or body, which is already the appropriate place to
register a paint invalidation.

BUG=416400

Review URL: https://codereview.chromium.org/601013002

git-svn-id: svn://svn.chromium.org/blink/trunk@182637 bbb929c8-8fbe-4397-9dbb-9b2b20218538