Mac: Deflake AppListServiceInteractiveTest.SwitchAppListProfilesDuringSearch
This is flaky because the app list ends up trying to read an ImageSkia
on one of Cocoa's background threads.

While NSView has a `canDrawConcurrently` function, AppKit is not calling
it in the case that is making this test flaky. The documentation says,
"The default is NO for most types of views.." [sic], and I've verified
it is `NO` and tried overriding it, but it was just never called.

However, having -[NSView canDraw] return `NO` seems to work fine.
Returning NO when there is no message loop ensures a draw only occurs on
a thread that Chrome created (not one of Cocoa's background threads),
and doesn't raise any DEPS issues.

BUG=417148, 415264

Review URL: https://codereview.chromium.org/599803002

Cr-Commit-Position: refs/heads/master@{#296591}