Remove the "snionly" concept from the HSTS preload.
The "snionly" tag only applies to preloaded values and only affects
people who are forcing SSLv3 on the command line. (We no longer have the
option in the preferences to disable TLS and the preference has been
renamed so even those who might have set it in the past shouldn't have
it now).

It also complicates pending work to reduce the size of the preload data.

BUG=102779
R=eroman@chromium.org, rsleevi@chromium.org

Review URL: https://codereview.chromium.org/578553004

Cr-Commit-Position: refs/heads/master@{#296821}