Fix crash when showing multiple popups.
The problem is that calling hide/cancel on a popup is async.

With two popups, you would get

Previously:
SHOW A
HIDE A
SHOW B
NOTIFY SELECTION A
HIDE B
NOTIFY SELECTION B

When notifying the selection of A in the above scenario, it was
clearing the frame pointer (and actually updating the wrong popup).
Then when hiding the second popup, it's reference was actually gone
already.

Fix this by making the call to hide sync instead of async.

BUG=416354

Review URL: https://codereview.chromium.org/603303002

Cr-Commit-Position: refs/heads/master@{#296826}