Maybe add CORS headers to data reduction proxy redirect
When the data reduction proxy responds with a bypass message,
chromium constructs a redirect response and feeds that through
the network stack. This results in the request being retried
directly to the origin. If the original request had an
Origin header, then the manufactured redirect should have the
appropriate CORS headers.

BUG=415644

Review URL: https://codereview.chromium.org/603323002

Cr-Commit-Position: refs/heads/master@{#296942}