Elim parent_window parameter from network_connect
We were passing a 'parent_window' parameter to the network_connect::ConnectToNetwork() in an effort to open network configuration dialogs from the desktop they were triggered from. This was never a good idea since gfx::NativeWindow is actually a pointer which may become invalid before the callback is triggered.

In practice this is unnecessary since we decided that all network configuration applies to the primary user, so we should always use the currently active desktop to host the dialog (which is what GetNativeWindow() does in ash_system_tray_delegate.cc).

BUG=413925

Review URL: https://codereview.chromium.org/607613002

Cr-Commit-Position: refs/heads/master@{#296969}