Accept inbound connections on unix domain socket
In order for CrOS Core services to speak Mojo, the mojo_shell will need to be able to accept incoming connections from externally-managed processes.

POR is to have the shell listen on a unix domain socket and convert incoming connections to a proper Mojo MessagePipe and then wire it to an InterfaceImpl<ExternalApplicationRegistrar>, which will make the appropriate calls on ApplicationManager.

BUG=418289
TEST=mojo_external_application_tests
STATUS=Fixed

Review URL: https://codereview.chromium.org/522443003

Cr-Commit-Position: refs/heads/master@{#297184}