Fix a crash when opening CronetTest app
If we install and open CronetTest app without CronetTestInstrumentation app,
CronetTest app will crash and complain that it cannot find
MockUrlRequestJobTest class. The reason is that this class is in
test/javatests/src/ and not in test/src/, so it does not get loaded to device
if we choose to only install and open CronetTest app.

BUG=415781

Review URL: https://codereview.chromium.org/612433002

Cr-Commit-Position: refs/heads/master@{#297261}