AppsGridView: Use specific types instead of views::View*.
Almost every views::View* in this class is actually an AppListItemView*,
and we are now specific about that. Also made a few explicit
PulsingBlockViews. This removes a mess of static_casts throughout the
file, and makes the interface to many methods type-safe and explicit.

Review URL: https://codereview.chromium.org/599933004

Cr-Commit-Position: refs/heads/master@{#297341}