Adjust maximum scroll bounds on FrameView to account for top controls. (Chromium-side)
The compositor adjusts the viewport bounds based on how far the top
controls are shown; however, the resize of the viewport on the Blink
side doesn't happen until a Resize event is received, which is heavily
throttled. This means that when a commit happens before the Resize,
the scroll offsets given to Blink may be outside the maximum bounds
and will be incorrectly clamped.

This CL adjusts FrameView's maximumScrollPosition to account for the
drift in where Blink and the Compositor think the top controls are.

Blink-side: https://codereview.chromium.org/560623002/

BUG=412581

Review URL: https://codereview.chromium.org/563513002

Cr-Commit-Position: refs/heads/master@{#297478}