Mac: Fix theme image drawing when building with >=10.9 SDK.
When linking with >=10.9 SDK, NSThemeFrame will have wantsLayer=YES and
wantsUpdateLayer=YES, resulting in the swizzled drawRect to be ineffective
(updateLayer is called instead of drawRect).
This patch moves the drawing of the theme image behind the tab strip area
to use a small subview instead. This subview is positioned at the top of the
window, behind its sibling views. The view will redraw when the window's
activation state changes because the theme image may be different.

NativeAppWindowCocoa is another user of the swizzled drawRect, so I'm not
removing the swizzling yet in this CL.

BUG=417872

Review URL: https://codereview.chromium.org/611453004

Cr-Commit-Position: refs/heads/master@{#297498}