[CSS Grid Layout] Fix positioned grid items crash
Modify RenderGrid::layoutGridItems() to take care of the positioned
items, otherwise when the tree is invalidated the list of positioned
objects would be empty and we would reach an assert.

More changes related to the absolute positioned items should be done
later, to adapt the behavior to the spec, but this is just a patch to
avoid the crash at this moment.

BUG=402504
TEST=fast/css-grid-layout/positioned-grid-items-crash.html

Review URL: https://codereview.chromium.org/594153002

git-svn-id: svn://svn.chromium.org/blink/trunk@183051 bbb929c8-8fbe-4397-9dbb-9b2b20218538