[Tizen][Temp] Allow a different ScreenOrientationProvider to be set.
At the moment, out-of-tree implementations of ScreenOrientationProvider
cannot be passed to ScreenOrientationDispatcherHost, so add a new method
SetProvider() to it to do that.

This was originally committed as pull request chromium-crosswalk/185,
and has been rewritten for M39.

It will not be necessary once https://codereview.chromium.org/546453004/
and https://code.google.com/p/chromium/issues/detail?id=396760 are fixed.