Move notifications to after the process has indeed closed.
Renderer process closed notification implies that the process is already gone. In reality, this is not always the case since calling ProcessDied(false) is supposed to tear down the process. It however sends the notification before doing so. This patch moves the notifications to after the process is closed and the IPC channel is torn down.

BUG=

Review URL: https://codereview.chromium.org/613173004

Cr-Commit-Position: refs/heads/master@{#297698}