media: Pass DemuxerStreamProvider in Renderer::Initialize().
Currently DemuxerStreamProvider is passed in the constructor of Renderer
implementations (e.g. RendererImpl and MojoRendererImpl). However, if we want to
create the Renderer earlier and pass it to WebMediaPlayerImpl or Pipeline, the
DemuxerStreamProvider may not be available yet. Meanwhile, when Renderer is
initialized by the Pipeline, DemuxerStreamProvider is always available.

This CL moves the DemuxerStreamProvider parameter to Renderer::Initialize() to
accommodate the above cases.

BUG=416570
TEST=Unit tests updated.

Review URL: https://codereview.chromium.org/591203003

Cr-Commit-Position: refs/heads/master@{#298006}