cc: Fix DCHECK in picture pile.
The DCHECK is trying to say that "If the tile isn't going
to get rerecorded, then it better be fully invalidated."

However the "isn't going to get recorded" part isn't quite
right. It compares the TileBounds to the interest_rect.
But the interest rect is iterated including borders. So
we should use the TileBoundsWithBorders or the
interest_rect_over_tiles to get the right DCHECK.

This uses TileBounds and interest_rect_over_tiles.

R=enne, vmpstr
BUG=418447

Review URL: https://codereview.chromium.org/624973002

Cr-Commit-Position: refs/heads/master@{#298045}