Add test coverage for event-targeting of the RootView and contents view
Adds test coverage to verify that the contents view
is targeted instead of the root view for gesture
events which should be targeted to the contents view.
Also tests that the root view is targeted when there
are no other eligible targets within the tree.

BUG=414339
TEST=ViewTargeterTest.TargetContentsAndRootView

Review URL: https://codereview.chromium.org/612093005

Cr-Commit-Position: refs/heads/master@{#298056}