Credential Manager: Add a blink::WebCredential-based constructor to CredentialInfo.
We current have a constructor that accepts specific data values, and
we're using it to verbosely convert a WebCredential object to a
CredentialInfo struct to pass between the renderer and browser. This
patch drops that constructor in favor of a constructor that accepts
a WebCredential directly.

BUG=400674
R=vabr@chromium.org

Review URL: https://codereview.chromium.org/635753002

Cr-Commit-Position: refs/heads/master@{#298442}