ChromeVox build fixes.
This fixes two variable expansion errors in the gyp files for ChromeVox.
This also reverts a CL that hid one of these bugs.

Revert "Fix the ChromeVox generate_deps.py to create the parent dir for the output file."

This reverts commit 76e477470adbfe3fb1ff6a99ab5feeb4ac57c13c.

Ninja creates the output directories for its known outputs, and scripts should
normally not do that.  In this case, the output was not specified correctly
which meant that a clean build would fail because ninja didn't create
the output directory.  Now that this is fixed, the quick fix isn't
needed any more.

R=dtseng@chromium.org
BUG=

Review URL: https://codereview.chromium.org/631893002

Cr-Commit-Position: refs/heads/master@{#298461}