[Android] Add a sync scrolling smoothness benchmark
Add a sync scrolling benchmark that disables threaded scrolling with the
key_mobile_sites test suite. This will be used to identify a list of
candidate sites that can then be used to guide and focus main thread
scrolling optimizations. In particular, the hope is to identify and
eliminate exhisting (and avoidable) hiccups in the pipeline that might
otherwise discourage efforts like the proposed 'beforescroll'
(crbug.com/410974).

BUG=417757

Review URL: https://codereview.chromium.org/623163002

Cr-Commit-Position: refs/heads/master@{#298544}