Simplify main resource referrer setting
There are several places where we set a referrer, but FrameLoader would have set it for us anyway. For the vast majority of cases, set referrer in FrameLoader::setReferrerForFrameRequest(). If FrameLoader::load() is called with a non-null originDocument on the FrameLoadRequest, this is the only place it will be set*

* There is a single exception: a certain class of plugin-initiated frame requests, which set the referrer from the chromium side and look to be somewhat complex to untangle.

BUG=

Review URL: https://codereview.chromium.org/604013002

git-svn-id: svn://svn.chromium.org/blink/trunk@183361 bbb929c8-8fbe-4397-9dbb-9b2b20218538