Refactoring the GLRenderer::GetBackgroundWithFilters method.
This patch prepares the existing code for the shader based
implementation of blending: https://codereview.chromium.org/555133002/.

It splits the GetBackgroundWithFilters() functions into multiple chunks,
so that we can have access to the backdrop bounding box and to
the backdrop itself, before being transformed with the inverse matrix.

This patch should not change the current behavior.

BUG=

Review URL: https://codereview.chromium.org/614153002

Cr-Commit-Position: refs/heads/master@{#298693}