[Win] Fix focus related issues in the new avatar bubble.
This does two things:

1. Removed the background color on the buttons when tabbing. This makes
it consistent with the Mac, where on tabbing, we only show an outline on
the buttons, and only when the button is hovered does its background change

2. Fixes the bug where when a button is pressed, it is disabled to prevent
double clicking. I don't think double clicks are an issue, but the disabling
is definitely an issue as it makes the FocusManager advance focus, which
end up looking weird.

BUG=413154
TEST=Start Chrome with --enable-new-avatar-menu. Clicking on the Switch
Person button should not look like it has temporarily focused the Incognito
button.

Review URL: https://codereview.chromium.org/632343002

Cr-Commit-Position: refs/heads/master@{#298735}