cc: Move RequiresHighResToDraw from tree to lthi.
This patch moves RequiresHighResToDraw setting (with setters) to
LTHI. The reason for this change is that we have situations in which
we need to access this property from the pending tree as well
(when deciding whether tiles are required for activation). Additionally,
most of the fiddling with the setting comes from LTHI anyway, so it
makes sense for the property to live there.

R=danakj, enne

Review URL: https://codereview.chromium.org/640873002

Cr-Commit-Position: refs/heads/master@{#298787}