Remove overzealous asserts in popup zoomer.
The cause for assert is "a double-show" that occurs due to
insufficient check in PopupZoomer.show():
  if (mShowing || mZoomedBitmap == null) return;
However, "fixing" it with
  if (isShowing() || mZoomedBitmap == null) return;
would break attempts to re-show while hide is still active,
which is undesirable.

Luckily, there is no reason to protect against the double-show
because we already protect ourselves with the view children checks
in ContentViewCore:
  if (mContainerViewAtCreation.indexOfChild(zoomer) == -1) etc.

Hence, this CL removes the overzealous asserts.

BUG=417512
NOTRY=True

Review URL: https://codereview.chromium.org/634113003

Cr-Commit-Position: refs/heads/master@{#298794}