As per the investigation in http://crbug.com/417756, it seems the Two* tests are susceptible to timeouts on GPU-less, slow bots.
This patch will make the test no longer detect that video is up for
both streams before proceeding - we assume other tests check that.
Instead, we're happy with just getting a stream before proceeding.
This removes the costly video checks from this test so it can test
what it wants to test, which is to ensure we can stop streams in an
orderly fashion.

BUG=417756
R=perkj@chromium.org

Review URL: https://codereview.chromium.org/640253002

Cr-Commit-Position: refs/heads/master@{#298884}