OOPIF: Don't use OpenURL for cross-site top-level navigations in the renderer.
Currently, with --site-per-process, DecidePolicyForNavigation returns an Ignore policy for cross-site top-level navigations and sends them to the browser via OpenURL.  This was breaking tests like RenderViewImplTest.DecideNavigationPolicy that don't expect an Ignore policy.  Doing the cross-site check and using OpenURL shouldn't be necessary here, since the browser is already making the process swap decision in CrossSiteResourceHandler.

This change also uncovered a problem with handling downloads with OOPIF, where we were deferring for a navigation policy check in CrossSiteResourceHandler before checking whether the request is for something that won't need a process swap.  This resulted in a DCHECK(!defer_ignored) being hit in BufferedResourceHandler::UseAlternateNextHandler when running DownloadContentTests.  This was fixed by moving the DeferForNavigationPolicyCheck to occur after checking for downloads, streams, and No Content.

This issue was also affecting the NavigateToURL function in browser tests, as it was stopping the load on a redirect and starting a new load.  With this change, it is properly one single load with a redirect in the middle.

BUG=417518
TEST=RenderViewImplTest.DecideNavigationPolicy passes on Site Isolation FYI bot.

Review URL: https://codereview.chromium.org/631393002

Cr-Commit-Position: refs/heads/master@{#298914}