Changed parser resource fetching to only allow priority increases.
If an image is referenced in multiple places on a page it may be
requested multiple times, even after other priority adjustments have
been made. This resulted in image requests thrashing their priority
as the parser discovered new instances that have already been laid out
and the priority increased.

Changing it to only allow increases in priority avoids the thrashing and
maintains the benefit of prioritizing visible images higher.  There are
no cases where the parser fetches should be allowed to decrease the
priority of requests for the same resource that are already in-flight.

The image priority when re-prioritizing visible images was also not
correctly accounting for the screen area for images referenced multiple
times so that has also been fixed.

BUG=415973

Review URL: https://codereview.chromium.org/584913003

git-svn-id: svn://svn.chromium.org/blink/trunk@183479 bbb929c8-8fbe-4397-9dbb-9b2b20218538