Change PDF scaling factor from double to float.
This ultimate consumers of this want a float anyway, so changing to be a float
all the way through is less misleading.  This also avoids some "value possibly
truncated" warnings (currently disabled) on MSVC.

This also removes a couple of scale-related functions in metafile_skia_wrapper.*
that seem to be entirely unused.

BUG=none
TEST=none

Review URL: https://codereview.chromium.org/641923002

Cr-Commit-Position: refs/heads/master@{#298980}