Skip audio frames with unknown timestamp in MPEG2TS
This is another issue with HLS streams produced by VLC. Even though it's recommended that each PES packet in MPEG2TS contain a timestamp, we have seen that in the real-world content streams that's not always the case. In order to be able to play such streams, instead of failing stream parsing when we don't have timing info at the beginning of playback, we will just skip some frames until we eventually  get a frame with timing info and from that point we should be able to proceed playing normally. Previous change by Damien fixed this issue for video streams. This change fixes it for AAC ADTS and MPEG1 audio streams in mpeg2 ts containers.

BUG=420227

Review URL: https://codereview.chromium.org/646673002

Cr-Commit-Position: refs/heads/master@{#298982}