Remove accelerated.de from HSTS list in Chrome 38.
This change landed a long time ago, on Aug 18th in r290306. However, in
the announcement email for the Chrome 38 branch, the revision number was
typoed and I concluded that it didn't need to be merged, but it did.

This change merges that removal to Chrome 38.

BUG=422109

Cr-Commit-Position: refs/branch-heads/2125@{#566}
Cr-Branched-From: b68026d94bda36dd106a3d91a098719f952a9477-refs/heads/master@{#290040}