cc: Ensure to skip the first tile that would be returned by twin.
Since we're using a spiral iterator, we run into a situations where
a shared tile is visited by an iterator, but it is meant to be returned
by the twin (since it has higher priority). We already skip these types
of tiles in the Pop function. However, it could be the case that
the very first tile is of this kind. The fix is to ensure we skip these
types of tiles in the constructor as well.

BUG=422178
R=reveman

Review URL: https://codereview.chromium.org/647683002

Cr-Commit-Position: refs/heads/master@{#299135}