Read HID report descriptor from sysfs and request permission on connect.
The HID report descriptor can be read from sysfs instead of executing an
ioctl on the device node. This means that on Chrome OS it isn't
necessary to ask the permission broker for access to the device on
enumeration. It can be delayed until the device is opened, as it is for
USB devices. This also resolves a long-standing issue with HID devices
not being enumerated on Chrome OS the first time an enumeration is
requested because of the extra asynchronous step required only on that
platform.

BUG=411899

Review URL: https://codereview.chromium.org/641203003

Cr-Commit-Position: refs/heads/master@{#299228}