Give up on threaded animations for the Mac App Launcher NSWindow
AppKit ignores -[NSView canDrawConcurrently:] and invokes redraws on the
animation thread.

The last attempt to make one part safe was r296591. However,
http://crbug.com/421952 links to undiagnosable crashes within AppKit
from a build at r298667. These are likely race conditions due to this
threading.

The downside is just that, when Chrome is not running, the initial
appearance of the app launcher window will be a bit delayed, and will
just "pop" in, rather than animate.

BUG=421952

Review URL: https://codereview.chromium.org/646353002

Cr-Commit-Position: refs/heads/master@{#299268}