cc: Bump up pending tree now tiles order for smoothness mode.
In smoothness takes priority mode, we should ensure to return tiles
that are required for activation sooner than eventually bin. The reason
for this is that we don't want to starve activation and the prepainting
required for smoothness should already be covered by the soon bin.

Additionally, we have to ensure not to mark these tiles as violating
memory policy, since required for activation tiles need special
treatment (unfortunately). The reason for this is that activation is
currently tied to updating our recordings and smoothness mode would
otherwise starve activation even if we're past the recording region.

BUG=422819
R=reveman

Review URL: https://codereview.chromium.org/651503004

Cr-Commit-Position: refs/heads/master@{#299386}