PasswordAutofillAgent: do not provisionally save forms in NULL frames
Based on a recent crash reports, there is a suspicion that PasswordAutofillAgent::provisionally_saved_forms_ gets NULL WebFrame pointers assigned as keys. That's likely happening in TextDidChangeInTextField(), when ProvisionallySavePassword() is called with element.document().frame(). That can be NULL, as the comment at blink::Document::frame() says.

This CL makes sure that forms for which we don't know the frame are not provisionally saved.

It is still unclear to me, how to reproduce this crash, so I have no idea how to reproduce it in tests. Therefore this CL does not add tests. There is a bug (400186) filed for making PasswordAutofillAgent observe frames directly, which should ultimately clean up this mess anyway.

BUG=420519

Review URL: https://codereview.chromium.org/649583003

Cr-Commit-Position: refs/heads/master@{#299443}