Make sure that a request with no-store response doesn't block next requests.
When reading a Stream[1][2], we fetch the data on demand. That makes a transaction
open until the client read all data. An HTTP cache transaction for a resource
blocks the subsequent transactions for the same resource while it reads from
and writes to the cache.
This CL adds a unittest that verifies that when cache-control: no-store is
on the response the request doesn't block following requests for the
same resource.

1: https://github.com/whatwg/streams
2: https://docs.google.com/a/chromium.org/document/d/1s3lodsbto_opSxHRFa0-gNIsi_4RaJOqPlmWBqwSS-c/edit

BUG=418879
R=ricea@chromium.org

Review URL: https://codereview.chromium.org/612313002

Cr-Commit-Position: refs/heads/master@{#299526}