OOPIF: Check whether localFrameRootTemporary is valid when processing resizing messages.
With --site-per-process, some content_browsertests were crashing in the renderer while processing a ViewMsg_Resize, because they accessed localFrameRootTemporary()->frameView() without first checking whether localFrameRootTemporary() is valid.  localFrameRootTemporary() won't be valid if site A opens a popup to site B: when we initialize the opener frame/view in B's process, the opener will just have one remote top-level frame.  For an example, see RenderFrameHostManagerTest.DontSwapProcessWithOnlyTargetBlank.  In this case, the resizing message is sent as part of initializing the swapped-out view for the opener, and should be ok to ignore.

BUG=417518

Review URL: https://codereview.chromium.org/613043005

git-svn-id: svn://svn.chromium.org/blink/trunk@183695 bbb929c8-8fbe-4397-9dbb-9b2b20218538