Pass correct client coordinates for mouse events synthesized from a GestureTap
After re-adding the extra hit-tests on GestureTap in crrev.com/429793004 the
scroll offset was being added twice to the synthesized mouse events. This went
unnoticed because the target had already been determined in earlier hit-tests
and so the event was dispatch to the proper target but with the wrong
coordinates. This CL fixes that to match the exact coordinates that native
mouse events would report.

BUG=418800

Review URL: https://codereview.chromium.org/641063003

git-svn-id: svn://svn.chromium.org/blink/trunk@183699 bbb929c8-8fbe-4397-9dbb-9b2b20218538