[MAC] Disable 'Import' button when visible checkboxes are unchecked.
In "Import bookmarks and settings" overlay, 'Import' button is disabled when
no checkboxes are checked. But for 'Safari' only 2 checkboxes are shown;
when other checkboxes are checked in background(not shown), it causes the button
to always be enabled regardless of visible checkboxes checked or unchecked.
This CL checks if a checkbox is both shown and checked, to enable the button.

BUG=275265
R=dbeam@chromium.org

TEST=
1.Launch chrome ,go to settings and click on 'Import bookmarks and settings' button under 'Users' section.
2.Select 'Safari' and Uncheck 'Browsing history' and 'Favorites/Bookmarks' option .
3.Observe the 'Import' button. The 'Import' button should be disabled.

Review URL: https://codereview.chromium.org/636933002

Cr-Commit-Position: refs/heads/master@{#299840}