Devtools Animations: Basic animation inspection & control in Styles pane
Add a new hidden devtools experiment for animation inspection within the
inspector. A new inspector domain is created for Animations.

CSS Animations, CSS Transitions and Web Animations can be viewed in an
Animations pane when an animation is in effect and targeting the element
being viewed in the Elements pane. Each animation can be paused, played
and scrubbed based on time position. Animation properties are shown for
all animations.

BUG=419269

Review URL: https://codereview.chromium.org/620783002

git-svn-id: svn://svn.chromium.org/blink/trunk@183847 bbb929c8-8fbe-4397-9dbb-9b2b20218538