Preparation of Promise<T> support in the Web IDL parser
We are about to support Promise<T> syntax in the Web IDL parser. As it is
specified in the Web IDL spec, I would like to implement it in
tools/idl_parser. This CL is a preparation for the change in chromium side,
in order not to break the existing behavior when the change lands.

BUG=421539

Review URL: https://codereview.chromium.org/654193002

git-svn-id: svn://svn.chromium.org/blink/trunk@183849 bbb929c8-8fbe-4397-9dbb-9b2b20218538