Do not create proxy hosts in the subtree of navigating frame.
This is an another attempt at landing https://codereview.chromium.org/536143002.

When a frame is navigating cross-process, its existing document will be completely destroyed and all child frames will be gone. This means that we don't need to create proxy objects for the new SiteInstance in the subtree of the navigating frame.

BUG=357747, 417030

Review URL: https://codereview.chromium.org/600483003

Cr-Commit-Position: refs/heads/master@{#300112}