Implements peer connection stats measurements in WebRTC call test.
The telemetry test for WebRTC has so far been measuring CPU and memory.
Since there's a stats API for WebRTC (see http://goo.gl/lfQKzE), we try
to use that do extract interesting properties about the call. The
properties we look at are such that we expect them to remain stable, so
a sudden change in any of them will warrant investigation.

This implementation is based on a trick where we instrument the
constructor for webkitRTCPeerConnection. Whenever a peer connection is
constructed, we attach a polling function which periodically polls the
connection for new stats (that's how the stats API works). Then we
make it easy for a measurement to extract this data, aggregate it into
time series and report it through Telemetry.

BUG=412787

Review URL: https://codereview.chromium.org/561803002

Cr-Commit-Position: refs/heads/master@{#300249}