net: show UI warnings for SSLv3.
This change causes sites using SSLv3 to show a "mixed content" level warning on
the lock icon and for the section in the connection status to go red.

There's no message added to the connection status because, if we plan on
removing SSLv3 by default in Chrome 40, then we want to merge this to Chrome 39
and thus there's no time to get a translation.

BUG=419870

Review URL: https://codereview.chromium.org/662883002

Cr-Commit-Position: refs/heads/master@{#300346}