Remove trivially-true checks in quic_framer.cc
There were several places in this file where an error code value
was checked to see if it was less than QUIC_NO_ERROR or
QUIC_STREAM_NO_ERROR, which are both 0. On some compilers this
causes a warning, because such checks are trivially true.

BUG=424334
TEST=net_unittests
R=rch@chromium.org

Review URL: https://codereview.chromium.org/662513007

Cr-Commit-Position: refs/heads/master@{#300368}