Add infrastructure for Chrome Actions
Chrome Actions are designed to be actions that live on the
toolbar, like extension actions, but are actually components
of Chrome. Examples of this can be ChromeCast, Translate,
Password Manager, etc. This change partially implements the basic
infrastructure for these actions.

(The main goal of this patch is to provide a shell that
the chromecast team can expand upon and use to prototype.)

* Create a generic ToolbarActionViewController, which can
  be implemented by a specific chrome action or to handle
  the general extensions case.
* (Mostly) abstract the concept of extensions out of the
  BrowserActionView, so it can be used for either extension
  actions or chrome actions.
* Create a simple ChromeActionsRegistry.
* Display chrome actions in the overflow menu of the browser
  actions container.

Known limitations:
* Chrome actions cannot be reordered, dragged, or shown in
  the main action toolbar (we need to consolidate chrome
  actions with the extension toolbar in the model somehow).

BUG=422976

Review URL: https://codereview.chromium.org/661493004

Cr-Commit-Position: refs/heads/master@{#300500}