Disable all pinch viewport machinery in site-per-process mode.
Currently, pinch viewport isn't designed with site-per-process in mind.
It will require some thought and work to get that to work. This patch
just ensures that we early-out in any non-trivial methods if the main
frame isn't local. Specifically, this patch is needed to early out in
registerLayersWithTreeView rather than ASSERT to keep the OOPI browser
tests from failing when virtual viewport is made the default.

BUG=425336

Review URL: https://codereview.chromium.org/671773002

git-svn-id: svn://svn.chromium.org/blink/trunk@184129 bbb929c8-8fbe-4397-9dbb-9b2b20218538