Fix support for multiple frames per SiteInstance in CreateProxiesForSiteInstance.
When a new cross-site frame is created, RFHM::CreateProxiesForSiteInstance walks the frame tree and creates proxies for the new frame in other nodes' SiteInstances.  This traversal currently hits this CHECK in RFHM::CreateRenderFrameProxy in cases where the new frame is added to a SiteInstance that already has other frames:

  CHECK_NE(instance, render_frame_host_->GetSiteInstance());

For example, if site A embeds two frames, B1 and B2, both pointing to site B, we hit this when we are visiting B1's frame tree node to create a proxy for B2, since B2's and B1's SiteInstance is the same.

This CL ensures that we aren't trying to create a proxy at frame tree nodes that share the new frame's SiteInstance.

BUG=None

Review URL: https://codereview.chromium.org/660083003

Cr-Commit-Position: refs/heads/master@{#300604}