[Android] Revert some code about file picker.
This patch is to revert some code about file picker.
The original implementation of file picker fails in some cases when
taking a picture from camera, then upstream made a fix to make it 100%
work but it needs to declare a content provider in AndroidManifest.xml,
otherwise, the whole file picker doesn't work.
We can't control how developer write their AndroidManifest.xml. If they
don't declare the content provider, it will crash when trying to take a
picture.

Related list of Chromium CLs:
   https://codereview.chromium.org/489053003
   https://codereview.chromium.org/591253002

TODO:This patch will be reverted when XWALK-2780 was resolved.

BUG=XWALK-2631, XWALK-2717