Pass needLocalPath flag to file chooser dialog.
blink::WebFileChooserParams has a flag to show whether the receiver
(WebFileChooserCompletion) needs local paths of files or not.

The CL conveys the flag to the file chooser dialog so that the dialog does not
create snapshot file when the receiver does not need local paths.

BUG=126902

Review URL: https://codereview.chromium.org/663293002

Cr-Commit-Position: refs/heads/master@{#300648}