chromeos: Choose monitor native mode as best match mode
Previously the UI chooses the highest pixel rate mode when multiple
modes exists for the same screen resolution.  If there exist a higher
pixel clock mode that's the same as the monitor's preferred mode, then
HDMI is first plugged in it will choose the preferred mode, but when
you later switch away to a different resolution and then switch back,
you will get the higher pixel clock version.
This patch changes FindDisplayModeMatchingSize() to choose the
preferred mode instead of the one with highest pixel clock.

BUG=chrome-os-partner:29514
TEST=Using the EDID of DELL U2410, there are two 1920x1200 modes
  1920x1200 (0x1d2)  154.0MHz -HSync +VSync *current +preferred
        h: width  1920 start 1968 end 2000 total 2080 skew    0 clock   74.0KHz
        v: height 1200 start 1203 end 1209 total 1235           clock   60.0Hz
  1920x1200 (0x1d3)  193.2MHz -HSync -VSync
        h: width  1920 start 2049 end 2256 total 2593 skew    0 clock   74.5KHz
        v: height 1200 start 1202 end 1205 total 1242           clock   60.0Hz

plug HDMI in, and obseve that it uses mode 0x1d2.
switch to 720p mode, then switch back to 1920x1200
observe that it does not use mode 0x1d3.

Review URL: https://codereview.chromium.org/606913002

Cr-Commit-Position: refs/heads/master@{#300976}