Implement the ability to layout test Web Worker-based Notifications.
Layout tests currently utilize a mocked NotificationPresenter implemented
in the test runner, but this won't be sufficient for Web Worker-based
notifications given that these won't be tied to a WebFrame.

Instead, pull up the layout test functionality to the browser process in
content_shell (only available when running layout tests!), similar to the
ability to request Notification permissions, so that we can allow displaying
and interacting with "displayed" notifications in layout tests as well.

The existing code path for Document-based Notifications is kept intact, but
will be removed once both Blink and Chromium switch over.

BUG=392187

Review URL: https://codereview.chromium.org/659043003

Cr-Commit-Position: refs/heads/master@{#301084}