cc: SolidColorLayerImpl::AppendSolidQuads uses visible content rect
Huge solid layers can cause a lot of memory to be allocated because
append quads function would use the content_bounds, not the
visible_content_rect.

PLI now uses visible rect, but SCLI is still using the content
bounds, this should probably be addressed in a follow up, but I'm
not too worried about this since it always worked this way and
this is mostly used by unit tests.  We can't simply pass the
visible rect because none of the unit tests set the visible rect,
and the visible rect is not passed from SCL to SCLI for tests that
don't directly interact with the SCLI.

BUG=423950

Review URL: https://codereview.chromium.org/675903002

Cr-Commit-Position: refs/heads/master@{#301134}