On Android, unlabeled link containing image should use image url as name.
This is just a heuristic to handle a situation where neither the link or
image are labeled. Since we can only export a single name attribute on
Android - unlike other systems where we export several fields and let the
assistive technology decide - this tries to do the sensible thing when
there's a link containing an image.

BUG=386623

Review URL: https://codereview.chromium.org/639773012

Cr-Commit-Position: refs/heads/master@{#301390}