cc: Make picture pile base thread safe.
This patch ensures that picture pile base is thread safe, since it is
not guaranteed that the ref is always adjusted on the same thread. In
particular, the worker thread (which refs picture pile on the cc thread)
might deref it (and delete it) on the worker thread.

R=reveman

Review URL: https://codereview.chromium.org/683543002

Cr-Commit-Position: refs/heads/master@{#301434}