[Sync] Fix bug where DeviceInfo is missing a backup timestamp
Remove last_synced_time_ member variable from ProfileSyncService.  This
variable acted as a cache of the underlying pref.  However, it was not
properly cleared in ProfileSyncService::DisableForUser when the
associated pref was cleared.  This caused a bug where device info may
not have a backup timestamp.

BUG=427096

Review URL: https://codereview.chromium.org/660023003

Cr-Commit-Position: refs/heads/master@{#301481}