MenuButtons enter Hover state upon Tap Down
MenuButton will now respond to TAP_DOWN gestures, entering the Hover state. It will return to a normal state if the gesture is cancelled. Tap gestures that activate the menu will still cause it to enter a Pressed state.

This experiment is behind the flag switches::kEnableTouchFeedback

TEST=MenuButtonTest.TouchFeedbackDuringTap, MenuButtonTest.TouchFeedbackDuringTapCancel
BUG=398404

Review URL: https://codereview.chromium.org/674163003

Cr-Commit-Position: refs/heads/master@{#301488}