cc: Added raster source.
This patch adds a raster source that is an opaque interface
for picture pile impl. It also removes the picture pile base
ref count and adds it to the raster source. This makes it
possible to use a member picture pile, instead of a ref
pointer with one owner.

What is missing here is the next step: removing picture pile
base in favour of keeping shared data on the pile and copying
it over. This should allow us to clean up the data and only
share what is actually shared.

R=danakj, enne, reveman

Review URL: https://codereview.chromium.org/666273002

Cr-Commit-Position: refs/heads/master@{#301667}