cc: Fix weakptr destruction of DelayedUniqueNotifier on shutdown.
During shutdown, the ThreadProxy is destroyed on the main thread, after
being Closed on the compositor thread.

The DelayedUniqueNotifier has posted weak ptrs inside the class, and
when the class is destroyed those weak ptrs are invalidated. But the
DelayedUniqueNotifier is destroyed on the main thread with ThreadProxy
while the weak ptrs were created on the compositor thread, leading to
CHECK failures.

This adds a new Shutdown() method to DelayedUniqueNotifier that will
immediately invalidate the weak ptrs (on the correct thread) and leave
the DelayedUniqueNotifier in a state that it's actually not able to
post/create new weak ptrs. This resolves the CHECK failures as when
the DelayedUniqueNotifier is later destroyed, there are no weak ptrs
left to invalidate.

R=vmpstr
BUG=428347

Review URL: https://codereview.chromium.org/692573002

Cr-Commit-Position: refs/heads/master@{#301881}