Remove unused render_gutter & gutter_width from menu_separator_win
Comments suggest this was Vista-only, I believe the small inset of
separator shown e.g. in
http://www.tomshw.it/guides/ictbusiness/business/20070102/images/vista_diary_1st_entry_classic_mode_with_start_menu_large.jpg .

This never appears to be set to true any more, I guess since the Aura-
style menus. (We're also looking to homogenize Vista & XP to be
as close as possible.)

R=sky@chromium.org
BUG=426573

Review URL: https://codereview.chromium.org/688073003

Cr-Commit-Position: refs/heads/master@{#301944}