Disable the polymer.load_topeka perf benchmark
There's something weird about this test or one of the other
test cases in the benchmark_smoke_unittest test suite that is
causing this test to hang when it is not run first.

I'm disabling this test for now so we can move on with the typ
cutover and get parallel test execution on the other platforms.

R=tonyg@chromium.org, ykyyip@chromium.org, dtu@chromium.org
BUG=428207

Review URL: https://codereview.chromium.org/683393002

Cr-Commit-Position: refs/heads/master@{#301946}