Support single-threaded impl-side painting
This makes the synchronous composite path support impl-side painting.
This initial version is intended for converting layout tests, as a real
version for the browser would commit to the active tree directly.

This path is currently not used anywhere (it DCHECKed) and the only
consumer after this patch is a unittest to verify that it works.

Depends on https://codereview.chromium.org/672673003/ so that
WaitForTasksRunningToComplete includes all visible tiles.

Depends on https://codereview.chromium.org/668123003/ for correctness,
as zero copy and masks aren't supported quite yet.

R=danakj@chromium.org,reveman@chromium.org,vmpstr@chromium.org
BUG=381919

Review URL: https://codereview.chromium.org/667793004

Cr-Commit-Position: refs/heads/master@{#301966}