IDL: union types support for method arguments
For method arguments, we can treat union type containers as the same way
as we do for IDL dictionaries.

A layout test is added for union type testing. Since we don't want to
generate union type containers just for testing, I used (double or
DOMString) in the test, which is already used in
core/animation/Timing.idl

BUG=240176

Review URL: https://codereview.chromium.org/691353002

git-svn-id: svn://svn.chromium.org/blink/trunk@184706 bbb929c8-8fbe-4397-9dbb-9b2b20218538