Fixing crashes in layout tests when virtual-viewport is enabled.
There are two fixes in this patch. The first is using the pinch viewport's
root layer when building a layer tree for printing in text. In virtual
viewport mode, the RenderLayerCompositor is no longer at the top of the
hierarchy as PinchViewport inserts layers above it. This caused exceptions
in devtools.js as layers were referencing parent layer id's that weren't
passed down.

The second is giving the transform layer in PinchViewport a name so that the
ASSERT_NOT_REACHED isn't triggered.

BUG=148816

Review URL: https://codereview.chromium.org/684373003

git-svn-id: svn://svn.chromium.org/blink/trunk@184736 bbb929c8-8fbe-4397-9dbb-9b2b20218538