[ServiceWorkerCache] Keep the ServiceWorkerCache alive when handling cache IPCs
With this CL, the ServiceWorkerCacheListener holds a reference to the
caches that it calls async functions on to ensure that the callbacks
are called.

Note: Keeping the cache alive longer does not alter cache deletion
behavior as cache deletion destroys the cache's backend but not the
ServiceWorkerCache object itself.

BUG=429161

Review URL: https://codereview.chromium.org/682383005

Cr-Commit-Position: refs/heads/master@{#302336}