Fix bug where AppInfoDialog was outliving ExtensionSettingsHandler
Fixed bug on Ubuntu Linux where the AppInfoDialog was outliving the
ExtensionSettingsHandler, causing Chrome to crash when the browser
window was closed while the dialog was open. This used to only happen
on Mac (where the dialog is disabled), but now happens on Linux Unity
due to Unity's always-accessible close/minimise/maximise toolbar.

TEST=Make Chrome fullscreen on Ubuntu with Unity enabled, then navigate
to chrome://extensions, click Details next to an app, and close Chrome
by clicking the X in Unity's fullscreen toolbar (not in Chrome). Chrome
should close gracefully and not crash.
BUG=427506

Review URL: https://codereview.chromium.org/683763003

Cr-Commit-Position: refs/heads/master@{#302566}