Fix folder icon animation in app list.
This CL fixes the animation of folders opening in the app list. This
regressed in r301308 because the ContentsView layer moved, exposing
a latent issue with the calculation of the top icon animation target.

The fix here is to use the view bounds as the target rather than the
layer bounds. The layer bounds are in the coordinate system of the
parent layer which is not what we want.

BUG=429079

Review URL: https://codereview.chromium.org/694743002

Cr-Commit-Position: refs/heads/master@{#302579}