android_webview: use chrome version information.
Replace the hardcoded version number used for the WebView's user agent
with the real version number of the corresponding Chrome build. This
uses the version information from src/chrome/VERSION; there is a pending
bug to move this out of chrome/ (http://crbug.com/345538) which will not
be impeded by this change as the location of the version info is
obtained from build/util/version.gypi.

BUG=297522

Review URL: https://codereview.chromium.org/695313002

Cr-Commit-Position: refs/heads/master@{#302614}