Remote assistance on Chrome OS Part V - It2MeHost Continue Window
This CL implements the ContinueWindow on ChromeOS, which will
be shown to the user for confirmation after 10 minutes into a Remote
Assistance session.

To implement the ContinueWindow, this CL introduces a new class
MessageBox, which is a re-usable component built on top of the
widget framework.

BUG=424908

Review URL: https://codereview.chromium.org/690183002

Cr-Commit-Position: refs/heads/master@{#302657}