Sync: Handle duplicating tab IDs in foreign session data.
This change prevents overwriting a foreign session SessionTab  with a data from another node if the SessionTab we already have has a newer timestamp. This ensures that if there are multiple tabs with the same ID, the tab that we end up associating with a window is the most recent one.

BUG=423501

Review URL: https://codereview.chromium.org/699033002

Cr-Commit-Position: refs/heads/master@{#302674}