Instrumenting copying string returned by GetRawDataResource to locate the source of jankiness.
Previous instrumentations showed that the jank is in this line:
*data = rb.GetRawDataResource(resource_id_).as_string();
And its average run time is 15.1 ms.
I also know that the expensive operation is outside of ResourceBundle::GetRawDataResourceForScale.
So, we have to suggest that it's indeed copying the string that is responsible for the jank.

BUG=422489

Review URL: https://codereview.chromium.org/702543005

Cr-Commit-Position: refs/heads/master@{#302686}