Disable -Wextra for all toolchains.
On clang, this flag is a no-op, and Blink code has a long way to go to be
clean with respect to these warnings on GCC toolchains. The GYP build has
never had -Wextra set, so let's just make GN in line with that.

BUG=424334
TEST=try bots
TBR=brettw

Review URL: https://codereview.chromium.org/692313002

Cr-Commit-Position: refs/heads/master@{#302693}