Fix double subtraction of sent_delta on sync tree activation.
When PushPropertiesTo sets the active layer's delta to delta - sent_delta,
LayerImpl attempts to mirror this to the pending twin. However, in doing
so it again subtracts pending_delta. This patch saves the sent_delta and
clears it before setting the scroll offset and delta on the active layer.

BUG=428327

Review URL: https://codereview.chromium.org/698003006

Cr-Commit-Position: refs/heads/master@{#302709}