Fix tests and enable virtual viewport on Android WebView.
In pinch virtual viewport mode, the viewport visible to Javascript is sized
to the width of the content and its height is that width multiplied by the
aspect ratio of the true viewport.  So I changed the Android WebView tests
to only do vertical scrolling, and also changed the width of the test page
to not be larger than the viewport (so that vertical scrolling also behaves as expected).

NOTRY=true
BUG=148816

Review URL: https://codereview.chromium.org/694113002

Cr-Commit-Position: refs/heads/master@{#302712}