Ignore ::first-letter from ancestors in grids and flexboxes
According to the grid and flexbox specs:
"the ::first-line and ::first-letter pseudo-elements do not apply to
grid/flex containers".
http://dev.w3.org/csswg/css-grid/#grid-containers
http://dev.w3.org/csswg/css-flexbox/#flex-containers

This was almost working right, except in the case that an ancestor was
setting the ::first-letter pseudo-element.

Modified RenderBlock::updateFirstLetter() in order to stop looking for
the first text child when you reach a grid or flexbox.
Added a few more cases to the current tests in order to check this
behavior.

TEST=css3/flexbox/flexbox-ignore-container-firstLetter.html
TEST=fast/css-grid-layout/grid-container-ignore-first-letter.html

BUG=430099

Review URL: https://codereview.chromium.org/686173006

git-svn-id: svn://svn.chromium.org/blink/trunk@184858 bbb929c8-8fbe-4397-9dbb-9b2b20218538