tools: Improve tough fast scrolling cases by running them at multiple speeds.
How fast we can scroll without checkerboarding is different for each
device. As it's not possible to pick one speed that is a good test on
all devices, instead run each tough scrolling page at a number of
speeds. The average amount of checkerbording is a good number to
monitor for improvements and the result of each individual run is also
very useful to determine how fast some specific device can scroll
without large amounts of checkerboarding.

BUG=

Review URL: https://codereview.chromium.org/699313003

Cr-Commit-Position: refs/heads/master@{#302788}