Don't use FormFieldData only as a key in a map
The FormFieldData was used in PasswordAutofillManagera::LoginToPasswordInfoMap merely as a map key, without needing the information in it. This CL changes the key type to int, and removes one IPC message completely in result.

Note: this CL is also relevant to http://crbug.com/377422 and http://crbug.com/118601, but does not directly contribute to their goals.

BUG=400186

Review URL: https://codereview.chromium.org/686653003

Cr-Commit-Position: refs/heads/master@{#302790}