Implement JS key event tester for apps v2.
This is intended to help us track down keyboard bugs in the web-app.
Eventually it will include a NaCl component for testing key events at
the PPAPI level, but for now it just uses JS key events.

Keypresses and releases are laid out horizontally, with a line break
every time the set of pressed keys is empty; this is similar to our
existing tool based on xev. We also trap "blur" to highlight keys that
remain pressed when the app loses focus.

BUG=

Review URL: https://codereview.chromium.org/682923002

Cr-Commit-Position: refs/heads/master@{#302823}