Fix no callback issue when other kind of streams try to set a larger buffer size.
The patch changes the code to set the buffer size if:
target buffer size is smaller than the current buffer size
current steam is the only active stream for both input and output.

This fix the issue that when input and output streams are sharing the same device, the output stream can set the buffer size to a large value which will results in no callback error for input stream.

BUG=428706
TEST=use whatever version of chrome
  # go to https://webrtc.googlecode.com/svn-history/r5497/trunk/samples/js/demos/html/pc1.html, click Start and allow the permission.
  # right click and select "show control"
  # you should hear audio if thing works.

Review URL: https://codereview.chromium.org/696833003

Cr-Commit-Position: refs/heads/master@{#302908}