Main-thread scroll if we have fixed-pos children and overflow:hidden.
If we are overflow:hidden and have fixed-pos children we would transition
to impl-side scrolling. This was exposing another bug (304810)
where on impl-side where we do not correctly stop body elements from 
scrolling. This lead us to scroll elements on the impl-thread even 
if they need to be main-thread scrolled.

Here we do not short-circuit the main-thread requirements if
the frame is overflow:hidden. This is equivalent to our old
implementation, and lets gmail continue to be on impl-thread,
which is what abarth had originally changed that caused this issue.

BUG=411767
R=vollick@chromium.org

Review URL: https://codereview.chromium.org/699183004

git-svn-id: svn://svn.chromium.org/blink/trunk@184891 bbb929c8-8fbe-4397-9dbb-9b2b20218538