Improve extensions version generation in GN
An error in the invocation of the exec_script call made GN re-run every time you compiled. The reason is that the VERSION file referenced in the deps of the script invocation had the wrong path. Since the file didn't exist, Ninja would always run the build step.

Forking scripts to run a GN-time is very unfortunate and slows everything down. This should only be done in extreme circumstances. There is already a template to do the version scripts at build time, we just need to move the definition of VERSION_FULL from the build file to the .in file.

Review URL: https://codereview.chromium.org/704903002

Cr-Commit-Position: refs/heads/master@{#302972}