DevTools: don't change current selection on Timeline flame chart when overview window changes
Shift+drag is more convenient for specifying the range of interest it's better to avoid also binding the selection to the current range in the overview pane.

Waterfall view will still populate its details view with the data for selected range in the overview pane.

BUG=426152

Review URL: https://codereview.chromium.org/695903002

git-svn-id: svn://svn.chromium.org/blink/trunk@184904 bbb929c8-8fbe-4397-9dbb-9b2b20218538