Context menu accelerators: R for Reload, P for Print?
Previous behaviour for "Reload" was accelerator "R" when the user clicks on the top part of tab and accelerator "L" when the user clicks on the page. Thus this was inconsistent. Now changing it to "R" everywhere is to my knowledge not causing any problems, because "Reload" and "Redo" both have accelerator "R", but they are not visible in the menu at the same time. "R" is also used for "Report an issue" and "Resume", but the same applies here.

Previous behaviour for "Print" was accelerator "R" (also sometimes "P"). Now changing it to "P" everywhere is also not going to cause problems, because "Paste" also has accelerator "P", but they are not visible in the menu at the same time. "P" is also used for "Play" and "Pause", but the same applies here.

BUG=113163

Review URL: https://codereview.chromium.org/680383005

Cr-Commit-Position: refs/heads/master@{#303000}