[Android] Support unfocusable container views.
If the ContentViewCore container view is not focusable (which can
happen when it's emebedded in an app via android webview), we
were clearing the RWHVA notion of focus when the RenderView was
swapped out. As the view is never focusable again, chromium never
considers this view focused and as such, text selection did not
function.

We support this type of container view by always treating it focused
from Chromium's point of view.

BUG=430859 (internal b/18066784)

Review URL: https://codereview.chromium.org/701233002

Cr-Commit-Position: refs/heads/master@{#303071}