[Most Visited] Only log impression metrics on initial NTP load.
The NTP view can refresh the data an arbitrary amount of times (due to changes in Top Sites, sync state, etc.). We observed that we can easily get the NTP to log impressions 10 times for a single NTP, just by staying on the page.

This change limits the logging of impressions to the initial load of the NTP, to avoid biasing the data too heavily. I think this will more closely track what's going on.

BUG=None
TEST=Manual

Review URL: https://codereview.chromium.org/701353004

Cr-Commit-Position: refs/heads/master@{#303077}