Cleanup: Use foreground rect in paintChildClippingMaskForFragments().
The clip rectangle was based on the foreground rectangle, while we
passed the background rectangle to PaintInfo.
Be consistent and use the foreground rectangle in both cases.

There was also a comment about skipping ourselves for border radius
clipping, although we did include ourselves. Removed the comment.

No behavioral changes intended, but added a test that would fail if we
had gone for the background rect instead of the foreground rect.

R=chrishtr@chromium.org

Review URL: https://codereview.chromium.org/706603002

git-svn-id: svn://svn.chromium.org/blink/trunk@184931 bbb929c8-8fbe-4397-9dbb-9b2b20218538