Add DeviceDisablingManager to manage device disabling
This CL moves the logic that determines whether the device is disabled
from the DeviceDisabledScreen into a new DeviceDisablingManager. For
now, this is a no-op refactoring. In a follow-up CL, the new manager
will also become responsible for disabling the device on the fly
when the appropriate signal is received during a device policy fetch.

BUG=425574
TEST=New unit tests; updated unit and browser tests

Review URL: https://codereview.chromium.org/692383005

Cr-Commit-Position: refs/heads/master@{#303105}