suppress link disambiguation popup when virtual keyboard requested
Currently Blink indicates the link disambiguation popup should be
shown even if one of the tap-able elements is the focused text input
element, which will also cause the virtual keyboard to be shown on
Windows touch input devices. Unfortunately showing the link disambig
popup breaks the virtual keyboard input. This CL suppresses the showing
of a link disambiguation popup whenever a virtual keyboard is requested
to be displayed.

Follow-on work is needed to tune the Blink heuristic to decide to not
show the link disambiguation popup based on the kind of selectable
elements within the ambiguous touch area.

BUG=428258

Review URL: https://codereview.chromium.org/712723002

Cr-Commit-Position: refs/heads/master@{#303501}