Revert paint-invalidation-container ASSERT
This CL reverts
https://codereview.chromium.org/414503002

The RenderLayer change was wrong and is the cause of the regression.
Unfortunately it's unclear what the path forward is: the ASSERT is
triggering, which means there are some bugs in the code, but it is
caused by having several definitions of paint invalidation container.
Fixing it will require aligning these definitions and it's unclear
which one is right.

BUG=415513

Review URL: https://codereview.chromium.org/706353002

git-svn-id: svn://svn.chromium.org/blink/trunk@185075 bbb929c8-8fbe-4397-9dbb-9b2b20218538