[Suggestions] Removing support for the delay before serving.
There was support for a delay before serving from cache, to allow for potential network requests to complete. We found through experimentation that "0", or no delay, is the thing that makes most sense.

The modification makes it so that there is no delay anymore. We will always serve from cache (if available), and issue a network request to update the cache.

Updated the method names so that it's clearer to the reader what's going on.

BUG=None
TEST=SuggestionsServiceTest

Review URL: https://codereview.chromium.org/683733011

Cr-Commit-Position: refs/heads/master@{#303663}