Fix content scaling bug on Mac
The root cause was that ImageTransportSurfaceFBO was using the pixel
size of the surface to early-out of surface reallocation. This is a problem
when scale factor changed and pixel size doesn't -- the change in scale
factor is never propagated to the underlying CALayer.

Move all logic for changing the size of the surface and for skipping
changes in the size of the surface to AllocateOrResizeFramebuffer.

Rename ambiguous sizes to be labeled as pixel_size or dip_size.

Remove a stray ScopedCAActionDisabler that does nothing (found by
inspection of nearby code).

Add myself as an OWNER for this path.

BUG=424713

Review URL: https://codereview.chromium.org/696823002

Cr-Commit-Position: refs/heads/master@{#302363}
(cherry picked from commit 712fa19c0fd1757fe041280086bf01f2e5ee3793)

Conflicts:
	content/common/gpu/image_transport_surface_fbo_mac.h

Review URL: https://codereview.chromium.org/717803002

Cr-Commit-Position: refs/branch-heads/2171@{#407}
Cr-Branched-From: 267aeeb8d85c8503a7fd12bd14654b8ea78d3974-refs/heads/master@{#297060}