gn: Support build directories outside the source tree.
The main use-case is to support using a fast SSD or ramdisk for the
build directory, but a slower persistent disk for the source code.

The general idea behind this change is modifying

 - RebaseSourceAbsolutePath()
 - SourceDir::ResolveRelativeFile()
 - SourceDir::ResolveRelativeDir()
 - GetOutputDirForSourceDirAsOutputFile()
 - PathOutput

to work with paths reaching out of the source directory.

Thanks to jam@ for the Windows-fixes.

BUG=343728
TEST=New unit tests + Unit tests pass.
TEST=`gn gen /ssd/out/Debug && ninja -C /ssd/out/Debug` work as expected.

Review URL: https://codereview.chromium.org/630223002

Cr-Commit-Position: refs/heads/master@{#303719}