Don't set WS_THICKFRAME if WS_EX_COMPOSITED is set -- it isn't a supported combination.
The previous fix to this function exposed some buggy behavior. This
change attempts to fix that. Some history:

Change 5150e6e (Sep 24, 2014) added this function.
Change 6561311 (Sep 30, 2014) tried to change it so that it would only
affect standard (WS_OVERLAPPED) windows, but actually caused it do nothing.
Change 26b57c2 (Nov 3, 2014) fixed the function so that it would actually
affect only standard (WS_OVERLAPPED) windows.
This was still incorrect because WS_THICKFRAME is not supposed to be set
for WS_EX_COMPOSITED windows. This change fixes that.

This function has never behaved correctly so it isn't known if this
change will finally fix it completely, but it does fix 431182 and
probably 432030.

BUG=431182,432030

Review URL: https://codereview.chromium.org/719453002

Cr-Commit-Position: refs/heads/master@{#303765}