command_buffer: Avoid uninitialized read in ShaderTranslatorCache
The key in ShaderTranslatorCache std::map is copied with memcpy and
compared with memcmp. However, the key is a class that does not
guarantee not having any padding.

Padding in ShBuiltInResources causes the comparison be invalid. Valgrind
reports uninitialized reads for this.

Fix by preserving usage of memcmp. Initialize the object
with memset. This should write any possible padding data to a
well-defined state.

Depends on an ANGLE fix where ShBuiltInResources is similarly
initialized with memset. The resource data is copied to the key
implicitly, and this may or may not copy the padding.

Review URL: https://codereview.chromium.org/654253002

Cr-Commit-Position: refs/heads/master@{#303817}