Devtools Animations: Show subtree animations for a given node
Animations are shown for the selected element making it difficult to
find and navigate between animations. This change introduces a persisted
checkbox setting to show all animations for the selected node's DOM tree.
When hovering on an animation section title the animating element on the
page is now highlighted. Clicking on the title to expand/collapse an
animation section is also enabled. The subtree setting is turned on by
default. This change is behind the Animation Inspection experiment.

BUG=419269

Review URL: https://codereview.chromium.org/682423002

git-svn-id: svn://svn.chromium.org/blink/trunk@185333 bbb929c8-8fbe-4397-9dbb-9b2b20218538