cc: Disable LayerTreeHostTestCrispUpAfterPinchEnds for being flaky.
When at a scale X and starting to pinch zoom to a lower scale, we create
a new tiling X/2. Whether we choose to draw with the X or X/2 tiling
appears to be flaky though. This matches what we've seen in the browser
manually as well.

TBR=enne,vmpstr
BUG=433208

Review URL: https://codereview.chromium.org/728833002

Cr-Commit-Position: refs/heads/master@{#304213}