Unify full-screen and fullscreen CSS.
ClientSession registers for full-screen events only while connected, which
can result in a race-condition whereby the property is not removed when the
app is windowed. When the apps v2 custom window frame was added, a separate
"fullscreen" property was added to the <body> node with the same semantics,
so the simplest fix is just to get rid of "full-screen" in favour of the
always-updated "fullscreen", and to update apps v1 to also maintain that
property.

Review URL: https://codereview.chromium.org/727563002

Cr-Commit-Position: refs/heads/master@{#304258}