Remove UsbDevice observers on ::Clear and service destruction.
This patch fixes a DCHECK in ObserverBase::AddObserver that catches us
adding the same observer to a UsbDevice more than once. This happens if
an app gets permission to access a device (ephemerally) and then the
application is restarted and the permission is granted again.

This issue could also result in a dangling pointer if the UsbDevice
object outlives the DevicePermissionManager service.

This patch contains the bug fixing component of
https://crrev.com/6ec7a5d370b1157b3ef5a37956f59d632f42c091

BUG=433122
R=scheib@chromium.org

Review URL: https://codereview.chromium.org/723363002

Cr-Commit-Position: refs/branch-heads/2214@{#65}
Cr-Branched-From: 03655fd3f6d72165dc3c9bd2c89807305316fe6c-refs/heads/master@{#303346}