Revert "Move some scroll invalidations to the paint invalidation phase"
This change partly reverts
https://codereview.chromium.org/498773007

This is blocked on fixing issues with incorrect paint
invalidation containers. Those are complicated matters
and require a lot of thinking to get right. It's also not
clear the amount of work required to fix this more fundamental
issue and we need something to fix the regression.

This change doesn't re-introduce the
updateFixedElementPaintInvalidationRectsAfterScroll
logic as it should be taken care by the paint
invalidation logic (or covered by some test case).

BUG=426507

Review URL: https://codereview.chromium.org/714083003

git-svn-id: svn://svn.chromium.org/blink/trunk@185463 bbb929c8-8fbe-4397-9dbb-9b2b20218538