Fix the way scrollbars are attached to LayerImpl.
A scrollbar is attached to the range of scrollable layers
between the clip and the scroll layer. The method that does this
had an early return but did not check for the changes in layers
in between, thus potentially missing some layers that need the
scrollbar.

This change attaches the scrollbar to all layers in the range,
whether scrollable or not.

BUG=419917

Review URL: https://codereview.chromium.org/726503004

Cr-Commit-Position: refs/heads/master@{#304561}