Viewport scrolls should always begin at the outer viewport.
This fixes an edge case on sites where the outer viewport is not scrollable.
In this case, when the use pinches in, the top controls can be manipulated.
When the user shows the top controls, the inner and outer viewports are
resized. This means that, until the main thread resizes the content layer,
the outer viewport becomes scrollable. This fix works by starting a scoll
delta meant for the inner viewport on the outer viewport.

BUG=432143

Review URL: https://codereview.chromium.org/729723002

Cr-Commit-Position: refs/heads/master@{#304617}