Always send enabled state in extension update checks
This changes the extension update system to always send a flag for
whether each extension is enabled or not when checking for updates, and
if it is disabled what the reasons are. The behavior before this patch is
to only send this metric when UMA is turned on. We already do not send cookies
with these requests, so the privacy review team gave sign off for this.

BUG=433376

Review URL: https://codereview.chromium.org/739033002

Cr-Commit-Position: refs/heads/master@{#304817}