Only disable main-frame scrollbars on Android.
Since virtual-viewport, we use the viewport scrollbars instead of the frame
scrollbars on Android. On desktop, we want to keep both sets, so we have
FrameView::scrollbarsDisabled that returns true if we shouldn't display
scrollbars. Previously, we keyed this off the scrollbar theme's
usesOverlayScrollbars() as Android was the only platform with overlay
scrollbars. However, with Mac getting overlay scrollbars and dev tools
emulation flipping the usesOverlayScrollbars() setting this is causing
problems.

Currently we'll key off an OS(ANDROID) define as this will need a merge but
this should really be behind a scrollbar theme setting of its own.

BUG=424102

Review URL: https://codereview.chromium.org/725373005

git-svn-id: svn://svn.chromium.org/blink/trunk@185615 bbb929c8-8fbe-4397-9dbb-9b2b20218538