cc: Rework how PLI cleans tilings up.
This patch changes how PLI cleans tilings up. Previously, we'd iterate
all tilings to find the ones that match the criteria, then iterate
those and remove them in PLI.

With this patch, we ask PLTS to do most of the work. We calculate
scales that use layer information and pass it down to PLTS to
do the cleanup. PLTS, in turn, constructs a list of things to remove
and removes it, also removing it from the passed twin sets.

This allows us to remove one function from PLI and one from PLTS,
as well as hides more of the implementation details of PLTS from PLI.

R=danakj, enne
BUG=433048

Review URL: https://codereview.chromium.org/742903002

Cr-Commit-Position: refs/heads/master@{#304972}