[SVG] Test framework for content model validation.
Test whether renderers are created for various parent/child SVG element
combinations (brute-forcing DOM instantiations across the whole SVG
domain).

This is the initial check-in of the framework, with a single element
validation for <filter>. Coverage is to be extended as more elements
start enforcing the SVG content model.

BUG=231555
R=fs@opera.com,pdr@chromium.org,schenney@chromium.org

Review URL: https://codereview.chromium.org/736263002

git-svn-id: svn://svn.chromium.org/blink/trunk@185697 bbb929c8-8fbe-4397-9dbb-9b2b20218538