Make display list canvases preserve the entire clips stack accross frames
This patch adds the ClipList class which stores a list of individual
clip operations on the GraphicsContext state stack, thus allowing
the full clip information to be accurately reproduced when transferring
the GraphicsContext to a new recording canvas.  Because this adds
overhead to GraphicsContext, clip tracking is disable by default
and only enable for 2d canvas contexts that use display lists.

BUG=408392

Review URL: https://codereview.chromium.org/724973002

git-svn-id: svn://svn.chromium.org/blink/trunk@185703 bbb929c8-8fbe-4397-9dbb-9b2b20218538