Get rid of LayerPainter::paintTransformedLayerIntoFragments().
Now we have a more unified code path for multicol and non-multicol.

Note that we no longer clip against the "paint dirty rect" here.
That was wrong anyway, since the "paint dirty rect" is in physical
pixels, while what it was clipped against (parent clip rectangle) was
not transformed. Instead of adding code that transforms or
inverse-transforms one of the rectangles, we can just as well trust
paintFragmentWithPhase() to do proper clipping when we finally get there.

We also make an effort not to meddle with layers that are not within
paintingInfo.rootLayer, since that's going to produce incorrect results
anyway.

R=chrishtr@chromium.org

Review URL: https://codereview.chromium.org/731933004

git-svn-id: svn://svn.chromium.org/blink/trunk@185777 bbb929c8-8fbe-4397-9dbb-9b2b20218538