SynchronousCompositorRegistry
The in_process directory connects browser and renderer
objects. These objects of course can have completely
independent lifetimes.

There was built in assumption that events happen in certain
order to make things work. In particular, it was assumed that
WebContentsImpl of a particular routing id is created on the
UI thread before the corresponding BeginFrameSource is bound
on the corresponding thread. This assumption is not true for
popup webview where the renderer side is created before
browser.

Create a SynchronousCompositorRegistry singleton that tracks
lifetime of these objects individually, and handles
connecting these objects only if all the objects are ready.

BUG=429853

Review URL: https://codereview.chromium.org/719653002

Cr-Commit-Position: refs/heads/master@{#305283}