Redesign of LayoutTest fast/canvas/image-pattern-rotate.html
This change redesigns fast/canvas/image-pattern-rotate.html to make it
use a text baseline and make it less sensitive to platform cofiguration.
Therefore it should no longer need to be rebaselined as frequently and
it should no longer fail when executed locally on a machine the differs
from the specific HW/OS configuration of the test bots.

BUG=435660

Review URL: https://codereview.chromium.org/745963004

git-svn-id: svn://svn.chromium.org/blink/trunk@185818 bbb929c8-8fbe-4397-9dbb-9b2b20218538