Create PageScaleAnimation after ImplTree is activated.
Since https://codereview.chromium.org/615953005/ the PageScaleAnimation
was created on the pending tree during a commit. This led to crbug.com/428327
since PageScaleAnimation is passed in the TotalScrollOffset of the viewport.
The viewport scroll offset on the pending tree will be different than the
offset on the active tree.

This patch effectively reverts most of https://codereview.chromium.org/615953005/.
Instead, the PendingPageScaleAnimation struct is given to the pending tree. Upon,
activation, this information is used in LayerTreeHostImpl to construct a new
PageScaleAnimation as before, using the active tree's state.

BUG=428327

Review URL: https://codereview.chromium.org/692213002

Cr-Commit-Position: refs/heads/master@{#305292}