Remove android_clang_dbg from PRESUBMIT.py
This bot is no longer in CQ and is subsumed by
android_clang_dbg_recipe.  It is a waste of resources to run two bots
by default.

Also, we are bringing up a replacement for android_clang_dbg_recipe
(see the linked bug), and this would allow us to recycle the shorter
bot name.

R=phajdan.jr@chromium.org
BUG=434516
TEST=git cl try: doesn't trigger android_clang_dbg

Review URL: https://codereview.chromium.org/751613004

Cr-Commit-Position: refs/heads/master@{#305460}