Remove calls to getRefCnt() in Chrome unit tests.
We've deprecated getRefCnt() for unique().  At this point, only unittests and
asserts are calling getRefCnt().  To really get rid of it, we'd like to turn
as many getRefCnt() calls as possible into calls to unique().

This CL is a fairly mechanical translation:
  EXPECT_EQ( 1, foo->getRefCnt()) ~~> EXPECT_TRUE(foo->unique())
  EXPECT_EQ(>1, foo->getRefCnt()) ~~> EXPECT_FALSE(foo->unique())

This results in a slight loss in precision in these modified tests, where we can no
longer distinguish between 2 and 3 refs without calling unref().

BUG=skia:3160

Review URL: https://codereview.chromium.org/757663002

Cr-Commit-Position: refs/heads/master@{#305477}