content: Refactor GPU memory buffer framework.
This reduces the complexity of the framework by removing
the indirection between GpuMemoryBufferImpl class and
GpuMemoryBufferFactory and limiting the number of
native GPU memory buffer types that can be used at
run time to one.

GpuMemoryBufferFactory is now used to allocate all native
GPU memory buffer types. The shared memory backed type is
unique in that it's the only type that can be allocated
without going through the factory.

BUG=

Review URL: https://codereview.chromium.org/732283002

Cr-Commit-Position: refs/heads/master@{#305502}