Add UMA counting for Sync Attachments response codes.
sync/util/data_type_histogram_unittest.cc had to be edited
because it initialized different types of histograms with
identical names. Previously this did not crash because the
tests had failed to invoke StatisticsRecorder::Initialize().
The HistogramTester this CL introduces in the attachments
unittests does that initialization, so things broke in fun
and unexpected ways.

BUG=436134

Review URL: https://codereview.chromium.org/728323003

Cr-Commit-Position: refs/heads/master@{#305514}