Fix tab focus in experimental app list.
This CL fixes tab focus being given to incorrect views. The start page's
WebView was taking focus and not returning it and the custom launcher
page's WebView was focusable when it was not being shown.

The dummy search box in the start page view still incorrectly gets focus
but this will be removed soon.

BUG=436327

Review URL: https://codereview.chromium.org/752393002

Cr-Commit-Position: refs/heads/master@{#305603}