IndexedDB: Persist lastModified time for constructed Files
The lastModified attribute of File objects is persisted
using the actual file directory entry metadata on disk.
When writing out a URL-backed file (created with new File())
we need to touch the file after streaming it to its
new home in the backing store to set the metadata.

TEST=https://codereview.chromium.org/725293008
R=cmumford@chromium.org
BUG=428338

Review URL: https://codereview.chromium.org/754013004

Cr-Commit-Position: refs/heads/master@{#305736}