Introduce StreamRegisterObserver.
We are planing to use Streams to send the data from the ServiceWorker to the browser process when FetchEvent.respondWith() is called. (crbug.com/436424)

When FetchEvent.respondWith() is called, we will create a blink::Stream and get the stream URL of it and send ServiceWorkerHostMsg_FetchEventFinished message with the stream URL to the browser process. And after that the ServiceWorker reads the body of the response and sends the data to the browser process using the Stream.

When we call blink::Stream::create() in the worker thread, StreamHostMsg_StartBuilding message is sent from the main thread.
And when the browser process receives this message, StreamRegistry::RegisterStream() is called.

But ServiceWorkerHostMsg_FetchEventFinished is sent from the worker thread.
So when the browser receives the FetchEventFinished message, StreamRegistry::RegisterStream() may not be called yet.

To support this condition we need to introduce StreamRegisterObserver.

When he browser process receives the FetchEventFinished message, it gets the Stream by calling StreamRegistry::GetStream(). And if it fails it registers a StreamRegisterObserver and waits for StreamHostMsg_StartBuilding message.

BUG=436424
TEST=content_unittests --gtest_filter=StreamTest.*

Review URL: https://codereview.chromium.org/755233002

Cr-Commit-Position: refs/heads/master@{#305792}