[WebView] Set back button to exit fullscreen.
This patch fixes a regression from K to L. Tapping the back button
while in fullscreen should exit fullscreen instead of the current
activity.

Key input handling in the WebView is, to put it nicely, anarchic.
This patch does not try to address any of the big issues, and instead
we attempt to follow K's behaviour as closely as possible, ie. key
back events in fullscreen:
1. are handled synchronously
2. are not propagated downstream to Chromium
3. do not fire WebViewClient.onUnhandledKeyEvent
4. are not propagated to the host app even if they
   override shouldOverrideKeyEvent

BUG=434045

Review URL: https://codereview.chromium.org/735883002

Cr-Commit-Position: refs/heads/master@{#305848}