Relay events from all HostZoomMaps to Chrome's ZoomEventManager.
Currently ZoomEventManager serves as a central aggregator for zoom
events originating from both the HostZoomMap and manual-mode zoom
events originating from the ZoomController.

With the advent of multiple HostZoomMaps, we can also use
ZoomEventManager to aggregate events from all sources.

In this CL we use ChromeZoomLevelPrefs, which already listens to the
zoom events from the HostZoomMaps, to relay those events to the
ZoomEventManager. One result is that WrenchMenu classes now only need
to subscribe to the ZoomEventManager to be notified of zoom events.

BUG=435017

Review URL: https://codereview.chromium.org/761583002

Cr-Commit-Position: refs/heads/master@{#305860}