PushEvents should be delivered with PushMessageData as the payload.
The PushMessageData object provides access to the data contained in a
push message in four different formats: as an ArrayBuffer, as a Blob,
as a JSON object or as a string. This gives developers more flexibility
in accessing the data they received from their server.

This was recently introduced in the specification:
    http://w3c.github.io/push-api/#pushmessagedata-interface
    https://github.com/w3c/push-api/issues/83

Follow-up patches will introduce support for the additional methods.

BUG=434808

Review URL: https://codereview.chromium.org/740003002

git-svn-id: svn://svn.chromium.org/blink/trunk@186120 bbb929c8-8fbe-4397-9dbb-9b2b20218538