Fix mis-type of getURLFromUUID()
Refer to where getURLFromUUID() is called, the parameter should be
std::string, not string. The compiler did not report this error because
"uring std::string" in net/quic/quic_stream_sequencer.h

Considering using in header is bad behaviour, so remove that together.

This change is part of CL
https://codereview.chromium.org/763833003

Conflicts:
	content/browser/indexed_db/indexed_db_backing_store.cc
	net/quic/quic_stream_sequencer.cc