Reland of Different fonts selected for width calculation Original Patch: https://codereview.chromium.org/593673003/ due to issues faced in XP for which test expectation was not updated and revert of patch was submitted
In some cases different fonts are selected to get width of string and
to draw text. In Font.cpp calls, font get changed due to fallback font
architecture. When making calls we need to check if Unicode is present
in primary font or not.
In case of Ellipsis, we need to check if Ellipsis Unicode is present in
primary font or not, if not then use "Dot" Unicode, this is also
suggested in CSS3 (http://www.w3.org/TR/2003/CR-css3-text-20030514/)

BUG=416425 TEST=fast/text/ellipsis-platform-font-change.html

Review URL: https://codereview.chromium.org/680213004

git-svn-id: svn://svn.chromium.org/blink/trunk@186162 bbb929c8-8fbe-4397-9dbb-9b2b20218538