GuestView: Decouple owner_web_contents from attachment
In the past, if a GuestView had an embedder WebContents, it was assumed to also
be attached to a container. These two concepts are no longer tightly coupled in
WTFrame.

In this CL, a GuestView's owner WebContents is the context that created the GuestView.

On attachment to a container, a GuestView's owner WebContents may change if the container
does not belong to the owner WebContents. The GuestView's attached state is managed separately.
An embedder WebContents is the owner WebContents if the GuestView is attached. Otherwise,
it's NULL.

A subsequent CL will introduce the notion of Detachment, where the attached state goes back
to false, but the owner WebContents does not change.

BUG=434226
TBR=nasko@chromium.org for trivial RenderWidgetHostViewGuest cleanup

Review URL: https://codereview.chromium.org/768693002

Cr-Commit-Position: refs/heads/master@{#306108}