[DevTools] Use proper frame metadata for screencast.
When OnSwapCompositorFrame is called, it has next frame metadata.
This frame is not yet rendered, so when we request a frame copy,
it actually has exactly the previous metadata.
As a result, screencast image is no longer stretched.

This patch also allows only a single frame processing at a time.

BUG=none

Review URL: https://codereview.chromium.org/770443002

Cr-Commit-Position: refs/heads/master@{#306175}