Fix NPE on rasterize and record benchmark with slimming paint enabled
The rasterize and record benchmark records the page with all graphics
context modes, including a disabled graphics context[1]. When slimming
paint is enabled a disabled graphics context will not emit an
SkPicture[2] which leads to a NPE in the DrawingRecorder.

This patch adds a null check of the picture in DrawingRecorder to
prevent this crash and prevent the display list from being built. An
assert has been added to the DrawingDisplayItem ctor to catch similar
cases in the future.

I considered adding a "!contextDisabled()" check in the DrawingRecorder
constructor but opted for the approach here because null graphics
contexts are an exceptional case and we shouldn't eat a pipeline stall
for it.

[1] https://code.google.com/p/chromium/codesearch#chromium/src/cc/debug/rasterize_and_record_benchmark.cc&q=RasterizeAndRecordBenchmark::RunOnLayer&sq=package:chromium&type=cs&l=117
[2] https://code.google.com/p/chromium/codesearch#chromium/src/third_party/WebKit/Source/platform/graphics/GraphicsContext.cpp&q=GraphicsContext::endRecording&sq=package:chromium&type=cs&l=549

TEST=rasterize_and_record perf test

Review URL: https://codereview.chromium.org/752113005

git-svn-id: svn://svn.chromium.org/blink/trunk@186249 bbb929c8-8fbe-4397-9dbb-9b2b20218538