[NaCl SDK] nacl_io: remove getcwd() implemenation (rely on glibc/newlib versions)
Previsouly we didn't have a working version of getcwd()
in newlib (libnacl) but we do now.

Secondly the resean for intercepting getcwd() directly
was due to gtest calling this in a static constructor.
gtest was recently changed to handle the failure of
getcwd() under NaCl so this workaround is no longer
needed.

BUG=437075

Review URL: https://codereview.chromium.org/756333005

Cr-Commit-Position: refs/heads/master@{#306220}