autoreload: don't suppress manual error pages
Simplify the ShouldSuppressErrorPage logic to *know* whether to suppress an
error page instead of guessing with heuristics. This avoids the situation where
a manual reload causes an OnStartLoading(NON_ERROR), which causes a
CancelPendingRequests(), which causes SuppressErrorPage() to falsely think
autoreload caused this load. The new unit test in
NetErrorHelperCoreAutoReloadTest demonstrates this sequence of events.

BUG=413606

Review URL: https://codereview.chromium.org/582243002

Cr-Commit-Position: refs/heads/master@{#306230}