Ensure layout on replaced elements in anonymous wrappers
When dimensions change on an ancestor we need to layout all the way down through
the percentage descendants. This means punching through any anonymous blocks that
sit between them. If a child is anonymous and its parent has a percentage height
then we need to layout that anonymous child so we can check for any percentage
children it is managing for the DOM parent.

BUG=414532

Review URL: https://codereview.chromium.org/764233002

git-svn-id: svn://svn.chromium.org/blink/trunk@186260 bbb929c8-8fbe-4397-9dbb-9b2b20218538