CodeGen: reuse far jumps when possible
For policies in sandbox_linux_unittests, the instruction count
reductions are:

BaselinePolicy: 420 -> 368 (-12%)
EqualityStressTestPolicy: 4050 -> 3316 (-18%)

The other test policies are too small to benefit, so they stay the
same size.

Also, I wanted to avoid needing to worry about

    jt = WithinRange(jt, kBranchRange);
    ...
    jt = WithinRange(jt, kBranchRange);

resulting in jump-to-jump sequences (e.g., if the first call were to
reuse a jump exactly 255 instructions away, and then jf needs to insert
a jump, which pushes the jump out-of-range).

There are a few ways I thought of solving this, but they seemed fairly
subtle and I didn't want to write unit tests for them either, so I went
with the current easy solution.

BUG=414363

Review URL: https://codereview.chromium.org/750773002

Cr-Commit-Position: refs/heads/master@{#306309}