Clean up TouchDispositionGestureFilter unit tests.
Previously, it was difficult to follow precisely where the location of
a gesture was coming from. This patch clearly separates the location of
synthesized gestures from non-synthesized gestures, and removes the
location from events for which location was unused, making the tests
easier to read.

Review URL: https://codereview.chromium.org/753743006

Cr-Commit-Position: refs/heads/master@{#306399}