Store binary flags in bool instead of int in WebInputEvent
Avoiding unnecessary size increase and improve readability by using the more
appropriate type. The comments in the code justifying the use of int type are
incorrect. In particular, these types are always initialized via the ctor that
does a memset on the sizeof, so padding DOES get initialized and shouldn't be
reported by valgrind as uninitialized memory.



 
BUG=435608

Review URL: https://codereview.chromium.org/743883005

git-svn-id: svn://svn.chromium.org/blink/trunk@186343 bbb929c8-8fbe-4397-9dbb-9b2b20218538