Implement style invalidation tracking
This patch wires up style invalidation tracking with support for
tracking invalidations created by the style invalidator. The goal
of this patch is to connect ScheduleStyleInvalidationTracking (which
has stacktrace, changedId, invalidationSet, etc.) and
StyleInvalidatorInvalidationTracking (which has reason, invalidationList,
and selectorPart). Instead of showing these details to the user we
synthesize StyleRecalcInvalidationTracking events from the above data.

The major changes in this patch are to add the two new events
(ScheduleStyleInvalidationTracking, StyleInvalidatorInvalidationTracking)
and wire up their logic and UI.

A screenshot of a Google query for science is at: http://pr.gg/inv34.png
(The wording will all change very soon so "Element is pending an
invalidation list" will change to something like "Style system invalidated")

I've added a new mechanism for dumping invalidation data and used it
in this patch. The other invalidation tests will be switched to this
approach in a followup patch.

BUG=410701

Review URL: https://codereview.chromium.org/731293006

git-svn-id: svn://svn.chromium.org/blink/trunk@186394 bbb929c8-8fbe-4397-9dbb-9b2b20218538