Fix FrameView::contentsInCompositedLayer() when the root layer is empty
When a page contains a composited negative-z element, we'll create an
extra level of GraphicsLayer to contain the actual background,
negative-z and foreground layers, and the RenderView's main graphics
layer is empty and not drawing contents.

FrameView::contentsInCompositedLayer() returned false in the above case,
causing missing invalidation of non-composited fixed-pos element when
the FrameView is scrolled.

Don't check if the RenderView's layer is drawing contents.

BUG=437123
TEST=compositing/overflow/fixed-scroll-in-empty-root-layer.html

Review URL: https://codereview.chromium.org/761853004

git-svn-id: svn://svn.chromium.org/blink/trunk@186457 bbb929c8-8fbe-4397-9dbb-9b2b20218538