Fixs a test bug exposed by the scheduler
Fixs a test bug exposed by the scheduler, where the promise throws a
NetworkError exception.  Normaly the test would end before the DOM
exception was thrown but with the scheduler (or a 100ms sleep in
BackgroundHTMLParser::sendTokensToMainThread) the order could change
leading to a spurious failure.

BUG=432129

Review URL: https://codereview.chromium.org/771653003

git-svn-id: svn://svn.chromium.org/blink/trunk@186504 bbb929c8-8fbe-4397-9dbb-9b2b20218538