PepperVideoDestinationHost: Don't make timestamps relative to Unix Epoch.
We would like PepperVideoDestinationHost to give out the same timestamp as is sent in to PepperVideoSourceHost. PepperVideoSourceHost uses TimeDelta::InSecondsF to create a PP_TimeTicks, but PepperVideoDestinationHost uses FromDoubleT to convert it back. FromDoubleT is relative to epoch (Jan 1, 1970), so it will add a big offset to the timestamp, causing it to be different from the one sent in.

Review URL: https://codereview.chromium.org/778993003

Cr-Commit-Position: refs/heads/master@{#306879}