Enlarge visible content rect by bounds delta.
Adjust the visible content rect in the root content layer
by the size of top controls. For this purpouse
we include bounds_delta in the visible rect calculaton.

This change is required for https://codereview.chromium.org/779113002
where we change the settings on Android platform to let the main
frame clip the content, after which the top controls should be
taken into account.

BUG=433687

Review URL: https://codereview.chromium.org/753243004

Cr-Commit-Position: refs/heads/master@{#306941}