Disk cache: make sure that data outside of a sparse block is dropped.
Non-sequential sparse data writes may end up writing before or after a given
sparse data block. Said data is only stored if the next write is sequential,
and it actually fills a given block.

This CL makes sure that data "before" the start of a given block is properly
discarded (as in, not accounted for when looking for stored data).

BUG=416895

Review URL: https://codereview.chromium.org/770153002

Cr-Commit-Position: refs/heads/master@{#307085}