Add framebuffer object related commands to command buffer.
glInvalidateFramebuffer, glInvalidateSubFramebuffer, glReadBuffer.

All other commands in this group have already been added.

Also, code generator has to be modified to remove the incorrect assumption
that the data pointer is the last arg of a function.  This assumption is not
true for glInvalidateSubFramebuffer.

Also, separate the bindings of glDiscardFramebufferEXT and glInvalidateFramebuffer so we have better control of them.

BUG=429053
TEST=gpu_unittests
R=kbr@chromium.org,bajones@chromium.org,sievers@chromium.org

Review URL: https://codereview.chromium.org/780433006

Cr-Commit-Position: refs/heads/master@{#307088}