Register for HID device notifications on Windows.
Instead of re-enumerating all HID devices whenever asked this change
registers with Windows to receive window messages whenever a HID device
is added or removed. This will allow Chrome to notify apps when devices
are connected so that they do not need to poll.

BUG=376719

Review URL: https://codereview.chromium.org/783773002

Cr-Commit-Position: refs/heads/master@{#307124}