Fix reloading of policy for extensions in public sessions.
Public sessions have a policy broker outside the public session that is
used to precache the policy for that public account before the session
is started (for example, right after enrollment). This broker also has
a tracker of the extensions for that account that updates the
SchemaRegistry, so that the ComponentCloudPolicyService fetches the
policy for those extensions before they are installed.

However, once the session starts and the extensions are loaded, the
SchemaRegistry will have the real Schemas that are used to validate
the policy data.

The bug is that the tracker at the broker is still overriding the
Schema with a stub whenever the policy is reloaded. This replaces
the valid Schema with an empty one, and all the policies get
filtered out due to this glitch.

BUG=437273

Review URL: https://codereview.chromium.org/761163002

Cr-Commit-Position: refs/heads/master@{#306173}
(cherry picked from commit 9b347e0d4de17c88860421831479c33968c1959a)

Review URL: https://codereview.chromium.org/787013002

Cr-Commit-Position: refs/branch-heads/2214@{#246}
Cr-Branched-From: 03655fd3f6d72165dc3c9bd2c89807305316fe6c-refs/heads/master@{#303346}