Changing setTimeout to not clamp delays to 1ms minimum
Changing setTimeout to not clamp delays to 1ms minimum, since a 1ms
clamp is not in the spec.  Clamping behavior for setInterval is
preserved, since TimerBase treats an interval of 0 as not repeating. A
layout test has been added to protect the behaviour of setInterval from
regressions.


ATTN Sheriffs: There is good probability that that this patch may have
unintended side effects.  For example layoyt tests that use setTimeout
and expect layout to occur are likely to become flaky (this is actually
a bug in the tests).  I tried to fix some of those here
https://codereview.chromium.org/644093003/

BUG=402694

Review URL: https://codereview.chromium.org/751093003

git-svn-id: svn://svn.chromium.org/blink/trunk@186786 bbb929c8-8fbe-4397-9dbb-9b2b20218538