[Telemetry] Disable testPowerMonitoringSync test on Mac & cleanup test macros
testPowerMonitoringSync is an end-end test for power measurement and requires hardware that supports power usage measurement.  The Telemetry unit tests are currently run on a VM on the bots and don't support power measurement.

Disable this test in the interim.

In addition this CL changes a bunch of uses of benchmark.Disabled/Enabled -> telemetry.unit_test_util.test.Disabled/Enabled in unit tests.

BUG=440666

Review URL: https://codereview.chromium.org/789043003

Cr-Commit-Position: refs/heads/master@{#307687}