cc: Push layer properties after tree/host properties during commit.
PictureLayer push properties will update tilings and maybe cause it
to choose a new tile size. If the viewport size is not pushed at that
time, it will choose the wrong tile size with GPU rasterization (where
tile sizes are based on the viewport size).

This test passes before/after this patch, but will fail with the
don't-swap-tilings patch without the changes in this CL.

BUG=387116

Review URL: https://codereview.chromium.org/796473002

Cr-Commit-Position: refs/heads/master@{#307786}