[Cast] Crash fix when converting audio in CastAudioSink.
This change rewrote a bit of the audio plumbing:
https://codereview.chromium.org/759383003/

Unfortunately, I was mistaken about the meaning of the value returned by
AudioConverter::ChunkSize().  With the correct understanding, I have
been able to not only resolve the root cause of the crash (a FIFO
underrun), but also greatly simplify the code.

BUG=440087

Review URL: https://codereview.chromium.org/787083005

Cr-Commit-Position: refs/heads/master@{#307800}