Main world CSP bypasses for javascript: URLs and font-src.
Resource fetches that are initiated from isolated world's should bypass
the CSP of the main world. This requires a lot of per-directive plumbing
to make sure the bypass is passed from the initiation point into
whatever task actually performs the fetch. This implements that plumbing
for javascript: URLs and for the font-src directive.

BUG=390329

Review URL: https://codereview.chromium.org/360863003

git-svn-id: svn://svn.chromium.org/blink/trunk@186916 bbb929c8-8fbe-4397-9dbb-9b2b20218538