Collect content:attr(...)-features in RuleFeatureSet.
We currently store the attribute names seen in content:attr(...) during
style application as a special case, and then afterwards ensure that we
have invalidation sets for those attributes. This is bad, because it's yet
another special thing we have to do when resolving a style.

This patch moves the collection of this feature to RuleFeatureSet, along
with the rest of the feature collection.

R=rune@opera.com
TEST=fast/css/invalidation/content-attr.html

Review URL: https://codereview.chromium.org/772103002

git-svn-id: svn://svn.chromium.org/blink/trunk@186954 bbb929c8-8fbe-4397-9dbb-9b2b20218538