Drop the VibrationManager service mojo connection after use.
Currently the Vibration API mojo service connection is kept alive at all times
after the first use (in render_blink_platform_impl) and is not explicitly
closed. As per crbug.com/440435 the fuzzer is complaining about a
"use after free" for a layout test using vibration. It appears the connection
and it's browser-side implementation is destroyed after shutdown..

It looks like we don't need to keep the connection anyway as the Vibration
API is a one-way api (i.e. from the renderer to the browser). This patch
addresses the above issue by resetting the mojo connection on the renderer
side after each call.

BUG=440435

Review URL: https://codereview.chromium.org/795703002

Cr-Commit-Position: refs/heads/master@{#307929}