Bump min PNaCl version to 14173.
We are changing the interface between Chrome and one of
the PNaCl translator nexes so that it uses the NaCl IRT
more (which is a more stable interface). However, this
means that we are not testing the old interface between
Chrome and the PNaCl translator nexe anymore on the chrome
waterfall, so the old interface may get broken. Bump the
min PNaCl version to try and get more users onto using the
new nexe, which uses the newer interface.

The new nexe also brings in a PNaCl translator based on
LLVM 3.5, so having that match the SDK (using LLVM 3.5)
is better for debugging (developer-only unstable debug
metadata has changed between 3.4 and 3.5).

BUG= https://code.google.com/p/nativeclient/issues/detail?id=3914

Review URL: https://codereview.chromium.org/789273002

Cr-Commit-Position: refs/heads/master@{#308030}