Remove kInstantProcess from renderer ContentSettingsObserver
The kInstantProcess flag was used to check if the current frame belongs to instant and should be whitelisted. This would cause layering troubles once ContentSettingsObserver is componentised.

Instead, this CL moves the check back to the construction of the observer, in ChromeContentRendererClient, and only records the Boolean flag for whitelisting in the observer.

BUG=384874

Review URL: https://codereview.chromium.org/797993002

Cr-Commit-Position: refs/heads/master@{#308106}