[Password Generation] Fix breakage that prevents filling of generated passwords
This bug was introduced in the move from RenderViewHost -> RenderFrameHost.
The Interactive UI test would have caught this, but was disabled for being
flaky. I'm actively working on fixing that test, but given the severity of the
problem it makes sense to check in the fix without waiting for the test to
be fixed.

BUG=440257

Review URL: https://codereview.chromium.org/790783003

Cr-Commit-Position: refs/heads/master@{#308173}