[Mac Extensions Toolbar] Make the container's animationEndFrame always valid
Previously, [BrowserActionsContainer animationEndFrame] only returned the
frame after animation once animation had occurred. Replace this with an always-
valid animationEndFrame method that returns the frame after animation, or the
current frame if the container is not animating.

Also, move the animation suppression for testing to the ToolbarActionsBar
so that it is platform-independent.

TBR=sky@chromium.org (moving animation suppression flag)
BUG=435518

Review URL: https://codereview.chromium.org/795663003

Cr-Commit-Position: refs/heads/master@{#308190}