[ServiceWorker] Create the blob handle from the BodyStreamBuffer in cache.put().
Currently we use BlobDataHandle as the body data of Request and Response object.
BlobDataHandle can handle only static data. It can't handle progressive data.
We will use BodyStreamBuffer for Response objects which will be created in FetchManager to support progressive loading.
https://codereview.chromium.org/786893004/

When cache.put() is called with the stream type response, it doesn't have the blob handle.
So we have to read the all body stream and create the blob handle before dispatching the put CacheBatchOperation.

This change depends on https://codereview.chromium.org/786893004/.

BUG=436424

Review URL: https://codereview.chromium.org/795513002

git-svn-id: svn://svn.chromium.org/blink/trunk@187129 bbb929c8-8fbe-4397-9dbb-9b2b20218538