views: Rename BookmarkExtensionBackground to something that does not have 'extension' on it.
Does not have much sense to have something that is not related to the
extensions system, have 'Extension' on its name.

Bookmarks system, and at least 'Bookmarks Bar' is a native concept and
native implementation.

Since this class is inheriting from views::Background and it is used to
paint the background of BookmarkBarView, we renamed it to
BookmarkBarViewBackground, so it is clear to what object it refers to.

BUG=None
TEST=There should be no regressions observed, no functionality changed.
R=sky@chromium.org

Review URL: https://codereview.chromium.org/804183002

Cr-Commit-Position: refs/heads/master@{#308500}