Put web app style frame behind a new flag.
The web app style frame is orthogonal to enabling the new bookmark app
creation flow, so it is being put behind a new flag. This will allow it
to be released at a later time and not delay the release of the new
bookmark apps flow.

BUG=442198

Review URL: https://codereview.chromium.org/806693002

Cr-Commit-Position: refs/heads/master@{#308544}