Remove NavigationAction
It's not really a useful abstraction. Most of the data it stores should be transient (m_eventTimeStamp, m_policy) or is redundant (m_resourceRequest). However, m_type is still useful, so instead of storing a NavigationAction on DocumentLoader, store a NavigationType.

The code that determines the NavigationPolcy and NavigationType lives in NavigationAction and is still necessary, so break that out into a couple of file-scope static helpers in FrameLoader.

Also, rename FrameLoader::loadWithNavigationAction() to startLoad() and have it take a FrameLoadRequest& instead of 5 parameters that were just read off of a FrameLoadRequest.

BUG=

Review URL: https://codereview.chromium.org/790383002

git-svn-id: svn://svn.chromium.org/blink/trunk@187310 bbb929c8-8fbe-4397-9dbb-9b2b20218538