Remove disabled tests session_restore.cold.startup_urls and session_restore.warm.startup_urls.
These are adding to the maintenance burden. If they need to be brought back to life, the measurement will need a substitute for CanRunForPage because that method is going away. https://codereview.chromium.org/808893002/

The CL that disabled the tests: https://codereview.chromium.org/738663002

BUG=389142,405386,440101

Review URL: https://codereview.chromium.org/809653004

Cr-Commit-Position: refs/heads/master@{#308737}