Main thread fractional scrolling for overflow div
When overflow div's scroll offset is changed, CompositedLayerMapping
compares the old and new scroll offset of scrollingContentsLayer to
decide if there is scrolling happening. Even though we save the
scroll offset as floating point in scrollableArea, but the comparison
is done on the integer type so small fractional scroll offset change won't
get propagated to compositor. This patch makes it that the detection of
scroll offset change is using floating point.

BUG=414283

Review URL: https://codereview.chromium.org/772403003

git-svn-id: svn://svn.chromium.org/blink/trunk@187338 bbb929c8-8fbe-4397-9dbb-9b2b20218538