Password Save prompt on in page navigation.
Check on in page navigation do we need to show password save prompt or save password. We use in page navigation since history.pushState causes it and we hope that in most cases when we have password form to store and in page navigation happens this is history.pushState.

BUG=357696

Review URL: https://codereview.chromium.org/802593005

Cr-Commit-Position: refs/heads/master@{#308797}