Tables with specific merge cell configuration render with extra height to tr elements.
Rows with only rowspan cells was not getting proper height. This row just
getting the one part of the existing height of rowspan cell.

Now, Calculating extra height required for the spanning cells in the row and
one portion of this height is allocated to it based on number of rows with
only rowspan cells having zero height.

R=jchaffraix@chromium.org, eseidel@chromium.org, leviw@chromium.org, rschoen@chromium.org
BUG=396655, 377518

Review URL: https://codereview.chromium.org/596823002

git-svn-id: svn://svn.chromium.org/blink/trunk@187448 bbb929c8-8fbe-4397-9dbb-9b2b20218538