[WebView] Move external video surface across container views.
When transitioning from/to fullscreen the WebView replaces the container
view. We need to move the ExternalVideoSurfaceContainer across container
views when this happens. Note that in fullscreen the video may still
be embedded in the page if the element in fullscreen is, for example, a <div>
containing the <video>.

BUG=440774,398485

TEST=Manual
1. Force use of external surface (set needs_external_surface_ to true
   in webmediaplayer_android.cc)
2. Load the following url on a test app using a WebView:
   http://robnyman.github.io/fullscreen/
3. Test with 2 different actions here:
   a. Go fullscreen on the <video> element by clicking on the fullscreen
      video control
   b. Go fullscreen on the <div> element by clicking on the “Fullscreen” button
4. Notice: video continues in fullscreen without frozen video and in the
   case of b the video still scrolls with the page

Review URL: https://codereview.chromium.org/799953002

Cr-Commit-Position: refs/heads/master@{#309000}