Clear matched properties cache for theme color changes.
A theme dependent color is represented by a keyword in CSS and the matched
property value won't change when the theme color changes. That lead to a
cache hit in MatchPropertiesCache so that a style recalc after changing the
theme wouldn't have an effect. So, we need to invalidate the cache on such
changes.

Changed method name and api layering to contain knowledge about necessary
changes in the StyleEngine.

BUG=441144

Review URL: https://codereview.chromium.org/804873003

git-svn-id: svn://svn.chromium.org/blink/trunk@187517 bbb929c8-8fbe-4397-9dbb-9b2b20218538