Fix bug where accessibility could "see behind" infobars.
This swallows all hover events that are unhandled by an infobar before
they reach the ContentView. This prevents a bug in accessibility mode
where if the user clicked on an empty space in the infobar then an
element from the webpage behind the infobar would be selected.

BUG=430701

Review URL: https://codereview.chromium.org/824763002

Cr-Commit-Position: refs/heads/master@{#309454}