Indexed DB: Replace transaction() overloads with IDL Unions, per spec
Now that IDL unions can support sequence types, replace overloads
of the IDBDatabase.transaction() method with a single method
accepting a union, which matches the spec.

(DOMStringList is not included in union in the spec, but is necessary since
sequence matching of iterable types is not yet fully implemented. This
was previously one of the overloads - there's no behavior change here.)

https://dvcs.w3.org/hg/IndexedDB/raw-file/tip/Overview.html#idl-def-IDBDatabase

BUG=430337
R=bashi@chromium.org

Review URL: https://codereview.chromium.org/814273002

git-svn-id: svn://svn.chromium.org/blink/trunk@187632 bbb929c8-8fbe-4397-9dbb-9b2b20218538