[Hotword] Don't show hotword options if search provider is not Google.
This CL moves the search provider check into the function which
determines how to display hotwording. Prior to this, it was only being
updated when the search provider was changed and other javascript could
come along and re-show the hotword section (even though it was now
supposed to be hidden). Now every time the hotwording should be
displayed, the search provider is checked beforehand. This is the
logically correct way to do it.

BUG=442938

Review URL: https://codereview.chromium.org/815683003

Cr-Commit-Position: refs/heads/master@{#309522}