Get basic SVG painting moved over to slimming paint
This patch brings basic SVG rendering to slimming paint. Most painting
now works with the exception of filters, clips, and masks. With this patch
we pass 2172/2700 of the tests in svg/ with slimming paint enabled, and
no longer crash in any test. I apologize ahead of time for this
mondrianic code review.

Major changes:
* A new GraphicsContext is now created for all nested painting (e.g.,
stamping out an SVG subtree for markers or patterns). In the future we
may want to change this pattern so the nested drawing bit is stored on
the GraphicsContext's m_recordingStateStack instead of creating a new
graphics context. This was a large change so it was left out of this
patch, but it would make filter/mask/clipPath support much simpler.
* FloatClipRecorder and FloatClipDisplayItem have been added because
SVG frequently clips to FloatRects.
* TransformRecorder has been changed to emit display items even when
the transform is identity. This fixed a crash where a cached drawing
was emitted but the transform changed and could not be looked up. A
bit has been added on the TransformDisplayItems to not save/restore
unnecessarily.
* DisplayItemList::replay has been changed to update the display
list. This only worked previously because callers happened to update
the paint list themselves.
* Three subdirectories have been added to the virtual suite for
test coverage.

BUG=412088

Review URL: https://codereview.chromium.org/770173004

git-svn-id: svn://svn.chromium.org/blink/trunk@187638 bbb929c8-8fbe-4397-9dbb-9b2b20218538