[ServiceWorker] Change the error handling of dispatching fetch event to the ServiceWorker.
In current implementation, when ServiceWorkerURLRequestJob
fails to dispatch the fetch event to the ServiceWorker,
the fetch request falls back to the network.

But it should return an error to the page. Because it is
hard to notice the unexpected worker termination. And also
CORS-preflight check may be bypassed using ServiceWorker
crash bugs.

BUG=437130

Review URL: https://codereview.chromium.org/826513003

Cr-Commit-Position: refs/heads/master@{#309657}