vda_unittest - Delete dismissed picture buffers after rendered.
For resolution changed videos, we may dismiss some picture buffers
in the middle of the stream. In the mean time, we may send picture
buffers to the rendering_helper. We can only delete the picture's
texture when it is dismissed and is not pending in the renderer.

BUG=319016
TEST=Run the test on lumpy and peach_pit.

Review URL: https://codereview.chromium.org/564373002

Cr-Commit-Position: refs/heads/master@{#309809}