[Telemetry] Switch empty action_name_to_run to 'RunPageInteractions' for 4 measurements:
+ ImageDecoding measurement:
This measurement is only paired with ImageDecodingMeasurementPageSet,
which contains pages with default RunPageInteractions which does nothing.

+ MemoryMultiTab measurement:
This measurement is not used by any benchmark. Though there are
ToughMemoryMultiTabPageSet which contains pages with default
RunPageInteractions which does nothing.

+PolymerLoadMeasurement measurement:
This measurement is paired with PicaPageSet & TopekaPageSet, both contain
only  pages with default RunPageInteractions which does nothing.

+RecordPerArea measurement:
This measurement is not used by any benchmark

BUG=418375

Review URL: https://codereview.chromium.org/804303003

Cr-Commit-Position: refs/heads/master@{#309832}