Prerender only the selected page in the app list.
On Windows, a warmup task is scheduled 60 seconds after Chrome startup
to improve app launcher show times. Part of this task is to cache text
labels on the app list.

This CL reduces the number of pages pre-rendered from (up to) 3 to just
1; the selected page. This is unlikely to have a big impact, since the
slowest parts of text layout are still touched, to layout items before
the prerender step. This CL is to verify that claim.

BUG=440224

Review URL: https://codereview.chromium.org/821113002

Cr-Commit-Position: refs/heads/master@{#309895}