Fix crash when signing out of managed account.
The bug: when signing out from a managed account, a dialog shows the
progress of clearing data. Occasionally, the activity showing the dialog
would be dismissed before the dialog was dismissed. When we then dismiss
the dialog, the app would crash.

This fixes the problem by using a DialogFragment instead of a Dialog.
The DialogFragment is managed by the Activity and is dismissed
automatically if the Activity finishes.

BUG=438876

Review URL: https://codereview.chromium.org/820503003

Cr-Commit-Position: refs/heads/master@{#309947}