Refactor the GestureEventQueue
Remove the |ShouldForward| public method, instead using a |QueueEvent|
method that forwards the event automatically, as necessary. This aligns
behavior with the TouchEventQueue.

Also allow pinch/scroll events to be sent when there is but one
in-flight scroll/pinch event already in the queue.

BUG=340075
TBR=sadrul@chromium.org

Review URL: https://codereview.chromium.org/803283004

Cr-Commit-Position: refs/heads/master@{#309950}