Clean-up: Remove base::TimeTicks::UnprotectedNow().
This method was added in r222396:
https://chromiumcodereview.appspot.com/23147002

However, with the current implementation of base::TimeTicks::Now(),
this logic appears to be bogus.  The comment in TrackedTime::Now()
thinks this method provides a lock-free execution of the timeGetTime()
function, but that is only the case on older systems without an
efficient QPC implementation.  Furthermore, for those older systems, the
overhead in processing for the "rollover-protected" timestamp is
negligible (~20ns, according to benchmarking; compare this to the
1.0 ms to 15.6 ms precision of the clock).

Review URL: https://codereview.chromium.org/829113002

Cr-Commit-Position: refs/heads/master@{#310039}