Fix a crash when a URLLoader request arrives during plugin instance deletion.
When a pepper plugin is removed from the document, it sends a
synchronous DidDestroy IPC to the plugin process. However, the plugin
can send IPCs which arrive while the renderer is waiting for the
DidDestroy response. If the plugin instance deletion is due to the frame
being deleted, these IPCs will be handled using the frame that is being
deleted, which eventually leads to a renderer crash. This CL fixes this
crash by treating the frame as being unavailable when handling URLLoader
requests while the plugin instance is being deleted.

Review URL: https://codereview.chromium.org/817443003

Cr-Commit-Position: refs/heads/master@{#310052}