Assigned ax role for table headers.
This patch makes <th> node have column header or row header.
The basic rule is as follows.
If header is preceded by header cells then it's a column header,
if it is preceded by cells then it's a row header.
If header is followed by header cells then it's a column header,
if it is followed by cells then it's a row header.

BUG=435081

Review URL: https://codereview.chromium.org/816423003

git-svn-id: svn://svn.chromium.org/blink/trunk@187901 bbb929c8-8fbe-4397-9dbb-9b2b20218538