Wrap ContentView in a parent view to fix infobar accessibility.
Previously, it was impossible to select infobars when using explore by
touch and traversing through the views using forward and backward
swipes. The problem was that ContentView contained both the
InfoBarContainer (a real View) as well as virtual views for all the
elements in the webpage -- but Android doesn't support ViewGroups that
contain both real and virtual views.

The fix is to add a FrameLayout, which is now a common parent of both
the ContentView and the InfoBarContainer. Here's the new view hierarchy:

  CompositorViewHolder
           |
      FrameLayout
       |        \
  ContentView  InfoBarContainer

Tab.getView() now returns the parent FrameLayout, rather than the
ContentView itself.

BUG=416663

Review URL: https://codereview.chromium.org/821393004

Cr-Commit-Position: refs/heads/master@{#310125}