Skip new trailing_whitespace_wrapping on Snowleopard Debug
The change in r187982 proved that the new tests for r187878 triggered an
existing problem and that the accompanying code change did not trigger a
regression. Mark the tests as skipped on Mac OS X 10.6 debug in order to
allow the code change from r187878 to be re-landed.

NOTRY=true
TBR=jsbell@chromium.org

Review URL: https://codereview.chromium.org/823683005

git-svn-id: svn://svn.chromium.org/blink/trunk@187991 bbb929c8-8fbe-4397-9dbb-9b2b20218538