Use DeviceMonitorWin to observe "UsbDevice" class devices.
The version of libusb embedded in Chrome does not support hotplug
events on Windows. As a workaround this patch sets up a class to
observer a DeviceMonitorWin instance that is registered to receive
events related to devices in the "UsbDevice" interface class. It
then triggers a full re-enumeration of connected devices.

BUG=411715

Review URL: https://codereview.chromium.org/837273003

Cr-Commit-Position: refs/heads/master@{#310422}