Cache the system font metrics on Windows in blink.
This will ensure that the renderer process does not crash on Windows 8+ if Win32K lockdown mode is enabled
for renderer processes. The crash occurs while retrieving the system font metrics which fails in this mode.

With this patch blink will retrieve the cached font metrics instead of querying the OS for the same.

BUG=442158

Review URL: https://codereview.chromium.org/840933002

Cr-Commit-Position: refs/heads/master@{#310497}