Reset mode on ancestors when a fullscreen node is removed.
Element::removedFrom() is called after a node has been detached and the
parent pointer has been cleared. Thus fullscreen mode has to be reset on
ancestors both in the removed tree and the tree it was linked into
before it was detached.

The patch fixes a corner case where :-webkit-full-screen-ancestor style
was still applied on ancestors after the parent of a fullscreen node was
removed.

BUG=447184

Review URL: https://codereview.chromium.org/820633002

git-svn-id: svn://svn.chromium.org/blink/trunk@188050 bbb929c8-8fbe-4397-9dbb-9b2b20218538