Refactor GradientGeneratedImage::drawPattern()
Instead of drawing gradient tiles explicitly, build a picture shader and use it to fill the whole area with a single drawRect call.

This yields significant performance gains when we draw lots of tiles, as Skia can generate a single gradient tile internally and use a fast bitmap shader for the actual fill.

R=schenney@chromium.org,senorblanco@chromium.org
BUG=447057,441025

Review URL: https://codereview.chromium.org/814573005

git-svn-id: svn://svn.chromium.org/blink/trunk@188072 bbb929c8-8fbe-4397-9dbb-9b2b20218538