Launch a new test instance for every test case in FormStructureBrowserTest.
This has a dual purpose:
1) it means less stuff to update every time you add a new test, and test names don't need to have some senseless number prepended to them
2) it will hopefully avert the timeouts that have plagued this test on several platforms

The one drawback is that it will take more time overall to run all tests. It takes about twice as long now. On my desktop, with 10 jobs, that's on the order of 40 seconds instead of 20 seconds.

If we need to disable an individual test, we can modify GetTestFiles() to leave out that one test.

BUG=423791,85276
R=mark@chromium.org, thestig@chromium.org

Review URL: https://codereview.chromium.org/811933002

Cr-Commit-Position: refs/heads/master@{#310623}