[MacViews] Disable WebDialogBrowserTest.SizeWindow.
The test crashes on Mac because it tries to create a views::Widget with
a WebContents view as parent, but the WebContents are still Cocoa based.

Disable the test for now since we don't yet use the Views WebDialog on
Mac. Once we eventually do, the above problem will be resolved.

BUG=404979, 447086

Review URL: https://codereview.chromium.org/833893004

Cr-Commit-Position: refs/heads/master@{#310648}