Upstream notification control for Google services
This upstreams a controller for notification related to sign-in and sync.

It has been cleaned up from the downstream implementation by splitting up
the parts related to sign-in and sync from the controller itself, and also
moved displaying some notifications into their respective controllers.

This included moving a drawable and a few strings upstream as well.

This CL is partly conflicting with
https://codereview.chromium.org/808153005/, but uses similar naming for things
like the NotificationManagerProxy so that it is easy to move forward with
merging the controllers in the future.

BUG=428882

Review URL: https://codereview.chromium.org/832173007

Cr-Commit-Position: refs/heads/master@{#310658}