App list: Fixed rendering issues upon opening the window in RTL mode.
This fixes two related bugs, one in the classic and one in the
experimental launcher in right-to-left languages. The classic launcher
would display a blank page upon opening the window; the experimental
launcher would display the start page layered on top of the apps grid.

The root cause of both bugs is http://crbug.com/446407; this is a quick
fix for merging back to M40 where this regressed.

BUG=441962
BUG=446978

Review URL: https://codereview.chromium.org/844883002

Cr-Commit-Position: refs/heads/master@{#310682}