::first-letter from ancestors should affect grids and flexboxes siblings
::first-letter from ancestors is already ignored in grids and flexboxes
since r184858. However, it's not propagated to the siblings, like in
floated or positioned elements.

From the spec:
"The first letter must occur on the first formatted line."

And:
"The "first formatted line" of an element may occur inside a block-level
descendant in the same flow."

http://www.w3.org/TR/CSS21/selector.html#pseudo-element-selectors

The patch fixes this and also updates the tests to check the new
behavior.

TEST=css3/flexbox/flexbox-ignore-container-firstLetter.html
TEST=fast/css-grid-layout/grid-container-ignore-first-letter.html

BUG=430099

Review URL: https://codereview.chromium.org/800843004

git-svn-id: svn://svn.chromium.org/blink/trunk@188131 bbb929c8-8fbe-4397-9dbb-9b2b20218538