LifecycleObservers can be destructed while LifecycleNotifier is iterating the observers
Note: This is a problem that happens in non-Oilpan.

LifecycleObservers can be destructed while LifecycleNotifier is iterating
the LifecycleObservers in LifecycleNotifier::contextDestroyed().
This is a problem because it confuses the iteration.

Specifically, the problem happens in AsyncCallTracker::ExecutionContextData.
See the crash report for more details.

At first, I tried to avoid AsyncCallTracker::ExecutionContextData from
getting destructed in contextDestroyed(), but noticed that it requires
a weird hack (e.g., delayed destruction etc). At the same time, I noticed that
we already have the same problem in ContextLifecycleNotifier and we've supported
destructions of ActiveDOMObjects while iterating the ActiveDOMObjects
(See ContextLifecycleNotifier.cpp). Thus I decided to use the same hack
for the LifcycleNotifier. In other words, this CL fixes the crash by
supporting destructions of LifecycleObservers during LifecycleNotifer::contextDestroyed().

As commented in the code, these hacks will be removed
once we ship Oilpan by default for LifecycleObservers.

BUG=446985
TEST=none, couldn't reproduce the crash.

Review URL: https://codereview.chromium.org/837233003

git-svn-id: svn://svn.chromium.org/blink/trunk@188142 bbb929c8-8fbe-4397-9dbb-9b2b20218538