Store serialized notification data as an Intent extra on Android.
When a displayed notification outlives the browser process, we need to
be able to wake up the browser and re-create all required objects and
data in order to fire the "notificationclick" event on the Service
Worker associated with a notification.

This patch serializes the notification data in a pickle, and passes
it as an opaque byte array to the Java side, where it's stored in the
Android intent. When the intent fires, we pass it back to C++ and
unserialize the data there.

BUG=90795, 432527

Review URL: https://codereview.chromium.org/839533004

Cr-Commit-Position: refs/heads/master@{#310770}