Move chrome_browser_backend's StopTracing logic to DevToolsClientBackend fully.
This paves the way to modify tracing_controller_backend to not rely on
browser_backend (and instead rely on DevToolsClientBackend) to issue
chrome tracing request.

This is accompanied with a rather major refactoring of
InspectorBackendList. Now the responsibility of creating and caching
InspectorBackendList belong solely to
DevToolsClientBackend/_DevToolsContextMapBackend.

InspectorBackendList is now in charge of filtering and creating wrapper.

BUG=423954,443897

Review URL: https://codereview.chromium.org/810883004

Cr-Commit-Position: refs/heads/master@{#310860}