[Android] Fix not showing bookmarks under "All items" when not signed in
On Enhanced bookmark mode, there was no bookmark shown under
"All bookmarks" even if user has bookmarks under "Uncategorized", when
user is not signed in.

The reason was that |BookmarksBridge#IsFolderAvailable| filtered out
desktop and others nodes if it's not signed in because we wanted to hide
them for the old bookmark UI.

Anyways, |BookmarksBridge#IsFolderAvailable| is not so relavent at
|BookmarksBridge::GetAllBookmarkIDsOrderedByCreationDate|'s context so
just remove it.

BUG=447360

Review URL: https://codereview.chromium.org/824913003

Cr-Commit-Position: refs/heads/master@{#311134}