CSS Parser : Move colorFromRGBColorString out of BisonCSSParser.
So that the call to parseColor gets wired to the new parser when
the flag is turned on (or to the old Bison if off). Thanks to that
we can safely remove parseColor methods from BisonCSSParser because
they are not used anymore (CSSParser::parseColor will call parseValue
in BisonCSSParser).

BUG=330389

Review URL: https://codereview.chromium.org/848743003

git-svn-id: svn://svn.chromium.org/blink/trunk@188284 bbb929c8-8fbe-4397-9dbb-9b2b20218538