::first-line should skip positioned elements
According to the spec:
"The first formatted line of an element may occur inside a block-level
descendant in the same flow (i.e., a block-level descendant that is not
out-of-flow due to floating or positioning)."
http://www.w3.org/TR/css3-selectors/#first-formatted-line

Modified condition in RenderBlock::firstLineBlock() in order to skip not
only floated elements but also positioned ones.

Add a new test to check this behavior and modified the counterpart test
for ::first-letter to increase coverage.

TEST=fast/css/first-letter-skip-out-of-flow.html

Review URL: https://codereview.chromium.org/835123003

git-svn-id: svn://svn.chromium.org/blink/trunk@188296 bbb929c8-8fbe-4397-9dbb-9b2b20218538