Push API: Replace buggy FutureData test helper with ResultQueue
Since FutureData didn't properly clear its state after a get(), it was possible
a piece of data provided via set() to be sent back twice to the test code!

Replaced it with a queue data structure that doesn't have these problems (and
also won't lose data in the event that test code causes two or more results to
be asynchronously generated before getting the results, though so far I think
we've been lucky on that front).

BUG=437277

Review URL: https://codereview.chromium.org/822203006

Cr-Commit-Position: refs/heads/master@{#311304}