Speed up Android resource extraction in the common case (nothing to do)
In the vast majority of cases (no update and no locale change), there is
actually nothing to extract at startup. This commit fixes the logic that
allows for a quick exit in this case. Instrumentation shows that the
cost of the useless checks that are currently done is very large (400+ms
on hammerhead with Lollipop for a cold start).

Add instrumentation to the Android resource extraction.

The instrumentation uses android.os.Trace instead of the more common
org.chromium.base.TraceEvent as the resource extraction is typically done before
the native library initialization, so TraceEvent calls are dropped (per
TraceEvent documentation).

BUG=445718

Review URL: https://codereview.chromium.org/808053005

Cr-Commit-Position: refs/heads/master@{#311312}