Add a test and a simple workaround for the WideViewport vs XHTML MP issue
As xhtmlmp.css sets viewport 'min-width' to 'auto' (via 'width:auto'),
we can not actually use it for setting the layout width in wide viewport
mode. Instead use already calculated 'page-defined width', which will
be 'device-width' in this case.

This is compatible with behaviour in JellyBean, but not in KitKat (which
was using 980px). We have agreed that conceptually, having an XHTMLMP
doctype is actually equivalent to viewport meta tag setting the viewport
to device-width.

BUG=447249

Review URL: https://codereview.chromium.org/837553005

git-svn-id: svn://svn.chromium.org/blink/trunk@188329 bbb929c8-8fbe-4397-9dbb-9b2b20218538