consumeDeclaration should check for CSSPropertyInvalid
This patch fixes consumeDeclaration to validate that the property is
valid. Note that validating the property actually applies in the given
rule type is left to the CSSPropertyParser, so we still allow e.g.
min-zoom with a style rule type or left with a viewport rule type.

Adding this checks fixes several tests, which were otherwise hitting
an assert when an invalid property would be set to initial or inherit.

BUG=330389

Review URL: https://codereview.chromium.org/791503003

git-svn-id: svn://svn.chromium.org/blink/trunk@188355 bbb929c8-8fbe-4397-9dbb-9b2b20218538