Change the WebCrypto behavior when exporting EC private keys that were imported without a public key.
Before such keys were exported to PKCS8 format, the publicKey would be missing in the ECPrivateKey.

Now instead the publicKey is written to exported bytes even if it was not provided during import.

This is more consistent with how export works for JWK, where the public key is a required parameter.

It also speeds up the structured (de) cloning of such keys, which uses PKCS8 format under the hood (and for which if the publicKey was not provided, BoringSSL computes it).

BUG=445635

Review URL: https://codereview.chromium.org/835633002

Cr-Commit-Position: refs/heads/master@{#311553}