Add ShowUnhandledTapUIIfNeeded for Contextual Search.
When a handler for a Click/Tap is called, we now check the DOM
version and after the handler is called we propagate a message
indicating if anything in the page changed (any DOM mutation).

This code is based on a CL by Yoichi (735743005).

BUG=412642

Review URL: https://codereview.chromium.org/819563004

git-svn-id: svn://svn.chromium.org/blink/trunk@188498 bbb929c8-8fbe-4397-9dbb-9b2b20218538