mac: Expose keychain access frequency to Telemetry.
Accessing the keychain is really slow. Telemetry tests use a mock keychain,
which is abnormally fast. This CL exposes the number of times the mock keychain
is accessed via a local histogram, and 4 telemetry measurements record that
number as a metric.
Those tests are:
 - speedometer
 - page_cycler
 - startup
 - tab_switching

Once this metric is collected, alerts will be added to the Telemetry dashboard
to help catch regressions to the number of times the keychain gets accessed.

BUG=

Review URL: https://codereview.chromium.org/714273004

Cr-Commit-Position: refs/heads/master@{#311744}