cc: Fix scroll offset after scroll animation is canceled
When a scroll offset animation is removed by the main thread and
a new scroll position is set, this new position should propagate
to the pending tree on commit, and then to the active tree on
activation. However, the removed animation continues running on
the compositor thread until activation, producing a scroll delta.
Currently, this scroll delta gets added to the new scroll
position (on both threads).

This CL prevents the scroll delta produced by a removed animation
from being added to the new post-animation scroll position.

BUG=243871

Review URL: https://codereview.chromium.org/845543007

Cr-Commit-Position: refs/heads/master@{#311895}