Workaround for TestMakeWebKeyboardEventWindowsKeyCode
Use ui::KeycodeConverter instead of X11 XKeysymToKeycode in
WebInputEventAuraTest.TestMakeWebKeyboardEventWindowsKeyCode.
A few bots (currently TSan and valgrind) still run an old X11
that doesn't provide the expected evdev-based code for some keys.
There used to be a workaround in ui::KeyEvent, but that had to
apply to every keystroke.

Also added a platform-independent version of the test.

BUG=449103

Review URL: https://codereview.chromium.org/853253002

Cr-Commit-Position: refs/heads/master@{#311945}