bluetooth: Shutdown BluetoothAdapter before DBus on ChromeOS.
The BluetoothAdapter is a lazy instance that may live due to references
until the shutdown of the browser. On ChromeOS the adapter and related
classes must release references to the DBusThreadManager before it is
shutdown. This patch adds an explicit call for that shutdown.

BUG=447387

Review URL: https://codereview.chromium.org/848613003

Cr-Commit-Position: refs/heads/master@{#311986}