Attempt to de-flake WebViewTest.Shim_TestResizeWebviewWithDisplayNoneResizesContent
In the test, webview.style.display = 'none' causes a callback to be put
on the js loop that changes the WebView's internal state. This needs to
run before webview.style.display = 'block'. Increasing the timeout
increases the probability (but does not guarantee) that will happen.

Review URL: https://codereview.chromium.org/800913003

Cr-Commit-Position: refs/heads/master@{#311992}