Remove the pointer reference in RenderSVGResource{Filter,Masker} functions
This patch removes the pointer references in RenderSVGResourceFilter and
RenderSVGResourceMasker. The pattern was first introduced in
http://trac.webkit.org/changeset/57886 to support swapping out the
context (an approach we'll re-introduce in followup patches, but with a
different pattern). There is no change in behavior with this patch as
it is just a cleanup.

Review URL: https://codereview.chromium.org/851313003

git-svn-id: svn://svn.chromium.org/blink/trunk@188575 bbb929c8-8fbe-4397-9dbb-9b2b20218538