Move origin url validation check to PermissionContextBase class.
Currently the validity check for the origin url and embedding url is performed
within the permissions related code on a case by case basis, i.e. for each API
separately. This patch moves the check to the base class to have a uniform
code path for all permissions in chrome. Also added a test.

BUG=

Review URL: https://codereview.chromium.org/815743003

Cr-Commit-Position: refs/heads/master@{#312217}