Allow use of checkstyle outside of main PRESUBMIT.py
This commit allows the use of checkstyle outside of the main PRESUBMIT
script in src. Since checkstyle can possibly be called from a subfolder
of src the absolute path to the java files are used instead of relative
to src.

We also honor the EXCLUDED_PATH blacklist for Java files.

BUG=446633

Review URL: https://codereview.chromium.org/839583006

Cr-Commit-Position: refs/heads/master@{#312287}