Don't match empty style rules.
If a style rule has no declarations, skip selector matching unless we are
matching for the inspector which expects empty rules to be included.

The render tree dump for the first-letter-first-line-hover.html test did
include a [color=#008000] for the first-letter renderer with this change.
Most likely because this change avoids a hover recalc since not trying to
match p:hover causes the hover flag not to be set on P. And that the
render tree is different after doing a recalc on the first-letter, for
some reason.

Instead of rebasing the result, I made it a ref-test.

R=esprehn@chromium.org
BUG=448610

Review URL: https://codereview.chromium.org/829393005

git-svn-id: svn://svn.chromium.org/blink/trunk@188731 bbb929c8-8fbe-4397-9dbb-9b2b20218538