Cast: Handle negative TimeTicks correctly in RTCP packet
RRTR in RTCP packets from the receiver contains timing information used
to reject old RTCP packets.

This NTP timestamp is translated to Chrome's internal TimeTicks object.
In the case when sender's NTP time is ahead of the receiver by a large
value, the TimeTicks value can be negative. This caused all RTCP packets
to be rejected.

This change fix this issue and added unit tests to cover this behavior.

BUG=447742

Review URL: https://codereview.chromium.org/860923002

Cr-Commit-Position: refs/heads/master@{#312335}