Moved DBus workaround to the run_gpu_test.py
This workaround was originally present in runtest.py and was applied to all tests (not just GPU). Since runtest.py is being deprecated and not being used on swarming, we need to move this workaround elsewhere. Although test_env.py is the most logical place for this, we chose to move it to run_gpu_test.py, because tests other than Linux GPU tests run just fine without this workaround and we would like to reduce risk breaking other tests with this change.

BUG=426897
R=kbr@chromium.org,maruel@chromium.org

Review URL: https://codereview.chromium.org/858103004

Cr-Commit-Position: refs/heads/master@{#312401}