Improve comments on WebContentsObserver RenderFrame* methods and rename.
This CL is the first step in a series of CLs to improve the WebContentsObserver interface and make it more consistent for RenderFrame(Host) notifications. It renames FrameDetached to FrameDeleted, to make it more generic and improved comments to reflect what objects the notification is about.

BUG=450799

Review URL: https://codereview.chromium.org/865723002

Cr-Commit-Position: refs/heads/master@{#312534}