Explicitly specify the ui::Compositor to DelegatedFrameHost
When DelegatedFrameHost was split off from RenderWidgetHostViewAura,
I had DelegatedFrameHost access the ui::Compositor via a call into the
client. This is harder to reason about than it should be, and leaves open
the possibility of keeping the DelegatedFrameHost as an observer of a
dead ui::Compositor.

Explicitly detach the DelegatedFrameHost from the ui::Compositor
when the ui::Compositor is being destroyed. Add a check to the
observer list, to ensure that it is empty at destruction.

BUG=(paying down technical debt)

Review URL: https://codereview.chromium.org/859423002

Cr-Commit-Position: refs/heads/master@{#312926}