Wrapped data reduction proxy initialization into its own class
The data reduction proxy code is initialized in profile code
in Chrome, primarily in ProfileImplIOData, and in the
AwBrowserContext and AwURLRequestContextGetter. This code is
rather involved. This change moves that initialization logic
to a new class, DataReductionProxyIOData, to avoid
duplication across platforms, and reduce the potential for regressions.

This CL also completely removes DRP support from the system network delegate.

BUG=429732, 447346

Review URL: https://codereview.chromium.org/778463002

Cr-Commit-Position: refs/heads/master@{#312975}