Fix to the GCM Account Tracker flooding on a stale refresh token
As a solution to a problem the Account Tracker will assume a stale
refresh token when it cannot get access token. Backoff logic in
token fetcher handles the case where service might not be reachale
and situations like that.

If the user signs in properly the account will be re-added.

BUG=452131
R=zea@chromium.org,dimich@chromium.org

Review URL: https://codereview.chromium.org/878713002

Cr-Commit-Position: refs/heads/master@{#313136}