Don't allow profile deletion in Metro mode.
Bad things happen if you delete the profile that started the Metro
process (at best, you crash), since relaunching Metro with a new user
is not supported. Hiding the button selectively depending on the
profile selected is a bit weird, so in Metro mode, hide it for all
profiles.

BUG=448352

Review URL: https://codereview.chromium.org/863063002

Cr-Commit-Position: refs/heads/master@{#313449}