cc: Remove SkSurface cache from Resource
The SkSurface that was being cached within the Resource was being held
and counted against skia's budget.  cc's budget is 512MB, and skia's is
~100MB, so cc would always blow skia's budget.  In the case of the ugam
page, we would end up redecoding, reupload and regenerating all of the
mip maps for each of the images on every frame.

Solution is to stop caching the SkSurface.

Additionally, we should investigate if this situation should even be
counted against skia's budget.  See crbug.com/452344

BUG=452138
BUG=447291

Review URL: https://codereview.chromium.org/874353007

Cr-Commit-Position: refs/heads/master@{#313547}