[Chromoting] Set ar_service_environment for Official builds
There is a check in build-webapp.py (line 197) that ensures that
non-dev builds have this set to 'prod'. Up 'til now, we've been
setting this manually (hence the check in build-webapp), but it
is better to set it correctly in the gyp file.

This also allows the bots to build Official builds without
triggering the build-webapp check.

BUG=

Review URL: https://codereview.chromium.org/886543002

Cr-Commit-Position: refs/heads/master@{#313558}