Allow new zoom bubble to override current one.
At present, if a user changes zoom on OSX while the zoom bubble is
already being shown, the new zoom bubble (with the most up to date
information) is not shown. This CL changes this behaviour by allowing
a new zoom bubble to be created, and replace the existing one.

Example: load a PDF and repeatedly click the zoom +/- controls in the
lower-right corner of the viewer.

BUG=444995

Review URL: https://codereview.chromium.org/849693006

Cr-Commit-Position: refs/heads/master@{#313564}