[Temp] Disable download_sdk_extras.py.
This script is only supposed to be used by Google's build slaves.
However, the way it checks whether it is being run by a Google slave
also works for our own build bots, but we do not have access to the
Google Storage bucket where the files are.

For now, disable the script to get our Android bots working again while
we deal with this upstream.