CSS Parser: Make consumeRuleList take a callback instead of returning a vector of rules
This patch makes changes consumeRuleList to take a callback to call on
each rule, instead of returning a vector of rules. This should be a
minor performance increase in the stylesheet case, since we will no
longer be building a vector.

This is to be followed by https://codereview.chromium.org/868593002/,
where we'll want to have namespace rules set on a style immediately
instead of queued up.

BUG=330389

Review URL: https://codereview.chromium.org/865783002

git-svn-id: svn://svn.chromium.org/blink/trunk@189164 bbb929c8-8fbe-4397-9dbb-9b2b20218538