Animations: Change transition fill behavior and representation of delay
Currently, transitions unnecessarily has a fill mode of 'both'. When a CSS
transition is replayed via the Web Animations API, new transitions cannot
be triggered on the same element since the transition was converted to
animation priority with the same fill behavior of 'both'. The fill mode cannot
simply be changed to 'none' since transition-delay needs to be represented.

The fill behavior is changed to 'none' and the transition-delay is represented
by adding an extra keyframe at the beginning of the animation if a delay is
present. This change also prevents the retargeting of compositor animations at
the end time since end-exclusivity means that no styles are applied at that time.
If we try to reach in and read the style at the end time we'll result in a crash.

This change does not change behavior and is therefore covered by existing tests.

BUG=437102

Committed: https://src.chromium.org/viewvc/blink?view=rev&revision=188712

Review URL: https://codereview.chromium.org/842563005

git-svn-id: svn://svn.chromium.org/blink/trunk@189171 bbb929c8-8fbe-4397-9dbb-9b2b20218538