ozone: evdev: Add a device event dispatcher that forwards to UI thread
This is part of a series moving IO on evdev devices to a fast thread,
since UI thread is too slow for the cursor.

Make a new class ForwardingDeviceEventDispatcher that wraps
EventFactoryOzone and can be safely passed to another thread. When the
other thread dispatches, we'll post to EventFactoryOzone on the UI thread
with the event.

BUG=449710
TEST=boot link_freon & move mouse

Review URL: https://codereview.chromium.org/874723002

Cr-Commit-Position: refs/heads/master@{#313875}