[Extensions] Simplify Extension Keybinding API Tests
The extension keybinding api tests currently include a lot of back-and-forth
by having an extension inject a script into a page, which changes the color, and
then injecting another script to see if the color did, in fact, change. Change
this to not rely on script injection.

Review URL: https://codereview.chromium.org/880523005

Cr-Commit-Position: refs/heads/master@{#313932}