Files.app: Split the startRequests method of metadata models into createRequests and startRequests.
Currently the startRequests method of metadata models has a responsibility to
create requests as well as to mark the cache as loading.

In the following patch, we need to apply the requests that are generated by
another cache set. Thus the method should be split into two methods (one is for
creating requests and another is for marking cache as loading).

BUG=410766
TEST=FileManagerJsTest.MetadataCacheSet,FileManagerJsTest.MetadataCacheItem
R=yawano@chromium.org

Review URL: https://codereview.chromium.org/895783003

Cr-Commit-Position: refs/heads/master@{#314270}