Reland: Reenable oobe spoken feedback test.
The test was disabled because it was flaky.  This CL tries to address that
and reenables the test.

This relands the previous attempt which was still flaky on slow enough environments (such as the memory bots).

BUG=346797
R=dtseng@chromium.org

Committed: https://crrev.com/e03f042d1c76262248e42334b1d0326b0e4319e7
Cr-Commit-Position: refs/heads/master@{#313467}

Review URL: https://codereview.chromium.org/866713002

Cr-Commit-Position: refs/heads/master@{#314298}