Fix browser launch on the login screen.
For the login screen, if there is some web content displayed, and the user clicks a link that is specified to open in a new window, a new browser window indeed opens, which is highly undesirable at the login screen.
To avoid this, a new delegate method was introduced, which should decide if we want to allow displaying any additional contents.

BUG=443096
TEST=manual

Review URL: https://codereview.chromium.org/812763003

Cr-Commit-Position: refs/heads/master@{#314311}