Prevent single-app kiosk mode auto-launch if device is disabled
If a device is disabled, it should show the disabled screen and not
start any sessions. Regular sessions, auto-launch for public sessions
and auto-launch for consumer kiosk mode respected this. Auto-launch
for enterprise kiosk mode takes a different code path that had
slipped through the cracks. The effect was that single-app kiosk mode
would auto-launch even if a device was disabled.

BUG=454005
TEST=New browser test

Review URL: https://codereview.chromium.org/882043006

Cr-Commit-Position: refs/heads/master@{#314330}