Add missing policies to to device policy proto
The |report_session_status| and |device_status_frequency| policies
were added on the server side in violation of our protocol:
Policies must always be added to the client-side proto first and
synced to the server.

In general, new policies should be added to the device policy proto
and policy_templates.json at the same time. As an emergency fix, this
CL gets a free pass for that rule as it is an emergency fix.

BUG=430908
TEST=None

Review URL: https://codereview.chromium.org/891323002

Cr-Commit-Position: refs/heads/master@{#314347}