[Extensions Toolbar] Ensure popup is closed at shutdown
Since the popup close process is asynchronous, it's possible that the popup
doesn't close by the time the platform delegate is deleted. If that's the
case, we need to be sure to close it, so that we do our cleanup (like
removing ourselves as an observer).

BUG=454361

Review URL: https://codereview.chromium.org/878813003

Cr-Commit-Position: refs/heads/master@{#314352}