Unify the three places that patch font loading for PDFium on Windows.
This also removed the code from content, which doesn't belong there because it's for a feature in Chrome.

In order to implement this in a unified way, I needed to create a public interface around ChildThread. A lot of the change is updating callers of ChildThread to use ChildThreadImpl instead.

BUG=448473

Review URL: https://codereview.chromium.org/897563002

Cr-Commit-Position: refs/heads/master@{#314367}