[Chromoting] Add ability to enable/disable GDrive support per-app in gyp.
Since not all AppRemoting apps will want or need GDrive support to be
enabled, we need to be able to control which apps enable this feature.

This cl adds 'app_features' to the AppRemoting gyp targets and passes
the list of features into the build_webapp script (so the manifest can
have the correct oauth scopes) and the webapp.

Also, because the 'cast' capability was "in the way" for these changes,
it has been converted into a feature as well (default: disabled).

BUG=

Review URL: https://codereview.chromium.org/895523004

Cr-Commit-Position: refs/heads/master@{#314408}