PlzNavigate: change behavior when navigating with a non-live RFH.
To fix a few tests the current behavior when navigating from a non-live RFH was changed so that the new, speculative one is immediately set as active matching what is done in RenderFrameHostManager::UpdateStateForNavigate. This replaced the behavior of simply bringing the current RFH back to life.

With this change 6 tests are fixed for PlzNavigate:

    RenderFrameHostManagerTest.WebUI
    WebContentsImplTest.NTPViewSource
    RenderFrameHostManagerTest.NewTabPageProcesses
    WebContentsImplTest.CrossSiteBoundariesAfterCrash
    WebContentsImplTest.CrossSiteComparesAgainstCurrentPage
    WebContentsImplTest.ShowInterstitialFromBrowserNewNavigationProceed

BUG=439423

Review URL: https://codereview.chromium.org/892123002

Cr-Commit-Position: refs/heads/master@{#314543}