Add PRESUBMIT check if modified UMA histogram name can be found (2)
This CL is an iteration of http://crrev.com/766713004, that got reverted
after http://crbug.com/445265 because it could not cope correctly
with structures of the type:

<histogram_suffixes name="SafeBrowsingStores.SBWhiteLists" separator=".">
  <suffix name="CSD" label="CSD"/>
  <suffix name="DownloadWhitelist" label="DownloadWhitelist"/>
  <affected-histogram name="SB2.DatabaseKilobytes"/>
</histogram_suffixes>

(in particular: affected-histogram-suffixes).

Limitation that should be solved now.

Original description ---------------------------------------------------

Add PRESUBMIT check if modified UMA histogram name can be found

This Presubmit checks if some diffs affect any UMA_HISTOGRAM_*
macro and, if so, checks if the histogram name is to be found
in either tools/metrics/histograms/histograms.xml or in the
CL diffs.

Addresses the problem of someone modifying code and
inadvertently forgetting a corresponding histograms.xml
adaptation, that has happened in the past.

BUG=434420, 445265

Review URL: https://codereview.chromium.org/885783007

Cr-Commit-Position: refs/heads/master@{#314557}