Create bookmark apps at the end of the process.
Prior to this change bookmark apps were created as soon as the user
started the process. If the user quit the process the app was then
removed. This made the flow feel strange, and also caused any matching
app that previously existed to be deleted if the user quit the process.

BUG=438019

Review URL: https://codereview.chromium.org/899443002

Cr-Commit-Position: refs/heads/master@{#314720}