Temporarily remove DCHECK that triggers during hardware teardown.
The logic that this DCHECK checks is correct, but during hardware
teardown the tracking information is also reset, even though the sync
RequestDrawGL has not set up the tracking info that it is resetting,
leaving the tracker in an unexpected state.

Ideally we should be differentiating the DrawGL calls required for
hardware teardown from the normal kModeProcess DrawGL calls, but that
seems to require touching framework code, making it difficult to fix.

BUG=451955

Review URL: https://codereview.chromium.org/902813003

Cr-Commit-Position: refs/heads/master@{#315002}