[CSS Grid Layout] Grid needs relayout if a style change mark it as dirty.
Whenever a style change affecting item's possition occurs the grid is
marked as dirty.

In some cases, like the one described in bug 452135's test, we reach a
race condition where painting is not yet completed but the style change
implies the grid is marked as dirty.

The GridPainter::paingChildren logic contains an assertion to verify the
grid is not dirty, which this specific case violates.

This StyleChange is classified to trigger relayout and full paint
invalidation in the RenderStyle logic, but we are doing it only for the
grid item. In order to avoid painting dirty grids, we need to mark the
grid container to relayout as well.

This change does not include test because it solves an issue caused by 
an unusual race condition not easy to reproduce with a layout test. 

BUG=452135

Review URL: https://codereview.chromium.org/887633002

git-svn-id: svn://svn.chromium.org/blink/trunk@189685 bbb929c8-8fbe-4397-9dbb-9b2b20218538