webrtc: Fixes WebRtcBrowserTest to use DOM attributes correctly.
WebRtcBrowserTest.RunsAudioVideoWebRTCCallInTwoTabs is assuming that DOM attributes are own properties on a JS obejct, which is not correct.  Blink (binding) team is working to move DOM attributes from own properties to properties on prototype chains, and then this test will break.

This CL fixes the test so that it stringifies not only own properties but also attributes on prototype chains.

BUG=43394

Review URL: https://codereview.chromium.org/904823002

Cr-Commit-Position: refs/heads/master@{#315198}