Quick fix for ChromeVox braille flakiness.
The libluis wrapper depends on the load event to fier when
the nacl component is loaded to initialize the braille
translator.  Somehow, this doesn't fire during startup unless
a message is sent to the component which causes braille translation
to not work at all, resulting in no braille output on the display
at all.

This is a minimal workaround intended for backporting to the M-41 branch while
the real cause can be investigated and addressed.

BUG=455889
R=dtseng@chromium.org

Review URL: https://codereview.chromium.org/905353002

Cr-Commit-Position: refs/heads/master@{#315319}