Script moving fixes related to context documents.
There were 2 bugs:

1) Moving script element to a new document whose context document is the same as
the current document: we ended up doing ScriptRunner::movePendingAsyncScript to
itself and that does the wrong thing (just removes the script).

2) Moving script element to a new document which is not its own context
document, and then moving it somewhere else. The script was not where we
expected it to be (it was in the context document, we expected it to be in the
non-context document).

BUG=456059

Review URL: https://codereview.chromium.org/913473002

git-svn-id: svn://svn.chromium.org/blink/trunk@189886 bbb929c8-8fbe-4397-9dbb-9b2b20218538