[Ozone-Dri] Move the ForceInitializationOfPrimaryDisplay from ScreenManager
Forcing the configuration of a display relies on the presence a
DRM device. Since we're moving towards hotplugable DRM devices this
removes the assumption of a DRM device from ScreenManager. It will
then allow us to initialize ScreenManager early, but add DRM devices
late, when the hotplug events are received.

BUG=427959

Review URL: https://codereview.chromium.org/897593004

Cr-Commit-Position: refs/heads/master@{#315572}