NaCl cleanup: Move ppb_nacl_private.h into components/nacl/renderer/
Since PPB_NaCl_Private is both implemented and consumed in
components/nacl/renderer/, there's no need for the interface to be
defined in ppapi/.  The fact that this is a PPAPI interface at all is
just a holdover from when the code for the NaCl trusted plugin lived
outside the Chromium repo.

We can drop the ppb_nacl_private.idl file.

Benefits:

 * This reduces the friction for changing ppb_nacl_private.h.  It's no
   longer necessary to change ppb_nacl_private.idl and re-run the
   generator.

 * This removes PPB_NaCl_Private from the PPAPI shims, reducing code
   size.

 * ppb_nacl_private.h no longer has to follow the PPAPI C header
   style.  For example, it would no longer need to define its own
   "struct PP_NaClFileInfo" type -- it could share a definition that's
   marshallable over Chrome IPC.

Changes to ppb_nacl_private.h:

 * Add comment to explain the status of this interface.
 * Remove unused "*_1_0" variants of #define and typedef.

BUG=239656
BUG=251460
TEST=trybots

Review URL: https://codereview.chromium.org/911463003

Cr-Commit-Position: refs/heads/master@{#315641}