Pepper: Always call status_cb in FileDownloader.
FileDownloader assumed that implementations of blink::WebURLLoader would always
call didFail and didFinishLoading, which is not the case. In the cases where
didFinishLoading was not called, FileDownloader would never run the status
callback, preventing any error reporting from happening.

This change deletes the implementation of blink::WebURLLoader in didFail and
invokes status_cb, allowing error reporting to work properly, and for memory to
be cleaned up.

BUG=440180

Review URL: https://codereview.chromium.org/909233002

Cr-Commit-Position: refs/heads/master@{#315649}