Autofill: Do not infer labels from divs with autofillable fields.
The label finding algorithm looks at previous divs to see if they contain text, and uses that text as the label. While this works for many webpages, not all webpages are structured this way. For a given field, when the neighbor divs contain input fields, they are likely not to be the label text.

BUG=454366

Review URL: https://codereview.chromium.org/906443002

Cr-Commit-Position: refs/heads/master@{#315688}