[New Multicolumn] Preparatory work for nested multicol support.
Introducing the class MultiColumnRow. A RenderMultiColumnSet will have one
or more of those. For now, only one row is supported. Moved a lot of code
from RenderMultiColumnSet into MultiColumnRow.

Additionally, since it would just be just cumbersome and waseful to store the
flow thread rectangle in RenderRegion (or RenderMultiColumnSet, for that
matter) now, create a rectangle on the fly when requested.

No behavioral changes are intended.

R=jchaffraix@chromium.org
BUG=447718

Review URL: https://codereview.chromium.org/883293004

git-svn-id: svn://svn.chromium.org/blink/trunk@189966 bbb929c8-8fbe-4397-9dbb-9b2b20218538