ozone: evdev: Delay notifying DeviceDataManager until all devices added
Using a counter to track the number of device additions requested vs
completed. When the all the devices have converters associated with them
then dispatch the events to indicate that the set of
touchscreens/mice/touchpads/keyboards is updated.

This will result in the DeviceDataManager only being updated when the
full set of scanned devices are ready. To ensure that the additions and
subtractions are paired it was necessary to run the OpenInputDevice callback on
failure too.

When removing devices the notification happens immediately (no change.)

BUG=455346
TEST=Interactive testing including hotplugging devices on link_freon
CQ_EXTRA_TRYBOTS=tryserver.chromium.linux:linux_ecs_ozone,linux_chromium_chromeos_ozone_dbg_ng

Review URL: https://codereview.chromium.org/915723004

Cr-Commit-Position: refs/heads/master@{#315809}