Chromium-side patch to rename "--dump-render-tree" to "--run-layout-test".
At this point, "--run-layout-test" is the flag that is actually functional, and using "--dump-render-tree" simply causes the "--run-layout-test" to be appended to the command line.

A subsequent Blink patch will change the scripts in Blink to use "--run-layout-test" instead of "--dump-render-tree".

This patch was originally uploaded here and then reverted: https://codereview.chromium.org/899103002/ . This patch introduces fixes for the issues that led to the original being reverted.

BUG=453252

Review URL: https://codereview.chromium.org/891903006

Cr-Commit-Position: refs/heads/master@{#315840}