Don't attempt to retrieve system font metrics on Windows if we fail to get the non client metrics from the OS.
On Windows 8 in win32k lockdown mode retrieving the non client metrics via the SystemParametersInfo API fails. While
we should not be entering that code path as the font metrics are cached during renderer startup, there are still some
crashes seen in the wild in extension processes in the lockdown mode on Windows 8. It is not clear why extensions enter
this code path.

In any case we should not be trying to retrieve font metrics if the LOGFONT structure is invalid. This patch achieves that. We set the font size as s_defaultFontSize in the LayoutThemeChromiumFontProvider::systemFont function.

BUG=442158

Review URL: https://codereview.chromium.org/921513003

git-svn-id: svn://svn.chromium.org/blink/trunk@190008 bbb929c8-8fbe-4397-9dbb-9b2b20218538