Shadow lists with different inset style shadows are not interpolable.
The transitions spec [1] says that, for shadow lists, if any pair of
shadows have different inset styles, then the entire shadow list is
not interpolable.

Currently, a pair of shadows with matching inset style may still
interpolate, even if there is a mismatching pair of shadows elsewhere
in the list. This patch changes this behaviour to follow the spec.

Firefox conforms to the spec in this case.

[1] http://dev.w3.org/csswg/css-transitions/

BUG=310982

Review URL: https://codereview.chromium.org/919643007

git-svn-id: svn://svn.chromium.org/blink/trunk@190029 bbb929c8-8fbe-4397-9dbb-9b2b20218538