Add mobile optimized hint to ContentViewCore
Currently, we guess if a page is mobile optimized by looking at rendered
frames. This is used to enable/disable double-tap zoom.

This change passes that information up to the ContentViewCore along with
the other frame information and adds a getIsMobileOptimizedHint() to
ContentViewCore to provide access to it. Since this value is just a
guess based on the frame info, it may change for repeated calls on the
same page (particularly between starting the load and receiving the
first 1-2 frame's infos).

BUG=444160

Review URL: https://codereview.chromium.org/902083003

Cr-Commit-Position: refs/heads/master@{#316046}