IndexedDB: Ack the receipt of blob's.
A recent change (crrev.com/813103002) mistakenly dropped the code to ack the
receipt of the blob resulting in a memory leak lasting the lifetime of the
IDB connection. This is the fix.

Tried writing a test for this, but considered not worth the effort as it would
require exposing lots of new API's. It was manually tested by adding a DCHECK
on the Chrome side to assert that it's blob map was empty at destruction.

BUG=457502

Review URL: https://codereview.chromium.org/917063005

git-svn-id: svn://svn.chromium.org/blink/trunk@190086 bbb929c8-8fbe-4397-9dbb-9b2b20218538