Remove inconsistent infinite rects to fix disappearing text in patterns.
LayoutRect has two infinite rects:
LayoutRect::infiniteRect - when pixel snapped will equal infiniteIntRect
LayoutRect::infiniteIntRect - almost infinite, modulo saturated arithmetic

The first of these is strange and surprising and will be removed in a
followup. An SVG patch[1] introduced a check for infiniteRect in
updatePaintInfoRect when it should have checked for infiniteIntRect.
This mistake could cause the PaintInfo rect to be incorrect when text
was transformed by a pattern because pattern painting relies on an
infinite rect passed from SVGPaintContext::paintSubtree.

This patch removes all but a single remaining user of LayoutRect::infiniteRect,
and crbug.com/440143 is tracking the last removal.

[1] https://chromium.googlesource.com/chromium/blink/+/f9bce18c04f077d1799c24330fa38f0f23f22604

TEST=svg/custom/transformed-text-pattern.html. Due to the transforms,
kerning is slightly different between the test and expectation so only
a single letter is used ('A') instead of the word PASS.

BUG=453295

Review URL: https://codereview.chromium.org/924733002

git-svn-id: svn://svn.chromium.org/blink/trunk@190134 bbb929c8-8fbe-4397-9dbb-9b2b20218538