PasswordAutofillAgent: only check visibility if needed
PasswordAutofillAgent::SendPasswordForms is called twice during page load:
On DidFinishDocumentLoad, when it is interested in parsed HTML forms,
and on DidFinishLoad, when it is interested in those forms, which are also visible.

Currently, the method does the visibility check all the times, and throuws out the result in the first situation.

This is not only wasting computational resources, but possibly causing crashes due to not finished rendering computations.

This CL fixes that, and only does the visibility check if needed.

BUG=457243

Review URL: https://codereview.chromium.org/925983002

Cr-Commit-Position: refs/heads/master@{#316218}