[Backport] Do not overwrite command when using the run_pie wrapper.
Some change starting in M41 has started causing an existing bug in
GetAndroidToolsStatusAndOutput() to fail in ICS devices. When the
run_pie wrapper needs to be used, the call to `adb push' ends up
overwriting the original command that was going to be run on the device,
and the instrumentation tests fail like this:

> LD_LIBRARY_PATH=/data/local/tmp/forwarder/ /data/local/tmp/run_pie push /path/in/host/out/Release/run_pie /data/local/tmp/run_pie; echo %$?
[...]
[PIE Loader] dlopen() failed: Cannot load library: load_library[1095]: Library 'push' not found.

Fix it by using a different variable name for the `adb push' command
that must be run on the host.

BUG=373219
R=primiano@chromium.org, torne@chromium.org

Review URL: https://codereview.chromium.org/926023002

Cr-Commit-Position: refs/heads/master@{#316233}