cc: Fix bug in generating picture layer pairs from 2 lists.
When splitting the picture pile list on LayerTreeHostImpl up into a
list on each LayerTreeImpl, a bug was introduced in the pair gathering
code.

If a layer is not in the RSLL on the active tree, we need to still
generate a pair for the pending twin if it is in the RSLL, so that we
will raster the layer and block activation on it.

R=enne, vmpstr
BUG=458411
NOTRY=true

Review URL: https://codereview.chromium.org/926023003

Cr-Commit-Position: refs/heads/master@{#316327}