Data Reduction Proxy Blocking page and resources
Adds the blocking page and resources for the Data Reduction Proxy
bypass interstitials. This intersitial is shown when the user
loads a resource that returns a data reduction proxy bypass on http.
The user can choose to continue loading the page or go back to the
previous page. If the user accepts loading the page, the intertitial
will not be shown again for another 5 minutes.

BUG=428408

Review URL: https://codereview.chromium.org/830503004

Cr-Commit-Position: refs/heads/master@{#316621}