Add a function to get Wifi SSID to net_util.
This CL adds |GetWifiSSID| to net_util to get the SSID of the associated
WiFi access point's, if there is one. The function is currently only
implemented for Linux (including ChromeOS and Android). On Linux, it retrieves
WiFi SSIDs for each interface and if all interfaces have the same SSID, returns it as
the current SSID. This is similar to what https://codereview.chromium.org/739983005/ does.

BUG=451272

Review URL: https://codereview.chromium.org/878513008

Cr-Commit-Position: refs/heads/master@{#316623}