IndexedDB: Ensure that blobs in unused prefetches are acked
When the browser sends a blob reference to the renderer it holds an
extra reference until the renderer "acks" receipt, to keep the blob
alive. Normally this is done when a "result" containing blobs is
handed off to a blink::IDBRequest.

When we "prefetch" for a cursor, multiple results are sent from the
browser to the renderer at once, but only seen by blink::IDBRequest
when script actually requests them. If script never does (e.g. it
abandons a cursor midway through iteration) then the blobs would never
be acked.

Add blob acking to the cleanup step when prefetch entries are dropped.

BUG=457498
R=cmumford@chromium.org

Review URL: https://codereview.chromium.org/924033003

Cr-Commit-Position: refs/heads/master@{#316688}