Pepper: Use the Mojo IRT interface in Chrome.
This change reenables the Mojo test in Chrome and uses the Mojo IRT interface
inside Mojo calls.

It touches a file in the Mojo repo, which corresponds to the following
change there:
  https://codereview.chromium.org/918243002/
That change is necessary for the mojo irt interface to be queried from C++
code.

BUG=454466

Review URL: https://codereview.chromium.org/923593002

Cr-Commit-Position: refs/heads/master@{#316694}