[tracing] Introduce memory tracing skeleton.
This is the first step towards the introduction of memory as a
metric in chromium tracing. This change introduces the core memory
dump manager which will coordinate the memory dumps.
Specifically, the following concepts are introduced by this CL:
 - memory_dump_manager: the front-end interface exposed to the rest of
   the codebase to create memory dump points.
 - memory_dump_provider: the contract interface that memory dumpers need
   to expose in order to coordinate with the manager.

This CL does NOT expose any user-visible tracing functionality.

BUG=458295

Review URL: https://codereview.chromium.org/933613002

Cr-Commit-Position: refs/heads/master@{#316700}