SVG: remove FIXME: Ensure that SVGSMILElement scheduled in SMILContainer is always valid
Before this CL, SMILTimeContainer allowed SVGSMILElement with invalid attribute type to be in the schedule.
These invalid SVGSMILElement was worked around in SMILTimeContainer::updateAnimations() loop (w/ a FIXME comment).

This CL fixes the issue by only allowing SVGSMILElement with a valid attribute type to be in the schedule.
+ a minor refactoring to dedupe m_container->{,un}schedule() calls into SVGSMILElement::{schedule,unscheduleIfScheduled}.
+ remove checkInvalidCSSAttributeType argument, which is always targetElement().

This suppresses unnecessary creation of SVGAnimatedTypeAnimator and speeds up Oilpan by reducing number of objects in the live objects graph.

TEST=LayoutTests/svg/animations/animate-css-xml-attributeType.html
BUG=None

Review URL: https://codereview.chromium.org/932543002

git-svn-id: svn://svn.chromium.org/blink/trunk@190389 bbb929c8-8fbe-4397-9dbb-9b2b20218538