Re-enable vertical flow in SimplePath
This patch changes the default code path of vertical flow text back to the
simple code path. This was changed to complex path in crbug.com/408992, but
it turned out that it degrades the performance by 100-150%.

The performance work for the complex path is under progress in
crbug.com/446011, but in order to ship better code, this patch makes a
temporary switch before the stable release. The complex path will be
re-enabled once we're back to work-mode.

BUG=445589

Review URL: https://codereview.chromium.org/932033002

git-svn-id: svn://svn.chromium.org/blink/trunk@190462 bbb929c8-8fbe-4397-9dbb-9b2b20218538