[Win] Disable minimize and maximize on translucent app windows that have shape.
If a window has both alpha and shape, and its bounds change,
translucency is lost. This includes minimizing. This CL tries to prevent
this to some extent by disabling minimize and maximize on these windows.

Note: this bug occurs even if the bounds change first, then the shape is
applied, so the logic should be whether the window ever intends to set a
shape, but the API does not express this at the moment.

BUG=459922, 417947

Review URL: https://codereview.chromium.org/943943002

Cr-Commit-Position: refs/heads/master@{#317250}