Make commit for swap promises do not unset can_cancel_commit.
When swap promises are enqueued we request a commit. If there
are no updates the promises should be broken and the commit
skipped (EarlyOut_NoUpdates). To achieve that we request
the commit by calling SetNeedsUpdateLayers instead of
SetNeedsCommit in order to not unset can_cancel_commit. Note
that SetNeedsUpdateLayers is a subset of SetNeedsCommit and a
no-op when a commit has already been requested.

BUG=458577

Review URL: https://codereview.chromium.org/925353002

Cr-Commit-Position: refs/heads/master@{#317292}