Add UMA histogram for whether page loads are mobile optimized
This adds a boolean histogram for whether Chrome assumes the web page is
mobile optimized. It is recorded as soon as the Tab gets the callback
for didFinishLoad(). On every call to didNavigateMainFrame(), the bit
the bit it is cleared again. This is only recorded for the main frame.

The viewport might not be correctly identified when didFinishLoad()
gets called, at which point the stat might be a false negative. This
means that this stat could be considered a lower bound for mobile
optimized pages.

BUG=None

Review URL: https://codereview.chromium.org/943823002

Cr-Commit-Position: refs/heads/master@{#317451}