[CSS Grid Layout] Remove old FIXME in RenderGrid::placeItemsOnGrid()
There was a FIXME about not re-resolving positions if the grid is grown during
auto-placement. But as the spec has changed [1] and you can only refer to the
explicit grid (even if you use for example grid-column: 1 / -1). It seems that
we can get rid of this FIXME as it's not a causing a problem anymore.

[1] http://dev.w3.org/csswg/css-grid/#grid-placement-int

Review URL: https://codereview.chromium.org/925593004

git-svn-id: svn://svn.chromium.org/blink/trunk@190608 bbb929c8-8fbe-4397-9dbb-9b2b20218538