Remove all remaining traces of chrome.infobars
And re-enable ExtensionApiTest.WindowOpen, to verify that the new test
is effective. This test was introduced by crbug.com/52298, but disabled
by crbug.com/64899. The test is supposed to check whether a tab that is
opened via window.open() from an extension process associated with a
particular Browser is opened in a non-incognito browser when 'incognito'
is not set to 'split' (but the default, 'spanning').

BUG=458325,64899
R=rdevlin.cronin@chromium.org,sky@chromium.org

Review URL: https://codereview.chromium.org/942873003

Cr-Commit-Position: refs/heads/master@{#317509}