Add notranslate meta tag on dir_header.html
dir_header.html is used to show a directory list on navigating
e.g., file:///. The listing page is already localized, but if
some files in the list is in a foreign language, translate
infobar may appear to suggest translation.

Translate infobar should not be shown on localized pages.

BUG=460732

Review URL: https://codereview.chromium.org/943193002

Cr-Commit-Position: refs/heads/master@{#317519}