Updating gypi after moving FocusType.h.
Source/core/page/FocusType was moved to
public/WebFocusType.h in https://codereview.chromium.org/852083002
This updates the gypi file to reflect that change and to avoid
a warning in the msvs-ninja gyp generator.

Note: there is no gypi file that tracks the files in public/platform.

R=fsamuel@chromium.org,jochen@chromium.org

BUG=448716,147644,459058

Review URL: https://codereview.chromium.org/950833002

git-svn-id: svn://svn.chromium.org/blink/trunk@190679 bbb929c8-8fbe-4397-9dbb-9b2b20218538