Use the version of clang shipped with Xcode for official build
Instead of checking which generator is used to decide whether to use
ToT clang or Xcode clang, make the decision based on the buildtype
instead.

Use the version of clang shipped with Xcode for official build, and
leave the default selected for developer builds (default to ToT clang).

BUG=459148

Review URL: https://codereview.chromium.org/936013002

Cr-Commit-Position: refs/heads/master@{#317627}