[Win] Revert the profile switcher code from the jumplist.
This was a failed experiment. The jumplist is way too fragile and finicky
to be a good, long term solution for a fast-profile-switcher (mostly because
it can only display up to 10 profiles).

This code was initially introduced in https://codereview.chromium.org/660813002/
(with a followup CL to fix it), but that CL also introduced some useful
jumplist related metrics, so I'll just manually revert the jumplist bits.

BUG=431392

Review URL: https://codereview.chromium.org/931853003

Cr-Commit-Position: refs/heads/master@{#317649}