Sync: Small refactoring of Directory::CheckTreeInvariants.
The purpose of this change is to minimize chances of hitting
crbug/456029.

1) Call ModelTypeToRootTag only when the root folder could
potentially be created on the client i.e. model type is
defined in client EntitySpecifics and it is one of supported
model types.
2) Rename is_type_root_folder to is_client_creatable_type_root_folder
to make it clear that the checks are specific to client
created type root folders.
3) Narrow the scope of where is_client_creatable_type_root_folder
has to be calculated.

BUG=456029

Review URL: https://codereview.chromium.org/948113005

Cr-Commit-Position: refs/heads/master@{#317742}