Make dispatchIfAlive ensure the life of its queue.
This is a speculative attempt to fix Issue 456636.

CustomElementSyncMicrotaskRunQueue::dispatchIfAlive checks that the
object is alive when it starts running callbacks, but doesn't ensure
the object lives for the duration of the callbacks. Since the
callbacks can run author JavaScript, the assumption that whatever kept
the object alive will continue to do so might be too optimistic.

Additionally, strengthen the assertion that queue processing is
non-reentrant.

BUG=456636

Review URL: https://codereview.chromium.org/916403003

git-svn-id: svn://svn.chromium.org/blink/trunk@190720 bbb929c8-8fbe-4397-9dbb-9b2b20218538