Update the UI when spinner is shown.
Since previously the UI was not updated even after the spinner is shown, items in the previous directory were visible with a spinner. This patch changes it to update the UI in order to remove items in the previous directory when spinner becomes visible.

BUG=460405
TEST=manually tested.

Review URL: https://codereview.chromium.org/950273002

Cr-Commit-Position: refs/heads/master@{#317773}