Plugin Power Saver: Fix PluginPreroller outliving the PluginInstanceThrottler
This fixes a use-after-destroy bug leading to crashes.

Object Lifetimes:
blink::WebPlugin >= PluginInstanceThrottler >= PluginPreroller.

The PluginPreroller is supposed to observe the Throttler and destroy itself when the Throttler gets destroyed.

Previously, the WebPlugin would sometimes destroy the Throttler before the PluginPreroller was ever created. The PluginPreroller would then try to observe a non-existent throttler, and crash.

This patch now attaches the PluginPreroller to the Throttler before the blink::WebPlugin has a chance to destroy the Throttler.

BUG=459920, 458326, 403800

Review URL: https://codereview.chromium.org/941803003

Cr-Commit-Position: refs/heads/master@{#317837}