Implement NativeViewAccessibilityWin using AXPlatformNodeWin.
AXPlatformNodeWin is an implementation of Windows accessibility APIs
that doesn't know anything about content/ or ui/views/ - it just
exposes an abstract tree of AXNodeData nodes and uses a delegate
interface for everything else.

This patch completely removes the Windows-specific views accessibility
code and implements it using AXPlatformNodeWin instead. A future
step will be to do the same with BrowserAccessibilityWin in content/.

This patch also includes a new unit test for AXPlatformNodeWin,
since it's now possible to test it without any views or web
content code. Hooray!

BUG=457564

Review URL: https://codereview.chromium.org/909143003

Cr-Commit-Position: refs/heads/master@{#317878}