Use complex path for vertical layout
This patch reverts the temporary switch of vertical layout to the simple code
path, which was done so that our work-in-progress performance work
should not affect the branch cut. Since the branch cut was done, we switch
to the complex path to work on its performance.

We acknowledge that this patch will degrade "vertical-japanese-kokoro-insert"
perf test by 400% or so. The plan is to switch to what we want to ship, and
make it faster by the next branch cut. If it does not reach close to the ref
by the next branch cut, we can discuss.

BUG=445589

Review URL: https://codereview.chromium.org/953733002

git-svn-id: svn://svn.chromium.org/blink/trunk@190789 bbb929c8-8fbe-4397-9dbb-9b2b20218538