Implement SVG masking with slimming paint
This patch moves SVG masking over to slimming paint. The Compositing
recorder has been heavily refactored to add a parameter for color
filters and a bounding rect, as well as merging the blend mode and
compositing ops.

With this patch we pass all tests in the newly-added virt/svg/masking
directory, as well as 3 additional tests.

This patch depends on: https://codereview.chromium.org/942403002

Review URL: https://codereview.chromium.org/947893002

git-svn-id: svn://svn.chromium.org/blink/trunk@190793 bbb929c8-8fbe-4397-9dbb-9b2b20218538