Update adb_remote_setup.sh to use git repository.
Currently the adb_remote_setup.sh uses the SVN repository to
update itself. Given that chromium has moved hosting to use
git, it should use that repository as its canonical source.

Gitiles currently does not have support for directly downloading
the raw contents of a file, but it is possible to download it as
base64 encoded data. This is a security precaution to ensure that
browsers do not accidentally act upon the downloaded data.
More info can be found in this issue:
https://code.google.com/p/gitiles/issues/detail?id=7

This CL updates the canonical script URL, and changes the
update-self function to base64 decode the downloaded data before
storing it in the new script file.

BUG=None

Review URL: https://codereview.chromium.org/926973002

Cr-Commit-Position: refs/heads/master@{#317998}