[Extensions] Add an ExtensionFunction::GetSenderWebContents() function
ExtensionFunction::GetAssociatedWebContents() is defined to return the "current"
web contents, but doesn't specify what that is. It's also defined in a bunch
of different ways. Make one GetSenderWebContents() that returns the web contents
of the associated render view host, and update a couple of APIs to use it.

BUG=461394

Review URL: https://codereview.chromium.org/950403002

Cr-Commit-Position: refs/heads/master@{#318079}