Fix official build in GN
Changes process_version template. In GYP this is used in two ways and the GN version only supported one of these. This change also deletes an unnecessary fork of the template file, and fixes the dependencies for the extra files (previously they were just passed with "-f" and not added to the inputs of the target.

Adds a cdm_adapter template which basically matches the GYP version. Uses this for the media clearkey adapter and the widevine one in the official build.

Fixes for the crypto targets when compiling with the official build's checked-in sysroots.

Review URL: https://codereview.chromium.org/949233003

Cr-Commit-Position: refs/heads/master@{#318083}