No need to suppress ZoomBubble if !window_->IsActive().
In Browser::OnZoomChanged() we currently filter on both (1) the event
having been generated by the active webcontents, and (2) the browser
window being active.

However, the window can receive <ctrl>-<scrollwheel> events even when
it is not active, and these will result in zoom changing without a zoom
bubble being shown.

This CL removes the check for window_->IsActive() so that the zoom
bubble always shows if the zoom changes for the active web contents by
a means other than the wrench menu.

TEST= Open chrome, make window inactive, hover over window and zoom
      with the mouse scroll-wheel; the zoom bubble should appear.
BUG=459322

Review URL: https://codereview.chromium.org/940673002

Cr-Commit-Position: refs/heads/master@{#318089}