Let runAfterDisplay ensure the callback is called after a paint
The previous version of runAfterDisplay seems not to ensure a paint
when content_shell runs headlessly. This causes less coverage of
tests in slimming paint mode about display item caching because the
many tested results are of the first paint, not the incremental
paint that the tests are intended to test.

Use testRunner.displayAsyncThen() instead to ensure the paint.

Also added autoNotifyDone mode, so that the test doesn't need to
call testRunner.notifyDone(), and ensures that the test finishes only
after the second paint.

BUG=449628, 450725

Review URL: https://codereview.chromium.org/959783002

git-svn-id: svn://svn.chromium.org/blink/trunk@190875 bbb929c8-8fbe-4397-9dbb-9b2b20218538