Rewrite AndroidSyncSettings to be significantly simpler.
The public methods no longer take an Account as a parameter.
Instead, when the account changes, updateAccount() is called
by ChromeSigninController. This removes the need for all of
the "didUpdate" logic.

Instead, updateCachedSettings() is called only when the
Android settings are changed or the account is changed.

CachedAccountSettings has been removed entirely. A separate
class is not necessary to cache three boolean values (which
it was already only handling two of!)

Review URL: https://codereview.chromium.org/879533004

Cr-Commit-Position: refs/heads/master@{#318196}