scheduler: Convert enums to enum classes
Convert enums to enum classes to make their purpose clearer and to make it
easier to expose them to the unit tests. Note that the task queue identifier is
maintained as a numeric enum because it serves as an index to the task queue
manager and selector. No functional changes.

Review URL: https://codereview.chromium.org/962633002

Cr-Commit-Position: refs/heads/master@{#318248}