ui: Pass correct image texture target and one-copy settings to cc.
The ui browser compositor should use onecopy for implside painting, as
we have enabled that for all platforms that use ui::Compositor.

The image texture target should also match the one used in the renderer
compositor. So have the ui::Compositor query it from the ContextFactory
and move the code that makes the decision out from RenderProcessHostImpl
to a shared location.

R=piman@chromium.org, reveman
BUG=314185

Review URL: https://codereview.chromium.org/963223002

Cr-Commit-Position: refs/heads/master@{#318551}