Fix Virtual Keyboard not being deployed in TouchView mode on Ozone.
Ozone includes the internal keyboard in the actice keyboard devices
even if the TV mode disables it due to the fact that the volume keys
can still be pressed. We now explicitly check the state of maximized
mode when deciding to show the keyboard.

TEST=VirtualKeyboardControllerAutoTest.EnabledDuringMaximizeMode, VirtualKeyboardControllerAutoTest.SuppressedInMaximizedMode
BUG=462666

Review URL: https://codereview.chromium.org/970983003

Cr-Commit-Position: refs/heads/master@{#318801}