Rename renderName() to name().
This CL updates the renderName() method after the work to move from
Render to Layout (see [1] for discussion). As well, the method is moved
in the header files to make it consistently public. The declaration in
LayoutObject.h is public and the rest are randomly public, private
or protected.

[1] https://groups.google.com/a/chromium.org/forum/#!topic/blink-dev/aQlkJ2bFbUU

BUG=462670,462959

Review URL: https://codereview.chromium.org/975433002

git-svn-id: svn://svn.chromium.org/blink/trunk@191197 bbb929c8-8fbe-4397-9dbb-9b2b20218538