Ensure that the on screen keyboard detection on Windows 8+ is more reliable.
The keyboard detection code is heuristic based as we don't have a reliable way to detect if a keyboard
is attached. We use the Setup API's for keyboard class GUID's and look for specific prefixes in the device names like
ACPI\PNP*, ACPI\MSF* and HID\VID*. If we find one of these we assume that a keyboard is present.

Fixed a bug in the keyboard detection function where in the order of params for the StartsWith helper was flipped.

This works on the Surface and Samsung touch screen laptops in my testing.

BUG=335735

Review URL: https://codereview.chromium.org/969293002

Cr-Commit-Position: refs/heads/master@{#318961}