Refactor away the Browser* dependency in exclusive_access
This change is needed to support the use of the exclusive
access bubble from extensions. There are two scenarios
under which the bubble will be used to:
- Display the bubble when keyboard lock is in effect.
- Display the fullscreen/mouse lock UI from Hosted app (not
  hooked up in this CL).

Exclusive access controllers such as fullscreen and mouse
lock are updated to no longer access Browser* object
directly, but instead use new ExclusiveAccessContext and
 ExclusiveAccessBubbleViewsContext interfaces. These are
implemented by BrowserView and will in the future be
implemented by NativeAppWindow.

Review URL: https://codereview.chromium.org/877413004

Cr-Commit-Position: refs/heads/master@{#318965}