Delete all uses of SpeechManager from the app list start page.
Constructing the SpeechManager creates a WebAudio instance which creates
an audio output stream that prevents ChromeOS from suspending. This
code is supposed to be dead and is slated for removal. This change gets
rid of the usage and is intended to be mergable into the M42 branch.

BUG=450081

Review URL: https://codereview.chromium.org/966983003

Cr-Commit-Position: refs/heads/master@{#318969}