Fix TrameTree::scopedChildCount when called for a remote frame.
The new OOPIF perf bot was crashing in
blink::FrameTree::scopedChildCount() when visiting
http://www.barnesandnoble.com/u/books-bestselling-books/379003057/.
(See bug for call stack).  The problem is this line:
    m_scopedChildCount = scopedChildCount(toLocalFrame(m_thisFrame)->document());

A simple repro is to call window.frames[0].length when the first frame
is remote.  This CL avoids the crash by passing in nullptr for
TreeScope when the frame in question is remote.

Note that because there is a null guard for TreeScope in
scopedChildCount(TreeScope*), currently scopedChildCount will return 0
instead of 1 when frames[0].length is called on a remote frame.  This
should be fixed when RemoteFrames get a TreeScope.

BUG=463346

Review URL: https://codereview.chromium.org/976803003

git-svn-id: svn://svn.chromium.org/blink/trunk@191226 bbb929c8-8fbe-4397-9dbb-9b2b20218538