Set whether context menu appears on mouse up from WebPreferences
Whether the context menu appears on mouse up can depend on run-time
settings (in the case of Opera). This reads the setting to
WebPreferences, so that the browser has the possibility of
communicating the appropriate setting to the renderer. If not
explicitly set, it defaults, as before, to 'true' on Windows and
'false' on other platforms.

BUG=

Review URL: https://codereview.chromium.org/819293002

Cr-Commit-Position: refs/heads/master@{#319040}