Draw correct repaint regions instead of a conservative bounding box
Previously, the update_rect() of the LayerImpl was used for the repaint
regions in the DebugRectHistory; now the base LayerImpl class defines a
virtual method giving back an invalidation Region that is overwritten by
PictureLayerImpl to use the already existent invalidations_ Region.
Regions give a more tighter bound on the invalidations than the
single conservative rect that was used before on a per-layer basis.

BUG=424682

Review URL: https://codereview.chromium.org/958843004

Cr-Commit-Position: refs/heads/master@{#319094}