Fix double scrolling of fixed child of transformed scrollable element
https://src.chromium.org/viewvc/blink?view=rev&revision=188245 was a
wrong fix to crbug.com/439930. The real issue is about the incorrect
offset of a fixed element under a relative-positioned element which is
a child of a transformed scrollable element.

This CL reverts that CL, and fix Layer::enclosingPositionedAncestor()
for fixed-position object, so that Layer::updateLayerPosition() can
calculate the location of the fixed-position object correctly.

BUG=462827,439930
TEST=fast/repaint/fixed-child-of-transformed-scrolled.html
TEST=fast/repaint/fixed-descendant-of-transformed-scrolled.html

Review URL: https://codereview.chromium.org/973383002

git-svn-id: svn://svn.chromium.org/blink/trunk@191301 bbb929c8-8fbe-4397-9dbb-9b2b20218538