key_mobile_sites_smooth: Reload pages before scrolling for LinkedIn and Wowwiki
We have shader compilation caching but telemetry can't benefit from it since
pages are tested with cache cleared.

We want to reflect the caching impact in our metrics. To enable the cache,
after the initial load of the page, we wait for a while for the page to
build some cache, and then reload the page and test the user interaction.

Linkedin and Wowwiki in key_mobile_sites_smooth are selected for this since
they are found to have expensive initial shader compilation cost.

BUG=430631,453861,451216

Review URL: https://codereview.chromium.org/959063002

Cr-Commit-Position: refs/heads/master@{#319102}