Recognize GC_PLUGIN_IGNORE() on 'operator new' declarations.
For the select few times an overriding "operator new" is needed for a
GC class, provide the GC_PLUGIN_IGNORE() escape hatch for these. The
responsibility that the "new"'s allocation will be on the Oilpan heap,
rests with the implementor of the class.

R=haraken
BUG=443854

Review URL: https://codereview.chromium.org/975393002

Cr-Commit-Position: refs/heads/master@{#319141}