Do not squash non-self-painting layers.
We erroneously try to squash layers that are
not self-painting. This causes a mismatch between
where invalidations are sent, and where repaints
need to occur. The end result of this is that
repaints tend to not happen when they should.
This CL ensures that we check for self-paintingness
before we try to squash, which makes squashing
behave more like regular compositing.

BUG=441542

Review URL: https://codereview.chromium.org/929353005

git-svn-id: svn://svn.chromium.org/blink/trunk@191382 bbb929c8-8fbe-4397-9dbb-9b2b20218538