Add support for escaped target names in isolate driver.
Currently the isolate_driver.py which creates the dependency files
used by the isolate system, does a simple split on all spaces when
trying to identify targets.

This can fail if the target name contains a space in the name. In
ninja, spaces are escaped with a $-prefix. An example would be
'Content$ Shell$ Helper.app'.

This CL adds support for such target names and ensures that they
stay as one item.

BUG=462248

Review URL: https://codereview.chromium.org/970203003

Cr-Commit-Position: refs/heads/master@{#319356}