Prevent a possible crash on omnibox navigation.
If the user enters a word that triggers the alternate nav infobar (i.e. is a
search by default but also happens to be an intranet hostname), and the omnibox
navigation synchronously triggers the extension system to navigate a background
page, the background page navigation could be mistaken for the omnibox
navigation, leading to a crash when trying to show the infobar.

This could also goof up shortcut backend statistics, in theory, if the
background page navigation succeeded but the omnibox navigation did not.

This CL also adds a different CHECK to verify that the navigations that we do
pay attention to are the ones we want.  I'll convert this to a DCHECK in a
couple of weeks if it doesn't turn up anything in the field.

BUG=363105
TEST=none

Review URL: https://codereview.chromium.org/865313002

(cherry picked from commit 8215378f588fa67251c277e326437188efda1dcc)

Cr-Original-Commit-Position: refs/heads/master@{#312950}
Cr-Commit-Position: refs/branch-heads/2272@{#415}
Cr-Branched-From: 827a380cfdb31aa54c8d56e63ce2c3fd8c3ba4d4-refs/heads/master@{#310958}