Replace uses of callOnMainThread with mainThread()->postTask()
This patch replaces the uses of callOnMainThread with
mainThread()->postTask() and moves the function an internal WTF
namespace so only the few places inside WTF that need it can call it.
The advantage is that all call sites are now annotated with FROM_HERE
which makes it easier to identify tasks in traces.

BUG=450977

Review URL: https://codereview.chromium.org/964523005

git-svn-id: svn://svn.chromium.org/blink/trunk@191398 bbb929c8-8fbe-4397-9dbb-9b2b20218538