Optimize BidiTextRun.cpp when hasStrongDirectionality is not needed
BreakingContext::handleText() calls textWidth() which in turn
calls directionForRun(), which consumes 5-7% of total layout
time in a perf test (vertical-japanese-kokoro-insert.html).

When callers do not need hasStrongDirectionality, and if the
string is 8 bit or of a single character, we could avoid calling
rather expensive functions such as BidiStatus::BidiStatus() and
BidiResolver::determineParagraphDirectionality().

This patch optimizes when the conditions meet by allowing callers
to declare whether hasStrongDirectionality is needed or not.

BUG=446011

Review URL: https://codereview.chromium.org/979673002

git-svn-id: svn://svn.chromium.org/blink/trunk@191400 bbb929c8-8fbe-4397-9dbb-9b2b20218538