Show console message for font decode error
Currently, if a webfont fails to parse or rejected by OTS, it is
discarded silently. That is not developer-friendly, since even valid
fonts can be rejected if OTS does not support it.

This patch adds console message on webfont decode error. The error
message is fixed for now. A followup patch will provide more detailed
error message.

BUG=97467

Review URL: https://codereview.chromium.org/939023002

git-svn-id: svn://svn.chromium.org/blink/trunk@191403 bbb929c8-8fbe-4397-9dbb-9b2b20218538