Use Length for the stroke-dasharray property in SVGLayoutStyle
This replaces the SVGLengthList used for strokeDashArray in
SVGLayoutStyle with a (reference-counted) Vector of Lengths.
SVGStrokeDasharrayStyleInterpolation is updated to treat a dash array in
a similar fashion, by having a list of LengthStyleInterpolations.
The storage of types in SVGStrokeDasharrayStyleInterpolation is dropped
for the same reason. AnimatableStrokeDasharrayList is updated to use
AnimatableLengths instead of AnimatableSVGLengths.

BUG=461375

Review URL: https://codereview.chromium.org/975733002

git-svn-id: svn://svn.chromium.org/blink/trunk@191432 bbb929c8-8fbe-4397-9dbb-9b2b20218538