[Cronet] Do not call into native adapter after it is destroyed in ChromiumUrlRequest.java
If consumer calls GetHttpStatusCode/GetHttpStatusText after
the request adapter has been destroyed, we will be in
trouble. The new APIs don't have this problem. This CL adds
null checks on the Java side to make sure we won't call into
the native adapter when it is destroyed.

BUG=460161

Review URL: https://codereview.chromium.org/945843003

Cr-Commit-Position: refs/heads/master@{#319522}