Delete unmasked credit cards when clearing data.
This hooks up unmasked server credit cards to the Chrome clear private data feature. When all data is cleared, all cards will be re-masked.

For time ranges, only cards unmasked in that time range will be re-masked. The use-case for re-masking cards unmasked in the past hour is very weak, but we just don't want to re-mask cards that you unmasked a long time ago, and this is how everything else works.

This adds a "phone number" column to the server address table. This will be hooked up in a future patch. Doing it here saves another database version bump.

Review URL: https://codereview.chromium.org/967453005

Cr-Commit-Position: refs/heads/master@{#319531}