Clean up HTMLObjectElement/HTMLEmbedElement layoutObjectIsNeeded
LayoutObjectIsNeeded is only called by shouldCreateRenderer, which is only
called by createLayoutObjectIfNeeded, which is only called by
Element::attach, which checks that the Document is in style recalc, so
not in detached state. This means we can remove the frame() checks as
the FIXME indicates.

Review URL: https://codereview.chromium.org/980273005

git-svn-id: svn://svn.chromium.org/blink/trunk@191463 bbb929c8-8fbe-4397-9dbb-9b2b20218538