CORS errors print error twice on preload.
Error messages on CORS checks in the ResourceFetcher did not
discriminate between preload scans and normal resource requests. Thus,
if there was a CORS failure and a preload for the resource was done, two
error messages would be printed to the console for the same failure.
This fix adds a check to canAccessResource in ResourceFetcher to make
sure that Blink is not in a preload scan when an error message is
printed.

BUG=464876

Review URL: https://codereview.chromium.org/984213002

git-svn-id: svn://svn.chromium.org/blink/trunk@191498 bbb929c8-8fbe-4397-9dbb-9b2b20218538