Sync HTML element interfaces E-F with the spec
Changing the HTMLFieldSetElement and HTMLFormElement.elements type from
HTMLCollection to HTMLFormControlsCollection does not change behavior,
since they already return a HTMLFormControlsCollection in C++ and that
object is wrapped correctly. Only some includes change in the generated
code.

Other than that, there are no changes to the generated code other than
the order.

HTMLFormElement.action is reflected as a URL by us and other tested
browsers, so a spec bug was filed to make the spec agree:
https://www.w3.org/Bugs/Public/show_bug.cgi?id=28169

BUG=460722

Review URL: https://codereview.chromium.org/980703006

git-svn-id: svn://svn.chromium.org/blink/trunk@191507 bbb929c8-8fbe-4397-9dbb-9b2b20218538