Delete PasswordStoreX::Get*LoginsImpl
As pointed out in https://codereview.chromium.org/906973007/diff/80001/chrome/browser/password_manager/password_store_x.cc#newcode157, instead of overriding Get*LoginsImpl in PasswordStoreX, it is enough to rely on the PasswordStoreDefault implementation of those methods to use PasswordStoreX::Fill*Logins.

The only current difference is that in addition to what Fill*Logins do, PasswordStoreX::Get*LoginsImpl also sorted the logins by origin, to match the order of the login database. There are two ways to fix this:
(1) Move the sorting from Get*LoginsImpl to Fill*Logins.
(2) Drop the sorting.

This CL chose (1). This means that now also PasswordSyncableService::ReadFromPasswordStore, which currently gets the logins unsorted, will perform the sorting. This call should be infrequent enough for this not to be noticeable.r

While (2) would also not break the insides of the password manager, it would degrade the UI (try to find a login in an unsorted list of them). For the reference, the CL which added the sorting was https://codereview.chromium.org/6953010.

BUG=324291

Review URL: https://codereview.chromium.org/986983002

Cr-Commit-Position: refs/heads/master@{#319617}