Android: Fix GPU.GPUProcessTerminationStatus UMA stat
ChildProcessConnectionImpl.mAlwaysInForeground is used
for the GPU process and indicates that it does not participate
in hiding like renderers. The 'important' binding gives it the
same priority as the main application, so upon termination we
should just use the app's foreground status to determine whether
the GPU process was in the foreground or not.

BUG=462422

Review URL: https://codereview.chromium.org/987193003

Cr-Commit-Position: refs/heads/master@{#319768}