Gallery.app : Truncates the file after a blob of new image is created.
Previously Gallery.app truncates the file before a blob of new image is created.
This means that when some error happens during the creation of new image,
the file is overwritten with 0 bytes.
To avoid it, this CL creates the blob before truncates the file.

BUG=463837
TEST=out/Release/browser_tests --gtest_filter=GalleryJsTest.GalleryItemTest

Review URL: https://codereview.chromium.org/988273002

Cr-Commit-Position: refs/heads/master@{#319817}