ServiceWorker: Send only a registration handle id instead of its entire object info
OnUpdateFound() needs only a handle id of an updated registration and does not
have to receive an entire registration object info.

In addition, this CL aims to make the rule of reference counting clearer.
ServiceWorkerRegistrationObjectInfo passed into other IPC handlers implicitly
conveys a reference of a registration handle, but this is not applicable for
OnUpdateFound().

BUG=452403
TEST=n/a (mechanical change only)

Review URL: https://codereview.chromium.org/979133004

Cr-Commit-Position: refs/heads/master@{#320005}