surfaces: Remove draw_and_swap_full_viewport_every_frame flag.
The draw_and_swap_full_viewport_every_frame flag causes the compositor
to try and draw even when its frame is incomplete. This was done in
order to support forced resource reclaiming, because when forced
resource reclaiming is on, we need to draw/swap even when there is no
damage (since the surface host has released the resources).

However, LayerTreeHostImpl::CalculateRenderPasses already checks
can_force_reclaim_resources and will not early out for no damage when it
is true. So this draw_and_swap_full_viewport_every_frame is not helping
the resource reclaiming case, and causes is to draw incomplete frames
instead.

R=jbauman
BUG=465861

Review URL: https://codereview.chromium.org/994223002

Cr-Commit-Position: refs/heads/master@{#320021}