Omnibox - Make Decision about Whether to Query Provider per-provider
In other words, inputs that the user is explicitly sending to a
keyword search engine shouldn't have to pass as extensive a list
of tests as those being sent to the default search engine.  There's
much less of a risk of data leakage here.

BUG=460636

Review URL: https://codereview.chromium.org/983493002

Cr-Commit-Position: refs/heads/master@{#320024}