cc: Prevent a layer with a copy request from overriding RequiresHighRes
Previously we were computing if we should skip the draw based on
requires high res for each layer. Then we'd decide to draw anyways if
there is a copy request.

Move this computation (and for animating checkerboards too) out of the
per-layer loop so we have a more obvious and deterministic order.

R=enne, vmpstr
BUG=465861

Review URL: https://codereview.chromium.org/996003002

Cr-Commit-Position: refs/heads/master@{#320029}