Fix ASSERT(m_canvas) failures when contextDisabled()
contextDisabled() is for performance tests to let GraphicsContext to
draw nothing. In either slimming-paint mode or non-slimming-paint mode,
we should check it at beginning of every drawing method of
GraphicsContext to ensure the performance when contextDisabled() is
as expected.

Especially in slimming-paint mode, missing the check causes null
dereference of m_canvas or ASSERT(m_canvas) failure.

- Ensure contextDisabled() is checked in all drawing methods of
  GraphicsContext;
- Use SkNullCanvas when contextDisabled() to ensure drawings directly
  using GraphicsContext::canvas() will not encounter null canvas and
  be also controled by contextDisabled().

Review URL: https://codereview.chromium.org/987383003

git-svn-id: svn://svn.chromium.org/blink/trunk@191673 bbb929c8-8fbe-4397-9dbb-9b2b20218538