Remove an old constructor from WebNotificationData, add a layout test.
Now that the users in Chromium pass the "silent" flag as well, we can
remove the old constructor.

This patch also introduces a test to verify that the "silent" attribute
will be properly reflected in the Notification object received in the
"notificationclick" event, as well as reflection in a newly created
Notification object through the constructor.

This is part of a three-sided patch:
  [1] https://codereview.chromium.org/960163003/
  [2] https://codereview.chromium.org/965573002/
  [3] This patch.

BUG=442134

Review URL: https://codereview.chromium.org/965463003

git-svn-id: svn://svn.chromium.org/blink/trunk@191705 bbb929c8-8fbe-4397-9dbb-9b2b20218538