Only allow non-deprecated IPv6 addresses which don't contain MAC to be used in WebRTC.
Original fix for this bug was not good for mobile since none of the IPv6 in mobile has temporary attributes. Instead, we look for FF:FE and the U/L bit to determine whether this has MAC or not and filter based on that.

BUG=465378

Review URL: https://codereview.chromium.org/1000643003

Cr-Commit-Position: refs/heads/master@{#320195}