Shut down the hotword detector if we ever need to stop it.
With a DSP, there's no way to pause the audio stream without shutting
it down completely. So always shut down the detector instead. With the
microphone, we probably want this anyways because stopping the detector
still runs the microphone input stream and the detector mearly ignores
the input.

BUG=465952

Review URL: https://codereview.chromium.org/997753002

Cr-Commit-Position: refs/heads/master@{#320261}