WebRtcVideoCapturerAdapter: Remove thread checker from wrapped media::VideoFrame
When all deep frame copies are removed, the wrapped media::VideoFrame will be passed all the way to the encoder, and it will be accessed by more than one thread. Therefore, the thread checker in that class has to go, but it's ok because the frame is read-only and the destructor in media::VideoFrame can handle being destroyed on another thread.

BUG=webrtc:1128

Review URL: https://codereview.chromium.org/1003663003

Cr-Commit-Position: refs/heads/master@{#320265}