Introduce the NotificationDatabase class.
This is the first part of the Web Notification database implementation,
which is now only capable of opening and creating empty in-memory and
file-based databases, as well as destroying them.

Together with this first bit of functionality come a few unit tests
to ensure that they work correctly. Until the database reaches sufficient
maturity, it won't be used for anything but unit tests.

Design document:
  http://goo.gl/TciXVp

BUG=447628

Review URL: https://codereview.chromium.org/988163003

Cr-Commit-Position: refs/heads/master@{#320285}