DevTools: [CSSP] rewrite WI.ComputedStylesSidebarPane
The patch re-writes WI.ComputedStylesSidebarPane so that it no longer
depends or inherits from TreeElement/PropertiesSection classes. This
also adds the "trace" icon to the left of every computed property; when
the button is clicked, the WI.StylesSidebarPane's filter gets populated
with the property name.

This patch not only simplifies WI.CSSP implementation, but also opens
a road for a lot of more simplifications:
- WI.StylePropertyTreeElementBase no longer needed and could be merged
into WI.StylePropertyTreeElement
- Method WI.SSP._fetchComputedCascade could return just a
CSSStyleDeclaration (not a dummy cascade like it does ATM). This means
that section cascades will be used for matched rules only, and thus
model editing operations of rule selectors/style properties could be
moved into section cascades.

R=pfeldman

BUG=466435

Review URL: https://codereview.chromium.org/986893002

git-svn-id: svn://svn.chromium.org/blink/trunk@191783 bbb929c8-8fbe-4397-9dbb-9b2b20218538