Put renderer contents in separate Surface on Android
When Surfaces are enabled on Android (still disabled by default) put the renderer contents into its own Surface instead of using a delegated layer on the browser compositor. This causes it to be scheduled and aggregated directly through the Display, which should reduce CPU usage and latency.

BUG=334090,449319

Review URL: https://codereview.chromium.org/725943003

Cr-Commit-Position: refs/heads/master@{#320372}