Mac: Use the same CGLContextObj for a GLSurface's CAOpenGLLayer and the GLContext that it is current to.
This is a step towards executing the command buffer for the browser
compositor's LayerTreeImpl inside the CAOpenGLLayer callback.

This exposes an issue where the GLSurface's CAOpenGLLayer's drawing
function, drawInCGLContext, changes some of the underlying
CGLContextObj's GL state. This is done partially inside the draw function,
and partially outside of the draw function.

To work around this, mark the context as having been dirtied whenever
the CAOpenGLLayer's drawInCGLContext function is called. Check if this
has happened whenever the context is made current, and also in places
where drawInCGLContext may be called while parsing the command buffer.

Because the CAOpenGLLayer may outlive its owning GLSurface and the
virtual GLContext that it was current to, have it hold a base::Closure
that it gets (ultimately) from the real GLContext. This way, it is guaranteed
that the real GLContext will always be informed when the CAOpenGLLayer's
drawInCGLContext method is called (even if the virtual GLContext which
hooked the two of them together has been destroyed).

BUG=423163

Review URL: https://codereview.chromium.org/995963002

Cr-Commit-Position: refs/heads/master@{#320470}