Introduce WTF::ConditionalDestructor policy class
The ConditionalDestructor class defines the destructor of the derived object. This base is used in order to completely avoid creating a destructor for an object that does not need to be destructed.
By doing so, the clang compiler will have correct information about whether or not the  object has a trivial destructor.

Introducing of this class reduces amount of the repeated code.

At the same time the patch is improving WTF::ListHashSet destruction code path: all the dedicated code is now encapsulated within its finalize() method.

Review URL: https://codereview.chromium.org/998583002

git-svn-id: svn://svn.chromium.org/blink/trunk@191832 bbb929c8-8fbe-4397-9dbb-9b2b20218538