Ensure the frame tree node id map is cleared in ResetForNewProcess().
The bug was that FrameTreeNode::ResetForNewProcess only ended up calling
RemoveFrame on its immediate children, not on deeper descendants. This
led to stale ids in the global frame tree node id map.

My proposed fix is to make it more foolproof by taking care of
updating the id map in the FrameTreeNode constructor and
destructor, and also triggering the FrameRemoveListener from the
FrameTreeNode destructor as well.

BUG=451099,463247

Review URL: https://codereview.chromium.org/985293003

Cr-Commit-Position: refs/heads/master@{#320580}