[New Multicolumn] Make positionForPoint() work.
When we decide to look "inside" a column set for text content, convert
the visual point to a flow thread point, and then redirect to the flow
thread. The flow thread is where the actual content is; column sets
are always childless. Also make sure that we don't enter flow threads
directly, since they are not visual.

This improves text selecting when hitting something outside the
columns (in the area around the multicol container, or in gaps between
columns).

Re-enable some old tests that now pass because of this change.

The old (current) multicol implementation deals with this in
LayoutBlock::adjustPointToColumnContents().

BUG=461352
R=dsinclair@chromium.org,jchaffraix@chromium.org

Review URL: https://codereview.chromium.org/978603003

git-svn-id: svn://svn.chromium.org/blink/trunk@191869 bbb929c8-8fbe-4397-9dbb-9b2b20218538