Use hash map for tracked_objects::location
Accessing the Location->Birth std::map takes up about 0.5% of the CPU cycles, according to recent data from ChromeOS-wide profiling. That is a LOT for Chrome, which contains a ton of stuff.

Replacing it with a hash map cuts that down by 1/3 to 1/2. See the bug for more details.

BUG=chromium:463700

Signed-off-by: Simon Que <sque@chromium.org>

Review URL: https://codereview.chromium.org/957573002

Cr-Commit-Position: refs/heads/master@{#320635}