[Temp][Android] Make the code work without SDK extras.
Requiring users to manually download Android SDK Extras is a step
backwards for automation. We do not use them in Crosswalk or in our
content_shell builds, so for now simply make things work without these
files:

 * Make download_sdk_extras.py do nothing instead of timing out, as
   Crosswalk's bots set CHROME_HEADLESS=1 but the files this script
   tries to fetch are only accessible from Google's network.
 * Make the calls to find(1) in java.gypi create the directories before
   entering them. This is what was causing gyp to fail before even if we
   never depended on the targets requiring the SDK Extras ourselves.

Hopefully we can work together with upstream to make the situation
easier to manage.