Define SkLONGLONG so that SkLONGLONG code in Skia can be removed.
With SkLONGLONG some multiplies floor the low bit, without it these will
truncate. Defining this on all platforms means all platforms will have
the same behavior and also the deviant behavior can later be removed.
Currently, Linux is the only platform which does not define this.

This is expected to mean rebaselining ~1200 layout tests on Linux only.
Once these are rebaselined, the Skia code using this define will be
removed and this define will then be removed as well.

BUG=skia:179
TBR=fmailita@chromium.org,reed@chromium.org
NOTRY=true
This is 'just' an expectations change. All modifications reviewed.

Review URL: https://codereview.chromium.org/1007003005

git-svn-id: svn://svn.chromium.org/blink/trunk@191956 bbb929c8-8fbe-4397-9dbb-9b2b20218538