bindings: Use AccessorName{Getter,Setter}Callback in AttributeConfiguration
This is a preparation for using Maybe version of SetAccessor().
V8 is deprecating a version of SetAccessor() which takes
Accessor{Getter,Setter}Callback, and will only provide a version
of SetAccesor() which takes AccessorName{Getter,Setter}Callback.
Use AccessorName{Getter,Setter}Callback and v8::Name in bindings
so that we can replace old API in the follow-up CL.

BUG=462402

Review URL: https://codereview.chromium.org/1002163002

git-svn-id: svn://svn.chromium.org/blink/trunk@191968 bbb929c8-8fbe-4397-9dbb-9b2b20218538