Allow creating background tabs in response to ctrl+tapping events
The problem is to work the foreground tab when tapping of
gesture with Ctrl key.
The click event is simulated on tapping of gesture, and it
should work to open the tab on the background like the left
button of the mouse with Ctrl key.
This patch is to allow opening the background tab on tapping
of gesture.

BUG=464718
R=jochen@chromium.org

Review URL: https://codereview.chromium.org/1010613002

git-svn-id: svn://svn.chromium.org/blink/trunk@192003 bbb929c8-8fbe-4397-9dbb-9b2b20218538