Split content settings pref related logic into a separate class
This CL is a pure refactoring which paves the way to allow each content setting type to be stored in a separate pref. A ContentSettingsPref class is introduced. Currently it contains all the logic for reading/writing all content setting types into a single pref. However in a subsequent change N of these objects will be created, one for each content setting type.

BUG=452388

Review URL: https://codereview.chromium.org/976243002

Cr-Commit-Position: refs/heads/master@{#321085}