Use StringKeyframes to make CSS Animations responsive to style changes
This patch switches from AnimatableValueKeyframe over to StringKeyframe
in CSS Animations. The only place that still uses AnimatableValueKeyframe
is transitions.

When an animated property is absent in the 0% or 100% keyframe, a neutral
keyframe is created which takes the value of the computed style. When
the computed style changes, the spec requires that the neutral keyframe
is updated accordingly to match the new style.

This patch also introduces behaviour changes for interpolating
SVG stroke-width, opacity and shape-image-transform. The new behaviours
of these interpolations in CSS Animations match that of Web Animations.

When there is a non-animation style change, neutral keyframes are updated.
Firefox handles this correctly, but IE11 doesn't.

BUG=361948

Review URL: https://codereview.chromium.org/948583003

git-svn-id: svn://svn.chromium.org/blink/trunk@192071 bbb929c8-8fbe-4397-9dbb-9b2b20218538