DevTools: [SSP] do not wrap computed style in SectionCascade
In the days when computed section inherited from StylePropertiesSection,
it had to provide a StyleSectionModel. For that reason, we had a
_fetchComputedCascade method in SSP, but now this all wrappers are
useless.

This patch makes SSP the only client of StyleSectionModels (as it is
intended to be). This makes it possible to encapsulate editing
operations in the StyleSectionModels in future patches.

R=pfeldman

Review URL: https://codereview.chromium.org/1015703002

git-svn-id: svn://svn.chromium.org/blink/trunk@192083 bbb929c8-8fbe-4397-9dbb-9b2b20218538