Refactor NavigationState to use NavigationParams
This CL separates NavigationState into a public interface with a limited number
of functions exposed outside of content/, and an implementation. The
implementation now owns a set of NavigationParams. For browser-initiated
navigations they are passed by the RenderView, and for renderer-intiated
navigations they have default values. This also allows to reset the
pending_navigation_params of RenderView with their values when the provisional
load fails for browser-initiated navigations.

BUG=459033

Review URL: https://codereview.chromium.org/1001043002

Cr-Commit-Position: refs/heads/master@{#321120}