Specify pnacl_newlib_dir to SCons when running nacl_integration
The PNaCl toolchain will soon be required by SCons to build the IRT. Even
though nacl_integration doesn't use the IRT built by SCons, it still wants to
set up the environment as part of the SCons global initialization, and that
fails if the compiler is not found.

Also remove the extra download, since the toolchain in the native_client directory (which is already synced by gclient) is now used.

R=ncbray@chromium.org
BUG= https://code.google.com/p/nativeclient/issues/detail?id=4088

Review URL: https://codereview.chromium.org/1021563002

Cr-Commit-Position: refs/heads/master@{#321286}