IDL: Add [Custom=CallEpilogue] to eliminate CG special cases
The effect is similar to having [Custom] on a method, but instead of
completely replacing the generated method implementation, the custom
callback is just called after the regular method implementation call.

Use this to reimplement EventTarget.{add,remove}EventListener and
MediaQueryList.{add,remove}Listener's code generation special cases.

Also use this to reduce the custom code for History.{push,replace}State
to such a call epilogue. While doing so, update History.idl so that the
now generated code is behaves the same as the old custom code.

BUG=345519

Review URL: https://codereview.chromium.org/1019453002

git-svn-id: svn://svn.chromium.org/blink/trunk@192146 bbb929c8-8fbe-4397-9dbb-9b2b20218538