[Webapp Refactor] Move key injection related functionality into the plugin layer.
This CL moves key injection logic setRemapKey() and injectKeyCombination() into
the plugin layer.

With this change, we no longer need to pass the default remap keys from the
ApplicationDelegate through the SessionConnector to the DesktopConnectedView
in order to setup a key remapping.

Instead, we can simply call plugin.setRemapKeys('a->b,b->c') on the delegate.

BUG=465878

Review URL: https://codereview.chromium.org/1022473004

Cr-Commit-Position: refs/heads/master@{#321419}