cc: Fix skewport math
Previously, if the viewport ever moved farther than the skewport limits
in a single frame, the skewport math would break.  It would first
intersect with the max skewport (becoming empty) and then unioning with
the viewport (becoming the viewport).  By moving the union first, it
prevents this empty condition, leaving a valid skewport in the right
direction.

This addresses a case where prepainting wasn't happening when it should
have.

BUG=446645

Review URL: https://codereview.chromium.org/1021663002

Cr-Commit-Position: refs/heads/master@{#321428}