Avoid potential invalid cast of AnimatableValues
This change makes CSSAnimatableValueFactory return nullptr instead of a
dummy value when we reach an unexpected state. The original motivation
for returning a dummy value was to avoid crashes in release however this
instead has the potential to result in invalid type casts in the
AnimatedStyleBuilder and cause security issues in Chrome.

Review URL: https://codereview.chromium.org/1023053002

git-svn-id: svn://svn.chromium.org/blink/trunk@192221 bbb929c8-8fbe-4397-9dbb-9b2b20218538