Move absolute path conversion back to the FILE thread.
In commit 3576d4645180fee42ab40475dd182fcd2afdebea the conversion of
the selected directory to an absolute path was moved from
FileSystemChooseEntryFunction::ConfirmDirectoryAccessOnFileThread to
FilesSelected. As the latter runs on the UI thread this violated I/O
restrictions and results in a DCHECK. This was not caught by tests
because in the test environment I/O appears to be allowed on the main
thread.

This change splits the call to IsUnderNonNativeLocalPath from the
conversion to an absolute path.

BUG=457841

Review URL: https://codereview.chromium.org/1023793002

Cr-Commit-Position: refs/heads/master@{#321510}