Chromevox2: make sure the braille cursor is shown for empty text fields.
The empty string in a boolean context led to the cursor not being
shown.  Changed to compare against undefined instead.
This also fixes the value start offset to always be 0. It should only be
non-zero for multiline text fields where only part of the field is included in
the braille output.

BUG=

Review URL: https://codereview.chromium.org/1019283003

Cr-Commit-Position: refs/heads/master@{#321543}