Fix serviceworker/chromium/windowclient-focus.html flakyness.
The test was trying to guess the URL ordering which isn't guaranteed.
It is now only checking whether the URL of the returned client matches
the one of the client that was focused.

It is also solving a bug in the test that was creating a disrepancy
between Mac and other platforms.

BUG=466943

Review URL: https://codereview.chromium.org/1020093003

git-svn-id: svn://svn.chromium.org/blink/trunk@192267 bbb929c8-8fbe-4397-9dbb-9b2b20218538