Fix for console error while selecting 'Keep local data only unitl you quit..' in cookies in content settings.
we was using 'cookies_session_only' as filter and while getting string value
for cookies we are passing cookiesSession from setContentFilterSettingsValue()
as we are fetching value and making its first character uppercase and concatenating with group
and passing this to get string via getString().
we are getting string value as undefined that causing console error.
Changes done so that proper value for cookiesSession will be fetched.

BUG=466976

Review URL: https://codereview.chromium.org/1005303002

Cr-Commit-Position: refs/heads/master@{#321750}