[Fetch] The body stream should be closed when the draining buffer is closed.
In the current implementation the body stream is closed when we create a
draining buffer which is used when text() is called, for example. That leads
to a confusing behavior that text() is running but the body stream gets closed.
This CL fixes the behavior: the body stream gets closed when the draining buffer
gets closed (e.g. text() is done) and gets errored when the draining buffer gets
errored.

BUG=435393
R=horo@chromium.org

Review URL: https://codereview.chromium.org/1028103002

git-svn-id: svn://svn.chromium.org/blink/trunk@192361 bbb929c8-8fbe-4397-9dbb-9b2b20218538