IDL: Add support for serializer definitions
A serializer definition is output as a 'Serializer' production, with
either no children, an 'Operation' child, or a 'Map'/'List' child. The
production has an attribute named 'ATTRIBUTE' in the

  serializer = identifier;

form (and no children), and the 'Map'/'List' productions have attributes
named 'GETTER', 'INHERIT' and 'ATTRIBUTES', where the latter is a list of
the attributes named in the definition.

BUG=469650

Review URL: https://codereview.chromium.org/1031593003

Cr-Commit-Position: refs/heads/master@{#321780}