Remove prerender cookie store, part 4.
This removes the ability to pass in a custom CookieStore to a URLRequest. It's
a small change to net/url_request which then balloons to everything in the
project that ever makes a URLRequest.

This reverts the rest of https://codereview.chromium.org/188693003 (and then
does a whole lot more because URLRequest constructors were unified recently to
always require passing in the fourth argument).

BUG=457344

Review URL: https://codereview.chromium.org/1003953008

Cr-Commit-Position: refs/heads/master@{#321820}