Only call OnActiveWindowChanged() for NotifyNormal FocusIn/FocusOut events.
This fixes flickering when switching the X11 input focus between non-chrome
windows, but having the pointer be above the chrome window (you need to use
keyboard shortcuts to switch focus to trigger this).

The reason for the flickering is that OnActiveWindowChanged() is called with
state ACTIVE, even though the window was not activated. Before my change, the
code called OnActiveWindowChanged() even for FocusIn/FocusOut events with
mode==NotifyGrab, which is not intended. For details on the events, see:
http://tronche.com/gui/x/xlib/events/input-focus/normal-and-grabbed.html

Review URL: https://codereview.chromium.org/1027883003

Cr-Commit-Position: refs/heads/master@{#321821}