For this specific layout test to perform and observe the test result, it needs the  long render duration to capture the disconnection of nodes which happens after starting the rendering.
This render length for successful test varies on the performance of machine, so the test might be flaky on different platforms. This CL makes the render length longer (20 sec) but the rendering happens non-realtime fashion. So it will not seriously lags the speed of layout test.

This is a hack and the only possible solution for this problem and we need a spec change to address this issue.

BUG=469767

Review URL: https://codereview.chromium.org/1029153002

git-svn-id: svn://svn.chromium.org/blink/trunk@192388 bbb929c8-8fbe-4397-9dbb-9b2b20218538