Fix line width calculation for leading spaces.
Currently, line width and leading space character width are each computed and
converted as floats and converted to LayoutUnit before subtracting the second
from the first.  This can lead to rounding errors.

Instead, do the arithmetic on the floats before converting to LayoutUnit, and
use ceilToFloat() to match the behavior of
LayoutBlockFlow::computeInlinePreferredLogicalWidths.

R=eae@chromium.org,hartmanng@chromium.org
BUG=321237

Review URL: https://codereview.chromium.org/1029203002

git-svn-id: svn://svn.chromium.org/blink/trunk@192404 bbb929c8-8fbe-4397-9dbb-9b2b20218538