Pass in a class name to WTF_MAKE_FAST_ALLOCATED
In preparation for getting more profiling information from PartitionAlloc (e.g., how many CSSValues are allocated in PartitionAlloc), I want to make fastMalloc understand a class name of the object that the fastMalloc is allocating. This CL passes in a class name to WTF_MAKE_FAST_ALLOCATED and introduces classNameOfThisObject() in FastAllocBase.h. (The classNameOfThisObject() is not yet used in this CL.)

BUG=469597

Review URL: https://codereview.chromium.org/1022383002

git-svn-id: svn://svn.chromium.org/blink/trunk@192405 bbb929c8-8fbe-4397-9dbb-9b2b20218538