Reset scroll status when a fling is ignored
In WebView, flings that target the root scrolling layer are ignored by
the compositor. In this case, we failed to clean up the scroll state in
InputHandlerProxy, leading it to believe that a scroll was still active
after the fling had been ignored. Handle this case, resetting the scroll
activity bit when a fling is ignored.

BUG=469369

Review URL: https://codereview.chromium.org/1028533004

Cr-Commit-Position: refs/heads/master@{#321932}