timer: PlatformThreadData::ThreadTimers should own SharedTimer for the thread.
In the main-thread, the SharedTimer object does not have any owner, and in a
worker-thread, the SharedTimer object is owned by the WorkerThread. But for
consistency between main and worker-threads, and to move functionalities that
are related to the platform-thread from WorkerThread (so that multiple workers
are able to share the same thread and use the same shared timer), make
PlatformThreadData::ThreadTimers the owner of the SharedTimer object for the
thread.

BUG=none

Review URL: https://codereview.chromium.org/1018073002

git-svn-id: svn://svn.chromium.org/blink/trunk@192413 bbb929c8-8fbe-4397-9dbb-9b2b20218538