[Win] Cleanup: The BrowserFrameView doesn't need to listen to profile name changes.
In the case on the new avatar button, the UpdateAvatar...() methods just updates
the visibility of the button, not its contents, and the visibility doesn't depend
on the name.

I've also tried to name the methods a little more consistently. There's
a giant refactor that needs to be done to rename avatar_button to avatar_icon and
new_avatar_button to avatar_button, but this isn't that.

BUG=466692

Review URL: https://codereview.chromium.org/1004803002

Cr-Commit-Position: refs/heads/master@{#322057}