ChannelMojo: Let MessagePipeReader wait "peer closed" signal as well.
This CL adds MOJO_SIGNAL_PEER_CLOSED for the MessagePipeReader's
waiting bits. Without this, PRECONDITION_FAILED error can be notified
at the end of the channel lifetime, and the browser process misinterprets
it as a renderer crash, which results sadface screen.

This rarely happens and is reproduced only with some underpowered CPU
with excessive pressure. So we cannot have any reliable test.

BUG=466814
R=viettrungluu@chromium.org

Review URL: https://codereview.chromium.org/1023043002

Cr-Commit-Position: refs/heads/master@{#322094}