Remove Mac native font type members from FontPlatformData
FontPlatformData keeps several Mac specific font members 
around, NSFont, ctFont, cgFont, while all other platforms
rely on just using an SkTypeface. This leads to a lot of 
ifdefs in the code, confusion about which member is currently
instantiated and a history of null-dereference crashes on Mac
due to hard to track failures of font instantiation.

After we got access to a CTFont object from an
SkTypeface for compatibility with HarfBuzz AAT shaping and
clipboard text export [1], I suggest to remove the native 
handles and focus on using SkTypefaces. This allows us as 
well to remove a lot of mostly redundant platform-specific
code on Mac.

[1] https://codereview.chromium.org/872963003/

BUG=414234

Review URL: https://codereview.chromium.org/879533003

git-svn-id: svn://svn.chromium.org/blink/trunk@192506 bbb929c8-8fbe-4397-9dbb-9b2b20218538