Support federated credentials in Chrome settings.
chrome://settings/passwords currently displays a blank input field for
federated credentials. This patch adjusts things to support displaying
the hostname of the federation with which the credential is bound. That
is, it will display "via facebook.com" if you've signed into a website
using Facebook, and saved that credential via the Credential Manager
API.

BUG=461800
R=vasilii@chromium.org

Review URL: https://codereview.chromium.org/960493004

Cr-Commit-Position: refs/heads/master@{#322146}