Create an interface for KeySystems, migrate WebEncryptedMediaClientImpl.
This does not move all of the previously static methods into the interface yet, just those used by WebEncryptedMediaClientImpl. The rest can be migrated as necessary.

This CL also includes a fixes for 457384 and 374751 (containers and codecs must match media type), because it was convenient along with the new interface for checking codec support.

BUG=468912,457384,374751

Review URL: https://codereview.chromium.org/1023863002

Cr-Commit-Position: refs/heads/master@{#322248}