Stop using the MEDIASTREAM content setting.
While the exceptions of the deprecated MEDIASTREAM are no longer used (and were migrated to MEDIASTREAM_MIC and MEDIASTREAM_CAMERA exceptions), the default value is still used.

This CL:
1. Replaces MEDIASTREAM default checks with MIC or CAMERA checks in contexts where only one of the two permissions is required.
2. Replaces MEDIASTREAM default checks with both MIC and CAMERA in contexts where both of the permissions are required.
3. Replaces any write of MEDIASTREAM default with write of both MIC and CAMERA.

This CL does not get rid of the MEDIASTREAM content setting entirely. This is because:
1. There is a mapping between content settings and sections in chrome://settings/content. The media part is mapped to MEDIASTREAM.
2. There is a mapping between content settings and icons shown in the omnibox when the content settings are queried. The icon shown for both camera and microphone is mapped to MEDIASTREAM.

However, after this CL is applied, the MEDIASTREAM content setting will only be used in the above circumstances, never to set or get a value.

To avoid complicated syncing between MEDIASTREAM value on old versions and MEDIASTREAM_MIC, MEDIASTREAM_CAMERA on new versions of Chrome, this CL will not land until crbug.com/452388 is resolved.

TESTED=UI,policy
BUG=464382

Review URL: https://codereview.chromium.org/965103004

Cr-Commit-Position: refs/heads/master@{#322364}