Only schedule a warmup when a user has the app launcher enabled.
This avoids a possibly janky delayed UI task for users that don't use
the app launcher.

After this, a user enabling the app list for the first time won't have a
warmed-up UI but that shouldn't matter because the education page
doesn't try to show it, and ShowForAppInstall() will warm it up.

BUG=431326, 292969

Review URL: https://codereview.chromium.org/1031223002

Cr-Commit-Position: refs/heads/master@{#322451}