Fixes a reuse WebUI bug because of the incorrect shadowing of a boolean value.
This only affects PlzNavigate. When a same-site navigation should reuse the current WebUI, in RFHM::GetFrameHostForNavigation the (shadowing) method-local |should_reuse_web_ui_| would be set to true and the speculative WebUI would (correctly) not be created. But later on as the instance-level |should_reuse_web_ui_| defaults to false, RFHM::CommitPending would try to get the |speculative_web_ui_| which would not be set and *could* cause a crash. As of now the crash doesn't happen because CommitPending is not called when the current RFH is kept for the navigation.

BUG=376094

Review URL: https://codereview.chromium.org/1012863004

Cr-Commit-Position: refs/heads/master@{#322485}