OOP PDF: Reset FittingType when manually setting the zoom.
Currently, when the zoom level is set to a particular value, the
viewport's fitting type is unchanged. This can cause the zoom to be
manually set to one value, but be overridden if the viewport changes
size as it will automatically re-apply the zoom for the fitting type.
This CL fixes this by setting the fitting type to none when the zoom is
set manually.

BUG=470427

Review URL: https://codereview.chromium.org/1038573003

Cr-Commit-Position: refs/heads/master@{#322502}