PlzNavigate: Send EnableViewSourceMode at commit time when reusing a RFH
This CL ensures that we properly send an EnableViewSourceMode IPC to the
RnederView of a reused RenderFrameHost at commit time when the request is for a
view source mode navigation. This is needed because on error a RenderView could
exit the view source mode. This fixes the
RenderFrameHostManagerTest.AlwaysSendEnableViewSourceMode test with PlzNavigate
enabled.

BUG=439423

Review URL: https://codereview.chromium.org/1035983002

Cr-Commit-Position: refs/heads/master@{#322561}