Blink: Pass 'unobscured' rect to WebPlugins.
This is to enable better throttling heuristics for plugins.

The 'unobscured' rect of a plugin is the visible portion of the plugin. It is the same as the 'clip' rect, except it is not clipped to the root window viewport.

That is to say - this still works for below-the-fold-plugins.

BUG=468142

Review URL: https://codereview.chromium.org/1033603007

git-svn-id: svn://svn.chromium.org/blink/trunk@192708 bbb929c8-8fbe-4397-9dbb-9b2b20218538