CopyFromCompositingSurface expects a size in pixels, not DIPs.
This caused an issue where HiDPI devices would request 212x142pixels no matter what.

This changes the code to get the proper scaled dimension for the device (in case of HiDPI, this would be 424x284 if user is on a 2x device, or more accordingly).

BUG=470875
TEST=ContentBased*, SimpleThumb* unit_tests
TBR=jbauman

Review URL: https://codereview.chromium.org/1028393003

Cr-Commit-Position: refs/heads/master@{#322705}