Lazily build the app list model.
This CL makes the AppListSyncableService build the app list model when
it is requested instead of building it when the extensions system is
ready.

This fixes an issue where the sync system was being brought up by the
app list on Chrome launch, likely causing startup slowdown.

BUG=462429

Review URL: https://codereview.chromium.org/981893002

Cr-Commit-Position: refs/heads/master@{#322731}