Remove Range usage in (Styled)MarkupAccumulator
This patch is basically same as crrev.com/1041463003.

A Range object assumes that two positions are in the same DOM tree and
this doesn't match 'selection for WebComponents' which will allow
position pair whose positions are in different DOM trees. To use
(Styled)MarkupAccumulator in both the DOM tree and the composed tree, we
need to eliminate methods which take Range objects.

Now (Styled)MarkupAccumulator works on a DOM tree. After commiting this
patch, we are going to templatize TextIterator to accept another
traversal strategy (ComposedTreeTraversal).

This patch is result of collaboration work with yosin@chromium.org
for selection of web components. We are going to add some features to
this class for the composed tree.

BUG=275851
TEST=n/a; No behavior changes.

Review URL: https://codereview.chromium.org/1046843002

git-svn-id: svn://svn.chromium.org/blink/trunk@192747 bbb929c8-8fbe-4397-9dbb-9b2b20218538