Remove hack to delay listPicker popup window resize
We were delaying layout because offsetHeight was unreliable while loading the page.
It turned out that this was due to PopupMenuImpl changing FontSelector after the page has been loaded.
This introduces PagePopupController.selectFontFromOwnerDocument so we can control the timing we change the FontSelector from JS.

BUG=None

Review URL: https://codereview.chromium.org/1042173003

git-svn-id: svn://svn.chromium.org/blink/trunk@192792 bbb929c8-8fbe-4397-9dbb-9b2b20218538