Remove the '--enable-file-cookies' flag.
Discussed at https://groups.google.com/a/chromium.org/forum/#!topic/net-dev/47jdSKRnbqA.

We can't drop the functionality entirely, as it's exposed as a public method
on Android's WebView, but we can limit the functionality to Android and remove
current callers. That's the approach this patch takes.

BUG=470482

Review URL: https://codereview.chromium.org/976553002

Cr-Commit-Position: refs/heads/master@{#322991}