[Chromoting] Fix base.inherits so parent methods can be used in parent ctor.
The current code calls the parent constructor before merging in the parent
prototypes. This means that when a child is created, the parent methods
aren't available in the parent ctor the first time that object is created.

Also, the base.inherits tests set up the Parent, Child and GrandChild
classes once before any tests are run. This means that the objects are
global and that the prototype chain for the parent is "fixed" the first time
a child object of that type is created. To fix this, the Class setup is now
done in a setupTestClasses function.

BUG=465878

Review URL: https://codereview.chromium.org/1035363002

Cr-Commit-Position: refs/heads/master@{#323076}