Re-land issue 1038783002: Add switch to defer creation of GPU process.
This relands https://codereview.chromium.org/1038783002 with the tests
only enabled on Linux.  These tests appear to be flaky on Windows and
failed on Mac after committing.

Original issue description:
For context - on Chromecast, we have the following scenario:
* We support running in the background (i.e. not rendering anything)
  with our service running in the browser process to listen for incoming
  cast connections.
* We cannot acquire any GPU resources (e.g. a GL context or surface)
  during this time as another external application needs it, and it's a
  global resource.

We have supported starting up in the foreground and switching to this mode
for a long time.

But now, we need to support starting up in this mode.  In other words, we
need to bring up the browser process and run our cast receiver service,
but we cannot run any of the GPU process's initialisation code until an
incoming cast arrives.

BUG=internal b/19898960

Review URL: https://codereview.chromium.org/1035383003

Cr-Commit-Position: refs/heads/master@{#323109}