bluetooth: Enable device_unittests target on Android GN build.
Initial configuration of GN to enable building the
device_unittests target. USB, Serial, and HID disabled as
they currently do not compile, and are not blocking the
dependent issues of this change (enabling Web Bluetooth).

Target builds with this change, but does not complete running
on devices.

BUG=471536

Review URL: https://codereview.chromium.org/1048983004

Cr-Commit-Position: refs/heads/master@{#323188}