Refactoring: Hide PendingSelection's properties
This patch moves some logic in LayoutView to PendingSelection and hides
PendingSelection's properties. This makes it easier to use the new
composed-tree version of Position on LayoutView and PendingSelection.

This patch is result of collaboration work with yosin@chromium.org
for selection of web components. After this patch, we are going to
fix those classes for the composed tree. We've already prepared that at
crrev.com/963693002.

BUG=275851
TEST=n/a; No behavior changes.

Review URL: https://codereview.chromium.org/1051743002

git-svn-id: svn://svn.chromium.org/blink/trunk@192916 bbb929c8-8fbe-4397-9dbb-9b2b20218538