[tracing] IPC messages and stubs for inter-process memory dumps
This CL introduces the four IPC messages and the message filter stubs
that will be used by the upcoming CLs to coordinate memory dumps
across processes.
The global coordination policy is simple: all the global dumps must
be orchestrated and handled by the browser process' MemoryDumpManager.
Memory dumps initiated by a child process must be routed through the
browser's MDM.
In other words this reads as: all the Chrome processes have a MDM,
but the browser's MDM is a more senior and responsible manager.

This CL introduces a total of four IPC messages:
(request, response) x (process, global), as follows:

Request local (i.e. current process) dumps to children:
 (browser -> child) TracingMsg_ProcessMemoryDumpRequest
 (child -> browser) TracingHostMsg_ProcessMemoryDumpResponse

Initiate global dumps from a child process:
 (child -> browser) TracingHostMsg_GlobalMemoryDumpRequest
 (browser -> child) TracingMsg_GlobalMemoryDumpResponse

If the global dump is initiated by the browser process, a total of
N_children x 2 (req/resp) IPC messages occur to perform the dump.

If the global dump is initiated by a child process, a further couple
of IPC messages (GlobalMemoryDump{Request,Response}) is required in
order to, respectively, tell the browser's MDM to initiate the dump
and get the final result back (all the child processes succeed / the
global dump was aborted because another one was in progress).

More context and design doc are available in the attached BUG.

BUG=462930

Review URL: https://codereview.chromium.org/1042723002

Cr-Commit-Position: refs/heads/master@{#323284}