cc: Trigger BeginImplFrameDeadline immediately when invisible.
When becoming invisible we are going to abort any pending draw / activate. We
should do that right now rather then waking up sometime in the future and doing
it. This increases the speed at which we become idle after going invisible.

A minor side effect is we will do more work if we become invisible and then
visible within a single frame, but this case is extremely unlikely while the
case compared to becoming invisible for a longer period.

Related to http://crrev.com/1050833002

BUG=371747
R=brianderson,jdduke

Review URL: https://codereview.chromium.org/1051123002

Cr-Commit-Position: refs/heads/master@{#323397}