Add NodeTraversal::isDecendantOf and NodeTraversal::index
This patch adds functions which will be used at TextIterator.

This patch is result of collaboration work with yosin@chromium.org
for selection of web components. We need to get TextIterator to work
not only on the DOM tree but also the Composed tree and are going to
fix TextIterator to switch node-traversing strategies. Now
TextIterator uses Node::isDecendant and Node::index directly, and
we need to replace them with traversal functions.

BUG=275851
TEST=n/a; No behavior changes.

Review URL: https://codereview.chromium.org/1052163002

git-svn-id: svn://svn.chromium.org/blink/trunk@193029 bbb929c8-8fbe-4397-9dbb-9b2b20218538