Add setNeedsLayoutAndFullPaintInvalidation to WebWidget.
Plugins need to invalidate layout on their WebView when
content changes due to factors that the view is not otherwise
aware of. Here we add a method to set the various bits necessary
to cause layout to perform work at the next opportunity.

In the current code, this will be used to force paint invalidation
to be issued on the plugin when placeholder content is loaded.

BUG=472348

Review URL: https://codereview.chromium.org/1056893003

git-svn-id: svn://svn.chromium.org/blink/trunk@193056 bbb929c8-8fbe-4397-9dbb-9b2b20218538