HTML version of distiller feedback footer
This change adds a HTML/CSS version of the YES/NO feedback form that
appears at the bottom of a distilled page. This form is either at
the bottom of the page or at the bottom of the distilled article
depending on which is larger.

This change depends on:
https://codereview.chromium.org/1015463004/

BUG=

Review URL: https://codereview.chromium.org/1049043003

Cr-Commit-Position: refs/heads/master@{#323582}