[MacViews] Remove stray/unguarded uses of ash/shell.h in cross-platform files
The shell.h header compiles on Mac, but its implementation isn't linked
in, resulting in errors later. This CL was made by putting an #error in
ash/shell.h which gets common cases (but there are more strays for
follow-ups).

Some of these #includes were simply redundant/unused.

app_info_footer_panel.cc has logic for pinning apps to the Ash shelf,
which needed a guard (it was already guarded, but not with #ifdefs).

BUG=425229

Review URL: https://codereview.chromium.org/1054793002

Cr-Commit-Position: refs/heads/master@{#323626}