Autofill OSX: Draw red outline around incorrect CVC input.
Change CVC text input field from NSTextField -> AutofillTextField,
and mark it as invalid (causing red border to be drawn) when a
retriable error occurs and no expiry date is showing. If an expiry date
is showing then it's not clear whether the date or CVC caused the error,
so don't mark either of them invalid.

BUG=448572

Review URL: https://codereview.chromium.org/1052513002

Cr-Commit-Position: refs/heads/master@{#323686}