Fix an exception caused by toast running on background thread
Chrome uses an AsyncTask to post the download notifications.
For completed downloads, Android DownloadManager will help us post the notification.
However, if DownloadManager failed to enqueue a download, an exception will be thrown to show a toast.
But toast cannot be run on a background thread, we need to use the UI thread to show a toast.
This change fixes the issue by using AsyncTask.onPostExecute to show the toast on.

BUG=469576

Review URL: https://codereview.chromium.org/1052313002

Cr-Commit-Position: refs/heads/master@{#323763}