Templatize some methods of StyledMarkupAccumulator
Now StyledMarkupAccumulator work only on the DOM tree. This patch
templatizes that and enables that to allow other traversal strategies.

This patch is result of collaboration work with yosin@chromium.org
for selection of web components. After this patch, we are going to
another StyledMarkupAccumulator strategy for the composed tree. We've
already prepared the new strategy at crrev.com/1058203002.

BUG=275851
TEST=n/a; No behavior changes.

Review URL: https://codereview.chromium.org/1065563004

git-svn-id: svn://svn.chromium.org/blink/trunk@193180 bbb929c8-8fbe-4397-9dbb-9b2b20218538