Init Mojo embedder for Windows NaCl plugins.
With this change (and another fix for NACL_BUILD_SUBARCH), we can test
Pepper/Mojo/NaCl support on Windows.

This also moves embedder initialization to nacl_helper for clarity:
The Mojo embedder must be initialized in nacl_helper when that process
is used. When the plugin is running in a chrome process, ContentMain()
handles Mojo embedder initialization.

This change doesn't enable the test MojoFailsWithoutFlag on Windows,
since the abort() triggers a dialog that causes the test to time out.

BUG=414804
R=dmichael@chromium.org, ncbray@chromium.org

Review URL: https://codereview.chromium.org/1059233004

Cr-Commit-Position: refs/heads/master@{#323908}