Make sure that min-size can not be negative.
computeMainAxisExtentForChild can return -1 when the width can't be resolved.
This means that adjustChildSizeForMinAndMax would not clamp the size
to a minimum of zero, even when we pass in a negative size from the
flex algorithm. This change makes sure that we always ensure to keep a 0 minimum.

Review URL: https://codereview.chromium.org/1059053002

git-svn-id: svn://svn.chromium.org/blink/trunk@193211 bbb929c8-8fbe-4397-9dbb-9b2b20218538