Fix a variety of omnibox bugs around zerosuggest, the escape key, or both.
* When we changed to OmniboxViewViews, we broke the behavior that the omnibox
could handle the escape key before it was processed as a "stop" accelerator.
Instead, Chrome trunk current processes escape as stop first (and exclusively),
then if/when there is no load to stop passes it through to the omnibox.  Restore
the other order, though not precisely with the old behavior.  Other than when
arrowing through the popup (in which cases escape just resets to the top entry):
** Escape when not editing (including when zerosuggest is open) will always both
reset the omnibox and stop any underlying load.
** Otherwise, escape will reset the omnibox but not stop any load.

* As part of the above fix, we now ensure that escape will always close any open
zerosuggest window regardless of which entry in it might be selected or what the
omnibox caret/selection currently is.

* When unfocusing the omnibox causes a zerosuggest popup to close, we also force
the omnibox to reset.  This ensures that, if zerosuggest caused us to
temporarily avoid displaying updates to the underlying page URL, they will now
be shown.  When combined with the changes above, this should mean that it's
impossible to leave zerosuggest mode but still be stuck seeing an out-of-date
URL.

* When copying the text of any item in the zerosuggest popup, including the
first one, and regardless of whether the visible URL is up-to-date, ensure that
the clipboard actually gets the correct text.  Previously, select-all + copy on
any zerosuggest item would copy whatever the current underlying URL of the page
was.

BUG=158913,423181,466565
TEST=See bugs

Review URL: https://codereview.chromium.org/1060613004

Cr-Commit-Position: refs/heads/master@{#324017}