Test gpu memory buffers drawing directly to the backbuffer.
To test the lifecycle of a memory buffer we used to use
glCopyTextureCHROMIUM. This has some limitations due to the extension
spec (i.e: limited subset of formats), and doesn't test the general
lifecycle of the memory buffer.

This patch changes GpuMemoryBufferTest.Lifecycle to test the lifecycle
of the GpuMemoryBuffer writing directly to the backbuffer.

BUG=

Review URL: https://codereview.chromium.org/1066073002

Cr-Commit-Position: refs/heads/master@{#324123}