Introduce commonAncestor() and countChildren() to NodeTraversal class
This patch introduces |commonAncestor()| and |countChildren()| to
|NodeTraversal| as |ComposedTreeTraversal| for implementing composed tree
version of |Position| and |PositionIterator| classes.

This patch is a preparation of templatize |Position| class:
http://crrev.com/1071473002

This patch is result of collaboration work with hajimehoshi@chromium.org
for selection of web components. After this patch, we are going to
implement another strategies for the composed tree. We've already
prepared that at http://crrev.com/1058203002.

BUG=275851
TEST=n/a; No behavior changes.

Review URL: https://codereview.chromium.org/1068153002

git-svn-id: svn://svn.chromium.org/blink/trunk@193333 bbb929c8-8fbe-4397-9dbb-9b2b20218538