Record add to homescreen events with URL from the owning web contents.
This change fixes a bug in 42 which is triggered if the start URL from
the manifest is not the same as the current page. In this case, the
fact that the shortcut was added is not remembered, causing the banner
to come up in future.

BUG=474631

Review URL: https://codereview.chromium.org/1065153002

Cr-Commit-Position: refs/heads/master@{#324198}