Added tests for mouse event ordering at tap.
On tap we used to fire mousemove, mouseout, mouseover, mousedown,
mouseup & click events but no mouseenter or mouseleave
events. The fix for another bug on mouse event
ordering (crrev.com/1047733002) added the missing
mouse-events-on-tap as a side-effect. The test in this CL
guarantees that mouse-events-on-tap are consistent with
pure-mouse-events in type & ordering.

BUG=457497

Review URL: https://codereview.chromium.org/984573004

git-svn-id: svn://svn.chromium.org/blink/trunk@193387 bbb929c8-8fbe-4397-9dbb-9b2b20218538