[S.P.] Tighten under-invalidation checking for debug build
After we enable slimming paint by default for devs, under-invalidation
checking will be a useful tool for devs to report and investigate
under-invalidation bugs. This change is a preparation for that.

In debug build, when slimmingPaintUnderInvalidationChecking is enabled,
- Quits when assured under-invalidation is discovered;
- Treats the following cases non-fatal:
  * picture dumps diff, but picture debug strings equal;
  * some display item disappears which is valid when we evict cached
    out-of-viewport subtrees; (we still print a warning because this
    may indicate a deleted-object-not-invalidated issue)

Exclude the under-invalidation-checking code in release build because it
is not so useful without client debug strings in release build, and to
reduce binary size.

BUG=450725

Review URL: https://codereview.chromium.org/1073473006

git-svn-id: svn://svn.chromium.org/blink/trunk@193410 bbb929c8-8fbe-4397-9dbb-9b2b20218538