ui: Add --ui-enable-slimming-paint and DisplayItem-backed PaintContext.
With the new flag enabled, ui::Layer will paint by emiting a
cc::DisplayItemList instead of painting into a gfx::Canvas supplied
by the compositor.

This patch adds a constructor to PaintContext that takes a
cc::DisplayItemList instead of a gfx::Canvas, and updates the various
Recorder classes to support emitting cc::DisplayItems if there is no
Canvas present in the PaintContext.

With this the UI paints correctly via the display-item backend (though
without caching to make it speedy).

R=piman@chromium.org, senorblanco@chromium.org, ajuma
BUG=466426

Review URL: https://codereview.chromium.org/1073603003

Cr-Commit-Position: refs/heads/master@{#324489}