Try switching from Lucida Console to Times to avoid startup regression
A previous change https://chromium.googlesource.com/chromium/src/+/04015b275eb5c775a5740722fd0805994dde8d2c
caused a regression on XP only in messageloop_start_time.

The previous change substituted fonts (rather than adding a new one) to try to avoid
this, but it apparently didn't always help. So, instead use Times New Roman in the font
dialog (which was happening before) to attempt to avoid this. This will make the
debug hud a little uglier on Windows, but not the end of the world.

There's unfortunately no perf trybots for this config, so it needs to actually land to
test.

R=jam@chromium.org
BUG=463613
TEST=run chrome with --enable-win32k-renderer-lockdown --show-fps-counter --ui-show-fps-counter

Review URL: https://codereview.chromium.org/1069163003

Cr-Commit-Position: refs/heads/master@{#324491}