Remove ni:/// prefix from Subresource Integrity
The latest version of the spec has eliminated the ni:/// prefix.
Additionally, this adds in a suite a new LayoutTests and fixes many old
ones.

This does not address two important changes that we'll implement in a
separate, later CL:

(1) Multiple strong hashes from a single  integrity attribute should be
stored and if any passes, the content should be valid. Currently, we
assume only one hash per integrity attribute.

(2) MIME type checking should be moved behind an experimental flag.  The
initial version of the spec does not include support for MIME types,
although it does contain a generic "options" postfix at the end of
hashes so we can implement MIME types in the future. Since this is a
very likely outcome, we will need to disable MIME type checking in
release, but it makes sense to keep it as experimental in preparation
for v2.

R=mkwst@chromium.org
BUG=471429

Review URL: https://codereview.chromium.org/1069883003

git-svn-id: svn://svn.chromium.org/blink/trunk@193495 bbb929c8-8fbe-4397-9dbb-9b2b20218538