Update WebDialogBrowserTest for better coverage / clarity, re-enable on Win
To re-enable the test on Windows, Windows enforces a minimum dialog
height at the OS level of ~59 pixels, so account for that.

Part of this CL is ensuring the behavior of modal dialogs on MacViews is
consistent with other platforms. A test is added to ensure closing the
parent window will also properly tear-down the modal dialog. This is
interesting on MacViews since initially it will have views::Widgets
parented off non-views (Browser) NSWindows, and the lifetime of the
dialog needs to tie in properly.

Other changes:

Usually DialogDelegates are owned by the dialog and freed in
OnDialogClosed(). (WebDialogBrowserTest had the DialogDelegate on the
stack). Update TestWebDialogDelegate to support this pattern for a more
authentic test.

The test was passing the test Browser's WebContents as the |source|
argument to WebContentsDelegate::MoveContents, rather than the
WebContents of the WebDialog. That's weird, but WebDialogView doesn't
care, so just pass nullptr. The test only cares about the test Browser's
WebContents for determining the NativeView |parent| argument for
Widget::CreateWindowWithParent().

Close the dialog using WebDialogView->web_contents()->Close() rather
than GetWidget::CloseNow(). This is asynchronous, but a better match for
how the dialog is actually closed. Note Widget::Close() is actually a
no-op since the WidgetDelegate returns false for CanClose(). CloseNow()
would bypass this, but (natively) Mac window-modal dialogs do a similar
thing (i.e. [NSWindow close] is ignored for a modal dialog "sheet"). So
do an "authentic" close via the WebContents.

Note the tests pass on MacViews with some changes in the pipeline, but
web_dialog_view_browsertest.cc needs to move in .gyp before it's run in
browser_tests on Mac.

BUG=52602, 447086

Review URL: https://codereview.chromium.org/1064373002

Cr-Commit-Position: refs/heads/master@{#324549}