Change logging to use syslog; fix CRD in Yosemite
Have the Chromoting host write logs to syslog. This works around the
issue where, in some Yosemite configurations, launchd fails to start the
agent if the user running the agent does not have write access to the
log or the log's directory.

Logging to syslog has additional advantages such as:
- Log rotation and overflow are now managed by the system.
- Messages are read-restricted to the user running the agent.

After this change, logs will no longer be stored in
/var/log/org.chromium.chromoting.log. To obtain logs, use the following
command:

sudo syslog -k Facility org.chromium.chromoting

If 'sudo' is omitted, messages logged in the login screen (while the
agent is running as root) will not be available.

N.B. The changes to ChromotingHostService.pkgproj were made using the 'Packages'
program. To test the changes, I built and installed a .dmg from scratch.

BUG=468121

Review URL: https://codereview.chromium.org/1073883002

Cr-Commit-Position: refs/heads/master@{#324550}