Added (indirect) testing of getClientBaseJid_.
This method was previously mocked, but now it is run as part of the
tests for HostController.start.

MockSignalStrategy had to be
changed because the approach of creating Sinon spies to implement
the methods of the mock class proved to be impractical for the
new tests, which need Sinon mocks rather than just spies.  Sinon
is not able to replace a spy with a mock, create a mock with
behavior for a nonexistent method, or to add behavior to a mock
after it has been created, so a default no-op implementation was
needed to enable mocking.  Creating the test doubles in the
unit test itself has the additional benefit of allowing the
test code to be statically type checked against Sinon's APIs,
so there was no need to turn off type checking in
host_controller_unittest.js, as was necessary in the older tests.

BUG=471928

Review URL: https://codereview.chromium.org/1069863002

Cr-Commit-Position: refs/heads/master@{#324552}