Allow WebView to directly observe focus change of its WebContents.
Currently WebContentsDelegate::WebContentsFocused exists, but not
WebContentsObserver::WebContentsFocused. WebContentsDelegate::WebContentsFocused
exists only to allow views::WebView to call View::RequestFocus. Unfortunately,
views::WebView is only a delegate of the WebContentses that it creates itself
(it only *observes* the WebContentses that it owns via SetWebContents). However,
this makes every delegate responsible for forwarding focus events to the
views::WebView, which is not robust.

This CL resolves a TODO in WebView to allow the views::WebView to *observe* the
focus change of its WebContents and eliminates
WebContentsDelegate::WebContentsFocused.

This fixes an issue in the app list where no delegate was forwarding focus
changes to the custom launcher page's WebView.

BUG=458371

Review URL: https://codereview.chromium.org/959613002

Cr-Commit-Position: refs/heads/master@{#324584}