Mac: Partially enable partial swap
We draw all contents in the IOSurface and remote CALayer paths into a
buffer, which we then draw to the screen. A partial swap here means that
we partially re-drew the buffer.

This isn't a complete partial swap, because, in both of these cases, we
still re-draw the full buffer to the screen.

While I'm in the neighborhood, clean up how we track the size of the
frame, since that is particularly confusing (especially when we have
partial swaps that would, more logically, pass in the damage rect).

BUG=474299

Review URL: https://codereview.chromium.org/1065123002

Cr-Commit-Position: refs/heads/master@{#324658}