[Extensions] Have the enum properties on the extension API consistent
Web style is to have enums as ALL_CAPS, but we have done a horrible job of this
in our own enums, so we can't change that.
But we can ensure that the enum properties exposed on the API are like that,
e.g. make sure it's chrome.runtime.RestartRequired.BECAUSE_I_SAID_SO, instead of
chrome.runtime.RestartRequired.because_i_said_so.

BUG=463184

Review URL: https://codereview.chromium.org/1076563002

Cr-Commit-Position: refs/heads/master@{#324659}