Make sure first-letter layoutObjects will re-create text runs.
Currently when changing the text effecting a first-letter layoutObject
we will dirty the text runs. This is fine for the remaining text
layoutObject as the ::setText call will mark the layoutObject as
needing layout. For the first-letter layoutObject the call to
force a layout, to re-create the text runs, was missing and the
first-letter would disappear until something forced a layout.

This CL adds the needed call into
FirstLetterPseudoElement::updateTextFragments to set the first-letter
layoutObject as needing layout and preferred width recalculation.

BUG=467636

Review URL: https://codereview.chromium.org/1078103003

git-svn-id: svn://svn.chromium.org/blink/trunk@193569 bbb929c8-8fbe-4397-9dbb-9b2b20218538