Fix integer overflow in ComputedStyle::computedLineHeight
When we compute huge values for line-height through
percentage or CSS calc, we'll overflow the integer and later
on ShapeOutsideInfo::computeDeltasForContainingBlockLine
will ASSERT because it expects non-negative line height.
So for the computed line height use an intermediate LayoutUnit
because it internally clamps to intMaxForLayoutUnit and the int
will not overflow. Note that the code path in computedLineHeight
for percentages does the same.

BUG=458461

Review URL: https://codereview.chromium.org/1016863002

git-svn-id: svn://svn.chromium.org/blink/trunk@193574 bbb929c8-8fbe-4397-9dbb-9b2b20218538