Explicitly call deprecated AX name calculation API.
This change explicitly calls deprecated AX name calculation functions
so that we can delete the ones without the deprecated prefix. It also
starts to add WebAXObjectProxy support for the new name calculation API
so that we can start to implement it and write layout tests for it.

BUG=473220

Review URL: https://codereview.chromium.org/1074953003

Cr-Commit-Position: refs/heads/master@{#324751}