leveldb: Always depend on snappy library.
We have this use_snappy variable, but for us it is always set to true,
so it does not make sense to have it anymore, if we always end up using
snappy.

So we just remove this variable now, and make this dependency
unconditionally.

Tested on Linux with the following command lines:

GYP:
$ ./build/gyp_chromium
$ ninja -C out/Debug leveldatabase

GN:
$ gn gen out-gn
$ ninja -C out-gn leveldatabase

BUG=None
TEST=see above
R=cmumford@chromium.org

Review URL: https://codereview.chromium.org/1065383003

Cr-Commit-Position: refs/heads/master@{#324788}