Better handling of broken applet codebase
A broken codebase would result in using an invalid url as base url in
the KURL constructor and that never results in anything useful (could
even crash). Since HTMLAppletElement is deprecated and badly documented
the goal of this patch is mostly to avoid the known badness and not to
make HTMLAppletElement perfect.

R=fs@opera.com

BUG=374110

Review URL: https://codereview.chromium.org/1079953002

git-svn-id: svn://svn.chromium.org/blink/trunk@193633 bbb929c8-8fbe-4397-9dbb-9b2b20218538