Handle removing scroll resizer inside a button.
With LayoutButton an anonymous inner layoutObject is created.
We then proxy all adds and removes to that inner object. If
we received a request to remove from the button itself we
would ASSERT.

This fails in the case where, inside the button, you have a
scrollable area. The scroll resizer will have its parent set
to the button, yet is never inserted into the tree. Then,
when we go to remove, we trigger the ASSERT.

This CL updates the remove code to handle this case.

BUG=467644

Review URL: https://codereview.chromium.org/1076383002

git-svn-id: svn://svn.chromium.org/blink/trunk@193654 bbb929c8-8fbe-4397-9dbb-9b2b20218538