Enable hardware acceleration for Nouveau and Stage3D
Nouveau has improved a LOT in the 3+ years since it was originally
blacklisted.  This patch removes the blacklist if the user is running
Mesa 10.1 or higher.  We can set the version requirement even higher
if you think that's best.  I've been testing with Mesa 10.5.1 and
haven't ran into any crashes or instability.  The same thing was done
for AMD GPUs running Mesa drivers last year:

https://codereview.chromium.org/215493002

This also removes the blacklist for Stage3D on Linux (was blacklisted
across all GPUs.)  Maybe at one time it was broken but it seems to work
just fine now on all the demos I've tried and with massive performance
increases from the hardware acceleration.

BUG=94103, 129848

Review URL: https://codereview.chromium.org/1027403002

Cr-Commit-Position: refs/heads/master@{#324984}