MacViews: Fix BubbleDelegateTest.VisibleWhenAnchorWidgetBoundsChanged
Fix the test by showing the anchor widget.

This change necessary on Mac because there are two windows, each with
their own root layer. Layers are not "flattened" on Mac so window
visibility is not determined by layers.

On Mac, Widgets are always hidden when they have a hidden parent.
Generally, this is true for aura::Windows as well but it is only
strictly true for aura::Windows without layers.

On Aura, once a layer is found on a visible window, the window is
considered visible if its own layer is drawn, and the layer in the
parent window is ignored. This might be a bug, but
VisibleWhenAnchorWidgetBoundsChanged needs to be fixed regardless.

BUG=378134
TEST=views_unittests

Review URL: https://codereview.chromium.org/1083613002

Cr-Commit-Position: refs/heads/master@{#325007}