DefaultProvider: lock only cache writes, not file writes.
In DefaultProvider, we're currently locking the entire operation of changing a default content setting:
1. Changing the setting in the cache
2. Changing the individual preference
3. Possibly changing the dictionary preference

In fact, only the cache access is multi-threaded and should be locked. Locking the preference access can lead to a situation where the same thread re-acquires the non-reentrant lock, such as when a preference is written by a non-incognito provider, what causes a preference read from the incognito provider, on the same (UI) thread.

BUG=476826

Review URL: https://codereview.chromium.org/1081103003

Cr-Commit-Position: refs/heads/master@{#325028}