Fixing UMA logging for GestureNav
After the recent refactoring patch (crrev.com/1072123002), some of the stats
were not being recorded propery. This patch fixes that and moves the logging to
OverscrollNavigationOverlay so that the stats are logged when the gesture nav
overlay slides on top of the screenshot as well as the live page.

BUG=476028

Review URL: https://codereview.chromium.org/1054963004

Cr-Commit-Position: refs/heads/master@{#325049}