[renderer]: Inform the RendererScheduler when the Renderer is hidden.
Add OnRendererVisible and OnRendererHidden to the RendererSheduler to
enable it to be informed if the renderer goes in and out of the
background. When backgrounded we start a long idle period, but set a
delayed task to end the idle period after 10 seconds to ensure we
eventually stop running idle tasks while in the background.

BUG=468414,469210

Review URL: https://codereview.chromium.org/1024123002

Cr-Commit-Position: refs/heads/master@{#325056}