Sets the word wrap behavior to tooltip aura for long tooltips.
r322681 changed the default word wrapping behavior of RenderText,
which causes a regression on tooltip. See the screenshot in
the reported bug.

This CL doesn't change the default behavior back to the original
one but specifies the correct word wrap behavior for tooltip.
This is because only tooltip code uses RenderText object directly
with multiline flag.

BUG=476875
R=sky@chromium.org

Review URL: https://codereview.chromium.org/1083113002

Cr-Commit-Position: refs/heads/master@{#325140}