New SELECT Popup: Support multiple displays for PagePopup
The render process only knows the WebScreenInfo for only one screen so we cannot handle adjusting popup position for multiple monitor setups.

This removes the ability to adjust popup position to fit inside the available screen space and show the popup always attached to the anchor element.

Popup size and whether the popup is displayed above or below the anchor element is still determined by screen info so this is not perfect but it is what the PopupContainer implimentation was doing.

BUG=475956

Review URL: https://codereview.chromium.org/1087743002

git-svn-id: svn://svn.chromium.org/blink/trunk@193744 bbb929c8-8fbe-4397-9dbb-9b2b20218538