[Download Notification] Use NotificationUIManager instead of MessageCenter
Direct use of message_center::MessageCenter may cause some problems (eg. crbug.com/469622) because it is not aware of profie.

This patch changes the code to use NotificationUIManager, which is a higher layer wrapper and is aware of profile, instead of message_center::MessageCenter, which is on lower layer and isn't aware of profile.

BUG=469622
TEST=manually tested

Review URL: https://codereview.chromium.org/1005393003

Cr-Commit-Position: refs/heads/master@{#325404}