Remove support for AllowBackgroundModeFromRenderer group in BackgroundRendererProcesses experiment.
This group was only intended to be ran on early channels to show that we are tracking the right stats and ensure that the latest background mode experiment isn't bringing back the regressions that mode had previously brought (see issue 381820 and issue 398103).

The analysis is complete and support for it in the client is no longer required (already ramped down server-side).

(this is essentially a partial revert of https://codereview.chromium.org/926663002)

BUG=459600

Review URL: https://codereview.chromium.org/1089873002

Cr-Commit-Position: refs/heads/master@{#325445}