Report the number of empty usernames without a nonempty pair.
In the process of investigation of empty username occurences in the password manager, we suspect re-authentication forms to be a frequent cause. This metric will show us in how many cases this is not the cause.

We interpret occurences of an empty and nonempty username pair on the same signon realm as a presence of both a first-time authentication and re-authentication form.

We report the number of password forms with an empty username in the login database, for which there isn't another password form with a nonempty username on the same signon realm.

Out of several possible approaches:
a) SELECT ... WHERE NOT EXISTS
b) SELECT DISTINCT ... FROM ... JOIN ... ON a.signon_realm=b.signon_realm
c) SELECT DISTINCT ... FROM ... WHERE ... AND a.signon_realm=b.signon_realm
d) SELECT, put into std::set<>, SELECT, iterate and search the set

we implemented a), as manual testing on 1000 random samples seemed to be the fastest.

BUG=456728

Review URL: https://codereview.chromium.org/1087233002

Cr-Commit-Position: refs/heads/master@{#325451}