Move device/usb classes from the FILE thread to UI thread.
Code that interacts with device/usb often lives on the UI thread. As
with the recent migration of device/hid (issue 422540) moving ownership
of these classes to the UI thread makes calling code substancially
simplier. Blocking operations are handled internally by posting tasks
to the FILE thread and returning a result to the UI thread
asynchronously.

This change paves the way for replacing libusb with platform-specific
implementations of these classes that may have different thread usage
requirements (as is the case in the device/hid code).

BUG=427985

Review URL: https://codereview.chromium.org/980023002

Cr-Commit-Position: refs/heads/master@{#325491}