Revert "[Merge to M43] Restore favicons to background tabs in lazy session restore."
This reverts commit e2adcde21a00d2a1edc79c9db41b452fa0ab43ec.

The above commit was landed on trunk after a refactoring in M44 and thus
doesn't compile on M43. Since we're not planning to go further with this
experiment on M43, we'll revert on M43 and move on.

TBR=sky
BUG=475530, 478154

Review URL: https://codereview.chromium.org/1072403006

Cr-Commit-Position: refs/branch-heads/2357@{#138}
Cr-Branched-From: 59d4494849b405682265ed5d3f5164573b9a939b-refs/heads/master@{#323860}