Do a basic origin check in WorkerGlobalScope::IsPrivilegedContext
There aren't currently any APIs that are available in Workers that also
require privileged contexts with ancestor checks. So instead of
implementing a full privileged context check for worker contexts (that
wouldn't be used), this CL just does a basic sanity check that the
responsible document's origin (the origin of the worker) is potentially
trustworthy.

BUG=474710

Review URL: https://codereview.chromium.org/1090993004

git-svn-id: svn://svn.chromium.org/blink/trunk@193961 bbb929c8-8fbe-4397-9dbb-9b2b20218538