Make chrome proxy client-type test expect trailing slashes after URLS.
Currently, the client-side tests receive a 301 Moved Permanently
response from the data reduction proxy, which redirects to the same URL
with a trailing slash that gets bypassed. This change causes the
client-type test to load the test URLs with the trailing slashes
directly, in order to short circuit the 301 redirect before the bypass.

BUG=477908

Review URL: https://codereview.chromium.org/1083223004

Cr-Commit-Position: refs/heads/master@{#325695}