Make always-on-top windows free-floating in Ash maximize mode
Currently windows in maximize mode are forced to be
maximized/centered, non-draggable, and non-resizable.
Applying these properties for always-on-top windows
leads to occlusion of content in other windows and thus
gives a poor user experience. This CL changes the behaviour
of MaximizeModeWindowManager to never manage windows
with the always-on-top property. As a result, any such
windows will be free-floating like in regular Ash.

BUG=458141
TEST=MaximizeModeWindowManagerTest.AlwaysOnTopWindows

Review URL: https://codereview.chromium.org/1063613003

Cr-Commit-Position: refs/heads/master@{#325749}