Show device asset id in version string.
Added GetDeviceAssetID() to browser policy connector which returns the asset identifier from the stored device policy, 'None Specified' if it is not set, or empty string if device is not enrolled.

Added asset-id string to version-labels that contains ID: <asset-id> of the device.

Changed OnEnterpriseInfoUpdated function in CoreOobeHandler which takes an additional argument asset_id and updates asset-id in version string.

BUG=464184

Review URL: https://codereview.chromium.org/1069713002

Cr-Commit-Position: refs/heads/master@{#325827}