Wrote a WebRTC test which runs inside an iframe.
This is a version of perkj's test in codereview.chromium.org/1073783003
which also covers a full peer connection call. It does not attempt to
verify the case where we nuke the iframe though since we don't know
what will happen past the gUM call in that call. The WebRTC call in
the iframe will just get torn down so there's not much we can test.

BUG=472617

Review URL: https://codereview.chromium.org/1058543004

Cr-Commit-Position: refs/heads/master@{#325896}