Allow use of metrics protobuf from java.
The protobuf compiler cannot currently generate the java classes needed
for reading and writing metrics messages. The java generator uses the
filename of the .proto to create an outer class which contains the
message as an inner class. Since nearly all of the metrics proto
messages have the same message name as filename protoc fails with this
error:

--javanano_out: chrome_user_metrics_extension.proto:
chrome_user_metrics_extension.proto: Cannot generate Java output because
the file's outer class name, "ChromeUserMetricsExtension", matches the
name of one of the types declared inside it.  Please either rename the
type or use the java_outer_classname option to specify a different outer
class name for the .proto file.

By manually specifying the outer classname the generator succeeds. Due
to some inconsistent naming of the messages (some have Proto at the end
of the message, others do not) we append "Protos" to each of the outer
class names. This makes their purpose clear and gives some consistency
when referring to them.

BUG=478977

Review URL: https://codereview.chromium.org/1085203004

Cr-Commit-Position: refs/heads/master@{#325964}