Fixes possible crash when dragging tabs and new tab inserted
The crash happens because ending the drag queries back to the TabStrip
for tabs before tabs_ is in sync with the model. If the newly inserted
tab is at the beginning of the model and the tab being dragged is at
the end of the model then GetTabsMatchingDraggedContents tries to
use an index that is outside of tabs_ and we crash.

The fix is to add the tab before ending the drag. By adding the tab
first we ensure the model and tabstrip are in sync.

BUG=474082
TEST=covered by test
R=pkotwicz@chromium.org

Review URL: https://codereview.chromium.org/1086893003

Cr-Commit-Position: refs/heads/master@{#326027}