Cronet: Fix IOBufferWithByteBuffer.
It was inheriting from IOBuffer, which automatically deletes its data_
pointer on destruction.  Switched to WrappedIOBuffer, which does not.

It's weird to a parent class to have a behavior its child class does
not.  Ideally, we'd fix IOBuffer, but given the number of consumers,
seems more effort than it's worth.

BUG=479206

Review URL: https://codereview.chromium.org/1099923002

Cr-Commit-Position: refs/heads/master@{#326080}