Ignore ViewHostMsg_UpdateFaviconURL messages for RenderViewHosts in the STATE_PENDING_SWAP_OUT state
ContentFaviconDriver assumes that it only gets DidUpdateFaviconURL() for the
navigation entry associated with the most recent main frame navigation. A
RenderViewHost is in the STATE_PENDING_SWAP_OUT state if the navigation
for a different page has been committed but the FrameHostMsg_SwapOut_ACK has
not yet been received. Drop ViewHostMsg_UpdateFaviconURL messages if the
RenderViewHosts is in the STATE_PENDING_SWAP_OUT state.

BUG=391585
TEST=None

Review URL: https://codereview.chromium.org/1053623005

Cr-Commit-Position: refs/heads/master@{#326087}