Support opening downloaded pdf in platform pdf viewer without showing intent picker
There are a couple places we will launch an intent when downloading content:
1. before the download starts
2. after download finishes.

In both cases, we should always honor the default activity associated with the intent type.
However, if there are multiple activities associated with the intent:
In case 1, passing the url to platform pdf viewer allows the pdf to render progressively.
In case 2, passing the pdf to platform pdf viewer saves the user from interacting with the intent picker.
In either cases, we check whether platform pdf viewer is available or not.
The platform pdf viewer is available through google drive app, which is part of GMS.

BUG=478304

Review URL: https://codereview.chromium.org/1092053006

Cr-Commit-Position: refs/heads/master@{#326104}