Adds PointerEvent type behind a flag.
This is our first patch on our way to implementing
PointerEvent in chromium:
  http://www.w3.org/TR/touch-events/#touchevent-interface

This CL adds the PointerEvent type in core/events, behind the RuntimeEnabledFeatures flag "PointerEvent". The CL also adds a test in the virtual/pointerevent test-suite (which will pass in the main test suite when we will flip the flag).

BUG=476563

Review URL: https://codereview.chromium.org/1087013003

git-svn-id: svn://svn.chromium.org/blink/trunk@194168 bbb929c8-8fbe-4397-9dbb-9b2b20218538