Add custom pylintrc files to perf and telemetry tools.
This CL adds minimal pylintrc files to the src/tools/perf and
src/tools/telemetry directories. These pylintrc files leave almost everything as
the sane default, with the exception of the indent size and lots of disabled
warnings and errors.

The set of disabled messages is currently identical to the set disabled by the
default pylintrc in depot_tools. As lint errors get fixed (in particular,
unused-argument and unused-import), they should be removed for stricter linting.

R=nednguyen@chromium.org, sullivan@chromium.org
BUG=475714

Review URL: https://codereview.chromium.org/1095423002

Cr-Commit-Position: refs/heads/master@{#326154}