Re-factor VisibleSelectionTest for adding tests for composed tree version
This patch changes |VisibleSelectionTest| for adding tests for composed tree
version easier with introducing shadow tree related tests.

This patch also introduces |Node| and |Position| for ease of writing and
debugging gTest.

This patch is result of collaboration work with hajimehoshi@chromium.org
for selection of web components. After this patch, we are going to
implement another strategies for the composed tree. We've already
prepared that at http://crrev.com/1009603011.

BUG=275851
TEST=n/a; No behavior changes.

Review URL: https://codereview.chromium.org/1052623005

git-svn-id: svn://svn.chromium.org/blink/trunk@194195 bbb929c8-8fbe-4397-9dbb-9b2b20218538