Merge M43:Fix stuck error screen after reconnecting WiFi in OOBE at sign-in.
The core of this issue was the split codepath in OobeUI::OobeUI(), which
checks if there is a WizardController available already, and creates an
ErrorScreen if it's not.

The problem is that in WizardController::CreateScreen() there is no
symmetrical check that an ErrorScreen was already created, although to
create a new one, it uses the same NetworkErrorView instance (which is
actually an ErrorScreenHandler), so it unbinds this channel from the old
ErrorScreen model thus making it unusable on three screens (Enrollment,
Sing-in and App Launch Splash) if there ever was a preceding call to
WizardController::CreateScreen(kErrorScreenName).

This fix makes sure that we only create one ErrorScreen for the one
ErrorScreenHandler that we have in OobeUI.

BUG=470977

Review URL: https://codereview.chromium.org/1094103002

Cr-Commit-Position: refs/heads/master@{#325995}
(cherry picked from commit ac8a005907d0e9bbb74d87c2f12b8610cf00f989)

Review URL: https://codereview.chromium.org/1094173003

Cr-Commit-Position: refs/branch-heads/2357@{#189}
Cr-Branched-From: 59d4494849b405682265ed5d3f5164573b9a939b-refs/heads/master@{#323860}