Reland: Introduce sys_sigprocmask and sys_sigaction. (patchset #5 id:160001 of https://codereview.chromium.org/1077143002/)
This CL is relanding the CL https://codereview.chromium.org/1077143002/,
which was reverted due to false alarm of the memory sanitizer.
Because the memory sanitizer does not know about direct kernel syscall
does, so it warns the arguments of the signal handler, unexpectedly.
This CL adds the msan unpoisoning to the arguemtns, so that the memory
sanitizer should understand it.

TEST=Ran sandbox_linux_unittests with msan=1. Ran bots.
BUG=358465
CQ_EXTRA_TRYBOTS=tryserver.chromium.linux:linux_chromium_trusty32_rel,linux_arm

Review URL: https://codereview.chromium.org/1092153005

Cr-Commit-Position: refs/heads/master@{#326270}