PlzNavigate: do not send FailedNavigation IPC when the navigation was aborted
This CL makes it so that the Navigator will not try to send a
FrameMsg_FailedNavigation when the request failed with an error code of
net::ERR_ABORTED. No error page should be shown in that case. This case can
happen when the NavigationResourceHandler is replaced by a
DownloadResourceHandler during a download.

BUG=475027

Review URL: https://codereview.chromium.org/1089603004

Cr-Commit-Position: refs/heads/master@{#326290}