Forget SSL error exceptions when good certs seen for regular requests.
Chrome remembers decisions by the user to proceed through SSL errors.
However, it remembers this even after a good certificate has been seen
for the given host. This change removes all previous exceptions for a
given host after a good certificate is seen on a regular request,
but not on redirects.

In the SSLPolicy, this adds a call to RevokeUserAllowExceptions() on the
SSLHostStateDelegate when a request response begins. This removes all
prior exceptions for the specified host. However, there is currently no
similar plumbing for redirects, so until that plumbing is built,
revocation will not occur when a valid cert for a host is seen on
redirect.

BUG=473390

Review URL: https://codereview.chromium.org/1058003004

Cr-Commit-Position: refs/heads/master@{#326332}