Don't implement snprintf when building with VS 2015.
VS 2015 (finally) implements snprintf so implementing it in user
code is no longer necessary or legal. This change makes the
definition of snprintf conditional on the compiler version.

VS 2015 also fixes the implementation of vsnprintf so that it
reliably null-terminates. This means that #defining it to a
custom implementation is no longer needed.

R=erik.corry@gmail.com,thakis@chromium.org,tkent@chromium.org,sigbjornf@opera.com
BUG=440500

Review URL: https://codereview.chromium.org/1090093002

git-svn-id: svn://svn.chromium.org/blink/trunk@194231 bbb929c8-8fbe-4397-9dbb-9b2b20218538