Refactoring: Introduce TextOffset for StyledMarkupAccumulator to replace Position
Now a StyledMarkupAccumulator has Position objects but they are
effective only when the positions are text offsets. To make matters
worse, a Position is designed only for the DOM tree and not for the
composed tree. As we need StyledMarkupAccumulator to work on both trees,
Position usage is not appropriate there.

This patch replaces them with a new class TextOffset to make it clear
that |m_start| and |m_end| represent text nodes' offsets.

BUG=n/a
TEST=n/a; no behavior change

Review URL: https://codereview.chromium.org/1084253003

git-svn-id: svn://svn.chromium.org/blink/trunk@194263 bbb929c8-8fbe-4397-9dbb-9b2b20218538