Increasing WebRTC AQ lead-in time to 10 sec to accommodate new slot bot
The new Mac bot in https://build.chromium.org/p/chromium.webrtc.fyi/
shows evidence of chopping of the intro of the ref file, which means
it is slower than the other bots in getting a call up. Since the
purpose of this test isn't to measure call setup time, I'm simply
extending the silence period at the start by five seconds to allow
for this. This will make the tests five seconds slower but I think we
can live with that since it only runs on WebRTC bots.

BUG=477653

Review URL: https://codereview.chromium.org/1099243003

Cr-Commit-Position: refs/heads/master@{#326518}