Fix VKEY determination for non-US layouts under Ozone.
We attempt to match Windows behaviour for the value of VKEY on
a keystroke. Some XKB layouts generate characters for AltGr+key
combinations that are not present in the reference (Windows-based)
sources, which led to mismtaches in the mappings, and incorrect VKEY
values for those keys. To avoid this, we match on any AltGr+key result
in cases the reference has none.

The corresponding test is revised to use a valid dummy character
for those cases, rather than 'NoSymbol'.

BUG=480529

Review URL: https://codereview.chromium.org/1100393005

Cr-Commit-Position: refs/heads/master@{#326695}