Stop flipping Help Icon in Hebrew Locale
Update the help button of the system tray to no longer flip while in the Hebrew
locale.

The asset used is a question mark, which is flipped in most RTL locales. Hebrew
is the exception, which uses the same question mark as LTR locales.

TEST=Manual testing on device
BUG=475237

Review URL: https://codereview.chromium.org/1072423004

Cr-Commit-Position: refs/heads/master@{#326699}