crazy linker: Improved fix for crbug/479220.
Remove IsSystemLibrary() and its list of NDK-exposed system libraries.
It is incomplete and not completable.

Replace it with a flag argument to LoadLibrary() that is true if the
library load is prompted by either preloads or dependencies.  Assume
the library being loaded is a system library and load through dlopen()
when the flag is true.

BUG=479220

Review URL: https://codereview.chromium.org/1105893002

Cr-Commit-Position: refs/heads/master@{#326838}