Fix flaky browser test SSLUITest.BadCertFollowedByGoodCert
A flaky test was added in https://codereview.chromium.org/1058003004/
that involved modifying the root cert cache, which did not happen
reliably. This changes and simplifies the test to use two different
servers on the same host, so the test passes since it relies on
same-host not same-origin, but avoids modifying any caches.

R=rsleevi@chromium.org
BUG=480667

Review URL: https://codereview.chromium.org/1080593006

Cr-Commit-Position: refs/heads/master@{#326840}