Remove the pending container size requests list
Trust ImageResource::setContainerSizeForLayoutObject to be called
through ImageResource::notifyObservers when the image has loaded.
Given that, we will already get notifications about the up-to-date
containerSize and containerZoom and there is no need for keeping
around a list of pending container sizes and zooms to send when the
image is created.

As to why we can trust ImageResource::notifyObservers to call back
into ImageResource::setContainerSizeForLayoutObject,
https://codereview.chromium.org/22482004 added a
updateInnerContentRect() call taking care of calling
ImageResource::setContainerSizeForLayoutObject when the image's paint
has been invalidated.

Review URL: https://codereview.chromium.org/1102803002

git-svn-id: svn://svn.chromium.org/blink/trunk@194439 bbb929c8-8fbe-4397-9dbb-9b2b20218538