Move the unit tests files side-by-side with the corresponding .cc files.
This is a mechanical change.

The idea here is that we want to simplify the directory tree of
the component updater files and easily see what compilation unit has
tests and which one does not.

This change includes some whitespace changes due to running git cl format.

BUG=479904

Review URL: https://codereview.chromium.org/1102103002

Cr-Commit-Position: refs/heads/master@{#327054}