Pass a TaskRunner through ViewsDelegate to put AX init on FILE thread
Adds an interface to ViewsDelegate to get a TaskRunner for a thread that
can be used for initialization tasks (implemented using the FILE thread),
and passes that TaskRunner to the ATK accessibility initialization code,
which currently does I/O on the main thread, causing a perf issue.

This change just adds the plumbing to get the TaskRunner there.
This will be followed up by this change that actually moves the initialization
to the other thread: https://codereview.chromium.org/1028553003/

BUG=468112

Review URL: https://codereview.chromium.org/1089003003

Cr-Commit-Position: refs/heads/master@{#327117}