Re-land Move ScopedFileDescriptor to dbus/file_descriptor.h
ScopedFileDescriptor was initially introduced to facilitate the
management of a 'lifeline' FD used when asking permission_broker to
open a port in the device's firewall on CrOS.  It's actually a scoped
wrapper around DBus::FileDescriptor, not a platform file descriptor. I
would like to use it to facilitate some other similar functionality in
which I create a pipe and pass one end over DBus, so it seems like a
good idea to move ScopedFileDescriptor into the dbus library directly.

This reverts commit ee9dc849ffa2df2820381c31d5ba8200774d287a.

BUG=chromium:481340
TEST=FirewallHole unit tests still pass
TBR=stevenjb, reillyg, keybuk

Review URL: https://codereview.chromium.org/1108083003

Cr-Commit-Position: refs/heads/master@{#327141}