[Slimming paint] Fix a crash when drawing SVG images
This patch creates a recording context for slimming paint, and reuses
the existing context when slimming paint is disabled. This fixes an
assertion about a null context when drawing svg background images.

This patch also removes the svg-as-background-4.html test which has
a 1px difference in debug vs release. This test actually covers
functionality we do not and will not support: javascript in SVG
background images. This is well tested elsewhere (both background
images and javascript-in-svg not working) so I've removed this test.

BUG=479081

Review URL: https://codereview.chromium.org/1109863002

git-svn-id: svn://svn.chromium.org/blink/trunk@194569 bbb929c8-8fbe-4397-9dbb-9b2b20218538