Mac: App menuBars: Ignore closing windows if they are not main
Currently when a window that is not main (e.g. the app list) completes a
fade-out animation and closes, it will trigger a mainMenu update. We
need to ensure the mainMenu says "Chrome" when the last window on the
active space is closed. But if that didn't have main status, it won't
have been influencing the mainMenu to begin with and also won't trigger
a later NSWindowDidBecomeMainNotification.

This regressed in r323157, but the bug was always present. It was
previously avoided because updates were skipped by anticipating whether
AppKit would make a subsequent window main. But that's hard to get right
and conflicts with behaviour used when closing panel windows.

BUG=481803
TEST=On Mac, open a packaged app window with the app launcher. When
launched for the first time, the menubar should say the apps name,
rather than "Chrome".

Review URL: https://codereview.chromium.org/1114803002

Cr-Commit-Position: refs/heads/master@{#327474}