cc: Remove tile sharing from tilings.
This patch removes sharing tiles from tilings. Previously, tiles
that were not invalidated were shared between the pending and
the active trees. With this patch, tiles that are exist on the
active tree and are not invalidated are not created on the pending
tree, resulting in fewer overall tiles. This improves performance
of updating and managing tiles.

Also, this patch opens up opportunities for a lot of code clean
up that can otherwise be confusing when two trees are involved.

R=danakj, enne

Committed: https://crrev.com/0061b75e810f088900a0a8fbdd29ba5595c73016
Cr-Commit-Position: refs/heads/master@{#326804}

Review URL: https://codereview.chromium.org/1051993002

Cr-Commit-Position: refs/heads/master@{#327553}