Allow Raw Input gamepads to persist across page refreshes
The problem was that we were clearing the controller collection when
pausing input, but only repopulating it when the connected device list
changed (a device was connected or removed). Now we only clear the list
on device list changed events or when the RawInputDataFetcher is deleted.

Also fixed an error windows was throwing that indicated we were passing
the wrong parameters to RegisterRawInputDevices when pausing input
monitoring.

BUG=413805

Review URL: https://codereview.chromium.org/1113913003

Cr-Commit-Position: refs/heads/master@{#327603}