Lazily create feedback reporters using application context.
Currently the feedback reporters for child accounts and DOM Distiller
require a call to their respective setExternalFeedbackReporter(...)
methods to happen before they can be used.

This CL changes this around so they are only created lazily whenever
the feedback reporter is needed the first time. The result of the creation
is still cached in a static member like before. The creation happens through
the use of the application context (ChromiumApplication). The upstream version
of these methods just return no-op implementation of the interface.

This new implementation still supports manually calling the set-method,
so rolling this to the downstream code base can happen without build
or functionality breakage. As soon as the downstream change has landed
and sticks, there should be a follow-up cleanup CL to remove the now
deprecated set-method.

BUG=428841

Review URL: https://codereview.chromium.org/1113453004

Cr-Commit-Position: refs/heads/master@{#327618}