Disable the chrome://apps page on ChromeOS.
This CL removes the chrome://apps page handler on ChromeOS as it was
already showing no apps and was confusing if you wound up there.

Ultimately this should be done in conjunction with hunting down the
resources on all platforms/scenarios and simplifying the
caching.

BUG=415225

Review URL: https://codereview.chromium.org/1076273003

Cr-Commit-Position: refs/heads/master@{#327649}