Register ContentViewCore for AccessibilityStateChangeListener
The state of AccessibilityManager is only being checked at particular
events like onShow() and onAttachedToWindow(). As a result, when the
state changes after a page is loaded, such as by UIAutomator,
accessibility will not be enabled until the WebView is reloaded or
background and foregrounded.

The interface for AccessibilityStateChangeListener was already
implemented in ContentViewCore with the onAccessibilityStateChanged
callback. However, ContentViewCore was never registered to receive the
callbacks.

BUG=482746

Review URL: https://codereview.chromium.org/1109393010

Cr-Commit-Position: refs/heads/master@{#327665}