[bindings] Remove custom setter used in CSSStyleDeclaration
This CL is a first in the series of CSSStyleDeclaration's usage of custom bindings.
Here we migrate the helper methods required from the custom bindings as is.
These duplicate helper functions would be eventually removed from the custom binding as we
finish the removal.

BUG=345519

R=haraken@chromium.org

Review URL: https://codereview.chromium.org/1119653002

git-svn-id: svn://svn.chromium.org/blink/trunk@194718 bbb929c8-8fbe-4397-9dbb-9b2b20218538