Initializes SecurityPolicy static locals during blink initialization
These static locals could be touched before we initialize threading
(which is necessary to call AtomicallyInitializedStaticReference)
in chrome build, so simply using AtomicallyInitializedStaticReference
doesn't work well. On the other hand in non-chromium build we need
to make sure they're initialized before or during the blink initialization.

This patch does:
- changes AtomicallyInitializedStaticReference back to DEFINE_STATIC_LOCAL
- adds explicit SecurityPolicy::init() method and call it in the blink
  initialization
so that static locals are properly initialized regardless of whether
it's in chrome build or not.

(Note: without this patch chrome starts to crash once we land this patch:
https://codereview.chromium.org/1072933006/ and give the whitelisting options)

BUG=441605
TEST=green bots

Review URL: https://codereview.chromium.org/1117513003

git-svn-id: svn://svn.chromium.org/blink/trunk@194750 bbb929c8-8fbe-4397-9dbb-9b2b20218538