[Mac] Skip a redundant deminiaturize in ui::FocusWindowSet
AppKit has a bug that can cause multiple windows to simultaneously
obtain key/main status under a hard-to-pin-down set of actions described
in http://crbug.com/459306.

FocusWindowSet may choose a miniaturized window to make the new
keyWindow, and (since r213829) it would first deminiaturize it before
making it key. However, this deminiaturize is redundant due to a later
call to [makeKeyAndOrderFront:]. Performing both seems to triggers a
race condition in AppKit in some circumstances. So remove it.

BUG=459306
TEST=See http://crbug.com/459306#c17

Review URL: https://codereview.chromium.org/1119533006

Cr-Commit-Position: refs/heads/master@{#327827}