Remove persistent notifications from the database when they close.
This CL implements said behavior both for Android (by directly
invoking PNSI::OnClose when an Intent is received) and for desktop
through the PersistentNotificationDelegate.

Tests will follow with a more elaborate suite for the database,
when it's clear what lives in //chrome and what lives in //content.

Depends on https://codereview.chromium.org/1099093003/
BUG=447628

Review URL: https://codereview.chromium.org/1071773003

Cr-Commit-Position: refs/heads/master@{#327912}