Fix GN visibility for a few targets that are compiled by default in a GN build.
When we build 'all' in GN, we can end up building targets that are
defined in various build files even if they are not transitively referenced
from 'gn_all' or 'gn_only'. This is undesirable because it makes it harder
to tell when there might be discrepancies between the GN and GYP builds.

Some of these targets may actually be hidden (using the visibility tag)
and only intended to be depended on by other targets that don't work yet.
By making this targets also visible to a top-level "//:gn_visibility"
target we can ensure that we're tracking them properly without breaking
the normal encapsulation rules.

The two targets affected in this patch are
- //build/config/sanitizers:options_sources
- //ui/resources:repack_ui_test_mac_locale_pack

R=brettw@chromium.org
BUG=431177
CQ_EXTRA_TRYBOTS=tryserver.chromium.mac:mac_chromium_gn_rel,mac_chromium_gn_dbg

Review URL: https://codereview.chromium.org/1121453004

Cr-Commit-Position: refs/heads/master@{#327922}