Restore simple fling status bookkeeping
Recently, an active fling counter was introduced to track whether a
fling was active in the renderer. This was useful both for filtering
fling cancel events as well as signalling the end of a fling gesture
for telemetry purposes.

However, there appear to be gaps in the fling accounting that don't
appear in telemetry but do appear in the wild, causing issues with
fling-related tap suppression logic. Restore the previous fling
bookkeeping for the GestureEventQueue until the in-flight fling
count is correct for all gesture types.

BUG=483037

Review URL: https://codereview.chromium.org/1113143002

Cr-Commit-Position: refs/heads/master@{#327993}