[S.P.] Fix under-invalidation of media volume slider on volume change
Normally paint invalidation should be driven by style change, but for
some custom-painted controls which paint themselves differently
according to element attributes/properties, we should invalidate
the layout object whenever the attribute/properties change.

Media volume slider paints itself differently according to volume.
Needs to invalidate it on volume change.

Was not a problem for non-slimming-paint because the invalidation rect
of the slider track on volume change covers the whole slider.

BTW fixed bug 282094 which was caused by a typo in the test.

BUG=482174,282094
TEST=virtual/slimmingpaint/fast/repaint/video-mute-repaint.html
TEST=virtual/slimmingpaint/fast/repaint/video-unmute-repaint.html

Review URL: https://codereview.chromium.org/1119143002

git-svn-id: svn://svn.chromium.org/blink/trunk@194843 bbb929c8-8fbe-4397-9dbb-9b2b20218538