blink: Enable the GPU trigger when viewport meta tag is not read
When we're on a device that isn't using the viewport meta tag 
(e.g. desktop), we turn on the GPU trigger.

If we're emulating a device on desktop, the trigger does respect the meta
tag.

One gotcha for this change, if you use --single-process, we don't read
the white/black list before figuring out if we should enable GPU, so
it gets enabled by default -- I'll log an issue for this.

Please see https://codereview.chromium.org/813773002/ for the other half
of this change.

BUG=440518

Review URL: https://codereview.chromium.org/792233007

git-svn-id: svn://svn.chromium.org/blink/trunk@194901 bbb929c8-8fbe-4397-9dbb-9b2b20218538