Fix SMIL UseCounter for cached SVG images
We cache complete SVG image documents in the cache which can skew
UseCounter numbers that need to be called for every SVG image. This
patch adds a new UseCounter entry for SMIL regardless of cache and
calls it on every SVG image load. The old UseCounter entry is wrong
and has been removed.

BUG=484004

Review URL: https://codereview.chromium.org/1118733003

git-svn-id: svn://svn.chromium.org/blink/trunk@194911 bbb929c8-8fbe-4397-9dbb-9b2b20218538