Load and apply a vcgt table from an ICC file to the internal display
When the command line switch --internal-display-color-profile-file is provided
load the given ICC file using qcms and extract the VCGT data. Use this VCGT
data to apply a gamma ramp to change the output on the internal display using
drmModeCrtcSetGammaRamp.

BUG=471749
TEST=On a link_freon device add the command line option to load a sample ICC
file (e.g. Bluish.icc to give a blue tint) and observe on startup that the
internal display is blue tinted.

Review URL: https://codereview.chromium.org/1028563003

Cr-Commit-Position: refs/heads/master@{#328318}