Don't allow WebViewPlugin to swallow GestureLongPress.
Blocked-plugins are represented by a subclass of WebViewPlugin. Prior to
https:// src.chromium.org/viewvc/blink?view=rev&revision=194630 this
plugin never received GestureLongPress events, but now it does and the
plugin presently consumes these events during the default event handler
phase, causing the context menu not to show when a GestureLongPress
occurs.

This CL modifies WebViewPlugin to return false from handleInputEvent()
when the event is GestureLongPress, allowing the context menu to be
shown.

BUG = 482842

Review URL: https://codereview.chromium.org/1124693002

Cr-Commit-Position: refs/heads/master@{#328319}