Invalidate ancestor resources in when relative lengths change
If a LayoutObject within a resource had relative lengths, and the viewport
changed, only the object itself would be invalidated - not any of the
clients of the resource. This would fail to invalidate the cached display
item with Slimming Paint.
Modify SVGClipPathElement::selfHasRelativeLengths() to indicate if any
of the children have relative lengths, to make sure the clients are
invalidated in that case. (SVGClipPathElement will otherwise never indicate
that it has relative lengths.)

Fixes svg/custom/viewport-resource-inval.svg with Slimming Paint enabled.

BUG=477743

Review URL: https://codereview.chromium.org/1111223002

git-svn-id: svn://svn.chromium.org/blink/trunk@194949 bbb929c8-8fbe-4397-9dbb-9b2b20218538