Prefer display lists over GPU for very large canvases
This change adds a preference for display lists over the old GPU
accelerated code path for 2d canvases with more than 4k-by-4k pixels.
This helps prevent very large canvases from hogging VRAM and allows
the rendering of the canvas contents to by culled based on visibility,
which is useful for large canvases. The threshold was chosen to match
the previous limit that was globally enforced for texture allocations,
which was working out well for 2D canvas.

BUG=117949

Review URL: https://codereview.chromium.org/1128163002

git-svn-id: svn://svn.chromium.org/blink/trunk@194988 bbb929c8-8fbe-4397-9dbb-9b2b20218538