Add Finch experiment to cookie monster.
The Finch experiment measures the effect of the cookie fetching strategy on
startup time. Under the current logic, as soon as any cookie is requested, all
cookies are fetched. In one of the paths of the Finch experiment, all cookies
are fetched only if all cookies are required.

BUG=473483

Review URL: https://codereview.chromium.org/1052373003

Cr-Commit-Position: refs/heads/master@{#328603}