[Media Router] Dialog controller implementation.
The MediaRouterDialogController is lazily created per WebContents using
the WebContentsUserData pattern. It provides APIs to create, query, and
to close the dialog.

The internals of the dialog also observe the states of the initiator
and dialog WebContents and handle WebContentes destruction/navigation
events accordingly.

I plan to upstream the unit tests/browser tests in another patch.

BUG=464216

TODO:
- initialize the dialog with info from the initiator once it has
  finished loading
- hook the controller to MediaRouterAction.

Review URL: https://codereview.chromium.org/1122433005

Cr-Commit-Position: refs/heads/master@{#328642}