Make Mac pinch thresholding apply only to zoom
Add a zoomDisabled flag (disabled so that it will default to false
everywhere except where we set it) to WebGestureEvent.

Note that the equivalent blink::Event does not have a corresponding
gesture sub-type that sends pinch scale information, so the
zoomDisabled information is not passed on either.

Chromium side:
https://codereview.chromium.org/1129693002

BUG=478981

Review URL: https://codereview.chromium.org/1127013002

git-svn-id: svn://svn.chromium.org/blink/trunk@195011 bbb929c8-8fbe-4397-9dbb-9b2b20218538