cc: Make HasImplThread() function work in AfterTest().
Before we run AfterTest() the layer_tree_host is destroyed (to allow AfterTest
to check that proper cleanup has occurred) which means that the proxy() object
is null. This means that HasImplThread() would previously return false in the
AfterTest() method even when running on a test which was created using
RunTest(threaded=true).

R=danakj

Review URL: https://codereview.chromium.org/1126973002

Cr-Commit-Position: refs/heads/master@{#328667}