Use Clang's update.py --print-revision everywhere.
Since yesterday, it works on Windows. This allows us to not
completely clobber the build from update.py each time, since the
build system tracks the dependency instead.

It should also be possible to use this from gn, so we can delete the
posix-print-revision.py script. The only problem was that the
hack we use for fixing the output in gclient runhooks didn't work.
I've adjusted that to not fail when the hack doesn't work.

BUG=483065
R=scottmg@chromium.org, thakis@chromium.org

Review URL: https://codereview.chromium.org/1127233002

Cr-Commit-Position: refs/heads/master@{#328699}