Add layout test to catch regression in focus visual
Though it was originally reported as <webview> issue on
crbug.com/479050, it was general issue on shadow hosts with
:focus CSS rule.  The symptom was that showing incorrect
focus ring on <webview> (or any shadow hosts with :focus
CSS-rule and tabindex) when any element inside shadow root
has focus.

The :focus matching code was fixed in https://codereview.chromium.org/1080923003/
and this CL adds the test for this case to prevent regression.

BUG=479050
TEST=fast/dom/shadow/css-focus-pseudo-match-shadow-host.html

Review URL: https://codereview.chromium.org/1114443002

git-svn-id: svn://svn.chromium.org/blink/trunk@195028 bbb929c8-8fbe-4397-9dbb-9b2b20218538