Add a layout test for a partially loaded BMP image
Adapt the jpeg test to use a bmp image and partially load the image to a
byte boundary near half the image size.

Note: the test bmp image loads bottom-to-top for some reason, so the top
half of the image test result should be a green.

TBR=pkasting@chromium.org
NOTRY=true
BUG=485527

Review URL: https://codereview.chromium.org/1136483002

git-svn-id: svn://svn.chromium.org/blink/trunk@195050 bbb929c8-8fbe-4397-9dbb-9b2b20218538