Fix a bug in web_drag_dest_mac_unittest.mm.
The test puts a string representation of a file path into PasteBoard, and then
extracts the string again. The logic that extracts the string first converts
the string into a file URL. The implementation for that logic in CoreFoundation
has different behavior depending on whether Chrome is linked against an OSX
10.9+ SDK. The expectation of the test should compare the string extracted from
PasteBoard against the initial string converted into an NSURL, and then back
into a string.

BUG=484203

Review URL: https://codereview.chromium.org/1128173004

Cr-Commit-Position: refs/heads/master@{#328791}