views: Avoid a malloc for every view being painted.
Instead of making a scoped_ptr<ClipTransformRecorder>, we just
instantiate one, and make the ClipTransformRecorder into a no-op if
no clips or transforms are set on it.

This patch changes the time to do UpdateLayers (after making
gfx::ImageSkiaReps immutable cuz they dominate times for now) for a
single tab loading spinner from 0.199594ms to 0.179203ms on a linux
z620, giving a 10% reduction in frame record time. I used a mean of
3000 samples in each run to reduce noise.

R=sky
BUG=466426

Review URL: https://codereview.chromium.org/1125363003

Cr-Commit-Position: refs/heads/master@{#328824}