Add a manifest property declaring USB printers supported by an app.
This new manifest property "usb_printers" declares a set of USB device
filters matching the printers supported by an app. This will be used by
the printerProvider API to allow the user to select a printer that can
be supported by the app and in so doing grant the app permission to
connect to the printer.

BUG=468955

Review URL: https://codereview.chromium.org/1123093003

Cr-Commit-Position: refs/heads/master@{#328845}