Fix key derivation for encrypted cert reports
1. Fix order of HKDF arguments: the salt was being passed as the label
and the label as the salt.
2. Pass sizeof(label) to the StringPiece constructor to match server.

This CL also adds a "known gold" test with a corresponding server-side
test. Hopefully if the client-side cert report encryption/decryption
code changes in such a way that the server would no longer be able to
decrypt reports, this test will catch it.

BUG=

Review URL: https://codereview.chromium.org/1131763005

Cr-Commit-Position: refs/heads/master@{#328872}