Properly notify ImageFrame of pixel changes from the BMP/ICO decoders
ImageFrame::setPixelsChanged()/notifyBitmapIfPixelsChanged() were added by
https://codereview.chromium.org/75303004 , which modified the other decoders to
call them.  Alpha asked if BMP and ICO needed to be modified, and the author
replied:

"I don't think we need any changes to the BMP and ICO decoders:

"It looks like BMP doesn't support progressive decoding or multiple frames.
BMPImageReader::decodeBMP initializes the ImageFrame (which then has
generationID==0). The function either completes decoding or it fails.

"ICO uses a BMP or a PNG decoder internally. For BMP we don't need to do
anything. For PNG, the ICO decoder calls PNGImageDecoder::frameBufferAtIndex,
which handles the change notification."

While the part about PNG-in-ICO is correct, the part about BMP is not.  BMP does
support progressive decoding, and decodeBMP() does not fail if the image cannot
yet be decoded completely (in the sense of calling setFailed()).  Instead, like
the other image decoders, it decodes as much as possible and then returns
without marking the frame complete.  (Hint to future authors: get the guy who
actually wrote the decoder you're analyzing to review your analysis of it...)

The consequence of this is not clear to me because I don't understand what
SkPixelRef generation IDs do.  I hypothesize that, for BMPs and BMP-in-ICOs, the
result is that a second decode after a first, incomplete decode will not always
result in everywhere using the updated pixel results.  Since I don't know,
though, I haven't filed a bug, and I don't know how to test this (manually or
otherwise).  Information welcome.

This change adds the appropriate bits to these decoders to do the notifications:
* ICOImageDecoder::frameBufferAtIndex() now calls notifyBitmapIfPixelsChanged().
  For PNG-in-ICO this is unnecessary as the nested
  PNGImageDecoder::frameBufferAtIndex() call will have done this, but it's
  harmless (will quickly no-op).
* BMPImageDecoder::frameBufferAtIndex() also calls this.
* Most importantly, BMPImageReader::decodeBMP() now calls setPixelsChanged()
  when calls to ProcessRLEData() or processNonRLEData() result in decoding more
  of the image.  Without this call, the above calls would do nothing.  This is
  implemented by means of a function-local scoping class.  (I could have just
  put the necessary conditionals in the code directly, but it would have
  involved repeated code blocks, and I would have needed to rework some of the
  existing code to have a shared exit point before which I could insert the
  appropriate call.  Hopefully my solution is not too overdesigned.)

It would also be possible to call setPixelsChanged() directly from e.g. setI()
and setRGBA(), but I'm concerned about the potential performance impact of doing
these calls (even cheap ones) on every pixel.  (In the original review, Alpha
had the same concerns about a similar such suggestion.)  Several other decoders
call this once per row, since that's architecturally convenient; here it was
just as easy to do them "almost optimally rarely".

TEST=None
BUG=485527

Review URL: https://codereview.chromium.org/1123363006

git-svn-id: svn://svn.chromium.org/blink/trunk@195107 bbb929c8-8fbe-4397-9dbb-9b2b20218538