Add support for escaped target names in isolate driver (3rd try)
Currently the isolate_driver.py which creates the dependency files
used by the isolate system, does a simple split on all spaces when
trying to identify targets.

This can fail if the target name contains a space in the name. In
ninja, spaces are escaped with a $-prefix. An example would be
'Content$ Shell$ Helper.app'.

This CL adds support for such target names and ensures that they
stay as one item. Doing this uncovered a few missing dependencies
and a missing file in some .isolate-files for the component build
on Mac.

1st try: https://codereview.chromium.org/970203003/
1st revert: https://codereview.chromium.org/985753002/
2nd try: https://codereview.chromium.org/1103793002/
2nd revert: https://codereview.chromium.org/1129493003/
BUG=462248

Review URL: https://codereview.chromium.org/1130523003

Cr-Commit-Position: refs/heads/master@{#329038}