Sync: add a test for downloading a tombstoned bookmark
This CL adds the first test that injects a tombstone into FakeServer (a
fancy way of saying deleting the entity).

As part of this CL, a util to wait for a bookmark count to match was
added. A util function to trigger a sync cycle given a ModelTypeSet was
also added.

BUG=365774

Review URL: https://codereview.chromium.org/1137543002

Cr-Commit-Position: refs/heads/master@{#329205}