Ensure WebTouchPoint ids remain in order in WebTouchEventBuilder.
Consumers of WebTouchEvents may expect that the touches are in the
same order from one event to the next. WebTouchEventBuilder is
encountered in some pathways but not others (e.g. WebTouchEvents
passed to plugins go through WebTouchEventBuilder).

This CL sorts the touches in WebTouchEvents that are constructed
by WebTouchEventBuilder.

BUG=463747

Review URL: https://codereview.chromium.org/1132643002

git-svn-id: svn://svn.chromium.org/blink/trunk@195224 bbb929c8-8fbe-4397-9dbb-9b2b20218538