Fix use-after-free on multiple same name CSS Animations with different timings
This change fixes an ASAN use-after-free crash when running two animations
that share the same @keyframes name on the same element with different
timings during a style update.
This scenario rides on an existing bug where we only store the last
animation for a given @keyframes name per element but process updates
for each specified animation on the same animation. This caused a scenario
where updates that replaced the KeyframeEffectModel were also triggering
synthetic keyframe updates on the same animation. The ordering of these
updates left a dangling RawPtr that was then used in appropriately
when updating expired synthetic keyframes.

BUG=484614

Review URL: https://codereview.chromium.org/1136203002

git-svn-id: svn://svn.chromium.org/blink/trunk@195229 bbb929c8-8fbe-4397-9dbb-9b2b20218538