Plumbs through screen size and scale factor to html_viewer
As ResourceBundle uses the scale factor we gate loading until we know
the scale factor.

I also made it so ViewportMetrics::device_scale_factor is 0 until we have a real value. This way ViewManager clients have a way to know when the real value is available.

BUG=none
TEST=none
R=ben@chromium.org

Review URL: https://codereview.chromium.org/1136743003

Cr-Commit-Position: refs/heads/master@{#329514}