Share backing canvases between browser compositors.
The software compositor's canvas is only used between BeginPaint and EndPaint. Those happen consecutively for a compositor, so we can share a single buffer to back multiple canvases, one per compositor. This cuts down on memory used to that necessary to contain just the largest window of the process.

Review URL: https://codereview.chromium.org/1132133004

Cr-Commit-Position: refs/heads/master@{#329525}