Disallow connecting ScriptProcessor with 0 output channels to any node
Do not allow a ScriptProcessor node that was created with 0 output
channels to be connected to any destination.  Since there are no
output channels, even though a ScriptProcessor always as one output,
it seems best not to allow any connections.  It's not clear what the
input to the node should be.

BUG=484606
TEST=scriptprocessonode-0-output-channels.html

Review URL: https://codereview.chromium.org/1131393002

git-svn-id: svn://svn.chromium.org/blink/trunk@195314 bbb929c8-8fbe-4397-9dbb-9b2b20218538