Remove the concept of a cleanup task
The only cleanup task in Blink which is not internal to a worker thread is
the task to close a WebSQL database. Because 1) only worker threads
support the concept of a cleanup task, and 2) WebSQL isn't supported on
worker threads anymore after https://codereview.chromium.org/561093003/,
we can safely remove the logic to handle cleanup tasks on worker
threads.

BUG=463143
TBR=haraken@chromium.org,jochen@chromium.org

Review URL: https://codereview.chromium.org/1127123010

git-svn-id: svn://svn.chromium.org/blink/trunk@195345 bbb929c8-8fbe-4397-9dbb-9b2b20218538