Update marking proxies as bad in ProxyList/ProxyService to use a std::vector
Currently, various methods related to marking proxies as bad in the ProxyList
and ProxyService classes take a single ProxyServer for optionally marking an
additional proxy as bad (i.e. ProxyList::UpdateRetryInfoOnFallback and
ProxyService::MarkProxiesAsBadUntil). This is linked to fundamental assumptions
in the Data Reduction Proxy that only two proxy servers are being used.

As part of moving to a dynamic configuration, it is possible that more than 2
proxy servers could be used for the Data Reduction Proxy, so using a vector
of ProxyServer objects makes more sense in the interface.

BUG=470587

Review URL: https://codereview.chromium.org/1128823005

Cr-Commit-Position: refs/heads/master@{#329861}