Fix latency calculation when multiple rendered buffers are in flight.
This is particularly important on ChromeOS with freon stack.
SwapBuffers ack is sent when the buffers actually swapped, not when the
call to swap them was made. This changed the scheduler behavior and delayed
beginning of the next frame, so to regain it we allowed multiple in flight
frames. The effect this has on latency info is that two frames arriving in
rapid succession would override incorrect latency info making it look like
things ultimately happened earlier than they did.

This patch associates latency info with a particular frame's callback
thus avoiding it being overridden.

BUG=485302

Review URL: https://codereview.chromium.org/1128323006

Cr-Commit-Position: refs/heads/master@{#329936}