Replace SW YUV conversion with higher quality Shader+FBO in pepper video path.
Improve quality of software rendering path by replacing call to SkCanvasVideoRenderer::ConvertVideoFrameToRGB in video_decoder_shim with a shader and render-to-texture.

Questions about this CL:

* gpu::gles2::GLESInterface
Would it be correct to send a base::WeakPtr<GLESInterface> to the YUVConv
class constructor to use for everything in this instance of the class?
ATM, it just sends a GLESInterface* to methods that need it (which is
probably wrong).  It was not clear to me whether the GLESInterface * could
change between the ::Initialize() method

* OpenGL state
YUVConv::convert() changes texture bindings and viewport.  I currently read
the old viewport so that I can restore it - is this necessary?  I don't save
former texture bindings either - is this necessary?

* OpenGL 'getters'
When compiled debug, why do all output variables to GL functions need to
contain either 0 or -1 before calling the function?  In my case GetIntegerv
and GetShaderiv.

* Errors
It is possible that shader compile could fail.  If it fails, an error string
can be retrieved.  What should I do in this case?
Same with opengl errors.  I thought it would be good to check for them in
debug builds, but what do I do if I catch one?

* Texture formats
I need a single-component texture format for uploading the Y,U,V components.
On desktop I would use GL_RED, but unextended GLES does not have GL_RED, you
have to use GL_LUMINANCE.  The current code uses GL_RED, and it works, but do
I need to check for GLES extension for GL_RED or can I assume it is there?

* Texture sizes
Even with the description in video_frame.h the distinction between coded_size
visible_size and natural_size was not entirely clear to me.  It seemed the
output texture size in VideoDecoderShim::OnOutputComplete followed "coded_size"
so that is what I used as well.  It was my intention to always decode the
entire image to the user's buffer, not some subregion - so is my code correct?

* FBO's
I see that DrawBuffers is DrawBuffersEXT, and that the default GLES
implementation does not support it.  Is it not necessary in this environment
to call DrawBuffers?

* ::SendPictures
If there is a format error yuv_converter will return false.  How can an error
be handled here?

* testing
I have tested with the video_decoder example program (using VPx) and our
application which uses h264.  I believe both use I420 format.  Do you have an
easy way to test other formats?  Is this class part of a unit test?

BUG=483183,477737
R=bbudge@chromium.org
TEST=Run pepper video_decoder example in ppapi/examples/video_decode

Review URL: https://codereview.chromium.org/1111653004

Cr-Commit-Position: refs/heads/master@{#329942}