Set network_accessed earlier, when network transaction creates stream.
HttpResponseInfo contains a "network_accessed" flag that is currently
set when HttpNetworkTransaction finishes sending the request. This is
not entirely accurate; it ends up false for requests that don't send a
request but do touch the network in other ways (e.g. DNS resolution
fails, or TCP connection is terminated before we finish sending the
request).

This flag is used only by Domain Reliability for checking whether a
request touched the network, and by ServiceWorker for histograms.
horo says ServiceWorker won't be affected, and I would prefer these
semantics for Domain Reliability's use, so I'm changing it.

Note that this also changes several instances of GetResponseInfo
to stop returning NULL.

See also for context:

https://groups.google.com/a/chromium.org/forum/#!topic/net-dev/Wtn65jGyot0

BUG=480565

Review URL: https://codereview.chromium.org/1072423005

Cr-Commit-Position: refs/heads/master@{#330011}