Change WebContents::last_active_time_ to Time instead of Timeticks.
For context, last_active_time_ is going to be used by session restore to order the loading of background tabs using MRU. In order for this to be robust, last_active_time_ must be saved and restore between sessions. Timeticks cannot be reliably restored as it's dependent on the current OS session. MRU code for session restore is being implemented in https://codereview.chromium.org/1131373003

Notes:
- In dev tools, replaced "activity" with "active" for consistency
- In OomPriorityManagerTest.Comparator, initialized last_active_time for all tabs, as this is necessary with Time (a default TimeTicks can go back in time, not the case with Time).

BUG=472772

Review URL: https://codereview.chromium.org/1140083004

Cr-Commit-Position: refs/heads/master@{#330029}