Add DataReductionProxyExperimentsStats and UMA for measuring potentially non-compressed bytes.
If no valid Data Reduction Proxy configuration is available to the client,
then the client does not get the benefit of the Data Reduction Proxy during
the period when a configuration is being retrieved. This CL sets up a
configurable simulation where all requests will continue to go through the
Data Reduction Proxy, but if the request is initiated during a period when
the client would request a configuration from the service, we will record
the reduction statistics to determine the overall effect on data compression.

BUG=484864

Review URL: https://codereview.chromium.org/1127893002

Cr-Commit-Position: refs/heads/master@{#330133}