[Android] Add a regression test for the JB RemoveHolderAndAdvanceLocked issue
This is a re-land of https://codereview.chromium.org/1134543002/ with
a workaround for http://crbug.com/486262, which was causing the initial
test instability.

The test reproduces the condition which was triggering an infinite loop
inside GinJavaBridgeDispatcherHost::RenderFrameDeleted.

Verified that the test indeed hangs up and timeouts if the fix is removed.

BUG=484927,486262
TBR=sgurun@chromium.org,tedchoc@chromium.org

Review URL: https://codereview.chromium.org/1134793005

Cr-Commit-Position: refs/heads/master@{#330135}