bisect-builds.py: Save cache in parent scope
Using 'cache = ...' always assigns ... to the local scope, while it
should be stored in the parent scope instead. Because of this oversight,
the cache only contained one list of revisions at any time, namely the
list of revisions for the current settings (e.g. "-a linux64" or "-l -a
linux64").
This patch fixed the issue by iterating over the cached dictionary and
copying the keys and values to the variable (in the parent scope).

R=rsesek@chromium.org

Review URL: https://codereview.chromium.org/1131403003

Cr-Commit-Position: refs/heads/master@{#330374}