Don't return wall clock times when time isn't ticking.
The video renderer may request wall clock times before the audio
renderer has notified the renderer that it has underflowed.  We
should refuse to convert wall clock times as soon as an underflow
state is detected, instead of waiting for the renderer to pause
everything on the media thread (which may be backed up for decoding
or other events).

BUG=489503
TEST=modified underflow test.

Review URL: https://codereview.chromium.org/1137323005

Cr-Commit-Position: refs/heads/master@{#330481}