Oilpan: Eagerly finalize WebSpeechSynthesizer
Once lazy sweeping is enabled, TtsDispatcher::synthesizer_client_ can become
a stale pointer if WebSpeechSynthesizerClientImpl is swept before the WebSpeechSynthesizer is swept.
(Remember that currently the TtsDispatcher::synthesizer_client_ is cleared in the destructor
of the WebSpeechSynthesizer.)

To avoid that from happening, this CL explicitly clears the TtsDispatcher::synthesizer_client_
in a pre-finalizer of the WebSpeechSynthesizer.

BUG=485843

Review URL: https://codereview.chromium.org/1136633004

git-svn-id: svn://svn.chromium.org/blink/trunk@195525 bbb929c8-8fbe-4397-9dbb-9b2b20218538