Don't detach whitespace nodes when the previous sibling is an out of flow block
A whitespace node was not being rendered after an out of flow block element was
inserted in front of it and the container was an inline element. Normally a
block element would cause the whitespace node to be detached as it should no
longer be rendered since the block element creates a new line. This doesn't
matter though if the block element is out of the normal flow of the document.
Now we correctly verify that the inserted block element is in the normal flow
the document before deciding if no renderer is needed.

BUG=477894

Review URL: https://codereview.chromium.org/1102703002

git-svn-id: svn://svn.chromium.org/blink/trunk@195585 bbb929c8-8fbe-4397-9dbb-9b2b20218538