Drop the document.charset part of fast/encoding/external-script-charset.xhtml
Simplify to only test the inheritance, but use the koi8-r for the XHTML
document to ensure that it's not just the default encoding being used to
decode the script.

The bots cannot apply this patch because of the encoding of
LayoutTests/fast/encoding/external-script-charset-koi-8.js.

BUG=438392
R=pdr@chromium.org

Review URL: https://codereview.chromium.org/1144913002

git-svn-id: svn://svn.chromium.org/blink/trunk@195595 bbb929c8-8fbe-4397-9dbb-9b2b20218538