Don't clip focus ring painting on Mac
Currently we clip focus rings to the visible viewport on Mac if
the focus ring is painted with NSCell::drawWithFrame(). This causes
inconsistent painting when the control is out-of-view (painted
without focus ring) and in-view (painted with focus ring).
With composited scrolling and/or slimming paint, we don't invalidate
the control when it scrolls, so we should not paint it differently.

The issue is revealed by https://codereview.chromium.org/1133363004/
which will paint scrolled contents in coordinates of contents instead
of container.

TBR=chrishtr@chromium.org (delegating to ccameron)
R=ccameron@chromium.org,wkorman@chromium.org
BUG=486790
TEST=fast/overflow/scrollRevealButton.html

Review URL: https://codereview.chromium.org/1131833004

git-svn-id: svn://svn.chromium.org/blink/trunk@195623 bbb929c8-8fbe-4397-9dbb-9b2b20218538