Invalidate SVG filter display items on subtree modification
When an SVG filtered subtree is modified, the SVGFilter display item
needs to be invalidated. For example:
<g filter='url(#filter)'>
    <rect ...>
</g>

This example will generate a display item list of a single SVGFilter
display item for the <g> because the <rect> subtree will be drawn
as an SkPicture as part of the filter. If the <rect> is changed, we
need to invalidate the SVGFilter display item. Without this patch, we
would re-use the (stale) cached SVGFilter display item.

With this change, we should be able to remove the filter caching
entirely in a followup and rely on the display list.

BUG=451606

Review URL: https://codereview.chromium.org/1148763002

git-svn-id: svn://svn.chromium.org/blink/trunk@195645 bbb929c8-8fbe-4397-9dbb-9b2b20218538