Split NTPLoginHandler across chrome://apps and chrome://history.
NTPLoginHandler was originally used to handle login from the NTP. It now
handles login on chrome://apps and chrome://history and employs a dirty
little hack to be re-used.

Undo that hack (exposing of ntp.updateLogin on chrome://history) and
split the handler logically, slimming the history handler
(which runs on mobile).

R=mad@chromium.org
BUG=329637

Review URL: https://codereview.chromium.org/1143963002

Cr-Commit-Position: refs/heads/master@{#330874}