Add ref-count on RenderViewHost for each new RenderFrameProxyHost.
Currently, only RenderFrameHosts are ref-counting the associated RenderViewHost.
Since it is possible to have a RenderViewHost with no associated RenderFrameHost
(for example a cross-process window opener/openee relationship), it should
be taken into account that the associated RenderFrameProxyHost is also
a ref-counting the RenderViewHost.
The goal of this CL is to allow for that, even though it isn't currently
the case. It is a small step towards removing swapped out RenderFrameHosts.

BUG=357747

Review URL: https://codereview.chromium.org/1150793002

Cr-Commit-Position: refs/heads/master@{#331016}