Replicate whether a frame is in a document or shadow tree.
This is needed for Blink to determine whether or not a frame can be
accessed via the window indexed getters (window[num]) and named
getters (window["name"]). Previously, Blink simply inspected the
tree scope of the frame's owner element and checked that it matched
the tree scope of the parent document. This breaks down in the case
of remote frames, since the document and owner element are not
available in the same process.

Since this information is determined at frame creation and cannot
change, Blink will rely on the embedder to forward this information
through frame replication. This CL is mostly straightforward
plumbing changes, but required updating lots of call sites and
function signatures.

BUG=473518
TBR=sky,tommycli,thestig,kalman,jrummell

Review URL: https://codereview.chromium.org/1141283002

Cr-Commit-Position: refs/heads/master@{#331059}