Fix Slimming Paint WebGL printing
Slimming Paint uses a recording canvas when printing, which triggered
an assert in GraphcisContext::drawBitmapRect which previously
did not trigger because we were not recording when printing.

This patch modifies the assert to check for printing, when we
are only recording temporarily before immediate playback into
a canvas on the main thread.

The patch also modifies DisplayItemListRecorderContext to copy
global properties fromt he primary context to the temporary
recording context.

R=danakj@chromium.org,chrishtr@chromium.org
BUG=471100

Review URL: https://codereview.chromium.org/1145843003

git-svn-id: svn://svn.chromium.org/blink/trunk@195818 bbb929c8-8fbe-4397-9dbb-9b2b20218538