Replace DCHECK in gl_helper_scaling.cc with logging.
The GLHelper interface doesn't have good ways to report errors to
callers (most methods return void), and the GLES2Interface it uses
doesn't have good support for detecting context loss.

It can't DCHECK about the results of OpenGL queries in the case that
the context is lost, so log if this happens. Examination of a few
users of GLHelper indicate that they already handle context loss
reasonably.

BUG=479975

Review URL: https://codereview.chromium.org/1152293003

Cr-Commit-Position: refs/heads/master@{#331213}