The CSSParser should parse pseudoTypes.
Instead of having a branch inside pseudoType() lets just do this inside the parser 
when constructing these objects. We now convert the pseudo type string into an
enum immediately and then use that value inside the parser as well to avoid
needing to compare strings or the number of colons.

BUG=490721

Review URL: https://codereview.chromium.org/1149853003

git-svn-id: svn://svn.chromium.org/blink/trunk@195824 bbb929c8-8fbe-4397-9dbb-9b2b20218538