Oilpan: Use a recursive mutex for ThreadState::threadAttachMutex
It is nasty to check isInGC in Heap::findPageFromAddress to determine
whether we should lock the mutex or not. It is better to use a recursive
mutex to make the lock re-entrant.

This mutex is not used in performance-sensitive places, so there would be
no performance concern.

BUG=475801

Review URL: https://codereview.chromium.org/1160593002

git-svn-id: svn://svn.chromium.org/blink/trunk@195884 bbb929c8-8fbe-4397-9dbb-9b2b20218538