Image elements can have inline layout objects
If it has a content: directive and no source then an image element
will get an inline renderer, so we can't assume that an image is always
a box. I'd resisted introducing this check as I felt that it wasn't possible
but clearly didn't experiment enough.

BUG=491582

Review URL: https://codereview.chromium.org/1146073003

git-svn-id: svn://svn.chromium.org/blink/trunk@195920 bbb929c8-8fbe-4397-9dbb-9b2b20218538