Refactoring: Introduce enum class ConvertBlocksToInlines
This CL introduces a new enum class ConvertBlocksToInlines replacing
current boolean values at createMarkup and so on.

This CL also moves the first parameter of StyledMarkupSerializer::
createMakup to the constructor because we are now trying to remove
all parameters from createMarkup as refactoring. Another CL for the
second parameter is crrev.com/1153473003

BUG=n/a
TEST=n/a; no behavior change

Review URL: https://codereview.chromium.org/1159753002

git-svn-id: svn://svn.chromium.org/blink/trunk@195966 bbb929c8-8fbe-4397-9dbb-9b2b20218538