Add early-out in SVGTextChunkBuilder::finalizeTransformMatrices()
This function will only do something when textLength is being applied.
If that's not the case, there'll just a lot of iteration and hash-
lookups for no good.
Not huge win performance-wise (in the noise), but a fairly simple and
obvious thing to do.

BUG=486669

Review URL: https://codereview.chromium.org/1161803002

git-svn-id: svn://svn.chromium.org/blink/trunk@195983 bbb929c8-8fbe-4397-9dbb-9b2b20218538