Fix HarfBuzzShaper::adjustSpacing buffer overrun when justify
crbug.com/444888 indicates that there are cases when HarfBuzzRun
is numGlyphs() == 0.

When adjustSpacing() adds a space before the first glyph of a run,
it adds the space to the advance of the last glyph of the previous run.
This logic will overrun if the previous run does not contain any glyphs.

This patch changes the logic to add the space to the offset of the first
glyph instead. I used the previousRun before because I could not make
the offset to work. Adding to the spacing as well when adding to the
offset works this time.

This patch also contains a fix where the offset is always added to X,
which must be added to Y if vertical flow.

BUG=486663
TEST=fast/text/justify-vertical.html

Review URL: https://codereview.chromium.org/1153653002

git-svn-id: svn://svn.chromium.org/blink/trunk@195987 bbb929c8-8fbe-4397-9dbb-9b2b20218538