[qcms] Add matrix transform flag and matrix extraction api
When a color transform is matrix-based (transforms using ICC V2 profiles
are commonly matrix-based), flag the color transform struct as matrix at
creation-time. Add an api to query if a color transform is matrix and an
api to extract the color transform matrix in row major order.

BUG=491826

Review URL: https://codereview.chromium.org/1155033003

Cr-Commit-Position: refs/heads/master@{#331615}