Move PlatformParentalControls Calculation to a Worker Thread
This change moves the PlatformParentalControlsValue calculation inside a
singleton to allow opportunistic calculation on a COM initialized worker
thread instead of the UI thread.

If the UI thread init is busy enough with other tasks (extensions and
profile loading) and another logical core is available, this should be able
to calculate without blocking the UI thread. If the UI thread runs really
quickly to getting the singleton, we would have needed to block anyways.

BUG=458388

Review URL: https://codereview.chromium.org/1143253004

Cr-Commit-Position: refs/heads/master@{#331653}