Drive: Limit the muximum number of child requests in a batch request.
Previously JobScheduler limits the total file size of batch request, but it does
not limit the total number of file. Even if the file size is 0 bytes, it takes
constant time to create a file entry on google drive. So it may take extreme
long time to process single batch request if we don't limit number of child
items.

BUG=491980
TEST=JobQueueTest.BatchRequstNumMaxJobs

Review URL: https://codereview.chromium.org/1159693002

Cr-Commit-Position: refs/heads/master@{#331720}