DevTools: speculative fix for crash in blink::InspectorDebuggerAgent::resolveBreakpoint
It looks like breakpointId is an empty string in some cases which causes crash when it is used as key in a HashMap. The crash call stack suggests that the empty string comes from the inspector state cookie where it was a key of JSON object. It is not clear to me how that may happen, probably deletion works incorrectly, I'm going to investigate this separately.

BUG=479096

Review URL: https://codereview.chromium.org/1151573007

git-svn-id: svn://svn.chromium.org/blink/trunk@196137 bbb929c8-8fbe-4397-9dbb-9b2b20218538