Prepare the infrastructure for password overrides.
Currently use the autofill field categories "USERNAME" and "USERNAME_AND_EMAIL" in our password form parsing heuristics. Use the same pipeline to pass the predictions of "PASSWORD" and "ACCOUNT_CREATION_PASSWORD" as well.

This is mostly refactoring. The new hints on password fields are not yet used in this CL.

BUG=492604

Review URL: https://codereview.chromium.org/1153023004

Cr-Commit-Position: refs/heads/master@{#331968}