[tracing] Expose getMemoryAllocatorDump to blink
This CL exposes a getMemoryAllocatorDump(name) to blink.
The code in chromium already had the API, now it is required in blink
for the BlinkGC dumper.
The plumbing code just keeps a reverse index (MAD -> WMAD) so that all
the logic about names remains in control of base/.
See crrev.com/1148683007 for the matching change in blink.

BUG=480500

Review URL: https://codereview.chromium.org/1163583003

Cr-Commit-Position: refs/heads/master@{#332194}