Remove --enable-ipv6 and --disable-ipv6 command line flags.
These were introduced for debugging and corner-case purposes. However, these were not permanently supported configurations - policy support was never added, which is a key indicator that these are options which will be supported going into the future.

--enable-ipv6 is hardly ever used. Internet searches for "--enable-ipv6 chrome" do not turn up many uses.

--disable-ipv6 is used more regularly though.

One place where --disable-ipv6 is used is in the SafeBrowsing tests. This was introduced in http://src.chromium.org/viewvc/chrome?revision=66196 to help accelerate test cycle times. The main issue was windows was resolving localhost to ::1 which made tests flaky.

The other general case where it has been used is when Chrome detects IPv6 availability but accesses are slow. Since this flag was added, Chrome added support for Happy Eyeballs, as well as improved the IPv6 probing for DNS, so the main motivations should be limited at this point.

BUG=344685

Review URL: https://codereview.chromium.org/1149303005

Cr-Commit-Position: refs/heads/master@{#332245}