Fix two NetworkQualityEstimator tests that were failing on Windows.
Current guess is that timer on Windows is low resolution
(~16 msec).
This can cause request.creation_time to be equal to
 last_connection_change_. This was causing following
2 tests to fail on Windows:
(1) NetworkQualityEstimatorTest.TestPeakKbpsFastestRTTUpdates
(2) URLRequestTestHTTP.NetworkQualityEstimator

If the guess is correct, these tests could potentially fail
on other platforms too.

Trying a fix which makes this check less stringent. Now,
 if request creation time is exactly equal to time of last
connection change, the request is still used to
estimate the network quality (unlike before, when the
former was required to be strictly larger than the latter).

BUG=492410

Review URL: https://codereview.chromium.org/1153023003

Cr-Commit-Position: refs/heads/master@{#332259}