Don't use field trials on the reference perf bots.
The current field trial config is used for an older build in the
case of reference. This is a problem because we are then subjecting
older builds to newer configurations that may crash.  To work
around this problem, don't add the field trials to reference builds.

Review URL: https://codereview.chromium.org/1162263005

Cr-Commit-Position: refs/heads/master@{#332318}