ServiceWorker: Make sure to write only an origin part in ServiceWorkerDatabase
An origin passed to ServiceWorkerDatabase should contain only an origin part,
but if it accidentally contains other parts like a path, they can also be
written in the database and cause database corruption.

This CL makes sure to write only an origin part in the database.

BUG=n/a
TEST=n/a

Review URL: https://codereview.chromium.org/1162543002

Cr-Commit-Position: refs/heads/master@{#332350}