Formalize a RenderFrameHost::AddMessageToConsole() API.
This CL adds a new API to RenderFrameHost allowing consumers to add a
new message to the DevTools console. There's currently one feature doing
this by manually sending an IPC message, but that doesn't scale well,
and I'm in the process of adding a second user.

BUG=

Review URL: https://codereview.chromium.org/1154673005

Cr-Commit-Position: refs/heads/master@{#332374}