Remove support for the "gcm_user_visible_only" manifest key.
Per the following Intent to blink-dev, support for this manifest key has
been deprecated per Chrome 44, and will be removed in Chrome 45. We'd like
to do this early in the release cycle, allowing developers to catch on
early when they run the Dev channel.

https://groups.google.com/a/chromium.org/d/topic/blink-dev/6OK5qm491Eg/discussion

A screenshot of the console warning can be seen here:
  http://peter.sh/files/devtools-deprecation-message.jpg

BUG=471534

Review URL: https://codereview.chromium.org/1158923002

Cr-Commit-Position: refs/heads/master@{#332386}