Made the CompositorAnimationObserver class usable
If you use CompositorAnimationObserver, you must check
compositor's shutdown. Otherwise there is a crash when
you unsubscribe because of compositor's delete.
There is workaround: to inheritance from
CompositorObserver and listen for OnCompositingShuttingDown.
But this is not graceful solution.
So I added OnCompositingShuttingDown to CompositorAnimationObserver.
I fixed potentional crash and wrote a test.

R=ajuma, danakj

Review URL: https://codereview.chromium.org/1147323003

Cr-Commit-Position: refs/heads/master@{#332480}