branch 2357 fork: cc: Allow synchronous compositor to PrepareTiles on each draw.
This is a cherry-pick of https://codereview.chromium.org/1163013002/
limited to OS_ANDROID only. The unit test is dropped.

Original description:

In OOM situations tiles which are not rasterizied by TileManager are
marked as rasterize on demand. This means that on the next draw we will
ask for PrepareTiles again. This can cause issues with the synchronous
compositor in certain situations where we receive two draws every vsync
and end up setting needs_prepare_tiles_ in each frame and asking for
vsyncs indefinitely.

To prevent this we allow the synchronous compositor to call PrepareTiles
on each draw so that needs_prepare_tiles_ is never true after a draw and
we don't unnecessarily ask for vsyncs.

BUG=495260
R=sunnyps@chromium.org

Review URL: https://codereview.chromium.org/1154363004

Cr-Commit-Position: refs/branch-heads/2357@{#454}
Cr-Branched-From: 59d4494849b405682265ed5d3f5164573b9a939b-refs/heads/master@{#323860}