Oilpan: We shouldn't schedule a GC before we complete sweeping
There is no reason to schedule a GC before we complete sweeping.
If we schedule a GC before completing sweeping, it is possible that
the next GC starts running before the unmarked objects get collected
(in fact this is frequently happening in memory.top_7_stress).

This CL moves scheduleGCIfNeeded() to after completeSweep().
One concern of this change would be that we might end up with delaying
scheduling a GC, but that cannot happen because completeSweep() is guaranteed
to run before we allocate any new page.

BUG=420515

Review URL: https://codereview.chromium.org/1158683003

git-svn-id: svn://svn.chromium.org/blink/trunk@196390 bbb929c8-8fbe-4397-9dbb-9b2b20218538