Fix crash in inspector-protocol/debugger/debugger-pause-dedicated-worker.html
Inspector uses v8 interrupts to execute devtools commands when inspected script is running (potentially for a long time). This is done by first posting command into the debug command queue and provisionally triggering the interrupt. However, inspector itself may be executing JS and in this case we will be running commands in a nested loop. This CL makes sure that no nested command dispatching is allowed on worker thread (but the case when the script is paused).

The downside of this change is that if a script evaluated in the console gets into tight loop we won't be able to pause in it.

BUG=495719

Review URL: https://codereview.chromium.org/1163923005

git-svn-id: svn://svn.chromium.org/blink/trunk@196412 bbb929c8-8fbe-4397-9dbb-9b2b20218538