[Contextual Search] Add metrics to track Taps before Open.
Adds a few more uma histogram metrics for Contextual Search.

1) We want to track how many promo peeks happen
before the user ever opens the panel.  We now track that value
in two histograms: one for users that have never opened, and
another for users that have opened (for archival purposes).

2) Another new metric is the total number of times the promo
was opened before the user decides (on the Opt-out).

3) Add the number of taps since any panel open.

4) We now track details of the search term resolution:
whether the base page was http, and whether the search term
was single-word or multi-word.

Also update the way our Flaky tests are annotated to make
the annotations more consistent and easier to maintain.

BUG=477827

Review URL: https://codereview.chromium.org/1162973004

Cr-Commit-Position: refs/heads/master@{#332640}