[Chromoting] Remove client_session from background.html
cl/1132133008 added jscompile checks for background.html and as part
of that removed client_session.js from the background.html GN build
because it was no longer needed (thanks to recent refactoring).

But the corresponding GYP build was not updated at that time. This
cl makes the change to the GYP build so that the GYP and GN builds
both produce identical results.

BUG=

Review URL: https://codereview.chromium.org/1152483005

Cr-Commit-Position: refs/heads/master@{#332691}