Support "live mode" in Mp4StreamParser.
According to the spec, when the duration is not specified, we treat the stream as a live stream. In Chromium, the media pipeline will switch to a low-delay mode for live streams. But the quality of implementation of low-delay mode is platform dependent.

Also with this change, we allow different streams in a playback to have different liveness. It's the audio/video decoder/renderer's responsibility to decide how to implement the live/low-delay mode.

BUG=420025
TEST=(1) Updated the unit tests to cover liveness detection. (2) Manually tested with the demo page attached in the bug (with a hack to force the duration to 0). Then I get zero frame delay when using software decoders.

Review URL: https://codereview.chromium.org/672623004

Cr-Commit-Position: refs/heads/master@{#332741}