Mac: Fix wrong sad tab background color after switching tabs.
When switching to another tab and back, the sad tab view gets removed from its
window and its layer gets recreated. We need to set the background color on the
recreated layer as well, not only the initial layer.
This was a regression caused by r331411.

BUG=495513

Review URL: https://codereview.chromium.org/1166843004

Cr-Commit-Position: refs/heads/master@{#332779}