Pass XImage ownership to SkBitmap in CopyAreaToCanvas
The SkBitmap's internal pixel ref is taken when the SkCavas is drawing
the image. SkCanvas deferrs the drawing operation. Since the XImage
would be destroyed when CopyAreaToCanvas()'s scope ends, the saved
pixel ref now points into deallocated memory.

This change pushes the ownership of the XImage into the SkBitmap such
that it is released when the SkCanvas releases the SkBitmap's pixel ref.

TEST=Ran ASAN build on x86-alex
BUG=492921

Review URL: https://codereview.chromium.org/1158233006

Cr-Commit-Position: refs/heads/master@{#332879}