Notify InputMethodBase on TextInputState change to send InputMethodObserver::OnTextInputStateChanged notification
InputMethodAuraLinux, which is derived from InputMethodBase, has implmented OnTextInputTypeChanged. And InputMethodAuraLinux is not propagating OntextInputTyepChanged to InputMethodBase, because of which InputMethodObserver's does not get the notification OnTextInputStateChanged.

The cl is to propagate the change event to InputMethodBase so that InputMethodObserver APIs are notified appropriately.

TBR=yukawa@chromium.org
BUG=None
TEST=None

Review URL: https://codereview.chromium.org/1160433006

Cr-Commit-Position: refs/heads/master@{#333008}