Oilpan: Introduce WeakProcessingVisitor
The main purpose of this CL is to remove a use of s_markingVisitor,
since I'm going to remove the global variable in a follow-up CL
(in preparation for implementing a visitor for a snapshot GC).

This CL also add a bunch of ASSERTs to verify that a visitor for
weak processing does not do things that the visitor is not allowed to do.

BUG=490087

Review URL: https://codereview.chromium.org/1168683002

git-svn-id: svn://svn.chromium.org/blink/trunk@196557 bbb929c8-8fbe-4397-9dbb-9b2b20218538