Use SequencedSocketData in HttpStreamParserTests.
SequencedSocketData is a bit cleaner and easier to use than the
alternatives, which we may want to get rid of, long term.

HttpStreamParser.AsyncChunkAndAsyncSocketWithMultipleChunks did rely on
DeterministicSocketData's behavior to get in a state where the previous
chunked upload body write was pending when data was appended to the
chunked upload data stream, but the updated test uses a new method to
ensure that happens.

BUG=492696

Review URL: https://codereview.chromium.org/1166693004

Cr-Commit-Position: refs/heads/master@{#333054}