[CSS Grid Layout] Support dots sequences in grid-template-areas
The CSS WG agreed to consider a sequence of "." characters a single null
cell token for grid-template-areas. That way authors can match the
number of dots to the width of the column.

For more information see:
https://lists.w3.org/Archives/Public/www-style/2015May/0175.html

Added new test cases to check the new behavior.

BUG=490088

Review URL: https://codereview.chromium.org/1160273004

git-svn-id: svn://svn.chromium.org/blink/trunk@196587 bbb929c8-8fbe-4397-9dbb-9b2b20218538