Allow CastContentBrowserClient to customise media pipeline device
We need a hook to customise how MediaPipelineDevice is created,
orthogonal to the vendor-specific aspect of creation.  Decided to
keep existing factory method for the vendor customisation so there's
no need to change vendor implementations for now (although this
will all get moved into libcast_media before long).  New hook is
in CastContentBrowserClient and can be substituted as the other
Platform* functions already are.

BUG=

Review URL: https://codereview.chromium.org/1168643004

Cr-Commit-Position: refs/heads/master@{#333090}