Update programmatic smooth scrolls on the RootFrameViewport.
Currently, the update occurs on the root FrameView so any programmatic
smooth scrolls on the RootFrameViewport wont get updated. This patch
causes the update to occur on the RootFrameViewport, which updates its
own animations, then calls the root FrameView's update method.

No test as smooth scrolls on RootFrameViewport aren't supported yet.
When that's enabled, any tests on that will implicitly test that
this works.

BUG=434497

Review URL: https://codereview.chromium.org/1148283011

git-svn-id: svn://svn.chromium.org/blink/trunk@196603 bbb929c8-8fbe-4397-9dbb-9b2b20218538