Add workerReady timing for ServiceWorker controlled requests [2/3]
To make ServiceWorker startup time measureable, in addition to
workerStart, we need another timing "workerReady" which is the point
ServiceWorker finished starting and is ready to handle events.

This timing is marked in DidPrepareFetchEvent, as well as send_start,
but unlike send_start, it's not cleared by request job restarts.

[1/3] https://codereview.chromium.org/1149273003/
[2/3] THIS
[3/3] https://codereview.chromium.org/1166953007/

BUG=465640
TEST=all existing tests should pass

Review URL: https://codereview.chromium.org/1146253004

Cr-Commit-Position: refs/heads/master@{#333240}