Merge into M44: Fixes the issue where overscroll animations were bleeding content over the browser window bounds.
The cause of the issue was that we were calling SetMasksToBounds(false) on the
web contents window's layer and its parent. The reason we were doing that was
that we wanted the screenshot to be captured for the entire page, not for
the visible part of it.

The fix is to get rid of SetMasksToBounds(false) and also to navigate only after
the overscroll animation finishes, not when it starts. This is unfortunate,
because it contributes to the perceived lag of the navigation, but this is
probably the best we can do for beta branch fix.

BUG=490817
TBR=sadrul

Review URL: https://codereview.chromium.org/1151723004

Cr-Commit-Position: refs/heads/master@{#331633}
(cherry picked from commit 43db64d9239b8965e2f9f741eebc3f0aa1616c76)

Review URL: https://codereview.chromium.org/1155493008

Cr-Commit-Position: refs/branch-heads/2403@{#235}
Cr-Branched-From: f54b8097a9c45ed4ad308133d49f05325d6c5070-refs/heads/master@{#330231}