Make power manager ignore first display configuration on Ozone
During the first display configuration the Ozone backend doesn't have
access to the GPU process to query for display resources. Thus it shows
only the internal display that Frecon provides the information for. This
causes the observer to notify powerd that there's only one display.
Since the display is closed, powerd suspends/shuts down the system.

On M44/ToT this issue is fixed by correctly scanning for the displays.
So, this change is only needed in M43 and once it is merged it should
be reverted from ToT.

BUG=478350

Review URL: https://codereview.chromium.org/1160143009

Cr-Commit-Position: refs/heads/master@{#333269}