Check for embedded PAC scripts when migrating proxy pref for DRP.
In M35 and earlier, the Data Reduction Proxy was enabled by setting the
proxy pref to an embedded PAC script. As of M41, Chrome would no longer
clear this embedded DRP PAC script, possibly causing the DRP to be used
when it shouldn't be if a client upgrades directly from M35- to M41+.

This change causes Chrome to clear the proxy pref if it represents an
embedded PAC script that contains the string ".googlezip.net:".

BUG=476610

Review URL: https://codereview.chromium.org/1148793008

Cr-Commit-Position: refs/heads/master@{#333319}