Flip the win8_chromium_ng recipe to 50% for testing the GN build.
We have flipped the "Win8 Aura" bot from GYP->GN.

We need to flip the matching CQ bot config over as well; the closest
match for this appears to be the "win8_chromium_ng" bot, which is
currently only in the CQ at 10%.

The "win8_chromium_rel" bot is in the CQ at 100%, but is running the
"chromium_trybot_legacy" recipe, which is doing something different.

As a part of the flip, we are effectively saying (for now) that we
don't care about the Win8 GYP build, and so it is okay to lose that
coverage.

Accordingly, this change drops the GYP bot from 100% of builds to
10%, and bumps the GN bot to 50%. If the GN bot seems to be green
at 50%, we'll move it ASAP to 100% to get full coverage.

R=sergeyberezin@chromium.org,
TBR=scottmg@chromium.org, phajdan.jr@chromium.org
BUG=497423
NOTRY=true

Review URL: https://codereview.chromium.org/1171823002

Cr-Commit-Position: refs/heads/master@{#333386}