Only use libv4l for video encoders
CrOS had a patch to return -1 from v4l2_fd_open when the plugin does
not exist. It broke the factory camera test because OpenCV used libv4l
to open the camera. For now, we'll only use libv4l for encoders so we
can revert the patch in CrOS and still avoid the libv4l overhead for
decoders. In the future, we either need to change libv4l in a
compatible way or not to use libv4l at all.

BUG=chrome-os-partner:40353
TEST=Run apprtc loopback on veyron jerry.

Review URL: https://codereview.chromium.org/1164213003

Cr-Commit-Position: refs/heads/master@{#333426}