Teach GPU command buffer whether a context is webgl.
Then we initialize it differently. For example, if it's WebGL 1, we disallow
NPOT support. Also, we initialize shader translator correctly once (before
we have to initialize it to the default ES spec and then reset a new one with
WebGL spec), so this is a speedup at WebGL context creation time.

Also, we separate the idea of whether a context is ES2/ES3 or WebGL1/WebGL2.

This is necessary because before we assume ES3==WebGL2, which isn't true, say,
we may always want to use ES3, for WebGL 2 and for compositor, etc.

BUG=497464
TEST=gpu_unittests, webgl_conformance, gl_tests
R=kbr@chromium.org,piman@chromium.org

Review URL: https://codereview.chromium.org/1143393007

Cr-Commit-Position: refs/heads/master@{#333439}