Ensure the BrowserPluginGuest is fully loaded before running PDF tests
For running some of the PDF tests we need the BrowserPluginGuest to be
loaded so that we can execute script in it. It should be good enough to
just navigate to the URL. But loading up the BrowserPluginGuest seems to
happen asynchronously as there was flakiness being seen due to the
BrowserPlugin guest not being available yet (see crbug.com/498077). So
instead use |LoadPdf| which ensures that the PDF is loaded before continuing.

BUG=498077

Review URL: https://codereview.chromium.org/1173553002

Cr-Commit-Position: refs/heads/master@{#333453}