Refactor FrameLoader loading interface
This CLs removes the reload and loadHistoryItem functions of FrameLoader in
favor of a unified load function. It also updates the public interface of
WebFrame to generate WebURLRequests for history navigations and reloads. The
end goal is to be able to only have a single function for loading in WebFrame
as well.

BUG=490713

Review URL: https://codereview.chromium.org/1156473002

git-svn-id: svn://svn.chromium.org/blink/trunk@196753 bbb929c8-8fbe-4397-9dbb-9b2b20218538