Oilpan: Remove Heap::s_lastGCWasConservative
I don't think we need the flag. This flag is used to determine
whether we should run a transitive closure after doing the conservative scanning.
However, we won't need the flag to make the decision.
If the marking stack is not empty, we need to run the transitive closure. Otherwise, we don't.

Thus this CL removes Heap::s_lastGCWasConservative.

BUG=420515

Review URL: https://codereview.chromium.org/1159143006

git-svn-id: svn://svn.chromium.org/blink/trunk@196760 bbb929c8-8fbe-4397-9dbb-9b2b20218538