Merging this to M44. Clear touch state in HWNDMessageHandler if the underlying child window on which the touch started changes.
This happens for e.g. in the LegacyRenderWidgetHostHWND window where in we can initiate a touch down context and in that context
create a New Tab which creates a new LegacyRenderWidgetHostHWND instance. Releasing the touch on the new window does not result in the new window
getting the touch up notification. The parent window HWNDMessageHandler is still in the state of expecting a touch up which causes it to ignore any mouse
input until that is received. In this case it is never received.

Fix is to notify HWNDMessageHandler via a WindowEventTarget function HandleParentChanged which is invoked by the LegacyRenderWidgetHostHWND class when its
parent changes. In the implementation in HWNDMessageHandler we clear the touch ids.

BUG=447153

Review URL: https://codereview.chromium.org/1151523011

TBR=sky

Cr-Commit-Position: refs/heads/master@{#332658}
(cherry picked from commit 6bdd5d57f9583c324808592138556896d3aca7d1)

Review URL: https://codereview.chromium.org/1171743003.

Cr-Commit-Position: refs/branch-heads/2403@{#260}
Cr-Branched-From: f54b8097a9c45ed4ad308133d49f05325d6c5070-refs/heads/master@{#330231}