Rename isNormalFlowOnly to something meaningful
The boolean was confusing and needed a lot of explanation for
new people. The idea behind it is really that we "treat some
elements as stacking context" for the purpose of painting and
hit testing, even if they are not (e.g. positioned elements).
This is a requirement by CSS 2.1, appendix E.

Review URL: https://codereview.chromium.org/1176503002

git-svn-id: svn://svn.chromium.org/blink/trunk@196798 bbb929c8-8fbe-4397-9dbb-9b2b20218538