Remove flaky test "synchronous-create-pattern.html"
This test uses setTimeout and is so flaky that it doesn't provide any
useful coverage. We should rewrite this test to avoid setTimeout and
to use CSS animations in 35ms-green-flash-at-35ms.svg instead of
SVG animations. For now I'm removing this test as it is causing
failures, but we should re-add it.

TBR=ojan
BUG=316730

Review URL: https://codereview.chromium.org/1175773002

git-svn-id: svn://svn.chromium.org/blink/trunk@196816 bbb929c8-8fbe-4397-9dbb-9b2b20218538