Add a scoping object to help block scripting during plugin destruction.
This doesn't block all scripting in plugin destruction: apparently,
being able to script the DOM in an inconsistent state is a "feature" of
Flash.

Instead, the blocks are scoped to instances where the Frame/Document
will be destroyed or replaced shortly afterwards: this is an extension
of the earlier efforts to block synchronous scripting by plugins in
frame detach. The eventual goal is to remove the need to run a nested
message loop in these circumstances, so that the frame navigation and
detach logic can be rewritten to be much easier to reason about.

BUG=498088

Review URL: https://codereview.chromium.org/1170413003

git-svn-id: svn://svn.chromium.org/blink/trunk@196832 bbb929c8-8fbe-4397-9dbb-9b2b20218538