Make sure multi chracter codepoints are deleted correctly
deleteSurroundingText() only deletes one character even for multi-character
codepoint.

On the blink side, we have InputMethodController::extendSelectionAndDelete()
to make sure selection and deletion respect Unicode boundaries. However,
AdapterInputConnection keeps track of selection region separately, and this
value is incorrectly updated.

On top of adding a new test for this case, it extends
waitAndVerifyEditorCallback to also check the outbound calls to
InputMethodManager.

The above extension found that testEnterKeyEventWhileComposingText fails
because there is hidden discrepancy between blink implementation and
what we report to InputMethodManager. So I've added a TODO for that.

BUG=497091

Review URL: https://codereview.chromium.org/1165793007

Cr-Commit-Position: refs/heads/master@{#333704}