DevTools: speculative fix of crash in PageDebuggerAgent::restore
It may happen that we navigate to a page where script execution is blocked. Debugger should not be anabled for such pages. This CL makes sure that we won't be restoring debugger state if we know for sure that the script execution is blocked.

restore() is called very early before the load starts hence at that point we don't know if the new page is blacklisted. This use case is not covered by this change and I'm hesitent complicating debugger code for such rare scenario.

BUG=498963

Review URL: https://codereview.chromium.org/1180713005

git-svn-id: svn://svn.chromium.org/blink/trunk@196937 bbb929c8-8fbe-4397-9dbb-9b2b20218538