[Telemetry] Modify value stringification to use display_name instead of name
This change extracts the value stringification tweaks from crrev.com/1075633002
to improve safety. In particular, not all pages have names, so display_name is
what we want to use instead for now, and as well, when there is no page, we
want to use the string 'None' since Python's None stringifies to ''.

BUG=461101
R=nednguyen

Review URL: https://codereview.chromium.org/1181713003

Cr-Commit-Position: refs/heads/master@{#334009}