Fixes crash in ViewManager during shutdown
The crash happens because the CommandBufferImpl can be destroyed on
another thread. This is because the CommandBufferImpl uses a
strongbinding that is created on another thread. This means if there
is a connection error the CommandBufferImpl is destroyed on the
background thread. CommandBufferImpl has references to a bunch of
things that need to be destroyed on the main thread.

BUG=none
TEST=none
R=jam@chromium.org

Review URL: https://codereview.chromium.org/1179933003

Cr-Commit-Position: refs/heads/master@{#334082}