Oilpan: Fix compile of GC_PROFILING build, and remove object graph dumper.
This patch removes the ability to dump object graphs, since it seems like
nobody is using it. The code does not compile anyway, and we can pull it back
whenever it becomes necessary in the future.

This patch also fixes a few places breaking compile of GC_PROFILING builds.

BUG=none
R=haraken@chromium.org, oilpan-reviews@chromium.org

Review URL: https://codereview.chromium.org/1179763005

git-svn-id: svn://svn.chromium.org/blink/trunk@196996 bbb929c8-8fbe-4397-9dbb-9b2b20218538