Sanitize SVG animation attributes which could set JavaScript URL values.
When pasting we remove the hrefs from links with javascript: URLs
(among other things) because the scripts may be malicious. But SVG
animations can update hrefs indirectly, and we were not scrutinizing
those. This extends the sanitization to remove attributes from SVG
animations that contain JavaScript URLs because the animation could be
directed to set a href.

BUG=452059
TEST=webkit_unit_tests UnsafeSVGAttributeSanitizationTest.*

Review URL: https://codereview.chromium.org/1171223004

git-svn-id: svn://svn.chromium.org/blink/trunk@197006 bbb929c8-8fbe-4397-9dbb-9b2b20218538