Move copyright scanner script to tools/
The script is used project-wide in Chromium. It used to be in android_webview/
because it was created as WebView-specific tool, but the kind of check it does
is not actually specific to WebView. And for a long time already it also runs
as part of global presubmit.

BUG=486158

Review URL: https://codereview.chromium.org/1167093004

Cr-Commit-Position: refs/heads/master@{#334196}