Fix possible DCHECKs during shutdown
I was occasionally hitting a DCHECK in destroying the WeakPtrFactory
owned by CommandBufferImpl. The DCHECK is because the WeakPtrFactory
was being destroyed on a thread other than the one where a reference
is still outstanding. The problem appears to be
CommandBufferDriverClientImpl can post tasks to CommandBufferImpl on
the background thread, which uses the WeakPtrFactory. As far as I can
tell there is no reason for these tasks (specifically the functions
UpdateVSyncParameters and DidLoseContext) to run on the background
thread. UpdateVSyncParameters is calling out to a
ViewportParameterListenerPtr, which should happen on the main thread,
not the background thread. Similarly DidLoseContext is going to
schedule destruction of the COmmandBufferImpl, which need not run on
the background thread either. I've nuked the WeakPtrFactory entirely.

BUG=none
TEST=none
R=jam@chromium.org

Review URL: https://codereview.chromium.org/1183673003

Cr-Commit-Position: refs/heads/master@{#334449}