cc: Remove maximum_number_of_failed_draws_before_draw_is_forced_ setting
This is never set by the embedder, it's only set in scheduler unit
tests, so we only need to have this available on the SchedulerSettings,
not the LayerTreeSettings.

Also remove the trailing underscore from the name.

R=brianderson
CQ_INCLUDE_TRYBOTS=tryserver.blink:linux_blink_rel

Review URL: https://codereview.chromium.org/1188613002

Cr-Commit-Position: refs/heads/master@{#334471}