Use replacedContentBox for LayoutHTMLCanvas when deciding PaintLayer's contentBox.
If a canvas size goes over the minimunAccelerated2dCanvasSize we set in Settings.in,
the canvas image is rendered in a paint layer. The problem is here the paint layer's
geomery is incorrect since a wrong content bounding size is used for the canvas.
LayoutHTMLCanvas is a kind of LayoutReplaced objects. Its geometry could be changed
after replacing an original content with an actual one or updating styles like object-fit.
Using LayoutBox::contentBox() for LayoutHTMLCanvas is such a carelessness. We can fix this
by using replacedContentBox() rather than using just contentBox() for a canvas.

BUG=467409

Review URL: https://codereview.chromium.org/1147363009

git-svn-id: svn://svn.chromium.org/blink/trunk@197160 bbb929c8-8fbe-4397-9dbb-9b2b20218538