Explicitly whitelist 'blob:' and 'filesystem:' in extensions' default CSP.
'blob:' and 'filesystem:' should not match 'self' in CSP source
expressions, but they currently do. In order to avoid breakage, this
patch whitelists them explicitly (which is a no-op at the moment) so that
we can change Blink's behavior without breaking extensions.

Perhaps we can re-evaluate this in v3. :)

BUG=473904

Review URL: https://codereview.chromium.org/1184353002

Cr-Commit-Position: refs/heads/master@{#334599}