Introduce CompositeDataConsumerHandle.
This CL introduces CompositeDataConsumerHandle, a utility class to composite
multiple WebDataConsumerHandle objects. This class is planned to be used in
the following use case, for example.

Imagine we want to provide a WebDataConsumerHandle representing a response body
gotten from a ThreadableLoader, but we need to create the handle BEFORE the
response body arrives. In such a case, we can create a
CompositeDataConsumerHandle initialized with a handle gotten from
CompositeDataConsumerHandle::createWaitingHandle. When the response body
arrives, calling CompositeDataConsumerHandle::update will update the handle so
that it represents the response body.

BUG=480746

Review URL: https://codereview.chromium.org/1162043007

git-svn-id: svn://svn.chromium.org/blink/trunk@197174 bbb929c8-8fbe-4397-9dbb-9b2b20218538