Change connectivity checks to be a task
Currently, the connectivity checks have been using a Future to
represent the data. Soon it is necessary to support callbacks
and other operations for this, so the usage of Future does not
make sense going forward.

This CL removes the dependency on the Future class, and instead
renames the collector to be a task.

As part of this, the methods that previously existed on the Future
could be moved to the outer class.

BUG=386395

Review URL: https://codereview.chromium.org/1190513004

Cr-Commit-Position: refs/heads/master@{#334611}