Remove --block-cross-site-documents (for now). We'll need to reimplement this browser-side.
Rename SiteIsolationPolicy::ShouldBlockResponse to ::OnReceivedFirstPacket, and
remove the logic that supported actually blocking requests in the renderer.

Rewrite SiteIsolationPolicyBrowserTest so that it's a direct test of the
histogram logging behavior, rather than of the response-blocking behavior. The
test does substantially the same operations, but iterating over the test docs
now has to be driven from C++ instead of from js, so it looks pretty different.

In a subsequent CL I'll rename these classes and split the classification logic
apart from the histogram logging logic.

BUG=481066, 419087, 436250

Review URL: https://codereview.chromium.org/1178443002

Cr-Commit-Position: refs/heads/master@{#334709}