Remove fixed header at chrome://inspect
The fixed header gets in the way when the user zooms in and attempts to
scroll the page. To fix this issue, the header is no longer at a
viewport-fixed position, but at a static position within the main
content. The location specified by the page header is already displayed
in the left navigation bar, so this change does not hurt usability.

BUG=492124

Review URL: https://codereview.chromium.org/1187813004

Cr-Commit-Position: refs/heads/master@{#334722}