Always add gnu.build.id to shared libraries
Breakpad and the crash server use ids to identify shared libraries.
These ids are calculated like so:

1. If the .so has gnu.build.id use that.
2. Else, use a hash of the first page of the .so.

When (2) is used, id collisions are rather common (particularly with
Android's use of an orderfile). This change makes the linker set
gnu.build.id to a hash of the output for every shared library.

BUG=498311

Review URL: https://codereview.chromium.org/1190833002

Cr-Commit-Position: refs/heads/master@{#334731}