Don't send JS->ObjC messages while doing URL verification
Because message handling with UIWebView involves a URL check, there's a
potential for infinite loops if any page overrides XHR and in that
override does something that causes a message to be sent (since that
will in turn trigger a URL check, which involves XHR, which triggers a
message, etc.).

This case has always existed, but the recently-added console.log interception
makes it much more likely.

BUG=500547
TEST=Open page referenced in bug; it should not hang or crash

Review URL: https://codereview.chromium.org/1186013005

Cr-Commit-Position: refs/heads/master@{#334899}