Parsing expectations is n^2 in the number of tests per line.
When merging expectations in different expectations files, we
do an n^2 operation. For each line/test combination, we merge
them. Instead, this algorithm only needs to merge a given
pair of lines once.

Add a cache of merged lines so that we reuse merged lines
when we've already seen a given pair.

This hits our current code base because we have ~25k
slimmingpaint tests that are skipped in TestExpectations.
Then running with "-i virtual" on the commandline creates
a fake expectations model to merge with that. On my z620,
this patch takes the time to merge the models from >4 minutes
to ~2 seconds.

Review URL: https://codereview.chromium.org/1182253003

git-svn-id: svn://svn.chromium.org/blink/trunk@197278 bbb929c8-8fbe-4397-9dbb-9b2b20218538