Use new Skia header file in test.
The content of the Skia header SkTypeface_android.h has moved into
SkFontMgr_android.h since the old name was somewhat misleading, as
it affected the font management, not the typefaces. The old header
is now deprecated and will be removed soon, so move to the new one.

Review URL: https://codereview.chromium.org/1181083005

Cr-Commit-Position: refs/heads/master@{#335037}