DevTools: Align flamechart ruler with overview
When the entire recording is selected make the flamechart
rulers match the overview ones. Some contents may be hidden
behind vertical scroller thought, but that's not a problem
as it is possible to zoom in.

Note that this schema is currently used for MacOS.

Drive-by: Fixed on-hover selection misalign on flamechart.

Review URL: https://codereview.chromium.org/1193593002

git-svn-id: svn://svn.chromium.org/blink/trunk@197397 bbb929c8-8fbe-4397-9dbb-9b2b20218538