Change cast status tray icon to mirror the screenshare icon.
This addresses some privacy concerns; using the cast icon implies that
whenever a cast is present the icon will show up; this is not the case
with DIAL-based casts.

It is currently possible for both screenshare and casting to be active
at the same time. See crbug.com/501554.

BUG=498208

Review URL: https://codereview.chromium.org/1187703003

Cr-Commit-Position: refs/heads/master@{#335195}