Fix an issue that OMA download doesn't start after download descriptor(dd) is downloaded
After a dd is downloaded, we launch a parser in the background to parse the descritor.
However, on the UI thread, we are also removing the dd entry.
These 2 operations will compete with each other.
If the dd is removed first, the parser will not be able to parse the file.
This change moved the remove operation after the parser has finished parsing the dd.
This CL also stop showing the NextURI dialog to user if user chooses to cancel the download.

BUG=414166

Review URL: https://codereview.chromium.org/1183143013

Cr-Commit-Position: refs/heads/master@{#335200}