Take 2: Moved logic for mapping child process FDs for ICU and V8 into child_process_launcher.cc
Used to be defined in each app's ContentBrowserClient, but since
content/ is the one that receives the FDs, it makes sense that it should
be the one to send them.

This also removes ChildProcessLauncher::AppendMappedFileCommandLineSwitches
as it is no longer needed.

Changes MemoryMappedFile::Region to be a POD so that it doesn't create require static initializers.
BUG=394502

Review URL: https://codereview.chromium.org/1185973003

Cr-Commit-Position: refs/heads/master@{#335207}