New SELECT Popup: Reduce the size of JSON data.
This CL reduces the size of JSON data which is passed from the origin document
to the popup. For example, we pass 'display' CSS property of an OPTION element
only if display:none.

This CL reduces time to open 43k OPTION popup by 0.5sec on my machine.

BUG=500401

Review URL: https://codereview.chromium.org/1193813002

git-svn-id: svn://svn.chromium.org/blink/trunk@197452 bbb929c8-8fbe-4397-9dbb-9b2b20218538