Use URL without fragment when query recent redirect list for favicons.
The hash fragment is often set with Javascript when a page loads, which causes the favicon mechanism to use the URL with fragment to query the recent redirects. However, the recent redirects are typically keyed to the URL without the hash.

We need to keep to lookup for the URL with an hash, though, in case the redirect is keyed to it, which happens if the user navigated to a URL with a hash fragment but was redirected.

BUG=498618

Review URL: https://codereview.chromium.org/1180403005

Cr-Commit-Position: refs/heads/master@{#335268}