cc, gpu: Use RGBA when using msaa on systems that don't support BGRA
The pixel is compiled to prefer BGRA, but doesn't support BGRA for
render buffers.  When we try to use msaa, it tries to create a BGRA
render buffer and fails.

Fix is, plumb the setting to the gpu raster code, and prefer RGBA
when using msaa on a system that doesn't support BGRA.

BUG=497478
CQ_INCLUDE_TRYBOTS=tryserver.blink:linux_blink_rel

Review URL: https://codereview.chromium.org/1174733003

Cr-Commit-Position: refs/heads/master@{#335395}