Move all clip display item parameters to the constructor
This patch changes ClipDisplayItem to take all rounded rects in through
the constructor. Because rounded rects are rarer than regular clips,
the rounded rects are now heap allocated. If this becomes a problem
in the future we may want to factor out a RoundedClipsDisplayItem.

No new tests as there should be no change in behavior.

BUG=484943

Review URL: https://codereview.chromium.org/1192363005

git-svn-id: svn://svn.chromium.org/blink/trunk@197515 bbb929c8-8fbe-4397-9dbb-9b2b20218538