Import cc/base/list_container.h
This patch temporarily adds ListContainer to platform/graphics to
optimize slimming paint's display list allocations. ListContainer is a
container type that handles allocating contiguous memory for new
elements and traversing through them with iterators.

This class has been commented to discourage its use. The location of
platform/graphics is also intended to minimize use. This class will
be removed once the repository merge stabalizes. The iterators
in this class are used in hot loops and calling through to cc was
found to be a performance issue:
https://groups.google.com/a/chromium.org/d/msg/paint-dev/YbGwc_1ncfY/ScLys7KiRloJ

Other than style, this class has been modified as little as possible.
There should be no functional differences, and the tests pass as expected.

BUG=484943

Review URL: https://codereview.chromium.org/1195303002

git-svn-id: svn://svn.chromium.org/blink/trunk@197599 bbb929c8-8fbe-4397-9dbb-9b2b20218538