Fix crash due to flash re-entrancy during instnace creation/scripting
Plugin creation causes some sync IPCs to be sent which can result in
the plugin executing script, which is also synchronous. If this
script interacts with the plugin that is still in creation it will
crash.

This CL adds a check to fail script execution in this case.

BUG=503401

Review URL: https://codereview.chromium.org/1202653004

Cr-Commit-Position: refs/heads/master@{#335829}