Fix IsPrecachingAllowed() to check Sync instead of Data Saver.
Fixed PrecacheManager::IsPrecachingAllowed() to check for the ability to sync
tabs rather than to check that Data Saver is enabled. This was used only by
ChromeNetworkDelegate to determine whether to report some request metrics to
UMA. This definition of the function became out-of-date as of commit dd40ab5.

Got rid of the internal IsSyncTabsEnabled() function, and had PrecacheLauncher
call PrecacheManager::IsPrecachingAllowed() instead. This should eliminate this
class of error in the future.

BUG=309216

Review URL: https://codereview.chromium.org/1176253002

Cr-Commit-Position: refs/heads/master@{#335835}