Open USB devices through the permission broker's OpenPath method.
The permission broker now supports opening devices on behalf of Chrome.
This change starts doing so for USB devices. This will allow the
permission broker to apply restrictions to the file descriptor before it
is passed to Chrome.

FakePermissionBrokerClient is given a simple implementation of OpenPath
for testing purposes.

BUG=496469

Review URL: https://codereview.chromium.org/1172533002

Cr-Commit-Position: refs/heads/master@{#335876}