Service Worker: Update stale workers
Before this patch we only update a worker on main resource navigation
inside its scope. This means a Push-only SW for which you don't navigate
to its scope is never updated.

This patch introduces a way to update "stale" workers.  A worker that
hasn't been checked for update from the network for over 24 hours is
considered stale. We check for staleness when a worker starts up and
periodically while it's running.  We try to update stale workers once
they stop (one reason is that skipWaiting() may kick out the worker
before it finished handling the event that triggered the staleness
check; also want to avoid network contention while the incumbent is
handling an event).  However if a worker is running too long, an update
may be triggered (otherwise, a constant stream of events keeping the
worker alive would prevent it from ever updating).

BUG=477598
TEST=ServiceWorkerVersionTest.StaleUpdate*

Review URL: https://codereview.chromium.org/1187623006

Cr-Commit-Position: refs/heads/master@{#335901}