Keep effectiveZoom finite
Through repeated application, it's quite possible to get the effective
zoom to overflow, resulting in Infinity in ComputedStyle.
This is a very uncommon case in reality (outside fuzzers), but make
sure to clamp the effective zoom value to a finite range to avoid the
simplest cases of Infinity havoc-wrecking.

BUG=490757,502997

Review URL: https://codereview.chromium.org/1193203003

git-svn-id: svn://svn.chromium.org/blink/trunk@197734 bbb929c8-8fbe-4397-9dbb-9b2b20218538