Remove LayoutTests/plugins/destroy-plugin-from-callback.html
Remove test that claims to pass unless it crashes yet times out or fail
with a different output more often than not. Given that it was added to
catch a regression in the webkit plugin architecture that isn't used by
Blink anymore the test is rather pointless at this ponint.

R=adamk@chromium.org
BUG=503848
NOTRY=true

Review URL: https://codereview.chromium.org/1207793002

git-svn-id: svn://svn.chromium.org/blink/trunk@197764 bbb929c8-8fbe-4397-9dbb-9b2b20218538