Should not crash when background image a display:none presents on the root
Since my previous background refactoring CL, BoxPainter::
calculateBackgroundImageGeometry() assumes that when it is called on
the LayoutView, the root element must have a box. However the background
obscuration code in LayoutBox::computeBackgroundIsKnownToBeObscured() was
not updated at the same time. This CL corrects the early out condition
to make sure calculateBackgroundImageGeometry() is only called with a
valid box.

BUG=501560

Review URL: https://codereview.chromium.org/1200263004

git-svn-id: svn://svn.chromium.org/blink/trunk@197789 bbb929c8-8fbe-4397-9dbb-9b2b20218538