DCHECK that hostname is not empty in MarkAlternativeServiceBroken.
The concept of brokenness for alternative services has been decoupled from the
origin to alternative service mapping.  MarkAlternativeServiceBroken marks an
actual alternative service broken by itself, not in relation to any particular
origin.  Therefore callers are expected to pass an alternative service with a
non-empty hostname.  (Empty hostnames are only used in the mapping to signify
"default host", which is relevant when alternative services of a canonical host
is used for other origins.)  This CL adds a DCHECK to document that.

BUG=392575

Review URL: https://codereview.chromium.org/1203233002

Cr-Commit-Position: refs/heads/master@{#336123}