Remove diagnostic code from BlockPainter.
There have been no crashes since the diagnostic code was added,
suggesting that the minor change in behavior that the code
introduced was sufficient to avoid the crash. Either that or
just having more code there avoided the crash, which would be
odd.

This change keeps the new behavior while removing the
diagnostics.

R=chrishtr
BUG=475698

Review URL: https://codereview.chromium.org/1205143003

git-svn-id: svn://svn.chromium.org/blink/trunk@197857 bbb929c8-8fbe-4397-9dbb-9b2b20218538