Compute the normal flow list on the fly
Storing the normal flow list in DeprecatedPaintLayerStackingNode
is not necessary and there are quite some complexity in the code
for invalidating this cache.
Ideally we shouldn't need the normal flow at all. It's not a CSS
concept but arose from our implementation of stacking contexts
and the multiple responsabilities of DeprecatedPaintLayer.

Review URL: https://codereview.chromium.org/1188363002

git-svn-id: svn://svn.chromium.org/blink/trunk@197871 bbb929c8-8fbe-4397-9dbb-9b2b20218538