Remove ChildAccountService.waitUntilFinished().
If the ground truth value of the child account state is not available yet, ChildAccountService will return a cached value from Preferences.

In practice, this should only become a problem if Chrome is opened straight to an activity that depends on the child account state (e.g. settings) and the state has been changed since the last time Chrome ran. In that case, Chrome will temporarily show the UI for regular users.

BUG=none

Review URL: https://codereview.chromium.org/1206503004

Cr-Commit-Position: refs/heads/master@{#336358}