Pass explicitly the frame to HistoryController::GoToEntry
When doing a session history navigation cross-process in a frame which
already has a proxy, the proxy is replaced by a new real frame. In this
case, HistoryController doesn't know about the new frame it needs to
interact with, so this CL adds it as an explicit parameter in GoToEntry.

This scenario becomes possible once we stop using swapped out frames
for navigation in regular Chrome.

BUG=357747

Review URL: https://codereview.chromium.org/1212363002

Cr-Commit-Position: refs/heads/master@{#336370}