PlzNavigate: add stream flush calls to speed up navigation.
When reading data from the network StreamWriter was not continuously
Flush()-ing the internal stream so the renderer was only being notified the
data was available when the whole body was fully read and the stream was
closed. This change adds an "immediate mode" to StreamWriter that makes it
Flush the internal stream whenever new data is available and makes PlzNavigate
navigations commit provisional load much faster.

Review URL: https://codereview.chromium.org/1205073002

Cr-Commit-Position: refs/heads/master@{#336552}