Split CookieSettings unit tests
This CL splits some of the tests in cookie_settings_unittest.cc in an
other file: cookie_settings_factory_unittest.cc.
These tests are related to the CookieSettings factory and test how
CookieSettings are managed in incognito. These tests are conceptually
Chrome level, and will not be moved to the component.

The remaining tests in cookie_settings_unittest.cc are cleaned from all
//chrome and //content dependencies in preparation for componentization.

The CookieSettings factory will be split to its own file in a follow up
change.

BUG=503973

Review URL: https://codereview.chromium.org/1216043002

Cr-Commit-Position: refs/heads/master@{#336566}