cc: Don't use page scale from CDP in PictureLayer
Rather than using DrawProperties to get the device and page scale,
instead get it from the LayerTree(Host)Impl.  The biggest change here is
that during the recursive PushPropertiesTo walk, LayerImpl figures out
whether it's under the page scale layer or not, and this flag is used
to figure out whether to apply page scale in PictureLayerImpl.

In the future, this same flag calculation will need to be done by any
layerization algorithm, so it seems forward looking.

The followups to this are to remove these flags from DrawProperties and
to make PictureLayerImpl calculate the contents scale itself.

R=danakj@chromium.org,vmpstr@chromium.org,vollick@chromium.org
CQ_INCLUDE_TRYBOTS=tryserver.blink:linux_blink_rel

Review URL: https://codereview.chromium.org/1214803003

Cr-Commit-Position: refs/heads/master@{#336697}