Reland: Optimizes IndexedDB's DeleteDatabase when no db exists
When deleting the non existing database in a origin with no existing
backing store, then in DeleteDatabse(), early exiting before
OpenBackingStore, so that it wont results in multiple transaction as
earlier.

Original patch by: payal.pandey

This was reverted (see crrev.com/1071553002) when it appeared to make
a test flaky, but the test continues to flake (now tracked as
crbug.com/505609) so let's try relanding.

BUG=460236

Review URL: https://codereview.chromium.org/1216573008

Cr-Commit-Position: refs/heads/master@{#336796}