Expose a SequencedTaskRunner that runs tasks on raster threads.
This is the first step to make the raster worker threads reusable
within the render.

The raster threads and the task graph runner have been moved in
a RasterWorkerPool class that implements the SequencedTaskRunner
interface.

The long term plan is to have a better defined TaskGraphRunner
interface and expose that one instead.

BUG=

Review URL: https://codereview.chromium.org/1211153002

Cr-Commit-Position: refs/heads/master@{#336862}