ServiceWorker: Check liveness of ServiceWorkerDatabase before deleting an old diskcache
After diskcache migration, ServiceWorkerStorage posts a delayed task to delete
an old diskcache and to record the result in ServiceWorkerDatabase. This task
accesses the database without any liveness check, so the task possibly crashes
if the database is closed before that.

This CL makes the task to confirm the liveness before accessing the database.

BUG=487482,505333

Review URL: https://codereview.chromium.org/1210073009

Cr-Commit-Position: refs/heads/master@{#336951}