Fix PermissionBubbleManager race conditions when merging requests on load.
The PermissionBubbleManager merges request that happened before the load
event when the load event happens. This was actually broken because the
boolean used was never reset to 0, even after the main frame was navigated.

Note that it means that this feature was never really working so we might
end up with permission prompts being delayed a lot because the load event
takes too long to happen.

BUG=434104

Review URL: https://codereview.chromium.org/1187923005

Cr-Commit-Position: refs/heads/master@{#336954}