PlzNavigate: reset the NavigationRequest in RFH::Commit/FailNavigation
This CL changes the moment the NavigationRequest is reset from
Navigator::DidNavigate to RenderFrameHost::CommitNavigation or
RenderFrameHost::FailedNavigation. This makes for an easier cancelation logic
and fixes issues in WebNavigationAPI tests where a navigation to about:blank is
shortly followed by a navigation to a real url.

BUG=504347

Review URL: https://codereview.chromium.org/1215763002

Cr-Commit-Position: refs/heads/master@{#337047}