Fixes bug in html_viewer that could lead to dropping interface requests
Connections other than the first to html viewer asynchronously create
HTMLDocument. HTMLDocument's constructor registers as an
InterfaceFactory for a handful of services. This means that any
requests for services between the time of the connection and creation
of HTMLDocument are dropped on the floor. This patch queues up all
requests and then processes them once HTMLDocument is created.

BUG=none
TEST=covered by tests
R=msw@chromium.org

Review URL: https://codereview.chromium.org/1222583002

Cr-Commit-Position: refs/heads/master@{#337138}