[Feedback] Move all tests to chrome/javatests
This CL moves the tests to a place not dependent on the Chrome shell.

Also, it updates the ConnectivityCheckerTestBase to instead just depend on
NativeLibraryTestBase, since it doesn't need anything else.

The FeedbackCollectorTest needs an activity, and as such it now uses
ChromeActivity and the ChromeActivityTestCaseBase. The profile is
retrieved from the last used one instead of from the current tab.

BUG=386395

Review URL: https://codereview.chromium.org/1216403003

Cr-Commit-Position: refs/heads/master@{#337148}