Mostly make webapps handle window.open in tabbed mode
Instead of throwing away the WebContents in webapp mode, pass it along
in the Intent.  This lets login windows and the like maintain parent-child
relationships between webapps and ChromeTabbedActivity, but it currently
fails to bring the WebappActivity back to the foreground when the child
tab closes.  Still need to come up with a good way to handle this.

* FullScreenActivity no longer uses the DocumentTabModelSelector's TabDelegates
  and instead maintains its own.  Might be worth revisiting this when there is
  a global TabManager.

* TabDelegate now knows how to fire Intents with WebContents to
  ChromeTabbedActivity.

* ChromeTabCreator is told how to accept Intents with WebContents and create
  Tabs with them.

* TabDelegate.getActivityTab() is deleted; its code is moved to the
  DocumentTabModelImpl since it was the only caller.

BUG=451453,490473

Review URL: https://codereview.chromium.org/1221643013

Cr-Commit-Position: refs/heads/master@{#337150}