Reintroduce DefaultBrowser.State UMA stat.
Take ownership of various other default browser stats in the process.

This is somewhat a revert of http://crrev.com/252953 plus some histograms enum fixes (note that this brings back a deprecated histogram's name but reporting with the exact same semantics as before so if anything this will add a continuity between the old and new data I guess).

BUG=505432

Review URL: https://codereview.chromium.org/1219743002

Cr-Commit-Position: refs/heads/master@{#337167}