Set start time for retargeted compositor transitions
When a CSS transition of a property occurs while the
previous transition of that same property has not completed,
it is possible for the animation to jump back if the
main thread is loaded.

The reason is that the animation curve for the new
transition is computed on the main thread using the
current value of the old transition.  However, if the
main thread is loaded, the compositor may have already
animated past the current value, so when the new
transition starts on the compositor, it will
animate the old values that the compositor has
already animated.

During the creation of the new animation curve, we
add the current time as the start time for this effect.
This will get passed down to the compositor and allow
it to account for the time lost between curve generation
and actual animation.

BUG=502278

Review URL: https://codereview.chromium.org/1190373002

git-svn-id: svn://svn.chromium.org/blink/trunk@198219 bbb929c8-8fbe-4397-9dbb-9b2b20218538