Speculative fix for ChromeWebViewGuestDelegate crash.
It is possible that when
ChromeWebViewGuestDelegate::HandleContextMenu() attempts to build a
menu to display that it will receive a null pointer (if the associated
web_contents' frame tree does not have a focussed frame). In the
current code this can lead to a crash. In the short term, while we
determine if this lack of a focussed frame is an error or not, early-out
if we receive a null menu.

BUG=487419

Review URL: https://codereview.chromium.org/1139193003

(cherry picked from commit 5fca2df48962e5812c8dc438c60434f1bcfae6a1)

Cr-Original-Commit-Position: refs/heads/master@{#330326}
Cr-Commit-Position: refs/branch-heads/2403@{#449}
Cr-Branched-From: f54b8097a9c45ed4ad308133d49f05325d6c5070-refs/heads/master@{#330231}