Oilpan: Eagerly finalize objects that have ExtraData
ExtraData can potentially have a back (raw) pointer to Blink objects,
which can cause use-after-free if the object that owns the ExtraData
is swept in an early phase of the lazy sweeping.

The ideal fix is to completely remove such raw pointers from content/
to Blink, but as a short-term fix, this CL marks those owner objects
as eagerly finalized.

BUG=

Review URL: https://codereview.chromium.org/1223793012

git-svn-id: svn://svn.chromium.org/blink/trunk@198361 bbb929c8-8fbe-4397-9dbb-9b2b20218538