Add --use-src-side-runtest-py flag to testing/scripts/common.py
This will make it a valid flag so that recipes can start
passing it.

For now it's no-op because default is going to be build-side
runtest.py (e.g. for chromium.perf infrastructure). Next step
will be making recipes pass that flag and change the wrapper
to make it take effect.

BUG=506498

Review URL: https://codereview.chromium.org/1222903003

Cr-Commit-Position: refs/heads/master@{#337579}