Assign cached test interfaces to the global and fix function ordering.
In http://crrev.com/1217633004 I changed the ordering of
V8Window::namedPropertyGetterCustom so the call to GetRealNamedProperty
was first, but this doesn't actually fix the problem with exposing
the test interfaces lazily.

Instead what we really want to do is keep a hidden Map of test interfaces
on the window to cache the instances.

This fixes the newly failing tests from my previous change:

http/tests/security/xss-DENIED-method-with-iframe-proto.html
webexposed/global-interface-listing.html
virtual/stable/webexposed/global-interface-listing.html

and makes the original test the change was for:
harness-tests/escape-bad-characters.html

continue to pass by caching the testRunner property.

I also correctly ordered the function so the checks for
local frames and the check for the frame being detached are respected.
This doesn't really matter for tests because we don't have testRunner
getter properties that have side effects on the Window.prototype, but
it makes the function look sane.

BUG=501648

Review URL: https://codereview.chromium.org/1223793019

git-svn-id: svn://svn.chromium.org/blink/trunk@198441 bbb929c8-8fbe-4397-9dbb-9b2b20218538