Reland: Add a public cpp targets for devices_app
This exposes the bare minimum stuff necessary to build a devices
app delegate, hiding the impl details behind a private
dependency.

Reland corrects a missing include in the app's main.cc.

This is a reland of the reland! The reland was reverted due
to suspicion of cast_shell_linux breakage, but that was
caused by something else.

BUG=None
TBR=reillyg@chromium.org,sky@chromium.org

Review URL: https://codereview.chromium.org/1213613018

Cr-Commit-Position: refs/heads/master@{#337725}