Remove old serializer
Remove old serializer (WebPageSerializerImpl) and its tests
(WebPageSerializerTest and resources) and use PageSerializer for all
serialization tasks in WebPageSerializer.

--

This review is part of set aiming to merge the two existing page
serializers, WebPageSerializerImpl and PageSerializer into the latter
one.

In addition to this it moves all the old tests from
WebPageNewSerializerTest and WebPageSerializerTest to the newer
PageSerializerTest structure and splits out a test for MHTML into the
MHTMLTest file.

Note: This is a rebase and split of
https://codereview.chromium.org/68613003 from Opera which was reverted
due to failing tests. In addition to the rebase some smaller fixes was
done to accommodate to changes in the code from when it was first
landed.

BUG=328354
R=hajimehoshi@chromium.org

Committed: https://src.chromium.org/viewvc/blink?view=rev&revision=198151

Review URL: https://codereview.chromium.org/1177733003

git-svn-id: svn://svn.chromium.org/blink/trunk@198503 bbb929c8-8fbe-4397-9dbb-9b2b20218538