Translate cancellation errors early for WKWebView with net stack.
The underlying error is not copied over from the IO thread when the net
stack is enabled for WKWebView, and the code in |-handleLoadError:
inMainFrame:| for aborting NSURLErrorCancelled errors incorrectly
assumed that these errors had underlying errors from the ent stack.  For
WKWebViews using the net stack, append the translated net error to the
NSURLErrorCancelled errors early so that the non-committed entries can
be properly discarded.

BUG=none

Review URL: https://codereview.chromium.org/1225503004

Cr-Commit-Position: refs/heads/master@{#337851}