Ensure ResourceScheduler starts requests even if the tab is closed.
Async revalidation requests are low priority, but they must be started
eventually even if the tab is closed.

Add unit tests to ensure that ResourceScheduler starts all requests
eventually, even if the client is deleted.

Also, modify ResourceScheduler to start any pending requests when the
client is deleted. This satisfies the requirement in the simplest way.

BUG=501805
TEST=content_unittests

Review URL: https://codereview.chromium.org/1192673002

Cr-Commit-Position: refs/heads/master@{#338058}