Clean up some detach logic in RenderFrameImpl.
There's no reason to gate RenderWidget frame unregistration
on the kSitePerProcess flag; it should be gated using the
same policy as the creation of the RenderWidget.

Moving the RenderWidget cleanup out of the is_subframe
block also paves the way for when the main frame also has a
RenderWidget.

BUG=419087
R=creis@chromium.org

Review URL: https://codereview.chromium.org/1229083002 .

Cr-Commit-Position: refs/heads/master@{#338145}