mac: Fix os_exchange_data_provider_mac.mm.
Provider::GetString is intended to return a plain-text representation of the
contents of the paste board. On OSX, it was possible for the paste board to
contain an NSURL, but not an NSString, and as a result Provider::GetString was
incorrectly returning the empty string.

This bug manifests itself when Chrome is compiled against the OSX 10.10 SDK. I
didn't trace down the exact source of the bug, although it is likely related to
https://codereview.chromium.org/1128173004.

BUG=463170

Review URL: https://codereview.chromium.org/1235713004

Cr-Commit-Position: refs/heads/master@{#338405}