Remove unused ExtraData support on exported mediastream objects.
The Blink exported interfaces RTCSessionDescriptionRequest, RTCVoidRequest,
and MediaStreamTrackSourcesRequest provide a facility to the embedder of
allowing 'ExtraData' to be associated and attached to object instances.

The embedder does not make use of this functionality for these mediastream objects,
so retire the unused ExtraData support. This simplifies the implementation of these
exported objects, in particular as regards to when they must be finalized as Blink
objects.

R=
BUG=

Review URL: https://codereview.chromium.org/1233793002

git-svn-id: svn://svn.chromium.org/blink/trunk@198741 bbb929c8-8fbe-4397-9dbb-9b2b20218538