Use permission_broker's OpenPath method to open serial ports.
This new method allows Chrome to open device nodes without having to
chmod them to a group of which Chrome is a member and allows the
permission broker to set opens on the file descriptor before passing it
to Chrome. This is a win for security and future extensibility.

BUG=None

Review URL: https://codereview.chromium.org/1235683003

Cr-Commit-Position: refs/heads/master@{#338478}