Only allow the -webkit-text color keyword in quirks/UA-sheet mode
This color keyword is mainly used to implement the 'quirk-inherit'
special value for 'color' on <table> in quirks mode [1].
Currently this keyword is treated differently for different properties,
but for the most common properties, the "only in quirks mode" rule is
applied. It's also marked as "QuirksOrUASheet" in CSSValueKeywords.in.
Extend the "only in quirks mode" rule to apply to all <color>
properties.

The test fast/css/webkit-text-display-none.html is converted to run in
quirks mode. It worked previously because setting inline style would use
the fast-path color parser which was inconsistent with the non-fast-path
color parser in accepting '-webkit-text' in standards mode.

[1] https://quirks.spec.whatwg.org/#the-tables-inherit-color-from-body-quirk

BUG=505410

Review URL: https://codereview.chromium.org/1216523006

git-svn-id: svn://svn.chromium.org/blink/trunk@198770 bbb929c8-8fbe-4397-9dbb-9b2b20218538