Introduce the PushSubscription.curve25519dh attribute (1/3)
This CL changes the WebPushSubscription object in the Blink API to accept
a new curve25519dh parameter in its constructor, and carry the member.

Mozilla's straw man has been used as a base for the behavior, with the
difference that we are experimenting with using Curve25519 rather than
P-256 as the ECDH curve. This is being discussed over at the IETF.

https://github.com/w3c/push-api/pull/130

The following Intent to Implement covers the newly exposed API:

https://groups.google.com/a/chromium.org/forum/#!topic/blink-dev/_LQX_cjAATA

This CL is part of a three-sided patch:
  [1] This CL.
  [2] https://codereview.chromium.org/1210283005
  [3] https://codereview.chromium.org/1237673002

BUG=486040

Review URL: https://codereview.chromium.org/1222903007

git-svn-id: svn://svn.chromium.org/blink/trunk@198784 bbb929c8-8fbe-4397-9dbb-9b2b20218538