Reland: Change the video color space default.
Reverted by: e735fb0c4b13b72561b8f9931a8788e97deced33
Original CL: https://codereview.chromium.org/1228843003
The new change is to media/cast/sender/h264_vt_encoder_unittest.cc.
The rest of the changes are due to rebasing on top of
11edd34b171b78462c1a27ac6587d977833122ea, which moves the color space
enum.

Previously video without color space metadata was assumed to be in
Rec601. Now the default depends on the kind of playback. Normal src=
defaults to Rec601 for SD sized video (<720 pixels high), and Rec709 for
HD. MSE will always default to Rec709. Using a size based heuristic
doesn't make sense for MSE where it is common for the resolution to
change mid playback.

This CL doesn't change the meaning of COLOR_SPACE_UNSPECIFIED. Instead,
it adds a color space field to VideoDecoderConfig, and updates the video
decoders to use this as the default if they don't find a more
authoritative value in the bitstream.

This also fixes a (year old!) bug causing the blackwhite
tests to always succeed, renames the rec709 blackwhite test
file to match the name in blackwhite.html, and re-encodes it
to contain the color space metadata (previously it had none).

BUG=333619

Review URL: https://codereview.chromium.org/1230593005

Cr-Commit-Position: refs/heads/master@{#338607}