DelegatedFrameHost should handle satisfied sequences in frame metadata.
When a RenderWidgetHostViewAura is acting as a guest view, it may need
to process satisifed sequences, received in the frame metadata, for its
embedder. This CL adds the satsfied sequence list to the data passed to
SwapDelegatedFrame() and includes it in the metadata subsequently passed
to SubmitFrame().

The CL also adds a dedicated BrowserPlugin satisfy callback in
ChildFrameCompositingHelper, just to make sure that these callbacks are
routed properly.

BUG=503594

Review URL: https://codereview.chromium.org/1229413003

Cr-Commit-Position: refs/heads/master@{#338696}