Refactor some browser-side chrome and webview printing code into component/printing/browser.
This patch refactors both the clank and webview code paths
on the browser-side. It moves the common bits to
components/printing/browser/, in particular the following changes
have been performed:
- significantly refactor print_manager and print_view_manager code
and extract common bits into a super class.
- use a callback to signal PdfWritingDone instead of PrintManagerDelegate.
This improves decoupling and also makes it possible to unify the clank side
of the code.
- extract common methods into functions in print_manager_utils.

BUG=465802

Review URL: https://codereview.chromium.org/1024883002

Cr-Commit-Position: refs/heads/master@{#338697}