Add data files for base and net. Move GN Python proto outputs.
The "data" for net_unittests and base_unittests should match the isolates.

To make this work, the python outputs for protoc is updated to match GYP (all Python outputs go into a "pyproto" directory).

Targets were added for Pytho-only dependencies listing the data files. This way one only needs to depend on the target and implicitly get the associated data files.

The location of the google protobuf helpers is moved to match GYP.

R=dpranke

Review URL: https://codereview.chromium.org/1227393002

Cr-Commit-Position: refs/heads/master@{#338741}