Fix keyboard accessibility feedback
Say we get an AccessibilityEvent of TYPE_VIEW_TEXT_CHANGED for a change
from "a" to "aa". The old implementations of GetTextChangeAddedCount
and GetTextChangeRemovedCount would match "a" to the first half of "aa"
from the left, then match "a" to the second half of "aa" from the right,
over-counting the number of matching letters. PopulateAccessibilityEvent
would then write the bogus added/removed letter counts to the event. On
Android, TalkBack would then ignore the malformed event, and the user
would miss audio feedback for that keystroke.

Now, we remove any matching letters on the left before matching on the
right, to avoid over-counting.

internal bug 20923568

BUG=510222

Review URL: https://codereview.chromium.org/1225113004

Cr-Commit-Position: refs/heads/master@{#338771}