Change KeepAliveImpl to handle extension unload and shutdown correctly.
Previously, if a KeepAlive was active (such as when a mojo-backed API
call is in-progress) when an extension is reloaded, the KeepAlive could
be destroyed after the ProcessManager resets the KeepAlive count,
leading to an incorrect KeepAlive count. If a KeepAlive was active
during browser shutdown, it would attempt to access its ProcessManager
after KeyedServices have been destroyed, which crashes.

BUG=509959

Review URL: https://codereview.chromium.org/1236133002

Cr-Commit-Position: refs/heads/master@{#338806}