[AndroidWebViewShell] Replace rebaseline script with a new version using test_runner.py
Now that test_runner (with isolate) can be used to run tests in WebViewLayoutTest
the old shell script is obsolete. This patch replaces the old shell script with a
simplified one and makes it consistent with how we run it on bot. If the rebaseline
parameter is supplied this script will make sure to recreate new expectation files.

BUG=497861

Review URL: https://codereview.chromium.org/1221433015

Cr-Commit-Position: refs/heads/master@{#338868}