Always include floating children into visual overflow
Previously we have conditions of whether to include floating children
into visual overflow. This makes the recalculation of visual overflow
depending on the change of the conditions. We could recalc visual
overflow on every change of the conditions, but that looks fragile.

With this CL we always include floating children into visual overflow.

This fixes the first half of bug 507590 (summarized in
https://code.google.com/p/chromium/issues/detail?id=507590#c5).
Because of the second half of the bug, the added test is still failing.

BUG=507590
TEST=fast/block/float/containing-block-change-compositing.html (still fails)

Review URL: https://codereview.chromium.org/1236793005

git-svn-id: svn://svn.chromium.org/blink/trunk@198977 bbb929c8-8fbe-4397-9dbb-9b2b20218538