Change Cast window transparency option from switch to API
The main use case for this feature involves starting a new Cast app
that requires transparency (not starting a new cast_shell process).
Also folded in the code to set the background to black for default
case - this is to prevent white flashes between apps.  Previously
was internal and it doesn't make sense to set this from two places.

BUG=internal b/22121206

Review URL: https://codereview.chromium.org/1239623004

Cr-Commit-Position: refs/heads/master@{#338907}