Attach a WebBluetooth to the LocalFrame and use it if it's available.
This lets us move the Bluetooth implementation from content/child to content/renderer, which lets us plumb the RenderFrame's ID through Blink.

We also add an error enum for content/ to use if the frame ID is wrong.

This is the first of a 3-patch sequence:
1. Attach a frame-specific WebBluetooth in Blink. (This patch)
2. Move Web Bluetooth from content/child/ to content/renderer/ (https://codereview.chromium.org/1234413006/)
3. Put the RenderFrame's ID in that WebBluetooth and plumb it back to the browser. (https://codereview.chromium.org/1228113004)

BUG=500989

Review URL: https://codereview.chromium.org/1228283003

git-svn-id: svn://svn.chromium.org/blink/trunk@199067 bbb929c8-8fbe-4397-9dbb-9b2b20218538