Move ImageSource enums into ImageDecoder
AlphaOption and GammaAndColorProfileOption are used only by
ImageDecoder. Prior to this CL, these enums lived in ImageSource,
which only passed them to the creation of a DeferredImageDecoder.

ImageSource is always created with the default options, so just pass
those directly to DeferredImageDecoder.

Change all users of the enums to point to ImageDecoder instead of
ImageSource.

No change in behavior, no new tests.

BUG=439655

Review URL: https://codereview.chromium.org/1242023004

git-svn-id: svn://svn.chromium.org/blink/trunk@199076 bbb929c8-8fbe-4397-9dbb-9b2b20218538