Invalidate paint of the start of continuations when needed
Changes of continuations affect layout (visual overflow) and paint of
the start of continuations with outlines. This CL address the paint
issue.

The test covers focus-ring only because I haven't got a test for normal
outlines. This might be because for now we have different painting paths
for normal outlines and focus rings. However, if we combined the paths
(which would greatly simplify outline painting), we would expose this
issue for normal outlines, so this CL covers both.

BUG=506669
TEST=fast/repaint/focus-ring-on-continuation-move.html

Review URL: https://codereview.chromium.org/1221423002

git-svn-id: svn://svn.chromium.org/blink/trunk@199140 bbb929c8-8fbe-4397-9dbb-9b2b20218538