Change HttpServerPropertiesManager::ParseAlternativeServiceDict() signature.
Change HttpServerPropertiesManager::ParseAlternativeServiceDict() to return a
bool indicating success, and return the parsed AlternativeServiceInfo as an
outparam.

This might be the first step on the long way to obsolete
UNINITIALIZED_ALTERNATIVE_PROTOCOL, which might pave the way to eventually
consolidate enum AlternativeProtocol and enum NextProto.

BUG=116575

Review URL: https://codereview.chromium.org/1218683006

Cr-Commit-Position: refs/heads/master@{#339436}