Precompile more in Blink in Windows for faster compilations
One reason Blink is slow to compile is that there is a lot of code
included in every compilation unit since everything depends on either
LayoutObject.h or Document.h and those in turn include huge portions
of the rest of Blink.

By precompiling LayoutObject.h and Document.h the compilation of core and
modules in Blink can be 4 times faster (4 minutes instead of
19 minutes on my computer).

The downside is that it will introduce Document.h and LayoutObject.h
also in compilation units that didn't expect it, for instance
XPathGrammer.y that suddenly will have both blink::Path and
blink::XPath::Path in scope (and blink::Filter / blink::XPath::Filter)

Note that distributed compilation system disables precompiled headers
globally so this will *not* make trybots faster.

This is a new attempt at this change after bug 511945 scared us into
reverting it. We still do not know the exact cause of 511945 but we
are making another attempt with this.

BUG=495697
R=thakis@chromium.org,sigbjornf@chromium.org

Review URL: https://codereview.chromium.org/1227953008

git-svn-id: svn://svn.chromium.org/blink/trunk@199168 bbb929c8-8fbe-4397-9dbb-9b2b20218538