download_gold_plugin.py: Redirect stderr to /dev/null when calling gsutil.
When called from a bot, runhooks' stdin will be redirected to /dev/null. The
hack we use in update.py that redirects stdin to stderr causes a read-only
/dev/null to eventually appear as gsutil's stderr here, which causes it
to blow up for an unknown reason. To avoid this problem, open a writable
/dev/null and redirect unconditionally here.

This prevents command line users from seeing gsutil's output, but at least
the thing runs.

BUG=453195
R=thakis@chromium.org

Review URL: https://codereview.chromium.org/1246783002

Cr-Commit-Position: refs/heads/master@{#339542}