mandoline: Fix character entry in html_viewer on Linux.
This reverts my previous change (https://codereview.chromium.org/1232703006),
which fixed the double character insertion in the mandoline omnibox, but
broke all text entry in the html_viewer. The correct change was to make
sure we're using an InputMethodMandoline.

In https://codereview.chromium.org/1155013005/, InputMethodMandoline
appears to have been removed by accident while removing
InputMethodEventFilter. However, at least on Linux, we need to *not* use
InputMethodAuraLinux, which operates on a stream of X11 events, instead
of a stream of events that is designed to look like Windows events on
all platforms.

We don't see these problems on Windows because InputMethodWin is already
expecting a stream of events that looks like what Windows natively sends.

BUG=none

Review URL: https://codereview.chromium.org/1234623004

Cr-Commit-Position: refs/heads/master@{#339672}