[NaCl SDK] Fix initialization of pthread_t in dlopen example
Initializing a pthread_t to NULL (or any value really)
is not valid.  This causes a compiler warning with the new
arm glibc toolchain so needs to be fixed because we
enable the new toolchain.

Turns out the pid_ member was not even needed so this
change removes it and adds a pthread_detach call which
allows for freeing on the thread's resources
(in the absence of pthread_join).

CQ_EXTRA_TRYBOTS=tryserver.chromium.linux:linux_nacl_sdk;tryserver.chromium.mac:mac_nacl_sdk
BUG=505885

Review URL: https://codereview.chromium.org/1243123002

Cr-Commit-Position: refs/heads/master@{#339720}