Reland unit tests for FontPlatformData::hasSpaceInLigaturesOrKerning
Add a set of unit tests for the hasSpaceInLigaturesOrKerning method that
verifies that the correct value is returned for a font that has space in
either the GPOS or GSUB tables vs for one that does not. Also tests that
the check is bypassed and false is returned if kerning and ligatures are
both disabled. Finally, it fixes the failure mode in that the method now
returns false in cases where the font tables cannot be accessed.

Originally committed as r199059 (bbb929c88fbe43979dbb9b2b20218538), then
reverted as r199186 (e3ec9c7d156ee82dc4f775031406db3407fb52b6) as it too
caused assert failures due to a bad rebase.

TBR=yhirano@chromium.org

Review URL: https://codereview.chromium.org/1245263003

git-svn-id: svn://svn.chromium.org/blink/trunk@199239 bbb929c8-8fbe-4397-9dbb-9b2b20218538