Track lazy-decoded pixel refs explicitly.
We are currently relying on SkBitmap::{set,get}URI() to tack discardable
info onto the SkBitmap itself.

This is convenient, but doesn't fit in an SkImage world (plus it is
inneficient to query via strcmp).

To get rid of the deprecatedBitmapForCurrentFrame() path, this CL reworks
lazy-decode tracking in a manner similar to the other frame attributes
(alpha, duration, etc.): the value is queried from DeferredImageDecoder
(via ImageSource), and cached in BitmapImage::m_frames.

BUG=449197
R=junov@chromium.org,noel@chromium.org,scroggo@chromium.org

Review URL: https://codereview.chromium.org/1253483004

git-svn-id: svn://svn.chromium.org/blink/trunk@199388 bbb929c8-8fbe-4397-9dbb-9b2b20218538