Fix a crasher in the GPU process caused by the DXVA code attempting to create an instance of the Video processor MFT
Merge to M45

The DXVA decoder object on Windows 8+ uses DX11 if available and in this mode it uses the VideoProcessor MFT for colorspace
conversion. This object exposes the IMFTransform interface and can be created like a traditional COM object via the
CoCreateInstance API. However this API expects CoInitialize to be called which is not the case on the main GPU thread and
fails with an error indicating that the COM apartment is not initialized. It is not clear as to why it is not failing in all
cases.

Proposed fix is to use DllGetClassObject and IClassFactory::CreateInstance to create the object directly and avoid using
CoCreateInstance. For this purpose added a helper function CreateCOMObjectFromDll in the dxva code.

BUG=493894
TBR=dalecurtis

Review URL: https://codereview.chromium.org/1235373003

Cr-Commit-Position: refs/heads/master@{#338975}
(cherry picked from commit f74c38420401d72fc8fdae2b8d03aece1279f7e7)

Conflicts:
	content/common/gpu/media/dxva_video_decode_accelerator.cc

Review URL: https://codereview.chromium.org/1254023002 .

Cr-Commit-Position: refs/branch-heads/2454@{#127}
Cr-Branched-From: 12bfc3360892ec53cd00fc239a47e5298beb063b-refs/heads/master@{#338390}