Add lossy casts between 'blink::WebSecurityOrigin' and 'url::Origin'.
This is a step towards improving the Origin concept between Blink and
its embedders. We'll still be losing information like 'document.domain'
when round-tripping an origin through the embedder, which we'll need
to fix in future CLs.

This depends on https://codereview.chromium.org/1251063002 adding
creation APIs to 'url::Origin'.

BUG=490074

Review URL: https://codereview.chromium.org/1251833002

git-svn-id: svn://svn.chromium.org/blink/trunk@199482 bbb929c8-8fbe-4397-9dbb-9b2b20218538