Add counters for deprecated HTML width attributes
The HTML width attribute traditionally supports values that contain
'%' based length values even if this is not standard. In Blink/WebKit
the presence of '*' also has an impact, causing the value to be ignored.
This CL adds a counter to account for these things, and to see how common
they are over time (with the hope of deprecating/removing them eventually).

BUG=501870

Review URL: https://codereview.chromium.org/1252403003

git-svn-id: svn://svn.chromium.org/blink/trunk@199510 bbb929c8-8fbe-4397-9dbb-9b2b20218538