Password manager internals page should make clear it is disabled in Incognito
Currently, chrome://password-manager-internals does not work in Incognito, by design [1]. But it does not says so. In fact, it indicates that it should work as in normal profiles.

Hence this CL changes the description of the internals page to notify it is turned off in Incognito.

Alternatively, it could redirect the user back to the normal profile as about:settings does, but I'm not sure it is worth the trouble. Such redirecting might also cause unexpected non-Incognito browsing, so I'm not comfortable doing that under the user's hands.

[1] https://docs.google.com/document/d/1ArDhTo0w-8tOPiTwqM1gG6ZGqODo8UTpXlJjjnCNK4s/edit#heading=h.jtqzjkc34g2x

R=mkwst@chromium.org
BUG=513226,347927

Review URL: https://codereview.chromium.org/1257793002

Cr-Commit-Position: refs/heads/master@{#340467}