[qcms] Check matrix transforms have the same PCS
qcms does not support transforms containing matrix profiles where the input
and output profile PCS (profile connections space) differs, but it does not
check the condition either (leading to the color cast effect in some cases,
see the bug for an example).

When a matrix color transform is created and the transform is matrix, check
that the PCS of the input / output profiles is the same, and fail transform
creation if not.

Update the internal profile creation code to assign an XYZ_SIGNATURE to the
profile PCS of color profiles created this way (the qcms sRGB color profile
for example) since their tristimuli (primaries) are in XYZ space.

CQ_INCLUDE_TRYBOTS=tryserver.blink:linux_blink_rel,mac_blink_rel,win_blink_rel

BUG=510682

Review URL: https://codereview.chromium.org/1256203002

Cr-Commit-Position: refs/heads/master@{#340482}