Allow empty chunks in the sync Cronet API
Currently the (deprecated) sync API of Cronet does not allow empty
chunks to be uploaded. Since uploads are finished with a final chunk
this means that the client app must know in advance for each chunk they
whether it is going to be the last one. This also means that it is not
possible to upload an empty request body this way.

Remove the length > 0 check from the appendChunk method.

BUG=

Review URL: https://codereview.chromium.org/1235883003

Cr-Commit-Position: refs/heads/master@{#340486}