Modify heuristics for main frame detection for URL requests
Change the heuristics for determining whether an external URL request
is on the main frame to not involve pending navigation items and
instead treat the requestURL as equal to the mainDocumentURL if they
have the same resource specifier (portion following the scheme). Also
adds a histogram to record what proportion of requests are blocked as
a part of subframe external protocol blocking.

BUG=348640

Review URL: https://codereview.chromium.org/1231223002

Cr-Commit-Position: refs/heads/master@{#340497}