Don't crash in RenderViewImpl::DidFlushPaint for views with no local frames.
If a RenderView is loading its main frame, and the main frame is asked
to navigate to a different SiteInstance before it paints, it's
possible to receive a DidFlushPaint after swapping out is complete,
which crashed in DidFlushPaint when trying to access a remote frame's
provisionalDataSource().  Currently, this only matters in
--site-per-process mode, but will also matter without it once
swapped-out is gone.

BUG=513552

Review URL: https://codereview.chromium.org/1255003005

Cr-Commit-Position: refs/heads/master@{#340612}