Use minimum scale to adjust viewport size (Blink-side).
When the top controls are moved, but before a full resize, we need to keep
track of the adjustment that CC will make on the visual and layout viewports.
The top controls directly affect the visual (pinch) viewport, but the layout
viewport needs to be adjusted as well. Previously this was done by matching
the aspect ratio. This works only if the top controls are the only thing that
can change the viewport size.

This patch makes it so that the top controls adjustment is scaled by the
minimum page scale factor before being subtracted out of the layout viewport.
We use the minimum page scale factor since the main frame is sized to be the
visual viewport at minimum scale.

BUG=464814

Review URL: https://codereview.chromium.org/1238423004

git-svn-id: svn://svn.chromium.org/blink/trunk@199585 bbb929c8-8fbe-4397-9dbb-9b2b20218538