Add support for /DEBUG:FASTLINK to gyp and gn files.
VC++ 2015 has a /DEBUG:FASTLINK option that can significantly
improve link performance on developer machines. Compared to normal
VS 2015 links it reduces elapsed time by 33-55% and reduces peak
working set by 65-90%. PDB sizes are reduced by 65-72%.

/DEBUG:FASTLINK works by having PDB files reference symbols in
.obj and .lib files instead of copying the debug information. This
means that PDBs only work on the machine they were created on, and
only as long as the .obj and .lib files exist, but this constraint
matches local-developer-builds very nicely.

/DEBUG:FASTLINK has an (undocumented) restriction that it is disabled
if /PROFILE is selected.

For GYP builds this feature is enabled with GYP_DEFINES=win_fastlink=1
For gn builds this feature is enabled with is_win_fastlink = true in
args.gn.

This switch is not supported on VC++ 2013

This connect bug requests a tool to de-fastlink a PDB so that it can
be shared/deployed for debugging on other machines:
https://connect.microsoft.com/VisualStudio/feedback/details/1598471

BUG=440500
R=jochen@chromium.org

Review URL: https://codereview.chromium.org/1254723012

Cr-Commit-Position: refs/heads/master@{#340733}