Remove unsafe async notify completion callback call.
This is a partial revert of a previous CL:
https://codereview.chromium.org/1241443005/

The original intention was to early return when there are no
pending async notifications, but it is unsafe to check if the
pending queries are empty there.

Because the notifications are fulfilled asynchronously we can
not reliably unit test the completion of that query. Another
method must be developed to unit test that query.

R=aelias@chromium.org, sievers@chromium.org
BUG=514124

Review URL: https://codereview.chromium.org/1259483003

Cr-Commit-Position: refs/heads/master@{#340766}