Inform WebWidget clients when layout size changes
A change to main frame layout size may mark the frame for layout,
and prior to this change that information was never passed up to
the WebWidget that may own the frame and control its layout.
This patch adds a callback on the WebWidgetClient informing it
of a change in layout size. Clients are expected to mark themselves
and/or their WebPluginContainer objects as needing layout.

This is the Blink side of a two sided patch.

TBR=pdr@chromium.org
BUG=493375

Committed: https://src.chromium.org/viewvc/blink?view=rev&revision=199587

Review URL: https://codereview.chromium.org/1258923005

git-svn-id: svn://svn.chromium.org/blink/trunk@199610 bbb929c8-8fbe-4397-9dbb-9b2b20218538