Avoid assertion when using calc() in -webkit-line-clamp
This patch makes the property parser reject calc() values when parsing
-webkit-line-clamp. Currently, using a calc() value will result in a
assertion when we apply the property. As the property is non-standard,
we simply disallow calc() rather than adding support for it.

This also re-orders the checks in parsing -webkit-line-clamp to avoid
trying to read fValue when it isn't initialized.

BUG=344966,474906

Review URL: https://codereview.chromium.org/1266513003

git-svn-id: svn://svn.chromium.org/blink/trunk@199647 bbb929c8-8fbe-4397-9dbb-9b2b20218538