Makes it possible for an html frame tree to have numerous local frames
When a frame is navigated we immediately swap it to remote and then
call Embed(). It's possible Embed() will callback to the same app, in
which case we need to convert the frame from remote to local. This
change does this.

The result of this change is that HTMLFrame implements FrameTreeClient
and forwards to HTMLFrameTreeManager. HTMLFrameTreeManager may end up
containing multiple local frames, each having it's own
connection. HTMLFrameTreeManager is destroyed only when the root is
destroyed, or there are no more local roots.

The code would likely be cleaner if the blink parts were separated
out and HTMLFrame* were effectively a client lib. I'll consider that
once I get everything working smoothly.

BUG=479172,490221
TEST=covered by tests
R=ben@chromium.org

Review URL: https://codereview.chromium.org/1254973006

Cr-Commit-Position: refs/heads/master@{#340954}