[Android WebViewShell] Add test for interfaces not exposed in webview using annotated exclusion file.
Add WebViewLayoutTest.testWebViewExcludedInterfaces which tests that certain
interfaces are not exposed in webview. The excluded interfaces are listed in
a text file to allow for easy modification and comments within the file are
supported to provide clarifications and to refer to crbugs.

The test also makes sure that the listed interfaces are actually exposed in
blink to avoid typos and non-existing interfaces in the exclusion file.

BUG=514217

Review URL: https://codereview.chromium.org/1260683002

Cr-Commit-Position: refs/heads/master@{#340956}