Unify code paths for fixed and absolutely positioned objects in accumulateOffsetTowardsAncestor().
Deleted most of the code specific to handling of fixed-positioned objects, and
have fixed-positioned objects use the code for absolutely positioned objects
instead. One piece needs to be kept, though: If the container of a fixed
positioned object is the viewport, we need to make sure that we call
LayoutObject::localToAbsolute() to calculate the correct position, since
DeprecatedPaintLayer::location() doesn't handle fixed-positioning properly.

This gets rid of an assertion failure that would trigger when we had a multicol
with a transform with a fixedpos.

BUG=496929
R=chrishtr@chromium.org,jchaffraix@chromium.org

Review URL: https://codereview.chromium.org/1268493002

git-svn-id: svn://svn.chromium.org/blink/trunk@199725 bbb929c8-8fbe-4397-9dbb-9b2b20218538