[css-grid] Grid containers reporting wrong preferred widths
LayoutGrid used to have its own overwritten version of
computePreferredLogicalWidths() because we didn't have an implementation of
computeIntrinsicLogicalWidths(). That implementation was not as complete as
LayoutBlock's because it was not taking into account min/max-width
restrictions.

computeIntrinsicLogicalWidths() has been there for some time so we can
safelly kill our overwrite and use LayoutBlock's version which addresses all
the FIXME's we had in our code.

Review URL: https://codereview.chromium.org/1258923009

git-svn-id: svn://svn.chromium.org/blink/trunk@199729 bbb929c8-8fbe-4397-9dbb-9b2b20218538