New SELECT Popup: Do not trigger style recalc and layout repeatedly.
In the original code, we trigger style recalc and layout N/1000 times where N is
the number of OPTIONs, and it caused janky user experience in a few seconds
after opening a popup.

This CL changes it so that we only trigger one style recalc and layout after
opening a popup.

BUG=510387

Review URL: https://codereview.chromium.org/1256103004

git-svn-id: svn://svn.chromium.org/blink/trunk@199798 bbb929c8-8fbe-4397-9dbb-9b2b20218538