Remove a race around NetworkChangeNotifier::test_notifications_only_
There is a possible race in which
NetworkChangeNotifier::test_notifications_only_ might be read on one
thread while being written on another. This CL fixes that by making
the variable static and only allowing writes to it before the
NetworkChangeNotifier is created (before other things might try to
read it).

BUG=454652

Review URL: https://codereview.chromium.org/1255263002

Cr-Commit-Position: refs/heads/master@{#341354}