Use ExpectTotalCount in CrossDevicePromoTest to check for empty samples
Instead of checking for null histogram samples, the absence of new ones should be checked with ExpectTotalCount being 0. No new samples can result in both null samples, or empty samples, and this depends on other tests running before. This can be a source of flakiness.

This was part of https://codereview.chromium.org/1264123002/, see that or the bug for more context.

BUG=473689
TBR=atwilson@chromium.org

Review URL: https://codereview.chromium.org/1268893004

Cr-Commit-Position: refs/heads/master@{#341371}