PlzNavigate: disable RenderFrameHostManagerTest::PageDoesBackAndReload.
PlzNavigate has a very different way of carrying out renderer initiated
navigations and of dealing with navigation cancellation. So this change disables
RenderFrameHostManagerTest::PageDoesBackAndReload because adapting it to work
with the two implementations would make it full of special cases. There are
other tests (listed in a comment in the change itself) that exercise the
equivalent logic in PlzNavigate.

BUG=439423

Review URL: https://codereview.chromium.org/1270843002

Cr-Commit-Position: refs/heads/master@{#341581}