Fix window.open to not navigate RemoteFrames when given an empty URL.
According to step 11 in
https://html.spec.whatwg.org/multipage/browsers.html#apis-for-creating-and-navigating-browsing-contexts-by-name,
if the url given to window.open is the empty string, and its target is
an existing frame, the right behavior is to return the WindowProxy and
never queue a navigation.

Currently, this is handled for LocalFrames in
NavigationScheduler::scheduleLocationChange, but RemoteFrames
incorrectly navigate to about:blank in this case.

This CL fixes that by moving the check for empty URLs into
createWindow().

BUG=486896

Review URL: https://codereview.chromium.org/1261143003

git-svn-id: svn://svn.chromium.org/blink/trunk@199929 bbb929c8-8fbe-4397-9dbb-9b2b20218538