Make it easier to test changes to a .icon file.
No need to recompile Chrome; views_examples_exe will read from a file. This would also be useful for the tool which will allow designers to create .icon files directly (which may or may not ever materialize).

Since this is potentially unsafe (could crash on the wrong inputs) and doesn't serve a purpose in Chrome itself, it's ifdef'd out of the header everywhere except those places that explicitly need it.

BUG=505953

Review URL: https://codereview.chromium.org/1268633002

Cr-Commit-Position: refs/heads/master@{#341667}