Remove empty alternative service vector from |alternative_service_map_|.
Remove empty alternative service vector from |alternative_service_map_| in
HttpServerPropertiesImpl::GetAlternativeServices().

The reporter of the bug confirmed that Patch Set 1 fixes the crash on their
local build.

The author of this CL confirmed that the regression test added in Patch Set 2
run without the fix in Patch Set 1 does crash on a local build.

Note that this is only half of the bug: the same problem still persists for
canonical hosts.  That deserves another regression test and another CL.

BUG=516486

Review URL: https://codereview.chromium.org/1256113005

Cr-Commit-Position: refs/heads/master@{#341708}