Componentize //chrome/browser/ui/omnibox
This CL componentizes the code in //chrome/browser/ui/omnibox that is shared
with iOS. The only files left in //chrome/browser/ui/omnibox are (a) specific
to the //chrome embedder of the omnibox component or (b) unittests, whose
componentization will occur as future work.

The code has no problematic dependencies, and so componentization is trivial.

BUG=511969
TBR=sky

Review URL: https://codereview.chromium.org/1257143004

Cr-Commit-Position: refs/heads/master@{#341898}