Change run-webkit-test's is_unexpected field
This field tries to capture unexpected flakes. In practice,
I don't think people actually use this output and it complicates a lot
of the downstream code. The downstream code should have a single
bit to check if the failure caused the bot to turn red.

Instead, make it represent whether the test was a regression (i.e.
turned the bot red).

We may want to rename this to have it make more sense, but at the moment
the tree is on fire due to removing flaky lines and we need this in
order to make sense of what's actually failing.

BUG=516908
R=joelo@chromium.org

Review URL: https://codereview.chromium.org/1279823002 .

git-svn-id: svn://svn.chromium.org/blink/trunk@200115 bbb929c8-8fbe-4397-9dbb-9b2b20218538