crnet: initialize net log earlier
The net log is completely threadsafe but needs to be used on the file user
blocking thread, so initializing it on the net thread can cause a race at
startup. Initialize it earlier on the main thread instead.

Also, bring back the graphical crnet_consumer, now that crnet_test exists.

BUG=514764

Review URL: https://codereview.chromium.org/1278923002

Cr-Commit-Position: refs/heads/master@{#342216}