Add another flaky line and s/NeedsManualRebaseline/NeedsRebaseline.
All these NeedsManualRebaselines are over a month old. Whatever
manual was required is passed.

There's four NeedsManualRebaselines left that are 2 weeks old.
It's not clear if those are being attended to.

TBR=pkasting

More cleanup after r200095.

Add back a bunch of lines that turned out to be needed.
This is filling in the holes for the bots that aren't
covered by the CQ.

TBR=pkasting@chromium.org

Review URL: https://codereview.chromium.org/1275773007 .

git-svn-id: svn://svn.chromium.org/blink/trunk@200124 bbb929c8-8fbe-4397-9dbb-9b2b20218538