Create MockUsbService in SetUp phase of Android USB tests.
If we wait to create the MockUsbService in SetUpOnMainThread() then the
BrowserContext has already been created and one of its KeyedServices
(such as PrinterDetector) may have already created a non-mock instance
of UsbService. This will trip the DCHECKs that ensure only one instance
of UsbService exists at a time. Instead create the MockUsbService in
SetUp().

I think this wasn't being hit on trybots because UsbService creation
fails if the system doesn't have the USB subsystem enabled in the
kernel (see crbug.com/517146). I am surprised no one noticed it on local
builds.

Fixing this is a prerequisite for resolving issue 517146.

BUG=517146

Review URL: https://codereview.chromium.org/1274033004

Cr-Commit-Position: refs/heads/master@{#342425}