Add URLRequest unit tests for HPKP violation reports
This CL adds URLRequest unit tests for sending HPKP violation
reports. The tests check that HPKP reports are sent when a URLRequest
violates a previously set pin, and that reports are sent when a
URLRequest violates the pins in a PKP-Report-Only header on that
request.

This is based on top of crrev.com/1266723003

BUG=445793

Review URL: https://codereview.chromium.org/1270043004

Cr-Commit-Position: refs/heads/master@{#342450}