Fix bug in box layout that affects size 0 views.
The layout code would reserve space for between-child-padding for a view
of size 0, but then when actually setting child bounds, ignore that 0
width child and not add the padding. The end result is that horizontal
box layouts with a zero width child would add an additional
between-child-padding amount of padding at the end of the box.

BUG=none

Review URL: https://codereview.chromium.org/1279763005

Cr-Commit-Position: refs/heads/master@{#342731}