Change handling of unmatched surrogate pairs
Change the way blink handles unmatched or invalid UTF16 surrogate pairs.
Formerly once an unmatched on invalid surrogate pair was encountered the
processing of the text node was aborted, resulting in the text following
the invalid surrogate pair to be trimmed.

With this change the unmatched/invalid pair is instead replaced with the
unicode replacement character and the rest of the node handled normally.

This matches the behavior of Firefox and Internet Explorer.

R=bsittler@chromium.org
BUG=518039
TEST=fast/text/unmatched-surrogate-pair.html

Review URL: https://codereview.chromium.org/1285633003

git-svn-id: svn://svn.chromium.org/blink/trunk@200348 bbb929c8-8fbe-4397-9dbb-9b2b20218538