[2454] Reintroduce the WebappActivity bouncer
Consolidating the old WebappManager with the ChromeLauncherActivity doesn't
work well on JB and KK because the ChromeLauncherActivity needs to be visible
in recents (things launch on top of it) while the the WebappManager shouldn't
have been visible.

Reintroduce the WebappManager as the WebappLauncherActivity and use the
taskAffinity="" flag to make sure nothing ever launches on top of it.

Also gets rid of the EXTRA that was used only for re-activating
WebappActivities; the action serves the same purpose.

BUG=517426

Review URL: https://codereview.chromium.org/1281403002

TBR=mariakhomenko
Cr-Commit-Position: refs/heads/master@{#342745}

Review URL: https://codereview.chromium.org/1281483013 .

Cr-Commit-Position: refs/branch-heads/2454@{#281}
Cr-Branched-From: 12bfc3360892ec53cd00fc239a47e5298beb063b-refs/heads/master@{#338390}