[MD settings] adding paper-dropdown-menu and paper-submenu
During the initial move from Polymer 0.5 to 1.0, the
paper-dropdown-menu and paper-submenu were left out.  They've now been added.  These elements will be used
in the Material Design settings pages.

The changes to paper-menu are to help support the
paper-submenu.

The change to paper-checkbox is because I accidentally
added a \n to the js file in my prior CL.

BUG=425627

Review URL: https://codereview.chromium.org/1284033003

Cr-Commit-Position: refs/heads/master@{#343066}