Ensure Task Manager's "sort" column is always a visible column.
If the Task Manager is sorting by the contents of the "Task" column,
say, and you hide that column, the Task Manager continues sorting
the table's contents by the Task column. Visually, this means no column
header displays an up/down sort arrow, and there's no apparent logic to
the table's sort order. This change forces the task manager to select a
new column for sorting the table in this situation.

BUG=517937

Review URL: https://codereview.chromium.org/1282593004

Cr-Commit-Position: refs/heads/master@{#343089}