Remove tab_utils.cc from android build
Also remove a couple pass-throughs from tab_utils.cc. The motivation for this change is to make it clearer that the code in this file is not used on Android, and to make it easier to change the file by removing the need to worry about the Android compile.

BUG=none

Review URL: https://codereview.chromium.org/1291473002

Cr-Commit-Position: refs/heads/master@{#343130}