Fix 2 ChromeVox next tests that passed without actually running.
The way to create callbacks for async functions in these e2e tests is to
invoke this.newCallback().  See https://codereview.chromium.org/938623003
for more background.  If this is not done, the tests will pass silently
before the asyc callbacks get called which means much of the tests don't
run.

This CL fixes two of the tests and marks the rest with TODOs because
they fail and need more debugging when this problem is fixed.

The live region tests are fixed by making paragraph nodes output their descendants
instead of their value property.

BUG=None

Review URL: https://codereview.chromium.org/1288043002

Cr-Commit-Position: refs/heads/master@{#343217}