[Download Notification] Update notification as progress type on complete/cancel
We have had the code showing notification as progress type once before showing it as base type. But it didn't work due to a bug (see the comment in the review for detail). This patch fixes it and refactor the code to make the code clear.

BUG=518698
TEST=manually tested

Review URL: https://codereview.chromium.org/1279593003

Cr-Commit-Position: refs/heads/master@{#343247}