Keep track of the timestamps for coalesced events
Events can be coalesced at browser side before being sent to
renderer. Currently when coalescing happens, some events'
timestamp are lost, i.e. we preserve the timestamp from
either the newest event or the oldest event.

In this CL we use the event's LatencyInfo to carry the timestamps of
a small set (up to 2) of the raw events. These timestamp are
used for input latency UMA metric. A heap based container to
carry the timestamp of all the raw input events seems excessive.
So lets use a stack based container with a small fixed-size and
see how it can change the graph pattern of our UMA metrics for
Event.Latency.Renderer2.MouseWheel/GestureScrollUpdate.

BUG=352193

Review URL: https://codereview.chromium.org/1231433006

Cr-Commit-Position: refs/heads/master@{#343255}