Mac Overlays: Make ImageTransportSurfaceOverlayMac observe GPU switches
When the GPU switches, create a new OpenGL context, to see which GL
renderer it uses. Whenever a context is made current to the surface,
ensure that the context is using the appropriate renderer.

Ideally, the GpuSwitchingManager should be able to tell the observers
which GL renderer to use, but that is not currently wired up.

BUG=519097

Review URL: https://codereview.chromium.org/1291703002

Cr-Commit-Position: refs/heads/master@{#343277}