Don't update the active cue set in detached Documents
After (or during) detaching the document, it makes very little sense to
keep updating the timeline and the set of active cues. Instead just punt
and ignore. Most importantly, this avoids trying to dispatch events to
Nodes (HTMLTrackElements precisely) during the disposal of the Document.
Remove the similar check in CueTimeline::endIgnoringUpdateRequests,
that was originally added for WK bug 105606 [1] which fixed a similar
issue, since the check in updateActiveCues should cover the relevant
part of that too.

[1] https://bugs.webkit.org/show_bug.cgi?id=105606

BUG=516298

Review URL: https://codereview.chromium.org/1282903003

git-svn-id: svn://svn.chromium.org/blink/trunk@200507 bbb929c8-8fbe-4397-9dbb-9b2b20218538