Reland: Enable ServiceWorkerPerformanceTimeline flag for stable
This ships User Timing and Resource Timing in Workers, and workerStart
attribute of Resource Timing.

This is a reland of r197722, which is reverted in r199493 to avoid
shipping prefixed Resource Timing APIs. That has been addressed in
r200301; now only unprefixed version of the APIs are available in worker
contexts.

Intent to Implement and Ship:
https://groups.google.com/a/chromium.org/forum/#!topic/blink-dev/htsW078UcFA

BUG=465640,465641,465643

Review URL: https://codereview.chromium.org/1289803002

git-svn-id: svn://svn.chromium.org/blink/trunk@200524 bbb929c8-8fbe-4397-9dbb-9b2b20218538