jni_generator: Wrap all native methods in stubs.
Instead of only wrapping native methods which call C++ methods, wrap
them all. This doesn't require any changes to the implementations, but
makes the handling of the two kinds of method more consistent in
preparation for introducing return type/parameter wrapping in future
changes.

This also changes the way that some of the stub code is generated to
make it less different depending whether native exports are in use, and
moves "extern "C"" declarations to be on only the specifically exported
symbols instead of applying to all forward declarations.

BUG=519562
R=rmcilroy@chromium.org

Review URL: https://codereview.chromium.org/1279163006 .

Cr-Commit-Position: refs/heads/master@{#343370}