Zero-height content should set a baseline on a table row
Per CSS "The table cell's 'height' property can influence the height of the row,
but it does not increase the height of the cell box." This means that we can't allow
height computed from the cell's style to set a baseline. We clarified this with
www-style: https://lists.w3.org/Archives/Public/www-style/2015Jul/0478.html.

This means a couple of tests we have are wrong, including the CSS test inline-table-valign-002.htm
which relies on the incorrect assumption that styled height on the cell contributes
to the height of the cell.

Our behaviour is now consistent with IE11 and Presto, but not FF - but dbaron
acknowledges in www-style that is awaiting a fix.

BUG=513319

TBR=mstensho@opera.com

Review URL: https://codereview.chromium.org/1258573002

(cherry picked from commit 392f48ee2c445a9e9093b2467e53fc69f09ddef3)

Review URL: https://codereview.chromium.org/1289353002 .

git-svn-id: svn://svn.chromium.org/blink/branches/chromium/2454@200549 bbb929c8-8fbe-4397-9dbb-9b2b20218538