Introduce WebPageImportanceSignals to judge state preservation needs
This CL introduces WebPageImportanceSignals, which is measured per Page.
We plan to propagate this signal to content/ so that Chromium can judge
the importance of the tab. Chromium can then use this signal to choose which
actions to take against the tab when we are approaching system resource limit.
For example:
- Simply discard the tab if user hadn't done anything with the tab.
- If he/she entered something on a form, we better preserve its state, etc.

To start with, we record if user has interacted with a HTML form on the page.

BUG=520838

Review URL: https://codereview.chromium.org/1296533003

git-svn-id: svn://svn.chromium.org/blink/trunk@200630 bbb929c8-8fbe-4397-9dbb-9b2b20218538