Merge 200246 "Make LoaderHolder ActiveDOMObject and cancel Fetch..."
> Make LoaderHolder ActiveDOMObject and cancel FetchDataLoader in stop()
>
> FetchDataLoader continue working after ExecutionContext is stopped and thus
> the owner/client of FetchDataLoader should take care of ExecutionContext's
> stop.
> This CL makes LoaderHolder, the only class in non-test code that calls
> FetchDataLoader::start(), to cancel FetchDataLoader when ExecutionContext
> is stopped.
>
> BUG=518154
>
> Review URL: https://codereview.chromium.org/1273773003

Due to refactoring by https://codereview.chromium.org/1233573002, the original
CL cannot be directly applicable to M-45 branch.

This CL renames ClientWithFinishNotification in BodyStreamBuffer.cpp to
LoaderHolder, and makes it to hold FetchDataLoader, to be ActiveDOMObject, and
to cancel FetchDataLoader when ExecutionContext is stopped.
LoaderHolder in this CL is now the only class in non-test code that calls
FetchDataLoader::start() and corresponds to BodyStreamBuffer::LoaderHolder in
the original CL.

This requires ExecutionContext in startLoading(), so the other changes supplies
ExecutionContext* as an argument of startLoading().
This CL also cleanup didFetchDataLoadFinishedFromDrainingStream() declarations.

R=nhiroki@chromium.org, yhirano@chromium.org

Review URL: https://codereview.chromium.org/1285833002

git-svn-id: svn://svn.chromium.org/blink/branches/chromium/2454@200631 bbb929c8-8fbe-4397-9dbb-9b2b20218538