Add GenericSharedMemoryId and use w/ GpuMemoryBuffer
Adds a GenericSharedMemoryId type which will be used to track various types of shared memory (SharedBitmap, IOSurface, base::SharedMemory).

Wire up GpuMemoryBufferManager so that GpuMemoryBufferId is just a typedef of this type, setting up for further integration in the future.

Note regarding message changes - We needed a way for all GenericSharedMemoryIds used by a process to be unique. Currently a renderer process sometimes internally allocates shared memory and sometimes asks the browser process to do so. In order to allow for consistent naming, we cause the Renderer process to provide an ID to the browser when asking the browser to allocate on its behalf.

+tsepez for message changes.
+piman for content/gpu changes
+danakj/reveman for other changes.

TBR=sky@chromium.org
BUG=512534
CQ_INCLUDE_TRYBOTS=tryserver.blink:linux_blink_rel

Review URL: https://codereview.chromium.org/1280513002

Cr-Commit-Position: refs/heads/master@{#343677}