[Extension ServiceWorkers] Blink: Passing v8::context to extensions dispatcher
This patch is to address a reverted patch (https://codereview.chromium.org/1264913002/).  At line 253 in WorkerThread.cpp, the original check was script()->scriptState()->contextIsValid() which can crash if the scriptState is null.  This patch was reverted and is now changed to script()->isContextInitialized().  isContextInitialized was also changed from a private to a public function.

BUG=501569

Review URL: https://codereview.chromium.org/1287103002

git-svn-id: svn://svn.chromium.org/blink/trunk@200696 bbb929c8-8fbe-4397-9dbb-9b2b20218538