cc: Ensure to keep the same low res tiling when activating.
Recent change caused a regression that would cause us to delete and
recreate the low res tilings for no reason. This happens because we
don't have a pending tree low res tiling, so we need to ensure that
we don't delete the active one when we can't match it to thw twin.

R=danakj, enne
BUG=520689
CQ_INCLUDE_TRYBOTS=tryserver.blink:linux_blink_rel

Review URL: https://codereview.chromium.org/1297793003

Cr-Commit-Position: refs/heads/master@{#343971}