Mac Overlays: Clean up animation timing
This fixes a few issues.

Ordinarily, we wait for content to finish processing on the GPU
before acknowledging a swap. If we get a call to discard the
backbuffer (which happens when closing or switching from a tab),
immediately acknowledge the swap.

Use a base::Timer to flush content to the screen. This can be
cancelled without waking up the run loop (this should fix some of
the regressions in idle wakeups).

If content has not finished processing on the GPU when the next
swap call comes in, wait for it to finish before processing the next
swap. This ensures more smooth animation, and makes disable-vsync
behave more appropriately.

BUG=520156

Review URL: https://codereview.chromium.org/1302433002

Cr-Commit-Position: refs/heads/master@{#344162}