Tolerate nonexistent scrollbar in LayoutBox::intrinsicScrollbarLogicalWidth.
DeprecatedPaintLayerScrollableArea will not create an overlay scrollbar if there
is no overflow on that axis, even if overflow:scroll is specified.  This is
intentional behavior added in http://crbug.com/415031.

It's correct for intrinsicScrollbarLogicalWidth to return 0 in this case, as it
already does for all overlay scrollbars.

BUG=520662
TEST=WebViewTest.GeolocationAPIEmbedderHasAccessAllow passes on Mac

Review URL: https://codereview.chromium.org/1292003007

git-svn-id: svn://svn.chromium.org/blink/trunk@200911 bbb929c8-8fbe-4397-9dbb-9b2b20218538