Use EventDispatcher instead of InputMethod to dispatch the simulated key events.
This reverts the changes of InputMethodEngine in the cl
https://codereview.chromium.org/1155013005.
Later in the cl https://codereview.chromium.org/1209663002, the
InputMethod became an EventHandler which is not the first guy to handle
the key events. So dispatching simulated key events to InputMethod
directly will skip the key event handling before InputMethod, such as
accelarator handler.

BUG=522782,521434
TEST=Verified the bugs not repro on local sandbox.

Review URL: https://codereview.chromium.org/1298143006

Cr-Commit-Position: refs/heads/master@{#344649}