Load tab metadata in AsyncTask.
When the DocumentTabModel is created, it loads tab state
from open tabs as well as saved tab metadata from the
previous Chrome instance. The saved tab metadata file is
necessary to know whether existing tabs have a navigation
history, in which case they cannot be retargeted when
home screen shortcuts are used.

This resulted in reading the metadata file on the UI
thread. This CL loads the metadata with an AsyncTask
instead, launched when originally the file was read
directly. This means that when users clicked on their
home screen shortcut and Chrome was not already open,
they would open a new tab instead of retargeting one
of the saved tabs since the metadata would not yet be
available.

The retargeting failure in this rare cold-start scenario
was deemed acceptable in favour of removing UI thread I/O.

BUG=496348

Review URL: https://codereview.chromium.org/1296673006

Cr-Commit-Position: refs/heads/master@{#345073}