Use top frame's origin when checking content settings for plugins.
When creating plugins, the renderer currently sends an IPC to the
browser process and passes the top frame's URL, to be used in a
content settings check on the IO thread.  However, with out-of-process
iframes, this doesn't work if the top frame is a RemoteFrame, which
has no document or URL available.  This CL changes the content
settings check to use the top frame's origin instead, which will use
replicated origins for RemoteFrames.

For discussion about this change, see https://groups.google.com/a/chromium.org/d/topic/site-isolation-dev/JPKPFfZpVHE/discussion

BUG=426658, 466297

Review URL: https://codereview.chromium.org/1168913002

Cr-Commit-Position: refs/heads/master@{#345104}