Fix Linux HID service and connection FileThreadHelper lifetimes.
Both HidServiceLinux and HidConnectionLinux need a helper class that is
owned by the FILE thread so that they can perform possibly blocking I/O
operations.

At some point HidServiceLinux::FileThreadHelper's lifetime became both
managed by the lifetime of the FILE thread message loop (as it was a
MessageLoop::DestructionObserver) and manually managed by the
HidServiceLinux that created it. Neither mechanism actually worked
because a) the helper was never actually added as a message loop
destruction observer and b) the |helper_| field of the HidServiceLinux
was never populated, so the helper would just leak.

The ideal solution is that these helpers must both managed by the
lifetime of the object they support, and also observe the lifetime of
the message loop on which they run.

Review URL: https://codereview.chromium.org/1304073005

Cr-Commit-Position: refs/heads/master@{#345388}