cc: Make SingleThreadProxy CommitComplete ordering match ThreadProxy
In ThreadProxy the order of operations is (when blocking main thead on
activation):
- CommitComplete on impl thread
- Activate
- CommitComplete on main thread

In SingleThreadProxy the order is:
- Activate
- CommitComplete on impl thread
- CommitComplete on main thread

This means that ordering is weird, both in tests and in production. I'd
like to move Animate() to CommitComplete but it happens at different
times so things get weird.

This changes the ordering to match. One thing this turned up is that we
need to activate animations during the commit since they are on the
active tree not the pending tree. And if we don't activate them, they
get ignored during UpdateDrawProperties as they appear to be on a
different tree.

This further turned up that currently when we activate an animation we
don't UpdateDrawProperties on the active tree afterward, leaving draw
properties incorrect for anything the animation would change for a
single frame.

R=ajuma, enne
BUG=522658
CQ_INCLUDE_TRYBOTS=tryserver.blink:linux_blink_rel

Review URL: https://codereview.chromium.org/1308393004

Cr-Commit-Position: refs/heads/master@{#345446}