mac: Remove fullscreen transition hack and associated logic bug.
The hack was introduced to fix the fullscreen transition back when the tab strip
and content view were sibling views. The hack is no longer necessary, since the
tab strip is not a subview of the content view, so I'm removing all the
associated logic.

There was a bug in the logic that caused the window's contentView to lose its
backing layer.

BUG=523250

Review URL: https://codereview.chromium.org/1311673005

Cr-Commit-Position: refs/heads/master@{#345463}