sandbox/linux/bpf_dsl: add golden files for compiled test policies
In preparation for removing the ad-hoc verifier, this CL adds a basic
golden-files regression framework for the bpf_dsl unit tests.  Policies
are compiled to byte code and expected to exactly match the pre-compiled
copies.  This means future changes to the policies or compiler may
require regenerating golden files, but that's not expected to be very
frequent.

The golden files are currently only for i386 and x86-64, but as most
of bpf_dsl is agnostic to the target platform except for word size,
that should suffice for now.  Followup CL will add golden files
for ARM.

BUG=414363

Review URL: https://codereview.chromium.org/1306723002

Cr-Commit-Position: refs/heads/master@{#345508}