Change pointer type of WebLocalFrameImpl from raw pointers to RefPtrWillBePersistent in WebEmbeddedWorkerImpl and WebSharedWorkerImpl.
Those member variables are instantiated with WebLocalFrame::create(),
and it actually creates a WebLocalFrameImpl instance.

On Oilpan build, WebLocalFrameImpl has a self-referential persistent to
keep itself alive, and it will be cleared on call of
WebLocalFrameImpl::close().

In those two classes, m_mainFrame->close() is called in destructors, so
the self referential persistent is cleared in destruction on Oilpan build.
Beside it, on both non-Oilpan build and Oilpan build, m_mainFrame is
also cleared and should be released.


BUG=509911

Review URL: https://codereview.chromium.org/1308313002

git-svn-id: svn://svn.chromium.org/blink/trunk@201191 bbb929c8-8fbe-4397-9dbb-9b2b20218538