Get ReadPixels to work with INT and UNSIGNED_INT format.
In order to get it to work and test it, glClearBuffer*() has to be strengthened.

The remaining issue for ReadPixles is to return reasonable implementation
specific format/type. Currently it always return RGBA and UNSIGNED_BYTE,
which won't work for INT and UNSIGNED_INT framebuffer.

With this CL, read-pixels-from-fbo-test.html finally work (minus the
implementation specific format/type).

BUG=429053
TEST=conformance2/reading/read-pixels-from-fbo-test.html
R=sievers@chromium.org

Review URL: https://codereview.chromium.org/1309333003

Cr-Commit-Position: refs/heads/master@{#345740}