[M46 Merge] Re-add camera and microphone settings to the extensions API.
The reason for removal of camera and microphone content settings from the extension API was that the UI could not correctly show this support.

As of https://codereview.chromium.org/1210173012/, we have a suitable UI, so this can be relanded.

In this CL, we first revert the removal CL (https://codereview.chromium.org/938513004/) in Patchset #1. In the following patchsets, we handle one change that has been made to content settings in the meantime, and that is the possibility to restrict the default content setting to a smaller subset of values than is supported for content setting exceptions of the same content type. This currently only happens in case of camera and microphone, so it was not necessary to handle it in the extensions' API until now.

==================================================

Revert details:

Original issue's description:
> Remove camera and microphone settings from the extensions API.
>
> These were added in CLs 880223002 and 889893002.
> Automatic revert in CLs 925303003 and 933273002 didn't work due to the way how kContentSettingsTypeNames in content_settings_helpers.cc works - we cannot remove an entry from the middle of the list.
>
> Reason for revert:
> Setting the camera and microphone default content settings currently cannot be shown in the UI and will not work exactly the way an app developer might expect it to (due to interference with the MEDIASTREAM content setting).
>
> Reverting this patch until the UI is ready.
>
> BUG=464410
>
> Committed: https://crrev.com/16991f7d102f1c2327587258eabd5e46fc362ec0
> Cr-Commit-Position: refs/heads/master@{#319286}

==================================================

TBR=mkwst@chromium.org,kalman@chromium.org,markusheintz@chromium.org
BUG=464410,455589

Review URL: https://codereview.chromium.org/1296983005

Cr-Commit-Position: refs/heads/master@{#345223}
(cherry picked from commit 270c118e934ce24372117e0b7fba55c6130f1012)

Review URL: https://codereview.chromium.org/1320623003 .

Cr-Commit-Position: refs/branch-heads/2490@{#40}
Cr-Branched-From: 7790a3535f2a81a03685eca31a32cf69ae0c114f-refs/heads/master@{#344925}