[AndroidWebViewShell] Make testGlobalInterface non-breaking
The testGlobalInterface layout test appears to need a rebaseline
relatively frequently. Since the bot running this test is on
chromium.android now it is not feasible for the sheriffs to expect
to fix this test every time it breaks. This patch makes this test
run in quiet mode (i.e. it always passes) but it does leave a trace
in logcat so that we can still track any changes to the interfaces
in webview.

BUG=523245,514217

Review URL: https://codereview.chromium.org/1317673002

Cr-Commit-Position: refs/heads/master@{#345886}