//components/variations: Create separate target for VariationsService
When VariationsService is componentized, it will bring in dependencies on //net
and //ui, which are unwanted in the //components/variations target. In
particular, Cronet depends on //components/variations and cannot depend on
//net.

This CL creates //components/variations/service, which currently holds
variations_service_client.h and will hold variations_service.* once it is
componentized.

BUG=516678
TBR=brettw

Review URL: https://codereview.chromium.org/1306893007

Cr-Commit-Position: refs/heads/master@{#345905}