Don't allow whitespace between elements with display:table-cell
In https://codereview.chromium.org/1065563005 we started allowing whitespace
inside anonymous table cells. We're only interested in anonymous table cells
that were generated to contain non-table-part content. If the anonymous cell
was generated to contain a form element with display:table-cell, for example, then
that is not a situation where we want to treat any whitespace before and
after the form element as renderable.
 
BUG=515771

Review URL: https://codereview.chromium.org/1280123004

git-svn-id: svn://svn.chromium.org/blink/trunk@201329 bbb929c8-8fbe-4397-9dbb-9b2b20218538