Revert of Turn VariationsService::GetVariationsServerURL into instance method (patchset #2 id:20001 of https://codereview.chromium.org/1312423003/ )
Reason for revert:
It turns out that this change is actually not safe as written: g_browser_process->variations_service() returns null in non-official builds unless a default variations server URL is set on the command line. The reason the patch did not fail is because the method being called was (a) not virtual and (b) didn't use any state of the object.

Original issue's description:
> Turn VariationsService::GetVariationsServerURL into instance method
>
> This will allow for followup work that uses |client_| within that
> method.
>
> BUG=516678
> TBR=bartfab
>
> Committed: https://crrev.com/642578a171c7acee3de6b514bf2738fcb0339aac
> Cr-Commit-Position: refs/heads/master@{#346111}

TBR=asvitkine@chromium.org,bartfab@chromium.org
NOPRESUBMIT=true
NOTREECHECKS=true
NOTRY=true
BUG=516678

Review URL: https://codereview.chromium.org/1320503005

Cr-Commit-Position: refs/heads/master@{#346141}