Speculative fix for recording texture-backed SkImages into SkPictures.
SkImages that are backed by textures are being recorded into SkPictures that are rendered in multiple threads. This causes multiple threads to access the same Gr- and GL context, neither of which are thread safe.

This change was suggested by junov@chromium.org as a possible fix.

BUG=524717

Review URL: https://codereview.chromium.org/1310573007

git-svn-id: svn://svn.chromium.org/blink/trunk@201410 bbb929c8-8fbe-4397-9dbb-9b2b20218538