Add BubbleManager to manage bubbles and ChromeBubbleManager for events.
BubbleManager - Responsible for showing bubbles.

BubbleDelegate - Should be implemented to provide the definition of a
specific bubble.

BubbleUI - This should be implemented for views and cocoa to provide a
bubble's UI.

BubbleController - Responsible for owning the BubbleUI and the
BubbleDelegate for a specific bubble.

ChromeBubbleManager - Knows about Chrome browser specifics to hook into
the right events.

BUG=496955

Review URL: https://codereview.chromium.org/1251633002

Cr-Commit-Position: refs/heads/master@{#346195}