cc: Turn LayerImpl's damage_rect into a gfx::Rect.
Currently it's a float rect so that if you're at some high scale, and
a tile is replaced, we only damage the exact physical pixels.

This would make that a bit more sloppy, so that we'd damage in layer
space. This should be inconsequential however in terms of cost of the
damage rect, and saves us some int->float->int converting every
time a tile is rastered, and some code complexity.

R=enne, vmpstr
BUG=342848
CQ_INCLUDE_TRYBOTS=tryserver.blink:linux_blink_rel

Review URL: https://codereview.chromium.org/1320163003

Cr-Commit-Position: refs/heads/master@{#346209}