Always set m_subtreeIsCompositing when the current layer is composited.
Previously, we only set m_subtreeIsCompositing (and pass it back up) when a layer
was directly composited. This would work only as long as there is some layer down
in the hierarchy was directly composited. The attached testcase is an example where
no layer is directly composited, only indirectly.

BUG=505873

Review URL: https://codereview.chromium.org/1311673010

git-svn-id: svn://svn.chromium.org/blink/trunk@201435 bbb929c8-8fbe-4397-9dbb-9b2b20218538