MacViews: Round corners for the Yosemite menu background paint for blur
For the blurred background to take effect, the non-layered parts of an
NSWindow can not be fully transparent. r345820 achieved this with a
window background color, but that gets square corners that mess with the
window shadow.

To fix, ask the content view to draw the almost-transparent background
after adding a rounded rectangle clip.

BUG=522365

Review URL: https://codereview.chromium.org/1326453002

Cr-Commit-Position: refs/heads/master@{#346342}