[Cronet] Make UrlRequestContextConfig.enableLegacyMode() public again
This method was previously made pacakage-private so it would be
hidden from the javadocs.  It only applies to the deprecated API.
It needs to be public for users of the deprecated API however.
It has been marked @deprecated which appears to hide it from
Cronet's javadocs.  We should keep in mind that not all javadoc
generation hides @deprecated APIs.

Review URL: https://codereview.chromium.org/1309813012

Cr-Commit-Position: refs/heads/master@{#346662}